Data: LeDang2022

A data frame of 107 variables and 773 participants. Labelled factors were converted from numerical to character representation through haven::as_factor(). The provided data frame is a fusion of the converted labelled factors for the demographic data and the unconverted Likert scale data for the psychometric data.

Usage

LeDang2022

Format

An object of class data.frame with 773 rows and 107 columns.

Source

The data were made available in Le and Dang (2022) .

Details

This data is based on self-report surveys collected from graduates of universities in Northern Vietnam (Le and Dang 2022) . The data has been used to analyze the relationship between measures of HEXACO personality traits and two job-search outcomes: (1) number of job interviews; and (2) job offers through job search networking behavior (Le and Dang 2022) .

The data were also analyzed using Integrated-Generalized Structured Component Analysis (IGSCA) in GSCAPro (Hwang et al. 2021) .

References

Hwang H, Cho G, Jung K, Falk C, Flake J, Jin MJ, Lee SH (2021). “An approach to structural equation modeling with both factors and components: Integrated generalized structured component analysis.” Psychological Methods, 26(3), 273–294. doi:10.1037/met0000336 .

Le S, Dang C (2022). “Data for the link between HEXACO personality traits and job search outcomes in Vietnam.” Data in Brief, 45, 108677. doi:10.1016/j.dib.2022.108677 .