Internal: Complete list of assess()'s ... arguments

A complete alphabetical list of all possible arguments accepted by assess()'s ... (dotdotdot) argument.

Arguments

.absolute
Logical. Should the absolute HTMT values be returned? Defaults to TRUE .

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.closed_form_ci
Logical. Should a closed-form confidence interval be computed? Defaults to FALSE.

.null_model
Logical. Should the degrees of freedom for the null model be computed? Defaults to FALSE.

.saturated
Logical. Should a saturated structural model be used? Defaults to FALSE.

.type_vcv
Character string. Which model-implied correlation matrix is calculated? One of "indicator" or "construct". Defaults to "indicator".

Details

Most arguments supplied to the ... argument of assess() are only accepted by a subset of the functions called by assess(). The following list shows which argument is passed to which (internal) function:

.absolute

Accepted by/Passed down to: calculateHTMT()

.alpha

Accepted by/Passed down to: calculateRhoT()

.closed_form_ci

Accepted by/Passed down to: calculateRhoT()

.null_model

Accepted by/Passed down to: calculateDf()

.saturated

Accepted by/Passed down to: calculateSRMR(), calculateDG(), calculateDL(), calculateDML()and subsequently fit().

.type_vcv

Accepted by/Passed down to: calculateSRMR(), calculateDG(), calculateDL(), calculateDML() and subsequently fit().