Internal: Complete list of csem()'s ... arguments

A complete alphabetical list of all possible arguments accepted by csem()'s ... (dotdotdot) argument.

Arguments

.approach_cor_robust
Character string. Approach used to obtain a robust indicator correlation matrix. One of: "none" in which case the standard Bravais-Person correlation is used, "spearman" for the Spearman rank correlation, or "mcd" via MASS::cov.rob() for a robust correlation matrix. Defaults to "none". Note that many postestimation procedures (such as testOMF() or fit() implicitly assume a continuous indicator correlation matrix (e.g. Bravais-Pearson correlation matrix). Only use if you know what you are doing.

.conv_criterion
Character string. The criterion to use for the convergence check. One of: "diff_absolute", "diff_squared", or "diff_relative". Defaults to "diff_absolute".

.dominant_indicators
A character vector of "construct_name" = "indicator_name" pairs, where "indicator_name" is a character string giving the name of the dominant indicator and "construct_name" a character string of the corresponding construct name. Dominant indicators may be specified for a subset of the constructs. Default to NULL.

.estimate_structural
Logical. Should the structural coefficients be estimated? Defaults to TRUE.

.iter_max
Integer. The maximum number of iterations allowed. If iter_max = 1 and .approach_weights = "PLS-PM" one-step weights are returned. If the algorithm exceeds the specified number, weights of iteration step .iter_max - 1 will be returned with a warning. Defaults to 100.

.PLS_modes
Either a named list specifying the mode that should be used for each construct in the form "construct_name" = mode, a single character string giving the mode that should be used for all constructs, or NULL. Possible choices for mode are: "modeA", "modeB", "modeBNNLS", "unit", "PCA", a single integer or a vector of fixed weights of the same length as there are indicators for the construct given by "construct_name". If only a single number is provided this is identical to using unit weights, as weights are rescaled such that the related composite has unit variance. Defaults to NULL. If NULL the appropriate mode according to the type of construct used is chosen. Ignored if .approach_weight is not PLS-PM.

.PLS_ignore_structural_model
Logical. Should the structural model be ignored when calculating the inner weights of the PLS-PM algorithm? Defaults to FALSE. Ignored if .approach_weights is not PLS-PM.

.PLS_weight_scheme_inner
Character string. The inner weighting scheme used by PLS-PM. One of: "centroid", "factorial", or "path". Defaults to "path". Ignored if .approach_weight is not PLS-PM.

.PLS_approach_cf
Character string. Approach used to obtain the correction factors for PLSc. One of: "dist_squared_euclid", "dist_euclid_weighted", "fisher_transformed", "mean_arithmetic", "mean_geometric", "mean_harmonic", "geo_of_harmonic". Defaults to "dist_squared_euclid". Ignored if .disattenuate = FALSE or if .approach_weights is not PLS-PM.

.tolerance
Double. The tolerance criterion for convergence. Defaults to 1e-05.

Details

Most arguments supplied to the ... argument of csem() are only accepted by a subset of the functions called by csem(). The following list shows which argument is passed to which (internal) function:

.approach_cor_robust

Accepted by/Passed down to: calculateIndicatorCor()

.conv_criterion

Accepted by/Passed down to: calculateWeightsPLS(), calculateWeightsGSCA(), calculateWeightsGSCAm() and subsequently checkConvergence().

.dominant_indicators

Accepted by/Passed down to: setDominantIndicator()

.estimate_structural

Accepted by/Passed down to: foreman()

.iter_max

Accepted by/Passed down to: calculateWeightsPLS(), calculateWeightsGSCA(), calculateWeightsGSCAm()

.PLS_modes, .PLS_ignore_structural_model, .PLS_weight_scheme_inner, .PLS_approach_cf

Accepted by/Passed down to: calculateWeightsPLS()

.tolerance

Accepted by/Passed down to: calculateWeightsPLS(), calculateWeightsGSCA(), calculateWeightsGSCAm(), calculateWeightsUnit()