cSEM: A package for composite-based structural equation modeling

Estimate, analyze, test, and study linear, nonlinear, hierarchical and multigroup structural equation models using composite-based approaches and procedures including estimation techniques such as partial least squares path modeling (PLS) and its derivatives (PLSc, ordPLSc, robustPLSc), generalized structured component analysis (GSCA), generalized structured component analysis with uniqueness terms (GSCAm), generalized canonical correlation analysis (GCCA) unit weights (sum score) and fixed weights, as well as several tests and typical postestimation procedures (e.g., assess the model fit, compute direct, indirect and total effects).

See also

Author

Maintainer: Florian Schuberth f.schuberth@utwente.nl (ORCID)

Authors:

Manuel E. Rademaker manuel-rademaker@outlook.de (ORCID)

Other contributors:

Tamara Schamberger tamara.schamberger@uni-wuerzburg.de (ORCID) [contributor]

Michael Klesel (ORCID) [contributor]

Huu Phuc Nguyen (ORCID) [contributor]

Theo K. Dijkstra [contributor]

Jörg Henseler (ORCID) [contributor]