HTMT

Computes either the heterotrait-monotrait ratio of correlations (HTMT) based on Henseler et al. (2015) or its advancement HTMT2. While the HTMT is a consistent estimator for the construct correlation in case of tau-equivalent measurement models, the HTMT2 is a consistent estimator for congeneric measurement models. In general, they are used to assess discriminant validity.

calculateHTMT(
 .object               = NULL,
 .type_htmt            = c('htmt','htmt2'),
 .absolute             = TRUE,
 .alpha                = 0.05,
 .ci                   = c("CI_percentile", "CI_standard_z", "CI_standard_t", 
                           "CI_basic", "CI_bc", "CI_bca", "CI_t_interval"),
 .handle_inadmissibles = c("drop", "ignore", "replace"),
 .inference            = FALSE,
 .only_common_factors  = TRUE,
 .R                    = 499,
 .seed                 = NULL,
 ...
)

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.type_htmt
Character string indicating the type of HTMT that should be calculated, i.e., the original HTMT ("htmt") or the HTMT2 ("htmt2"). Defaults to "htmt"

.absolute
Logical. Should the absolute HTMT values be returned? Defaults to TRUE .

.alpha
A numeric value giving the significance level. Defaults to 0.05.

.ci
A character strings naming the type of confidence interval to use to compute the 1-alpha% quantile of the bootstrap HTMT values. For possible choices see infer(). Ignored if .inference = FALSE. Defaults to "CI_percentile".

.handle_inadmissibles
Character string. How should inadmissible results be treated? One of "drop", "ignore", or "replace". If "drop", all replications/resamples yielding an inadmissible result will be dropped (i.e. the number of results returned will potentially be less than .R). For "ignore" all results are returned even if all or some of the replications yielded inadmissible results (i.e. number of results returned is equal to .R). For "replace" resampling continues until there are exactly .R admissible solutions. Depending on the frequency of inadmissible solutions this may significantly increase computing time. Defaults to "drop".

.inference
Logical. Should critical values be computed? Defaults to FALSE.

.only_common_factors
Logical. Should only concepts modeled as common factors be included when calculating one of the following quality critera: AVE, the Fornell-Larcker criterion, HTMT, and all reliability estimates. Defaults to TRUE.

.R
Integer. The number of bootstrap replications. Defaults to 499.

.seed
Integer or NULL. The random seed to use. Defaults to NULL in which case an arbitrary seed is chosen. Note that the scope of the seed is limited to the body of the function it is used in. Hence, the global seed will not be altered!

...
Ignored.

Value

A lower tringular matrix of HTMT values. If .inference = TRUE the upper tringular part is the 1-.alpha%-quantile of the HTMT's bootstrap distribution.

Details

Computation of the HTMT assumes that all intra-block and inter-block correlations between indicators are either all-positive or all-negative. A warning is given if this is not the case. If all intra-block or inter-block correlations are negative the absolute HTMT values are returned (.absolute = TRUE).

To obtain the 1-alpha%-quantile of the bootstrap distribution for each HTMT value set .inference = TRUE. To choose the type of confidence interval to use to compute the 1-alpha%-quantile, use .ci. To control the bootstrap process, arguments .handle_inadmissibles, .R and .seed are available.

Since the HTMT is defined with respect to a classical true score measurement model only concepts modeled as common factors are considered by default. For concepts modeled as composites the HTMT may be computed by setting .only_common_factors = FALSE, however, it is unclear how to interpret values in this case.

References

Henseler J, Ringle CM, Sarstedt M (2015). “A New Criterion for Assessing Discriminant Validity in Variance-based Structural Equation Modeling.” Journal of the Academy of Marketing Science, 43(1), 115--135. doi: 10.1007/s11747-014-0403-8 , https://doi.org/10.1007/s11747-014-0403-8.

See also