Data: corp_rep_data

A data frame with 344 observations and 41 variables. It was used in Hair et al. (2017) to demonstrate PLS-PM.

Usage

corp_rep_data

Format

An object of class data.frame with 344 rows and 41 columns.

Source

The seminr package (version 2.4.0).

Details

The dataset contains 344 observations on the following 41 variables:

serviceprovider

A categorical variable for the service provider: 1, 2, 3, or 4.

servicetype

A categorical variable for the service type: 1=Prepaid plan (n=125); 2=Contract plan (n=219).

csor_1

The company behaves in a socially conscious way.

csor_2

The company is forthright in giving information to the public.

csor_3

The company has a fair attitude toward competitors.

csor_4

The company is concerned about the preservation of the environment.

csor_5

The company is not only concerned about the profits.

csor_global

Please assess the extent to which the company acts in socially conscious ways 0 (not at all) to 7 (definitely).

attr_1

The company is succesful in attracting high-quality employees.

attr_2

I could see myself working at the company.

attr_3

I like the physical appearance of the company/buildings/shops, etc.

attr_global

Please assess the company’s overall attractiveness; 0=very low; 7=very high.

perf_1

The company is a very well managed company.

perf_2

The company is an economically stable company.

perf_3

The business risk for the company is modest compared to its competitors.

perf_4

The company has growth potential.

perf_5

The company has a clear vision about the future of the company.

perf_global

Please assess the company’s overall performance; 0=very low; 7=very high.

qual_1

The products/services offered by the company are of high quality.

qual_2

The company is an innovator, rather than an imitator with respect to industry.

qual_3

The company's services/products offer good quality for money.

qual_4

The services the company offered are good.

qual_5

Customer concerns are held in high regard at the company.

qual_6

The company is a reliable partner for customers.

qual_7

The company is a trustworthy company.

qual_8

I have a lot of respect for the company.

qual_global

Please assess the overall quality of the company’s activities; 0=very low; 7=very high.

like_1

The company is a copany that I can better identify with than other companies.

like_2

The company is a company that I would regret more not having if it no longer existed than other companies.

like_3

I regard the company as a likeable company.

comp_1

The company is a top competitor in its market.

comp_2

As far as I know, the company is recognized worldwide.

comp_3

I believe that the company performs at a premium level.

cusl_1

I would recommend the company to friends and relatives.

cusl_2

If I had to choose again, I would choose the company as my mobile phone services provider.

cusl_3

I will remain a customer of the company in the future.

cusa

I am satisfied with company.

switch_1

It takes me a great deal of time to switch to another company.

switch_2

It costs me too much to switch to another company.

switch_3

It takes a lot of effort to get used to a new company with its specific "rules" and practices.

switch_4

In general, it would be a hassle switching to another company.

References

Hair JF, Hult GTM, Ringle CM, Sarstedt M (2017). A Primer on Partial Least Squares Structural Equation Modeling (PLS-SEM), 2nd edition. Sage, London.