Composite-based SEM

Estimate linear, nonlinear, hierachical or multigroup structural equation models using a composite-based approach. In cSEM any method or approach that involves linear compounts (scores/proxies/composites) of observables (indicators/items/manifest variables) is defined as composite-based. See the the cSEM vignette for details.

csem(
  .data                        = NULL,
  .model                       = NULL,
  .approach_nl                 = c("sequential", "replace"),
  .approach_paths              = c("OLS", "2SLS"),
  .approach_weights            = c("PLS-PM", "SUMCORR", "MAXVAR", "SSQCORR", "MINVAR", "GENVAR",
                                   "GSCA", "PCA", "unit", "bartlett", "regression"),
  .disattenuate                = TRUE,
  .id                          = NULL,
  .normality                   = TRUE,
  .reliabilities               = NULL,
  .resample_method             = c("none", "bootstrap", "jackknife"),
  .resample_method2            = c("none", "bootstrap", "jackknife"),
  .R                           = 499,
  .R2                          = 199,
  .handle_inadmissibles        = c("drop", "ignore", "replace"),
  .user_funs                   = NULL,
  .eval_plan                   = c("sequential", "multiprocess"),
  .seed                        = NULL,
  ...
  )

Arguments

.data
A data.frame or a matrix of standardized or unstandarized data (indicators/items/manifest variables). Additionally, a list of data.frame(s) or matrice(s) is accepted in which case estimation is repeated for each data set. Possible column types or classes of the data provided are: "logical", "numeric" ("double" or "integer"), "factor" ("ordered" and/or "unordered") or a mix of several types. The data may also include one character column whose column name must be given to .id. This column is assumed to contain group identifiers used to split the data into groups.

.model
A model in lavaan model syntax or a cSEMModel list.

.approach_2ndorder
Character string. Approach used for models containing second order constructs. One of: "3stage" or "repeated_indicators". Defaults to "3stage".

.approach_nl
Character string. Approach used to estimate nonlinear structural relationships. One of: "sequential" or "replace". Defaults to "sequential".

.approach_paths
Character string. Approach used to estimate the structural coefficients. One of: "OLS" or "2SLS" (not yet implemented). Defaults to "OLS".

.approach_weights
Character string. Approach used to obtain composite weights. One of: "PLS-PM", "SUMCORR", "MAXVAR", "SSQCORR", "MINVAR", "GENVAR", "GSCA", "PCA", "unit", "bartlett", or "regression". Defaults to "PLS-PM".

.disattenuate
Logical. Should composite/proxy correlations be disattenuated to yield consisten loadings and path estimates if at least one of the construct is modeled as a common factor? Defaults to TRUE.

.id
Character string or integer. The name or position of the column of .data used to split the data into groups. Defaults to NULL.

.normality
Logical. Should joint normality of \([\eta_{1:p}; \zeta; \epsilon]\) be assumed in the nonlinear model? See (Dijkstra and Schermelleh-Engel 2014) for details. Defaults to TRUE. Ignored if the model is linear.

.reliabilities
A character vector of "name" = value pairs, where value is a number between 0 and 1 and "name" a character string of the corresponding construct name, or NULL. Reliabilities may be given for a subset of the constructs. Defaults to NULL in which case reliabilities are estimated by csem(). Currently, only supported for .approach_weights = "PLS-PM".

.resample_method
Character string. The resampling method to use. One of: "bootstrap" or "jackknife". Defaults to "bootstrap".

.resample_method2
Character string. The resampling method to use when resampling from a resample. One of: "none", "bootstrap" or "jackknife". For "bootstrap" the number of draws is provided via .R2. Currently, resampling from each resample is only required for the studentized confidence intervall computed by the infer() function. Defaults to "none".

.R
Integer. The number of bootstrap replications. Defaults to 499.

.R2
Integer. The number of bootstrap replications to use when resampling from a resample. Defaults to 199.

.handle_inadmissibles
Character string. How should inadmissible results be treated? One of "drop", "ignore", or "replace". If "drop", all replications/resamples yielding an inadmissible result will be dropped (i.e. the number of results returned will be less than .R). For "ignore" all results are returned even if they are inadmissible (i.e. number of results returned = .R). For "replace" resampling continues until there are exactly .R admissible solutions. Defaults to "drop".

.user_funs
A function or a (named) list of functions to apply to every resample. Takes .object as an input (e.g., myFun <- function(.object) {...}). Output should preferably be a (named) vector but matrices are also accepted. However, the output will be vectorized (columnwise) in this case. See the examples section for details.

.eval_plan
Character string. The evaluation plan to use. One of "sequential" or "multiprocess". In the latter case all available cores will be used. Defaults to "sequential".

.seed
Integer. The random seed to use. Defaults to NULL in which case an arbitrary seed is choosen.

...
Further arguments to be passed down to lower level functions of csem(). See csem_dotdotdot for a complete list of available arguments.

Value

An object of class cSEMResults with methods for all postestimation generics. Note that, technically, a call to csem() results in an object with at least two class attributes. The first class attribute is always cSEMResults. The second is one of cSEMResults_default, cSEMResults_multi, or cSEMResults_2ndorder and depends on the estimated model and/or the type of data provided to the .model and .data arguments. The third class attribute cSEMResults_resampled is only added if resampling was conducted. Technically, method dispatch for all postestimation functions is based on the second class attribute. For a details see the cSEMResults helpfile .

Details

csem() estimates linear, nonlinear, hierachical or multigroup structural equation models using a composite-based approach.

Data and model:
The .data and .model arguments are required. Data must be provided as either a matrix or a data.frame with column names matching the indicator names used in the model description of the measurement model. Alternatively, a list of matrices or data.frame's may be provided in which case estimation is repeated for each data set. The data provided via .data may contain one character column whose column name must be provided to .id. Values of this column are interpreted as group identifiers and csem() will split the data by levels of that column and run the estimation for each level separately. To provide a model use the lavaan model syntax with two notable extensions/changes. First: the "<~" operator in cSEM is used to define a composite instead of a causal formative common factor. Second: the "." is used to indicate interactions between constructs as in e.g., construct1.construct2. Alternatively a standardized (possibly incomplete) cSEMModel-list may be supplied.
Weights and path coefficients:
By default weights are estimated using the partial least squares (path) algorithm ("PLS-PM"). A broad range of alternative weightning algorithms may be supplied to .appraoch_weights. Currently the following approaches are implemented

(Default) Partial least squares path modeling ("PLS"). The algorithm can be customized. See calculateWeightsPLS() for details.

Generalized structured component analysis ("GSCA")

Generalized canoncial correlation analysis (GCCA), including "SUMCORR", "MAXVAR", "SSQCORR", "MINVAR", "GENVAR"

Principal component analysis ("PCA")

Factor score regression using sum scores ("unit"), regression ("regression") or bartlett scores ("bartlett")

Composite-indicator and composite-composite correlations are properly disattenuated by default to yield consistent loadings, construct correlations, and path coefficients if any of the constructs in the model are modeled as a common factor. For PLS-PM disattenuation is done using PLSc (Dijkstra and Henseler 2015) . For GSCA disattenuation is done implicitly by using GSCAm. Weights obtained by GCCA, unit weights and fixed weights are disattenuated using a Disattenuation my be suppressed by setting .disattenuate = FALSE. Note, however, that quantities in this case are inconsistent estimates for their construct level counterparts if any of the constructs in the structural model is modeled as a common factor!
Nonlinear models:
If the model is nonlinear csem() estimates a polynomial structural equation model using a non-iterative method of moments approach described in (Dijkstra and Schermelleh-Engel 2014) . Non linear terms include interactions and exponential terms. The latter is described in model syntax as an "interaction with itself", e.g., x_1^3 = x1.x1.x1. Currently only exponential terms up to a power of three (i.e. three-way interactions) are allowed. The current version of the package allows two kinds of estimation: estimation of the reduced form equation (.approach_nl = "reduced") and sequential estimation (.approach_nl = "sequential"). The latter does not not allow for multivariate normality of all exogenous variables, i.e., the latent variables and the error terms. Distributional assumptions are kept to a minimum (an i.i.d. sample from a population with finite moments for the relevant order); for higher order models, that go beyond interaction, we work in this version with the assumption that as far as the relevant moments are concerned certain combinations of measurement errors behave as if they were Gaussian. For details see: Dijkstra and Schermelleh-Engel (2014) .
Inference:
Inference is done via resampling. See resamplecSEMResults for details.
References

Dijkstra TK, Henseler J (2015). “Consistent and Asymptotically Normal PLS Estimators for Linear Structural Equations.” Computational Statistics \& Data Analysis, 81, 10--23. Dijkstra TK, Schermelleh-Engel K (2014). “Consistent Partial Least Squares For Nonlinear Structural Equation Models.” Psychometrika, 79(4), 585--604.

See also

Examples

# NOT RUN {
# }