A parseModel extractor function for the purposes of running I-GSCA code example

A parseModel extractor function for the purposes of running I-GSCA code example

A parseModel extractor function for the purposes of running I-GSCA code example

extract_parseModel(model, data, ind_domi_as_first = TRUE)

extract_parseModel(model, data, ind_domi_as_first = TRUE)

Arguments

model

Specified Model in lavaan style

data

Dataframe

ind_domi_as_first

Boolean for whether the first indicator for each latent factor should be chosen as the dominant indicator

Value

Returns a list of matrices required for igsca_sim() to run: z0, W0, B0, C0, lv_type, ov_type, ind_domi.

Returns a list of matrices required for igsca_sim() to run: z0, W0, B0, C0, lv_type, ov_type, ind_domi.

Examples

tutorial_igsca_model <- "
# Composite Model
NetworkingBehavior <~ Behavior1 + Behavior2 + Behavior3 + Behavior5 + Behavior7 + Behavior8 + Behavior9
Numberofjobinterviews <~ Interview1 + Interview2
Numberofjoboffers <~ Offer1 + Offer2 

# Reflective Measurement Model
HonestyHumility =~ Honesty1 + Honesty2 + Honesty3 + Honesty4 + Honesty5 + Honesty6 + Honesty7 + Honesty8 + Honesty9 + Honesty10
Emotionality =~ Emotion1 + Emotion2 + Emotion3 + Emotion4 + Emotion5 + Emotion6 + Emotion8 + Emotion10
Extraversion =~ Extraver2 + Extraver3 + Extraver4 + Extraver5 + Extraver6 + Extraver7 + Extraver8 + Extraver9 + Extraver10
Agreeableness =~ Agreeable1 + Agreeable3 + Agreeable4 + Agreeable5 + Agreeable7 + Agreeable8 + Agreeable9 + Agreeable10
Conscientiousness =~ Conscientious1 + Conscientious3 + Conscientious4 + Conscientious6 + Conscientious7 + Conscientious8 + Conscientious9 + Conscientious10
OpennesstoExperience =~ Openness1 + Openness2 + Openness3 + Openness5 + Openness7 + Openness8 + Openness9 + Openness10

# Structural Model
NetworkingBehavior ~ HonestyHumility + Emotionality + Extraversion + Agreeableness + Conscientiousness + OpennesstoExperience
Numberofjobinterviews ~ NetworkingBehavior
Numberofjoboffers ~ NetworkingBehavior
"

data("LeDang2022")

extract_parseModel(model = tutorial_igsca_model, data = LeDang2022, ind_domi_as_first = TRUE)
#> $z0
#> # A tibble: 773 × 62
#>      Honesty1   Honesty2   Honesty3 Honesty4 Honesty5 Honesty6 Honesty7 Honesty8
#>    <hvn_lbll> <hvn_lbll> <hvn_lbll> <hvn_lb> <hvn_lb> <hvn_lb> <hvn_lb> <hvn_lb>
#>  1          4          4          4        4        4        4        4        4
#>  2          3          3          5        3        5        5        5        5
#>  3          2          2          3        2        3        3        3        3
#>  4          1          1          2        1        2        2        2        2
#>  5          5          5          5        5        5        4        5        5
#>  6          4          4          3        4        3        3        3        3
#>  7          2          2          3        2        3        3        3        3
#>  8          2          2          3        2        3        3        3        3
#>  9          2          2          3        2        3        3        3        3
#> 10          4          4          4        4        4        4        4        4
#> # ℹ 763 more rows
#> # ℹ 54 more variables: Honesty9 <hvn_lbll>, Honesty10 <hvn_lbll>,
#> #   Emotion1 <hvn_lbll>, Emotion2 <hvn_lbll>, Emotion3 <hvn_lbll>,
#> #   Emotion4 <hvn_lbll>, Emotion5 <hvn_lbll>, Emotion6 <hvn_lbll>,
#> #   Emotion8 <hvn_lbll>, Emotion10 <hvn_lbll>, Extraver2 <hvn_lbll>,
#> #   Extraver3 <hvn_lbll>, Extraver4 <hvn_lbll>, Extraver5 <hvn_lbll>,
#> #   Extraver6 <hvn_lbll>, Extraver7 <hvn_lbll>, Extraver8 <hvn_lbll>, …
#> 
#> $B0
#>                       HonestyHumility Emotionality Extraversion Agreeableness
#> HonestyHumility                     0            0            0             0
#> Emotionality                        0            0            0             0
#> Extraversion                        0            0            0             0
#> Agreeableness                       0            0            0             0
#> Conscientiousness                   0            0            0             0
#> OpennesstoExperience                0            0            0             0
#> NetworkingBehavior                  0            0            0             0
#> Numberofjobinterviews               0            0            0             0
#> Numberofjoboffers                   0            0            0             0
#>                       Conscientiousness OpennesstoExperience NetworkingBehavior
#> HonestyHumility                       0                    0                  1
#> Emotionality                          0                    0                  1
#> Extraversion                          0                    0                  1
#> Agreeableness                         0                    0                  1
#> Conscientiousness                     0                    0                  1
#> OpennesstoExperience                  0                    0                  1
#> NetworkingBehavior                    0                    0                  0
#> Numberofjobinterviews                 0                    0                  0
#> Numberofjoboffers                     0                    0                  0
#>                       Numberofjobinterviews Numberofjoboffers
#> HonestyHumility                           0                 0
#> Emotionality                              0                 0
#> Extraversion                              0                 0
#> Agreeableness                             0                 0
#> Conscientiousness                         0                 0
#> OpennesstoExperience                      0                 0
#> NetworkingBehavior                        1                 1
#> Numberofjobinterviews                     0                 0
#> Numberofjoboffers                         0                 0
#> 
#> $W0
#>                 HonestyHumility Emotionality Extraversion Agreeableness
#> Honesty1                      1            0            0             0
#> Honesty2                      1            0            0             0
#> Honesty3                      1            0            0             0
#> Honesty4                      1            0            0             0
#> Honesty5                      1            0            0             0
#> Honesty6                      1            0            0             0
#> Honesty7                      1            0            0             0
#> Honesty8                      1            0            0             0
#> Honesty9                      1            0            0             0
#> Honesty10                     1            0            0             0
#> Emotion1                      0            1            0             0
#> Emotion2                      0            1            0             0
#> Emotion3                      0            1            0             0
#> Emotion4                      0            1            0             0
#> Emotion5                      0            1            0             0
#> Emotion6                      0            1            0             0
#> Emotion8                      0            1            0             0
#> Emotion10                     0            1            0             0
#> Extraver2                     0            0            1             0
#> Extraver3                     0            0            1             0
#> Extraver4                     0            0            1             0
#> Extraver5                     0            0            1             0
#> Extraver6                     0            0            1             0
#> Extraver7                     0            0            1             0
#> Extraver8                     0            0            1             0
#> Extraver9                     0            0            1             0
#> Extraver10                    0            0            1             0
#> Agreeable1                    0            0            0             1
#> Agreeable3                    0            0            0             1
#> Agreeable4                    0            0            0             1
#> Agreeable5                    0            0            0             1
#> Agreeable7                    0            0            0             1
#> Agreeable8                    0            0            0             1
#> Agreeable9                    0            0            0             1
#> Agreeable10                   0            0            0             1
#> Conscientious1                0            0            0             0
#> Conscientious3                0            0            0             0
#> Conscientious4                0            0            0             0
#> Conscientious6                0            0            0             0
#> Conscientious7                0            0            0             0
#> Conscientious8                0            0            0             0
#> Conscientious9                0            0            0             0
#> Conscientious10               0            0            0             0
#> Openness1                     0            0            0             0
#> Openness2                     0            0            0             0
#> Openness3                     0            0            0             0
#> Openness5                     0            0            0             0
#> Openness7                     0            0            0             0
#> Openness8                     0            0            0             0
#> Openness9                     0            0            0             0
#> Openness10                    0            0            0             0
#> Behavior1                     0            0            0             0
#> Behavior2                     0            0            0             0
#> Behavior3                     0            0            0             0
#> Behavior5                     0            0            0             0
#> Behavior7                     0            0            0             0
#> Behavior8                     0            0            0             0
#> Behavior9                     0            0            0             0
#> Interview1                    0            0            0             0
#> Interview2                    0            0            0             0
#> Offer1                        0            0            0             0
#> Offer2                        0            0            0             0
#>                 Conscientiousness OpennesstoExperience NetworkingBehavior
#> Honesty1                        0                    0                  0
#> Honesty2                        0                    0                  0
#> Honesty3                        0                    0                  0
#> Honesty4                        0                    0                  0
#> Honesty5                        0                    0                  0
#> Honesty6                        0                    0                  0
#> Honesty7                        0                    0                  0
#> Honesty8                        0                    0                  0
#> Honesty9                        0                    0                  0
#> Honesty10                       0                    0                  0
#> Emotion1                        0                    0                  0
#> Emotion2                        0                    0                  0
#> Emotion3                        0                    0                  0
#> Emotion4                        0                    0                  0
#> Emotion5                        0                    0                  0
#> Emotion6                        0                    0                  0
#> Emotion8                        0                    0                  0
#> Emotion10                       0                    0                  0
#> Extraver2                       0                    0                  0
#> Extraver3                       0                    0                  0
#> Extraver4                       0                    0                  0
#> Extraver5                       0                    0                  0
#> Extraver6                       0                    0                  0
#> Extraver7                       0                    0                  0
#> Extraver8                       0                    0                  0
#> Extraver9                       0                    0                  0
#> Extraver10                      0                    0                  0
#> Agreeable1                      0                    0                  0
#> Agreeable3                      0                    0                  0
#> Agreeable4                      0                    0                  0
#> Agreeable5                      0                    0                  0
#> Agreeable7                      0                    0                  0
#> Agreeable8                      0                    0                  0
#> Agreeable9                      0                    0                  0
#> Agreeable10                     0                    0                  0
#> Conscientious1                  1                    0                  0
#> Conscientious3                  1                    0                  0
#> Conscientious4                  1                    0                  0
#> Conscientious6                  1                    0                  0
#> Conscientious7                  1                    0                  0
#> Conscientious8                  1                    0                  0
#> Conscientious9                  1                    0                  0
#> Conscientious10                 1                    0                  0
#> Openness1                       0                    1                  0
#> Openness2                       0                    1                  0
#> Openness3                       0                    1                  0
#> Openness5                       0                    1                  0
#> Openness7                       0                    1                  0
#> Openness8                       0                    1                  0
#> Openness9                       0                    1                  0
#> Openness10                      0                    1                  0
#> Behavior1                       0                    0                  1
#> Behavior2                       0                    0                  1
#> Behavior3                       0                    0                  1
#> Behavior5                       0                    0                  1
#> Behavior7                       0                    0                  1
#> Behavior8                       0                    0                  1
#> Behavior9                       0                    0                  1
#> Interview1                      0                    0                  0
#> Interview2                      0                    0                  0
#> Offer1                          0                    0                  0
#> Offer2                          0                    0                  0
#>                 Numberofjobinterviews Numberofjoboffers
#> Honesty1                            0                 0
#> Honesty2                            0                 0
#> Honesty3                            0                 0
#> Honesty4                            0                 0
#> Honesty5                            0                 0
#> Honesty6                            0                 0
#> Honesty7                            0                 0
#> Honesty8                            0                 0
#> Honesty9                            0                 0
#> Honesty10                           0                 0
#> Emotion1                            0                 0
#> Emotion2                            0                 0
#> Emotion3                            0                 0
#> Emotion4                            0                 0
#> Emotion5                            0                 0
#> Emotion6                            0                 0
#> Emotion8                            0                 0
#> Emotion10                           0                 0
#> Extraver2                           0                 0
#> Extraver3                           0                 0
#> Extraver4                           0                 0
#> Extraver5                           0                 0
#> Extraver6                           0                 0
#> Extraver7                           0                 0
#> Extraver8                           0                 0
#> Extraver9                           0                 0
#> Extraver10                          0                 0
#> Agreeable1                          0                 0
#> Agreeable3                          0                 0
#> Agreeable4                          0                 0
#> Agreeable5                          0                 0
#> Agreeable7                          0                 0
#> Agreeable8                          0                 0
#> Agreeable9                          0                 0
#> Agreeable10                         0                 0
#> Conscientious1                      0                 0
#> Conscientious3                      0                 0
#> Conscientious4                      0                 0
#> Conscientious6                      0                 0
#> Conscientious7                      0                 0
#> Conscientious8                      0                 0
#> Conscientious9                      0                 0
#> Conscientious10                     0                 0
#> Openness1                           0                 0
#> Openness2                           0                 0
#> Openness3                           0                 0
#> Openness5                           0                 0
#> Openness7                           0                 0
#> Openness8                           0                 0
#> Openness9                           0                 0
#> Openness10                          0                 0
#> Behavior1                           0                 0
#> Behavior2                           0                 0
#> Behavior3                           0                 0
#> Behavior5                           0                 0
#> Behavior7                           0                 0
#> Behavior8                           0                 0
#> Behavior9                           0                 0
#> Interview1                          1                 0
#> Interview2                          1                 0
#> Offer1                              0                 1
#> Offer2                              0                 1
#> 
#> $lv_type
#>       HonestyHumility          Emotionality          Extraversion 
#>                  TRUE                  TRUE                  TRUE 
#>         Agreeableness     Conscientiousness  OpennesstoExperience 
#>                  TRUE                  TRUE                  TRUE 
#>    NetworkingBehavior Numberofjobinterviews     Numberofjoboffers 
#>                 FALSE                 FALSE                 FALSE 
#> 
#> $ov_type
#>        Honesty1        Honesty2        Honesty3        Honesty4        Honesty5 
#>            TRUE            TRUE            TRUE            TRUE            TRUE 
#>        Honesty6        Honesty7        Honesty8        Honesty9       Honesty10 
#>            TRUE            TRUE            TRUE            TRUE            TRUE 
#>        Emotion1        Emotion2        Emotion3        Emotion4        Emotion5 
#>            TRUE            TRUE            TRUE            TRUE            TRUE 
#>        Emotion6        Emotion8       Emotion10       Extraver2       Extraver3 
#>            TRUE            TRUE            TRUE            TRUE            TRUE 
#>       Extraver4       Extraver5       Extraver6       Extraver7       Extraver8 
#>            TRUE            TRUE            TRUE            TRUE            TRUE 
#>       Extraver9      Extraver10      Agreeable1      Agreeable3      Agreeable4 
#>            TRUE            TRUE            TRUE            TRUE            TRUE 
#>      Agreeable5      Agreeable7      Agreeable8      Agreeable9     Agreeable10 
#>            TRUE            TRUE            TRUE            TRUE            TRUE 
#>  Conscientious1  Conscientious3  Conscientious4  Conscientious6  Conscientious7 
#>            TRUE            TRUE            TRUE            TRUE            TRUE 
#>  Conscientious8  Conscientious9 Conscientious10       Openness1       Openness2 
#>            TRUE            TRUE            TRUE            TRUE            TRUE 
#>       Openness3       Openness5       Openness7       Openness8       Openness9 
#>            TRUE            TRUE            TRUE            TRUE            TRUE 
#>      Openness10       Behavior1       Behavior2       Behavior3       Behavior5 
#>            TRUE           FALSE           FALSE           FALSE           FALSE 
#>       Behavior7       Behavior8       Behavior9      Interview1      Interview2 
#>           FALSE           FALSE           FALSE           FALSE           FALSE 
#>          Offer1          Offer2 
#>           FALSE           FALSE 
#> 
#> $C0
#>                 HonestyHumility Emotionality Extraversion Agreeableness
#> Honesty1                      1            0            0             0
#> Honesty2                      1            0            0             0
#> Honesty3                      1            0            0             0
#> Honesty4                      1            0            0             0
#> Honesty5                      1            0            0             0
#> Honesty6                      1            0            0             0
#> Honesty7                      1            0            0             0
#> Honesty8                      1            0            0             0
#> Honesty9                      1            0            0             0
#> Honesty10                     1            0            0             0
#> Emotion1                      0            1            0             0
#> Emotion2                      0            1            0             0
#> Emotion3                      0            1            0             0
#> Emotion4                      0            1            0             0
#> Emotion5                      0            1            0             0
#> Emotion6                      0            1            0             0
#> Emotion8                      0            1            0             0
#> Emotion10                     0            1            0             0
#> Extraver2                     0            0            1             0
#> Extraver3                     0            0            1             0
#> Extraver4                     0            0            1             0
#> Extraver5                     0            0            1             0
#> Extraver6                     0            0            1             0
#> Extraver7                     0            0            1             0
#> Extraver8                     0            0            1             0
#> Extraver9                     0            0            1             0
#> Extraver10                    0            0            1             0
#> Agreeable1                    0            0            0             1
#> Agreeable3                    0            0            0             1
#> Agreeable4                    0            0            0             1
#> Agreeable5                    0            0            0             1
#> Agreeable7                    0            0            0             1
#> Agreeable8                    0            0            0             1
#> Agreeable9                    0            0            0             1
#> Agreeable10                   0            0            0             1
#> Conscientious1                0            0            0             0
#> Conscientious3                0            0            0             0
#> Conscientious4                0            0            0             0
#> Conscientious6                0            0            0             0
#> Conscientious7                0            0            0             0
#> Conscientious8                0            0            0             0
#> Conscientious9                0            0            0             0
#> Conscientious10               0            0            0             0
#> Openness1                     0            0            0             0
#> Openness2                     0            0            0             0
#> Openness3                     0            0            0             0
#> Openness5                     0            0            0             0
#> Openness7                     0            0            0             0
#> Openness8                     0            0            0             0
#> Openness9                     0            0            0             0
#> Openness10                    0            0            0             0
#> Behavior1                     0            0            0             0
#> Behavior2                     0            0            0             0
#> Behavior3                     0            0            0             0
#> Behavior5                     0            0            0             0
#> Behavior7                     0            0            0             0
#> Behavior8                     0            0            0             0
#> Behavior9                     0            0            0             0
#> Interview1                    0            0            0             0
#> Interview2                    0            0            0             0
#> Offer1                        0            0            0             0
#> Offer2                        0            0            0             0
#>                 Conscientiousness OpennesstoExperience NetworkingBehavior
#> Honesty1                        0                    0                  0
#> Honesty2                        0                    0                  0
#> Honesty3                        0                    0                  0
#> Honesty4                        0                    0                  0
#> Honesty5                        0                    0                  0
#> Honesty6                        0                    0                  0
#> Honesty7                        0                    0                  0
#> Honesty8                        0                    0                  0
#> Honesty9                        0                    0                  0
#> Honesty10                       0                    0                  0
#> Emotion1                        0                    0                  0
#> Emotion2                        0                    0                  0
#> Emotion3                        0                    0                  0
#> Emotion4                        0                    0                  0
#> Emotion5                        0                    0                  0
#> Emotion6                        0                    0                  0
#> Emotion8                        0                    0                  0
#> Emotion10                       0                    0                  0
#> Extraver2                       0                    0                  0
#> Extraver3                       0                    0                  0
#> Extraver4                       0                    0                  0
#> Extraver5                       0                    0                  0
#> Extraver6                       0                    0                  0
#> Extraver7                       0                    0                  0
#> Extraver8                       0                    0                  0
#> Extraver9                       0                    0                  0
#> Extraver10                      0                    0                  0
#> Agreeable1                      0                    0                  0
#> Agreeable3                      0                    0                  0
#> Agreeable4                      0                    0                  0
#> Agreeable5                      0                    0                  0
#> Agreeable7                      0                    0                  0
#> Agreeable8                      0                    0                  0
#> Agreeable9                      0                    0                  0
#> Agreeable10                     0                    0                  0
#> Conscientious1                  1                    0                  0
#> Conscientious3                  1                    0                  0
#> Conscientious4                  1                    0                  0
#> Conscientious6                  1                    0                  0
#> Conscientious7                  1                    0                  0
#> Conscientious8                  1                    0                  0
#> Conscientious9                  1                    0                  0
#> Conscientious10                 1                    0                  0
#> Openness1                       0                    1                  0
#> Openness2                       0                    1                  0
#> Openness3                       0                    1                  0
#> Openness5                       0                    1                  0
#> Openness7                       0                    1                  0
#> Openness8                       0                    1                  0
#> Openness9                       0                    1                  0
#> Openness10                      0                    1                  0
#> Behavior1                       0                    0                  1
#> Behavior2                       0                    0                  1
#> Behavior3                       0                    0                  1
#> Behavior5                       0                    0                  1
#> Behavior7                       0                    0                  1
#> Behavior8                       0                    0                  1
#> Behavior9                       0                    0                  1
#> Interview1                      0                    0                  0
#> Interview2                      0                    0                  0
#> Offer1                          0                    0                  0
#> Offer2                          0                    0                  0
#>                 Numberofjobinterviews Numberofjoboffers
#> Honesty1                            0                 0
#> Honesty2                            0                 0
#> Honesty3                            0                 0
#> Honesty4                            0                 0
#> Honesty5                            0                 0
#> Honesty6                            0                 0
#> Honesty7                            0                 0
#> Honesty8                            0                 0
#> Honesty9                            0                 0
#> Honesty10                           0                 0
#> Emotion1                            0                 0
#> Emotion2                            0                 0
#> Emotion3                            0                 0
#> Emotion4                            0                 0
#> Emotion5                            0                 0
#> Emotion6                            0                 0
#> Emotion8                            0                 0
#> Emotion10                           0                 0
#> Extraver2                           0                 0
#> Extraver3                           0                 0
#> Extraver4                           0                 0
#> Extraver5                           0                 0
#> Extraver6                           0                 0
#> Extraver7                           0                 0
#> Extraver8                           0                 0
#> Extraver9                           0                 0
#> Extraver10                          0                 0
#> Agreeable1                          0                 0
#> Agreeable3                          0                 0
#> Agreeable4                          0                 0
#> Agreeable5                          0                 0
#> Agreeable7                          0                 0
#> Agreeable8                          0                 0
#> Agreeable9                          0                 0
#> Agreeable10                         0                 0
#> Conscientious1                      0                 0
#> Conscientious3                      0                 0
#> Conscientious4                      0                 0
#> Conscientious6                      0                 0
#> Conscientious7                      0                 0
#> Conscientious8                      0                 0
#> Conscientious9                      0                 0
#> Conscientious10                     0                 0
#> Openness1                           0                 0
#> Openness2                           0                 0
#> Openness3                           0                 0
#> Openness5                           0                 0
#> Openness7                           0                 0
#> Openness8                           0                 0
#> Openness9                           0                 0
#> Openness10                          0                 0
#> Behavior1                           0                 0
#> Behavior2                           0                 0
#> Behavior3                           0                 0
#> Behavior5                           0                 0
#> Behavior7                           0                 0
#> Behavior8                           0                 0
#> Behavior9                           0                 0
#> Interview1                          1                 0
#> Interview2                          1                 0
#> Offer1                              0                 1
#> Offer2                              0                 1
#> 
#> $ind_domi
#>       HonestyHumility          Emotionality          Extraversion 
#>                     1                    11                    19 
#>         Agreeableness     Conscientiousness  OpennesstoExperience 
#>                    28                    36                    44 
#>    NetworkingBehavior Numberofjobinterviews     Numberofjoboffers 
#>                    52                    59                    61 
#> 

require(here)
#> Loading required package: here
#> here() starts at /home/runner/work/cSEM/cSEM
require(readxl)
#> Loading required package: readxl

tutorial_igsca_model <- "
# Composite Model
NetworkingBehavior <~ Behavior1 + Behavior2 + Behavior3 + Behavior5 + Behavior7 + Behavior8 + Behavior9
Numberofjobinterviews <~ Interview1 + Interview2
Numberofjoboffers <~ Offer1 + Offer2 

# Reflective Measurement Model
HonestyHumility =~ Honesty1 + Honesty2 + Honesty3 + Honesty4 + Honesty5 + Honesty6 + Honesty7 + Honesty8 + Honesty9 + Honesty10
Emotionality =~ Emotion1 + Emotion2 + Emotion3 + Emotion4 + Emotion5 + Emotion6 + Emotion8 + Emotion10
Extraversion =~ Extraver2 + Extraver3 + Extraver4 + Extraver5 + Extraver6 + Extraver7 + Extraver8 + Extraver9 + Extraver10
Agreeableness =~ Agreeable1 + Agreeable3 + Agreeable4 + Agreeable5 + Agreeable7 + Agreeable8 + Agreeable9 + Agreeable10
Conscientiousness =~ Conscientious1 + Conscientious3 + Conscientious4 + Conscientious6 + Conscientious7 + Conscientious8 + Conscientious9 + Conscientious10
OpennesstoExperience =~ Openness1 + Openness2 + Openness3 + Openness5 + Openness7 + Openness8 + Openness9 + Openness10

# Structural Model
NetworkingBehavior ~ HonestyHumility + Emotionality + Extraversion + Agreeableness + Conscientiousness + OpennesstoExperience
Numberofjobinterviews ~ NetworkingBehavior
Numberofjoboffers ~ NetworkingBehavior
"

dat <- readxl::read_excel(here::here("dev", "Notes", "data", "mmc1.xlsx")) 
#> Error: `path` does not exist: ‘/home/runner/work/cSEM/cSEM/dev/Notes/data/mmc1.xlsx’

extract_parseModel(model = tutorial_igsca_model, data = dat, ind_domi_as_first = TRUE)
#> Error in eval(expr, envir, enclos): object 'dat' not found