Composite-based SEM

The central hub of the cSEM package. It acts like a foreman by collecting all (estimation) tasks, distributing them to lower level package functions, and eventually recollecting all of their results. It is called by csem() to manage the actual calculations. It may be called directly by the user, however, in most cases it will likely be more convenient to use csem() instead.

foreman(
  .data                        = NULL,
  .model                       = NULL,
  .approach_cor_robust         = c("none", "mcd", "spearman"),
  .approach_nl                 = c("sequential", "replace"),
  .approach_paths              = c("OLS", "2SLS"),
  .approach_weights            = c("PLS-PM", "SUMCORR", "MAXVAR", "SSQCORR", "MINVAR", "GENVAR",
                                   "GSCA", "unit", "bartlett", "regression"),
  .conv_criterion              = c("diff_absolute", "diff_squared", "diff_relative"),
  .disattenuate                = TRUE,
  .dominant_indicators         = NULL,
  .estimate_structural         = TRUE,
  .id                          = NULL,
  .iter_max                    = 100,
  .normality                   = TRUE,
  .PLS_approach_cf             = c("dist_squared_euclid", "dist_euclid_weighted",
                                   "fisher_transformed", "mean_arithmetic",
                                   "mean_geometric", "mean_harmonic",
                                   "geo_of_harmonic"),
  .PLS_ignore_structural_model = FALSE,
  .PLS_modes                   = NULL,
  .PLS_weight_scheme_inner     = c("path", "centroid", "factorial"),
  .reliabilities               = NULL,
  .tolerance                   = 1e-05
  )

Arguments

.data
A data.frame or a matrix of standardized or unstandarized data. Possible column types or classes of the data provided are: logical, numeric (double or integer), factor (ordered and unordered) or a mix of several types. The data may also include one character column whose column name must be given to .id. This column is assumed to contain group identifiers used to split the data into groups.

.model
A model in lavaan model syntax or a cSEMModel list.

.approach_cor_robust
Character string. Approach used to obtain a robust indicator correlation matrix. One of: "none" in which case the standard bravais-person correlation is used, "spearman" for the spearman correlation, or "mcd" via MASS::cov.rob() for a robust correlation matrix. Defaults to "none".

.approach_nl
Character string. Approach used to estimate nonlinear structural relationships. One of: "sequential" or "replace". Defaults to "sequential".

.approach_paths
Character string. Approach used to estimate the structural coefficients. One of: "OLS" or "2SLS" (not yet implemented). Defaults to "OLS".

.approach_weights
Character string. Approach used to obtain composite weights. One of: "PLS-PM", "SUMCORR", "MAXVAR", "SSQCORR", "MINVAR", "GENVAR", "GSCA", "PCA", "unit", "bartlett", or "regression". Defaults to "PLS-PM".

.conv_criterion
Character string. The criterion to use for the convergence check. One of: "diff_absolute", "diff_squared", or "diff_relative". Defaults to "diff_absolute".

.disattenuate
Logical. Should composite/proxy correlations be disattenuated to yield consisten loadings and path estimates if at least one of the construct is modeled as a common factor? Defaults to TRUE.

.dominant_indicators
A character vector of "name" = "value" pairs, where "value" is a character string giving the name of the dominant indicator and "name" a character string of the corresponding construct name. Dominant indicators may be specified for a subset of the constructs.

.estimate_structural
Logical. Should the structural coefficients be estimated? Defaults to TRUE.

.id
Character string or integer. The name or position of the column of .data used to split the data into groups. Defaults to NULL.

.iter_max
Integer. The maximum number of iterations allowed. If iter_max = 1 and .approach_weights = "PLS-PM" one-step weights are returned. If the algorithm exceeds the specified number, weights of iteration step .iter_max - 1 will be returned with a warning. Defaults to 100.

.normality
Logical. Should joint normality of \([\eta_{1:p}; \zeta; \epsilon]\) be assumed in the nonlinear model? See (Dijkstra and Schermelleh-Engel 2014) for details. Defaults to TRUE. Ignored if the model is linear.

.PLS_approach_cf
Character string. Approach used to obtain the correction factors for PLSc. One of: "dist_squared_euclid", "dist_euclid_weighted", "fisher_transformed", "mean_arithmetic", "mean_geometric", "mean_harmonic", "geo_of_harmonic". Defaults to "dist_squared_euclid". Ignored if .disattenuate = FALSE or if .approach_weights is not PLS-PM.

.PLS_ignore_structural_model
Logical. Should the structural model be ignored when calculating the inner weights of the PLS-PM algorithm? Defaults to FALSE. Ignored if .approach_weights is not PLS-PM.

.PLS_modes
Either a named list specifying the mode that should be used for each construct in the form "name" = "mode", a single character string giving the mode that should be used for all constructs, or NULL. Possible choices for "mode" are: "modeA", "modeB", "modeBNNLS", "unit", a single number (weight) or a vector of fixed weights of the same length as there are indicators for the construct given by "name". If only a single number is provided this is identical to using unit weights, as weights are rescaled such that the related composite has unit variance. Defaults to NULL. If NULL the appropriate mode according to the type of construct used is choosen. Ignored if .approach_weight is not PLS-PM.

.PLS_weight_scheme_inner
Character string. The inner weighting scheme used in PLS-PM. One of: "centroid", "factorial", or "path". Defaults to "path". Ignored if .approach_weight is not PLS-PM.

.reliabilities
A character vector of "name" = value pairs, where value is a number between 0 and 1 and "name" a character string of the corresponding construct name, or NULL. Reliabilities may be given for a subset of the constructs. Defaults to NULL in which case reliabilities are estimated by csem(). Currently, only supported for .approach_weights = "PLS-PM".

.tolerance
Double. The tolerance criterion for convergence. Defaults to 1e-05.

.starting_values
list of named vectors containig the starting values. Defaults to "NULL".

See also