One-to-One R Translation of gsca_inione.m from Heungsun Hwang

Internal I-GSCA Function

Initializes the values for I-GSCA

gsca_inione(z0, W0, B0)

gsca_inione(z0, W0, B0)

Arguments

z0

Data Matrix

W0

Indicator matrix of weights

B0

Indicator matrix of Path Coefficients.

Value

When used in the context of igsca_sim(), it returns a list of the starting values for Weights, Loadings and Path Coefficients. In principle, otherwise, it is a slightly modified implementation of ordinary Generalised Structured Component Analysis (GSCA).

When used in the context of igsca_sim(), it returns a list of the starting values for Weights, Loadings and Path Coefficients. In principle, otherwise, it is a slightly modified implementation of ordinary Generalised Structured Component Analysis (GSCA).

Details

Initializes the values for I-GSCA

Author

Michael S. Truong

Examples

tutorial_gsca_model <- "
# Composite Model
NetworkingBehavior <~ Behavior1 + Behavior2 + Behavior3 + Behavior5 + Behavior7 + Behavior8 + Behavior9
Numberofjobinterviews <~ Interview1 + Interview2
Numberofjoboffers <~ Offer1 + Offer2 

# Reflective Measurement Model Forced into Composite
HonestyHumility <~ Honesty1 + Honesty2 + Honesty3 + Honesty4 + Honesty5 + Honesty6 + Honesty7 + Honesty8 + Honesty9 + Honesty10
Emotionality <~ Emotion1 + Emotion2 + Emotion3 + Emotion4 + Emotion5 + Emotion6 + Emotion8 + Emotion10
Extraversion <~ Extraver2 + Extraver3 + Extraver4 + Extraver5 + Extraver6 + Extraver7 + Extraver8 + Extraver9 + Extraver10
Agreeableness <~ Agreeable1 + Agreeable3 + Agreeable4 + Agreeable5 + Agreeable7 + Agreeable8 + Agreeable9 + Agreeable10
Conscientiousness <~ Conscientious1 + Conscientious3 + Conscientious4 + Conscientious6 + Conscientious7 + Conscientious8 + Conscientious9 + Conscientious10
OpennesstoExperience <~ Openness1 + Openness2 + Openness3 + Openness5 + Openness7 + Openness8 + Openness9 + Openness10

# Structural Model
NetworkingBehavior ~ HonestyHumility + Emotionality + Extraversion + Agreeableness + Conscientiousness + OpennesstoExperience
Numberofjobinterviews ~ NetworkingBehavior
Numberofjoboffers ~ NetworkingBehavior
"

data("LeDang2022")

gsca_in <- extract_parseModel(model = tutorial_gsca_model,
                                   data = LeDang2022,
                                   ind_domi_as_first = TRUE)
                         
(gsca_inione_test <- with(gsca_in,
                          gsca_inione(z0 = z0,
                                      W0 = W0,
                                      B0 = B0)
                                      )
                                      )
#> $W
#>                 HonestyHumility Emotionality Extraversion Agreeableness
#> Honesty1              0.1304854    0.0000000    0.0000000     0.0000000
#> Honesty2              0.1315566    0.0000000    0.0000000     0.0000000
#> Honesty3              0.1345806    0.0000000    0.0000000     0.0000000
#> Honesty4              0.1393564    0.0000000    0.0000000     0.0000000
#> Honesty5              0.1291586    0.0000000    0.0000000     0.0000000
#> Honesty6              0.1344402    0.0000000    0.0000000     0.0000000
#> Honesty7              0.1330054    0.0000000    0.0000000     0.0000000
#> Honesty8              0.1416233    0.0000000    0.0000000     0.0000000
#> Honesty9              0.1459216    0.0000000    0.0000000     0.0000000
#> Honesty10             0.1334691    0.0000000    0.0000000     0.0000000
#> Emotion1              0.0000000    0.1784883    0.0000000     0.0000000
#> Emotion2              0.0000000    0.1768872    0.0000000     0.0000000
#> Emotion3              0.0000000    0.1525551    0.0000000     0.0000000
#> Emotion4              0.0000000    0.1525590    0.0000000     0.0000000
#> Emotion5              0.0000000    0.1666775    0.0000000     0.0000000
#> Emotion6              0.0000000    0.1509965    0.0000000     0.0000000
#> Emotion8              0.0000000    0.1774802    0.0000000     0.0000000
#> Emotion10             0.0000000    0.1665313    0.0000000     0.0000000
#> Extraver2             0.0000000    0.0000000    0.1352698     0.0000000
#> Extraver3             0.0000000    0.0000000    0.1274749     0.0000000
#> Extraver4             0.0000000    0.0000000    0.1505977     0.0000000
#> Extraver5             0.0000000    0.0000000    0.1538935     0.0000000
#> Extraver6             0.0000000    0.0000000    0.1433667     0.0000000
#> Extraver7             0.0000000    0.0000000    0.1441101     0.0000000
#> Extraver8             0.0000000    0.0000000    0.1487139     0.0000000
#> Extraver9             0.0000000    0.0000000    0.1565432     0.0000000
#> Extraver10            0.0000000    0.0000000    0.1540750     0.0000000
#> Agreeable1            0.0000000    0.0000000    0.0000000     0.1763759
#> Agreeable3            0.0000000    0.0000000    0.0000000     0.1382801
#> Agreeable4            0.0000000    0.0000000    0.0000000     0.1545001
#> Agreeable5            0.0000000    0.0000000    0.0000000     0.1844769
#> Agreeable7            0.0000000    0.0000000    0.0000000     0.1830864
#> Agreeable8            0.0000000    0.0000000    0.0000000     0.1440676
#> Agreeable9            0.0000000    0.0000000    0.0000000     0.1474328
#> Agreeable10           0.0000000    0.0000000    0.0000000     0.1536665
#> Conscientious1        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious3        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious4        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious6        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious7        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious8        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious9        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious10       0.0000000    0.0000000    0.0000000     0.0000000
#> Openness1             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness2             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness3             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness5             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness7             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness8             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness9             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness10            0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior1             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior2             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior3             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior5             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior7             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior8             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior9             0.0000000    0.0000000    0.0000000     0.0000000
#> Interview1            0.0000000    0.0000000    0.0000000     0.0000000
#> Interview2            0.0000000    0.0000000    0.0000000     0.0000000
#> Offer1                0.0000000    0.0000000    0.0000000     0.0000000
#> Offer2                0.0000000    0.0000000    0.0000000     0.0000000
#>                 Conscientiousness OpennesstoExperience NetworkingBehavior
#> Honesty1                0.0000000            0.0000000          0.0000000
#> Honesty2                0.0000000            0.0000000          0.0000000
#> Honesty3                0.0000000            0.0000000          0.0000000
#> Honesty4                0.0000000            0.0000000          0.0000000
#> Honesty5                0.0000000            0.0000000          0.0000000
#> Honesty6                0.0000000            0.0000000          0.0000000
#> Honesty7                0.0000000            0.0000000          0.0000000
#> Honesty8                0.0000000            0.0000000          0.0000000
#> Honesty9                0.0000000            0.0000000          0.0000000
#> Honesty10               0.0000000            0.0000000          0.0000000
#> Emotion1                0.0000000            0.0000000          0.0000000
#> Emotion2                0.0000000            0.0000000          0.0000000
#> Emotion3                0.0000000            0.0000000          0.0000000
#> Emotion4                0.0000000            0.0000000          0.0000000
#> Emotion5                0.0000000            0.0000000          0.0000000
#> Emotion6                0.0000000            0.0000000          0.0000000
#> Emotion8                0.0000000            0.0000000          0.0000000
#> Emotion10               0.0000000            0.0000000          0.0000000
#> Extraver2               0.0000000            0.0000000          0.0000000
#> Extraver3               0.0000000            0.0000000          0.0000000
#> Extraver4               0.0000000            0.0000000          0.0000000
#> Extraver5               0.0000000            0.0000000          0.0000000
#> Extraver6               0.0000000            0.0000000          0.0000000
#> Extraver7               0.0000000            0.0000000          0.0000000
#> Extraver8               0.0000000            0.0000000          0.0000000
#> Extraver9               0.0000000            0.0000000          0.0000000
#> Extraver10              0.0000000            0.0000000          0.0000000
#> Agreeable1              0.0000000            0.0000000          0.0000000
#> Agreeable3              0.0000000            0.0000000          0.0000000
#> Agreeable4              0.0000000            0.0000000          0.0000000
#> Agreeable5              0.0000000            0.0000000          0.0000000
#> Agreeable7              0.0000000            0.0000000          0.0000000
#> Agreeable8              0.0000000            0.0000000          0.0000000
#> Agreeable9              0.0000000            0.0000000          0.0000000
#> Agreeable10             0.0000000            0.0000000          0.0000000
#> Conscientious1          0.1781865            0.0000000          0.0000000
#> Conscientious3          0.1357296            0.0000000          0.0000000
#> Conscientious4          0.1710166            0.0000000          0.0000000
#> Conscientious6          0.1663097            0.0000000          0.0000000
#> Conscientious7          0.1843887            0.0000000          0.0000000
#> Conscientious8          0.1746693            0.0000000          0.0000000
#> Conscientious9          0.1531889            0.0000000          0.0000000
#> Conscientious10         0.1449009            0.0000000          0.0000000
#> Openness1               0.0000000            0.1630387          0.0000000
#> Openness2               0.0000000            0.1645995          0.0000000
#> Openness3               0.0000000            0.1646811          0.0000000
#> Openness5               0.0000000            0.1679694          0.0000000
#> Openness7               0.0000000            0.1841669          0.0000000
#> Openness8               0.0000000            0.1601390          0.0000000
#> Openness9               0.0000000            0.1738690          0.0000000
#> Openness10              0.0000000            0.1569210          0.0000000
#> Behavior1               0.0000000            0.0000000          0.1685460
#> Behavior2               0.0000000            0.0000000          0.1722992
#> Behavior3               0.0000000            0.0000000          0.1900044
#> Behavior5               0.0000000            0.0000000          0.1467963
#> Behavior7               0.0000000            0.0000000          0.1571797
#> Behavior8               0.0000000            0.0000000          0.1820792
#> Behavior9               0.0000000            0.0000000          0.1689299
#> Interview1              0.0000000            0.0000000          0.0000000
#> Interview2              0.0000000            0.0000000          0.0000000
#> Offer1                  0.0000000            0.0000000          0.0000000
#> Offer2                  0.0000000            0.0000000          0.0000000
#>                 Numberofjobinterviews Numberofjoboffers
#> Honesty1                    0.0000000         0.0000000
#> Honesty2                    0.0000000         0.0000000
#> Honesty3                    0.0000000         0.0000000
#> Honesty4                    0.0000000         0.0000000
#> Honesty5                    0.0000000         0.0000000
#> Honesty6                    0.0000000         0.0000000
#> Honesty7                    0.0000000         0.0000000
#> Honesty8                    0.0000000         0.0000000
#> Honesty9                    0.0000000         0.0000000
#> Honesty10                   0.0000000         0.0000000
#> Emotion1                    0.0000000         0.0000000
#> Emotion2                    0.0000000         0.0000000
#> Emotion3                    0.0000000         0.0000000
#> Emotion4                    0.0000000         0.0000000
#> Emotion5                    0.0000000         0.0000000
#> Emotion6                    0.0000000         0.0000000
#> Emotion8                    0.0000000         0.0000000
#> Emotion10                   0.0000000         0.0000000
#> Extraver2                   0.0000000         0.0000000
#> Extraver3                   0.0000000         0.0000000
#> Extraver4                   0.0000000         0.0000000
#> Extraver5                   0.0000000         0.0000000
#> Extraver6                   0.0000000         0.0000000
#> Extraver7                   0.0000000         0.0000000
#> Extraver8                   0.0000000         0.0000000
#> Extraver9                   0.0000000         0.0000000
#> Extraver10                  0.0000000         0.0000000
#> Agreeable1                  0.0000000         0.0000000
#> Agreeable3                  0.0000000         0.0000000
#> Agreeable4                  0.0000000         0.0000000
#> Agreeable5                  0.0000000         0.0000000
#> Agreeable7                  0.0000000         0.0000000
#> Agreeable8                  0.0000000         0.0000000
#> Agreeable9                  0.0000000         0.0000000
#> Agreeable10                 0.0000000         0.0000000
#> Conscientious1              0.0000000         0.0000000
#> Conscientious3              0.0000000         0.0000000
#> Conscientious4              0.0000000         0.0000000
#> Conscientious6              0.0000000         0.0000000
#> Conscientious7              0.0000000         0.0000000
#> Conscientious8              0.0000000         0.0000000
#> Conscientious9              0.0000000         0.0000000
#> Conscientious10             0.0000000         0.0000000
#> Openness1                   0.0000000         0.0000000
#> Openness2                   0.0000000         0.0000000
#> Openness3                   0.0000000         0.0000000
#> Openness5                   0.0000000         0.0000000
#> Openness7                   0.0000000         0.0000000
#> Openness8                   0.0000000         0.0000000
#> Openness9                   0.0000000         0.0000000
#> Openness10                  0.0000000         0.0000000
#> Behavior1                   0.0000000         0.0000000
#> Behavior2                   0.0000000         0.0000000
#> Behavior3                   0.0000000         0.0000000
#> Behavior5                   0.0000000         0.0000000
#> Behavior7                   0.0000000         0.0000000
#> Behavior8                   0.0000000         0.0000000
#> Behavior9                   0.0000000         0.0000000
#> Interview1                  0.5576583         0.0000000
#> Interview2                  0.5468182         0.0000000
#> Offer1                      0.0000000         0.5695984
#> Offer2                      0.0000000         0.5659101
#> 
#> $C
#>                        Honesty1  Honesty2  Honesty3  Honesty4  Honesty5
#> HonestyHumility       0.7136025 0.7140461 0.7288227 0.7645021 0.6972383
#> Emotionality          0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Extraversion          0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Agreeableness         0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Conscientiousness     0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> OpennesstoExperience  0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> NetworkingBehavior    0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Numberofjobinterviews 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Numberofjoboffers     0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>                        Honesty6  Honesty7  Honesty8  Honesty9 Honesty10
#> HonestyHumility       0.7203683 0.7366868 0.7690265 0.8035588 0.7287935
#> Emotionality          0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Extraversion          0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Agreeableness         0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Conscientiousness     0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> OpennesstoExperience  0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> NetworkingBehavior    0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Numberofjobinterviews 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Numberofjoboffers     0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>                        Emotion1 Emotion2  Emotion3  Emotion4  Emotion5
#> HonestyHumility       0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#> Emotionality          0.8102323 0.828708 0.6875224 0.6883683 0.7818643
#> Extraversion          0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#> Agreeableness         0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#> Conscientiousness     0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#> OpennesstoExperience  0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#> NetworkingBehavior    0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#> Numberofjobinterviews 0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#> Numberofjoboffers     0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#>                        Emotion6  Emotion8 Emotion10 Extraver2 Extraver3
#> HonestyHumility       0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Emotionality          0.6693734 0.7897156 0.7646741 0.0000000 0.0000000
#> Extraversion          0.0000000 0.0000000 0.0000000 0.7365427 0.6593729
#> Agreeableness         0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Conscientiousness     0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> OpennesstoExperience  0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> NetworkingBehavior    0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Numberofjobinterviews 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Numberofjoboffers     0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>                       Extraver4 Extraver5 Extraver6 Extraver7 Extraver8
#> HonestyHumility       0.0000000  0.000000 0.0000000 0.0000000   0.00000
#> Emotionality          0.0000000  0.000000 0.0000000 0.0000000   0.00000
#> Extraversion          0.7667772  0.755162 0.7651834 0.7604894   0.75179
#> Agreeableness         0.0000000  0.000000 0.0000000 0.0000000   0.00000
#> Conscientiousness     0.0000000  0.000000 0.0000000 0.0000000   0.00000
#> OpennesstoExperience  0.0000000  0.000000 0.0000000 0.0000000   0.00000
#> NetworkingBehavior    0.0000000  0.000000 0.0000000 0.0000000   0.00000
#> Numberofjobinterviews 0.0000000  0.000000 0.0000000 0.0000000   0.00000
#> Numberofjoboffers     0.0000000  0.000000 0.0000000 0.0000000   0.00000
#>                       Extraver9 Extraver10 Agreeable1 Agreeable3 Agreeable4
#> HonestyHumility       0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> Emotionality          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> Extraversion          0.8097646  0.8227442  0.0000000  0.0000000  0.0000000
#> Agreeableness         0.0000000  0.0000000  0.8455926  0.6631494  0.7388701
#> Conscientiousness     0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> OpennesstoExperience  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> NetworkingBehavior    0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> Numberofjobinterviews 0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> Numberofjoboffers     0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#>                       Agreeable5 Agreeable7 Agreeable8 Agreeable9 Agreeable10
#> HonestyHumility         0.000000  0.0000000  0.0000000   0.000000    0.000000
#> Emotionality            0.000000  0.0000000  0.0000000   0.000000    0.000000
#> Extraversion            0.000000  0.0000000  0.0000000   0.000000    0.000000
#> Agreeableness           0.888436  0.8784568  0.6971122   0.716613    0.743096
#> Conscientiousness       0.000000  0.0000000  0.0000000   0.000000    0.000000
#> OpennesstoExperience    0.000000  0.0000000  0.0000000   0.000000    0.000000
#> NetworkingBehavior      0.000000  0.0000000  0.0000000   0.000000    0.000000
#> Numberofjobinterviews   0.000000  0.0000000  0.0000000   0.000000    0.000000
#> Numberofjoboffers       0.000000  0.0000000  0.0000000   0.000000    0.000000
#>                       Conscientious1 Conscientious3 Conscientious4
#> HonestyHumility            0.0000000      0.0000000      0.0000000
#> Emotionality               0.0000000      0.0000000      0.0000000
#> Extraversion               0.0000000      0.0000000      0.0000000
#> Agreeableness              0.0000000      0.0000000      0.0000000
#> Conscientiousness          0.8187044      0.6257072      0.7976585
#> OpennesstoExperience       0.0000000      0.0000000      0.0000000
#> NetworkingBehavior         0.0000000      0.0000000      0.0000000
#> Numberofjobinterviews      0.0000000      0.0000000      0.0000000
#> Numberofjoboffers          0.0000000      0.0000000      0.0000000
#>                       Conscientious6 Conscientious7 Conscientious8
#> HonestyHumility             0.000000      0.0000000      0.0000000
#> Emotionality                0.000000      0.0000000      0.0000000
#> Extraversion                0.000000      0.0000000      0.0000000
#> Agreeableness               0.000000      0.0000000      0.0000000
#> Conscientiousness           0.767274      0.8542736      0.8120271
#> OpennesstoExperience        0.000000      0.0000000      0.0000000
#> NetworkingBehavior          0.000000      0.0000000      0.0000000
#> Numberofjobinterviews       0.000000      0.0000000      0.0000000
#> Numberofjoboffers           0.000000      0.0000000      0.0000000
#>                       Conscientious9 Conscientious10 Openness1 Openness2
#> HonestyHumility            0.0000000       0.0000000 0.0000000 0.0000000
#> Emotionality               0.0000000       0.0000000 0.0000000 0.0000000
#> Extraversion               0.0000000       0.0000000 0.0000000 0.0000000
#> Agreeableness              0.0000000       0.0000000 0.0000000 0.0000000
#> Conscientiousness          0.7078072       0.6721158 0.0000000 0.0000000
#> OpennesstoExperience       0.0000000       0.0000000 0.7245915 0.7451747
#> NetworkingBehavior         0.0000000       0.0000000 0.0000000 0.0000000
#> Numberofjobinterviews      0.0000000       0.0000000 0.0000000 0.0000000
#> Numberofjoboffers          0.0000000       0.0000000 0.0000000 0.0000000
#>                       Openness3 Openness5 Openness7 Openness8 Openness9
#> HonestyHumility       0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Emotionality          0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Extraversion          0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Agreeableness         0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Conscientiousness     0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> OpennesstoExperience  0.7297157 0.7585541 0.8283719 0.7179428 0.7746969
#> NetworkingBehavior    0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Numberofjobinterviews 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Numberofjoboffers     0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>                       Openness10 Behavior1 Behavior2 Behavior3 Behavior5
#> HonestyHumility        0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Emotionality           0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Extraversion           0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Agreeableness          0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Conscientiousness      0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> OpennesstoExperience   0.6971585 0.0000000 0.0000000 0.0000000 0.0000000
#> NetworkingBehavior     0.0000000 0.9047628 0.7993825 0.8749522 0.7306764
#> Numberofjobinterviews  0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Numberofjoboffers      0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>                       Behavior7 Behavior8 Behavior9 Interview1 Interview2
#> HonestyHumility       0.0000000 0.0000000 0.0000000  0.0000000  0.0000000
#> Emotionality          0.0000000 0.0000000 0.0000000  0.0000000  0.0000000
#> Extraversion          0.0000000 0.0000000 0.0000000  0.0000000  0.0000000
#> Agreeableness         0.0000000 0.0000000 0.0000000  0.0000000  0.0000000
#> Conscientiousness     0.0000000 0.0000000 0.0000000  0.0000000  0.0000000
#> OpennesstoExperience  0.0000000 0.0000000 0.0000000  0.0000000  0.0000000
#> NetworkingBehavior    0.7372587 0.9124828 0.9130486  0.0000000  0.0000000
#> Numberofjobinterviews 0.0000000 0.0000000 0.0000000  0.9073254  0.9034493
#> Numberofjoboffers     0.0000000 0.0000000 0.0000000  0.0000000  0.0000000
#>                         Offer1    Offer2
#> HonestyHumility       0.000000 0.0000000
#> Emotionality          0.000000 0.0000000
#> Extraversion          0.000000 0.0000000
#> Agreeableness         0.000000 0.0000000
#> Conscientiousness     0.000000 0.0000000
#> OpennesstoExperience  0.000000 0.0000000
#> NetworkingBehavior    0.000000 0.0000000
#> Numberofjobinterviews 0.000000 0.0000000
#> Numberofjoboffers     0.881474 0.8798462
#> 
#> $B
#>                       HonestyHumility Emotionality Extraversion Agreeableness
#> HonestyHumility                     0            0            0             0
#> Emotionality                        0            0            0             0
#> Extraversion                        0            0            0             0
#> Agreeableness                       0            0            0             0
#> Conscientiousness                   0            0            0             0
#> OpennesstoExperience                0            0            0             0
#> NetworkingBehavior                  0            0            0             0
#> Numberofjobinterviews               0            0            0             0
#> Numberofjoboffers                   0            0            0             0
#>                       Conscientiousness OpennesstoExperience NetworkingBehavior
#> HonestyHumility                       0                    0         0.14236612
#> Emotionality                          0                    0        -0.41118701
#> Extraversion                          0                    0         0.38967537
#> Agreeableness                         0                    0         0.13340329
#> Conscientiousness                     0                    0         0.08717214
#> OpennesstoExperience                  0                    0         0.14874473
#> NetworkingBehavior                    0                    0         0.00000000
#> Numberofjobinterviews                 0                    0         0.00000000
#> Numberofjoboffers                     0                    0         0.00000000
#>                       Numberofjobinterviews Numberofjoboffers
#> HonestyHumility                   0.0000000        0.00000000
#> Emotionality                      0.0000000        0.00000000
#> Extraversion                      0.0000000        0.00000000
#> Agreeableness                     0.0000000        0.00000000
#> Conscientiousness                 0.0000000        0.00000000
#> OpennesstoExperience              0.0000000        0.00000000
#> NetworkingBehavior                0.1310452        0.09478699
#> Numberofjobinterviews             0.0000000        0.00000000
#> Numberofjoboffers                 0.0000000        0.00000000
#> 
#> $it
#> [1] 5
#> 

require(here)
require(readxl)

tutorial_gsca_model <- "
# Composite Model
NetworkingBehavior <~ Behavior1 + Behavior2 + Behavior3 + Behavior5 + Behavior7 + Behavior8 + Behavior9
Numberofjobinterviews <~ Interview1 + Interview2
Numberofjoboffers <~ Offer1 + Offer2 

# Reflective Measurement Model Forced into Composite
HonestyHumility <~ Honesty1 + Honesty2 + Honesty3 + Honesty4 + Honesty5 + Honesty6 + Honesty7 + Honesty8 + Honesty9 + Honesty10
Emotionality <~ Emotion1 + Emotion2 + Emotion3 + Emotion4 + Emotion5 + Emotion6 + Emotion8 + Emotion10
Extraversion <~ Extraver2 + Extraver3 + Extraver4 + Extraver5 + Extraver6 + Extraver7 + Extraver8 + Extraver9 + Extraver10
Agreeableness <~ Agreeable1 + Agreeable3 + Agreeable4 + Agreeable5 + Agreeable7 + Agreeable8 + Agreeable9 + Agreeable10
Conscientiousness <~ Conscientious1 + Conscientious3 + Conscientious4 + Conscientious6 + Conscientious7 + Conscientious8 + Conscientious9 + Conscientious10
OpennesstoExperience <~ Openness1 + Openness2 + Openness3 + Openness5 + Openness7 + Openness8 + Openness9 + Openness10

# Structural Model
NetworkingBehavior ~ HonestyHumility + Emotionality + Extraversion + Agreeableness + Conscientiousness + OpennesstoExperience
Numberofjobinterviews ~ NetworkingBehavior
Numberofjoboffers ~ NetworkingBehavior
"

dat <- readxl::read_excel(here::here("dev", "Notes", "data", "mmc1.xlsx")) 
#> Error: `path` does not exist: ‘/home/runner/work/cSEM/cSEM/dev/Notes/data/mmc1.xlsx’

gsca_sim_in <- extract_parseModel(model = tutorial_gsca_model,
                                   data = dat,
                                   ind_domi_as_first = TRUE)
#> Error in eval(expr, envir, enclos): object 'dat' not found
                         
(gsca_inione_test <- gsca_inione(z0 = gsca_sim_in$z0, W0 = gsca_sim_in$W0, B0 = gsca_sim_in$B0))
#> Error in eval(expr, envir, enclos): object 'gsca_sim_in' not found