R Implementation of igsca_sim.m

This is a streamlined R implementation of I-GSCA, based on the matlab implementation in igsca_sim.m by Heungsun Hwang.

igsca(Z0, W0, C0, B0, con_type, ov_type, ind_domi, itmax = 100, ceps = 0.001)

Arguments

Z0

Input matrix. Should contain named columns in this form.

W0

Indicator matrix of weights: indicators (rows) and their corresponding construct variable (columns).

C0

Indicator matrix of loadings: indicators (rows) and their corresponding construct variable (columns).

B0

Square indicator matrix of path coefficients: from-construct-variable (rows) and to-construct-variable (columns). The order of construct variables should match the order in C0 and W0.

con_type

A vector that denotes whether each construct variable (columns in W0 and C0) is a common factor or composite. Its length should be equal to the number of columns of W0 and C0.

ov_type

An indicator vector that indices whether an indicator (rows of W0 and C0) corresponds to a common factor variable (1) or a composite variable (0). This vector is important for computing the uniqueness terms (D) because it zeros the entries for composite indicators.

ind_domi

A numeric vector that indices the dominant indicator for each construct variable. It is to be clarified whether this should only apply to factor latent variables or also composite latent variables. This is important for ensuring that the signs of the path-coefficients and loadings are consistent. It is sometimes the case in composite-based structural equation modelling methods that loadings/path-coefficients may have the opposite sign. The length of this vector should be equal to the number of construct variables and each value should represent the row number of the dominant indicator for that construct variable.

itmax

Maximum number of iterations of the Alternating Least Squares (ALS) algorithm.

ceps

Minimum amount of absolute change in the estimates of the path-coefficients (if B0 is non-zero) or the loadings (if B0 is all zero, meaning ther are no path-coefficients) between ALS iterations before ending the optimization.

Value

List of 4 matrices that make up a fitted I-GSCA Model: (1) Weights, (2) Loadings, (3) Uniqueness Terms D^2, and (4) Path Coefficients. TODO: Cite the GSCA Pro SEM reference instead of GSCA Pro's Example

Details

The code here generally assumes that every indicator corresponds to a single construct. (A construct is a superset consisting of latent and emergent variables. Latent variables are sometimes referred to as common-factor/factor variables; and emergent variables are sometimes referred to as component/composite variables.)

Author

Michael S. Truong

Examples

# Specify the model according to GSCA Pro's example
tutorial_igsca_model <- "
# Composite Model
NetworkingBehavior <~ Behavior1 + Behavior2 + Behavior3 + Behavior5 + Behavior7 + Behavior8 + Behavior9
Numberofjobinterviews <~ Interview1 + Interview2
Numberofjoboffers <~ Offer1 + Offer2 

# Reflective Measurement Model
HonestyHumility =~ Honesty1 + Honesty2 + Honesty3 + Honesty4 + Honesty5 + Honesty6 + Honesty7 + Honesty8 + Honesty9 + Honesty10
Emotionality =~ Emotion1 + Emotion2 + Emotion3 + Emotion4 + Emotion5 + Emotion6 + Emotion8 + Emotion10
Extraversion =~ Extraver2 + Extraver3 + Extraver4 + Extraver5 + Extraver6 + Extraver7 + Extraver8 + Extraver9 + Extraver10
Agreeableness =~ Agreeable1 + Agreeable3 + Agreeable4 + Agreeable5 + Agreeable7 + Agreeable8 + Agreeable9 + Agreeable10
Conscientiousness =~ Conscientious1 + Conscientious3 + Conscientious4 + Conscientious6 + Conscientious7 + Conscientious8 + Conscientious9 + Conscientious10
OpennesstoExperience =~ Openness1 + Openness2 + Openness3 + Openness5 + Openness7 + Openness8 + Openness9 + Openness10

# Structural Model
NetworkingBehavior ~ HonestyHumility + Emotionality + Extraversion + Agreeableness + Conscientiousness + OpennesstoExperience
Numberofjobinterviews ~ NetworkingBehavior
Numberofjoboffers ~ NetworkingBehavior
"

# Load the data
data(LeDang2022)

# Pre-generate the required matrices for the algorithm and assume that the first indicator for each latent variable is the dominant indicator.
igsca_in <- extract_parseModel(model = tutorial_igsca_model,
                                   data = LeDang2022,
                                   ind_domi_as_first = TRUE)

# Fit the I-GSCA model
(igsca_out <- with(igsca_in, igsca(Z0 = Z0,
                            W0 = W0,
                            C0 = C0,
                            B0 = B0,
                            con_type = con_type,
                            ov_type = ov_type,
                            ind_domi = ind_domi)
                            ))
#> $Weights
#>                 HonestyHumility Emotionality Extraversion Agreeableness
#> Honesty1              0.1312522    0.0000000    0.0000000     0.0000000
#> Honesty2              0.1312351    0.0000000    0.0000000     0.0000000
#> Honesty3              0.1417287    0.0000000    0.0000000     0.0000000
#> Honesty4              0.1486670    0.0000000    0.0000000     0.0000000
#> Honesty5              0.1355867    0.0000000    0.0000000     0.0000000
#> Honesty6              0.1400846    0.0000000    0.0000000     0.0000000
#> Honesty7              0.1432580    0.0000000    0.0000000     0.0000000
#> Honesty8              0.1495468    0.0000000    0.0000000     0.0000000
#> Honesty9              0.1521847    0.0000000    0.0000000     0.0000000
#> Honesty10             0.1417230    0.0000000    0.0000000     0.0000000
#> Emotion1              0.0000000    0.1893749    0.0000000     0.0000000
#> Emotion2              0.0000000    0.1936932    0.0000000     0.0000000
#> Emotion3              0.0000000    0.1606940    0.0000000     0.0000000
#> Emotion4              0.0000000    0.1608917    0.0000000     0.0000000
#> Emotion5              0.0000000    0.1827445    0.0000000     0.0000000
#> Emotion6              0.0000000    0.1560745    0.0000000     0.0000000
#> Emotion8              0.0000000    0.1710157    0.0000000     0.0000000
#> Emotion10             0.0000000    0.1787266    0.0000000     0.0000000
#> Extraver2             0.0000000    0.0000000    0.1492444     0.0000000
#> Extraver3             0.0000000    0.0000000    0.1336076     0.0000000
#> Extraver4             0.0000000    0.0000000    0.1553708     0.0000000
#> Extraver5             0.0000000    0.0000000    0.1530172     0.0000000
#> Extraver6             0.0000000    0.0000000    0.1550478     0.0000000
#> Extraver7             0.0000000    0.0000000    0.1513613     0.0000000
#> Extraver8             0.0000000    0.0000000    0.1446676     0.0000000
#> Extraver9             0.0000000    0.0000000    0.1640812     0.0000000
#> Extraver10            0.0000000    0.0000000    0.1667113     0.0000000
#> Agreeable1            0.0000000    0.0000000    0.0000000     0.1839949
#> Agreeable3            0.0000000    0.0000000    0.0000000     0.1442966
#> Agreeable4            0.0000000    0.0000000    0.0000000     0.1607728
#> Agreeable5            0.0000000    0.0000000    0.0000000     0.1933173
#> Agreeable7            0.0000000    0.0000000    0.0000000     0.1881124
#> Agreeable8            0.0000000    0.0000000    0.0000000     0.1504076
#> Agreeable9            0.0000000    0.0000000    0.0000000     0.1559299
#> Agreeable10           0.0000000    0.0000000    0.0000000     0.1616924
#> Conscientious1        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious3        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious4        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious6        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious7        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious8        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious9        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious10       0.0000000    0.0000000    0.0000000     0.0000000
#> Openness1             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness2             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness3             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness5             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness7             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness8             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness9             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness10            0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior1             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior2             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior3             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior5             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior7             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior8             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior9             0.0000000    0.0000000    0.0000000     0.0000000
#> Interview1            0.0000000    0.0000000    0.0000000     0.0000000
#> Interview2            0.0000000    0.0000000    0.0000000     0.0000000
#> Offer1                0.0000000    0.0000000    0.0000000     0.0000000
#> Offer2                0.0000000    0.0000000    0.0000000     0.0000000
#>                 Conscientiousness OpennesstoExperience NetworkingBehavior
#> Honesty1                0.0000000            0.0000000          0.0000000
#> Honesty2                0.0000000            0.0000000          0.0000000
#> Honesty3                0.0000000            0.0000000          0.0000000
#> Honesty4                0.0000000            0.0000000          0.0000000
#> Honesty5                0.0000000            0.0000000          0.0000000
#> Honesty6                0.0000000            0.0000000          0.0000000
#> Honesty7                0.0000000            0.0000000          0.0000000
#> Honesty8                0.0000000            0.0000000          0.0000000
#> Honesty9                0.0000000            0.0000000          0.0000000
#> Honesty10               0.0000000            0.0000000          0.0000000
#> Emotion1                0.0000000            0.0000000          0.0000000
#> Emotion2                0.0000000            0.0000000          0.0000000
#> Emotion3                0.0000000            0.0000000          0.0000000
#> Emotion4                0.0000000            0.0000000          0.0000000
#> Emotion5                0.0000000            0.0000000          0.0000000
#> Emotion6                0.0000000            0.0000000          0.0000000
#> Emotion8                0.0000000            0.0000000          0.0000000
#> Emotion10               0.0000000            0.0000000          0.0000000
#> Extraver2               0.0000000            0.0000000          0.0000000
#> Extraver3               0.0000000            0.0000000          0.0000000
#> Extraver4               0.0000000            0.0000000          0.0000000
#> Extraver5               0.0000000            0.0000000          0.0000000
#> Extraver6               0.0000000            0.0000000          0.0000000
#> Extraver7               0.0000000            0.0000000          0.0000000
#> Extraver8               0.0000000            0.0000000          0.0000000
#> Extraver9               0.0000000            0.0000000          0.0000000
#> Extraver10              0.0000000            0.0000000          0.0000000
#> Agreeable1              0.0000000            0.0000000          0.0000000
#> Agreeable3              0.0000000            0.0000000          0.0000000
#> Agreeable4              0.0000000            0.0000000          0.0000000
#> Agreeable5              0.0000000            0.0000000          0.0000000
#> Agreeable7              0.0000000            0.0000000          0.0000000
#> Agreeable8              0.0000000            0.0000000          0.0000000
#> Agreeable9              0.0000000            0.0000000          0.0000000
#> Agreeable10             0.0000000            0.0000000          0.0000000
#> Conscientious1          0.1863041            0.0000000          0.0000000
#> Conscientious3          0.1423857            0.0000000          0.0000000
#> Conscientious4          0.1815149            0.0000000          0.0000000
#> Conscientious6          0.1746006            0.0000000          0.0000000
#> Conscientious7          0.1905536            0.0000000          0.0000000
#> Conscientious8          0.1847620            0.0000000          0.0000000
#> Conscientious9          0.1610684            0.0000000          0.0000000
#> Conscientious10         0.1529464            0.0000000          0.0000000
#> Openness1               0.0000000            0.1714281          0.0000000
#> Openness2               0.0000000            0.1762978          0.0000000
#> Openness3               0.0000000            0.1726404          0.0000000
#> Openness5               0.0000000            0.1794632          0.0000000
#> Openness7               0.0000000            0.1949658          0.0000000
#> Openness8               0.0000000            0.1698551          0.0000000
#> Openness9               0.0000000            0.1832823          0.0000000
#> Openness10              0.0000000            0.1649378          0.0000000
#> Behavior1               0.0000000            0.0000000          0.1700226
#> Behavior2               0.0000000            0.0000000          0.1707887
#> Behavior3               0.0000000            0.0000000          0.1885848
#> Behavior5               0.0000000            0.0000000          0.1479338
#> Behavior7               0.0000000            0.0000000          0.1565416
#> Behavior8               0.0000000            0.0000000          0.1811013
#> Behavior9               0.0000000            0.0000000          0.1707307
#> Interview1              0.0000000            0.0000000          0.0000000
#> Interview2              0.0000000            0.0000000          0.0000000
#> Offer1                  0.0000000            0.0000000          0.0000000
#> Offer2                  0.0000000            0.0000000          0.0000000
#>                 Numberofjobinterviews Numberofjoboffers
#> Honesty1                    0.0000000         0.0000000
#> Honesty2                    0.0000000         0.0000000
#> Honesty3                    0.0000000         0.0000000
#> Honesty4                    0.0000000         0.0000000
#> Honesty5                    0.0000000         0.0000000
#> Honesty6                    0.0000000         0.0000000
#> Honesty7                    0.0000000         0.0000000
#> Honesty8                    0.0000000         0.0000000
#> Honesty9                    0.0000000         0.0000000
#> Honesty10                   0.0000000         0.0000000
#> Emotion1                    0.0000000         0.0000000
#> Emotion2                    0.0000000         0.0000000
#> Emotion3                    0.0000000         0.0000000
#> Emotion4                    0.0000000         0.0000000
#> Emotion5                    0.0000000         0.0000000
#> Emotion6                    0.0000000         0.0000000
#> Emotion8                    0.0000000         0.0000000
#> Emotion10                   0.0000000         0.0000000
#> Extraver2                   0.0000000         0.0000000
#> Extraver3                   0.0000000         0.0000000
#> Extraver4                   0.0000000         0.0000000
#> Extraver5                   0.0000000         0.0000000
#> Extraver6                   0.0000000         0.0000000
#> Extraver7                   0.0000000         0.0000000
#> Extraver8                   0.0000000         0.0000000
#> Extraver9                   0.0000000         0.0000000
#> Extraver10                  0.0000000         0.0000000
#> Agreeable1                  0.0000000         0.0000000
#> Agreeable3                  0.0000000         0.0000000
#> Agreeable4                  0.0000000         0.0000000
#> Agreeable5                  0.0000000         0.0000000
#> Agreeable7                  0.0000000         0.0000000
#> Agreeable8                  0.0000000         0.0000000
#> Agreeable9                  0.0000000         0.0000000
#> Agreeable10                 0.0000000         0.0000000
#> Conscientious1              0.0000000         0.0000000
#> Conscientious3              0.0000000         0.0000000
#> Conscientious4              0.0000000         0.0000000
#> Conscientious6              0.0000000         0.0000000
#> Conscientious7              0.0000000         0.0000000
#> Conscientious8              0.0000000         0.0000000
#> Conscientious9              0.0000000         0.0000000
#> Conscientious10             0.0000000         0.0000000
#> Openness1                   0.0000000         0.0000000
#> Openness2                   0.0000000         0.0000000
#> Openness3                   0.0000000         0.0000000
#> Openness5                   0.0000000         0.0000000
#> Openness7                   0.0000000         0.0000000
#> Openness8                   0.0000000         0.0000000
#> Openness9                   0.0000000         0.0000000
#> Openness10                  0.0000000         0.0000000
#> Behavior1                   0.0000000         0.0000000
#> Behavior2                   0.0000000         0.0000000
#> Behavior3                   0.0000000         0.0000000
#> Behavior5                   0.0000000         0.0000000
#> Behavior7                   0.0000000         0.0000000
#> Behavior8                   0.0000000         0.0000000
#> Behavior9                   0.0000000         0.0000000
#> Interview1                  0.5553846         0.0000000
#> Interview2                  0.5490997         0.0000000
#> Offer1                      0.0000000         0.5694425
#> Offer2                      0.0000000         0.5660663
#> 
#> $Loadings
#>                 HonestyHumility Emotionality Extraversion Agreeableness
#> Honesty1             0.67564593   0.01961251  0.003675978  -0.055809887
#> Honesty2             0.67434029   0.00000000  0.000000000   0.000000000
#> Honesty3             0.72882271   0.00000000  0.000000000   0.000000000
#> Honesty4             0.76450210   0.00000000  0.000000000   0.000000000
#> Honesty5             0.69723834   0.00000000  0.000000000   0.000000000
#> Honesty6             0.72036834   0.00000000  0.000000000   0.000000000
#> Honesty7             0.73668680   0.00000000  0.000000000   0.000000000
#> Honesty8             0.76902650   0.00000000  0.000000000   0.000000000
#> Honesty9             0.78521235  -0.03240507  0.041935702   0.004028337
#> Honesty10            0.72879355   0.00000000  0.000000000   0.000000000
#> Emotion1             0.00000000   0.81023233  0.000000000   0.000000000
#> Emotion2             0.00000000   0.82870795  0.000000000   0.000000000
#> Emotion3             0.00000000   0.68752242  0.000000000   0.000000000
#> Emotion4             0.00000000   0.68836835  0.000000000   0.000000000
#> Emotion5             0.00000000   0.78186431  0.000000000   0.000000000
#> Emotion6             0.00000000   0.66937340  0.000000000   0.000000000
#> Emotion8            -0.02234061   0.72883969  0.056053437   0.018721425
#> Emotion10            0.00000000   0.76467406  0.000000000   0.000000000
#> Extraver2            0.00000000   0.00000000  0.736542658   0.000000000
#> Extraver3            0.00000000   0.00000000  0.659372878   0.000000000
#> Extraver4            0.00000000   0.00000000  0.766777172   0.000000000
#> Extraver5            0.00000000   0.00000000  0.755162039   0.000000000
#> Extraver6            0.00000000   0.00000000  0.765183443   0.000000000
#> Extraver7            0.00000000   0.00000000  0.760489441   0.000000000
#> Extraver8           -0.06318632  -0.03913248  0.714201944   0.018793733
#> Extraver9            0.00000000   0.00000000  0.809764603   0.000000000
#> Extraver10           0.00000000   0.00000000  0.822744178   0.000000000
#> Agreeable1           0.00000000   0.00000000  0.000000000   0.845592583
#> Agreeable3           0.00000000   0.00000000  0.000000000   0.663149391
#> Agreeable4           0.00000000   0.00000000  0.000000000   0.738870098
#> Agreeable5           0.00000000   0.00000000  0.000000000   0.888436029
#> Agreeable7           0.00000000   0.00000000  0.000000000   0.878456827
#> Agreeable8          -0.01584033  -0.07264533  0.148137134   0.697112171
#> Agreeable9           0.00000000   0.00000000  0.000000000   0.716612994
#> Agreeable10          0.00000000   0.00000000  0.000000000   0.743095972
#> Conscientious1       0.00000000   0.00000000  0.000000000   0.000000000
#> Conscientious3       0.00000000   0.00000000  0.000000000   0.000000000
#> Conscientious4       0.00000000   0.00000000  0.000000000   0.000000000
#> Conscientious6       0.00000000   0.00000000  0.000000000   0.000000000
#> Conscientious7       0.00000000   0.00000000 -0.012736299   0.054842083
#> Conscientious8       0.01450482   0.00000000  0.000000000   0.000000000
#> Conscientious9       0.00000000   0.00000000  0.000000000   0.000000000
#> Conscientious10      0.00000000   0.00000000  0.000000000   0.000000000
#> Openness1            0.00000000   0.00000000  0.000000000   0.000000000
#> Openness2            0.00000000   0.00000000  0.000000000   0.000000000
#> Openness3            0.00000000   0.00000000  0.000000000   0.000000000
#> Openness5            0.00000000   0.00000000  0.000000000   0.000000000
#> Openness7            0.04416162   0.03799825  0.010000227  -0.024692611
#> Openness8            0.00000000   0.00000000  0.000000000   0.000000000
#> Openness9            0.00000000   0.00000000  0.000000000   0.000000000
#> Openness10           0.00000000   0.00000000  0.000000000   0.000000000
#> Behavior1            0.00000000   0.00000000  0.000000000   0.000000000
#> Behavior2            0.00000000   0.00000000  0.000000000   0.000000000
#> Behavior3            0.00000000   0.00000000  0.000000000   0.000000000
#> Behavior5            0.00000000   0.00000000  0.000000000   0.000000000
#> Behavior7            0.00000000   0.00000000  0.000000000   0.000000000
#> Behavior8            0.00000000   0.00000000  0.000000000   0.000000000
#> Behavior9            0.00000000   0.00000000  0.000000000   0.000000000
#> Interview1           0.00000000   0.00000000  0.000000000   0.000000000
#> Interview2           0.00000000   0.00000000  0.000000000   0.000000000
#> Offer1               0.00000000   0.00000000  0.000000000   0.000000000
#> Offer2               0.00000000   0.00000000  0.000000000   0.000000000
#>                 Conscientiousness OpennesstoExperience NetworkingBehavior
#> Honesty1               0.05185169          0.028029112        0.008624634
#> Honesty2               0.00000000          0.000000000        0.000000000
#> Honesty3               0.00000000          0.000000000        0.000000000
#> Honesty4               0.00000000          0.000000000        0.000000000
#> Honesty5               0.00000000          0.000000000        0.000000000
#> Honesty6               0.00000000          0.000000000        0.000000000
#> Honesty7               0.00000000          0.000000000        0.000000000
#> Honesty8               0.00000000          0.000000000        0.000000000
#> Honesty9              -0.01938104         -0.003024050       -0.046830650
#> Honesty10              0.00000000          0.000000000        0.000000000
#> Emotion1               0.00000000          0.000000000        0.000000000
#> Emotion2               0.00000000          0.000000000        0.000000000
#> Emotion3               0.00000000          0.000000000        0.000000000
#> Emotion4               0.00000000          0.000000000        0.000000000
#> Emotion5               0.00000000          0.000000000        0.000000000
#> Emotion6               0.00000000          0.000000000        0.000000000
#> Emotion8              -0.00551681          0.000361025       -0.042087979
#> Emotion10              0.00000000          0.000000000        0.000000000
#> Extraver2              0.00000000          0.000000000        0.000000000
#> Extraver3              0.00000000          0.000000000        0.000000000
#> Extraver4              0.00000000          0.000000000        0.000000000
#> Extraver5              0.00000000          0.000000000        0.000000000
#> Extraver6              0.00000000          0.000000000        0.000000000
#> Extraver7              0.00000000          0.000000000        0.333264635
#> Extraver8             -0.02162690          0.000000000        0.000000000
#> Extraver9              0.00000000          0.000000000        0.000000000
#> Extraver10             0.00000000          0.000000000        0.000000000
#> Agreeable1             0.00000000          0.000000000        0.000000000
#> Agreeable3             0.00000000          0.000000000        0.000000000
#> Agreeable4             0.00000000          0.000000000        0.000000000
#> Agreeable5             0.00000000          0.000000000        0.000000000
#> Agreeable7            -0.02722624         -0.017479505        0.234898612
#> Agreeable8             0.00000000          0.000000000        0.000000000
#> Agreeable9             0.00000000          0.000000000        0.000000000
#> Agreeable10            0.00000000          0.000000000        0.000000000
#> Conscientious1         0.81870444          0.000000000        0.000000000
#> Conscientious3         0.62570715          0.000000000        0.000000000
#> Conscientious4         0.79765851          0.000000000        0.000000000
#> Conscientious6         0.76727402          0.000000000        0.000000000
#> Conscientious7         0.84064236          0.025202828       -0.006460250
#> Conscientious8         0.81202705          0.000000000        0.000000000
#> Conscientious9         0.70780723          0.000000000        0.000000000
#> Conscientious10        0.67211580          0.000000000        0.000000000
#> Openness1              0.00000000          0.724591501        0.000000000
#> Openness2              0.00000000          0.745174708        0.000000000
#> Openness3              0.00000000          0.729715688        0.000000000
#> Openness5              0.00000000          0.758554109        0.000000000
#> Openness7              0.04572697          0.828707949       -0.036025518
#> Openness8              0.00000000          0.717942758        0.000000000
#> Openness9              0.00000000          0.774696871        0.000000000
#> Openness10             0.00000000          0.697158505        0.000000000
#> Behavior1              0.00000000          0.000000000        0.904762760
#> Behavior2              0.00000000          0.000000000        0.799382523
#> Behavior3              0.00000000          0.000000000        0.874952203
#> Behavior5              0.00000000          0.000000000        0.733945955
#> Behavior7              0.00000000          0.000000000        0.737258676
#> Behavior8              0.00000000          0.000000000        0.912482807
#> Behavior9              0.00000000          0.000000000        0.913048617
#> Interview1             0.00000000          0.000000000        0.000000000
#> Interview2             0.00000000          0.000000000        0.000000000
#> Offer1                 0.00000000          0.000000000        0.000000000
#> Offer2                 0.00000000          0.000000000        0.000000000
#>                 Numberofjobinterviews Numberofjoboffers
#> Honesty1                  0.015990836       0.009000068
#> Honesty2                  0.000000000       0.000000000
#> Honesty3                  0.000000000       0.000000000
#> Honesty4                  0.000000000       0.000000000
#> Honesty5                  0.000000000       0.000000000
#> Honesty6                  0.000000000       0.000000000
#> Honesty7                  0.000000000       0.000000000
#> Honesty8                  0.000000000       0.000000000
#> Honesty9                 -0.003794615       0.000000000
#> Honesty10                 0.000000000       0.000000000
#> Emotion1                  0.000000000       0.000000000
#> Emotion2                  0.000000000       0.000000000
#> Emotion3                  0.000000000       0.000000000
#> Emotion4                  0.000000000       0.000000000
#> Emotion5                  0.000000000       0.000000000
#> Emotion6                 -0.064209981       0.047164573
#> Emotion8                  0.000000000       0.000000000
#> Emotion10                 0.000000000       0.000000000
#> Extraver2                 0.000000000       0.000000000
#> Extraver3                 0.000000000       0.000000000
#> Extraver4                 0.000000000       0.000000000
#> Extraver5                 0.000000000       0.000000000
#> Extraver6                 0.000000000       0.000000000
#> Extraver7                -0.036718422      -0.024185441
#> Extraver8                 0.000000000       0.000000000
#> Extraver9                 0.000000000       0.000000000
#> Extraver10                0.000000000       0.000000000
#> Agreeable1                0.000000000       0.000000000
#> Agreeable3                0.000000000       0.000000000
#> Agreeable4                0.000000000       0.000000000
#> Agreeable5                0.000000000       0.000000000
#> Agreeable7                0.107065699       0.081987308
#> Agreeable8                0.000000000       0.000000000
#> Agreeable9                0.000000000       0.000000000
#> Agreeable10               0.000000000       0.000000000
#> Conscientious1            0.000000000       0.000000000
#> Conscientious3            0.000000000       0.000000000
#> Conscientious4            0.000000000       0.000000000
#> Conscientious6            0.000000000       0.000000000
#> Conscientious7           -0.024597125       0.011574961
#> Conscientious8            0.000000000       0.000000000
#> Conscientious9            0.000000000       0.000000000
#> Conscientious10           0.000000000       0.000000000
#> Openness1                 0.000000000       0.000000000
#> Openness2                 0.000000000       0.000000000
#> Openness3                 0.000000000       0.000000000
#> Openness5                 0.000000000       0.000000000
#> Openness7                 0.000000000       0.000000000
#> Openness8                 0.000000000       0.000000000
#> Openness9                 0.000000000       0.000000000
#> Openness10                0.000000000       0.000000000
#> Behavior1                 0.000000000       0.000000000
#> Behavior2                 0.000000000       0.000000000
#> Behavior3                 0.000000000       0.000000000
#> Behavior5                -0.021237258       0.000000000
#> Behavior7                 0.000000000       0.000000000
#> Behavior8                 0.000000000       0.000000000
#> Behavior9                 0.000000000       0.000000000
#> Interview1                0.907325362       0.000000000
#> Interview2                0.903449302       0.000000000
#> Offer1                    0.000000000       0.881473951
#> Offer2                   -0.001135717       0.880310267
#> 
#> $`Path Coefficients`
#>                       HonestyHumility Emotionality Extraversion Agreeableness
#> HonestyHumility                     0            0            0             0
#> Emotionality                        0            0            0             0
#> Extraversion                        0            0            0             0
#> Agreeableness                       0            0            0             0
#> Conscientiousness                   0            0            0             0
#> OpennesstoExperience                0            0            0             0
#> NetworkingBehavior                  0            0            0             0
#> Numberofjobinterviews               0            0            0             0
#> Numberofjoboffers                   0            0            0             0
#>                       Conscientiousness OpennesstoExperience NetworkingBehavior
#> HonestyHumility                       0                    0         0.13659105
#> Emotionality                          0                    0        -0.43143824
#> Extraversion                          0                    0         0.40467767
#> Agreeableness                         0                    0         0.12650076
#> Conscientiousness                     0                    0         0.08191758
#> OpennesstoExperience                  0                    0         0.14672563
#> NetworkingBehavior                    0                    0         0.00000000
#> Numberofjobinterviews                 0                    0         0.00000000
#> Numberofjoboffers                     0                    0         0.00000000
#>                       Numberofjobinterviews Numberofjoboffers
#> HonestyHumility                   0.0000000        0.00000000
#> Emotionality                      0.0000000        0.00000000
#> Extraversion                      0.0000000        0.00000000
#> Agreeableness                     0.0000000        0.00000000
#> Conscientiousness                 0.0000000        0.00000000
#> OpennesstoExperience              0.0000000        0.00000000
#> NetworkingBehavior                0.1308596        0.09473541
#> Numberofjobinterviews             0.0000000        0.00000000
#> Numberofjoboffers                 0.0000000        0.00000000
#> 
#> $`Uniqueness Terms`
#>        Honesty1        Honesty2        Honesty3        Honesty4        Honesty5 
#>       0.5256890       0.5289661       0.4784132       0.4232852       0.5398616 
#>        Honesty6        Honesty7        Honesty8        Honesty9       Honesty10 
#>       0.5049946       0.4646338       0.4137913       0.3653486       0.4814213 
#>        Emotion1        Emotion2        Emotion3        Emotion4        Emotion5 
#>       0.3607437       0.3309335       0.5861916       0.5798525       0.4250220 
#>        Emotion6        Emotion8       Emotion10       Extraver2       Extraver3 
#>       0.6085836       0.4188019       0.4620516       0.4865669       0.6108463 
#>       Extraver4       Extraver5       Extraver6       Extraver7       Extraver8 
#>       0.4420375       0.4603679       0.4251869       0.4649918       0.4612005 
#>       Extraver9      Extraver10      Agreeable1      Agreeable3      Agreeable4 
#>       0.3627198       0.3290539       0.3077245       0.6152270       0.4980621 
#>      Agreeable5      Agreeable7      Agreeable8      Agreeable9     Agreeable10 
#>       0.1941859       0.2160481       0.5565306       0.5288929       0.4920430 
#>  Conscientious1  Conscientious3  Conscientious4  Conscientious6  Conscientious7 
#>       0.3589430       0.6462198       0.3994692       0.4535376       0.2681892 
#>  Conscientious8  Conscientious9 Conscientious10       Openness1       Openness2 
#>       0.3490078       0.5433395       0.6017636       0.5250522       0.4929726 
#>       Openness3       Openness5       Openness7       Openness8       Openness9 
#>       0.5136302       0.4683197       0.3257443       0.5410549       0.4469085 
#>      Openness10       Behavior1       Behavior2       Behavior3       Behavior5 
#>       0.5666319       0.0000000       0.0000000       0.0000000       0.0000000 
#>       Behavior7       Behavior8       Behavior9      Interview1      Interview2 
#>       0.0000000       0.0000000       0.0000000       0.0000000       0.0000000 
#>          Offer1          Offer2 
#>       0.0000000       0.0000000 
#>