One-to-One R Translation of igsca_sim.m by Heungsun Hwang

This is a one-to-one R translation of the matlab implementation of I-GSCA by Heungsun Hwang. This implementation was also re-factored for ease of interpretation.

igsca_sim(
  z0,
  W0,
  C0,
  B0,
  lv_type,
  ov_type,
  ind_domi,
  nbt,
  swap_step = c("noswap", "prepare_for_ALS", "first_iteration_update_theta_Gamma_W_V",
    "first_iteration_update_C_B_D_uniqueD_est", "first_factor_update",
    "first_composite_update", "flip_signs_ind_domi"),
  itmax = 100,
  ceps = 0.001,
  devmode = FALSE,
  devdir = list("tests", "comparisons", "igsca_translation"),
  devmode_checkobj = FALSE
)

Arguments

z0

Input data-frame/matrix. Should contain named columns.

W0

Indicator matrix of weights: indicators (rows) and their corresponding latent variable (columns).

C0

Indicator matrix of loadings: indicators (rows) and their corresponding latent variable (columns).

B0

Square indicator matrix of path coefficients: from-latent-variable (rows) and to-latent-variable (columns). The order of latent variables should match the order in C0 and W0.

lv_type

A logical vector that indices whether a latent variable (columns in W0 and C0) is a factor (TRUE) or composite (FALSE). Its length should be equal to the number of columns of W0 and C0.

ov_type

An indicator vector that indices whether an indicator (rows of W0 and C0) corresponds to a factor latent variable (1) or a composite latent variable (0). This vector is important for computing the uniqueness terms (D) because it zeros the entries for composite indicators.

ind_domi

A numeric vector that indices the dominant indicator for each latent variable. It is to be clarified whether this should only apply to factor latent variables or also composite latent variables. This is important for ensuring that the signs of the path-coefficients and loadings are consistent. It is sometimes the case in composite-based structural equation modelling methods that loadings/path-coefficients may have the opposite sign. The length of this vector should be equal to the number of latent variables and each value should represent the row number of the dominant indicator for that latent variable.

nbt

Non-functional argument for the number of bootstrap-iterations, intended for computation of standard-errors, 95% confidence intervals and p-values.

swap_step

This is an argument for different key stages of the algorithm. When selected, it swaps the overlapping objects between the R and Matlab environment at that stage of the algorithm. Requires devmode to be TRUE to be functional. 'noswap' means that no swapping will be made.

itmax

Maximum number of iterations in the Alternating Least Squares (ALS) algorithm.

ceps

Minimum amount of absolute change in the estimates of the path-coefficients (if B0 is non-zero) or the loadings (if B0 is all zero, meaning ther are no path-coefficients) between ALS iterations before ending the optimization.

devmode

Logical argument (TRUE/FALSE) for whether devmode functionality should be enabled. When enabled, the function saves the function's environment as a .Rds at various stages of the algorithm. This was intended to facilitate the comparison between the R and Matlab implementations by narrowing down sources of discrepancy.

devdir

A list of folders to be provided to here::here() to store the saved environments from swap_step.

devmode_checkobj

A development argument for checking whether the list of overlapping and non-overlapping objects between the R and Matlab implementations of I-GSCA are the same as before.

Value

List of 4 matrices that make up a fitted I-GSCA Model: (1) Weights, (2) Loadings, (3) Uniqueness Terms D^2, and (4) Path Coefficients.

Details

The code here generally assumes that every indicator corresponds to a single latent variable.

Author

Michael S. Truong

Examples

require(here)
require(readxl)

# Specify the model according to GSCA Pro's example
tutorial_igsca_model <- "
# Composite Model
NetworkingBehavior <~ Behavior1 + Behavior2 + Behavior3 + Behavior5 + Behavior7 + Behavior8 + Behavior9
Numberofjobinterviews <~ Interview1 + Interview2
Numberofjoboffers <~ Offer1 + Offer2 

# Reflective Measurement Model
HonestyHumility =~ Honesty1 + Honesty2 + Honesty3 + Honesty4 + Honesty5 + Honesty6 + Honesty7 + Honesty8 + Honesty9 + Honesty10
Emotionality =~ Emotion1 + Emotion2 + Emotion3 + Emotion4 + Emotion5 + Emotion6 + Emotion8 + Emotion10
Extraversion =~ Extraver2 + Extraver3 + Extraver4 + Extraver5 + Extraver6 + Extraver7 + Extraver8 + Extraver9 + Extraver10
Agreeableness =~ Agreeable1 + Agreeable3 + Agreeable4 + Agreeable5 + Agreeable7 + Agreeable8 + Agreeable9 + Agreeable10
Conscientiousness =~ Conscientious1 + Conscientious3 + Conscientious4 + Conscientious6 + Conscientious7 + Conscientious8 + Conscientious9 + Conscientious10
OpennesstoExperience =~ Openness1 + Openness2 + Openness3 + Openness5 + Openness7 + Openness8 + Openness9 + Openness10

# Structural Model
NetworkingBehavior ~ HonestyHumility + Emotionality + Extraversion + Agreeableness + Conscientiousness + OpennesstoExperience
Numberofjobinterviews ~ NetworkingBehavior
Numberofjoboffers ~ NetworkingBehavior
"

# Load the data
dat <- readxl::read_excel(here::here("dev", "Notes", "data", "mmc1.xlsx")) 
#> Error: `path` does not exist: ‘/home/runner/work/cSEM/cSEM/dev/Notes/data/mmc1.xlsx’

# Pre-generate the required matrices for the algorithm and assume that the first indicator for each latent variable is the dominant indicator.
igsca_sim_in <- extract_parseModel(model = tutorial_igsca_model,
                                   data = dat,
                                   ind_domi_as_first = TRUE)
#> Error in eval(expr, envir, enclos): object 'dat' not found

# Fit the I-GSCA model
(igsca_sim_out <- igsca_sim(z0 = igsca_sim_in$z0,
                            W0 = igsca_sim_in$W0,
                            C0 = igsca_sim_in$C0,
                            B0 = igsca_sim_in$B0,
                            lv_type = igsca_sim_in$lv_type,
                            ov_type = igsca_sim_in$ov_type,
                            ind_domi = igsca_sim_in$ind_domi,
                            nbt = 0,
                            devmode = FALSE,
                            swap_step = "noswap")
                            )
#> Error in eval(expr, envir, enclos): object 'igsca_sim_in' not found