Inference

Calculate common inferential quantities. For users interested in the estimated standard errors and/or confidences intervals summarize() will usually be more helpful as it has a much more user-friendly print method.

MeanResample(.first_resample)

SdResample(.first_resample, .resample_method, .n)

BiasResample(.first_resample, .resample_method, .n)

StandardCIResample(
  .first_resample,
  .bias_corrected,
  .dist = c("z", "t"),
  .df = c("type1", "type2"),
  .resample_method,
  .n,
  .probs
)

PercentilCIResample(.first_resample, .probs)

BasicCIResample(.first_resample, .bias_corrected, .probs)

TStatCIResample(
  .first_resample,
  .second_resample,
  .bias_corrected,
  .resample_method,
  .resample_method2,
  .n,
  .probs
)

BcCIResample(.first_resample, .probs)

BcaCIResample(.object, .first_resample, .probs)

infer(
 .object            = NULL,
 .alpha             = 0.05,
 .bias_corrected    = TRUE,
 .quantity          = c("all", "mean", "sd", "bias", "CI_standard_z",
                        "CI_standard_t", "CI_percentile", "CI_basic",
                        "CI_bc", "CI_bca", "CI_t_interval")
)

Arguments

.first_resample
A list containing the .R resamples based on the original data obtained by resamplecSEMResults().

.resample_method
Character string. The resampling method to use. One of: "none", "bootstrap" or "jackknife". Defaults to "none".

.n
Integer. The number of observations of the original data.

.bias_corrected
Logical. Should the standard and the tStat confidence interval be bias-corrected using the bootstrapped bias estimate? If TRUE the confidence interval for some estimated parameter theta is centered at 2*theta - theta*_hat, where theta*_hat is the average over all .R bootstrap estimates of theta. Defaults to TRUE

.dist
Character string. The distribution to use for the critical value. One of "t" for Student's t-distribution or "z" for the standard normal distribution. Defaults to "z".

.df
Character string. The method for obtaining the degrees of freedom. Choices are "type1" and "type2". Defaults to "type1" .

.probs
A vector of probabilities.

.second_resample
A list containing .R2 resamples for each of the .R resamples of the first run.

.resample_method2
Character string. The resampling method to use when resampling from a resample. One of: "none", "bootstrap" or "jackknife". For "bootstrap" the number of draws is provided via .R2. Currently, resampling from each resample is only required for the studentized confidence intervall ("CI_t_interval") computed by the infer() function. Defaults to "none".

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.quantity
Character string. Which statistic should be returned? One of "all", "mean", "sd", "bias", "CI_standard_z", "CI_standard_t", "CI_percentile", "CI_basic", "CI_bc", "CI_bca", "CI_t_interval" Defaults to "all" in which case all quantities that do not require additional resampling are returned, i.e., all quantities but "CI_bca", "CI_t_interval".

Value

A list of class cSEMInfer.

Details

Calculate common inferential quantities (e.g., estimated standard errors, estimated bias, confidence intervals) based on a cSEMResults_resampled object as obtained by calling resamplecSEMResults() or by setting .resample_method = "bootstrap" or "jackknife" when calling csem(). Currently, the following quantities are returned by default (.quantity = "all"):

"mean", "sd" and "bias"

The mean, the standard deviation and the estimated bias (defined as the difference between the resample mean and the original estimate).

"CI_standard_z" and "CI_standard_t"

The standard confidence interval with standard errors estimated by the resample standard deviation. While "CI_standard_z" assumes a standard normally distributed statistic, "CI_standard_t" assumes a t-statistic with N - 1 degrees of freedom.

"CI_percentile"

The percentile confidence interval

"CI_basic"

The basic confidence interval

"CI_bc"

The bias corrected confidence interval

In addition, the bias-corrected and accelerated (.quantity = "CI_bca") and/or the "studentized" confidence interval (".quantity = CI_t_interval") may be computed. The former requires jackknife estimates to compute influence values and the latter requires double bootstrap. Both can potentially be time consuming. Hence, computation is triggerd only if explicitly chosen.

Functions

MeanResample: Computes the mean over all resamples for each resampled statistic/parameter.

SdResample: Computes the standard deviation over all resamples for each resampled statistic/estimator. This is usually taken to be the estimate of the standard error of the statistic/estimator.

BiasResample: Computes the estimated bias for each resampled statistic/estimator.

StandardCIResample: Computes the Standard CI with bootstrap SE's. Critical quantiles can be based on both the t- or the standard normal distribution (z). The former may perform better in small samples but there is no clear consenus on what the degrees of freedom should be. We use N - 1 ("type1").

PercentilCIResample: Computes the Percentile CI. The function takes the distribution F* (the CDF) of the resamples as an estimator for the true distribution F of the statistic/estimator of interest. Quantiles of the estimated distribution are then used as lower and upper bound.

BasicCIResample: Computes the Basic CI.

TStatCIResample: Computes the Studentized or t-statistic CI The function computes a boostrap t-statisic (since it is roughly pivotal) and constructs the CI based on bootstraped t-values and bootstraped/jackknife SE's

BcCIResample: Computes the Basic bias-corrected CI.

BcaCIResample: Computes the Basic bias-corrected and accelerated CI.

See also

Examples

model <- "
# Structural model
QUAL ~ EXPE
EXPE ~ IMAG
SAT  ~ IMAG + EXPE + QUAL + VAL
LOY  ~ IMAG + SAT
VAL  ~ EXPE + QUAL

# Measurement model
EXPE =~ expe1 + expe2 + expe3 + expe4 + expe5
IMAG =~ imag1 + imag2 + imag3 + imag4 + imag5
LOY  =~ loy1  + loy2  + loy3  + loy4
QUAL =~ qual1 + qual2 + qual3 + qual4 + qual5
SAT  =~ sat1  + sat2  + sat3  + sat4
VAL  =~ val1  + val2  + val3  + val4
"

## Estimate the model with bootstrap resampling 
a <- csem(satisfaction, model, .resample_method = "bootstrap", .R = 50)

## Compute inferential quantities
inf <- infer(a)

inf$Path_estimates$CI_basic

#>      EXPE ~ IMAG QUAL ~ EXPE VAL ~ EXPE VAL ~ QUAL SAT ~ IMAG SAT ~ EXPE
#> 95%L   0.5525000    1.013127  -5.454185   5.175456 -0.3153886   1.686469
#> 95%U   0.7264779    1.061032  -4.324367   6.294038  0.3088265   3.005721
#>      SAT ~ QUAL SAT ~ VAL LOY ~ IMAG LOY ~ SAT
#> 95%L  -4.002051  1.328586 0.09527189 0.3709789
#> 95%U  -2.347409  2.166338 0.42447341 0.7600598

inf$Indirect_effect$sd

#> QUAL ~ IMAG  VAL ~ IMAG  VAL ~ EXPE  SAT ~ IMAG  SAT ~ EXPE  SAT ~ QUAL 
#>  0.05860713  0.06697901  0.32751340  0.12872148  0.30829995  0.59440821 
#>  LOY ~ IMAG  LOY ~ EXPE  LOY ~ QUAL   LOY ~ VAL 
#>  0.06637711  0.12333398  0.31055299  0.19156874 

### To compute the bias-corrected and accelerated and/or the studentized t-inverval
### confidence interval:
inf <- infer(a, .quantity = c("all", "CI_bca")) # requires jackknife estimates

## For the studentied t-interval confidence interval, a double bootstrap is required:
## Estimate the model with double bootstrap resampling:
# Notes:
#   1. The .resample_method2 arguments triggers a bootstrap of each bootstrap sample
#   2. The double bootstrap is is very time consuming, consider setting 
#      `.eval_plan = "multiprocess`. 
#      To speed things up .R and .R2 are reduced for the example. Results are
#      therefore rather unreliable.
a1 <- csem(satisfaction, model, .resample_method = "bootstrap", .R = 40,
          .resample_method2 = "bootstrap", .R2 = 20, .handle_inadmissibles = "replace")
infer(a1, .quantity = "CI_t_interval")

#> $Path_estimates
#> $Path_estimates$CI_t_interval
#>      EXPE ~ IMAG QUAL ~ EXPE VAL ~ EXPE VAL ~ QUAL SAT ~ IMAG SAT ~ EXPE
#> 95%L   0.5709381    1.002436  -5.841294   3.652894 -0.3201541  0.9514138
#> 95%U   0.7604730    1.040795  -2.788526   6.762881  0.1929500  3.1862292
#>      SAT ~ QUAL SAT ~ VAL LOY ~ IMAG LOY ~ SAT
#> 95%L  -3.628159  1.519443  0.1081783 0.4385908
#> 95%U  -1.247221  2.022240  0.4061811 0.7318056
#> 
#> 
#> $Loading_estimates
#> $Loading_estimates$CI_t_interval
#>      IMAG =~ imag1 IMAG =~ imag2 IMAG =~ imag3 IMAG =~ imag4 IMAG =~ imag5
#> 95%L     0.4250932     0.6907819     0.6891490     0.1560079     0.6985790
#> 95%U     0.6743345     0.8728232     0.9973199     0.6831919     0.8946883
#>      EXPE =~ expe1 EXPE =~ expe2 EXPE =~ expe3 EXPE =~ expe4 EXPE =~ expe5
#> 95%L     0.5625977     0.7750777     0.5244897     0.6263616     0.6339677
#> 95%U     0.7729064     0.8739597     0.7011314     0.7772365     0.7996086
#>      QUAL =~ qual1 QUAL =~ qual2 QUAL =~ qual3 QUAL =~ qual4 QUAL =~ qual5
#> 95%L     0.5130296     0.7622344     0.5030588     0.6267507     0.6993810
#> 95%U     0.8738486     0.8713576     0.7733479     0.8072062     0.8292855
#>      VAL =~ val1 VAL =~ val2 VAL =~ val3 VAL =~ val4 SAT =~ sat1 SAT =~ sat2
#> 95%L   0.8018450   0.4656814   0.2772224   0.7182411   0.8913141   0.8258381
#> 95%U   0.9009933   0.7732456   0.6989665   0.8466651   1.0012777   0.9358386
#>      SAT =~ sat3 SAT =~ sat4 LOY =~ loy1 LOY =~ loy2 LOY =~ loy3 LOY =~ loy4
#> 95%L   0.6409567   0.6399087   0.7837733   0.4032661   0.8557978   0.1438706
#> 95%U   0.7941787   0.8121614   1.0039962   0.7150912   0.9637432   0.6582587
#> 
#> 
#> $Weight_estimates
#> $Weight_estimates$CI_t_interval
#>      IMAG <~ imag1 IMAG <~ imag2 IMAG <~ imag3 IMAG <~ imag4 IMAG <~ imag5
#> 95%L     0.1536362     0.2681209     0.2793628    0.06936659     0.2488499
#> 95%U     0.2475962     0.3441956     0.3825852    0.23688740     0.3700438
#>      EXPE <~ expe1 EXPE <~ expe2 EXPE <~ expe3 EXPE <~ expe4 EXPE <~ expe5
#> 95%L     0.2022975     0.2648649     0.1970659     0.2344334     0.2324375
#> 95%U     0.2736494     0.3329858     0.2454850     0.2729355     0.2876348
#>      QUAL <~ qual1 QUAL <~ qual2 QUAL <~ qual3 QUAL <~ qual4 QUAL <~ qual5
#> 95%L     0.2067232     0.2471991     0.1786856     0.2211848     0.2224629
#> 95%U     0.2720910     0.3065868     0.2546872     0.2660513     0.2925513
#>      VAL <~ val1 VAL <~ val2 VAL <~ val3 VAL <~ val4 SAT <~ sat1 SAT <~ sat2
#> 95%L   0.3205125   0.2734681   0.1760381   0.3010635   0.3044321   0.2877775
#> 95%U   0.4672240   0.3066602   0.2732199   0.3934064   0.3516010   0.3307024
#>      SAT <~ sat3 SAT <~ sat4 LOY <~ loy1 LOY <~ loy2 LOY <~ loy3 LOY <~ loy4
#> 95%L   0.2308317   0.2380994   0.3425282   0.2147647   0.3704460  0.07168159
#> 95%U   0.2667038   0.2724502   0.4498589   0.3018190   0.4376592  0.26409198
#> 
#> 
#> $Indirect_effect
#> $Indirect_effect$CI_t_interval
#>      QUAL ~ IMAG VAL ~ IMAG VAL ~ EXPE SAT ~ IMAG SAT ~ EXPE SAT ~ QUAL
#> 95%L   0.5703654  0.4737995   3.609875  0.4120868 -2.0608897   5.543814
#> 95%U   0.7873832  0.6416883   6.846364  0.6490695 -0.3643754   9.526895
#>      LOY ~ IMAG LOY ~ EXPE LOY ~ QUAL LOY ~ VAL
#> 95%L  0.1515964  0.3210373   1.900550 0.6863774
#> 95%U  0.4549105  0.5989155   3.478843 1.2204499
#> 
#> 
#> $Total_effect
#> $Total_effect$CI_t_interval
#>      EXPE ~ IMAG QUAL ~ IMAG QUAL ~ EXPE VAL ~ IMAG VAL ~ EXPE VAL ~ QUAL
#> 95%L   0.5709381   0.5703654    1.002436  0.4737995  0.7755923   3.652894
#> 95%U   0.7604730   0.7873832    1.040795  0.6416883  0.8738292   6.762881
#>      SAT ~ IMAG SAT ~ EXPE SAT ~ QUAL SAT ~ VAL LOY ~ IMAG LOY ~ EXPE
#> 95%L  0.2671629  0.6059686   3.708538  1.519443  0.4063114  0.3210373
#> 95%U  0.6559925  0.9377843   5.787784  2.022240  0.6640082  0.5989155
#>      LOY ~ QUAL LOY ~ VAL LOY ~ SAT
#> 95%L   1.900550 0.6863774 0.4385908
#> 95%U   3.478843 1.2204499 0.7318056
#> 
#> 
#> attr(,"class")
#> [1] "cSEMInfer"