Helper Function for parsing GSCA Pro V1.2.1 Full Results Files

The commented headers require more work to get working

The commented headers require more work to get working

parse_GSCAPro_FullResults(
  file_path = list("tests", "comparisons", "igsca_translation", "GSCAPro_1_2_1Output",
    "Tutorial_IGSCA_model1_full_result_ver0.csv"),
  headers = c("Weights", "Loadings", "Path coefficients", "Unique D^2")
)

parse_GSCAPro_FullResults(
  file_path = list("tests", "comparisons", "igsca_translation", "GSCAPro_1_2_1Output",
    "Tutorial_IGSCA_model1_full_result_ver0.csv"),
  headers = c("Weights", "Loadings", "Path coefficients", "Unique D^2")
)

Value

List of extracted matrices from GSCA Pro Full results

List of extracted matrices from GSCA Pro Full results

Examples

parse_GSCAPro_FullResults()
#> $Weights
#>                          V1  Estimate        SE  95%CI(L)  95%CI(U) V6
#> 1       Networking Behavior        NA        NA        NA        NA NA
#> 2                 Behavior1 0.1671936 0.0099055 0.1539514 0.1909371 NA
#> 3                 Behavior2 0.1726433 0.0066667 0.1587636 0.1830990 NA
#> 4                 Behavior3 0.1911421 0.0086230 0.1665411 0.2058762 NA
#> 5                 Behavior5 0.1440346 0.0071702 0.1331092 0.1612546 NA
#> 6                 Behavior7 0.1596119 0.0073981 0.1414242 0.1707143 NA
#> 7                 Behavior8 0.1815757 0.0107682 0.1591219 0.2051062 NA
#> 8                 Behavior9 0.1696251 0.0113066 0.1418340 0.1894732 NA
#> 9  Number of job interviews        NA        NA        NA        NA NA
#> 10               Interview1 0.5576638 0.0064227 0.5464545 0.5730207 NA
#> 11               Interview2 0.5468126 0.0065723 0.5343719 0.5594989 NA
#> 12     Number of job offers        NA        NA        NA        NA NA
#> 13                   Offer1 0.5696112 0.0077011 0.5565174 0.5834127 NA
#> 14                   Offer2 0.5658973 0.0075644 0.5512911 0.5816959 NA
#> 15         Honesty-Humility        NA        NA        NA        NA NA
#> 16                 Honesty1 0.1349733 0.0044542 0.1262036 0.1440515 NA
#> 17                 Honesty2 0.1350386 0.0048016 0.1252625 0.1445645 NA
#> 18                 Honesty3 0.1392710 0.0049692 0.1303633 0.1500562 NA
#> 19                 Honesty4 0.1474210 0.0043937 0.1374298 0.1547556 NA
#> 20                 Honesty5 0.1317405 0.0047460 0.1230254 0.1423623 NA
#> 21                 Honesty6 0.1376587 0.0042891 0.1303873 0.1460119 NA
#> 22                 Honesty7 0.1419123 0.0068544 0.1294019 0.1557427 NA
#> 23                 Honesty8 0.1489751 0.0051467 0.1397968 0.1618976 NA
#> 24                 Honesty9 0.1576230 0.0049096 0.1488768 0.1678034 NA
#> 25                Honesty10 0.1399914 0.0049802 0.1307772 0.1490573 NA
#> 26             Emotionality        NA        NA        NA        NA NA
#> 27                 Emotion1 0.1928581 0.0065152 0.1804106 0.2048190 NA
#> 28                 Emotion2 0.1983629 0.0057328 0.1860129 0.2082276 NA
#> 29                 Emotion3 0.1534060 0.0070156 0.1401976 0.1693963 NA
#> 30                 Emotion4 0.1550656 0.0063297 0.1413497 0.1657459 NA
#> 31                 Emotion5 0.1832932 0.0065394 0.1722218 0.1958798 NA
#> 32                 Emotion6 0.1479708 0.0060471 0.1344830 0.1579674 NA
#> 33                 Emotion8 0.1836255 0.0053606 0.1739570 0.1958031 NA
#> 34                Emotion10 0.1740474 0.0056151 0.1638572 0.1840936 NA
#> 35             Extraversion        NA        NA        NA        NA NA
#> 36                Extraver2 0.1469662 0.0066977 0.1350997 0.1587938 NA
#> 37                Extraver3 0.1273698 0.0071747 0.1169543 0.1439670 NA
#> 38                Extraver4 0.1535417 0.0059790 0.1410227 0.1636863 NA
#> 39                Extraver5 0.1501212 0.0066615 0.1373258 0.1647098 NA
#> 40                Extraver6 0.1554642 0.0046926 0.1464919 0.1649889 NA
#> 41                Extraver7 0.1514070 0.0041721 0.1452309 0.1599156 NA
#> 42                Extraver8 0.1501175 0.0057552 0.1393428 0.1620640 NA
#> 43                Extraver9 0.1660452 0.0054149 0.1560904 0.1767925 NA
#> 44               Extraver10 0.1703625 0.0064440 0.1566846 0.1819763 NA
#> 45            Agreeableness        NA        NA        NA        NA NA
#> 46               Agreeable1 0.1839944 0.0056785 0.1745753 0.1972824 NA
#> 47               Agreeable3 0.1359169 0.0129041 0.1058523 0.1549573 NA
#> 48               Agreeable4 0.1562316 0.0061708 0.1441099 0.1679014 NA
#> 49               Agreeable5 0.2040870 0.0068218 0.1874830 0.2174982 NA
#> 50               Agreeable7 0.1990366 0.0062715 0.1879625 0.2125708 NA
#> 51               Agreeable8 0.1480126 0.0062727 0.1345683 0.1580970 NA
#> 52               Agreeable9 0.1483438 0.0047706 0.1387289 0.1577470 NA
#> 53              Agreeable10 0.1547643 0.0061173 0.1428565 0.1677494 NA
#> 54        Conscientiousness        NA        NA        NA        NA NA
#> 55           Conscientious1 0.1869531 0.0050317 0.1784698 0.1993210 NA
#> 56           Conscientious3 0.1352216 0.0075690 0.1207304 0.1503276 NA
#> 57           Conscientious4 0.1808409 0.0056776 0.1710754 0.1929719 NA
#> 58           Conscientious6 0.1712721 0.0061318 0.1583916 0.1847007 NA
#> 59           Conscientious7 0.2026445 0.0061786 0.1906399 0.2125484 NA
#> 60           Conscientious8 0.1890013 0.0061771 0.1778275 0.2019886 NA
#> 61           Conscientious9 0.1571068 0.0076107 0.1424904 0.1743967 NA
#> 62          Conscientious10 0.1453154 0.0075222 0.1310376 0.1595066 NA
#> 63   Openness to Experience        NA        NA        NA        NA NA
#> 64                Openness1 0.1693864 0.0057547 0.1562103 0.1810029 NA
#> 65                Openness2 0.1759010 0.0069362 0.1622506 0.1923783 NA
#> 66                Openness3 0.1706805 0.0069817 0.1571838 0.1849877 NA
#> 67                Openness5 0.1801300 0.0061647 0.1688753 0.1916710 NA
#> 68                Openness7 0.2043809 0.0054475 0.1939947 0.2137997 NA
#> 69                Openness8 0.1661168 0.0070050 0.1536792 0.1813385 NA
#> 70                Openness9 0.1837819 0.0060944 0.1705079 0.1946385 NA
#> 71               Openness10 0.1600255 0.0071424 0.1441736 0.1730915 NA
#> 72                                 NA        NA        NA        NA NA
#> 
#> $Loadings
#>                          V1  Estimate        SE  95%CI(L)  95%CI(U) V6
#> 1       Networking Behavior        NA        NA        NA        NA NA
#> 2                 Behavior1 0.9046829 0.0148949 0.8715773 0.9271598 NA
#> 3                 Behavior2 0.7994980 0.0220873 0.7471305 0.8337132 NA
#> 4                 Behavior3 0.8749911 0.0173729 0.8392425 0.9045295 NA
#> 5                 Behavior5 0.7296377 0.0229329 0.6851776 0.7770384 NA
#> 6                 Behavior7 0.7380472 0.0364928 0.6663535 0.8007315 NA
#> 7                 Behavior8 0.9124444 0.0080646 0.8982692 0.9276207 NA
#> 8                 Behavior9 0.9131622 0.0076771 0.8976023 0.9279955 NA
#> 9  Number of job interviews        NA        NA        NA        NA NA
#> 10               Interview1 0.9073432 0.0070093 0.8932917 0.9225212 NA
#> 11               Interview2 0.9034311 0.0072442 0.8887800 0.9169500 NA
#> 12     Number of job offers        NA        NA        NA        NA NA
#> 13                   Offer1 0.8814935 0.0108241 0.8587240 0.9032240 NA
#> 14                   Offer2 0.8798265 0.0110266 0.8578285 0.9013256 NA
#> 15         Honesty-Humility        NA        NA        NA        NA NA
#> 16                 Honesty1 0.6726453 0.0289471 0.6181610 0.7289371 NA
#> 17                 Honesty2 0.6729669 0.0266330 0.6231794 0.7213007 NA
#> 18                 Honesty3 0.6940607 0.0216417 0.6541770 0.7430362 NA
#> 19                 Honesty4 0.7346629 0.0255258 0.6864575 0.7839045 NA
#> 20                 Honesty5 0.6565293 0.0258370 0.6023431 0.7048155 NA
#> 21                 Honesty6 0.6860204 0.0258578 0.6230655 0.7269823 NA
#> 22                 Honesty7 0.7072245 0.0263630 0.6568493 0.7572211 NA
#> 23                 Honesty8 0.7424046 0.0202610 0.7041661 0.7812717 NA
#> 24                 Honesty9 0.7854853 0.0178506 0.7518543 0.8193960 NA
#> 25                Honesty10 0.6976471 0.0281672 0.6472320 0.7484519 NA
#> 26             Emotionality        NA        NA        NA        NA NA
#> 27                 Emotion1 0.7916473 0.0177419 0.7574219 0.8202426 NA
#> 28                 Emotion2 0.8142535 0.0177223 0.7772647 0.8455244 NA
#> 29                 Emotion3 0.6296732 0.0317257 0.5707160 0.6903520 NA
#> 30                 Emotion4 0.6364922 0.0282551 0.5755661 0.6924075 NA
#> 31                 Emotion5 0.7523806 0.0230923 0.7095760 0.7954332 NA
#> 32                 Emotion6 0.6073636 0.0309599 0.5359383 0.6596498 NA
#> 33                 Emotion8 0.7536960 0.0212556 0.7111993 0.7930121 NA
#> 34                Emotion10 0.7143856 0.0275660 0.6565207 0.7673382 NA
#> 35             Extraversion        NA        NA        NA        NA NA
#> 36                Extraver2 0.6993185 0.0322083 0.6283271 0.7636305 NA
#> 37                Extraver3 0.6060653 0.0386622 0.5343619 0.6768038 NA
#> 38                Extraver4 0.7305910 0.0300952 0.6691501 0.7775481 NA
#> 39                Extraver5 0.7143110 0.0294604 0.6516207 0.7673753 NA
#> 40                Extraver6 0.7397646 0.0218056 0.6952885 0.7788172 NA
#> 41                Extraver7 0.7204402 0.0266672 0.6699985 0.7671337 NA
#> 42                Extraver8 0.7142999 0.0298465 0.6534033 0.7758096 NA
#> 43                Extraver9 0.7900990 0.0174506 0.7497601 0.8165177 NA
#> 44               Extraver10 0.8106963 0.0229368 0.7607017 0.8503707 NA
#> 45            Agreeableness        NA        NA        NA        NA NA
#> 46               Agreeable1 0.8151671 0.0211218 0.7765273 0.8571380 NA
#> 47               Agreeable3 0.6019611 0.0495275 0.5086651 0.6816085 NA
#> 48               Agreeable4 0.6920310 0.0379855 0.6211914 0.7669604 NA
#> 49               Agreeable5 0.9020809 0.0149907 0.8614063 0.9206177 NA
#> 50               Agreeable7 0.8813253 0.0156511 0.8599418 0.9175347 NA
#> 51               Agreeable8 0.6553498 0.0329894 0.5838173 0.7060341 NA
#> 52               Agreeable9 0.6570083 0.0313352 0.5973889 0.7182614 NA
#> 53              Agreeable10 0.6855452 0.0383433 0.6185974 0.7514370 NA
#> 54        Conscientiousness        NA        NA        NA        NA NA
#> 55           Conscientious1 0.7861901 0.0246077 0.7249477 0.8238033 NA
#> 56           Conscientious3 0.5685257 0.0287022 0.5203846 0.6152199 NA
#> 57           Conscientious4 0.7604740 0.0219012 0.7216313 0.8053899 NA
#> 58           Conscientious6 0.7202210 0.0273935 0.6663930 0.7694239 NA
#> 59           Conscientious7 0.8511911 0.0139611 0.8213115 0.8738885 NA
#> 60           Conscientious8 0.7946675 0.0188292 0.7550659 0.8254577 NA
#> 61           Conscientious9 0.6606040 0.0356693 0.5932370 0.7407418 NA
#> 62          Conscientious10 0.6110580 0.0399238 0.5442812 0.6813451 NA
#> 63   Openness to Experience        NA        NA        NA        NA NA
#> 64                Openness1 0.6776794 0.0271176 0.6141487 0.7216211 NA
#> 65                Openness2 0.7037400 0.0262765 0.6571034 0.7642079 NA
#> 66                Openness3 0.6828585 0.0304850 0.6314594 0.7419502 NA
#> 67                Openness5 0.7206599 0.0206272 0.6719613 0.7568806 NA
#> 68                Openness7 0.8174478 0.0185773 0.7792504 0.8547072 NA
#> 69                Openness8 0.6646166 0.0292087 0.6195105 0.7281679 NA
#> 70                Openness9 0.7352702 0.0242531 0.6867222 0.7799879 NA
#> 71               Openness10 0.6402401 0.0291230 0.5933387 0.7049633 NA
#> 72                                 NA        NA        NA        NA NA
#> 
#> $`Path coefficients`
#>                                              V1 V2   Estimate        SE
#> 1         Honesty-Humility->Networking Behavior NA  0.1361644 0.0309411
#> 2             Emotionality->Networking Behavior NA -0.4311118 0.0340716
#> 3             Extraversion->Networking Behavior NA  0.4051107 0.0362823
#> 4            Agreeableness->Networking Behavior NA  0.1268461 0.0336086
#> 5        Conscientiousness->Networking Behavior NA  0.0818501 0.0284513
#> 6   Openness to Experience->Networking Behavior NA  0.1463257 0.0273398
#> 7 Networking Behavior->Number of job interviews NA  0.1310967 0.0361253
#> 8     Networking Behavior->Number of job offers NA  0.0948720 0.0363889
#> 9                                               NA         NA        NA
#>     95%CI(L)   95%CI(U)
#> 1  0.0823789  0.1988008
#> 2 -0.4975993 -0.3714208
#> 3  0.3317080  0.4769753
#> 4  0.0665760  0.2012249
#> 5  0.0312144  0.1347351
#> 6  0.0948065  0.2024902
#> 7  0.0672806  0.2064333
#> 8  0.0209743  0.1622576
#> 9         NA         NA
#> 
#> $`Unique D^2`
#>    Honesty1  Honesty2  Honesty3  Honesty4  Honesty5 Honesty6  Honesty7
#> 1 0.5273984 0.5306635 0.4787401 0.4222894 0.5405153 0.504849 0.4636667
#>    Honesty8  Honesty9 Honesty10 Emotion1 Emotion2  Emotion3  Emotion4  Emotion5
#> 1 0.4133138 0.3626471 0.4813632 0.355008 0.325329 0.5879229 0.5801572 0.4234307
#>    Emotion6  Emotion8 Emotion10 Extraver2 Extraver3 Extraver4 Extraver5
#> 1 0.6114926 0.4203239 0.4667098 0.4872904 0.6124864  0.443869 0.4627561
#>   Extraver6 Extraver7 Extraver8 Extraver9 Extraver10 Agreeable1 Agreeable3
#> 1 0.4222565 0.4668947 0.4622508 0.3604529  0.3234635  0.3152288  0.6182034
#>   Agreeable4 Agreeable5 Agreeable7 Agreeable8 Agreeable9 Agreeable10
#> 1  0.5026326  0.1692086  0.2094651   0.555478  0.5351004   0.4995825
#>   Conscientious1 Conscientious3 Conscientious4 Conscientious6 Conscientious7
#> 1      0.3623469      0.6460999      0.4025884      0.4577823      0.2557225
#>   Conscientious8 Conscientious9 Conscientious10 Openness1 Openness2 Openness3
#> 1      0.3450169      0.5441084        0.605072 0.5260668  0.493418 0.5147944
#>   Openness5 Openness7 Openness8 Openness9 Openness10 V52
#> 1 0.4682718 0.3139757 0.5436591 0.4486768  0.5695222  NA
#> 
parse_GSCAPro_FullResults()
#> $Weights
#>                          V1  Estimate        SE  95%CI(L)  95%CI(U) V6
#> 1       Networking Behavior        NA        NA        NA        NA NA
#> 2                 Behavior1 0.1671936 0.0099055 0.1539514 0.1909371 NA
#> 3                 Behavior2 0.1726433 0.0066667 0.1587636 0.1830990 NA
#> 4                 Behavior3 0.1911421 0.0086230 0.1665411 0.2058762 NA
#> 5                 Behavior5 0.1440346 0.0071702 0.1331092 0.1612546 NA
#> 6                 Behavior7 0.1596119 0.0073981 0.1414242 0.1707143 NA
#> 7                 Behavior8 0.1815757 0.0107682 0.1591219 0.2051062 NA
#> 8                 Behavior9 0.1696251 0.0113066 0.1418340 0.1894732 NA
#> 9  Number of job interviews        NA        NA        NA        NA NA
#> 10               Interview1 0.5576638 0.0064227 0.5464545 0.5730207 NA
#> 11               Interview2 0.5468126 0.0065723 0.5343719 0.5594989 NA
#> 12     Number of job offers        NA        NA        NA        NA NA
#> 13                   Offer1 0.5696112 0.0077011 0.5565174 0.5834127 NA
#> 14                   Offer2 0.5658973 0.0075644 0.5512911 0.5816959 NA
#> 15         Honesty-Humility        NA        NA        NA        NA NA
#> 16                 Honesty1 0.1349733 0.0044542 0.1262036 0.1440515 NA
#> 17                 Honesty2 0.1350386 0.0048016 0.1252625 0.1445645 NA
#> 18                 Honesty3 0.1392710 0.0049692 0.1303633 0.1500562 NA
#> 19                 Honesty4 0.1474210 0.0043937 0.1374298 0.1547556 NA
#> 20                 Honesty5 0.1317405 0.0047460 0.1230254 0.1423623 NA
#> 21                 Honesty6 0.1376587 0.0042891 0.1303873 0.1460119 NA
#> 22                 Honesty7 0.1419123 0.0068544 0.1294019 0.1557427 NA
#> 23                 Honesty8 0.1489751 0.0051467 0.1397968 0.1618976 NA
#> 24                 Honesty9 0.1576230 0.0049096 0.1488768 0.1678034 NA
#> 25                Honesty10 0.1399914 0.0049802 0.1307772 0.1490573 NA
#> 26             Emotionality        NA        NA        NA        NA NA
#> 27                 Emotion1 0.1928581 0.0065152 0.1804106 0.2048190 NA
#> 28                 Emotion2 0.1983629 0.0057328 0.1860129 0.2082276 NA
#> 29                 Emotion3 0.1534060 0.0070156 0.1401976 0.1693963 NA
#> 30                 Emotion4 0.1550656 0.0063297 0.1413497 0.1657459 NA
#> 31                 Emotion5 0.1832932 0.0065394 0.1722218 0.1958798 NA
#> 32                 Emotion6 0.1479708 0.0060471 0.1344830 0.1579674 NA
#> 33                 Emotion8 0.1836255 0.0053606 0.1739570 0.1958031 NA
#> 34                Emotion10 0.1740474 0.0056151 0.1638572 0.1840936 NA
#> 35             Extraversion        NA        NA        NA        NA NA
#> 36                Extraver2 0.1469662 0.0066977 0.1350997 0.1587938 NA
#> 37                Extraver3 0.1273698 0.0071747 0.1169543 0.1439670 NA
#> 38                Extraver4 0.1535417 0.0059790 0.1410227 0.1636863 NA
#> 39                Extraver5 0.1501212 0.0066615 0.1373258 0.1647098 NA
#> 40                Extraver6 0.1554642 0.0046926 0.1464919 0.1649889 NA
#> 41                Extraver7 0.1514070 0.0041721 0.1452309 0.1599156 NA
#> 42                Extraver8 0.1501175 0.0057552 0.1393428 0.1620640 NA
#> 43                Extraver9 0.1660452 0.0054149 0.1560904 0.1767925 NA
#> 44               Extraver10 0.1703625 0.0064440 0.1566846 0.1819763 NA
#> 45            Agreeableness        NA        NA        NA        NA NA
#> 46               Agreeable1 0.1839944 0.0056785 0.1745753 0.1972824 NA
#> 47               Agreeable3 0.1359169 0.0129041 0.1058523 0.1549573 NA
#> 48               Agreeable4 0.1562316 0.0061708 0.1441099 0.1679014 NA
#> 49               Agreeable5 0.2040870 0.0068218 0.1874830 0.2174982 NA
#> 50               Agreeable7 0.1990366 0.0062715 0.1879625 0.2125708 NA
#> 51               Agreeable8 0.1480126 0.0062727 0.1345683 0.1580970 NA
#> 52               Agreeable9 0.1483438 0.0047706 0.1387289 0.1577470 NA
#> 53              Agreeable10 0.1547643 0.0061173 0.1428565 0.1677494 NA
#> 54        Conscientiousness        NA        NA        NA        NA NA
#> 55           Conscientious1 0.1869531 0.0050317 0.1784698 0.1993210 NA
#> 56           Conscientious3 0.1352216 0.0075690 0.1207304 0.1503276 NA
#> 57           Conscientious4 0.1808409 0.0056776 0.1710754 0.1929719 NA
#> 58           Conscientious6 0.1712721 0.0061318 0.1583916 0.1847007 NA
#> 59           Conscientious7 0.2026445 0.0061786 0.1906399 0.2125484 NA
#> 60           Conscientious8 0.1890013 0.0061771 0.1778275 0.2019886 NA
#> 61           Conscientious9 0.1571068 0.0076107 0.1424904 0.1743967 NA
#> 62          Conscientious10 0.1453154 0.0075222 0.1310376 0.1595066 NA
#> 63   Openness to Experience        NA        NA        NA        NA NA
#> 64                Openness1 0.1693864 0.0057547 0.1562103 0.1810029 NA
#> 65                Openness2 0.1759010 0.0069362 0.1622506 0.1923783 NA
#> 66                Openness3 0.1706805 0.0069817 0.1571838 0.1849877 NA
#> 67                Openness5 0.1801300 0.0061647 0.1688753 0.1916710 NA
#> 68                Openness7 0.2043809 0.0054475 0.1939947 0.2137997 NA
#> 69                Openness8 0.1661168 0.0070050 0.1536792 0.1813385 NA
#> 70                Openness9 0.1837819 0.0060944 0.1705079 0.1946385 NA
#> 71               Openness10 0.1600255 0.0071424 0.1441736 0.1730915 NA
#> 72                                 NA        NA        NA        NA NA
#> 
#> $Loadings
#>                          V1  Estimate        SE  95%CI(L)  95%CI(U) V6
#> 1       Networking Behavior        NA        NA        NA        NA NA
#> 2                 Behavior1 0.9046829 0.0148949 0.8715773 0.9271598 NA
#> 3                 Behavior2 0.7994980 0.0220873 0.7471305 0.8337132 NA
#> 4                 Behavior3 0.8749911 0.0173729 0.8392425 0.9045295 NA
#> 5                 Behavior5 0.7296377 0.0229329 0.6851776 0.7770384 NA
#> 6                 Behavior7 0.7380472 0.0364928 0.6663535 0.8007315 NA
#> 7                 Behavior8 0.9124444 0.0080646 0.8982692 0.9276207 NA
#> 8                 Behavior9 0.9131622 0.0076771 0.8976023 0.9279955 NA
#> 9  Number of job interviews        NA        NA        NA        NA NA
#> 10               Interview1 0.9073432 0.0070093 0.8932917 0.9225212 NA
#> 11               Interview2 0.9034311 0.0072442 0.8887800 0.9169500 NA
#> 12     Number of job offers        NA        NA        NA        NA NA
#> 13                   Offer1 0.8814935 0.0108241 0.8587240 0.9032240 NA
#> 14                   Offer2 0.8798265 0.0110266 0.8578285 0.9013256 NA
#> 15         Honesty-Humility        NA        NA        NA        NA NA
#> 16                 Honesty1 0.6726453 0.0289471 0.6181610 0.7289371 NA
#> 17                 Honesty2 0.6729669 0.0266330 0.6231794 0.7213007 NA
#> 18                 Honesty3 0.6940607 0.0216417 0.6541770 0.7430362 NA
#> 19                 Honesty4 0.7346629 0.0255258 0.6864575 0.7839045 NA
#> 20                 Honesty5 0.6565293 0.0258370 0.6023431 0.7048155 NA
#> 21                 Honesty6 0.6860204 0.0258578 0.6230655 0.7269823 NA
#> 22                 Honesty7 0.7072245 0.0263630 0.6568493 0.7572211 NA
#> 23                 Honesty8 0.7424046 0.0202610 0.7041661 0.7812717 NA
#> 24                 Honesty9 0.7854853 0.0178506 0.7518543 0.8193960 NA
#> 25                Honesty10 0.6976471 0.0281672 0.6472320 0.7484519 NA
#> 26             Emotionality        NA        NA        NA        NA NA
#> 27                 Emotion1 0.7916473 0.0177419 0.7574219 0.8202426 NA
#> 28                 Emotion2 0.8142535 0.0177223 0.7772647 0.8455244 NA
#> 29                 Emotion3 0.6296732 0.0317257 0.5707160 0.6903520 NA
#> 30                 Emotion4 0.6364922 0.0282551 0.5755661 0.6924075 NA
#> 31                 Emotion5 0.7523806 0.0230923 0.7095760 0.7954332 NA
#> 32                 Emotion6 0.6073636 0.0309599 0.5359383 0.6596498 NA
#> 33                 Emotion8 0.7536960 0.0212556 0.7111993 0.7930121 NA
#> 34                Emotion10 0.7143856 0.0275660 0.6565207 0.7673382 NA
#> 35             Extraversion        NA        NA        NA        NA NA
#> 36                Extraver2 0.6993185 0.0322083 0.6283271 0.7636305 NA
#> 37                Extraver3 0.6060653 0.0386622 0.5343619 0.6768038 NA
#> 38                Extraver4 0.7305910 0.0300952 0.6691501 0.7775481 NA
#> 39                Extraver5 0.7143110 0.0294604 0.6516207 0.7673753 NA
#> 40                Extraver6 0.7397646 0.0218056 0.6952885 0.7788172 NA
#> 41                Extraver7 0.7204402 0.0266672 0.6699985 0.7671337 NA
#> 42                Extraver8 0.7142999 0.0298465 0.6534033 0.7758096 NA
#> 43                Extraver9 0.7900990 0.0174506 0.7497601 0.8165177 NA
#> 44               Extraver10 0.8106963 0.0229368 0.7607017 0.8503707 NA
#> 45            Agreeableness        NA        NA        NA        NA NA
#> 46               Agreeable1 0.8151671 0.0211218 0.7765273 0.8571380 NA
#> 47               Agreeable3 0.6019611 0.0495275 0.5086651 0.6816085 NA
#> 48               Agreeable4 0.6920310 0.0379855 0.6211914 0.7669604 NA
#> 49               Agreeable5 0.9020809 0.0149907 0.8614063 0.9206177 NA
#> 50               Agreeable7 0.8813253 0.0156511 0.8599418 0.9175347 NA
#> 51               Agreeable8 0.6553498 0.0329894 0.5838173 0.7060341 NA
#> 52               Agreeable9 0.6570083 0.0313352 0.5973889 0.7182614 NA
#> 53              Agreeable10 0.6855452 0.0383433 0.6185974 0.7514370 NA
#> 54        Conscientiousness        NA        NA        NA        NA NA
#> 55           Conscientious1 0.7861901 0.0246077 0.7249477 0.8238033 NA
#> 56           Conscientious3 0.5685257 0.0287022 0.5203846 0.6152199 NA
#> 57           Conscientious4 0.7604740 0.0219012 0.7216313 0.8053899 NA
#> 58           Conscientious6 0.7202210 0.0273935 0.6663930 0.7694239 NA
#> 59           Conscientious7 0.8511911 0.0139611 0.8213115 0.8738885 NA
#> 60           Conscientious8 0.7946675 0.0188292 0.7550659 0.8254577 NA
#> 61           Conscientious9 0.6606040 0.0356693 0.5932370 0.7407418 NA
#> 62          Conscientious10 0.6110580 0.0399238 0.5442812 0.6813451 NA
#> 63   Openness to Experience        NA        NA        NA        NA NA
#> 64                Openness1 0.6776794 0.0271176 0.6141487 0.7216211 NA
#> 65                Openness2 0.7037400 0.0262765 0.6571034 0.7642079 NA
#> 66                Openness3 0.6828585 0.0304850 0.6314594 0.7419502 NA
#> 67                Openness5 0.7206599 0.0206272 0.6719613 0.7568806 NA
#> 68                Openness7 0.8174478 0.0185773 0.7792504 0.8547072 NA
#> 69                Openness8 0.6646166 0.0292087 0.6195105 0.7281679 NA
#> 70                Openness9 0.7352702 0.0242531 0.6867222 0.7799879 NA
#> 71               Openness10 0.6402401 0.0291230 0.5933387 0.7049633 NA
#> 72                                 NA        NA        NA        NA NA
#> 
#> $`Path coefficients`
#>                                              V1 V2   Estimate        SE
#> 1         Honesty-Humility->Networking Behavior NA  0.1361644 0.0309411
#> 2             Emotionality->Networking Behavior NA -0.4311118 0.0340716
#> 3             Extraversion->Networking Behavior NA  0.4051107 0.0362823
#> 4            Agreeableness->Networking Behavior NA  0.1268461 0.0336086
#> 5        Conscientiousness->Networking Behavior NA  0.0818501 0.0284513
#> 6   Openness to Experience->Networking Behavior NA  0.1463257 0.0273398
#> 7 Networking Behavior->Number of job interviews NA  0.1310967 0.0361253
#> 8     Networking Behavior->Number of job offers NA  0.0948720 0.0363889
#> 9                                               NA         NA        NA
#>     95%CI(L)   95%CI(U)
#> 1  0.0823789  0.1988008
#> 2 -0.4975993 -0.3714208
#> 3  0.3317080  0.4769753
#> 4  0.0665760  0.2012249
#> 5  0.0312144  0.1347351
#> 6  0.0948065  0.2024902
#> 7  0.0672806  0.2064333
#> 8  0.0209743  0.1622576
#> 9         NA         NA
#> 
#> $`Unique D^2`
#>    Honesty1  Honesty2  Honesty3  Honesty4  Honesty5 Honesty6  Honesty7
#> 1 0.5273984 0.5306635 0.4787401 0.4222894 0.5405153 0.504849 0.4636667
#>    Honesty8  Honesty9 Honesty10 Emotion1 Emotion2  Emotion3  Emotion4  Emotion5
#> 1 0.4133138 0.3626471 0.4813632 0.355008 0.325329 0.5879229 0.5801572 0.4234307
#>    Emotion6  Emotion8 Emotion10 Extraver2 Extraver3 Extraver4 Extraver5
#> 1 0.6114926 0.4203239 0.4667098 0.4872904 0.6124864  0.443869 0.4627561
#>   Extraver6 Extraver7 Extraver8 Extraver9 Extraver10 Agreeable1 Agreeable3
#> 1 0.4222565 0.4668947 0.4622508 0.3604529  0.3234635  0.3152288  0.6182034
#>   Agreeable4 Agreeable5 Agreeable7 Agreeable8 Agreeable9 Agreeable10
#> 1  0.5026326  0.1692086  0.2094651   0.555478  0.5351004   0.4995825
#>   Conscientious1 Conscientious3 Conscientious4 Conscientious6 Conscientious7
#> 1      0.3623469      0.6460999      0.4025884      0.4577823      0.2557225
#>   Conscientious8 Conscientious9 Conscientious10 Openness1 Openness2 Openness3
#> 1      0.3450169      0.5441084        0.605072 0.5260668  0.493418 0.5147944
#>   Openness5 Openness7 Openness8 Openness9 Openness10 V52
#> 1 0.4682718 0.3139757 0.5436591 0.4486768  0.5695222  NA
#>