Internal: Reliability

Compute several reliability measures. See the vignette for details.

calculateRhoC(.object = NULL, .only_common_factors = TRUE)

calculateRhoT(.object = NULL, .alpha = args_default()$.alpha,
  .calculate_ci = args_default()$.calculate_ci,
  .only_common_factors = TRUE)

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.only_common_factors
Logical. Should only constructs modeled as common factors be included? Defaults to TRUE.

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.calculate_ci
Logical. Should a closed-form confidence interval be computed? Defaults to FALSE.

Value

For calculateRhoC() a named numeric vector containing the reliability estimates. For calculateRhoT() a data.frame with as many rows as there are constructs modeled as common factors in the model (unless .only_common_factors = FALSE in which case the number of rows equals the total number of constructs in the model). The first column contains the name of the construct. The second column the reliability estimate. If .calculate_ci = TRUE the remaining columns contain lower and upper bounds for the (1 - .alpha) confidence interval(s).

Details

Reliability is the consistency of measurement. Practically, reliability is empirically assessed based on the classical true score measurement theory and defined as the ratio of the true score variance relative to the proxy (test score) variance, where the later is a weighted linear combinations of the indicators (i.e. a proxy or stand-in for the true score).

The current literatue provides numerous reliability measures, based on (seemingly) different formulae with no systematic naming conventions. This package follows Cho (2016) 's attempt for a systematic approach to reliability by adopting his proposed unified naming convention. See Cho (2016) for details.

Since reliability is defined with respect to a classical true score measurement model only constructs modeled as common factors are considered by default. For constructs modeled as composites reliability may be estimated by setting .only_common_factors = FALSE, however, it is unclear how to interpret reliability in this case.

For the the tau-equivalent reliability (rhoT) a closed-form confidence interval may be computed (Trinchera et al. 2018) by setting .calculate_ci = TRUE (default is FALSE). If .alpha is a vector several CI's are returned.

The function is only applicable to objects of class cSEMResults_default. For other object classes use assess().

Functions

calculateRhoC: Calculate the congeneric reliability, also known as composite reliability or rho_A.

calculateRhoT: Calculate the tau-equivalent reliability, also known as Cronbach alpha or coefficient alpha. Since indicators are always standarized in cSEM, tau-equivalent reliability is identical to the parallel reliability.

References

Cho E (2016). “Making Reliability Reliable.” Organizational Research Methods, 19(4), 651--682. doi: 10.1177/1094428116656239 . Trinchera L, Marie N, Marcoulides GA (2018). “A Distribution Free Interval Estimate for Coefficient Alpha.” Structural Equation Modeling: A Multidisciplinary Journal, 25(6), 876--887. doi: 10.1080/10705511.2018.1431544 .

See also