savePlot

This function saves a given plot of a cSEMResults object to a specified file format.

savePlot(
 .plot_object,
 .filename,
 .path = NULL)

Arguments

.plot_object

Object returned by one of the following functions plot.cSEMResults_default(), plot.cSEMResults_multi(), or plot.cSEMResults_2ndorder().

.filename

Character string. The name of the file to save the plot to (supports 'pdf', 'png', 'svg', and 'dot' formats).

.path

Character string. Path of the directory to save the file to. Defaults to the current working directory.

See also