tidy Method for cSEMResults Objects

This method and documentation is based on the tidy method for lavaan, as found in broom:::tidy.lavaan().

# S3 method for class 'cSEMResults'
tidy(
  x,
  parameters = c("all", "Path_estimates", "Loadings_estimates", "Weight_estimates", "D2",
    "Effect_estimates", "Indicator_correlation", "Exo_construct_correlation"),
  ...
)

Arguments

x

Fitted csem model with class cSEMResults as returned from csem()

parameters

Character string. Selects the desired parameter estimates as outputted from summarize(). Defaults to all.

...

Arguments passed on to summarize

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

.object

An R object of class cSEMResults resulting from a call to csem().

Value

A data.frame with one row for each estimated parameter. The description of the column names are as follows:

term

The result of paste(lhs, op, rhs)

op

The operator in the model syntax (e.g., ~~ for correlation, <~ for weights, =~ for loadings, and ~ for path estimates)

estimate

The parameter estimate

construct.type

Whether the modelled construct is a common factor or composite variable

std.error

statistic

The t-statistic as returned from summarize() and cSEM:::addInfer()

p_value

Details

Cautionary Note: As with all tidy() methods, mis-specified arguments may be silently ignored.