Update Pseudo Weights for Composites and Common-Factors

Computation of WW matrix was converted between Matlab to R based on page 44 of Hiebeler (2015) .

update_X_WW_pseudo_weights(Z, U, D, C, n_constructs, B)

Arguments

Z

Standardized data matrix

U

Matrix of estimates related to unique error

D

Matrix of estimated unique error

C

Matrix of estimated loadings

n_constructs

Number of constructs

B

Matrix of path coefficients

Value

Two matrices:

X: Remaining part of data (Z) after accounting for uniqueness terms (U) and (D), used for estimating composite loadings. Also used for standardizing theta when updating Gamma, W and V

WW: Weights after accounting for current Loading and Path-Coefficients values, used for estimating common-factor loadings

References

Hiebeler DE (2015). R and MATLAB. CRC Press.