Data: BergamiBagozzi2000

A data frame containing 22 variables with 305 observations.

BergamiBagozzi2000

Format

An object of class data.frame with 305 rows and 22 columns.

Source

Survey among South Korean employees conducted and reported by Bergami and Bagozzi (2000) .

Details

The dataset contains 22 variables and originates from a larger survey among South Korean employees conducted and reported by Bergami and Bagozzi (2000) . It is also used in Hwang and Takane (2014) and Henseler (2020) for demonstration purposes, see the corresponding tutorial.

References

Bergami M, Bagozzi RP (2000). “Self-categorization, affective commitment and group self-esteem as distinct aspects of social identity in the organization.” British Journal of Social Psychology, 39, 555--577. doi: 10.1348/014466600164633 .

Henseler J (2020). Composite-Based Structural Equation Modeling: An Introduction to Partial Least Squares & Co. Using ADANCO. Guilford Press.

Hwang H, Takane Y (2014). Generalized Structured Component Analysis: A Component-Based Approach to Structural Equation Modeling. Chapman and Hall/CRC.

Examples

#============================================================================
# Example is taken from Henseler (2020)
#============================================================================
model_Bergami_Bagozzi="
# Measurement models
OrgPres =~ cei1 + cei2 + cei3 + cei4 + cei5 + cei6 + cei7 + cei8 
OrgIden =~ ma1 + ma2 + ma3 + ma4 + ma5 + ma6
AffLove =~ orgcmt1 + orgcmt2 + orgcmt3 + orgcmt7
AffJoy  =~ orgcmt5 + orgcmt8
Gender  <~ gender

# Structural model 
OrgIden ~ OrgPres
AffLove ~ OrgPres + OrgIden + Gender 
AffJoy  ~ OrgPres + OrgIden + Gender 
"

out <- csem(.data = BergamiBagozzi2000, .model = model_Bergami_Bagozzi,
            .PLS_weight_scheme_inner = 'factorial',
            .tolerance = 1e-06
)