Data: BergamiBagozzi2000

A data frame containing 22 variables with 305 observations.

BergamiBagozzi2000

Format

An object of class data.frame with 305 rows and 22 columns.

Source

Survey among South Korean employees conducted and reported by Bergami and Bagozzi (2000) .

Details

The dataset contains 22 variables and originates from a larger survey among South Korean employees conducted and reported by Bergami and Bagozzi (2000) . It is also used in Hwang and Takane (2004) and Henseler (2021) for demonstration purposes, see the corresponding tutorial.

References

Bergami M, Bagozzi RP (2000). “Self-categorization, affective commitment and group self-esteem as distinct aspects of social identity in the organization.” British Journal of Social Psychology, 39(4), 555–577. doi:10.1348/014466600164633 .

Henseler J (2021). Composite-Based Structural Equation Modeling: Analyzing Latent and Emergent Variables. Guilford Press, New York.

Hwang H, Takane Y (2004). “Generalized Structured Component Analysis.” Psychometrika, 69(1), 81–99.

Examples

#============================================================================
# Example is taken from Henseler (2021)
#============================================================================
model_Bergami_Bagozzi_Henseler="
# Measurement models
OrgPres =~ cei1 + cei2 + cei3 + cei4 + cei5 + cei6 + cei7 + cei8 
OrgIden =~ ma1 + ma2 + ma3 + ma4 + ma5 + ma6
AffLove =~ orgcmt1 + orgcmt2 + orgcmt3 + orgcmt7
AffJoy  =~ orgcmt5 + orgcmt8
Gender  <~ gender

# Structural model 
OrgIden ~ OrgPres
AffLove ~ OrgPres + OrgIden + Gender 
AffJoy  ~ OrgPres + OrgIden + Gender 
"

out <- csem(.data = BergamiBagozzi2000, 
            .model = model_Bergami_Bagozzi_Henseler,
            .PLS_weight_scheme_inner = 'factorial',
            .tolerance = 1e-06
)

#============================================================================
# Example is taken from Hwang et al. (2004)
#============================================================================ 

model_Bergami_Bagozzi_Hwang="
# Measurement models
OrgPres =~ cei1 + cei2 + cei3 + cei4 + cei5 + cei6 + cei7 + cei8 
OrgIden =~ ma1 + ma2 + ma3 + ma4 + ma5 + ma6
AffJoy =~ orgcmt1 + orgcmt2 + orgcmt3 + orgcmt7
AffLove  =~ orgcmt5 + orgcmt6 + orgcmt8

# Structural model 
OrgIden ~ OrgPres 
AffLove ~ OrgIden
AffJoy  ~ OrgIden"

out_Hwang <- csem(.data = BergamiBagozzi2000, 
                 .model = model_Bergami_Bagozzi_Hwang,
                 .approach_weights = "GSCA",
                 .disattenuate = FALSE,
                 .id = "gender",
                 .tolerance = 1e-06) 
#> Error in (function (.data = args_default()$.data, .model = args_default()$.model,     .approach_cor_robust = args_default()$.approach_cor_robust,     .approach_nl = args_default()$.approach_nl, .approach_paths = args_default()$.approach_paths,     .approach_weights = args_default()$.approach_weights, .conv_criterion = args_default()$.conv_criterion,     .disattenuate = args_default()$.disattenuate, .dominant_indicators = args_default()$.dominant_indicators,     .estimate_structural = args_default()$.estimate_structural,     .id = args_default()$.id, .instruments = args_default()$.instruments,     .iter_max = args_default()$.iter_max, .normality = args_default()$.normality,     .PLS_approach_cf = args_default()$.PLS_approach_cf, .PLS_ignore_structural_model = args_default()$.PLS_ignore_structural_model,     .PLS_modes = args_default()$.PLS_modes, .PLS_weight_scheme_inner = args_default()$.PLS_weight_scheme_inner,     .reliabilities = args_default()$.reliabilities, .starting_values = args_default()$.starting_values,     .tolerance = args_default()$.tolerance) {    args_used <- c(as.list(environment(), all.names = TRUE))    csem_model <- parseModel(.model, .instruments = .instruments)    X_cleaned <- processData(.data = .data, .model = csem_model,         .instruments = NULL)    Cor <- calculateIndicatorCor(.X_cleaned = X_cleaned, .approach_cor_robust = .approach_cor_robust)    S <- Cor$S    X <- scale(data.matrix(X_cleaned))    if (.approach_weights == "PLS-PM") {        W <- calculateWeightsPLS(.S = S, .csem_model = csem_model,             .iter_max = .iter_max, .PLS_modes = .PLS_modes, .PLS_ignore_structural_model = .PLS_ignore_structural_model,             .PLS_weight_scheme_inner = .PLS_weight_scheme_inner,             .data = X, .conv_criterion = .conv_criterion, .tolerance = .tolerance,             .starting_values = .starting_values)    }    else if (.approach_weights %in% c("SUMCORR", "MAXVAR", "SSQCORR",         "MINVAR", "GENVAR")) {        W <- calculateWeightsKettenring(.S = S, .csem_model = csem_model,             .approach_gcca = .approach_weights)    }    else if (.approach_weights == "GSCA") {        if (csem_model$model_type == "Nonlinear") {            stop2("cSEM currently does not support GSCA and GSCAm for models containing nonlinear terms.")        }        if (.disattenuate & all(csem_model$construct_type ==             "Common factor")) {            W <- calculateWeightsGSCAm(.X = X, .csem_model = csem_model,                 .conv_criterion = .conv_criterion, .iter_max = .iter_max,                 .tolerance = .tolerance, .starting_values = .starting_values)            W$W <- scaleWeights(S, W$W)        }        else if (all(csem_model$construct_type == "Composite")) {            W <- calculateWeightsGSCA(.X = X, .S = S, .csem_model = csem_model,                 .conv_criterion = .conv_criterion, .iter_max = .iter_max,                 .tolerance = .tolerance, .starting_values = .starting_values)        }    }    else if (.approach_weights == "IGSCA") {        W <- calculateWeightsIGSCA(.data = X_cleaned, .csem_model = csem_model,             .tolerance = .tolerance, .iter_max = .iter_max, .dominant_indicators = .dominant_indicators)        return(W)    }    else if (.approach_weights == "unit") {        W <- calculateWeightsUnit(.S = S, .csem_model = csem_model,             .starting_values = .starting_values)    }    else if (.approach_weights %in% c("bartlett", "regression")) {        W <- list(W = csem_model$measurement, E = NULL, Modes = NULL,             Conv_status = NULL, Iterations = 0)    }    else if (.approach_weights == "PCA") {        W <- calculateWeightsPCA(.S = S, .csem_model = csem_model)    }    if (!is.null(.dominant_indicators)) {        W$W <- setDominantIndicator(.W = W$W, .dominant_indicators = .dominant_indicators,             .S = S)    }    LambdaQ2W <- calculateReliabilities(.X = X, .S = S, .W = W,         .approach_weights = .approach_weights, .csem_model = csem_model,         .disattenuate = .disattenuate, .PLS_approach_cf = .PLS_approach_cf,         .reliabilities = .reliabilities)    Weights <- LambdaQ2W$W    Lambda <- LambdaQ2W$Lambda    Q <- sqrt(LambdaQ2W$Q2)    args_used$.disattenuate <- LambdaQ2W$.disattenuate    Theta <- S - t(Lambda) %*% Lambda    H <- X %*% t(Weights)    C <- calculateCompositeVCV(.S = S, .W = Weights)    P <- calculateConstructVCV(.C = C, .Q = Q)    if (sum(rowSums(csem_model$structural)) == 0) {        .estimate_structural <- FALSE    }    if (.estimate_structural) {        estim_results <- estimatePath(.approach_nl = .approach_nl,             .approach_paths = .approach_paths, .csem_model = csem_model,             .H = H, .normality = .normality, .P = P, .Q = Q)    }    else {        estim_results <- NULL    }    out <- list(Estimates = list(Path_estimates = if (.estimate_structural) {        estim_results$Path_estimates    } else {        estim_results    }, Loading_estimates = Lambda, Weight_estimates = Weights,         Inner_weight_estimates = W$E, Residual_correlation = Theta,         Construct_scores = H, Indicator_VCV = S, Proxy_VCV = C,         Construct_VCV = P, Reliabilities = Q^2, R2 = if (.estimate_structural) {            estim_results$R2        } else {            estim_results        }, R2adj = if (.estimate_structural) {            estim_results$R2adj        } else {            estim_results        }, VIF = if (.estimate_structural) {            estim_results$VIF        } else {            estim_results        }, SE = if (.estimate_structural) {            estim_results$SE        } else {            estim_results        }), Information = list(Data = X, Model = csem_model,         Arguments = args_used, Type_of_indicator_correlation = Cor$cor_type,         Threshold_parameter_estimates = Cor$thres_est, Weight_info = list(Modes = W$Modes,             Number_iterations = W$Iterations, Convergence_status = W$Conv_status),         Approach_2ndorder = NA))    class(out) <- c("cSEMResults", "cSEMResults_default")    invisible(out)})(.PLS_approach_cf = "dist_squared_euclid", .PLS_ignore_structural_model = FALSE,     .PLS_modes = NULL, .PLS_weight_scheme_inner = "path", .approach_cor_robust = "none",     .approach_nl = "sequential", .approach_paths = "OLS", .approach_weights = "GSCA",     .conv_criterion = "diff_absolute", .data = structure(list(        cei1 = c(3, 4, 5, 4, 4, 4, 4, 4, 4, 4, 5, 4, 4, 4, 4,         2, 4, 5, 4, 4, 5, 5, 1, 5, 5, 5, 4, 4, 4, 4, 4, 4, 1,         4, 4, 4, 5, 4, 4, 5, 5, 4, 4, 3, 5, 5, 4, 4, 5, 5, 5,         5, 5, 5, 5, 5, 4, 5, 5, 4, 2, 5, 4, 5, 4, 4, 5, 4, 4,         4, 2, 5, 4, 4, 4, 5, 4, 5, 4, 5, 4, 5, 4, 5, 5, 5, 4,         4, 5, 4, 5, 4, 4, 4, 4, 5, 3, 4, 5, 5, 4, 4, 5, 5, 5,         5, 5, 4, 3, 5, 4, 5, 4, 4, 5, 4, 5, 5, 5, 5, 3, 4, 4,         5, 5, 5, 5, 5, 4, 4, 5, 4, 5, 5, 4, 5, 5, 4, 5, 4, 5,         4, 3, 4, 4, 4, 2, 4, 5, 5, 4, 5, 5, 4, 4, 5, 5), cei2 = c(3,         5, 5, 4, 4, 4, 4, 3, 4, 5, 4, 4, 4, 4, 4, 2, 4, 4, 4,         4, 4, 5, 3, 5, 4, 5, 4, 4, 4, 4, 4, 4, 1, 5, 4, 4, 4,         4, 4, 5, 5, 3, 4, 1, 5, 5, 4, 4, 5, 4, 5, 4, 5, 4, 5,         4, 4, 5, 5, 4, 3, 5, 2, 3, 4, 4, 5, 4, 4, 4, 3, 5, 3,         3, 3, 5, 3, 4, 4, 5, 3, 5, 4, 4, 5, 5, 3, 3, 4, 4, 5,         3, 3, 4, 3, 3, 3, 4, 5, 4, 4, 4, 5, 4, 4, 4, 5, 4, 4,         5, 4, 4, 4, 4, 4, 4, 4, 4, 3, 4, 3, 5, 4, 4, 4, 5, 4,         4, 4, 3, 4, 4, 4, 3, 4, 5, 4, 4, 4, 4, 5, 4, 4, 3, 4,         4, 3, 4, 5, 4, 4, 4, 4, 4, 4, 5, 3), cei3 = c(3, 3, 4,         4, 4, 4, 3, 3, 4, 3, 4, 4, 4, 3, 3, 3, 3, 4, 4, 3, 4,         4, 3, 5, 4, 5, 4, 3, 3, 3, 4, 4, 1, 4, 4, 3, 4, 4, 4,         5, 5, 3, 3, 2, 5, 5, 4, 4, 4, 4, 5, 4, 5, 4, 5, 3, 4,         5, 5, 4, 3, 4, 2, 3, 3, 4, 5, 3, 4, 3, 2, 5, 5, 2, 3,         3, 4, 4, 4, 4, 3, 3, 3, 5, 5, 3, 4, 3, 4, 4, 5, 3, 3,         4, 2, 2, 3, 4, 4, 4, 3, 4, 4, 3, 4, 3, 4, 5, 4, 4, 3,         4, 5, 4, 4, 4, 3, 4, 3, 3, 4, 5, 3, 4, 4, 5, 4, 3, 4,         4, 4, 4, 4, 3, 4, 4, 4, 4, 4, 4, 5, 3, 4, 3, 3, 4, 3,         3, 5, 3, 3, 3, 4, 4, 4, 5, 3), cei4 = c(4, 4, 4, 4, 4,         4, 5, 4, 4, 5, 5, 4, 5, 4, 4, 3, 4, 4, 4, 4, 4, 5, 2,         5, 4, 5, 4, 4, 5, 5, 4, 4, 1, 5, 4, 4, 4, 4, 4, 5, 5,         4, 4, 5, 5, 5, 4, 4, 5, 4, 5, 4, 5, 5, 5, 5, 4, 5, 5,         4, 3, 4, 4, 4, 3, 4, 5, 4, 4, 4, 3, 5, 4, 4, 4, 5, 4,         5, 4, 5, 4, 4, 4, 4, 5, 5, 4, 3, 5, 4, 5, 4, 4, 4, 4,         4, 3, 4, 4, 5, 4, 4, 5, 5, 4, 5, 5, 5, 4, 5, 4, 5, 5,         4, 4, 4, 5, 4, 4, 4, 5, 5, 4, 5, 5, 5, 5, 4, 4, 4, 4,         4, 4, 5, 4, 4, 5, 4, 4, 4, 5, 3, 4, 3, 4, 4, 2, 4, 5,         5, 4, 4, 5, 4, 4, 5, 3), cei5 = c(3, 4, 5, 4, 4, 4, 5,         4, 4, 5, 5, 4, 5, 4, 5, 3, 4, 4, 4, 4, 5, 5, 1, 5, 3,         4, 4, 4, 4, 4, 4, 4, 1, 5, 4, 4, 4, 4, 4, 5, 5, 4, 4,         4, 5, 5, 4, 4, 5, 4, 5, 4, 5, 5, 5, 5, 4, 5, 5, 4, 3,         5, 4, 5, 4, 4, 5, 4, 4, 5, 1, 4, 4, 4, 4, 5, 4, 5, 5,         5, 4, 4, 4, 5, 5, 5, 4, 4, 4, 4, 5, 5, 4, 4, 4, 5, 3,         4, 4, 5, 4, 4, 5, 5, 4, 5, 5, 4, 3, 5, 4, 5, 5, 4, 4,         5, 5, 5, 5, 5, 5, 5, 4, 5, 5, 5, 5, 5, 4, 4, 5, 4, 5,         5, 4, 4, 5, 4, 4, 3, 3, 4, 3, 4, 2, 4, 2, 4, 5, 5, 4,         4, 5, 4, 4, 5, 4), cei6 = c(3, 3, 4, 4, 4, 4, 3, 3, 4,         3, 4, 4, 5, 3, 5, 2, 4, 4, 4, 4, 4, 5, 2, 5, 3, 4, 4,         4, 3, 3, 4, 4, 1, 4, 4, 4, 4, 4, 4, 5, 3, 4, 4, 3, 5,         5, 4, 3, 4, 3, 4, 4, 5, 4, 5, 4, 4, 5, 5, 4, 3, 5, 2,         2, 4, 4, 5, 4, 4, 3, 2, 4, 3, 2, 4, 5, 4, 5, 4, 4, 4,         4, 4, 4, 5, 5, 3, 3, 3, 4, 4, 4, 3, 4, 5, 3, 3, 4, 4,         5, 3, 4, 5, 4, 4, 4, 4, 4, 3, 5, 4, 4, 4, 4, 4, 4, 4,         3, 4, 4, 3, 5, 4, 5, 4, 5, 4, 4, 4, 4, 4, 4, 4, 3, 4,         4, 4, 4, 4, 3, 4, 4, 4, 4, 3, 4, 2, 4, 5, 5, 3, 4, 5,         4, 4, 5, 4), cei7 = c(1, 3, 4, 4, 4, 4, 3, 3, 4, 3, 4,         4, 5, 4, 4, 3, 4, 4, 4, 3, 3, 5, 2, 5, 3, 4, 4, 3, 3,         3, 3, 4, 1, 4, 4, 4, 4, 4, 4, 5, 3, 4, 3, 2, 5, 5, 4,         2, 4, 3, 4, 4, 5, 4, 5, 3, 4, 5, 4, 4, 3, 4, 3, 3, 3,         3, 5, 4, 4, 3, 3, 4, 4, 2, 3, 3, 4, 5, 4, 3, 4, 3, 4,         5, 5, 3, 4, 4, 3, 4, 4, 4, 3, 4, 4, 2, 3, 4, 4, 5, 3,         4, 4, 3, 4, 3, 4, 4, 3, 5, 3, 4, 4, 4, 4, 4, 2, 3, 3,         3, 3, 5, 4, 4, 5, 5, 4, 3, 4, 4, 4, 5, 4, 3, 4, 4, 4,         4, 4, 3, 3, 3, 4, 4, 2, 4, 2, 3, 5, 5, 3, 4, 5, 4, 4,         5, 4), cei8 = c(3, 3, 5, 4, 4, 4, 4, 4, 4, 4, 5, 4, 5,         4, 5, 3, 4, 4, 4, 4, 4, 5, 2, 5, 4, 4, 4, 4, 4, 2, 4,         4, 1, 5, 4, 4, 4, 4, 4, 5, 4, 4, 4, 1, 5, 5, 4, 3, 4,         4, 5, 4, 5, 4, 5, 5, 4, 5, 5, 4, 3, 4, 4, 3, 3, 4, 5,         4, 4, 5, 4, 4, 4, 3, 4, 5, 4, 5, 5, 5, 4, 4, 4, 4, 5,         5, 4, 4, 4, 4, 3, 4, 4, 4, 4, 4, 3, 4, 4, 5, 4, 4, 5,         5, 4, 4, 5, 4, 3, 5, 4, 4, 5, 4, 4, 4, 4, 4, 4, 3, 3,         5, 4, 5, 5, 5, 4, 4, 4, 4, 4, 5, 4, 5, 4, 5, 5, 4, 4,         4, 3, 3, 4, 4, 3, 4, 2, 4, 5, 5, 4, 4, 5, 4, 4, 5, 4),         ma1 = c(4, 2, 3, 4, 4, 3, 3, 4, 4, 3, 4, 3, 4, 4, 4,         3, 4, 4, 4, 4, 4, 2, 3, 3, 4, 4, 2, 3, 3, 3, 4, 3, 1,         4, 3, 3, 4, 4, 4, 4, 5, 3, 4, 2, 3, 4, 4, 4, 4, 4, 5,         4, 5, 5, 3, 4, 4, 4, 4, 4, 4, 2, 3, 4, 4, 4, 2, 4, 3,         4, 2, 2, 2, 2, 3, 5, 4, 3, 4, 4, 3, 3, 3, 5, 5, 4, 4,         4, 4, 4, 5, 4, 4, 4, 4, 4, 3, 4, 5, 3, 3, 4, 4, 2, 5,         4, 4, 5, 4, 3, 4, 4, 5, 4, 4, 5, 4, 2, 3, 2, 3, 5, 4,         4, 5, 4, 3, 4, 3, 3, 5, 4, 2, 4, 4, 4, 5, 4, 4, 4, 4,         4, 4, 4, 1, 4, 4, 4, 4, 4, 5, 4, 3, 3, 4, 5, 4), ma2 = c(4,         2, 2, 4, 4, 3, 4, 4, 4, 4, 4, 3, 4, 4, 4, 3, 4, 3, 4,         4, 4, 3, 3, 3, 4, 4, 4, 3, 2, 4, 4, 2, 2, 4, 3, 3, 4,         4, 4, 4, 5, 4, 4, 1, 1, 5, 4, 4, 5, 4, 5, 4, 5, 5, 3,         4, 4, 4, 4, 3, 3, 3, 1, 5, 4, 4, 4, 2, 2, 3, 3, 3, 2,         2, 3, 5, 3, 4, 4, 4, 4, 4, 4, 4, 5, 3, 3, 3, 4, 4, 5,         5, 3, 5, 4, 4, 3, 4, 4, 4, 3, 4, 4, 3, 4, 4, 4, 4, 3,         4, 3, 4, 4, 3, 2, 4, 4, 2, 3, 3, 4, 5, 4, 4, 5, 5, 2,         4, 4, 4, 4, 4, 3, 4, 3, 4, 4, 4, 5, 3, 4, 4, 3, 4, 1,         4, 5, 3, 5, 4, 5, 4, 3, 4, 4, 5, 4), ma3 = c(4, 3, 4,         4, 4, 4, 4, 5, 4, 5, 4, 3, 3, 4, 4, 2, 4, 4, 4, 4, 5,         5, 4, 5, 5, 4, 4, 4, 3, 2, 4, 3, 1, 3, 3, 3, 5, 4, 4,         4, 5, 5, 4, 5, 4, 5, 4, 2, 4, 4, 4, 5, 4, 5, 4, 4, 4,         4, 4, 4, 3, 4, 2, 5, 4, 3, 4, 2, 5, 4, 2, 3, 3, 3, 4,         5, 4, 4, 5, 4, 2, 5, 4, 4, 4, 4, 5, 4, 5, 4, 4, 5, 4,         5, 5, 3, 3, 4, 5, 5, 3, 5, 4, 3, 5, 5, 4, 5, 5, 4, 4,         3, 5, 4, 4, 4, 3, 3, 4, 4, 3, 4, 4, 4, 4, 4, 2, 4, 4,         4, 5, 4, 4, 4, 4, 4, 5, 5, 4, 4, 5, 5, 5, 5, 3, 4, 5,         4, 5, 4, 4, 3, 4, 4, 4, 5, 3), ma4 = c(4, 3, 3, 4, 4,         3, 3, 4, 5, 4, 4, 3, 4, 3, 4, 3, 3, 3, 4, 4, 4, 4, 3,         5, 4, 3, 4, 3, 1, 4, 4, 3, 2, 5, 3, 3, 5, 4, 4, 4, 4,         3, 3, 2, 2, 4, 4, 3, 3, 4, 3, 5, 4, 4, 3, 4, 4, 4, 4,         3, 3, 2, 1, 5, 3, 4, 3, 3, 3, 3, 3, 1, 2, 2, 2, 5, 4,         4, 4, 4, 3, 4, 3, 4, 5, 3, 3, 4, 4, 4, 5, 5, 4, 4, 3,         3, 2, 4, 4, 3, 2, 3, 4, 1, 3, 5, 4, 5, 4, 4, 3, 2, 4,         5, 4, 4, 4, 2, 2, 2, 4, 5, 4, 4, 5, 4, 1, 3, 4, 4, 4,         3, 2, 3, 4, 4, 5, 4, 5, 4, 5, 4, 4, 4, 1, 4, 4, 4, 4,         4, 5, 3, 3, 1, 4, 5, 2), ma5 = c(3, 2, 2, 4, 4, 3, 4,         4, 4, 4, 4, 3, 3, 3, 3, 5, 4, 4, 4, 4, 5, 3, 3, 5, 4,         4, 4, 3, 1, 3, 4, 4, 1, 4, 3, 3, 4, 3, 3, 4, 3, 4, 3,         1, 3, 4, 4, 3, 4, 4, 4, 4, 4, 4, 3, 4, 4, 4, 4, 3, 3,         1, 2, 5, 4, 4, 4, 4, 3, 3, 2, 1, 3, 2, 2, 5, 5, 4, 5,         4, 3, 3, 3, 5, 5, 3, 3, 4, 3, 3, 5, 5, 4, 4, 4, 4, 2,         4, 3, 2, 1, 2, 4, 1, 3, 4, 3, 4, 4, 3, 3, 3, 4, 4, 4,         4, 2, 1, 2, 2, 3, 5, 4, 4, 5, 4, 2, 2, 4, 4, 4, 4, 2,         3, 4, 4, 5, 4, 4, 4, 5, 3, 4, 5, 1, 4, 4, 3, 4, 4, 5,         4, 3, 1, 4, 5, 3), ma6 = c(4, 3, 4, 4, 4, 2, 2, 4, 4,         4, 4, 3, 4, 4, 4, 4, 4, 4, 5, 4, 4, 3, 4, 4, 4, 4, 4,         3, 3, 3, 4, 3, 1, 2, 3, 3, 4, 3, 3, 5, 5, 4, 4, 4, 1,         4, 4, 3, 3, 3, 5, 4, 5, 4, 4, 4, 4, 4, 4, 4, 4, 3, 4,         4, 4, 4, 4, 4, 3, 3, 1, 3, 3, 2, 3, 5, 5, 3, 4, 4, 3,         3, 3, 4, 5, 3, 4, 3, 4, 4, 5, 4, 4, 4, 3, 4, 2, 4, 3,         3, 3, 3, 4, 2, 4, 4, 4, 4, 4, 3, 3, 4, 4, 4, 4, 4, 4,         2, 4, 3, 4, 5, 4, 4, 4, 5, 3, 4, 4, 3, 5, 4, 3, 4, 4,         4, 4, 4, 4, 4, 5, 4, 4, 3, 1, 4, 5, 3, 4, 3, 5, 4, 4,         4, 4, 5, 3), orgcmt1 = c(3, 3, 1, 3, 3, 2, 3, 3, 3, 4,         4, 3, 3, 3, 4, 2, 3, 4, 3, 3, 3, 3, 4, 5, 4, 4, 3, 2,         1, 2, 4, 3, 2, 4, 4, 3, 4, 4, 4, 4, 3, 1, 3, 4, 2, 4,         4, 2, 3, 3, 3, 4, 4, 2, 3, 3, 4, 3, 4, 3, 3, 3, 1, 5,         3, 4, 3, 2, 3, 3, 3, 2, 3, 2, 2, 4, 4, 3, 3, 3, 1, 3,         2, 5, 5, 3, 3, 4, 4, 3, 5, 2, 3, 5, 3, 2, 2, 3, 3, 2,         1, 3, 3, 1, 4, 5, 4, 4, 2, 4, 3, 3, 3, 3, 3, 4, 3, 2,         3, 2, 3, 4, 3, 4, 5, 4, 2, 4, 3, 2, 5, 1, 2, 3, 4, 4,         4, 3, 3, 3, 4, 2, 3, 3, 2, 3, 1, 3, 3, 4, 4, 3, 4, 4,         3, 3, 2), orgcmt2 = c(2, 1, 3, 3, 3, 2, 4, 3, 4, 4, 4,         3, 4, 3, 3, 1, 3, 3, 4, 4, 3, 2, 3, 4, 3, 4, 4, 3, 1,         4, 4, 3, 2, 4, 4, 2, 4, 4, 4, 3, 4, 4, 4, 5, 2, 4, 4,         2, 3, 2, 3, 2, 4, 3, 3, 2, 4, 4, 3, 3, 3, 2, 2, 5, 4,         4, 4, 3, 4, 3, 2, 2, 4, 2, 3, 4, 3, 4, 2, 3, 2, 2, 2,         3, 3, 3, 2, 3, 4, 3, 3, 4, 3, 3, 3, 2, 2, 3, 4, 3, 1,         2, 3, 2, 3, 4, 3, 4, 3, 4, 4, 3, 3, 2, 3, 3, 4, 2, 3,         3, 2, 3, 3, 4, 4, 4, 2, 2, 3, 3, 3, 2, 2, 4, 4, 4, 4,         3, 4, 3, 4, 3, 3, 3, 1, 4, 3, 3, 3, 3, 4, 4, 3, 2, 4,         4, 3), orgcmt3 = c(3, 3, 3, 2, 3, 2, 3, 4, 4, 4, 4, 3,         3, 3, 4, 3, 3, 3, 4, 4, 3, 2, 3, 4, 4, 3, 4, 2, 1, 3,         4, 3, 2, 4, 4, 2, 4, 4, 4, 4, 4, 3, 3, 5, 2, 4, 4, 2,         3, 4, 4, 4, 4, 4, 3, 4, 4, 3, 3, 2, 4, 3, 1, 5, 4, 4,         3, 3, 3, 4, 3, 2, 3, 2, 2, 4, 4, 4, 4, 4, 3, 3, 2, 3,         5, 3, 5, 3, 4, 3, 5, 5, 3, 3, 2, 4, 2, 4, 3, 3, 1, 5,         4, 2, 4, 4, 4, 4, 3, 3, 3, 3, 4, 4, 3, 4, 3, 1, 3, 4,         1, 3, 4, 4, 5, 4, 3, 3, 3, 4, 4, 3, 2, 4, 4, 4, 4, 3,         4, 3, 4, 2, 3, 3, 2, 4, 2, 3, 4, 4, 4, 3, 3, 2, 4, 5,         2), orgcmt7 = c(4, 3, 4, 3, 3, 2, 3, 3, 4, 4, 4, 3, 3,         3, 4, 2, 3, 4, 3, 3, 3, 4, 3, 4, 3, 3, 4, 2, 2, 3, 4,         3, 1, 4, 3, 2, 4, 4, 4, 3, 5, 4, 3, 2, 1, 5, 4, 3, 4,         4, 4, 4, 4, 4, 4, 4, 4, 3, 4, 3, 4, 3, 3, 4, 3, 4, 4,         3, 3, 4, 2, 3, 3, 2, 3, 4, 2, 4, 4, 3, 4, 3, 3, 2, 5,         4, 4, 3, 4, 3, 5, 5, 3, 3, 3, 2, 2, 4, 3, 3, 2, 3, 3,         3, 4, 4, 4, 4, 2, 4, 4, 3, 3, 3, 3, 4, 4, 2, 4, 3, 4,         4, 4, 4, 5, 4, 4, 2, 3, 4, 4, 1, 3, 3, 4, 4, 4, 3, 4,         3, 5, 3, 4, 4, 3, 4, 3, 3, 4, 5, 4, 4, 4, 3, 4, 4, 3),         orgcmt5 = c(3, 3, 4, 3, 3, 3, 3, 2, 2, 3, 2, 3, 3, 3,         3, 3, 3, 3, 2, 3, 2, 3, 2, 2, 3, 2, 3, 4, 4, 2, 2, 2,         2, 3, 4, 4, 2, 2, 2, 2, 3, 5, 3, 1, 2, 2, 3, 3, 3, 2,         2, 4, 2, 2, 3, 3, 2, 3, 2, 3, 3, 4, 5, 2, 3, 2, 2, 3,         2, 3, 2, 3, 4, 4, 3, 3, 2, 2, 3, 3, 5, 3, 3, 3, 1, 3,         4, 2, 3, 4, 3, 3, 3, 3, 3, 3, 3, 3, 4, 2, 5, 4, 2, 3,         3, 1, 3, 2, 3, 3, 3, 3, 2, 3, 3, 2, 2, 4, 3, 3, 3, 2,         3, 2, 1, 1, 2, 3, 3, 3, 2, 3, 3, 2, 2, 2, 1, 3, 3, 3,         1, 2, 3, 3, 3, 2, 4, 3, 4, 2, 2, 3, 3, 4, 3, 2, 4), orgcmt6 = c(3,         3, 3, 3, 3, 4, 3, 3, 2, 4, 4, 2, 4, 4, 3, 4, 3, 3, 3,         4, 2, 3, 3, 3, 3, 3, 3, 3, 4, 2, 2, 2, 2, 3, 4, 4, 2,         3, 3, 2, 3, 4, 3, 4, 2, 2, 4, 3, 4, 3, 3, 3, 2, 2, 4,         3, 3, 3, 3, 3, 3, 3, 5, 3, 3, 3, 4, 3, 2, 4, 2, 3, 4,         4, 3, 2, 4, 3, 3, 3, 5, 3, 3, 2, 3, 3, 4, 2, 3, 4, 1,         2, 3, 3, 2, 3, 2, 4, 3, 3, 5, 4, 2, 3, 2, 1, 3, 2, 3,         3, 4, 3, 3, 3, 3, 2, 2, 4, 3, 3, 3, 4, 3, 2, 4, 2, 4,         4, 3, 3, 2, 2, 3, 4, 2, 3, 3, 3, 2, 3, 2, 3, 3, 3, 4,         3, 3, 3, 4, 3, 2, 4, 3, 4, 2, 3, 3), orgcmt8 = c(3, 3,         2, 3, 3, 2, 3, 2, 2, 3, 2, 2, 3, 3, 3, 3, 2, 3, 2, 2,         2, 2, 3, 1, 3, 2, 2, 3, 4, 3, 2, 2, 1, 2, 3, 4, 2, 2,         2, 2, 2, 2, 3, 1, 2, 2, 3, 3, 3, 2, 2, 2, 1, 2, 3, 2,         2, 3, 2, 3, 3, 3, 3, 2, 2, 2, 4, 3, 3, 3, 3, 4, 3, 3,         3, 2, 3, 2, 2, 3, 5, 3, 3, 1, 1, 2, 4, 1, 2, 3, 1, 2,         2, 1, 4, 3, 3, 2, 3, 2, 4, 4, 2, 3, 2, 1, 2, 2, 3, 2,         3, 3, 2, 4, 2, 2, 2, 2, 3, 3, 3, 2, 2, 2, 1, 1, 3, 3,         3, 2, 2, 3, 4, 2, 2, 2, 1, 3, 2, 3, 3, 3, 2, 2, 3, 2,         4, 3, 1, 3, 2, 2, 4, 3, 2, 2, 3)), class = "data.frame", row.names = c(NA,     157L)), .disattenuate = FALSE, .dominant_indicators = NULL,     .estimate_structural = TRUE, .id = "gender", .instruments = NULL,     .iter_max = 100, .model = structure(list(structural = structure(c(0,     1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0), dim = c(4L,     4L), dimnames = list(c("OrgPres", "OrgIden", "AffLove", "AffJoy"    ), c("OrgPres", "OrgIden", "AffLove", "AffJoy"))), measurement = structure(c(1,     0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0,     1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0,     0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0,     1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0,     0, 0, 1, 0, 0, 0, 1), dim = c(4L, 21L), dimnames = list(c("OrgPres",     "OrgIden", "AffLove", "AffJoy"), c("cei1", "cei2", "cei3",     "cei4", "cei5", "cei6", "cei7", "cei8", "ma1", "ma2", "ma3",     "ma4", "ma5", "ma6", "orgcmt5", "orgcmt6", "orgcmt8", "orgcmt1",     "orgcmt2", "orgcmt3", "orgcmt7"))), error_cor = structure(c(0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,     0, 0, 0), dim = c(21L, 21L), dimnames = list(c("cei1", "cei2",     "cei3", "cei4", "cei5", "cei6", "cei7", "cei8", "ma1", "ma2",     "ma3", "ma4", "ma5", "ma6", "orgcmt5", "orgcmt6", "orgcmt8",     "orgcmt1", "orgcmt2", "orgcmt3", "orgcmt7"), c("cei1", "cei2",     "cei3", "cei4", "cei5", "cei6", "cei7", "cei8", "ma1", "ma2",     "ma3", "ma4", "ma5", "ma6", "orgcmt5", "orgcmt6", "orgcmt8",     "orgcmt1", "orgcmt2", "orgcmt3", "orgcmt7"))), cor_specified = structure(numeric(0), dim = c(0L,     0L), dimnames = list(NULL, NULL)), construct_type = c(OrgPres = "Common factor",     OrgIden = "Common factor", AffLove = "Common factor", AffJoy = "Common factor"    ), construct_order = c(OrgPres = "First order", OrgIden = "First order",     AffLove = "First order", AffJoy = "First order"), model_type = "Linear",         indicators = c("cei1", "cei2", "cei3", "cei4", "cei5",         "cei6", "cei7", "cei8", "ma1", "ma2", "ma3", "ma4", "ma5",         "ma6", "orgcmt5", "orgcmt6", "orgcmt8", "orgcmt1", "orgcmt2",         "orgcmt3", "orgcmt7"), cons_exo = "OrgPres", cons_endo = c("OrgIden",         "AffLove", "AffJoy"), vars_2nd = character(0), vars_attached_to_2nd = character(0),         vars_not_attached_to_2nd = c("OrgIden", "AffLove", "AffJoy",         "OrgPres")), class = "cSEMModel"), .normality = FALSE,     .reliabilities = NULL, .starting_values = NULL, .tolerance = 1e-06): object 'W' not found