cSEMArguments

An alphabetical list of all arguments used by functions of the cSEM package, including their description and defaults. Mainly used for internal purposes (parameter inheritance). To list all arguments and their defaults, use args_default(). To list all arguments and their possible choices, use args_default(.choices = TRUE).

Arguments

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.approach
Character string. The Kettenring approach to use. One of "SUMCORR", "MAXVAR", "SSQCORR", "MINVAR" or "GENVAR". Defaults to "SUMCORR".

.approach_2ndorder
Character string. Approach used for models containing second order constructs. One of: "3stage" or "repeated_indicators". Defaults to "3stage".

.approach_cor_robust
Character string. Approach used to obtain a robust indicator correlation matrix. One of: "none" in which case the standard bravais-person correlation is used, "spearman" for the spearman correlation, or "mcd" via MASS::cov.rob() for a robust correlation matrix. Defaults to "none".

.approach_nl
Character string. Approach used to estimate nonlinear structural relationships. One of: "sequential" or "replace". Defaults to "sequential".

.approach_paths
Character string. Approach used to estimate the structural coefficients. One of: "OLS" or "2SLS" (not yet implemented). Defaults to "OLS".

.approach_weights
Character string. Approach used to obtain composite weights. One of: "PLS-PM", "SUMCORR", "MAXVAR", "SSQCORR", "MINVAR", "GENVAR", "GSCA", "PCA", "unit", "bartlett", or "regression". Defaults to "PLS-PM".

.args_used
A list of function argument names to fun() whose value was modified by the user.

.bias_corrected
Logical. Should the standard and the tStat confidence intervall be bias-corrected using the bootstraped bias estimate? If TRUE the confidence intervall for some estimated parameter theta is centered at 2*theta - theta*_hat, where theta*_hat is the average over all .R bootstrap estimates of theta. Defaults to TRUE

.C
A (J x J) composite variance-covariance matrix.

.calculate_ci
Logical. Should a closed-form confidence interval be computed? Defaults to FALSE.

.choices
Logical. Should candidate values for the arguments be returned? Defaults to FALSE.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

.conv_criterion
Character string. The criterion to use for the convergence check. One of: "diff_absolute", "diff_squared", or "diff_relative". Defaults to "diff_absolute".

.csem_model
A (possibly incomplete) cSEMModel-list.

.csem_resample
A list resulting from a call to resamplecSEMResults().

.cv_folds
Integer. The number of cross-validation folds to use. Setting .cv_folds to N (the number of observations) produces leave-one-out cross-validation samples. Defaults to 10.

.data
A data.frame or a matrix of standardized or unstandarized data. Possible column types or classes of the data provided are: logical, numeric (double or integer), factor (ordered and unordered) or a mix of several types. The data may also include one character column whose column name must be given to .id. This column is assumed to contain group identifiers used to split the data into groups.

.disattenuate
Logical. Should composite/proxy correlations be disattenuated to yield consisten loadings and path estimates if at least one of the construct is modeled as a common factor? Defaults to TRUE.

.dist
Character string. The distribution to use for the critical value. One of "t" for Student's t-distribution or "z" for the standard normal distribution. Defaults to "z".

.distance
Character string. A distance measure. One of: "geodesic" or "squared_euclidian". Defaults to "geodesic".

.df
Character string. The method for obtaining the degrees of freedom. Choices are "type1" and "type2". Defaults to "type1" .

.dominant_indicators
A character vector of "name" = "value" pairs, where "value" is a character string giving the name of the dominant indicator and "name" a character string of the corresponding construct name. Dominant indicators may be specified for a subset of the constructs.

.E
A (J x J) matrix of inner weights.

.estimate_structural
Logical. Should the structural coefficients be estimated? Defaults to TRUE.

.eval_plan
Character string. The evaluation plan to use. One of "sequential" or "multiprocess". In the latter case all available cores will be used. Defaults to "sequential".

.first_resample
A list containing the .R resamples based on the original data obtained by resamplecSEMResults()

.second_resample
A list containing .R2 resamples for each of the .R resamples of the first run.

.H
The (N x J) matrix of construct scores.

.handle_inadmissibles
Character string. How should inadmissible results be treated? One of "drop", "ignore", or "replace". If "drop", all replications/resamples yielding an inadmissible result will be dropped (i.e. the number of results returned will be less than .R). For "ignore" all results are returned even if they are inadmissible (i.e. number of results returned = .R). For "replace" resampling continues until there are exactly .R admissible solutions. Defaults to "drop".

.id
Character string or integer. The name or position of the column of .data used to split the data into groups. Defaults to NULL.

.iter_max
Integer. The maximum number of iterations allowed. If iter_max = 1 and .approach_weights = "PLS-PM" one-step weights are returned. If the algorithm exceeds the specified number, weights of iteration step .iter_max - 1 will be returned with a warning. Defaults to 100.

.n
Integer. The number of observations of the original data.

.matrix1
A matrix to compare.

.matrix2
A matrix to compare.

.matrices
A list of at least two matrices.

.model
A model in lavaan model syntax or a cSEMModel list.

.modes
A vector giving the mode for each construct in the form "name" = "mode". Only used internally.

.normality
Logical. Should joint normality of \([\eta_{1:p}; \zeta; \epsilon]\) be assumed in the nonlinear model? See (Dijkstra and Schermelleh-Engel 2014) for details. Defaults to TRUE. Ignored if the model is linear.

.object
An R object of class cSEMResults resulting from a call to csem().

.only_common_factors
Logical. Should only constructs modeled as common factors be included? Defaults to TRUE.

.P
A (J x J) construct variance-covariance matrix (possibly disattenuated).

.PLS_approach_cf
Character string. Approach used to obtain the correction factors for PLSc. One of: "dist_squared_euclid", "dist_euclid_weighted", "fisher_transformed", "mean_arithmetic", "mean_geometric", "mean_harmonic", "geo_of_harmonic". Defaults to "dist_squared_euclid". Ignored if .disattenuate = FALSE or if .approach_weights is not PLS-PM.

.PLS_ignore_structural_model
Logical. Should the structural model be ignored when calculating the inner weights of the PLS-PM algorithm? Defaults to FALSE. Ignored if .approach_weights is not PLS-PM.

.PLS_modes
Either a named list specifying the mode that should be used for each construct in the form "name" = "mode", a single character string giving the mode that should be used for all constructs, or NULL. Possible choices for "mode" are: "modeA", "modeB", "modeBNNLS", "unit", a single number (weight) or a vector of fixed weights of the same length as there are indicators for the construct given by "name". If only a single number is provided this is identical to using unit weights, as weights are rescaled such that the related composite has unit variance. Defaults to NULL. If NULL the appropriate mode according to the type of construct used is choosen. Ignored if .approach_weight is not PLS-PM.

.PLS_weight_scheme_inner
Character string. The inner weighting scheme used in PLS-PM. One of: "centroid", "factorial", or "path". Defaults to "path". Ignored if .approach_weight is not PLS-PM.

.probs
A vector of probabilities.

.quantity
Character string. Which statistic should be returned? One of (TODO) Defaults to (TODO).

.Q
A vector of composite-construct correlations with element names equal to the names of the J construct names used in the measurement model. Note Q^2 is also called the reliability coefficient.

.reliabilities
A character vector of "name" = value pairs, where value is a number between 0 and 1 and "name" a character string of the corresponding construct name, or NULL. Reliabilities may be given for a subset of the constructs. Defaults to NULL in which case reliabilities are estimated by csem(). Currently, only supported for .approach_weights = "PLS-PM".

.resample_method
Character string. The resampling method to use. One of: "bootstrap" or "jackknife". Defaults to "bootstrap".

.resample_method2
Character string. The resampling method to use when resampling from a resample. One of: "none", "bootstrap" or "jackknife". For "bootstrap" the number of draws is provided via .R2. Currently, resampling from each resample is only required for the studentized confidence intervall computed by the infer() function. Defaults to "none".

`.resample_object`
An R object of class cSEMResults_resampled obtained from resamplecSEMResults() or by setting .resample_method = "bootstrap" or "jackknife" when calling csem().

.R
Integer. The number of bootstrap replications. Defaults to 499.

.R2
Integer. The number of bootstrap replications to use when resampling from a resample. Defaults to 199.

.S
The (K x K) empirical indicator correlation matrix.

.saturated
Logical. Should a saturated structural model be used? Defaults to FALSE.

.sign_change_option
Character string. Indicates which sign change option should be used in the resample function. One of "no" "individual" "construct". Defaults to "no".

.seed
Integer. The random seed to use. Defaults to NULL in which case an arbitrary seed is choosen.

.stage
Character string. The stage the model is need for. One of "first" or "second". Defaults to "first".

.starting_values
list of named vectors containig the starting values. Defaults to "NULL".

.terms
A vector of construct names to be classified.

.tolerance
Double. The tolerance criterion for convergence. Defaults to 1e-05.

.type_vcv
Character string. Indicates which model-implied correlation matrix is calcuted One of "indicator" or "construct". Defaults to "indicator".

.verbose
Logical. Should information be printed to the console? Defaults to TRUE.

.user_funs
A function or a (named) list of functions to apply to every resample. Takes .object as an input (e.g., myFun <- function(.object) {...}). Output should preferably be a (named) vector but matrices are also accepted. However, the output will be vectorized (columnwise) in this case. See the examples section for details.

.vector1
A vector of numeric values.

.vector2
A vector of numeric values.

.W
A (J x K) matrix of weights.

.W_new
A (J x K) matrix of weights.

.W_old
A (J x K) matrix of weights.

.X
A matrix of processed data (scaled, cleaned and ordered).

.X_cleaned
A data.frame of processed data (cleaned and ordered). Note: X_cleaned may not be scaled!