R Implementation of igsca_sim.m

This R implementation of I-GSCA is based on the Matlab implementation in igsca_sim.m by Dr. Heungsun Hwang.

igsca(
  Z0,
  W0,
  C0,
  B0,
  con_type,
  indicator_type,
  .dominant_indicators = .dominant_indicators,
  .iter_max = 100,
  .tolerance = 1e-04
)

Arguments

Z0

Data matrix of N cases (measurements) x J indicators with named columns, unstandardized.

W0

Indicator matrix of weights: J indicators (rows) and their corresponding Gamma construct variables (columns).

C0

Indicator matrix of loadings: J indicators (rows) and their corresponding Gamma construct variable (columns).

B0

Square indicator matrix of path coefficients: from-construct-variable (rows) and to-construct-variable (columns). The order of Gamma construct variables should match the order in C0 and W0.

con_type

A vector that denotes whether each construct variable (columns in W0 and C0) is a common factor or composite. Its length should be equal to the number of columns of W0 and C0.

indicator_type

An indicator vector that indices whether a j indicator (rows of W0 and C0) corresponds to a common factor variable (1) or a composite variable (0). This vector is important for computing the uniqueness terms (D) because it zeros the entries for composite indicators.

.dominant_indicators

A named vector that indices the dominant indicator for each construct variable. It is to be clarified whether this should only apply to factor latent variables or also composite latent variables. This is important for ensuring that the signs of the path-coefficients and loadings are consistent. It is sometimes the case in composite-based structural equation modelling methods that loadings/path-coefficients may have the opposite sign. The length of this vector should be equal to the number of construct variables and each value should represent the row number of the dominant indicator for that construct variable.

.iter_max

Maximum number of iterations of the Alternating Least Squares (ALS) algorithm.

.tolerance

Minimum amount of absolute change in the estimates of the path-coefficients (if B0 is non-zero) or the loadings (if B0 is all zero, meaning ther are no path-coefficients) between ALS iterations before ending the optimization.

Value

List of 4 matrices that make up a fitted I-GSCA Model:

(1) Weights

(2) Loadings

(3) Uniqueness Terms D^2

(4) Path Coefficients.

Details

In the example section, the specified model is based on the tutorial I-GSCA model associated with GSCA Pro (Hwang et al. 2023) .

References

Hwang H, Cho G, Choo H (2023). “GSCA Pro—Free Stand-Alone Software for Structural Equation Modeling.” Structural Equation Modeling A Multidisciplinary Journal.

Author

Michael S. Truong

Examples

# Specify the model according to GSCA Pro's example
tutorial_igsca_model <- "
# Composite Model
NetworkingBehavior <~ Behavior1 + Behavior2 + Behavior3 + Behavior5 + Behavior7 + Behavior8 + Behavior9
Numberofjobinterviews <~ Interview1 + Interview2
Numberofjoboffers <~ Offer1 + Offer2 

# Reflective Measurement Model
HonestyHumility =~ Honesty1 + Honesty2 + Honesty3 + Honesty4 + Honesty5 + Honesty6 + Honesty7 + Honesty8 + Honesty9 + Honesty10
Emotionality =~ Emotion1 + Emotion2 + Emotion3 + Emotion4 + Emotion5 + Emotion6 + Emotion8 + Emotion10
Extraversion =~ Extraver2 + Extraver3 + Extraver4 + Extraver5 + Extraver6 + Extraver7 + Extraver8 + Extraver9 + Extraver10
Agreeableness =~ Agreeable1 + Agreeable3 + Agreeable4 + Agreeable5 + Agreeable7 + Agreeable8 + Agreeable9 + Agreeable10
Conscientiousness =~ Conscientious1 + Conscientious3 + Conscientious4 + Conscientious6 + Conscientious7 + Conscientious8 + Conscientious9 + Conscientious10
OpennesstoExperience =~ Openness1 + Openness2 + Openness3 + Openness5 + Openness7 + Openness8 + Openness9 + Openness10

# Structural Model
NetworkingBehavior ~ HonestyHumility + Emotionality + Extraversion + Agreeableness + Conscientiousness + OpennesstoExperience
Numberofjobinterviews ~ NetworkingBehavior
Numberofjoboffers ~ NetworkingBehavior
"

data(LeDang2022)

csem(.data = LeDang2022, tutorial_igsca_model, .approach_weights = "IGSCA",
.dominant_indicators = NULL, .tolerance = 0.0001)
#> Warning: This unsafe exit needs to be fixed before merge
#> $W
#>                 HonestyHumility Emotionality Extraversion Agreeableness
#> Honesty1              0.1311898    0.0000000    0.0000000     0.0000000
#> Honesty2              0.1311064    0.0000000    0.0000000     0.0000000
#> Honesty3              0.1416990    0.0000000    0.0000000     0.0000000
#> Honesty4              0.1486359    0.0000000    0.0000000     0.0000000
#> Honesty5              0.1355583    0.0000000    0.0000000     0.0000000
#> Honesty6              0.1400553    0.0000000    0.0000000     0.0000000
#> Honesty7              0.1432280    0.0000000    0.0000000     0.0000000
#> Honesty8              0.1495155    0.0000000    0.0000000     0.0000000
#> Honesty9              0.1525276    0.0000000    0.0000000     0.0000000
#> Honesty10             0.1416933    0.0000000    0.0000000     0.0000000
#> Emotion1              0.0000000    0.1894210    0.0000000     0.0000000
#> Emotion2              0.0000000    0.1937403    0.0000000     0.0000000
#> Emotion3              0.0000000    0.1607331    0.0000000     0.0000000
#> Emotion4              0.0000000    0.1609309    0.0000000     0.0000000
#> Emotion5              0.0000000    0.1827889    0.0000000     0.0000000
#> Emotion6              0.0000000    0.1561125    0.0000000     0.0000000
#> Emotion8              0.0000000    0.1707045    0.0000000     0.0000000
#> Emotion10             0.0000000    0.1787701    0.0000000     0.0000000
#> Extraver2             0.0000000    0.0000000    0.1492672     0.0000000
#> Extraver3             0.0000000    0.0000000    0.1336280     0.0000000
#> Extraver4             0.0000000    0.0000000    0.1553945     0.0000000
#> Extraver5             0.0000000    0.0000000    0.1530406     0.0000000
#> Extraver6             0.0000000    0.0000000    0.1550715     0.0000000
#> Extraver7             0.0000000    0.0000000    0.1513869     0.0000000
#> Extraver8             0.0000000    0.0000000    0.1444577     0.0000000
#> Extraver9             0.0000000    0.0000000    0.1641063     0.0000000
#> Extraver10            0.0000000    0.0000000    0.1667367     0.0000000
#> Agreeable1            0.0000000    0.0000000    0.0000000     0.1839844
#> Agreeable3            0.0000000    0.0000000    0.0000000     0.1442883
#> Agreeable4            0.0000000    0.0000000    0.0000000     0.1607636
#> Agreeable5            0.0000000    0.0000000    0.0000000     0.1933063
#> Agreeable7            0.0000000    0.0000000    0.0000000     0.1881003
#> Agreeable8            0.0000000    0.0000000    0.0000000     0.1503936
#> Agreeable9            0.0000000    0.0000000    0.0000000     0.1559209
#> Agreeable10           0.0000000    0.0000000    0.0000000     0.1616831
#> Conscientious1        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious3        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious4        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious6        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious7        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious8        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious9        0.0000000    0.0000000    0.0000000     0.0000000
#> Conscientious10       0.0000000    0.0000000    0.0000000     0.0000000
#> Openness1             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness2             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness3             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness5             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness7             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness8             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness9             0.0000000    0.0000000    0.0000000     0.0000000
#> Openness10            0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior1             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior2             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior3             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior5             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior7             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior8             0.0000000    0.0000000    0.0000000     0.0000000
#> Behavior9             0.0000000    0.0000000    0.0000000     0.0000000
#> Interview1            0.0000000    0.0000000    0.0000000     0.0000000
#> Interview2            0.0000000    0.0000000    0.0000000     0.0000000
#> Offer1                0.0000000    0.0000000    0.0000000     0.0000000
#> Offer2                0.0000000    0.0000000    0.0000000     0.0000000
#>                 Conscientiousness OpennesstoExperience NetworkingBehavior
#> Honesty1                0.0000000            0.0000000          0.0000000
#> Honesty2                0.0000000            0.0000000          0.0000000
#> Honesty3                0.0000000            0.0000000          0.0000000
#> Honesty4                0.0000000            0.0000000          0.0000000
#> Honesty5                0.0000000            0.0000000          0.0000000
#> Honesty6                0.0000000            0.0000000          0.0000000
#> Honesty7                0.0000000            0.0000000          0.0000000
#> Honesty8                0.0000000            0.0000000          0.0000000
#> Honesty9                0.0000000            0.0000000          0.0000000
#> Honesty10               0.0000000            0.0000000          0.0000000
#> Emotion1                0.0000000            0.0000000          0.0000000
#> Emotion2                0.0000000            0.0000000          0.0000000
#> Emotion3                0.0000000            0.0000000          0.0000000
#> Emotion4                0.0000000            0.0000000          0.0000000
#> Emotion5                0.0000000            0.0000000          0.0000000
#> Emotion6                0.0000000            0.0000000          0.0000000
#> Emotion8                0.0000000            0.0000000          0.0000000
#> Emotion10               0.0000000            0.0000000          0.0000000
#> Extraver2               0.0000000            0.0000000          0.0000000
#> Extraver3               0.0000000            0.0000000          0.0000000
#> Extraver4               0.0000000            0.0000000          0.0000000
#> Extraver5               0.0000000            0.0000000          0.0000000
#> Extraver6               0.0000000            0.0000000          0.0000000
#> Extraver7               0.0000000            0.0000000          0.0000000
#> Extraver8               0.0000000            0.0000000          0.0000000
#> Extraver9               0.0000000            0.0000000          0.0000000
#> Extraver10              0.0000000            0.0000000          0.0000000
#> Agreeable1              0.0000000            0.0000000          0.0000000
#> Agreeable3              0.0000000            0.0000000          0.0000000
#> Agreeable4              0.0000000            0.0000000          0.0000000
#> Agreeable5              0.0000000            0.0000000          0.0000000
#> Agreeable7              0.0000000            0.0000000          0.0000000
#> Agreeable8              0.0000000            0.0000000          0.0000000
#> Agreeable9              0.0000000            0.0000000          0.0000000
#> Agreeable10             0.0000000            0.0000000          0.0000000
#> Conscientious1          0.1862130            0.0000000          0.0000000
#> Conscientious3          0.1423161            0.0000000          0.0000000
#> Conscientious4          0.1814261            0.0000000          0.0000000
#> Conscientious6          0.1745152            0.0000000          0.0000000
#> Conscientious7          0.1910078            0.0000000          0.0000000
#> Conscientious8          0.1846711            0.0000000          0.0000000
#> Conscientious9          0.1609896            0.0000000          0.0000000
#> Conscientious10         0.1528716            0.0000000          0.0000000
#> Openness1               0.0000000            0.1713169          0.0000000
#> Openness2               0.0000000            0.1761834          0.0000000
#> Openness3               0.0000000            0.1725284          0.0000000
#> Openness5               0.0000000            0.1793467          0.0000000
#> Openness7               0.0000000            0.1956255          0.0000000
#> Openness8               0.0000000            0.1697449          0.0000000
#> Openness9               0.0000000            0.1831634          0.0000000
#> Openness10              0.0000000            0.1648308          0.0000000
#> Behavior1               0.0000000            0.0000000          0.1700159
#> Behavior2               0.0000000            0.0000000          0.1707564
#> Behavior3               0.0000000            0.0000000          0.1886047
#> Behavior5               0.0000000            0.0000000          0.1479393
#> Behavior7               0.0000000            0.0000000          0.1565448
#> Behavior8               0.0000000            0.0000000          0.1810763
#> Behavior9               0.0000000            0.0000000          0.1707645
#> Interview1              0.0000000            0.0000000          0.0000000
#> Interview2              0.0000000            0.0000000          0.0000000
#> Offer1                  0.0000000            0.0000000          0.0000000
#> Offer2                  0.0000000            0.0000000          0.0000000
#>                 Numberofjobinterviews Numberofjoboffers
#> Honesty1                    0.0000000         0.0000000
#> Honesty2                    0.0000000         0.0000000
#> Honesty3                    0.0000000         0.0000000
#> Honesty4                    0.0000000         0.0000000
#> Honesty5                    0.0000000         0.0000000
#> Honesty6                    0.0000000         0.0000000
#> Honesty7                    0.0000000         0.0000000
#> Honesty8                    0.0000000         0.0000000
#> Honesty9                    0.0000000         0.0000000
#> Honesty10                   0.0000000         0.0000000
#> Emotion1                    0.0000000         0.0000000
#> Emotion2                    0.0000000         0.0000000
#> Emotion3                    0.0000000         0.0000000
#> Emotion4                    0.0000000         0.0000000
#> Emotion5                    0.0000000         0.0000000
#> Emotion6                    0.0000000         0.0000000
#> Emotion8                    0.0000000         0.0000000
#> Emotion10                   0.0000000         0.0000000
#> Extraver2                   0.0000000         0.0000000
#> Extraver3                   0.0000000         0.0000000
#> Extraver4                   0.0000000         0.0000000
#> Extraver5                   0.0000000         0.0000000
#> Extraver6                   0.0000000         0.0000000
#> Extraver7                   0.0000000         0.0000000
#> Extraver8                   0.0000000         0.0000000
#> Extraver9                   0.0000000         0.0000000
#> Extraver10                  0.0000000         0.0000000
#> Agreeable1                  0.0000000         0.0000000
#> Agreeable3                  0.0000000         0.0000000
#> Agreeable4                  0.0000000         0.0000000
#> Agreeable5                  0.0000000         0.0000000
#> Agreeable7                  0.0000000         0.0000000
#> Agreeable8                  0.0000000         0.0000000
#> Agreeable9                  0.0000000         0.0000000
#> Agreeable10                 0.0000000         0.0000000
#> Conscientious1              0.0000000         0.0000000
#> Conscientious3              0.0000000         0.0000000
#> Conscientious4              0.0000000         0.0000000
#> Conscientious6              0.0000000         0.0000000
#> Conscientious7              0.0000000         0.0000000
#> Conscientious8              0.0000000         0.0000000
#> Conscientious9              0.0000000         0.0000000
#> Conscientious10             0.0000000         0.0000000
#> Openness1                   0.0000000         0.0000000
#> Openness2                   0.0000000         0.0000000
#> Openness3                   0.0000000         0.0000000
#> Openness5                   0.0000000         0.0000000
#> Openness7                   0.0000000         0.0000000
#> Openness8                   0.0000000         0.0000000
#> Openness9                   0.0000000         0.0000000
#> Openness10                  0.0000000         0.0000000
#> Behavior1                   0.0000000         0.0000000
#> Behavior2                   0.0000000         0.0000000
#> Behavior3                   0.0000000         0.0000000
#> Behavior5                   0.0000000         0.0000000
#> Behavior7                   0.0000000         0.0000000
#> Behavior8                   0.0000000         0.0000000
#> Behavior9                   0.0000000         0.0000000
#> Interview1                  0.5553840         0.0000000
#> Interview2                  0.5491002         0.0000000
#> Offer1                      0.0000000         0.5694471
#> Offer2                      0.0000000         0.5660617
#> 
#> $C
#>                 HonestyHumility Emotionality Extraversion Agreeableness
#> Honesty1             0.67560660   0.01982923  0.003703151  -0.056055040
#> Honesty2             0.67408215   0.00000000  0.000000000   0.000000000
#> Honesty3             0.72882271   0.00000000  0.000000000   0.000000000
#> Honesty4             0.76450210   0.00000000  0.000000000   0.000000000
#> Honesty5             0.69723834   0.00000000  0.000000000   0.000000000
#> Honesty6             0.72036834   0.00000000  0.000000000   0.000000000
#> Honesty7             0.73668680   0.00000000  0.000000000   0.000000000
#> Honesty8             0.76902650   0.00000000  0.000000000   0.000000000
#> Honesty9             0.78662626  -0.03241749  0.042314749   0.004102341
#> Honesty10            0.72879355   0.00000000  0.000000000   0.000000000
#> Emotion1             0.00000000   0.81023233  0.000000000   0.000000000
#> Emotion2             0.00000000   0.82870795  0.000000000   0.000000000
#> Emotion3             0.00000000   0.68752242  0.000000000   0.000000000
#> Emotion4             0.00000000   0.68836835  0.000000000   0.000000000
#> Emotion5             0.00000000   0.78186431  0.000000000   0.000000000
#> Emotion6             0.00000000   0.66937340  0.000000000   0.000000000
#> Emotion8            -0.02259233   0.72790237  0.056761551   0.018706474
#> Emotion10            0.00000000   0.76467406  0.000000000   0.000000000
#> Extraver2            0.00000000   0.00000000  0.736542658   0.000000000
#> Extraver3            0.00000000   0.00000000  0.659372878   0.000000000
#> Extraver4            0.00000000   0.00000000  0.766777172   0.000000000
#> Extraver5            0.00000000   0.00000000  0.755162039   0.000000000
#> Extraver6            0.00000000   0.00000000  0.765183443   0.000000000
#> Extraver7            0.00000000   0.00000000  0.760489441   0.000000000
#> Extraver8           -0.06370461  -0.03934736  0.713568089   0.018909051
#> Extraver9            0.00000000   0.00000000  0.809764603   0.000000000
#> Extraver10           0.00000000   0.00000000  0.822744178   0.000000000
#> Agreeable1           0.00000000   0.00000000  0.000000000   0.845592583
#> Agreeable3           0.00000000   0.00000000  0.000000000   0.663149391
#> Agreeable4           0.00000000   0.00000000  0.000000000   0.738870098
#> Agreeable5           0.00000000   0.00000000  0.000000000   0.888436029
#> Agreeable7           0.00000000   0.00000000  0.000000000   0.878456827
#> Agreeable8          -0.01580236  -0.07255399  0.148385102   0.697112171
#> Agreeable9           0.00000000   0.00000000  0.000000000   0.716612994
#> Agreeable10          0.00000000   0.00000000  0.000000000   0.743095972
#> Conscientious1       0.00000000   0.00000000  0.000000000   0.000000000
#> Conscientious3       0.00000000   0.00000000  0.000000000   0.000000000
#> Conscientious4       0.00000000   0.00000000  0.000000000   0.000000000
#> Conscientious6       0.00000000   0.00000000  0.000000000   0.000000000
#> Conscientious7       0.00000000   0.00000000 -0.012758653   0.055195596
#> Conscientious8       0.01460733   0.00000000  0.000000000   0.000000000
#> Conscientious9       0.00000000   0.00000000  0.000000000   0.000000000
#> Conscientious10      0.00000000   0.00000000  0.000000000   0.000000000
#> Openness1            0.00000000   0.00000000  0.000000000   0.000000000
#> Openness2            0.00000000   0.00000000  0.000000000   0.000000000
#> Openness3            0.00000000   0.00000000  0.000000000   0.000000000
#> Openness5            0.00000000   0.00000000  0.000000000   0.000000000
#> Openness7            0.04495655   0.03892170  0.010124145  -0.024861033
#> Openness8            0.00000000   0.00000000  0.000000000   0.000000000
#> Openness9            0.00000000   0.00000000  0.000000000   0.000000000
#> Openness10           0.00000000   0.00000000  0.000000000   0.000000000
#> Behavior1            0.00000000   0.00000000  0.000000000   0.000000000
#> Behavior2            0.00000000   0.00000000  0.000000000   0.000000000
#> Behavior3            0.00000000   0.00000000  0.000000000   0.000000000
#> Behavior5            0.00000000   0.00000000  0.000000000   0.000000000
#> Behavior7            0.00000000   0.00000000  0.000000000   0.000000000
#> Behavior8            0.00000000   0.00000000  0.000000000   0.000000000
#> Behavior9            0.00000000   0.00000000  0.000000000   0.000000000
#> Interview1           0.00000000   0.00000000  0.000000000   0.000000000
#> Interview2           0.00000000   0.00000000  0.000000000   0.000000000
#> Offer1               0.00000000   0.00000000  0.000000000   0.000000000
#> Offer2               0.00000000   0.00000000  0.000000000   0.000000000
#>                 Conscientiousness OpennesstoExperience NetworkingBehavior
#> Honesty1              0.052213190         0.0283305172        0.008668912
#> Honesty2              0.000000000         0.0000000000        0.000000000
#> Honesty3              0.000000000         0.0000000000        0.000000000
#> Honesty4              0.000000000         0.0000000000        0.000000000
#> Honesty5              0.000000000         0.0000000000        0.000000000
#> Honesty6              0.000000000         0.0000000000        0.000000000
#> Honesty7              0.000000000         0.0000000000        0.000000000
#> Honesty8              0.000000000         0.0000000000        0.000000000
#> Honesty9             -0.019372793        -0.0027401008       -0.047448723
#> Honesty10             0.000000000         0.0000000000        0.000000000
#> Emotion1              0.000000000         0.0000000000        0.000000000
#> Emotion2              0.000000000         0.0000000000        0.000000000
#> Emotion3              0.000000000         0.0000000000        0.000000000
#> Emotion4              0.000000000         0.0000000000        0.000000000
#> Emotion5              0.000000000         0.0000000000        0.000000000
#> Emotion6              0.000000000         0.0000000000        0.000000000
#> Emotion8             -0.005629317         0.0002579131       -0.042802434
#> Emotion10             0.000000000         0.0000000000        0.000000000
#> Extraver2             0.000000000         0.0000000000        0.000000000
#> Extraver3             0.000000000         0.0000000000        0.000000000
#> Extraver4             0.000000000         0.0000000000        0.000000000
#> Extraver5             0.000000000         0.0000000000        0.000000000
#> Extraver6             0.000000000         0.0000000000        0.000000000
#> Extraver7             0.000000000         0.0000000000        0.333264897
#> Extraver8            -0.021850140         0.0000000000        0.000000000
#> Extraver9             0.000000000         0.0000000000        0.000000000
#> Extraver10            0.000000000         0.0000000000        0.000000000
#> Agreeable1            0.000000000         0.0000000000        0.000000000
#> Agreeable3            0.000000000         0.0000000000        0.000000000
#> Agreeable4            0.000000000         0.0000000000        0.000000000
#> Agreeable5            0.000000000         0.0000000000        0.000000000
#> Agreeable7           -0.027428065        -0.0176572840        0.234978851
#> Agreeable8            0.000000000         0.0000000000        0.000000000
#> Agreeable9            0.000000000         0.0000000000        0.000000000
#> Agreeable10           0.000000000         0.0000000000        0.000000000
#> Conscientious1        0.818704438         0.0000000000        0.000000000
#> Conscientious3        0.625707155         0.0000000000        0.000000000
#> Conscientious4        0.797658514         0.0000000000        0.000000000
#> Conscientious6        0.767274023         0.0000000000        0.000000000
#> Conscientious7        0.842426287         0.0253958087       -0.006924628
#> Conscientious8        0.812027054         0.0000000000        0.000000000
#> Conscientious9        0.707807226         0.0000000000        0.000000000
#> Conscientious10       0.672115803         0.0000000000        0.000000000
#> Openness1             0.000000000         0.7245915007        0.000000000
#> Openness2             0.000000000         0.7451747078        0.000000000
#> Openness3             0.000000000         0.7297156879        0.000000000
#> Openness5             0.000000000         0.7585541085        0.000000000
#> Openness7             0.046222876         0.8312083841       -0.036497125
#> Openness8             0.000000000         0.7179427584        0.000000000
#> Openness9             0.000000000         0.7746968706        0.000000000
#> Openness10            0.000000000         0.6971585049        0.000000000
#> Behavior1             0.000000000         0.0000000000        0.904762760
#> Behavior2             0.000000000         0.0000000000        0.799382523
#> Behavior3             0.000000000         0.0000000000        0.874952203
#> Behavior5             0.000000000         0.0000000000        0.733949610
#> Behavior7             0.000000000         0.0000000000        0.737258676
#> Behavior8             0.000000000         0.0000000000        0.912482807
#> Behavior9             0.000000000         0.0000000000        0.913048617
#> Interview1            0.000000000         0.0000000000        0.000000000
#> Interview2            0.000000000         0.0000000000        0.000000000
#> Offer1                0.000000000         0.0000000000        0.000000000
#> Offer2                0.000000000         0.0000000000        0.000000000
#>                 Numberofjobinterviews Numberofjoboffers
#> Honesty1                  0.015947725       0.009005715
#> Honesty2                  0.000000000       0.000000000
#> Honesty3                  0.000000000       0.000000000
#> Honesty4                  0.000000000       0.000000000
#> Honesty5                  0.000000000       0.000000000
#> Honesty6                  0.000000000       0.000000000
#> Honesty7                  0.000000000       0.000000000
#> Honesty8                  0.000000000       0.000000000
#> Honesty9                 -0.004016987       0.000000000
#> Honesty10                 0.000000000       0.000000000
#> Emotion1                  0.000000000       0.000000000
#> Emotion2                  0.000000000       0.000000000
#> Emotion3                  0.000000000       0.000000000
#> Emotion4                  0.000000000       0.000000000
#> Emotion5                  0.000000000       0.000000000
#> Emotion6                 -0.064209702       0.047163964
#> Emotion8                  0.000000000       0.000000000
#> Emotion10                 0.000000000       0.000000000
#> Extraver2                 0.000000000       0.000000000
#> Extraver3                 0.000000000       0.000000000
#> Extraver4                 0.000000000       0.000000000
#> Extraver5                 0.000000000       0.000000000
#> Extraver6                 0.000000000       0.000000000
#> Extraver7                -0.036718309      -0.024185356
#> Extraver8                 0.000000000       0.000000000
#> Extraver9                 0.000000000       0.000000000
#> Extraver10                0.000000000       0.000000000
#> Agreeable1                0.000000000       0.000000000
#> Agreeable3                0.000000000       0.000000000
#> Agreeable4                0.000000000       0.000000000
#> Agreeable5                0.000000000       0.000000000
#> Agreeable7                0.107109224       0.082003558
#> Agreeable8                0.000000000       0.000000000
#> Agreeable9                0.000000000       0.000000000
#> Agreeable10               0.000000000       0.000000000
#> Conscientious1            0.000000000       0.000000000
#> Conscientious3            0.000000000       0.000000000
#> Conscientious4            0.000000000       0.000000000
#> Conscientious6            0.000000000       0.000000000
#> Conscientious7           -0.024779504       0.011497266
#> Conscientious8            0.000000000       0.000000000
#> Conscientious9            0.000000000       0.000000000
#> Conscientious10           0.000000000       0.000000000
#> Openness1                 0.000000000       0.000000000
#> Openness2                 0.000000000       0.000000000
#> Openness3                 0.000000000       0.000000000
#> Openness5                 0.000000000       0.000000000
#> Openness7                 0.000000000       0.000000000
#> Openness8                 0.000000000       0.000000000
#> Openness9                 0.000000000       0.000000000
#> Openness10                0.000000000       0.000000000
#> Behavior1                 0.000000000       0.000000000
#> Behavior2                 0.000000000       0.000000000
#> Behavior3                 0.000000000       0.000000000
#> Behavior5                -0.021237338       0.000000000
#> Behavior7                 0.000000000       0.000000000
#> Behavior8                 0.000000000       0.000000000
#> Behavior9                 0.000000000       0.000000000
#> Interview1                0.907325362       0.000000000
#> Interview2                0.903449302       0.000000000
#> Offer1                    0.000000000       0.881473951
#> Offer2                   -0.001134881       0.880307901
#> 
#> $B
#>                       HonestyHumility Emotionality Extraversion Agreeableness
#> HonestyHumility                     0            0            0             0
#> Emotionality                        0            0            0             0
#> Extraversion                        0            0            0             0
#> Agreeableness                       0            0            0             0
#> Conscientiousness                   0            0            0             0
#> OpennesstoExperience                0            0            0             0
#> NetworkingBehavior                  0            0            0             0
#> Numberofjobinterviews               0            0            0             0
#> Numberofjoboffers                   0            0            0             0
#>                       Conscientiousness OpennesstoExperience NetworkingBehavior
#> HonestyHumility                       0                    0         0.13658084
#> Emotionality                          0                    0        -0.43143587
#> Extraversion                          0                    0         0.40466125
#> Agreeableness                         0                    0         0.12649885
#> Conscientiousness                     0                    0         0.08191108
#> OpennesstoExperience                  0                    0         0.14671568
#> NetworkingBehavior                    0                    0         0.00000000
#> Numberofjobinterviews                 0                    0         0.00000000
#> Numberofjoboffers                     0                    0         0.00000000
#>                       Numberofjobinterviews Numberofjoboffers
#> HonestyHumility                    0.000000        0.00000000
#> Emotionality                       0.000000        0.00000000
#> Extraversion                       0.000000        0.00000000
#> Agreeableness                      0.000000        0.00000000
#> Conscientiousness                  0.000000        0.00000000
#> OpennesstoExperience               0.000000        0.00000000
#> NetworkingBehavior                 0.130859        0.09473457
#> Numberofjobinterviews              0.000000        0.00000000
#> Numberofjoboffers                  0.000000        0.00000000
#> 
#> $E
#>        Honesty1        Honesty2        Honesty3        Honesty4        Honesty5 
#>       0.5257438       0.5291813       0.4785980       0.4232111       0.5404090 
#>        Honesty6        Honesty7        Honesty8        Honesty9       Honesty10 
#>       0.5055015       0.4644715       0.4132550       0.3639625       0.4815995 
#>        Emotion1        Emotion2        Emotion3        Emotion4        Emotion5 
#>       0.3599573       0.3290173       0.5869259       0.5805218       0.4243817 
#>        Emotion6        Emotion8       Emotion10       Extraver2       Extraver3 
#>       0.6092235       0.4192445       0.4622169       0.4862893       0.6112057 
#>       Extraver4       Extraver5       Extraver6       Extraver7       Extraver8 
#>       0.4424327       0.4612551       0.4250687       0.4650675       0.4617015 
#>       Extraver9      Extraver10      Agreeable1      Agreeable3      Agreeable4 
#>       0.3622495       0.3271373       0.3077996       0.6158193       0.4991113 
#>      Agreeable5      Agreeable7      Agreeable8      Agreeable9     Agreeable10 
#>       0.1912282       0.2136827       0.5573519       0.5296857       0.4929461 
#>  Conscientious1  Conscientious3  Conscientious4  Conscientious6  Conscientious7 
#>       0.3587697       0.6465926       0.3995023       0.4540452       0.2661082 
#>  Conscientious8  Conscientious9 Conscientious10       Openness1       Openness2 
#>       0.3483827       0.5440482       0.6023712       0.5255073       0.4930969 
#>       Openness3       Openness5       Openness7       Openness8       Openness9 
#>       0.5139456       0.4682003       0.3230560       0.5414559       0.4470597 
#>      Openness10       Behavior1       Behavior2       Behavior3       Behavior5 
#>       0.5672990       0.0000000       0.0000000       0.0000000       0.0000000 
#>       Behavior7       Behavior8       Behavior9      Interview1      Interview2 
#>       0.0000000       0.0000000       0.0000000       0.0000000       0.0000000 
#>          Offer1          Offer2 
#>       0.0000000       0.0000000 
#> 
#> $Modes
#> [1] "gsca"
#> 
#> $Conv_status
#> [1] TRUE
#> 
#> $Iterations
#> [1] 6
#>