Inference

Calculate common inferencial quantities (e.g. estimated standard error, estimates bias, several confidence intervals) based on a cSEMResults_resampled object as obtained from resamplecSEMResults() or by setting .resample_method = "bootstrap" or "jackknife" when calling csem(). Currently, the following quantities are returned by default (.quantity = "all"):

"mean", "sd" and "bias"

The mean, the standard deviation and the bias (defined as the difference between the resample mean and the original estimate).

"CI_standard_z" and "CI_standard_t"

The standard confidence interval with standard errors estimated by the resample standard deviation. While "CI_standard_z" assumes a standard-normally distributed statistic, "CI_standard_t" assumes a t-statistic with .df = c("type1", "type2")

"CI_percentile"

The percentile confidence interval

"CI_basic"

The basic confidence interval

"CI_bc"

The bias corrected confidence interval

"CI_bca"

The bias corrected and accelerated confidence interval. NOTE: only possible if .resample_method = "bootstrap" and will be slow as jackknife estimates need to be computed.

"CI_t_interval"

The "studentized" confidence interval

See xxx for details on their use and calculation.

infer(
 .resample_object   = NULL,
 .alpha             = 0.05,
 .bias_corrected    = TRUE,
 .quantity          = c("all", "mean", "sd", "bias", "CI_standard_z",
                        "CI_standard_t", "CI_percentile", "CI_basic",
                        "CI_bc", "CI_bca", "CI_t_intervall")
)

MeanResample(.first_resample)

SdResample(.first_resample, .resample_method, .n)

BiasResample(.first_resample, .resample_method, .n)

StandardCIResample(.first_resample, .bias_corrected, .dist = c("z", "t"),
  .df = c("type1", "type2"), .resample_method, .n, .probs)

PercentilCIResample(.first_resample, .probs)

BasicCIResample(.first_resample, .bias_corrected, .probs)

TStatCIResample(.first_resample, .second_resample, .bias_corrected,
  .resample_method, .resample_method2, .n, .probs)

BcCIResample(.first_resample, .probs)

BcaCIResample(.object, .first_resample, .probs)

Arguments

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.bias_corrected
Logical. Should the standard and the tStat confidence intervall be bias-corrected using the bootstraped bias estimate? If TRUE the confidence intervall for some estimated parameter theta is centered at 2*theta - theta*_hat, where theta*_hat is the average over all .R bootstrap estimates of theta. Defaults to TRUE

.quantity
Character string. Which statistic should be returned? One of (TODO) Defaults to (TODO).

.first_resample
A list containing the .R resamples based on the original data obtained by resamplecSEMResults()

.resample_method
Character string. The resampling method to use. One of: "bootstrap" or "jackknife". Defaults to "bootstrap".

.n
Integer. The number of observations of the original data.

.dist
Character string. The distribution to use for the critical value. One of "t" for Student's t-distribution or "z" for the standard normal distribution. Defaults to "z".

.df
Character string. The method for obtaining the degrees of freedom. Choices are "type1" and "type2". Defaults to "type1" .

.probs
A vector of probabilities.

.second_resample
A list containing .R2 resamples for each of the .R resamples of the first run.

.resample_method2
Character string. The resampling method to use when resampling from a resample. One of: "none", "bootstrap" or "jackknife". For "bootstrap" the number of draws is provided via .R2. Currently, resampling from each resample is only required for the studentized confidence intervall computed by the infer() function. Defaults to "none".

.object
An R object of class cSEMResults resulting from a call to csem().

Functions

MeanResample: Computes the mean over all resamples for each resampled statistic/parameter.

SdResample: Computes the standard deviation over all resamples for each resampled statistic/estimator This is usually taken to be the estimate of the standard error of the statistic/estimator.

BiasResample: Computes the estimated bias for each resampled statistic/estimator.

StandardCIResample: Computes the Standard CI with bootstrap SE's. Critical quantiles can be based on both the t- or the standard normal distribution (z). The former may perform better in small samples but there is no clear consenus on what the degrees of freedom should be.

PercentilCIResample: Computes the Percentile CI. The function takes the distribution F* (the CDF) of the resamples as an estimator for the true distribution F of the statistic/estimator of interest. Quantiles of the estimated distribution are then used as lower and upper bound.

BasicCIResample: Computes the Basic CI.

TStatCIResample: Computes the Studentized or t-statistic CI The function computes a boostrap t-statisic (since it is roughly pivotal) and constructs the CI based on bootstraped t-values and bootstraped/jackknife SE's

BcCIResample: (TODO)

BcaCIResample: (TODO)

See also

Examples

# NOT RUN {
# still to implement
# }