Parse lavaan model

Turns a model written in lavaan model syntax into a cSEMModel list.

parseModel(.model)

Arguments

.model
A model in lavaan model syntax or a cSEMModel list.

Value

An object of class cSEMModel is a standardized list containing the following components. J stands for the number of constructs and K for the number of indicators.

$structural

A matrix mimicking the structural relationship between constructs. If constructs are only linearly related, structural is of dimension (J x J) with row- and column names equal to the construct names. If the structural model contains nonlinear relationships structural is (J x (J + J*)) where J* is the number of nonlinear terms. Rows are ordered such that exogenous constructs are always first, followed by constructs that only depend on exogenous constructs and/or previously ordered constructs.

$measurement

A (J x K) matrix mimicking the measurement relationship between constructs and their related indicators. Rows are in the same order as the matrix $structural with row names equal to the construct names. The order of the columns is such that $measurement forms a block diagonal matrix.

$error_cor

A (K x K) matrix mimicking the measurement error correlation relationship. The row and column order is identical to $measurement.

$construct_type

A named vector containing the names of each construct and their respective type ("Common factor" or "Composite").

$construct_order

A named vector containing the names of each construct and their respective order ("First order" or "Second order").

$model_type

The type of model ("Linear" or "Nonlinear").

Note: it is possible to supply an incomplete cSEMModel list to all functions that require .csem_model as a mandatory argument. Currently, only the structural and the measurement matrix are required. However, specifying an incomplete cSEMModel list may lead to unexpected behavior and errors. Use with care.

Examples

model <- '
# Structural model
y1 ~ y2 + y3

# Measurement model
y1 =~ x1 + x2 + x3
y2 =~ x4 + x5
y3 =~ x6 + x7

# Error correlation
x1 ~~ x2
'

(m <- parseModel(model))

#> $structural
#>    y2 y3 y1
#> y2  0  0  0
#> y3  0  0  0
#> y1  1  1  0
#> 
#> $measurement
#>    x4 x5 x6 x7 x1 x2 x3
#> y2  1  1  0  0  0  0  0
#> y3  0  0  1  1  0  0  0
#> y1  0  0  0  0  1  1  1
#> 
#> $error_cor
#>    x4 x5 x6 x7 x1 x2 x3
#> x4  0  0  0  0  0  0  0
#> x5  0  0  0  0  0  0  0
#> x6  0  0  0  0  0  0  0
#> x7  0  0  0  0  0  0  0
#> x1  0  0  0  0  0  1  0
#> x2  0  0  0  0  1  0  0
#> x3  0  0  0  0  0  0  0
#> 
#> $construct_type
#>              y2              y3              y1 
#> "Common factor" "Common factor" "Common factor" 
#> 
#> $construct_order
#>            y2            y3            y1 
#> "First order" "First order" "First order" 
#> 
#> $model_type
#> [1] "Linear"
#> 
#> attr(,"class")
#> [1] "cSEMModel"

# If the model is already a cSEMModel object, the model is returned as is.

identical(m, parseModel(m)) # TRUE

#> [1] TRUE