Resample data

Resample data from a dataset using common resampling methods. For bootstrap or jackknife resampling, package users usually dont need to call this function but directly use resamplecSEMResults() instead.

resampleData(
 .object          = NULL,
 .resample_method = c("bootstrap", "jackknife", "permutation", "cross-validation"),
 .cv_folds        = 10,
 .data            = NULL,
 .id              = NULL,
 .R               = 499,
 .seed            = sample(.Random.seed, 1)
)

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.resample_method
Character string. The resampling method to use. One of: "bootstrap", "jackknife", "permutation", or "cross-validation". Defaults to "bootstrap".

.cv_folds
Integer. The number of cross-validation folds to use. Setting .cv_folds to N (the number of observations) produces leave-one-out cross-validation samples. Defaults to 10.

.data
A data.frame, a matrix or a list of data of either type. Possible column types or classes of the data provided are: "logical", "numeric" ("double" or "integer"), "factor" (ordered and unordered) or a mix of several types. The data may also include one character column whose column name must be given to .id. This column is assumed to contain group identifiers used to split the data into groups. If .data is provided, .object is ignored. Defaults to NULL.

.id
Character string or integer. The name or position of the column of .data used to split the data into groups. Defaults to NULL.

.R
Integer. The number of bootstrap replications, permutation runs or cross-validation repetitions to use. Defaults to 499.

.seed
Integer. The random seed to use. Defaults to NULL in which case an arbitrary seed is choosen.

Value

The structure of the output depends on the type of input and the resampling method:

Bootstrap

If a matrix or data.frame without grouping variable is provided (.id = NULL), the result is a list of length .R (default 499). Each element of that list is a bootstrap (re)sample. If a grouping variable is specified or a list of data is provided (where each list element is assumed to contain data for one group), resampling is done by group. Hence, the result is a list of length equal to the number of group levels with each list element containing .R bootstrap samples based on the N_g observations of group g.

Jackknife

If a matrix or data.frame without grouping variable is provided (.id = NULL), the result is a list of length equal to the number of observations/rows (N) of the dataset provided. Each element of that list is a jackknife (re)sample. If a grouping variable is specified or a list of data is provided (where each list element is assumed to contain data for one group), resampling is done by group. Hence, the result is a list of length equal to the number of group levels with each list element containing N jackknife samples based on the N_g observations of group g.

Permutation

If a matrix or data.frame without grouping variable is provided an error is returned as permutation will simply reorder the observations. If a grouping variable is specified or a list of data is provided (where each list element is assumed to contain data for one group), group membership is permutated. Hence, the result is a list of length .R where each element of that list is a permutation (re)sample.

Cross-validation

If a matrix or data.frame without grouping variable is provided a list of length .R is returned. Each list element contains a list containing the k splits/folds subsequently used as test and training datasets. If a grouping variable is specified or a list of data is provided (where each list element is assumed to contain data for one group), cross-validation is repeated .R times for each group. Hence, the result is a list of length equal to the number of group levels, each containing .R list element (the repetitions) which in turn contain the k splits/folds.

Details

The function resampleData() is general purpose. It simply resamples data from a given dataset according to the resampling method provided via the .resample_method argument. Currently, bootstrap, jackknife, permutation, and cross-validation (both leave-one-out (LOOCV) and k-fold cross-validation) are implemented.

The user may provide a dataset via .data to be resampled or a cSEMResults object in which case the original data used in the call to csem() is used for resampling. If both, a cSEMResults object and a dataset via .data are provided the former is ignored.

As csem() accepts a single dataset, a list of datasets as well as datasets that contain a column name used to split the data into groups, the cSEMResults object may contain multiple datasets. In this case, resampling is done by dataset or group. Note that depending on the number of datasets/groups provided this computation may be slower as resampling will be repeated for each dataset/group.

To split data provided via the .data argument into groups, the column name or the column index of the column containing the group levels to split the data by must be given to .id. If data that contains grouping is taken from a cSEMResults object, .id is taken from within the object. Hence, providing .id is redundant and therefore ignored.

The number of bootstrap or permutation runs as well as the number of cross-validation repetitions is given by .R. The default is 499 but should be increased in real applications. See e.g., Hesterberg (2015) , p.380 for recommendations concerning the bootstrap. For jackknife .R is ignored.

Choosing resample_method = "permutation" for ungrouped data causes an error as permutation will simply reorder the observations. If a list of data is provided each list element is assumed to represent the observations belonging to one group. In this case, data is pooled and group adherance permutated.

For cross-validation the number of folds (k) defaults to 10 and can be changed via the .cv_folds argument. Setting k = N (where N is the number of observations) produces leave-one-out cross-validation samples. Note: 1. k can not be larger than the N. 2. If N/k is not not an integer the last fold will have less observations.

Random number generation (RNG) uses the L'Ecuyer-CRMR RGN stream as implemented in the future.apply package (Bengtsson 2018) . See ?future_lapply for details.

References

Bengtsson H (2018). future.apply: Apply Function to Elements in Parallel using Futures. R package version 1.0.1, https://CRAN.R-project.org/package=future.apply. Hesterberg TC (2015). “What Teachers Should Know About the Bootstrap: Resampling in the Undergraduate Statistics Curriculum.” The American Statistician, 69(4), 371--386. doi: 10.1080/00031305.2015.1089789 .

See also

Examples

require(cSEM)

# ===========================================================================
### Using the raw data 
# ===========================================================================
## Create .R bootstrap samples ----------------------------------------------

res_boot1 <- resampleData(.data = satisfaction)
str(res_boot1, max.level = 3, list.len = 3)

#> List of 499
#>  $ :'data.frame':	250 obs. of  27 variables:
#>   ..$ imag1: int [1:250] 7 8 8 10 8 7 10 10 9 10 ...
#>   ..$ imag2: int [1:250] 7 9 8 3 8 7 10 10 9 10 ...
#>   ..$ imag3: int [1:250] 8 9 8 7 8 7 10 9 10 10 ...
#>   .. [list output truncated]
#>  $ :'data.frame':	250 obs. of  27 variables:
#>   ..$ imag1: int [1:250] 8 8 8 8 9 8 8 10 7 5 ...
#>   ..$ imag2: int [1:250] 8 4 8 8 9 7 7 10 8 8 ...
#>   ..$ imag3: int [1:250] 9 3 8 8 7 8 8 10 9 8 ...
#>   .. [list output truncated]
#>  $ :'data.frame':	250 obs. of  27 variables:
#>   ..$ imag1: int [1:250] 9 8 8 9 8 9 8 8 8 7 ...
#>   ..$ imag2: int [1:250] 9 9 8 10 7 10 6 8 8 9 ...
#>   ..$ imag3: int [1:250] 10 9 8 9 8 9 8 4 8 8 ...
#>   .. [list output truncated]
#>   [list output truncated]

# res_boot1 is the same as if a seed is set
res_boot1  <- resampleData(.data = satisfaction, .seed = 2364)
res_boot1a <- resampleData(.data = satisfaction,
                           .resample_method = "bootstrap", .seed = 2364)

identical(res_boot1, res_boot1a) # TRUE

#> [1] TRUE

## Create cross-validation samples repeated .R = 100 ------------------------

dat <- data.frame(
  "x1" = rnorm(100),
  "x2" = rnorm(100),
  "group" = sample(c("male", "female"), size = 100, replace = TRUE),
  stringsAsFactors = FALSE)

# 10-fold cross-validation
cv_10 <- resampleData(.data = dat, .resample_method = "cross-validation",
                      .id = "group", .R = 100)
str(cv_10, max.level = 3, list.len = 3)

#> List of 2
#>  $ female:List of 100
#>   ..$ :List of 10
#>   .. ..$ 1 :'data.frame':	5 obs. of  3 variables:
#>   .. ..$ 2 :'data.frame':	5 obs. of  3 variables:
#>   .. ..$ 3 :'data.frame':	5 obs. of  3 variables:
#>   .. .. [list output truncated]
#>   ..$ :List of 10
#>   .. ..$ 1 :'data.frame':	5 obs. of  3 variables:
#>   .. ..$ 2 :'data.frame':	5 obs. of  3 variables:
#>   .. ..$ 3 :'data.frame':	5 obs. of  3 variables:
#>   .. .. [list output truncated]
#>   ..$ :List of 10
#>   .. ..$ 1 :'data.frame':	5 obs. of  3 variables:
#>   .. ..$ 2 :'data.frame':	5 obs. of  3 variables:
#>   .. ..$ 3 :'data.frame':	5 obs. of  3 variables:
#>   .. .. [list output truncated]
#>   .. [list output truncated]
#>  $ male  :List of 100
#>   ..$ :List of 10
#>   .. ..$ 1 :'data.frame':	6 obs. of  3 variables:
#>   .. ..$ 2 :'data.frame':	6 obs. of  3 variables:
#>   .. ..$ 3 :'data.frame':	6 obs. of  3 variables:
#>   .. .. [list output truncated]
#>   ..$ :List of 10
#>   .. ..$ 1 :'data.frame':	6 obs. of  3 variables:
#>   .. ..$ 2 :'data.frame':	6 obs. of  3 variables:
#>   .. ..$ 3 :'data.frame':	6 obs. of  3 variables:
#>   .. .. [list output truncated]
#>   ..$ :List of 10
#>   .. ..$ 1 :'data.frame':	6 obs. of  3 variables:
#>   .. ..$ 2 :'data.frame':	6 obs. of  3 variables:
#>   .. ..$ 3 :'data.frame':	6 obs. of  3 variables:
#>   .. .. [list output truncated]
#>   .. [list output truncated]

# Leave-one-out-cross-validation
cv_loocv  <- resampleData(.data = dat[, -3],
                          .resample_method = "cross-validation",
                          .cv_folds = nrow(dat),
                          .R = 50)
str(cv_loocv, max.level = 2, list.len = 3)

#> List of 50
#>  $ :List of 100
#>   ..$ 1  :'data.frame':	1 obs. of  2 variables:
#>   ..$ 2  :'data.frame':	1 obs. of  2 variables:
#>   ..$ 3  :'data.frame':	1 obs. of  2 variables:
#>   .. [list output truncated]
#>  $ :List of 100
#>   ..$ 1  :'data.frame':	1 obs. of  2 variables:
#>   ..$ 2  :'data.frame':	1 obs. of  2 variables:
#>   ..$ 3  :'data.frame':	1 obs. of  2 variables:
#>   .. [list output truncated]
#>  $ :List of 100
#>   ..$ 1  :'data.frame':	1 obs. of  2 variables:
#>   ..$ 2  :'data.frame':	1 obs. of  2 variables:
#>   ..$ 3  :'data.frame':	1 obs. of  2 variables:
#>   .. [list output truncated]
#>   [list output truncated]

## Create permuation samples ------------------------------------------------

res_perm <- resampleData(.data = dat, .resample_method = "permutation",
                         .id = "group")
str(res_perm, max.level = 2, list.len = 3)

#> List of 499
#>  $ :'data.frame':	100 obs. of  3 variables:
#>   ..$ x1: num [1:100] 1.5621 -1.0942 -0.0259 -1.5146 -0.3071 ...
#>   ..$ x2: num [1:100] 1.227 0.461 -1.075 -0.25 -0.488 ...
#>   ..$ id: Factor w/ 2 levels "female","male": 1 2 2 2 1 1 1 1 1 2 ...
#>  $ :'data.frame':	100 obs. of  3 variables:
#>   ..$ x1: num [1:100] 1.5621 -1.0942 -0.0259 -1.5146 -0.3071 ...
#>   ..$ x2: num [1:100] 1.227 0.461 -1.075 -0.25 -0.488 ...
#>   ..$ id: Factor w/ 2 levels "female","male": 1 1 2 1 2 1 2 2 2 2 ...
#>  $ :'data.frame':	100 obs. of  3 variables:
#>   ..$ x1: num [1:100] 1.5621 -1.0942 -0.0259 -1.5146 -0.3071 ...
#>   ..$ x2: num [1:100] 1.227 0.461 -1.075 -0.25 -0.488 ...
#>   ..$ id: Factor w/ 2 levels "female","male": 1 1 2 1 2 2 1 1 1 1 ...
#>   [list output truncated]

# Forgetting to set .id causes an error

# NOT RUN {
res_perm <- resampleData(.data = dat, .resample_method = "permutation")
# }

# ===========================================================================
### Using a cSEMResults object
# ===========================================================================

model <- "
# Structural model
QUAL ~ EXPE
EXPE ~ IMAG
SAT  ~ IMAG + EXPE + QUAL + VAL
LOY  ~ IMAG + SAT
VAL  ~ EXPE + QUAL

# Measurement model
EXPE =~ expe1 + expe2 + expe3 + expe4 + expe5
IMAG =~ imag1 + imag2 + imag3 + imag4 + imag5
LOY  =~ loy1  + loy2  + loy3  + loy4
QUAL =~ qual1 + qual2 + qual3 + qual4 + qual5
SAT  =~ sat1  + sat2  + sat3  + sat4
VAL  =~ val1  + val2  + val3  + val4
"
a <- csem(satisfaction, model)

# Create bootstrap and jackknife samples
res_boot <- resampleData(a, .resample_method = "bootstrap", .R = 999)
res_jack <- resampleData(a, .resample_method = "jackknife")

# Since `satisfaction` is the dataset used the following approaches yield
# identical results.
res_boot_data   <- resampleData(.data = satisfaction, .seed = 2364)
res_boot_object <- resampleData(a, .seed = 2364)

identical(res_boot_data, res_boot_object) # TRUE

#> [1] TRUE