Resample cSEMResults

Resample a cSEMResults object using bootstrap or jackknife resampling. The function is called by csem() if the user sets csem(..., .resample_method = "bootstrap") or csem(..., .resample_method = "jackknife") but may also be called seperately. Technically, resamplecSEMResults() is a generic function with methods for classes cSEMResults_default, cSEMResults_multi and cSEMResults_2ndorder.

resamplecSEMResults(
 .object                = NULL,
 .resample_method       = c("bootstrap", "jackknife), 
 .resample_method2      = c("none", "bootstrap", "jackknife), 
 .R                     = 499,
 .R2                    = 199,
 .handle_inadmissibles  = c("drop", "ignore", "replace"),
 .user_funs             = NULL,
 .eval_plan             = c("sequential", "multiprocess"),
 .seed                  = sample(.Random.seed, 1),
 .sign_change_option    = args_default()$.sign_change_option
)

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.resample_method
Character string. The resampling method to use. One of: "bootstrap" or "jackknife". Defaults to "bootstrap".

.resample_method2
Character string. The resampling method to use when resampling from a resample. One of: "none", "bootstrap" or "jackknife". For "bootstrap" the number of draws is provided via .R2. Currently, resampling from each resample is only required for the studentized confidence intervall computed by the infer() function. Defaults to "none".

.R
Integer. The number of bootstrap replications. Defaults to 499.

.R2
Integer. The number of bootstrap replications to use when resampling from a resample. Defaults to 199.

.handle_inadmissibles
Character string. How should inadmissible results be treated? One of "drop", "ignore", or "replace". If "drop", all replications/resamples yielding an inadmissible result will be dropped (i.e. the number of results returned will be less than .R). For "ignore" all results are returned even if they are inadmissible (i.e. number of results returned = .R). For "replace" resampling continues until there are exactly .R admissible solutions. Defaults to "drop".

.user_funs
A function or a (named) list of functions to apply to every resample. Takes .object as an input (e.g., myFun <- function(.object) {...}). Output should preferably be a (named) vector but matrices are also accepted. However, the output will be vectorized (columnwise) in this case. See the examples section for details.

.eval_plan
Character string. The evaluation plan to use. One of "sequential" or "multiprocess". In the latter case all available cores will be used. Defaults to "sequential".

.seed
Integer. The random seed to use. Defaults to NULL in which case an arbitrary seed is choosen.

.sign_change_option
Character string. Indicates which sign change option should be used in the resample function. One of "no" "individual" "construct". Defaults to "no".

Value

The core structure is the same structure as that of .object with the following elements added:

$Estimates_resamples: A list containing the .R resamples and the original estimates for each of the resampled quantities (Path_estimates, Loading_estimates, Weight_estimates, user defined functions). Each list element is a list containing elements $Resamples and $Original. $Resamples is a (.R x K) matrix with each row representing one resample for each of the K parameters/statistics. $Original contains the original estimates (vectorized by column if the output of the user provided function is a matrix.

$Information_resamples: A list containing addtional information.

Use str(<.object>, list.len = 3) for on the resulting object for an overview.

Details

Given M resamples (for bootstrap M = .R and for jackknife M = N, where N is the number of observations) based on the data used to compute the cSEMResults object provided via .object, resamplecSEMResults() essentially calls csem() on each resample using the arguments of the origianl call (ignoring any arguments related to resampling) and returns estimates for each of a subset of practically useful resampled parameters/statistics computed by csem(). Currently, the following quantities are computed and returned based on each resample: Path estimates, Loading estimates, Weight estimates.

In practical application users may need to resample a specific statistic (e.g, the heterotrait-monotrait ratio of correlations (HTMT) or restrictions on path coefficients such as beta_1 = beta_2). Such statistics may be provided by a function f(.object) or a list of such functions via the .user_funs argument. The only accepted argument of these functions is .object which must be an object of class cSEMResults. Internally, the function will be applied on each resample to produce the desired statistic. Hence, arbitrary complicated statistics may be resampled as long as the body of the function draws on elements contained in the cSEMResults object only. See ?cSEMResults and the examples section for details.

Both resampling the origianl cSEMResults object (call it "first resample") and resampling based on a resampled cSEMResults object (call it "second resample") are supported. Choices for the former are "bootstrap" and "jackknife". Resampling based on a resample is turned off by default (.resample_method2 = "none") as this significantly increases computation time (there are now M * M2 resamples to compute, where M2 is .R2 or N). Currently, resamples of a resample are only required for the studentized confidence intervall computed by the infer() function. Typically, bootstrap resamples are used in this case (Davison and Hinkley 1997) .

As csem() accepts a single dataset, a list of datasets as well as datasets that contain a column name used to split the data into groups, the cSEMResults object may contain multiple datasets. In this case, resampling is done by dataset or group. Note that depending on the number of datasets/groups provided this computation may be considerably slower as resampling will be repeated for each dataset/group. However, apart from speed considerations users dont need to worry about the type of input used to compute the cSEMResults object as resamplecSEMResults() is able to deal with each case.

The number of bootstrap runs for the first and second run are given by .R and .R2. The default is 499 for the first and 199 for the second run but should be increased in real applications. See e.g., Hesterberg (2015) , p.380, Davison and Hinkley (1997) , and (Efron and Hastie 2016) for recommendations. For jackknife .R are .R2 are ignored.

Resampling may produce inadmissble results (as checked by verify()). By default these results are dropped however users may choose to "ignore" or "replace" inadmissble results in which case resampling continious until the necessary number of admissble results is reached.

The cSEM package supports (multi)processing via the future framework (Bengtsson 2018) . Users may simply choose an evaluation plan via .eval_plan and the package takes care of all the complicated backend issues. Currently, users may chose between standard single-core/single-session evaluation ("sequential") and multiprocessing ("multiprocess"). The future package provides other options (e.g. "cluster" or "remote"), however, they probably wont be needed in the context of the cSEM package as simulations usually dont require high-performance clusters. Depeding on the platform, the future package will manage to distribute tasks to multiple R sessions (Windows) or multiple cores. Note that multiprocessing is not necessary always faster when only a "small" number of replications is required as the overhead of initializing new sessions or distributing tasks to different cores will not immediatley be compensated by the avaiability of multiple sessions/cores. As a rule of thumb, the number of resamples to should be larger than 100 to offset the overhead.

Random number generation (RNG) uses the L'Ecuyer-CRMR RGN stream as implemented in the future.apply package (Bengtsson 2018) . It is independent of the evaluation plan. Hence, setting e.g., .seed = 123 will generate the same random number and replicates for both .eval_plan = "sequential" and .eval_plan = "multiprocess". See ?future_lapply for details.

References

Bengtsson H (2018). future: Unified Parallel and Distributed Processing in R for Everyone. R package version 1.10.0, https://CRAN.R-project.org/package=future. Bengtsson H (2018). future.apply: Apply Function to Elements in Parallel using Futures. R package version 1.0.1, https://CRAN.R-project.org/package=future.apply. Davison AC, Hinkley DV (1997). Bootstrap Methods and their Application. Cambridge University Press. doi: 10.1017/cbo9780511802843 . Efron B, Hastie T (2016). Computer Age Statistical Inference. Cambridge University Pr. ISBN 1107149894. Hesterberg TC (2015). “What Teachers Should Know About the Bootstrap: Resampling in the Undergraduate Statistics Curriculum.” The American Statistician, 69(4), 371--386. doi: 10.1080/00031305.2015.1089789 .

See also

Examples

require(cSEM)

model <- "
# Structural model
QUAL ~ EXPE
EXPE ~ IMAG
SAT  ~ IMAG + EXPE + QUAL + VAL
LOY  ~ IMAG + SAT
VAL  ~ EXPE + QUAL

# Measurement model
EXPE =~ expe1 + expe2 + expe3 + expe4 + expe5
IMAG =~ imag1 + imag2 + imag3 + imag4 + imag5
LOY  =~ loy1  + loy2  + loy3  + loy4
QUAL =~ qual1 + qual2 + qual3 + qual4 + qual5
SAT  =~ sat1  + sat2  + sat3  + sat4
VAL  =~ val1  + val2  + val3  + val4
"

a <- csem(satisfaction, model)

# Bootstrap and jackknife estimation
boot <- resamplecSEMResults(a)
jack <- resamplecSEMResults(a, .resample_method = "jackknife")

# Bootstrap resampling with resampling each resample (totals .R*R2 runs).
boot1 <- resamplecSEMResults(a, .resample_method = "bootstrap", .R = 50,
                             .resample_method2 = "bootstrap", .R2 = 20,
                              .seed = 1303)

# To get inferencial quanitites such as the estimates standard error or
# the percentile confidence intervall for each resampled quantity use infer()
inference <- infer(boot1)
inference$Path_estimates$Standard_deviation

#> NULL

inference$Path_estimates$CI_percentile

#>      EXPE ~ IMAG QUAL ~ EXPE VAL ~ EXPE VAL ~ QUAL SAT ~ IMAG SAT ~ EXPE
#> 95%L   0.5895055   0.9415908 -1.9471578  0.9747211 -0.3665258 -0.8227780
#> 95%U   0.7511700   0.9829358 -0.1137174  2.7703402  0.3034237  0.4847504
#>      SAT ~ QUAL SAT ~ VAL LOY ~ IMAG LOY ~ SAT
#> 95%L -0.8250451 0.6044351  0.2197090 0.3103386
#> 95%U  0.8708847 1.4934751  0.5155106 0.6258471

# As usual summarize() can be called directly
summarize(boot1)

#> ______________________________________________________________________________________________________________
#> -------------------------------------------------- Overview --------------------------------------------------
#> 
#> 	Number of observations           = 250
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = path
#> 	Type of indicator correlation    = Bravais-Pearson
#> 	Path model estimator             = OLS
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode 
#> 
#> 	IMAG  Common factor  First order   modeA
#> 	EXPE  Common factor  First order   modeA
#> 	QUAL  Common factor  First order   modeA
#> 	VAL   Common factor  First order   modeA
#> 	SAT   Common factor  First order   modeA
#> 	LOY   Common factor  First order   modeA
#> 
#> -------------------------------------------------- Estimates -------------------------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>   Path           Estimate  Std. error   t-stat.   p-value
#>   EXPE ~ IMAG      0.6559      0.0541   12.1321    0.0000
#>   QUAL ~ EXPE      0.9799      0.0148   66.1356    0.0000
#>   VAL ~ EXPE      -1.7377      0.5804   -2.9939    0.0028
#>   VAL ~ QUAL       2.5851      0.5789    4.4654    0.0000
#>   SAT ~ IMAG      -0.0046      0.2177   -0.0209    0.9833
#>   SAT ~ EXPE       0.6244      0.4137    1.5093    0.1312
#>   SAT ~ QUAL      -0.9575      0.5485   -1.7458    0.0809
#>   SAT ~ VAL        1.2865      0.2803    4.5896    0.0000
#>   LOY ~ IMAG       0.3113      0.0950    3.2790    0.0010
#>   LOY ~ SAT        0.5152      0.0959    5.3717    0.0000
#> 
#> Estimated Loadings:
#> ===================
#>   Loading          Estimate  Std. error   t-stat.   p-value
#>   IMAG =~ imag1      0.5603      0.0656    8.5382    0.0000
#>   IMAG =~ imag2      0.8097      0.0401   20.1776    0.0000
#>   IMAG =~ imag3      0.8386      0.0348   24.0777    0.0000
#>   IMAG =~ imag4      0.4927      0.0885    5.5695    0.0000
#>   IMAG =~ imag5      0.7758      0.0614   12.6452    0.0000
#>   EXPE =~ expe1      0.6758      0.0642   10.5263    0.0000
#>   EXPE =~ expe2      0.8036      0.0450   17.8425    0.0000
#>   EXPE =~ expe3      0.6403      0.0689    9.2938    0.0000
#>   EXPE =~ expe4      0.7402      0.0534   13.8703    0.0000
#>   EXPE =~ expe5      0.7569      0.0362   20.9208    0.0000
#>   QUAL =~ qual1      0.7484      0.0617   12.1345    0.0000
#>   QUAL =~ qual2      0.8305      0.0371   22.3757    0.0000
#>   QUAL =~ qual3      0.6969      0.0556   12.5334    0.0000
#>   QUAL =~ qual4      0.7577      0.0416   18.2165    0.0000
#>   QUAL =~ qual5      0.7621      0.0310   24.5485    0.0000
#>   VAL =~ val1        0.8538      0.0224   38.0809    0.0000
#>   VAL =~ val2        0.7180      0.0528   13.6062    0.0000
#>   VAL =~ val3        0.6123      0.0638    9.6035    0.0000
#>   VAL =~ val4        0.7950      0.0537   14.8132    0.0000
#>   SAT =~ sat1        0.9270      0.0217   42.7357    0.0000
#>   SAT =~ sat2        0.8866      0.0350   25.3578    0.0000
#>   SAT =~ sat3        0.7090      0.0360   19.6902    0.0000
#>   SAT =~ sat4        0.7707      0.0378   20.3885    0.0000
#>   LOY =~ loy1        0.8983      0.0388   23.1511    0.0000
#>   LOY =~ loy2        0.5884      0.0921    6.3877    0.0000
#>   LOY =~ loy3        0.8908      0.0410   21.7278    0.0000
#>   LOY =~ loy4        0.5183      0.1063    4.8736    0.0000
#> ______________________________________________________________________________________________________________

# In the example above .R x .R2 = 50 x 20 = 1000. Multiprocessing will be
# much here and is therefore recommended. Note that multiprocessing
# does not affect the random number generation
boot2 <- resamplecSEMResults(a, .resample_method = "bootstrap", .R = 50,
                             .resample_method2 = "bootstrap", .R2 = 20,
                             .eval_plan = "multiprocess", .seed = 1303)

identical(boot1, boot2)

#> [1] TRUE