Summarize model

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

summarize(
 .object = NULL,
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 50)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = path
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample methode                 = bootstrap
#> 	Number of resamples              = 50
#> 	Number of admissible results     = 50
#> 	Approach to handle inadmissibles = drop
#> 	Sign change option               = none
#> 	Random seed                      = 91078442
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode 
#> 
#> 	eta1  Common factor  First order   modeA
#> 	eta2  Common factor  First order   modeA
#> 	eta3  Common factor  First order   modeA
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0377   17.8002    0.0000 [ 0.6035; 0.7386 ] 
#>   eta3 ~ eta1      0.4585      0.0859    5.3364    0.0000 [ 0.2834; 0.6115 ] 
#>   eta3 ~ eta2      0.3052      0.0847    3.6048    0.0003 [ 0.1424; 0.4247 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0456   14.5447    0.0000 [ 0.5733; 0.7579 ] 
#>   eta1 =~ y12      0.6493      0.0447   14.5269    0.0000 [ 0.5729; 0.7259 ] 
#>   eta1 =~ y13      0.7613      0.0385   19.7800    0.0000 [ 0.6832; 0.8200 ] 
#>   eta2 =~ y21      0.5165      0.0556    9.2950    0.0000 [ 0.3976; 0.6022 ] 
#>   eta2 =~ y22      0.7554      0.0361   20.9132    0.0000 [ 0.6690; 0.8098 ] 
#>   eta2 =~ y23      0.7997      0.0346   23.1006    0.0000 [ 0.7360; 0.8769 ] 
#>   eta3 =~ y31      0.8223      0.0334   24.6265    0.0000 [ 0.7581; 0.8710 ] 
#>   eta3 =~ y32      0.6581      0.0444   14.8358    0.0000 [ 0.5862; 0.7318 ] 
#>   eta3 =~ y33      0.7474      0.0436   17.1564    0.0000 [ 0.6568; 0.8268 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weights        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0256   15.4681    0.0000 [ 0.3487; 0.4446 ] 
#>   eta1 <~ y12      0.3873      0.0241   16.0687    0.0000 [ 0.3569; 0.4411 ] 
#>   eta1 <~ y13      0.4542      0.0220   20.6695    0.0000 [ 0.4132; 0.4891 ] 
#>   eta2 <~ y21      0.3058      0.0292   10.4619    0.0000 [ 0.2411; 0.3567 ] 
#>   eta2 <~ y22      0.4473      0.0212   21.0671    0.0000 [ 0.4013; 0.4840 ] 
#>   eta2 <~ y23      0.4735      0.0220   21.5687    0.0000 [ 0.4380; 0.5166 ] 
#>   eta3 <~ y31      0.4400      0.0222   19.7819    0.0000 [ 0.4014; 0.4802 ] 
#>   eta3 <~ y32      0.3521      0.0188   18.7065    0.0000 [ 0.3218; 0.3865 ] 
#>   eta3 <~ y33      0.3999      0.0224   17.8373    0.0000 [ 0.3673; 0.4368 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0377   17.8002    0.0000 [ 0.6035; 0.7386 ] 
#>   eta3 ~ eta1       0.6634      0.0437   15.1660    0.0000 [ 0.5645; 0.7331 ] 
#>   eta3 ~ eta2       0.3052      0.0847    3.6048    0.0003 [ 0.1424; 0.4247 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0541    3.7900    0.0002 [ 0.0989; 0.2986 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04558837 14.544714  6.309377e-48
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04469492 14.526884  8.186048e-48
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03849059 19.780048  4.422708e-87
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05556262  9.295005  1.471989e-20
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03612017 20.913183  4.062078e-97
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03461661 23.100576 4.568683e-118
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03339000 24.626452 6.580282e-134
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04435679 14.835811  8.595932e-50
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04356532 17.156402  5.629508e-66
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5732888          0.7579235
#> 2          0.5729274          0.7259029
#> 3          0.6832154          0.8199860
#> 4          0.3976429          0.6021569
#> 5          0.6690153          0.8098178
#> 6          0.7359876          0.8769234
#> 7          0.7580604          0.8710317
#> 8          0.5861734          0.7317998
#> 9          0.6567992          0.8268426

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"),
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = path
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample methode                 = bootstrap
#> 	Number of resamples              = 50
#> 	Number of admissible results     = 50
#> 	Approach to handle inadmissibles = drop
#> 	Sign change option               = none
#> 	Random seed                      = 91078442
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode 
#> 
#> 	eta1  Common factor  First order   modeA
#> 	eta2  Common factor  First order   modeA
#> 	eta3  Common factor  First order   modeA
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0377   17.8002    0.0000 [ 0.5712; 0.7662 ] 
#>   eta3 ~ eta1      0.4585      0.0859    5.3364    0.0000 [ 0.2255; 0.6698 ] 
#>   eta3 ~ eta2      0.3052      0.0847    3.6048    0.0003 [ 0.0902; 0.5280 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0456   14.5447    0.0000 [ 0.5434; 0.7791 ] 
#>   eta1 =~ y12      0.6493      0.0447   14.5269    0.0000 [ 0.5327; 0.7638 ] 
#>   eta1 =~ y13      0.7613      0.0385   19.7800    0.0000 [ 0.6692; 0.8682 ] 
#>   eta2 =~ y21      0.5165      0.0556    9.2950    0.0000 [ 0.3733; 0.6607 ] 
#>   eta2 =~ y22      0.7554      0.0361   20.9132    0.0000 [ 0.6646; 0.8514 ] 
#>   eta2 =~ y23      0.7997      0.0346   23.1006    0.0000 [ 0.7055; 0.8845 ] 
#>   eta3 =~ y31      0.8223      0.0334   24.6265    0.0000 [ 0.7413; 0.9140 ] 
#>   eta3 =~ y32      0.6581      0.0444   14.8358    0.0000 [ 0.5423; 0.7717 ] 
#>   eta3 =~ y33      0.7474      0.0436   17.1564    0.0000 [ 0.6348; 0.8601 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weights        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0256   15.4681    0.0000 [ 0.3273; 0.4595 ] 
#>   eta1 <~ y12      0.3873      0.0241   16.0687    0.0000 [ 0.3234; 0.4481 ] 
#>   eta1 <~ y13      0.4542      0.0220   20.6695    0.0000 [ 0.4005; 0.5141 ] 
#>   eta2 <~ y21      0.3058      0.0292   10.4619    0.0000 [ 0.2316; 0.3828 ] 
#>   eta2 <~ y22      0.4473      0.0212   21.0671    0.0000 [ 0.3949; 0.5047 ] 
#>   eta2 <~ y23      0.4735      0.0220   21.5687    0.0000 [ 0.4150; 0.5285 ] 
#>   eta3 <~ y31      0.4400      0.0222   19.7819    0.0000 [ 0.3840; 0.4990 ] 
#>   eta3 <~ y32      0.3521      0.0188   18.7065    0.0000 [ 0.3022; 0.3996 ] 
#>   eta3 <~ y33      0.3999      0.0224   17.8373    0.0000 [ 0.3410; 0.4569 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0377   17.8002    0.0000 [ 0.5712; 0.7662 ] 
#>   eta3 ~ eta1       0.6634      0.0437   15.1660    0.0000 [ 0.5423; 0.7685 ] 
#>   eta3 ~ eta2       0.3052      0.0847    3.6048    0.0003 [ 0.0902; 0.5280 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0541    3.7900    0.0002 [ 0.0680; 0.3475 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03771494 17.800200 7.041653e-71
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08592050  5.336407 9.480626e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08465077  3.604824 3.123646e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.57120289          0.7662432          0.5946234          0.7428227
#> 2         0.22550367          0.6698360          0.2788593          0.6164803
#> 3         0.09023011          0.5279961          0.1427972          0.4754290
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5916740          0.7452722          0.6035071          0.7385602
#> 2          0.2228695          0.6599376          0.2834209          0.6114943
#> 3          0.1245752          0.5283276          0.1424138          0.4247350