Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1078560872
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0530   12.6742    0.0000 [ 0.5731; 0.7623 ] 
#>   eta3 ~ eta1      0.4585      0.0849    5.3987    0.0000 [ 0.2842; 0.5869 ] 
#>   eta3 ~ eta2      0.3052      0.0941    3.2414    0.0012 [ 0.1667; 0.5410 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0413   16.0744    0.0000 [ 0.5598; 0.7151 ] 
#>   eta1 =~ y12      0.6493      0.0399   16.2821    0.0000 [ 0.5762; 0.7169 ] 
#>   eta1 =~ y13      0.7613      0.0378   20.1551    0.0000 [ 0.6855; 0.8210 ] 
#>   eta2 =~ y21      0.5165      0.0482   10.7199    0.0000 [ 0.4512; 0.6195 ] 
#>   eta2 =~ y22      0.7554      0.0370   20.4291    0.0000 [ 0.6954; 0.8178 ] 
#>   eta2 =~ y23      0.7997      0.0344   23.2478    0.0000 [ 0.7391; 0.8493 ] 
#>   eta3 =~ y31      0.8223      0.0359   22.9300    0.0000 [ 0.7593; 0.8771 ] 
#>   eta3 =~ y32      0.6581      0.0413   15.9178    0.0000 [ 0.5906; 0.7221 ] 
#>   eta3 =~ y33      0.7474      0.0357   20.9550    0.0000 [ 0.6698; 0.7997 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0233   16.9728    0.0000 [ 0.3479; 0.4306 ] 
#>   eta1 <~ y12      0.3873      0.0215   18.0243    0.0000 [ 0.3552; 0.4260 ] 
#>   eta1 <~ y13      0.4542      0.0228   19.9081    0.0000 [ 0.4118; 0.4899 ] 
#>   eta2 <~ y21      0.3058      0.0238   12.8279    0.0000 [ 0.2686; 0.3491 ] 
#>   eta2 <~ y22      0.4473      0.0197   22.7024    0.0000 [ 0.4034; 0.4800 ] 
#>   eta2 <~ y23      0.4735      0.0213   22.1825    0.0000 [ 0.4357; 0.5063 ] 
#>   eta3 <~ y31      0.4400      0.0182   24.1395    0.0000 [ 0.4010; 0.4712 ] 
#>   eta3 <~ y32      0.3521      0.0201   17.5213    0.0000 [ 0.3195; 0.3896 ] 
#>   eta3 <~ y33      0.3999      0.0194   20.6031    0.0000 [ 0.3641; 0.4348 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0530   12.6742    0.0000 [ 0.5731; 0.7623 ] 
#>   eta3 ~ eta1       0.6634      0.0402   16.4832    0.0000 [ 0.6088; 0.7465 ] 
#>   eta3 ~ eta2       0.3052      0.0941    3.2414    0.0012 [ 0.1667; 0.5410 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0602    3.4030    0.0007 [ 0.1159; 0.3337 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04125011 16.07438  3.858457e-58
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03987687 16.28207  1.323208e-59
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03777431 20.15512  2.426663e-90
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04817720 10.71990  8.209074e-27
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03697599 20.42914  9.210521e-93
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03439745 23.24776 1.498710e-119
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03586029 22.93003 2.331971e-116
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04134172 15.91779  4.769067e-57
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03566798 20.95505  1.687643e-97
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5597716          0.7151191
#> 2          0.5761606          0.7169274
#> 3          0.6854547          0.8209543
#> 4          0.4512116          0.6195393
#> 5          0.6953967          0.8177862
#> 6          0.7390503          0.8493449
#> 7          0.7592945          0.8770865
#> 8          0.5906402          0.7221164
#> 9          0.6698355          0.7997110

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1078560872
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0530   12.6742    0.0000 [ 0.5353; 0.8092 ] 
#>   eta3 ~ eta1      0.4585      0.0849    5.3987    0.0000 [ 0.2368; 0.6760 ] 
#>   eta3 ~ eta2      0.3052      0.0941    3.2414    0.0012 [ 0.0570; 0.5439 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0413   16.0744    0.0000 [ 0.5680; 0.7813 ] 
#>   eta1 =~ y12      0.6493      0.0399   16.2821    0.0000 [ 0.5538; 0.7600 ] 
#>   eta1 =~ y13      0.7613      0.0378   20.1551    0.0000 [ 0.6626; 0.8579 ] 
#>   eta2 =~ y21      0.5165      0.0482   10.7199    0.0000 [ 0.3890; 0.6382 ] 
#>   eta2 =~ y22      0.7554      0.0370   20.4291    0.0000 [ 0.6664; 0.8576 ] 
#>   eta2 =~ y23      0.7997      0.0344   23.2478    0.0000 [ 0.7148; 0.8927 ] 
#>   eta3 =~ y31      0.8223      0.0359   22.9300    0.0000 [ 0.7332; 0.9186 ] 
#>   eta3 =~ y32      0.6581      0.0413   15.9178    0.0000 [ 0.5494; 0.7632 ] 
#>   eta3 =~ y33      0.7474      0.0357   20.9550    0.0000 [ 0.6617; 0.8462 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0233   16.9728    0.0000 [ 0.3375; 0.4580 ] 
#>   eta1 <~ y12      0.3873      0.0215   18.0243    0.0000 [ 0.3317; 0.4428 ] 
#>   eta1 <~ y13      0.4542      0.0228   19.9081    0.0000 [ 0.3889; 0.5069 ] 
#>   eta2 <~ y21      0.3058      0.0238   12.8279    0.0000 [ 0.2411; 0.3644 ] 
#>   eta2 <~ y22      0.4473      0.0197   22.7024    0.0000 [ 0.3978; 0.4997 ] 
#>   eta2 <~ y23      0.4735      0.0213   22.1825    0.0000 [ 0.4180; 0.5283 ] 
#>   eta3 <~ y31      0.4400      0.0182   24.1395    0.0000 [ 0.3924; 0.4867 ] 
#>   eta3 <~ y32      0.3521      0.0201   17.5213    0.0000 [ 0.2974; 0.4013 ] 
#>   eta3 <~ y33      0.3999      0.0194   20.6031    0.0000 [ 0.3510; 0.4514 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0530   12.6742    0.0000 [ 0.5353; 0.8092 ] 
#>   eta3 ~ eta1       0.6634      0.0402   16.4832    0.0000 [ 0.5558; 0.7639 ] 
#>   eta3 ~ eta2       0.3052      0.0941    3.2414    0.0012 [ 0.0570; 0.5439 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0602    3.4030    0.0007 [ 0.0478; 0.3592 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05296864 12.674167 8.222740e-37
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08492843  5.398743 6.710947e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09414050  3.241444 1.189260e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.53526744          0.8091913          0.5681603          0.7762984
#> 2         0.23676427          0.6759662          0.2895038          0.6232266
#> 3         0.05702747          0.5438690          0.1154876          0.4854088
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5340017          0.7634611          0.5730628          0.7623260
#> 2          0.2513151          0.6486937          0.2842448          0.5868815
#> 3          0.1609772          0.5441692          0.1666943          0.5410327