Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1515345800
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0470   14.2920    0.0000 [ 0.5887; 0.7521 ] 
#>   eta3 ~ eta1      0.4585      0.0790    5.8048    0.0000 [ 0.3438; 0.5968 ] 
#>   eta3 ~ eta2      0.3052      0.0858    3.5550    0.0004 [ 0.1608; 0.4454 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0466   14.2262    0.0000 [ 0.5674; 0.7322 ] 
#>   eta1 =~ y12      0.6493      0.0396   16.3987    0.0000 [ 0.5893; 0.7227 ] 
#>   eta1 =~ y13      0.7613      0.0325   23.4286    0.0000 [ 0.7022; 0.8042 ] 
#>   eta2 =~ y21      0.5165      0.0596    8.6596    0.0000 [ 0.4043; 0.6220 ] 
#>   eta2 =~ y22      0.7554      0.0392   19.2712    0.0000 [ 0.6775; 0.8404 ] 
#>   eta2 =~ y23      0.7997      0.0364   21.9934    0.0000 [ 0.7490; 0.8790 ] 
#>   eta3 =~ y31      0.8223      0.0318   25.8186    0.0000 [ 0.7787; 0.8840 ] 
#>   eta3 =~ y32      0.6581      0.0425   15.4808    0.0000 [ 0.5795; 0.7276 ] 
#>   eta3 =~ y33      0.7474      0.0418   17.8916    0.0000 [ 0.6557; 0.8074 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0231   17.1483    0.0000 [ 0.3477; 0.4287 ] 
#>   eta1 <~ y12      0.3873      0.0208   18.6338    0.0000 [ 0.3540; 0.4213 ] 
#>   eta1 <~ y13      0.4542      0.0219   20.7180    0.0000 [ 0.4160; 0.4910 ] 
#>   eta2 <~ y21      0.3058      0.0317    9.6434    0.0000 [ 0.2463; 0.3599 ] 
#>   eta2 <~ y22      0.4473      0.0270   16.5714    0.0000 [ 0.4055; 0.5001 ] 
#>   eta2 <~ y23      0.4735      0.0184   25.7657    0.0000 [ 0.4413; 0.5055 ] 
#>   eta3 <~ y31      0.4400      0.0187   23.5054    0.0000 [ 0.4117; 0.4735 ] 
#>   eta3 <~ y32      0.3521      0.0186   18.9820    0.0000 [ 0.3205; 0.3817 ] 
#>   eta3 <~ y33      0.3999      0.0224   17.8823    0.0000 [ 0.3544; 0.4348 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0470   14.2920    0.0000 [ 0.5887; 0.7521 ] 
#>   eta3 ~ eta1       0.6634      0.0334   19.8749    0.0000 [ 0.6027; 0.7092 ] 
#>   eta3 ~ eta2       0.3052      0.0858    3.5550    0.0004 [ 0.1608; 0.4454 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0639    3.2058    0.0013 [ 0.1012; 0.3350 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04660898 14.22623  6.299008e-46
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03959315 16.39874  1.952322e-60
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03249638 23.42863 2.183226e-121
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05963937  8.65963  4.732998e-18
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03919781 19.27117  9.378817e-83
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03635932 21.99336 3.333660e-107
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03184824 25.81861 5.481120e-147
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04250866 15.48082  4.674663e-54
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04177504 17.89164  1.370077e-71
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5674215          0.7321519
#> 2          0.5893364          0.7226703
#> 3          0.7022413          0.8042161
#> 4          0.4043193          0.6219508
#> 5          0.6775094          0.8403771
#> 6          0.7489961          0.8789942
#> 7          0.7787013          0.8839935
#> 8          0.5795234          0.7276446
#> 9          0.6557060          0.8073684

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1515345800
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0470   14.2920    0.0000 [ 0.5478; 0.7908 ] 
#>   eta3 ~ eta1      0.4585      0.0790    5.8048    0.0000 [ 0.2541; 0.6626 ] 
#>   eta3 ~ eta2      0.3052      0.0858    3.5550    0.0004 [ 0.0894; 0.5333 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0466   14.2262    0.0000 [ 0.5444; 0.7854 ] 
#>   eta1 =~ y12      0.6493      0.0396   16.3987    0.0000 [ 0.5469; 0.7516 ] 
#>   eta1 =~ y13      0.7613      0.0325   23.4286    0.0000 [ 0.6789; 0.8469 ] 
#>   eta2 =~ y21      0.5165      0.0596    8.6596    0.0000 [ 0.3637; 0.6721 ] 
#>   eta2 =~ y22      0.7554      0.0392   19.2712    0.0000 [ 0.6579; 0.8606 ] 
#>   eta2 =~ y23      0.7997      0.0364   21.9934    0.0000 [ 0.7075; 0.8955 ] 
#>   eta3 =~ y31      0.8223      0.0318   25.8186    0.0000 [ 0.7322; 0.8969 ] 
#>   eta3 =~ y32      0.6581      0.0425   15.4808    0.0000 [ 0.5452; 0.7650 ] 
#>   eta3 =~ y33      0.7474      0.0418   17.8916    0.0000 [ 0.6443; 0.8604 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0231   17.1483    0.0000 [ 0.3365; 0.4558 ] 
#>   eta1 <~ y12      0.3873      0.0208   18.6338    0.0000 [ 0.3329; 0.4404 ] 
#>   eta1 <~ y13      0.4542      0.0219   20.7180    0.0000 [ 0.3974; 0.5107 ] 
#>   eta2 <~ y21      0.3058      0.0317    9.6434    0.0000 [ 0.2239; 0.3879 ] 
#>   eta2 <~ y22      0.4473      0.0270   16.5714    0.0000 [ 0.3778; 0.5174 ] 
#>   eta2 <~ y23      0.4735      0.0184   25.7657    0.0000 [ 0.4253; 0.5204 ] 
#>   eta3 <~ y31      0.4400      0.0187   23.5054    0.0000 [ 0.3893; 0.4861 ] 
#>   eta3 <~ y32      0.3521      0.0186   18.9820    0.0000 [ 0.3044; 0.4003 ] 
#>   eta3 <~ y33      0.3999      0.0224   17.8823    0.0000 [ 0.3465; 0.4621 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0470   14.2920    0.0000 [ 0.5478; 0.7908 ] 
#>   eta3 ~ eta1       0.6634      0.0334   19.8749    0.0000 [ 0.5795; 0.7521 ] 
#>   eta3 ~ eta2       0.3052      0.0858    3.5550    0.0004 [ 0.0894; 0.5333 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0639    3.2058    0.0013 [ 0.0422; 0.3727 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04697274 14.291978 2.455350e-46
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07898808  5.804759 6.445864e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08583723  3.554997 3.779831e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54784265          0.7907592          0.5770122          0.7615896
#> 2         0.25409979          0.6625815          0.3031505          0.6135309
#> 3         0.08938093          0.5332826          0.1426849          0.4799787
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5765272          0.7625097          0.5886880          0.7520500
#> 2          0.2922953          0.6100682          0.3438160          0.5968163
#> 3          0.1428173          0.4526071          0.1607687          0.4454351