Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1036861017
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0443   15.1629    0.0000 [ 0.5977; 0.7411 ] 
#>   eta3 ~ eta1      0.4585      0.0763    6.0104    0.0000 [ 0.3401; 0.6115 ] 
#>   eta3 ~ eta2      0.3052      0.0769    3.9664    0.0001 [ 0.1454; 0.4237 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0400   16.5587    0.0000 [ 0.5835; 0.7313 ] 
#>   eta1 =~ y12      0.6493      0.0373   17.3838    0.0000 [ 0.5479; 0.6883 ] 
#>   eta1 =~ y13      0.7613      0.0352   21.6137    0.0000 [ 0.6804; 0.8093 ] 
#>   eta2 =~ y21      0.5165      0.0541    9.5533    0.0000 [ 0.4267; 0.6079 ] 
#>   eta2 =~ y22      0.7554      0.0358   21.0777    0.0000 [ 0.6941; 0.8198 ] 
#>   eta2 =~ y23      0.7997      0.0343   23.3146    0.0000 [ 0.7416; 0.8599 ] 
#>   eta3 =~ y31      0.8223      0.0282   29.1768    0.0000 [ 0.7640; 0.8657 ] 
#>   eta3 =~ y32      0.6581      0.0368   17.8697    0.0000 [ 0.5896; 0.7322 ] 
#>   eta3 =~ y33      0.7474      0.0437   17.1019    0.0000 [ 0.6599; 0.8392 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0246   16.0877    0.0000 [ 0.3605; 0.4519 ] 
#>   eta1 <~ y12      0.3873      0.0189   20.4534    0.0000 [ 0.3472; 0.4169 ] 
#>   eta1 <~ y13      0.4542      0.0189   24.0782    0.0000 [ 0.4286; 0.5055 ] 
#>   eta2 <~ y21      0.3058      0.0290   10.5560    0.0000 [ 0.2577; 0.3615 ] 
#>   eta2 <~ y22      0.4473      0.0210   21.2792    0.0000 [ 0.4081; 0.4824 ] 
#>   eta2 <~ y23      0.4735      0.0196   24.1516    0.0000 [ 0.4468; 0.5035 ] 
#>   eta3 <~ y31      0.4400      0.0181   24.3444    0.0000 [ 0.4061; 0.4763 ] 
#>   eta3 <~ y32      0.3521      0.0189   18.5842    0.0000 [ 0.3256; 0.3830 ] 
#>   eta3 <~ y33      0.3999      0.0188   21.2837    0.0000 [ 0.3654; 0.4349 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0443   15.1629    0.0000 [ 0.5977; 0.7411 ] 
#>   eta3 ~ eta1       0.6634      0.0392   16.9287    0.0000 [ 0.6075; 0.7372 ] 
#>   eta3 ~ eta2       0.3052      0.0769    3.9664    0.0001 [ 0.1454; 0.4237 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0521    3.9341    0.0001 [ 0.0913; 0.2899 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04004357 16.558712  1.385288e-61
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03734957 17.383812  1.094309e-67
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03522522 21.613654 1.336431e-103
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05406045  9.553283  1.256498e-21
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03583816 21.077746  1.273035e-98
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03429885 23.314589 3.153363e-120
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02818253 29.176849 3.814504e-187
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03682592 17.869719  2.030130e-71
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04370420 17.101884  1.436828e-65
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5835349          0.7312621
#> 2          0.5479079          0.6882641
#> 3          0.6803735          0.8092899
#> 4          0.4266701          0.6079262
#> 5          0.6940853          0.8197677
#> 6          0.7415606          0.8599085
#> 7          0.7640085          0.8656579
#> 8          0.5896171          0.7322221
#> 9          0.6599424          0.8392053

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1036861017
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0443   15.1629    0.0000 [ 0.5575; 0.7864 ] 
#>   eta3 ~ eta1      0.4585      0.0763    6.0104    0.0000 [ 0.2433; 0.6378 ] 
#>   eta3 ~ eta2      0.3052      0.0769    3.9664    0.0001 [ 0.1215; 0.5193 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0400   16.5587    0.0000 [ 0.5645; 0.7715 ] 
#>   eta1 =~ y12      0.6493      0.0373   17.3838    0.0000 [ 0.5656; 0.7588 ] 
#>   eta1 =~ y13      0.7613      0.0352   21.6137    0.0000 [ 0.6685; 0.8506 ] 
#>   eta2 =~ y21      0.5165      0.0541    9.5533    0.0000 [ 0.3747; 0.6543 ] 
#>   eta2 =~ y22      0.7554      0.0358   21.0777    0.0000 [ 0.6707; 0.8560 ] 
#>   eta2 =~ y23      0.7997      0.0343   23.3146    0.0000 [ 0.7103; 0.8876 ] 
#>   eta3 =~ y31      0.8223      0.0282   29.1768    0.0000 [ 0.7506; 0.8964 ] 
#>   eta3 =~ y32      0.6581      0.0368   17.8697    0.0000 [ 0.5582; 0.7486 ] 
#>   eta3 =~ y33      0.7474      0.0437   17.1019    0.0000 [ 0.6383; 0.8643 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0246   16.0877    0.0000 [ 0.3306; 0.4578 ] 
#>   eta1 <~ y12      0.3873      0.0189   20.4534    0.0000 [ 0.3421; 0.4400 ] 
#>   eta1 <~ y13      0.4542      0.0189   24.0782    0.0000 [ 0.3995; 0.4970 ] 
#>   eta2 <~ y21      0.3058      0.0290   10.5560    0.0000 [ 0.2290; 0.3789 ] 
#>   eta2 <~ y22      0.4473      0.0210   21.2792    0.0000 [ 0.3962; 0.5049 ] 
#>   eta2 <~ y23      0.4735      0.0196   24.1516    0.0000 [ 0.4208; 0.5222 ] 
#>   eta3 <~ y31      0.4400      0.0181   24.3444    0.0000 [ 0.3937; 0.4872 ] 
#>   eta3 <~ y32      0.3521      0.0189   18.5842    0.0000 [ 0.3006; 0.3986 ] 
#>   eta3 <~ y33      0.3999      0.0188   21.2837    0.0000 [ 0.3536; 0.4508 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0443   15.1629    0.0000 [ 0.5575; 0.7864 ] 
#>   eta3 ~ eta1       0.6634      0.0392   16.9287    0.0000 [ 0.5551; 0.7577 ] 
#>   eta3 ~ eta2       0.3052      0.0769    3.9664    0.0001 [ 0.1215; 0.5193 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0521    3.9341    0.0001 [ 0.0812; 0.3505 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04427462 15.162940 6.222079e-52
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07628572  6.010388 1.850802e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07693346  3.966429 7.295735e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5574762          0.7864395          0.5849702          0.7589455
#> 2          0.2433049          0.6378116          0.2906774          0.5904390
#> 3          0.1214758          0.5193322          0.1692506          0.4715574
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5927521          0.7508654          0.5976821          0.7410815
#> 2          0.3167065          0.6159483          0.3401348          0.6114646
#> 3          0.1157025          0.4483057          0.1453663          0.4237078