Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1326984296
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0452   14.8399    0.0000 [ 0.5830; 0.7699 ] 
#>   eta3 ~ eta1      0.4585      0.0815    5.6253    0.0000 [ 0.2926; 0.5829 ] 
#>   eta3 ~ eta2      0.3052      0.0830    3.6754    0.0002 [ 0.1701; 0.4607 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0408   16.2396    0.0000 [ 0.6033; 0.7520 ] 
#>   eta1 =~ y12      0.6493      0.0300   21.6152    0.0000 [ 0.5976; 0.6975 ] 
#>   eta1 =~ y13      0.7613      0.0368   20.6785    0.0000 [ 0.7033; 0.8200 ] 
#>   eta2 =~ y21      0.5165      0.0521    9.9154    0.0000 [ 0.4274; 0.6217 ] 
#>   eta2 =~ y22      0.7554      0.0311   24.2521    0.0000 [ 0.6837; 0.8041 ] 
#>   eta2 =~ y23      0.7997      0.0388   20.6287    0.0000 [ 0.7177; 0.8509 ] 
#>   eta3 =~ y31      0.8223      0.0392   20.9572    0.0000 [ 0.7505; 0.8808 ] 
#>   eta3 =~ y32      0.6581      0.0312   21.0919    0.0000 [ 0.6164; 0.7114 ] 
#>   eta3 =~ y33      0.7474      0.0478   15.6344    0.0000 [ 0.6458; 0.8258 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0208   19.0256    0.0000 [ 0.3650; 0.4285 ] 
#>   eta1 <~ y12      0.3873      0.0138   28.0958    0.0000 [ 0.3625; 0.4121 ] 
#>   eta1 <~ y13      0.4542      0.0232   19.6004    0.0000 [ 0.4160; 0.4945 ] 
#>   eta2 <~ y21      0.3058      0.0250   12.2550    0.0000 [ 0.2557; 0.3478 ] 
#>   eta2 <~ y22      0.4473      0.0214   20.9064    0.0000 [ 0.4055; 0.4874 ] 
#>   eta2 <~ y23      0.4735      0.0181   26.2323    0.0000 [ 0.4416; 0.5028 ] 
#>   eta3 <~ y31      0.4400      0.0199   22.0760    0.0000 [ 0.3995; 0.4687 ] 
#>   eta3 <~ y32      0.3521      0.0180   19.5896    0.0000 [ 0.3253; 0.3978 ] 
#>   eta3 <~ y33      0.3999      0.0225   17.7683    0.0000 [ 0.3545; 0.4468 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0452   14.8399    0.0000 [ 0.5830; 0.7699 ] 
#>   eta3 ~ eta1       0.6634      0.0380   17.4501    0.0000 [ 0.5889; 0.7301 ] 
#>   eta3 ~ eta2       0.3052      0.0830    3.6754    0.0002 [ 0.1701; 0.4607 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0576    3.5553    0.0004 [ 0.1085; 0.3129 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04083031 16.239649  2.644328e-59
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03003797 21.615242 1.291245e-103
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03681832 20.678452  5.415029e-95
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05208627  9.915373  3.569210e-23
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03114730 24.252105 6.282600e-130
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03876464 20.628690  1.517035e-94
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03923605 20.957189  1.613389e-97
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03120010 21.091888  9.441798e-99
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04780646 15.634376  4.246121e-55
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6033109          0.7520044
#> 2          0.5976341          0.6974667
#> 3          0.7032580          0.8200333
#> 4          0.4273805          0.6216847
#> 5          0.6836730          0.8041132
#> 6          0.7177350          0.8508866
#> 7          0.7505459          0.8807830
#> 8          0.6163936          0.7113956
#> 9          0.6458301          0.8257514

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1326984296
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0452   14.8399    0.0000 [ 0.5513; 0.7853 ] 
#>   eta3 ~ eta1      0.4585      0.0815    5.6253    0.0000 [ 0.2600; 0.6815 ] 
#>   eta3 ~ eta2      0.3052      0.0830    3.6754    0.0002 [ 0.0800; 0.5094 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0408   16.2396    0.0000 [ 0.5517; 0.7629 ] 
#>   eta1 =~ y12      0.6493      0.0300   21.6152    0.0000 [ 0.5732; 0.7285 ] 
#>   eta1 =~ y13      0.7613      0.0368   20.6785    0.0000 [ 0.6671; 0.8575 ] 
#>   eta2 =~ y21      0.5165      0.0521    9.9154    0.0000 [ 0.3805; 0.6499 ] 
#>   eta2 =~ y22      0.7554      0.0311   24.2521    0.0000 [ 0.6715; 0.8326 ] 
#>   eta2 =~ y23      0.7997      0.0388   20.6287    0.0000 [ 0.7026; 0.9031 ] 
#>   eta3 =~ y31      0.8223      0.0392   20.9572    0.0000 [ 0.7275; 0.9304 ] 
#>   eta3 =~ y32      0.6581      0.0312   21.0919    0.0000 [ 0.5719; 0.7332 ] 
#>   eta3 =~ y33      0.7474      0.0478   15.6344    0.0000 [ 0.6152; 0.8624 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0208   19.0256    0.0000 [ 0.3395; 0.4470 ] 
#>   eta1 <~ y12      0.3873      0.0138   28.0958    0.0000 [ 0.3538; 0.4250 ] 
#>   eta1 <~ y13      0.4542      0.0232   19.6004    0.0000 [ 0.3959; 0.5157 ] 
#>   eta2 <~ y21      0.3058      0.0250   12.2550    0.0000 [ 0.2414; 0.3705 ] 
#>   eta2 <~ y22      0.4473      0.0214   20.9064    0.0000 [ 0.3904; 0.5011 ] 
#>   eta2 <~ y23      0.4735      0.0181   26.2323    0.0000 [ 0.4295; 0.5229 ] 
#>   eta3 <~ y31      0.4400      0.0199   22.0760    0.0000 [ 0.3943; 0.4974 ] 
#>   eta3 <~ y32      0.3521      0.0180   19.5896    0.0000 [ 0.3045; 0.3974 ] 
#>   eta3 <~ y33      0.3999      0.0225   17.7683    0.0000 [ 0.3394; 0.4558 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0452   14.8399    0.0000 [ 0.5513; 0.7853 ] 
#>   eta3 ~ eta1       0.6634      0.0380   17.4501    0.0000 [ 0.5693; 0.7659 ] 
#>   eta3 ~ eta2       0.3052      0.0830    3.6754    0.0002 [ 0.0800; 0.5094 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0576    3.5553    0.0004 [ 0.0478; 0.3458 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04523829 14.839938 8.083114e-50
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08150739  5.625340 1.851431e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08302449  3.675435 2.374444e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55132078          0.7852677          0.5794132          0.7571752
#> 2         0.25997346          0.6814837          0.3105886          0.6308685
#> 3         0.08004008          0.5093959          0.1315973          0.4578386
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5723232          0.7837369          0.5829701          0.7698951
#> 2          0.2776316          0.6300120          0.2925817          0.5828932
#> 3          0.1227381          0.4831321          0.1700678          0.4606905