Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1073400290
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0412   16.2955    0.0000 [ 0.6144; 0.7587 ] 
#>   eta3 ~ eta1      0.4585      0.0925    4.9575    0.0000 [ 0.2975; 0.6140 ] 
#>   eta3 ~ eta2      0.3052      0.0962    3.1733    0.0015 [ 0.1335; 0.4559 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0378   17.5433    0.0000 [ 0.5982; 0.7391 ] 
#>   eta1 =~ y12      0.6493      0.0431   15.0771    0.0000 [ 0.5569; 0.7034 ] 
#>   eta1 =~ y13      0.7613      0.0326   23.3751    0.0000 [ 0.7241; 0.8267 ] 
#>   eta2 =~ y21      0.5165      0.0509   10.1501    0.0000 [ 0.4028; 0.5819 ] 
#>   eta2 =~ y22      0.7554      0.0357   21.1623    0.0000 [ 0.7060; 0.8352 ] 
#>   eta2 =~ y23      0.7997      0.0367   21.7946    0.0000 [ 0.7371; 0.8644 ] 
#>   eta3 =~ y31      0.8223      0.0307   26.7953    0.0000 [ 0.7818; 0.8793 ] 
#>   eta3 =~ y32      0.6581      0.0371   17.7171    0.0000 [ 0.5814; 0.7119 ] 
#>   eta3 =~ y33      0.7474      0.0335   22.3212    0.0000 [ 0.6654; 0.8012 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0243   16.2502    0.0000 [ 0.3575; 0.4484 ] 
#>   eta1 <~ y12      0.3873      0.0201   19.2954    0.0000 [ 0.3429; 0.4224 ] 
#>   eta1 <~ y13      0.4542      0.0162   28.0778    0.0000 [ 0.4258; 0.4799 ] 
#>   eta2 <~ y21      0.3058      0.0270   11.3091    0.0000 [ 0.2417; 0.3419 ] 
#>   eta2 <~ y22      0.4473      0.0174   25.7546    0.0000 [ 0.4231; 0.4796 ] 
#>   eta2 <~ y23      0.4735      0.0218   21.7605    0.0000 [ 0.4403; 0.5157 ] 
#>   eta3 <~ y31      0.4400      0.0147   30.0166    0.0000 [ 0.4161; 0.4702 ] 
#>   eta3 <~ y32      0.3521      0.0179   19.7150    0.0000 [ 0.3177; 0.3857 ] 
#>   eta3 <~ y33      0.3999      0.0183   21.8407    0.0000 [ 0.3709; 0.4378 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0412   16.2955    0.0000 [ 0.6144; 0.7587 ] 
#>   eta3 ~ eta1       0.6634      0.0417   15.8952    0.0000 [ 0.5888; 0.7379 ] 
#>   eta3 ~ eta2       0.3052      0.0962    3.1733    0.0015 [ 0.1335; 0.4559 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0668    3.0689    0.0021 [ 0.0943; 0.3050 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03779611 17.54334  6.688259e-69
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04306392 15.07707  2.292060e-51
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03257086 23.37506 7.665470e-121
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05088190 10.15007  3.311278e-24
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03569496 21.16230  2.125810e-99
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03669085 21.79464 2.608461e-105
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03068742 26.79526 3.669961e-158
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03714305 17.71715  3.091855e-70
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03348487 22.32125 2.297820e-110
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5981700          0.7390912
#> 2          0.5568578          0.7034233
#> 3          0.7241060          0.8266538
#> 4          0.4027993          0.5819025
#> 5          0.7059637          0.8351925
#> 6          0.7370968          0.8643827
#> 7          0.7817618          0.8792849
#> 8          0.5813644          0.7118900
#> 9          0.6653969          0.8012027

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1073400290
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0412   16.2955    0.0000 [ 0.5563; 0.7693 ] 
#>   eta3 ~ eta1      0.4585      0.0925    4.9575    0.0000 [ 0.2110; 0.6893 ] 
#>   eta3 ~ eta2      0.3052      0.0962    3.1733    0.0015 [ 0.0595; 0.5568 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0378   17.5433    0.0000 [ 0.5552; 0.7507 ] 
#>   eta1 =~ y12      0.6493      0.0431   15.0771    0.0000 [ 0.5346; 0.7574 ] 
#>   eta1 =~ y13      0.7613      0.0326   23.3751    0.0000 [ 0.6706; 0.8390 ] 
#>   eta2 =~ y21      0.5165      0.0509   10.1501    0.0000 [ 0.4014; 0.6646 ] 
#>   eta2 =~ y22      0.7554      0.0357   21.1623    0.0000 [ 0.6614; 0.8460 ] 
#>   eta2 =~ y23      0.7997      0.0367   21.7946    0.0000 [ 0.7060; 0.8958 ] 
#>   eta3 =~ y31      0.8223      0.0307   26.7953    0.0000 [ 0.7406; 0.8993 ] 
#>   eta3 =~ y32      0.6581      0.0371   17.7171    0.0000 [ 0.5638; 0.7559 ] 
#>   eta3 =~ y33      0.7474      0.0335   22.3212    0.0000 [ 0.6691; 0.8423 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0243   16.2502    0.0000 [ 0.3323; 0.4581 ] 
#>   eta1 <~ y12      0.3873      0.0201   19.2954    0.0000 [ 0.3394; 0.4432 ] 
#>   eta1 <~ y13      0.4542      0.0162   28.0778    0.0000 [ 0.4152; 0.4989 ] 
#>   eta2 <~ y21      0.3058      0.0270   11.3091    0.0000 [ 0.2432; 0.3831 ] 
#>   eta2 <~ y22      0.4473      0.0174   25.7546    0.0000 [ 0.3974; 0.4872 ] 
#>   eta2 <~ y23      0.4735      0.0218   21.7605    0.0000 [ 0.4136; 0.5261 ] 
#>   eta3 <~ y31      0.4400      0.0147   30.0166    0.0000 [ 0.3987; 0.4745 ] 
#>   eta3 <~ y32      0.3521      0.0179   19.7150    0.0000 [ 0.3053; 0.3976 ] 
#>   eta3 <~ y33      0.3999      0.0183   21.8407    0.0000 [ 0.3550; 0.4497 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0412   16.2955    0.0000 [ 0.5563; 0.7693 ] 
#>   eta3 ~ eta1       0.6634      0.0417   15.8952    0.0000 [ 0.5467; 0.7625 ] 
#>   eta3 ~ eta2       0.3052      0.0962    3.1733    0.0015 [ 0.0595; 0.5568 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0668    3.0689    0.0021 [ 0.0319; 0.3771 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04119748 16.295496 1.062385e-59
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09248667  4.957544 7.138983e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09616145  3.173321 1.507060e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55629668          0.7693468          0.5818798          0.7437637
#> 2         0.21097971          0.6892686          0.2684129          0.6318354
#> 3         0.05946381          0.5567565          0.1191789          0.4970414
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.6075200          0.7787045          0.6143854          0.7587239
#> 2          0.2517407          0.6447561          0.2974785          0.6140379
#> 3          0.1308214          0.4957834          0.1334544          0.4558797