Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 488057933
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0487   13.7874    0.0000 [ 0.5727; 0.7616 ] 
#>   eta3 ~ eta1      0.4585      0.0714    6.4213    0.0000 [ 0.3487; 0.6183 ] 
#>   eta3 ~ eta2      0.3052      0.0798    3.8250    0.0001 [ 0.1534; 0.4278 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0378   17.5617    0.0000 [ 0.6030; 0.7451 ] 
#>   eta1 =~ y12      0.6493      0.0328   19.8060    0.0000 [ 0.5967; 0.7014 ] 
#>   eta1 =~ y13      0.7613      0.0318   23.9116    0.0000 [ 0.7114; 0.8172 ] 
#>   eta2 =~ y21      0.5165      0.0528    9.7814    0.0000 [ 0.4305; 0.6102 ] 
#>   eta2 =~ y22      0.7554      0.0377   20.0312    0.0000 [ 0.6857; 0.8158 ] 
#>   eta2 =~ y23      0.7997      0.0305   26.2171    0.0000 [ 0.7457; 0.8531 ] 
#>   eta3 =~ y31      0.8223      0.0360   22.8303    0.0000 [ 0.7647; 0.8905 ] 
#>   eta3 =~ y32      0.6581      0.0385   17.1089    0.0000 [ 0.5811; 0.7228 ] 
#>   eta3 =~ y33      0.7474      0.0387   19.3061    0.0000 [ 0.6716; 0.8177 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0208   18.9852    0.0000 [ 0.3659; 0.4402 ] 
#>   eta1 <~ y12      0.3873      0.0169   22.9620    0.0000 [ 0.3589; 0.4168 ] 
#>   eta1 <~ y13      0.4542      0.0186   24.3961    0.0000 [ 0.4116; 0.4888 ] 
#>   eta2 <~ y21      0.3058      0.0263   11.6065    0.0000 [ 0.2645; 0.3525 ] 
#>   eta2 <~ y22      0.4473      0.0201   22.2707    0.0000 [ 0.4121; 0.4824 ] 
#>   eta2 <~ y23      0.4735      0.0213   22.1856    0.0000 [ 0.4339; 0.5032 ] 
#>   eta3 <~ y31      0.4400      0.0202   21.7425    0.0000 [ 0.4000; 0.4744 ] 
#>   eta3 <~ y32      0.3521      0.0179   19.6971    0.0000 [ 0.3179; 0.3837 ] 
#>   eta3 <~ y33      0.3999      0.0172   23.2128    0.0000 [ 0.3668; 0.4243 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0487   13.7874    0.0000 [ 0.5727; 0.7616 ] 
#>   eta3 ~ eta1       0.6634      0.0369   17.9611    0.0000 [ 0.6058; 0.7302 ] 
#>   eta3 ~ eta2       0.3052      0.0798    3.8250    0.0001 [ 0.1534; 0.4278 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0523    3.9139    0.0001 [ 0.1053; 0.2970 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03775669 17.561651  4.844521e-69
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03278195 19.805959  2.644805e-87
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03184005 23.911573 2.321374e-126
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05279960  9.781416  1.353046e-22
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03771059 20.031179  2.945990e-89
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03050165 26.217065 1.697982e-151
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03601692 22.830306 2.293442e-115
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03846347 17.108931  1.273147e-65
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03871431 19.306147  4.768491e-83
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6030360          0.7451245
#> 2          0.5966650          0.7013542
#> 3          0.7113665          0.8171991
#> 4          0.4305034          0.6102493
#> 5          0.6857393          0.8157819
#> 6          0.7457172          0.8531227
#> 7          0.7646686          0.8905152
#> 8          0.5810978          0.7227885
#> 9          0.6715943          0.8177313

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 488057933
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0487   13.7874    0.0000 [ 0.5464; 0.7982 ] 
#>   eta3 ~ eta1      0.4585      0.0714    6.4213    0.0000 [ 0.2668; 0.6361 ] 
#>   eta3 ~ eta2      0.3052      0.0798    3.8250    0.0001 [ 0.1102; 0.5227 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0378   17.5617    0.0000 [ 0.5655; 0.7608 ] 
#>   eta1 =~ y12      0.6493      0.0328   19.8060    0.0000 [ 0.5624; 0.7320 ] 
#>   eta1 =~ y13      0.7613      0.0318   23.9116    0.0000 [ 0.6762; 0.8408 ] 
#>   eta2 =~ y21      0.5165      0.0528    9.7814    0.0000 [ 0.3721; 0.6452 ] 
#>   eta2 =~ y22      0.7554      0.0377   20.0312    0.0000 [ 0.6594; 0.8545 ] 
#>   eta2 =~ y23      0.7997      0.0305   26.2171    0.0000 [ 0.7257; 0.8834 ] 
#>   eta3 =~ y31      0.8223      0.0360   22.8303    0.0000 [ 0.7239; 0.9102 ] 
#>   eta3 =~ y32      0.6581      0.0385   17.1089    0.0000 [ 0.5615; 0.7604 ] 
#>   eta3 =~ y33      0.7474      0.0387   19.3061    0.0000 [ 0.6464; 0.8466 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0208   18.9852    0.0000 [ 0.3433; 0.4510 ] 
#>   eta1 <~ y12      0.3873      0.0169   22.9620    0.0000 [ 0.3441; 0.4313 ] 
#>   eta1 <~ y13      0.4542      0.0186   24.3961    0.0000 [ 0.4061; 0.5024 ] 
#>   eta2 <~ y21      0.3058      0.0263   11.6065    0.0000 [ 0.2337; 0.3700 ] 
#>   eta2 <~ y22      0.4473      0.0201   22.2707    0.0000 [ 0.3967; 0.5006 ] 
#>   eta2 <~ y23      0.4735      0.0213   22.1856    0.0000 [ 0.4214; 0.5318 ] 
#>   eta3 <~ y31      0.4400      0.0202   21.7425    0.0000 [ 0.3860; 0.4906 ] 
#>   eta3 <~ y32      0.3521      0.0179   19.6971    0.0000 [ 0.3085; 0.4009 ] 
#>   eta3 <~ y33      0.3999      0.0172   23.2128    0.0000 [ 0.3561; 0.4452 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0487   13.7874    0.0000 [ 0.5464; 0.7982 ] 
#>   eta3 ~ eta1       0.6634      0.0369   17.9611    0.0000 [ 0.5696; 0.7606 ] 
#>   eta3 ~ eta2       0.3052      0.0798    3.8250    0.0001 [ 0.1102; 0.5227 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0523    3.9139    0.0001 [ 0.0783; 0.3490 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04869195 13.787360 3.036716e-43
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07140449  6.421260 1.351509e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07977873  3.824968 1.307889e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5464110          0.7982183          0.5766482          0.7679812
#> 2          0.2667952          0.6360589          0.3111366          0.5917176
#> 3          0.1101686          0.5227392          0.1597103          0.4731975
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5698107          0.8014157          0.5726874          0.7616253
#> 2          0.3169513          0.6225863          0.3486751          0.6183069
#> 3          0.1241483          0.4569952          0.1534327          0.4278321