Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -577253996
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0408   16.4427    0.0000 [ 0.5993; 0.7560 ] 
#>   eta3 ~ eta1      0.4585      0.0695    6.5991    0.0000 [ 0.3462; 0.6004 ] 
#>   eta3 ~ eta2      0.3052      0.0671    4.5449    0.0000 [ 0.1811; 0.4083 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0403   16.4558    0.0000 [ 0.5913; 0.7256 ] 
#>   eta1 =~ y12      0.6493      0.0407   15.9618    0.0000 [ 0.5781; 0.7193 ] 
#>   eta1 =~ y13      0.7613      0.0328   23.1951    0.0000 [ 0.6918; 0.8002 ] 
#>   eta2 =~ y21      0.5165      0.0598    8.6299    0.0000 [ 0.3926; 0.5935 ] 
#>   eta2 =~ y22      0.7554      0.0310   24.3841    0.0000 [ 0.7016; 0.8122 ] 
#>   eta2 =~ y23      0.7997      0.0325   24.6194    0.0000 [ 0.7381; 0.8535 ] 
#>   eta3 =~ y31      0.8223      0.0332   24.7538    0.0000 [ 0.7781; 0.8782 ] 
#>   eta3 =~ y32      0.6581      0.0401   16.4151    0.0000 [ 0.5882; 0.7293 ] 
#>   eta3 =~ y33      0.7474      0.0347   21.5374    0.0000 [ 0.6826; 0.8060 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0203   19.4501    0.0000 [ 0.3732; 0.4453 ] 
#>   eta1 <~ y12      0.3873      0.0201   19.3132    0.0000 [ 0.3488; 0.4261 ] 
#>   eta1 <~ y13      0.4542      0.0197   23.0916    0.0000 [ 0.4214; 0.4828 ] 
#>   eta2 <~ y21      0.3058      0.0301   10.1454    0.0000 [ 0.2501; 0.3471 ] 
#>   eta2 <~ y22      0.4473      0.0208   21.5504    0.0000 [ 0.4078; 0.4794 ] 
#>   eta2 <~ y23      0.4735      0.0203   23.3150    0.0000 [ 0.4278; 0.5138 ] 
#>   eta3 <~ y31      0.4400      0.0164   26.8877    0.0000 [ 0.4067; 0.4709 ] 
#>   eta3 <~ y32      0.3521      0.0175   20.1524    0.0000 [ 0.3171; 0.3795 ] 
#>   eta3 <~ y33      0.3999      0.0193   20.7106    0.0000 [ 0.3636; 0.4264 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0408   16.4427    0.0000 [ 0.5993; 0.7560 ] 
#>   eta3 ~ eta1       0.6634      0.0343   19.3236    0.0000 [ 0.6274; 0.7396 ] 
#>   eta3 ~ eta2       0.3052      0.0671    4.5449    0.0000 [ 0.1811; 0.4083 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0465    4.4039    0.0000 [ 0.1262; 0.2934 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04029401 16.455793  7.621771e-61
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04067703 15.961782  2.358981e-57
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03282356 23.195103 5.102079e-119
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05984490  8.629889  6.141146e-18
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03097874 24.384069 2.523982e-131
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03248106 24.619384 7.833375e-134
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03321825 24.753786 2.822544e-135
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04008913 16.415146  1.490215e-60
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03470354 21.537401 6.949888e-103
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5912798          0.7255647
#> 2          0.5781249          0.7192672
#> 3          0.6917738          0.8002081
#> 4          0.3925501          0.5934583
#> 5          0.7015675          0.8122271
#> 6          0.7380864          0.8535169
#> 7          0.7780998          0.8782446
#> 8          0.5881616          0.7293076
#> 9          0.6826427          0.8059762

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -577253996
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0408   16.4427    0.0000 [ 0.5548; 0.7659 ] 
#>   eta3 ~ eta1      0.4585      0.0695    6.5991    0.0000 [ 0.2603; 0.6196 ] 
#>   eta3 ~ eta2      0.3052      0.0671    4.5449    0.0000 [ 0.1467; 0.4939 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0403   16.4558    0.0000 [ 0.5572; 0.7656 ] 
#>   eta1 =~ y12      0.6493      0.0407   15.9618    0.0000 [ 0.5556; 0.7659 ] 
#>   eta1 =~ y13      0.7613      0.0328   23.1951    0.0000 [ 0.6870; 0.8567 ] 
#>   eta2 =~ y21      0.5165      0.0598    8.6299    0.0000 [ 0.3584; 0.6679 ] 
#>   eta2 =~ y22      0.7554      0.0310   24.3841    0.0000 [ 0.6733; 0.8335 ] 
#>   eta2 =~ y23      0.7997      0.0325   24.6194    0.0000 [ 0.7085; 0.8764 ] 
#>   eta3 =~ y31      0.8223      0.0332   24.7538    0.0000 [ 0.7326; 0.9044 ] 
#>   eta3 =~ y32      0.6581      0.0401   16.4151    0.0000 [ 0.5458; 0.7532 ] 
#>   eta3 =~ y33      0.7474      0.0347   21.5374    0.0000 [ 0.6610; 0.8404 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0203   19.4501    0.0000 [ 0.3364; 0.4415 ] 
#>   eta1 <~ y12      0.3873      0.0201   19.3132    0.0000 [ 0.3370; 0.4407 ] 
#>   eta1 <~ y13      0.4542      0.0197   23.0916    0.0000 [ 0.4030; 0.5047 ] 
#>   eta2 <~ y21      0.3058      0.0301   10.1454    0.0000 [ 0.2296; 0.3854 ] 
#>   eta2 <~ y22      0.4473      0.0208   21.5504    0.0000 [ 0.3968; 0.5041 ] 
#>   eta2 <~ y23      0.4735      0.0203   23.3150    0.0000 [ 0.4214; 0.5264 ] 
#>   eta3 <~ y31      0.4400      0.0164   26.8877    0.0000 [ 0.3985; 0.4831 ] 
#>   eta3 <~ y32      0.3521      0.0175   20.1524    0.0000 [ 0.3048; 0.3951 ] 
#>   eta3 <~ y33      0.3999      0.0193   20.7106    0.0000 [ 0.3542; 0.4540 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0408   16.4427    0.0000 [ 0.5548; 0.7659 ] 
#>   eta3 ~ eta1       0.6634      0.0343   19.3236    0.0000 [ 0.5634; 0.7409 ] 
#>   eta3 ~ eta2       0.3052      0.0671    4.5449    0.0000 [ 0.1467; 0.4939 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0465    4.4039    0.0000 [ 0.0919; 0.3325 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04082865 16.442703 9.460292e-61
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.06948062  6.599060 4.137732e-11
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.06714090  4.544936 5.495189e-06
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5547637          0.7659065          0.5801179          0.7405524
#> 2          0.2603042          0.6196188          0.3034509          0.5764721
#> 3          0.1466820          0.4938968          0.1883757          0.4522031
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5871431          0.7677490          0.5992877          0.7559712
#> 2          0.3219413          0.6112924          0.3461588          0.6003614
#> 3          0.1662897          0.4092559          0.1810539          0.4082967