Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -197093687
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0440   15.2457    0.0000 [ 0.5927; 0.7396 ] 
#>   eta3 ~ eta1      0.4585      0.1010    4.5396    0.0000 [ 0.2503; 0.6381 ] 
#>   eta3 ~ eta2      0.3052      0.0994    3.0713    0.0021 [ 0.1261; 0.4772 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0356   18.6083    0.0000 [ 0.5875; 0.7182 ] 
#>   eta1 =~ y12      0.6493      0.0402   16.1452    0.0000 [ 0.5630; 0.7070 ] 
#>   eta1 =~ y13      0.7613      0.0364   20.9441    0.0000 [ 0.6963; 0.8160 ] 
#>   eta2 =~ y21      0.5165      0.0456   11.3223    0.0000 [ 0.4456; 0.5893 ] 
#>   eta2 =~ y22      0.7554      0.0339   22.3151    0.0000 [ 0.6649; 0.7997 ] 
#>   eta2 =~ y23      0.7997      0.0281   28.4580    0.0000 [ 0.7470; 0.8570 ] 
#>   eta3 =~ y31      0.8223      0.0358   22.9695    0.0000 [ 0.7487; 0.8717 ] 
#>   eta3 =~ y32      0.6581      0.0421   15.6178    0.0000 [ 0.5759; 0.7258 ] 
#>   eta3 =~ y33      0.7474      0.0400   18.6978    0.0000 [ 0.6777; 0.8254 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0188   21.0568    0.0000 [ 0.3568; 0.4330 ] 
#>   eta1 <~ y12      0.3873      0.0219   17.6759    0.0000 [ 0.3532; 0.4305 ] 
#>   eta1 <~ y13      0.4542      0.0210   21.6286    0.0000 [ 0.4213; 0.4922 ] 
#>   eta2 <~ y21      0.3058      0.0226   13.5560    0.0000 [ 0.2691; 0.3368 ] 
#>   eta2 <~ y22      0.4473      0.0167   26.7825    0.0000 [ 0.4098; 0.4709 ] 
#>   eta2 <~ y23      0.4735      0.0208   22.8020    0.0000 [ 0.4407; 0.5148 ] 
#>   eta3 <~ y31      0.4400      0.0175   25.1635    0.0000 [ 0.4072; 0.4603 ] 
#>   eta3 <~ y32      0.3521      0.0234   15.0710    0.0000 [ 0.3112; 0.3908 ] 
#>   eta3 <~ y33      0.3999      0.0194   20.6380    0.0000 [ 0.3713; 0.4433 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0440   15.2457    0.0000 [ 0.5927; 0.7396 ] 
#>   eta3 ~ eta1       0.6634      0.0463   14.3315    0.0000 [ 0.5838; 0.7388 ] 
#>   eta3 ~ eta2       0.3052      0.0994    3.0713    0.0021 [ 0.1261; 0.4772 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0648    3.1595    0.0016 [ 0.0892; 0.3166 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03563306 18.60828  2.753391e-77
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04021484 16.14523  1.226897e-58
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03635136 20.94408  2.124564e-97
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04561405 11.32227  1.017965e-29
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03385100 22.31508 2.637713e-110
#> 6 eta2 =~ y23  Common factor 0.7996637 0.02809983 28.45795 3.885041e-178
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03579871 22.96947 9.415541e-117
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04213575 15.61783  5.504861e-55
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03997394 18.69779  5.160313e-78
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5875324          0.7181638
#> 2          0.5629533          0.7070401
#> 3          0.6962848          0.8159733
#> 4          0.4456175          0.5892938
#> 5          0.6648752          0.7997302
#> 6          0.7470019          0.8570349
#> 7          0.7486606          0.8717089
#> 8          0.5759303          0.7258132
#> 9          0.6776545          0.8254432

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -197093687
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0440   15.2457    0.0000 [ 0.5499; 0.7776 ] 
#>   eta3 ~ eta1      0.4585      0.1010    4.5396    0.0000 [ 0.1783; 0.7006 ] 
#>   eta3 ~ eta2      0.3052      0.0994    3.0713    0.0021 [ 0.0671; 0.5809 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0356   18.6083    0.0000 [ 0.5766; 0.7609 ] 
#>   eta1 =~ y12      0.6493      0.0402   16.1452    0.0000 [ 0.5457; 0.7536 ] 
#>   eta1 =~ y13      0.7613      0.0364   20.9441    0.0000 [ 0.6717; 0.8597 ] 
#>   eta2 =~ y21      0.5165      0.0456   11.3223    0.0000 [ 0.3997; 0.6355 ] 
#>   eta2 =~ y22      0.7554      0.0339   22.3151    0.0000 [ 0.6747; 0.8498 ] 
#>   eta2 =~ y23      0.7997      0.0281   28.4580    0.0000 [ 0.7196; 0.8649 ] 
#>   eta3 =~ y31      0.8223      0.0358   22.9695    0.0000 [ 0.7375; 0.9226 ] 
#>   eta3 =~ y32      0.6581      0.0421   15.6178    0.0000 [ 0.5577; 0.7756 ] 
#>   eta3 =~ y33      0.7474      0.0400   18.6978    0.0000 [ 0.6339; 0.8406 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0188   21.0568    0.0000 [ 0.3477; 0.4448 ] 
#>   eta1 <~ y12      0.3873      0.0219   17.6759    0.0000 [ 0.3282; 0.4415 ] 
#>   eta1 <~ y13      0.4542      0.0210   21.6286    0.0000 [ 0.3992; 0.5078 ] 
#>   eta2 <~ y21      0.3058      0.0226   13.5560    0.0000 [ 0.2486; 0.3652 ] 
#>   eta2 <~ y22      0.4473      0.0167   26.7825    0.0000 [ 0.4084; 0.4948 ] 
#>   eta2 <~ y23      0.4735      0.0208   22.8020    0.0000 [ 0.4154; 0.5228 ] 
#>   eta3 <~ y31      0.4400      0.0175   25.1635    0.0000 [ 0.3975; 0.4879 ] 
#>   eta3 <~ y32      0.3521      0.0234   15.0710    0.0000 [ 0.2950; 0.4158 ] 
#>   eta3 <~ y33      0.3999      0.0194   20.6380    0.0000 [ 0.3431; 0.4433 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0440   15.2457    0.0000 [ 0.5499; 0.7776 ] 
#>   eta3 ~ eta1       0.6634      0.0463   14.3315    0.0000 [ 0.5364; 0.7757 ] 
#>   eta3 ~ eta2       0.3052      0.0994    3.0713    0.0021 [ 0.0671; 0.5809 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0648    3.1595    0.0016 [ 0.0490; 0.3843 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04403417 15.245738 1.757354e-52
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.10100240  4.539563 5.637085e-06
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09935516  3.071316 2.131172e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54991925          0.7776391          0.5772640          0.7502944
#> 2         0.17826905          0.7005964          0.2409904          0.6378751
#> 3         0.06707144          0.5808802          0.1287698          0.5191818
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5656360          0.7531224          0.5927341          0.7395692
#> 2          0.2436545          0.6515373          0.2502768          0.6381128
#> 3          0.1159778          0.4816263          0.1260933          0.4771503