Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1784386869
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0523   12.8400    0.0000 [ 0.5737; 0.7574 ] 
#>   eta3 ~ eta1      0.4585      0.0984    4.6617    0.0000 [ 0.3167; 0.6998 ] 
#>   eta3 ~ eta2      0.3052      0.0989    3.0859    0.0020 [ 0.0900; 0.4538 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0407   16.2909    0.0000 [ 0.5985; 0.7127 ] 
#>   eta1 =~ y12      0.6493      0.0319   20.3812    0.0000 [ 0.5915; 0.7061 ] 
#>   eta1 =~ y13      0.7613      0.0358   21.2791    0.0000 [ 0.7052; 0.8312 ] 
#>   eta2 =~ y21      0.5165      0.0623    8.2867    0.0000 [ 0.3600; 0.6114 ] 
#>   eta2 =~ y22      0.7554      0.0365   20.7045    0.0000 [ 0.7061; 0.8182 ] 
#>   eta2 =~ y23      0.7997      0.0303   26.3883    0.0000 [ 0.7556; 0.8543 ] 
#>   eta3 =~ y31      0.8223      0.0326   25.2214    0.0000 [ 0.7511; 0.8711 ] 
#>   eta3 =~ y32      0.6581      0.0340   19.3279    0.0000 [ 0.5948; 0.7166 ] 
#>   eta3 =~ y33      0.7474      0.0523   14.2782    0.0000 [ 0.6426; 0.8433 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0219   18.0573    0.0000 [ 0.3610; 0.4392 ] 
#>   eta1 <~ y12      0.3873      0.0219   17.6827    0.0000 [ 0.3536; 0.4265 ] 
#>   eta1 <~ y13      0.4542      0.0169   26.8308    0.0000 [ 0.4217; 0.4874 ] 
#>   eta2 <~ y21      0.3058      0.0335    9.1327    0.0000 [ 0.2245; 0.3475 ] 
#>   eta2 <~ y22      0.4473      0.0239   18.7164    0.0000 [ 0.4113; 0.5029 ] 
#>   eta2 <~ y23      0.4735      0.0179   26.5166    0.0000 [ 0.4434; 0.5094 ] 
#>   eta3 <~ y31      0.4400      0.0214   20.5384    0.0000 [ 0.4023; 0.4845 ] 
#>   eta3 <~ y32      0.3521      0.0170   20.6676    0.0000 [ 0.3241; 0.3910 ] 
#>   eta3 <~ y33      0.3999      0.0213   18.7341    0.0000 [ 0.3593; 0.4432 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0523   12.8400    0.0000 [ 0.5737; 0.7574 ] 
#>   eta3 ~ eta1       0.6634      0.0452   14.6653    0.0000 [ 0.5882; 0.7579 ] 
#>   eta3 ~ eta2       0.3052      0.0989    3.0859    0.0020 [ 0.0900; 0.4538 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0667    3.0720    0.0021 [ 0.0581; 0.3171 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04070189 16.290887  1.145552e-59
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03185674 20.381178  2.456737e-92
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03577912 21.279055 1.774852e-100
#> 4 eta2 =~ y21  Common factor 0.5164548 0.06232324  8.286713  1.164209e-16
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03648414 20.704550  3.151649e-95
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03030372 26.388305 1.866483e-153
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03260239 25.221383 2.334473e-140
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03404758 19.327918  3.127880e-83
#> 9 eta3 =~ y33  Common factor 0.7474241 0.05234712 14.278229  2.991045e-46
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5985334          0.7127019
#> 2          0.5914981          0.7060881
#> 3          0.7052067          0.8312377
#> 4          0.3599837          0.6113789
#> 5          0.7061262          0.8182371
#> 6          0.7556139          0.8542691
#> 7          0.7511178          0.8710643
#> 8          0.5947974          0.7166006
#> 9          0.6426194          0.8432711

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1784386869
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0523   12.8400    0.0000 [ 0.5362; 0.8066 ] 
#>   eta3 ~ eta1      0.4585      0.0984    4.6617    0.0000 [ 0.2002; 0.7088 ] 
#>   eta3 ~ eta2      0.3052      0.0989    3.0859    0.0020 [ 0.0504; 0.5618 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0407   16.2909    0.0000 [ 0.5680; 0.7785 ] 
#>   eta1 =~ y12      0.6493      0.0319   20.3812    0.0000 [ 0.5683; 0.7331 ] 
#>   eta1 =~ y13      0.7613      0.0358   21.2791    0.0000 [ 0.6654; 0.8504 ] 
#>   eta2 =~ y21      0.5165      0.0623    8.2867    0.0000 [ 0.3749; 0.6972 ] 
#>   eta2 =~ y22      0.7554      0.0365   20.7045    0.0000 [ 0.6594; 0.8481 ] 
#>   eta2 =~ y23      0.7997      0.0303   26.3883    0.0000 [ 0.7214; 0.8781 ] 
#>   eta3 =~ y31      0.8223      0.0326   25.2214    0.0000 [ 0.7428; 0.9114 ] 
#>   eta3 =~ y32      0.6581      0.0340   19.3279    0.0000 [ 0.5686; 0.7447 ] 
#>   eta3 =~ y33      0.7474      0.0523   14.2782    0.0000 [ 0.6197; 0.8904 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0219   18.0573    0.0000 [ 0.3431; 0.4564 ] 
#>   eta1 <~ y12      0.3873      0.0219   17.6827    0.0000 [ 0.3294; 0.4426 ] 
#>   eta1 <~ y13      0.4542      0.0169   26.8308    0.0000 [ 0.4062; 0.4937 ] 
#>   eta2 <~ y21      0.3058      0.0335    9.1327    0.0000 [ 0.2284; 0.4016 ] 
#>   eta2 <~ y22      0.4473      0.0239   18.7164    0.0000 [ 0.3801; 0.5037 ] 
#>   eta2 <~ y23      0.4735      0.0179   26.5166    0.0000 [ 0.4228; 0.5152 ] 
#>   eta3 <~ y31      0.4400      0.0214   20.5384    0.0000 [ 0.3837; 0.4945 ] 
#>   eta3 <~ y32      0.3521      0.0170   20.6676    0.0000 [ 0.3047; 0.3928 ] 
#>   eta3 <~ y33      0.3999      0.0213   18.7341    0.0000 [ 0.3462; 0.4566 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0523   12.8400    0.0000 [ 0.5362; 0.8066 ] 
#>   eta3 ~ eta1       0.6634      0.0452   14.6653    0.0000 [ 0.5440; 0.7779 ] 
#>   eta3 ~ eta2       0.3052      0.0989    3.0859    0.0020 [ 0.0504; 0.5618 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0667    3.0720    0.0021 [ 0.0340; 0.3789 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05228460 12.839983 9.789536e-38
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09835528  4.661740 3.135468e-06
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09888574  3.085896 2.029398e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.53624963          0.8066361          0.5687178          0.7741679
#> 2         0.20017554          0.7088135          0.2612530          0.6477360
#> 3         0.05043882          0.5618201          0.1118457          0.5004132
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.55511991          0.7779839         0.57373840          0.7574072
#> 2         0.30274203          0.7251265         0.31674854          0.6998019
#> 3         0.05460404          0.4728001         0.09002176          0.4538113