Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -969804716
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0420   15.9916    0.0000 [ 0.5698; 0.7238 ] 
#>   eta3 ~ eta1      0.4585      0.0720    6.3718    0.0000 [ 0.3572; 0.6068 ] 
#>   eta3 ~ eta2      0.3052      0.0857    3.5591    0.0004 [ 0.1447; 0.4376 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0459   14.4529    0.0000 [ 0.5762; 0.7669 ] 
#>   eta1 =~ y12      0.6493      0.0345   18.8277    0.0000 [ 0.5737; 0.6961 ] 
#>   eta1 =~ y13      0.7613      0.0401   18.9882    0.0000 [ 0.7063; 0.8302 ] 
#>   eta2 =~ y21      0.5165      0.0596    8.6651    0.0000 [ 0.4020; 0.5966 ] 
#>   eta2 =~ y22      0.7554      0.0422   17.9066    0.0000 [ 0.6633; 0.8097 ] 
#>   eta2 =~ y23      0.7997      0.0465   17.1954    0.0000 [ 0.7164; 0.8718 ] 
#>   eta3 =~ y31      0.8223      0.0306   26.8389    0.0000 [ 0.7666; 0.8685 ] 
#>   eta3 =~ y32      0.6581      0.0460   14.3052    0.0000 [ 0.5637; 0.7241 ] 
#>   eta3 =~ y33      0.7474      0.0353   21.1878    0.0000 [ 0.6715; 0.8015 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0251   15.7360    0.0000 [ 0.3621; 0.4536 ] 
#>   eta1 <~ y12      0.3873      0.0203   19.0660    0.0000 [ 0.3515; 0.4243 ] 
#>   eta1 <~ y13      0.4542      0.0208   21.8733    0.0000 [ 0.4197; 0.4943 ] 
#>   eta2 <~ y21      0.3058      0.0291   10.5170    0.0000 [ 0.2544; 0.3558 ] 
#>   eta2 <~ y22      0.4473      0.0298   15.0017    0.0000 [ 0.4001; 0.5048 ] 
#>   eta2 <~ y23      0.4735      0.0216   21.9191    0.0000 [ 0.4477; 0.5221 ] 
#>   eta3 <~ y31      0.4400      0.0171   25.6823    0.0000 [ 0.4079; 0.4696 ] 
#>   eta3 <~ y32      0.3521      0.0199   17.6643    0.0000 [ 0.3170; 0.3835 ] 
#>   eta3 <~ y33      0.3999      0.0196   20.4253    0.0000 [ 0.3680; 0.4478 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0420   15.9916    0.0000 [ 0.5698; 0.7238 ] 
#>   eta3 ~ eta1       0.6634      0.0311   21.3022    0.0000 [ 0.6033; 0.7071 ] 
#>   eta3 ~ eta2       0.3052      0.0857    3.5591    0.0004 [ 0.1447; 0.4376 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0566    3.6197    0.0003 [ 0.1010; 0.2941 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04587798 14.452901  2.403449e-47
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03448523 18.827712  4.476920e-79
#> 3 eta1 =~ y13  Common factor 0.7613458 0.04009568 18.988226  2.134168e-80
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05960171  8.665101  4.511137e-18
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04218492 17.906581  1.047765e-71
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04650465 17.195350  2.877128e-66
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03063756 26.838866 1.137755e-158
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04600221 14.305158  2.031758e-46
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03527623 21.187758  1.238568e-99
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5762433          0.7669270
#> 2          0.5736849          0.6961158
#> 3          0.7062898          0.8302229
#> 4          0.4020028          0.5965647
#> 5          0.6632867          0.8097137
#> 6          0.7163869          0.8717697
#> 7          0.7666141          0.8684613
#> 8          0.5636745          0.7240859
#> 9          0.6715068          0.8014649

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -969804716
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0420   15.9916    0.0000 [ 0.5751; 0.7922 ] 
#>   eta3 ~ eta1      0.4585      0.0720    6.3718    0.0000 [ 0.2531; 0.6252 ] 
#>   eta3 ~ eta2      0.3052      0.0857    3.5591    0.0004 [ 0.1107; 0.5541 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0459   14.4529    0.0000 [ 0.5439; 0.7812 ] 
#>   eta1 =~ y12      0.6493      0.0345   18.8277    0.0000 [ 0.5672; 0.7455 ] 
#>   eta1 =~ y13      0.7613      0.0401   18.9882    0.0000 [ 0.6520; 0.8593 ] 
#>   eta2 =~ y21      0.5165      0.0596    8.6651    0.0000 [ 0.3694; 0.6776 ] 
#>   eta2 =~ y22      0.7554      0.0422   17.9066    0.0000 [ 0.6601; 0.8782 ] 
#>   eta2 =~ y23      0.7997      0.0465   17.1954    0.0000 [ 0.6827; 0.9232 ] 
#>   eta3 =~ y31      0.8223      0.0306   26.8389    0.0000 [ 0.7501; 0.9085 ] 
#>   eta3 =~ y32      0.6581      0.0460   14.3052    0.0000 [ 0.5482; 0.7861 ] 
#>   eta3 =~ y33      0.7474      0.0353   21.1878    0.0000 [ 0.6622; 0.8446 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0251   15.7360    0.0000 [ 0.3304; 0.4604 ] 
#>   eta1 <~ y12      0.3873      0.0203   19.0660    0.0000 [ 0.3390; 0.4441 ] 
#>   eta1 <~ y13      0.4542      0.0208   21.8733    0.0000 [ 0.3972; 0.5046 ] 
#>   eta2 <~ y21      0.3058      0.0291   10.5170    0.0000 [ 0.2305; 0.3808 ] 
#>   eta2 <~ y22      0.4473      0.0298   15.0017    0.0000 [ 0.3708; 0.5250 ] 
#>   eta2 <~ y23      0.4735      0.0216   21.9191    0.0000 [ 0.4121; 0.5238 ] 
#>   eta3 <~ y31      0.4400      0.0171   25.6823    0.0000 [ 0.3929; 0.4815 ] 
#>   eta3 <~ y32      0.3521      0.0199   17.6643    0.0000 [ 0.3006; 0.4037 ] 
#>   eta3 <~ y33      0.3999      0.0196   20.4253    0.0000 [ 0.3466; 0.4478 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0420   15.9916    0.0000 [ 0.5751; 0.7922 ] 
#>   eta3 ~ eta1       0.6634      0.0311   21.3022    0.0000 [ 0.5857; 0.7467 ] 
#>   eta3 ~ eta2       0.3052      0.0857    3.5591    0.0004 [ 0.1107; 0.5541 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0566    3.6197    0.0003 [ 0.0808; 0.3734 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04198037 15.991605 1.462143e-57
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07195872  6.371803 1.868186e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08573919  3.559062 3.721814e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5750942          0.7921930          0.6011635          0.7661236
#> 2          0.2530558          0.6251857          0.2977413          0.5805001
#> 3          0.1107423          0.5541370          0.1639853          0.5008939
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5660286          0.7327535          0.5698279          0.7237938
#> 2          0.3319887          0.6136829          0.3571688          0.6068010
#> 3          0.1348666          0.4522826          0.1446885          0.4375697