Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 242345779
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0429   15.6367    0.0000 [ 0.5893; 0.7553 ] 
#>   eta3 ~ eta1      0.4585      0.0828    5.5384    0.0000 [ 0.3428; 0.6538 ] 
#>   eta3 ~ eta2      0.3052      0.0804    3.7950    0.0001 [ 0.1265; 0.4099 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0355   18.6985    0.0000 [ 0.6152; 0.7302 ] 
#>   eta1 =~ y12      0.6493      0.0506   12.8228    0.0000 [ 0.5503; 0.7267 ] 
#>   eta1 =~ y13      0.7613      0.0268   28.4099    0.0000 [ 0.7174; 0.8045 ] 
#>   eta2 =~ y21      0.5165      0.0414   12.4811    0.0000 [ 0.4315; 0.5766 ] 
#>   eta2 =~ y22      0.7554      0.0306   24.7253    0.0000 [ 0.7012; 0.8168 ] 
#>   eta2 =~ y23      0.7997      0.0350   22.8619    0.0000 [ 0.7446; 0.8709 ] 
#>   eta3 =~ y31      0.8223      0.0330   24.9428    0.0000 [ 0.7496; 0.8638 ] 
#>   eta3 =~ y32      0.6581      0.0486   13.5409    0.0000 [ 0.5678; 0.7388 ] 
#>   eta3 =~ y33      0.7474      0.0415   18.0309    0.0000 [ 0.6777; 0.8260 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0215   18.4112    0.0000 [ 0.3647; 0.4445 ] 
#>   eta1 <~ y12      0.3873      0.0197   19.6441    0.0000 [ 0.3433; 0.4151 ] 
#>   eta1 <~ y13      0.4542      0.0213   21.2854    0.0000 [ 0.4098; 0.4880 ] 
#>   eta2 <~ y21      0.3058      0.0206   14.8375    0.0000 [ 0.2587; 0.3366 ] 
#>   eta2 <~ y22      0.4473      0.0150   29.7804    0.0000 [ 0.4178; 0.4726 ] 
#>   eta2 <~ y23      0.4735      0.0211   22.4183    0.0000 [ 0.4374; 0.5088 ] 
#>   eta3 <~ y31      0.4400      0.0179   24.5819    0.0000 [ 0.4105; 0.4746 ] 
#>   eta3 <~ y32      0.3521      0.0234   15.0355    0.0000 [ 0.3126; 0.4043 ] 
#>   eta3 <~ y33      0.3999      0.0216   18.4750    0.0000 [ 0.3586; 0.4357 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0429   15.6367    0.0000 [ 0.5893; 0.7553 ] 
#>   eta3 ~ eta1       0.6634      0.0451   14.7189    0.0000 [ 0.5829; 0.7430 ] 
#>   eta3 ~ eta2       0.3052      0.0804    3.7950    0.0001 [ 0.1265; 0.4099 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0532    3.8494    0.0001 [ 0.0888; 0.2744 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03546116 18.69848  5.093413e-78
#> 2 eta1 =~ y12  Common factor 0.6492779 0.05063478 12.82277  1.222596e-37
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02679864 28.40987 1.527204e-177
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04137890 12.48112  9.465146e-36
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03055124 24.72527 5.721540e-135
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03497794 22.86194 1.111665e-115
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03296655 24.94278 2.557719e-137
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04859875 13.54086  8.973163e-42
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04145239 18.03091  1.114508e-72
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6151855          0.7301820
#> 2          0.5503281          0.7267218
#> 3          0.7173658          0.8044648
#> 4          0.4315426          0.5765506
#> 5          0.7012013          0.8167965
#> 6          0.7446257          0.8709101
#> 7          0.7495544          0.8637640
#> 8          0.5678305          0.7388215
#> 9          0.6776576          0.8260180

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 242345779
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0429   15.6367    0.0000 [ 0.5530; 0.7751 ] 
#>   eta3 ~ eta1      0.4585      0.0828    5.5384    0.0000 [ 0.2406; 0.6687 ] 
#>   eta3 ~ eta2      0.3052      0.0804    3.7950    0.0001 [ 0.1036; 0.5195 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0355   18.6985    0.0000 [ 0.5628; 0.7462 ] 
#>   eta1 =~ y12      0.6493      0.0506   12.8228    0.0000 [ 0.5237; 0.7855 ] 
#>   eta1 =~ y13      0.7613      0.0268   28.4099    0.0000 [ 0.6944; 0.8329 ] 
#>   eta2 =~ y21      0.5165      0.0414   12.4811    0.0000 [ 0.4144; 0.6284 ] 
#>   eta2 =~ y22      0.7554      0.0306   24.7253    0.0000 [ 0.6795; 0.8375 ] 
#>   eta2 =~ y23      0.7997      0.0350   22.8619    0.0000 [ 0.6978; 0.8787 ] 
#>   eta3 =~ y31      0.8223      0.0330   24.9428    0.0000 [ 0.7400; 0.9105 ] 
#>   eta3 =~ y32      0.6581      0.0486   13.5409    0.0000 [ 0.5305; 0.7818 ] 
#>   eta3 =~ y33      0.7474      0.0415   18.0309    0.0000 [ 0.6332; 0.8476 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0215   18.4112    0.0000 [ 0.3352; 0.4463 ] 
#>   eta1 <~ y12      0.3873      0.0197   19.6441    0.0000 [ 0.3405; 0.4425 ] 
#>   eta1 <~ y13      0.4542      0.0213   21.2854    0.0000 [ 0.4006; 0.5109 ] 
#>   eta2 <~ y21      0.3058      0.0206   14.8375    0.0000 [ 0.2568; 0.3634 ] 
#>   eta2 <~ y22      0.4473      0.0150   29.7804    0.0000 [ 0.4119; 0.4896 ] 
#>   eta2 <~ y23      0.4735      0.0211   22.4183    0.0000 [ 0.4138; 0.5230 ] 
#>   eta3 <~ y31      0.4400      0.0179   24.5819    0.0000 [ 0.3972; 0.4898 ] 
#>   eta3 <~ y32      0.3521      0.0234   15.0355    0.0000 [ 0.2923; 0.4134 ] 
#>   eta3 <~ y33      0.3999      0.0216   18.4750    0.0000 [ 0.3420; 0.4540 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0429   15.6367    0.0000 [ 0.5530; 0.7751 ] 
#>   eta3 ~ eta1       0.6634      0.0451   14.7189    0.0000 [ 0.5459; 0.7790 ] 
#>   eta3 ~ eta2       0.3052      0.0804    3.7950    0.0001 [ 0.1036; 0.5195 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0532    3.8494    0.0001 [ 0.0702; 0.3454 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04293311 15.636729 4.092170e-55
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08278662  5.538416 3.052191e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08040919  3.794978 1.476564e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5530260          0.7750518          0.5796869          0.7483908
#> 2          0.2405691          0.6686947          0.2919786          0.6172852
#> 3          0.1036495          0.5194805          0.1535827          0.4695473
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5878458          0.7560995          0.5893314          0.7553340
#> 2          0.3246163          0.6608688          0.3427641          0.6538093
#> 3          0.1212629          0.4131233          0.1265429          0.4098999