Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1524831060
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0390   17.2242    0.0000 [ 0.6173; 0.7477 ] 
#>   eta3 ~ eta1      0.4585      0.1025    4.4753    0.0000 [ 0.2798; 0.6984 ] 
#>   eta3 ~ eta2      0.3052      0.0978    3.1193    0.0018 [ 0.1048; 0.4927 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0371   17.8614    0.0000 [ 0.5790; 0.7089 ] 
#>   eta1 =~ y12      0.6493      0.0354   18.3532    0.0000 [ 0.5897; 0.7008 ] 
#>   eta1 =~ y13      0.7613      0.0367   20.7704    0.0000 [ 0.7006; 0.8362 ] 
#>   eta2 =~ y21      0.5165      0.0403   12.8185    0.0000 [ 0.4522; 0.5949 ] 
#>   eta2 =~ y22      0.7554      0.0364   20.7263    0.0000 [ 0.7003; 0.8226 ] 
#>   eta2 =~ y23      0.7997      0.0323   24.7730    0.0000 [ 0.7418; 0.8653 ] 
#>   eta3 =~ y31      0.8223      0.0353   23.3261    0.0000 [ 0.7574; 0.8971 ] 
#>   eta3 =~ y32      0.6581      0.0470   14.0137    0.0000 [ 0.5571; 0.7305 ] 
#>   eta3 =~ y33      0.7474      0.0347   21.5686    0.0000 [ 0.6915; 0.8169 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0187   21.1268    0.0000 [ 0.3536; 0.4154 ] 
#>   eta1 <~ y12      0.3873      0.0147   26.3336    0.0000 [ 0.3621; 0.4141 ] 
#>   eta1 <~ y13      0.4542      0.0253   17.9569    0.0000 [ 0.4180; 0.5097 ] 
#>   eta2 <~ y21      0.3058      0.0180   17.0107    0.0000 [ 0.2687; 0.3377 ] 
#>   eta2 <~ y22      0.4473      0.0206   21.7526    0.0000 [ 0.4138; 0.4831 ] 
#>   eta2 <~ y23      0.4735      0.0208   22.7427    0.0000 [ 0.4319; 0.5054 ] 
#>   eta3 <~ y31      0.4400      0.0194   22.6602    0.0000 [ 0.4102; 0.4840 ] 
#>   eta3 <~ y32      0.3521      0.0219   16.0920    0.0000 [ 0.3040; 0.3848 ] 
#>   eta3 <~ y33      0.3999      0.0175   22.8838    0.0000 [ 0.3749; 0.4388 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0390   17.2242    0.0000 [ 0.6173; 0.7477 ] 
#>   eta3 ~ eta1       0.6634      0.0460   14.4202    0.0000 [ 0.6033; 0.7645 ] 
#>   eta3 ~ eta2       0.3052      0.0978    3.1193    0.0018 [ 0.1048; 0.4927 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0659    3.1078    0.0019 [ 0.0809; 0.3379 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03712297 17.86144  2.354660e-71
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03537683 18.35320  3.112509e-75
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03665535 20.77038  8.021390e-96
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04028992 12.81846  1.292408e-37
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03644583 20.72631  2.005941e-95
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03227961 24.77303 1.751297e-135
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03525139 23.32610 2.409928e-120
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04695903 14.01368  1.285618e-44
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03465338 21.56858 3.544191e-103
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5789955          0.7088988
#> 2          0.5897006          0.7008086
#> 3          0.7006360          0.8362422
#> 4          0.4521961          0.5949275
#> 5          0.7003063          0.8226023
#> 6          0.7418326          0.8653326
#> 7          0.7574095          0.8970521
#> 8          0.5571195          0.7305171
#> 9          0.6914775          0.8168704

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1524831060
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0390   17.2242    0.0000 [ 0.5586; 0.7602 ] 
#>   eta3 ~ eta1      0.4585      0.1025    4.4753    0.0000 [ 0.1660; 0.6958 ] 
#>   eta3 ~ eta2      0.3052      0.0978    3.1193    0.0018 [ 0.0693; 0.5752 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0371   17.8614    0.0000 [ 0.5787; 0.7706 ] 
#>   eta1 =~ y12      0.6493      0.0354   18.3532    0.0000 [ 0.5520; 0.7350 ] 
#>   eta1 =~ y13      0.7613      0.0367   20.7704    0.0000 [ 0.6604; 0.8500 ] 
#>   eta2 =~ y21      0.5165      0.0403   12.8185    0.0000 [ 0.4087; 0.6171 ] 
#>   eta2 =~ y22      0.7554      0.0364   20.7263    0.0000 [ 0.6618; 0.8502 ] 
#>   eta2 =~ y23      0.7997      0.0323   24.7730    0.0000 [ 0.7232; 0.8902 ] 
#>   eta3 =~ y31      0.8223      0.0353   23.3261    0.0000 [ 0.7304; 0.9127 ] 
#>   eta3 =~ y32      0.6581      0.0470   14.0137    0.0000 [ 0.5445; 0.7873 ] 
#>   eta3 =~ y33      0.7474      0.0347   21.5686    0.0000 [ 0.6506; 0.8298 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0187   21.1268    0.0000 [ 0.3545; 0.4513 ] 
#>   eta1 <~ y12      0.3873      0.0147   26.3336    0.0000 [ 0.3464; 0.4225 ] 
#>   eta1 <~ y13      0.4542      0.0253   17.9569    0.0000 [ 0.3852; 0.5160 ] 
#>   eta2 <~ y21      0.3058      0.0180   17.0107    0.0000 [ 0.2566; 0.3496 ] 
#>   eta2 <~ y22      0.4473      0.0206   21.7526    0.0000 [ 0.3931; 0.4994 ] 
#>   eta2 <~ y23      0.4735      0.0208   22.7427    0.0000 [ 0.4222; 0.5299 ] 
#>   eta3 <~ y31      0.4400      0.0194   22.6602    0.0000 [ 0.3897; 0.4901 ] 
#>   eta3 <~ y32      0.3521      0.0219   16.0920    0.0000 [ 0.3002; 0.4134 ] 
#>   eta3 <~ y33      0.3999      0.0175   22.8838    0.0000 [ 0.3512; 0.4416 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0390   17.2242    0.0000 [ 0.5586; 0.7602 ] 
#>   eta3 ~ eta1       0.6634      0.0460   14.4202    0.0000 [ 0.5257; 0.7636 ] 
#>   eta3 ~ eta2       0.3052      0.0978    3.1193    0.0018 [ 0.0693; 0.5752 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0659    3.1078    0.0019 [ 0.0433; 0.3841 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03897627 17.224156 1.749607e-66
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.10245338  4.475272 7.631395e-06
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09782562  3.119338 1.812581e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55864976          0.7602130          0.5828536          0.7360092
#> 2         0.16601700          0.6958480          0.2296393          0.6322257
#> 3         0.06925593          0.5751548          0.1300045          0.5144062
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.61129860          0.7687480          0.6172691          0.7476519
#> 2         0.24518092          0.6989731          0.2798427          0.6983900
#> 3         0.08757376          0.5089185          0.1047584          0.4927424