Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 2023138785
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0437   15.3705    0.0000 [ 0.6053; 0.7443 ] 
#>   eta3 ~ eta1      0.4585      0.0902    5.0814    0.0000 [ 0.3089; 0.6471 ] 
#>   eta3 ~ eta2      0.3052      0.0963    3.1672    0.0015 [ 0.1141; 0.5014 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0296   22.3803    0.0000 [ 0.5883; 0.7013 ] 
#>   eta1 =~ y12      0.6493      0.0378   17.1930    0.0000 [ 0.5979; 0.7318 ] 
#>   eta1 =~ y13      0.7613      0.0316   24.1154    0.0000 [ 0.6962; 0.8229 ] 
#>   eta2 =~ y21      0.5165      0.0547    9.4444    0.0000 [ 0.4143; 0.5918 ] 
#>   eta2 =~ y22      0.7554      0.0339   22.3004    0.0000 [ 0.6898; 0.8049 ] 
#>   eta2 =~ y23      0.7997      0.0361   22.1786    0.0000 [ 0.7192; 0.8404 ] 
#>   eta3 =~ y31      0.8223      0.0363   22.6424    0.0000 [ 0.7641; 0.8896 ] 
#>   eta3 =~ y32      0.6581      0.0370   17.8063    0.0000 [ 0.5785; 0.7136 ] 
#>   eta3 =~ y33      0.7474      0.0457   16.3433    0.0000 [ 0.6606; 0.8187 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0184   21.4469    0.0000 [ 0.3558; 0.4237 ] 
#>   eta1 <~ y12      0.3873      0.0151   25.7352    0.0000 [ 0.3711; 0.4168 ] 
#>   eta1 <~ y13      0.4542      0.0190   23.8575    0.0000 [ 0.4166; 0.4910 ] 
#>   eta2 <~ y21      0.3058      0.0284   10.7677    0.0000 [ 0.2516; 0.3522 ] 
#>   eta2 <~ y22      0.4473      0.0203   22.0159    0.0000 [ 0.4193; 0.4891 ] 
#>   eta2 <~ y23      0.4735      0.0198   23.9699    0.0000 [ 0.4353; 0.5026 ] 
#>   eta3 <~ y31      0.4400      0.0211   20.8242    0.0000 [ 0.4137; 0.4939 ] 
#>   eta3 <~ y32      0.3521      0.0178   19.8261    0.0000 [ 0.3201; 0.3769 ] 
#>   eta3 <~ y33      0.3999      0.0229   17.4950    0.0000 [ 0.3518; 0.4435 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0437   15.3705    0.0000 [ 0.6053; 0.7443 ] 
#>   eta3 ~ eta1       0.6634      0.0472   14.0446    0.0000 [ 0.5784; 0.7303 ] 
#>   eta3 ~ eta2       0.3052      0.0963    3.1672    0.0015 [ 0.1141; 0.5014 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0658    3.1114    0.0019 [ 0.0793; 0.3179 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.02962742 22.380278 6.126129e-111
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03776415 17.192972  2.997642e-66
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03157090 24.115428 1.722194e-128
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05468352  9.444432  3.573379e-21
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03387323 22.300429 3.659718e-110
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03605570 22.178567 5.531218e-109
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03631582 22.642400 1.657688e-113
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03695708 17.806303  6.314561e-71
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04573275 16.343301  4.855135e-60
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5882981          0.7013010
#> 2          0.5978583          0.7317884
#> 3          0.6962122          0.8228833
#> 4          0.4143304          0.5917847
#> 5          0.6898184          0.8049089
#> 6          0.7191862          0.8404068
#> 7          0.7641314          0.8895641
#> 8          0.5785462          0.7135887
#> 9          0.6605862          0.8187259

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 2023138785
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0437   15.3705    0.0000 [ 0.5541; 0.7799 ] 
#>   eta3 ~ eta1      0.4585      0.0902    5.0814    0.0000 [ 0.2353; 0.7019 ] 
#>   eta3 ~ eta2      0.3052      0.0963    3.1672    0.0015 [ 0.0440; 0.5423 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0296   22.3803    0.0000 [ 0.5937; 0.7469 ] 
#>   eta1 =~ y12      0.6493      0.0378   17.1930    0.0000 [ 0.5412; 0.7365 ] 
#>   eta1 =~ y13      0.7613      0.0316   24.1154    0.0000 [ 0.6780; 0.8413 ] 
#>   eta2 =~ y21      0.5165      0.0547    9.4444    0.0000 [ 0.3763; 0.6590 ] 
#>   eta2 =~ y22      0.7554      0.0339   22.3004    0.0000 [ 0.6711; 0.8463 ] 
#>   eta2 =~ y23      0.7997      0.0361   22.1786    0.0000 [ 0.7148; 0.9012 ] 
#>   eta3 =~ y31      0.8223      0.0363   22.6424    0.0000 [ 0.7174; 0.9052 ] 
#>   eta3 =~ y32      0.6581      0.0370   17.8063    0.0000 [ 0.5680; 0.7591 ] 
#>   eta3 =~ y33      0.7474      0.0457   16.3433    0.0000 [ 0.6348; 0.8713 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0184   21.4469    0.0000 [ 0.3535; 0.4489 ] 
#>   eta1 <~ y12      0.3873      0.0151   25.7352    0.0000 [ 0.3439; 0.4218 ] 
#>   eta1 <~ y13      0.4542      0.0190   23.8575    0.0000 [ 0.4056; 0.5040 ] 
#>   eta2 <~ y21      0.3058      0.0284   10.7677    0.0000 [ 0.2308; 0.3777 ] 
#>   eta2 <~ y22      0.4473      0.0203   22.0159    0.0000 [ 0.3928; 0.4979 ] 
#>   eta2 <~ y23      0.4735      0.0198   23.9699    0.0000 [ 0.4233; 0.5254 ] 
#>   eta3 <~ y31      0.4400      0.0211   20.8242    0.0000 [ 0.3797; 0.4890 ] 
#>   eta3 <~ y32      0.3521      0.0178   19.8261    0.0000 [ 0.3095; 0.4013 ] 
#>   eta3 <~ y33      0.3999      0.0229   17.4950    0.0000 [ 0.3442; 0.4624 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0437   15.3705    0.0000 [ 0.5541; 0.7799 ] 
#>   eta3 ~ eta1       0.6634      0.0472   14.0446    0.0000 [ 0.5424; 0.7867 ] 
#>   eta3 ~ eta2       0.3052      0.0963    3.1672    0.0015 [ 0.0440; 0.5423 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0658    3.1114    0.0019 [ 0.0257; 0.3662 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04367687 15.370458 2.583365e-53
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09023152  5.081448 3.745682e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09634755  3.167191 1.539189e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55407457          0.7799467          0.5811974          0.7528238
#> 2         0.23530758          0.7019340          0.2913403          0.6459013
#> 3         0.04399997          0.5422551          0.1038307          0.4824244
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.56465871          0.7699346          0.6053482          0.7442806
#> 2         0.30132399          0.6790376          0.3089117          0.6470777
#> 3         0.06492103          0.5046158          0.1140561          0.5013776