Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -633853889
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0521   12.8852    0.0000 [ 0.5621; 0.7614 ] 
#>   eta3 ~ eta1      0.4585      0.0867    5.2882    0.0000 [ 0.2982; 0.6069 ] 
#>   eta3 ~ eta2      0.3052      0.0886    3.4445    0.0006 [ 0.1576; 0.4593 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0450   14.7205    0.0000 [ 0.5902; 0.7426 ] 
#>   eta1 =~ y12      0.6493      0.0339   19.1756    0.0000 [ 0.5987; 0.7141 ] 
#>   eta1 =~ y13      0.7613      0.0382   19.9225    0.0000 [ 0.7072; 0.8263 ] 
#>   eta2 =~ y21      0.5165      0.0535    9.6559    0.0000 [ 0.4109; 0.6032 ] 
#>   eta2 =~ y22      0.7554      0.0421   17.9275    0.0000 [ 0.6914; 0.8355 ] 
#>   eta2 =~ y23      0.7997      0.0429   18.6414    0.0000 [ 0.7114; 0.8563 ] 
#>   eta3 =~ y31      0.8223      0.0298   27.5592    0.0000 [ 0.7682; 0.8764 ] 
#>   eta3 =~ y32      0.6581      0.0376   17.5088    0.0000 [ 0.5939; 0.7288 ] 
#>   eta3 =~ y33      0.7474      0.0379   19.7266    0.0000 [ 0.6492; 0.8096 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0212   18.6822    0.0000 [ 0.3542; 0.4250 ] 
#>   eta1 <~ y12      0.3873      0.0190   20.3469    0.0000 [ 0.3517; 0.4185 ] 
#>   eta1 <~ y13      0.4542      0.0239   19.0020    0.0000 [ 0.4026; 0.4943 ] 
#>   eta2 <~ y21      0.3058      0.0319    9.5986    0.0000 [ 0.2486; 0.3678 ] 
#>   eta2 <~ y22      0.4473      0.0232   19.3142    0.0000 [ 0.4113; 0.4810 ] 
#>   eta2 <~ y23      0.4735      0.0205   23.1515    0.0000 [ 0.4363; 0.5160 ] 
#>   eta3 <~ y31      0.4400      0.0179   24.5434    0.0000 [ 0.4039; 0.4748 ] 
#>   eta3 <~ y32      0.3521      0.0182   19.3826    0.0000 [ 0.3155; 0.3852 ] 
#>   eta3 <~ y33      0.3999      0.0147   27.1246    0.0000 [ 0.3723; 0.4257 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0521   12.8852    0.0000 [ 0.5621; 0.7614 ] 
#>   eta3 ~ eta1       0.6634      0.0382   17.3628    0.0000 [ 0.5725; 0.7171 ] 
#>   eta3 ~ eta2       0.3052      0.0886    3.4445    0.0006 [ 0.1576; 0.4593 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0621    3.3015    0.0010 [ 0.1076; 0.3209 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04504405 14.720475  4.763017e-49
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03385966 19.175559  5.922821e-82
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03821534 19.922518  2.596020e-88
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05348580  9.655924  4.639590e-22
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04213579 17.927460  7.199393e-72
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04289713 18.641427  1.482341e-77
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02983677 27.559193 3.433977e-167
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03758511 17.508765  1.228301e-68
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03788910 19.726625  1.273974e-86
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5901851          0.7426481
#> 2          0.5987053          0.7141442
#> 3          0.7072434          0.8263119
#> 4          0.4109412          0.6032085
#> 5          0.6913787          0.8355147
#> 6          0.7114114          0.8563354
#> 7          0.7682053          0.8763506
#> 8          0.5939200          0.7287717
#> 9          0.6491643          0.8095997

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -633853889
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0521   12.8852    0.0000 [ 0.5355; 0.8049 ] 
#>   eta3 ~ eta1      0.4585      0.0867    5.2882    0.0000 [ 0.2557; 0.7040 ] 
#>   eta3 ~ eta2      0.3052      0.0886    3.4445    0.0006 [ 0.0601; 0.5182 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0450   14.7205    0.0000 [ 0.5462; 0.7791 ] 
#>   eta1 =~ y12      0.6493      0.0339   19.1756    0.0000 [ 0.5532; 0.7283 ] 
#>   eta1 =~ y13      0.7613      0.0382   19.9225    0.0000 [ 0.6687; 0.8664 ] 
#>   eta2 =~ y21      0.5165      0.0535    9.6559    0.0000 [ 0.3794; 0.6560 ] 
#>   eta2 =~ y22      0.7554      0.0421   17.9275    0.0000 [ 0.6405; 0.8584 ] 
#>   eta2 =~ y23      0.7997      0.0429   18.6414    0.0000 [ 0.6922; 0.9141 ] 
#>   eta3 =~ y31      0.8223      0.0298   27.5592    0.0000 [ 0.7398; 0.8941 ] 
#>   eta3 =~ y32      0.6581      0.0376   17.5088    0.0000 [ 0.5717; 0.7661 ] 
#>   eta3 =~ y33      0.7474      0.0379   19.7266    0.0000 [ 0.6540; 0.8499 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0212   18.6822    0.0000 [ 0.3417; 0.4512 ] 
#>   eta1 <~ y12      0.3873      0.0190   20.3469    0.0000 [ 0.3339; 0.4323 ] 
#>   eta1 <~ y13      0.4542      0.0239   19.0020    0.0000 [ 0.3969; 0.5205 ] 
#>   eta2 <~ y21      0.3058      0.0319    9.5986    0.0000 [ 0.2247; 0.3895 ] 
#>   eta2 <~ y22      0.4473      0.0232   19.3142    0.0000 [ 0.3849; 0.5046 ] 
#>   eta2 <~ y23      0.4735      0.0205   23.1515    0.0000 [ 0.4239; 0.5296 ] 
#>   eta3 <~ y31      0.4400      0.0179   24.5434    0.0000 [ 0.3879; 0.4806 ] 
#>   eta3 <~ y32      0.3521      0.0182   19.3826    0.0000 [ 0.3089; 0.4028 ] 
#>   eta3 <~ y33      0.3999      0.0147   27.1246    0.0000 [ 0.3620; 0.4382 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0521   12.8852    0.0000 [ 0.5355; 0.8049 ] 
#>   eta3 ~ eta1       0.6634      0.0382   17.3628    0.0000 [ 0.5749; 0.7725 ] 
#>   eta3 ~ eta2       0.3052      0.0886    3.4445    0.0006 [ 0.0601; 0.5182 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0621    3.3015    0.0010 [ 0.0334; 0.3543 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05210118 12.885186 5.454350e-38
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08670424  5.288170 1.235462e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08859050  3.444513 5.720894e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5354556          0.8048935          0.5678098          0.7725393
#> 2          0.2556591          0.7040445          0.3095015          0.6502022
#> 3          0.0600993          0.5182394          0.1151130          0.4632257
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5483363          0.7657407          0.5620528          0.7614467
#> 2          0.2970954          0.6526725          0.2981548          0.6068767
#> 3          0.1137827          0.4815994          0.1575956          0.4592940