Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2011863314
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0468   14.3311    0.0000 [ 0.5793; 0.7406 ] 
#>   eta3 ~ eta1      0.4585      0.0687    6.6707    0.0000 [ 0.3087; 0.5543 ] 
#>   eta3 ~ eta2      0.3052      0.0818    3.7292    0.0002 [ 0.1752; 0.4739 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0405   16.3526    0.0000 [ 0.5988; 0.7393 ] 
#>   eta1 =~ y12      0.6493      0.0418   15.5338    0.0000 [ 0.5802; 0.7160 ] 
#>   eta1 =~ y13      0.7613      0.0375   20.3160    0.0000 [ 0.6973; 0.8293 ] 
#>   eta2 =~ y21      0.5165      0.0640    8.0669    0.0000 [ 0.4099; 0.6529 ] 
#>   eta2 =~ y22      0.7554      0.0321   23.5408    0.0000 [ 0.7096; 0.8024 ] 
#>   eta2 =~ y23      0.7997      0.0332   24.0944    0.0000 [ 0.7375; 0.8574 ] 
#>   eta3 =~ y31      0.8223      0.0266   30.9117    0.0000 [ 0.7718; 0.8714 ] 
#>   eta3 =~ y32      0.6581      0.0353   18.6358    0.0000 [ 0.5862; 0.7198 ] 
#>   eta3 =~ y33      0.7474      0.0372   20.0795    0.0000 [ 0.6764; 0.8170 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0230   17.1977    0.0000 [ 0.3554; 0.4312 ] 
#>   eta1 <~ y12      0.3873      0.0222   17.4471    0.0000 [ 0.3430; 0.4188 ] 
#>   eta1 <~ y13      0.4542      0.0203   22.4197    0.0000 [ 0.4133; 0.4869 ] 
#>   eta2 <~ y21      0.3058      0.0289   10.5902    0.0000 [ 0.2504; 0.3571 ] 
#>   eta2 <~ y22      0.4473      0.0212   21.1037    0.0000 [ 0.4084; 0.4787 ] 
#>   eta2 <~ y23      0.4735      0.0254   18.6624    0.0000 [ 0.4241; 0.5278 ] 
#>   eta3 <~ y31      0.4400      0.0168   26.1432    0.0000 [ 0.4139; 0.4767 ] 
#>   eta3 <~ y32      0.3521      0.0170   20.6716    0.0000 [ 0.3193; 0.3796 ] 
#>   eta3 <~ y33      0.3999      0.0170   23.5315    0.0000 [ 0.3706; 0.4247 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0468   14.3311    0.0000 [ 0.5793; 0.7406 ] 
#>   eta3 ~ eta1       0.6634      0.0276   24.0366    0.0000 [ 0.6135; 0.7101 ] 
#>   eta3 ~ eta2       0.3052      0.0818    3.7292    0.0002 [ 0.1752; 0.4739 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0592    3.4632    0.0005 [ 0.1207; 0.3402 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04054836 16.352569  4.170245e-60
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04179782 15.533773  2.049568e-54
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03747519 20.315996  9.284771e-92
#> 4 eta2 =~ y21  Common factor 0.5164548 0.06402168  8.066874  7.212083e-16
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03208850 23.540763 1.560968e-122
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03318874 24.094427 2.859639e-128
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02660085 30.911694 8.318547e-210
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03531208 18.635801  1.646712e-77
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03722327 20.079488  1.115387e-89
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5988231          0.7393085
#> 2          0.5801814          0.7159922
#> 3          0.6972956          0.8293009
#> 4          0.4098993          0.6529449
#> 5          0.7095674          0.8023513
#> 6          0.7374892          0.8573654
#> 7          0.7717687          0.8714411
#> 8          0.5861847          0.7198471
#> 9          0.6764199          0.8170000

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2011863314
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0468   14.3311    0.0000 [ 0.5564; 0.7987 ] 
#>   eta3 ~ eta1      0.4585      0.0687    6.6707    0.0000 [ 0.2801; 0.6355 ] 
#>   eta3 ~ eta2      0.3052      0.0818    3.7292    0.0002 [ 0.0888; 0.5120 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0405   16.3526    0.0000 [ 0.5609; 0.7706 ] 
#>   eta1 =~ y12      0.6493      0.0418   15.5338    0.0000 [ 0.5413; 0.7575 ] 
#>   eta1 =~ y13      0.7613      0.0375   20.3160    0.0000 [ 0.6601; 0.8539 ] 
#>   eta2 =~ y21      0.5165      0.0640    8.0669    0.0000 [ 0.3468; 0.6779 ] 
#>   eta2 =~ y22      0.7554      0.0321   23.5408    0.0000 [ 0.6667; 0.8326 ] 
#>   eta2 =~ y23      0.7997      0.0332   24.0944    0.0000 [ 0.7175; 0.8891 ] 
#>   eta3 =~ y31      0.8223      0.0266   30.9117    0.0000 [ 0.7544; 0.8919 ] 
#>   eta3 =~ y32      0.6581      0.0353   18.6358    0.0000 [ 0.5678; 0.7504 ] 
#>   eta3 =~ y33      0.7474      0.0372   20.0795    0.0000 [ 0.6525; 0.8450 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0230   17.1977    0.0000 [ 0.3384; 0.4573 ] 
#>   eta1 <~ y12      0.3873      0.0222   17.4471    0.0000 [ 0.3308; 0.4456 ] 
#>   eta1 <~ y13      0.4542      0.0203   22.4197    0.0000 [ 0.4001; 0.5048 ] 
#>   eta2 <~ y21      0.3058      0.0289   10.5902    0.0000 [ 0.2311; 0.3805 ] 
#>   eta2 <~ y22      0.4473      0.0212   21.1037    0.0000 [ 0.3913; 0.5009 ] 
#>   eta2 <~ y23      0.4735      0.0254   18.6624    0.0000 [ 0.4121; 0.5433 ] 
#>   eta3 <~ y31      0.4400      0.0168   26.1432    0.0000 [ 0.3960; 0.4830 ] 
#>   eta3 <~ y32      0.3521      0.0170   20.6716    0.0000 [ 0.3081; 0.3961 ] 
#>   eta3 <~ y33      0.3999      0.0170   23.5315    0.0000 [ 0.3560; 0.4439 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0468   14.3311    0.0000 [ 0.5564; 0.7987 ] 
#>   eta3 ~ eta1       0.6634      0.0276   24.0366    0.0000 [ 0.5897; 0.7325 ] 
#>   eta3 ~ eta2       0.3052      0.0818    3.7292    0.0002 [ 0.0888; 0.5120 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0592    3.4632    0.0005 [ 0.0503; 0.3562 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04684467 14.331052 1.399843e-46
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.06873413  6.670729 2.545354e-11
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08182767  3.729192 1.920945e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55643240          0.7986866          0.5855224          0.7695966
#> 2         0.28008009          0.6355342          0.3227632          0.5928511
#> 3         0.08884484          0.5120114          0.1396589          0.4611973
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5594097          0.7701376          0.5792738          0.7406355
#> 2          0.2837060          0.5692092          0.3087240          0.5543033
#> 3          0.1700453          0.4757020          0.1752337          0.4738512