Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -111460701
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0430   15.6029    0.0000 [ 0.5789; 0.7381 ] 
#>   eta3 ~ eta1      0.4585      0.0714    6.4195    0.0000 [ 0.3106; 0.5560 ] 
#>   eta3 ~ eta2      0.3052      0.0731    4.1756    0.0000 [ 0.2169; 0.4436 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0405   16.3791    0.0000 [ 0.5839; 0.7329 ] 
#>   eta1 =~ y12      0.6493      0.0443   14.6665    0.0000 [ 0.5727; 0.7345 ] 
#>   eta1 =~ y13      0.7613      0.0293   25.9854    0.0000 [ 0.7109; 0.8142 ] 
#>   eta2 =~ y21      0.5165      0.0475   10.8720    0.0000 [ 0.4082; 0.5833 ] 
#>   eta2 =~ y22      0.7554      0.0356   21.2138    0.0000 [ 0.6727; 0.8033 ] 
#>   eta2 =~ y23      0.7997      0.0374   21.3790    0.0000 [ 0.7234; 0.8550 ] 
#>   eta3 =~ y31      0.8223      0.0414   19.8547    0.0000 [ 0.7391; 0.8859 ] 
#>   eta3 =~ y32      0.6581      0.0447   14.7252    0.0000 [ 0.5969; 0.7421 ] 
#>   eta3 =~ y33      0.7474      0.0392   19.0470    0.0000 [ 0.6907; 0.8221 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0207   19.0822    0.0000 [ 0.3565; 0.4309 ] 
#>   eta1 <~ y12      0.3873      0.0179   21.6012    0.0000 [ 0.3579; 0.4222 ] 
#>   eta1 <~ y13      0.4542      0.0237   19.1939    0.0000 [ 0.4146; 0.5054 ] 
#>   eta2 <~ y21      0.3058      0.0257   11.8804    0.0000 [ 0.2475; 0.3399 ] 
#>   eta2 <~ y22      0.4473      0.0202   22.1142    0.0000 [ 0.4185; 0.4793 ] 
#>   eta2 <~ y23      0.4735      0.0183   25.9159    0.0000 [ 0.4449; 0.5070 ] 
#>   eta3 <~ y31      0.4400      0.0202   21.8210    0.0000 [ 0.4081; 0.4748 ] 
#>   eta3 <~ y32      0.3521      0.0232   15.1953    0.0000 [ 0.3197; 0.3995 ] 
#>   eta3 <~ y33      0.3999      0.0211   18.9840    0.0000 [ 0.3688; 0.4442 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0430   15.6029    0.0000 [ 0.5789; 0.7381 ] 
#>   eta3 ~ eta1       0.6634      0.0391   16.9614    0.0000 [ 0.5939; 0.7300 ] 
#>   eta3 ~ eta2       0.3052      0.0731    4.1756    0.0000 [ 0.2169; 0.4436 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0497    4.1192    0.0000 [ 0.1437; 0.2943 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04048277 16.37906  2.698675e-60
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04426947 14.66649  1.056653e-48
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02929897 25.98541 7.240085e-149
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04750326 10.87199  1.567490e-27
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03560833 21.21379 7.123888e-100
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03740419 21.37899 2.096224e-101
#> 7 eta3 =~ y31  Common factor 0.8222773 0.04141475 19.85470  1.003641e-87
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04468994 14.72521  4.440737e-49
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03924113 19.04696  6.963535e-81
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5838765          0.7328799
#> 2          0.5727123          0.7344709
#> 3          0.7108861          0.8141584
#> 4          0.4082198          0.5833230
#> 5          0.6727480          0.8033366
#> 6          0.7233874          0.8549627
#> 7          0.7390909          0.8858824
#> 8          0.5968749          0.7421172
#> 9          0.6907078          0.8221487

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -111460701
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0430   15.6029    0.0000 [ 0.5627; 0.7852 ] 
#>   eta3 ~ eta1      0.4585      0.0714    6.4195    0.0000 [ 0.2835; 0.6529 ] 
#>   eta3 ~ eta2      0.3052      0.0731    4.1756    0.0000 [ 0.1092; 0.4871 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0405   16.3791    0.0000 [ 0.5576; 0.7670 ] 
#>   eta1 =~ y12      0.6493      0.0443   14.6665    0.0000 [ 0.5286; 0.7575 ] 
#>   eta1 =~ y13      0.7613      0.0293   25.9854    0.0000 [ 0.6850; 0.8365 ] 
#>   eta2 =~ y21      0.5165      0.0475   10.8720    0.0000 [ 0.4039; 0.6496 ] 
#>   eta2 =~ y22      0.7554      0.0356   21.2138    0.0000 [ 0.6643; 0.8484 ] 
#>   eta2 =~ y23      0.7997      0.0374   21.3790    0.0000 [ 0.7077; 0.9012 ] 
#>   eta3 =~ y31      0.8223      0.0414   19.8547    0.0000 [ 0.7177; 0.9318 ] 
#>   eta3 =~ y32      0.6581      0.0447   14.7252    0.0000 [ 0.5441; 0.7753 ] 
#>   eta3 =~ y33      0.7474      0.0392   19.0470    0.0000 [ 0.6492; 0.8522 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0207   19.0822    0.0000 [ 0.3439; 0.4511 ] 
#>   eta1 <~ y12      0.3873      0.0179   21.6012    0.0000 [ 0.3399; 0.4326 ] 
#>   eta1 <~ y13      0.4542      0.0237   19.1939    0.0000 [ 0.3949; 0.5173 ] 
#>   eta2 <~ y21      0.3058      0.0257   11.8804    0.0000 [ 0.2426; 0.3757 ] 
#>   eta2 <~ y22      0.4473      0.0202   22.1142    0.0000 [ 0.3910; 0.4956 ] 
#>   eta2 <~ y23      0.4735      0.0183   25.9159    0.0000 [ 0.4244; 0.5189 ] 
#>   eta3 <~ y31      0.4400      0.0202   21.8210    0.0000 [ 0.3873; 0.4915 ] 
#>   eta3 <~ y32      0.3521      0.0232   15.1953    0.0000 [ 0.2915; 0.4114 ] 
#>   eta3 <~ y33      0.3999      0.0211   18.9840    0.0000 [ 0.3454; 0.4543 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0430   15.6029    0.0000 [ 0.5627; 0.7852 ] 
#>   eta3 ~ eta1       0.6634      0.0391   16.9614    0.0000 [ 0.5684; 0.7707 ] 
#>   eta3 ~ eta2       0.3052      0.0731    4.1756    0.0000 [ 0.1092; 0.4871 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0497    4.1192    0.0000 [ 0.0727; 0.3299 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04302620 15.602899 6.956239e-55
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07142396  6.419509 1.367144e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07307912  4.175626 2.971671e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5626995          0.7852068          0.5894183          0.7584880
#> 2          0.2835431          0.6529075          0.3278966          0.6085541
#> 3          0.1091933          0.4871173          0.1545746          0.4417360
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5640947          0.7462125          0.5788901          0.7381281
#> 2          0.3080583          0.6351228          0.3105694          0.5559898
#> 3          0.1354653          0.4440392          0.2168588          0.4435541