Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1167420725
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0404   16.6178    0.0000 [ 0.5847; 0.7357 ] 
#>   eta3 ~ eta1      0.4585      0.0747    6.1371    0.0000 [ 0.3170; 0.5604 ] 
#>   eta3 ~ eta2      0.3052      0.0761    4.0120    0.0001 [ 0.2049; 0.4588 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0348   19.0673    0.0000 [ 0.6052; 0.7290 ] 
#>   eta1 =~ y12      0.6493      0.0388   16.7267    0.0000 [ 0.5555; 0.6911 ] 
#>   eta1 =~ y13      0.7613      0.0320   23.8084    0.0000 [ 0.7075; 0.8068 ] 
#>   eta2 =~ y21      0.5165      0.0648    7.9691    0.0000 [ 0.3966; 0.6157 ] 
#>   eta2 =~ y22      0.7554      0.0382   19.7938    0.0000 [ 0.6916; 0.8329 ] 
#>   eta2 =~ y23      0.7997      0.0390   20.4859    0.0000 [ 0.7273; 0.8788 ] 
#>   eta3 =~ y31      0.8223      0.0305   26.9753    0.0000 [ 0.7751; 0.8653 ] 
#>   eta3 =~ y32      0.6581      0.0313   21.0046    0.0000 [ 0.6107; 0.7120 ] 
#>   eta3 =~ y33      0.7474      0.0379   19.7436    0.0000 [ 0.6908; 0.8217 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0226   17.5095    0.0000 [ 0.3523; 0.4360 ] 
#>   eta1 <~ y12      0.3873      0.0178   21.7795    0.0000 [ 0.3432; 0.4074 ] 
#>   eta1 <~ y13      0.4542      0.0180   25.2427    0.0000 [ 0.4224; 0.4827 ] 
#>   eta2 <~ y21      0.3058      0.0316    9.6823    0.0000 [ 0.2397; 0.3560 ] 
#>   eta2 <~ y22      0.4473      0.0273   16.4011    0.0000 [ 0.4028; 0.5078 ] 
#>   eta2 <~ y23      0.4735      0.0198   23.9464    0.0000 [ 0.4364; 0.5134 ] 
#>   eta3 <~ y31      0.4400      0.0186   23.6775    0.0000 [ 0.4045; 0.4651 ] 
#>   eta3 <~ y32      0.3521      0.0154   22.9121    0.0000 [ 0.3260; 0.3779 ] 
#>   eta3 <~ y33      0.3999      0.0169   23.6980    0.0000 [ 0.3712; 0.4295 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0404   16.6178    0.0000 [ 0.5847; 0.7357 ] 
#>   eta3 ~ eta1       0.6634      0.0394   16.8339    0.0000 [ 0.5790; 0.7254 ] 
#>   eta3 ~ eta2       0.3052      0.0761    4.0120    0.0001 [ 0.2049; 0.4588 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0527    3.8880    0.0001 [ 0.1367; 0.3046 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03477517 19.067338  4.717120e-81
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03881675 16.726746  8.368761e-63
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03197803 23.808403 2.733353e-125
#> 4 eta2 =~ y21  Common factor 0.5164548 0.06480691  7.969132  1.597924e-15
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03816280 19.793824  3.365150e-87
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03903483 20.485904  2.875946e-93
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03048266 26.975252 2.884690e-160
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03132974 21.004608  5.952215e-98
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03785650 19.743612  9.103174e-87
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6052063          0.7290113
#> 2          0.5555054          0.6910810
#> 3          0.7075070          0.8068496
#> 4          0.3966368          0.6156853
#> 5          0.6916338          0.8329275
#> 6          0.7272525          0.8787952
#> 7          0.7751012          0.8653411
#> 8          0.6106898          0.7120302
#> 9          0.6908118          0.8217075

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1167420725
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0404   16.6178    0.0000 [ 0.5676; 0.7765 ] 
#>   eta3 ~ eta1      0.4585      0.0747    6.1371    0.0000 [ 0.2629; 0.6493 ] 
#>   eta3 ~ eta2      0.3052      0.0761    4.0120    0.0001 [ 0.1145; 0.5078 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0348   19.0673    0.0000 [ 0.5754; 0.7553 ] 
#>   eta1 =~ y12      0.6493      0.0388   16.7267    0.0000 [ 0.5566; 0.7574 ] 
#>   eta1 =~ y13      0.7613      0.0320   23.8084    0.0000 [ 0.6805; 0.8458 ] 
#>   eta2 =~ y21      0.5165      0.0648    7.9691    0.0000 [ 0.3626; 0.6978 ] 
#>   eta2 =~ y22      0.7554      0.0382   19.7938    0.0000 [ 0.6528; 0.8502 ] 
#>   eta2 =~ y23      0.7997      0.0390   20.4859    0.0000 [ 0.6897; 0.8915 ] 
#>   eta3 =~ y31      0.8223      0.0305   26.9753    0.0000 [ 0.7461; 0.9037 ] 
#>   eta3 =~ y32      0.6581      0.0313   21.0046    0.0000 [ 0.5735; 0.7356 ] 
#>   eta3 =~ y33      0.7474      0.0379   19.7436    0.0000 [ 0.6489; 0.8446 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0226   17.5095    0.0000 [ 0.3352; 0.4521 ] 
#>   eta1 <~ y12      0.3873      0.0178   21.7795    0.0000 [ 0.3432; 0.4351 ] 
#>   eta1 <~ y13      0.4542      0.0180   25.2427    0.0000 [ 0.4051; 0.4982 ] 
#>   eta2 <~ y21      0.3058      0.0316    9.6823    0.0000 [ 0.2337; 0.3971 ] 
#>   eta2 <~ y22      0.4473      0.0273   16.4011    0.0000 [ 0.3752; 0.5163 ] 
#>   eta2 <~ y23      0.4735      0.0198   23.9464    0.0000 [ 0.4182; 0.5205 ] 
#>   eta3 <~ y31      0.4400      0.0186   23.6775    0.0000 [ 0.3938; 0.4899 ] 
#>   eta3 <~ y32      0.3521      0.0154   22.9121    0.0000 [ 0.3110; 0.3905 ] 
#>   eta3 <~ y33      0.3999      0.0169   23.6980    0.0000 [ 0.3566; 0.4438 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0404   16.6178    0.0000 [ 0.5676; 0.7765 ] 
#>   eta3 ~ eta1       0.6634      0.0394   16.8339    0.0000 [ 0.5635; 0.7673 ] 
#>   eta3 ~ eta2       0.3052      0.0761    4.0120    0.0001 [ 0.1145; 0.5078 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0527    3.8880    0.0001 [ 0.0730; 0.3455 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04039838 16.617832 5.177156e-62
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07471068  6.137098 8.404233e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07605869  4.012048 6.019425e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5675853          0.7765028          0.5926722          0.7514159
#> 2          0.2629015          0.6492629          0.3092959          0.6028685
#> 3          0.1144705          0.5078031          0.1617020          0.4605715
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5730834          0.7547657          0.5846812          0.7356618
#> 2          0.2655961          0.5755875          0.3169675          0.5604405
#> 3          0.1921174          0.4840890          0.2048972          0.4588472