Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -1935354347
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0463   14.4956    0.0000 [ 0.5652; 0.7349 ] 
#>   eta3 ~ eta1      0.4585      0.0830    5.5247    0.0000 [ 0.2986; 0.6109 ] 
#>   eta3 ~ eta2      0.3052      0.0878    3.4768    0.0005 [ 0.1501; 0.4729 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0379   17.4928    0.0000 [ 0.6102; 0.7403 ] 
#>   eta1 =~ y12      0.6493      0.0441   14.7111    0.0000 [ 0.5633; 0.7294 ] 
#>   eta1 =~ y13      0.7613      0.0301   25.2901    0.0000 [ 0.7223; 0.8378 ] 
#>   eta2 =~ y21      0.5165      0.0674    7.6623    0.0000 [ 0.3927; 0.6222 ] 
#>   eta2 =~ y22      0.7554      0.0419   18.0129    0.0000 [ 0.6865; 0.8354 ] 
#>   eta2 =~ y23      0.7997      0.0399   20.0187    0.0000 [ 0.7220; 0.8589 ] 
#>   eta3 =~ y31      0.8223      0.0381   21.5956    0.0000 [ 0.7561; 0.8945 ] 
#>   eta3 =~ y32      0.6581      0.0396   16.6226    0.0000 [ 0.5775; 0.7252 ] 
#>   eta3 =~ y33      0.7474      0.0415   17.9987    0.0000 [ 0.6723; 0.8213 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0242   16.3291    0.0000 [ 0.3582; 0.4440 ] 
#>   eta1 <~ y12      0.3873      0.0189   20.5466    0.0000 [ 0.3474; 0.4195 ] 
#>   eta1 <~ y13      0.4542      0.0164   27.7032    0.0000 [ 0.4289; 0.4796 ] 
#>   eta2 <~ y21      0.3058      0.0321    9.5124    0.0000 [ 0.2440; 0.3545 ] 
#>   eta2 <~ y22      0.4473      0.0265   16.8824    0.0000 [ 0.3984; 0.4872 ] 
#>   eta2 <~ y23      0.4735      0.0265   17.8599    0.0000 [ 0.4287; 0.5245 ] 
#>   eta3 <~ y31      0.4400      0.0188   23.4093    0.0000 [ 0.4119; 0.4827 ] 
#>   eta3 <~ y32      0.3521      0.0186   18.9282    0.0000 [ 0.3195; 0.3831 ] 
#>   eta3 <~ y33      0.3999      0.0217   18.4139    0.0000 [ 0.3584; 0.4398 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0463   14.4956    0.0000 [ 0.5652; 0.7349 ] 
#>   eta3 ~ eta1       0.6634      0.0354   18.7348    0.0000 [ 0.5834; 0.7163 ] 
#>   eta3 ~ eta2       0.3052      0.0878    3.4768    0.0005 [ 0.1501; 0.4729 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0589    3.4807    0.0005 [ 0.1023; 0.3152 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03790524 17.492829  1.624878e-68
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04413534 14.711066  5.473861e-49
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03010451 25.290095 4.105348e-141
#> 4 eta2 =~ y21  Common factor 0.5164548 0.06740182  7.662327  1.825934e-14
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04193600 18.012871  1.544091e-72
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03994578 20.018728  3.782537e-89
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03807613 21.595611 1.975136e-103
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03958885 16.622583  4.782714e-62
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04152658 17.998690  1.994808e-72
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6101861          0.7402828
#> 2          0.5633407          0.7293651
#> 3          0.7222926          0.8377970
#> 4          0.3927010          0.6222157
#> 5          0.6864906          0.8353722
#> 6          0.7219890          0.8588780
#> 7          0.7561309          0.8944601
#> 8          0.5775031          0.7252110
#> 9          0.6722526          0.8212629

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -1935354347
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0463   14.4956    0.0000 [ 0.5557; 0.7952 ] 
#>   eta3 ~ eta1      0.4585      0.0830    5.5247    0.0000 [ 0.2594; 0.6886 ] 
#>   eta3 ~ eta2      0.3052      0.0878    3.4768    0.0005 [ 0.0666; 0.5205 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0379   17.4928    0.0000 [ 0.5601; 0.7562 ] 
#>   eta1 =~ y12      0.6493      0.0441   14.7111    0.0000 [ 0.5306; 0.7589 ] 
#>   eta1 =~ y13      0.7613      0.0301   25.2901    0.0000 [ 0.6776; 0.8333 ] 
#>   eta2 =~ y21      0.5165      0.0674    7.6623    0.0000 [ 0.3429; 0.6915 ] 
#>   eta2 =~ y22      0.7554      0.0419   18.0129    0.0000 [ 0.6467; 0.8636 ] 
#>   eta2 =~ y23      0.7997      0.0399   20.0187    0.0000 [ 0.6959; 0.9025 ] 
#>   eta3 =~ y31      0.8223      0.0381   21.5956    0.0000 [ 0.7204; 0.9173 ] 
#>   eta3 =~ y32      0.6581      0.0396   16.6226    0.0000 [ 0.5613; 0.7660 ] 
#>   eta3 =~ y33      0.7474      0.0415   17.9987    0.0000 [ 0.6402; 0.8550 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0242   16.3291    0.0000 [ 0.3343; 0.4596 ] 
#>   eta1 <~ y12      0.3873      0.0189   20.5466    0.0000 [ 0.3407; 0.4382 ] 
#>   eta1 <~ y13      0.4542      0.0164   27.7032    0.0000 [ 0.4134; 0.4982 ] 
#>   eta2 <~ y21      0.3058      0.0321    9.5124    0.0000 [ 0.2244; 0.3906 ] 
#>   eta2 <~ y22      0.4473      0.0265   16.8824    0.0000 [ 0.3791; 0.5161 ] 
#>   eta2 <~ y23      0.4735      0.0265   17.8599    0.0000 [ 0.4051; 0.5422 ] 
#>   eta3 <~ y31      0.4400      0.0188   23.4093    0.0000 [ 0.3892; 0.4864 ] 
#>   eta3 <~ y32      0.3521      0.0186   18.9282    0.0000 [ 0.3068; 0.4030 ] 
#>   eta3 <~ y33      0.3999      0.0217   18.4139    0.0000 [ 0.3435; 0.4558 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0463   14.4956    0.0000 [ 0.5557; 0.7952 ] 
#>   eta3 ~ eta1       0.6634      0.0354   18.7348    0.0000 [ 0.5816; 0.7647 ] 
#>   eta3 ~ eta2       0.3052      0.0878    3.4768    0.0005 [ 0.0666; 0.5205 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0589    3.4807    0.0005 [ 0.0470; 0.3513 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04631276 14.495648 1.290785e-47
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08299243  5.524682 3.300837e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08776906  3.476751 5.075299e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55572980          0.7952332          0.5844895          0.7664736
#> 2         0.25938253          0.6885725          0.3109199          0.6370352
#> 3         0.06657823          0.5204702          0.1210818          0.4659667
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5527321          0.7677205          0.5652433          0.7349272
#> 2          0.2707900          0.6128401          0.2986256          0.6108676
#> 3          0.1220492          0.4899792          0.1501278          0.4729289