Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1215101660
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0436   15.3927    0.0000 [ 0.5948; 0.7409 ] 
#>   eta3 ~ eta1      0.4585      0.0680    6.7429    0.0000 [ 0.3059; 0.5707 ] 
#>   eta3 ~ eta2      0.3052      0.0656    4.6528    0.0000 [ 0.1834; 0.4446 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0421   15.7602    0.0000 [ 0.5860; 0.7283 ] 
#>   eta1 =~ y12      0.6493      0.0417   15.5862    0.0000 [ 0.5663; 0.7164 ] 
#>   eta1 =~ y13      0.7613      0.0327   23.2824    0.0000 [ 0.7097; 0.8156 ] 
#>   eta2 =~ y21      0.5165      0.0517    9.9864    0.0000 [ 0.4349; 0.6139 ] 
#>   eta2 =~ y22      0.7554      0.0370   20.4009    0.0000 [ 0.6905; 0.8236 ] 
#>   eta2 =~ y23      0.7997      0.0467   17.1082    0.0000 [ 0.7076; 0.8809 ] 
#>   eta3 =~ y31      0.8223      0.0333   24.6746    0.0000 [ 0.7661; 0.8781 ] 
#>   eta3 =~ y32      0.6581      0.0395   16.6422    0.0000 [ 0.5817; 0.7186 ] 
#>   eta3 =~ y33      0.7474      0.0410   18.2093    0.0000 [ 0.6511; 0.7985 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0203   19.4521    0.0000 [ 0.3673; 0.4396 ] 
#>   eta1 <~ y12      0.3873      0.0183   21.1929    0.0000 [ 0.3568; 0.4129 ] 
#>   eta1 <~ y13      0.4542      0.0230   19.7772    0.0000 [ 0.4227; 0.4966 ] 
#>   eta2 <~ y21      0.3058      0.0282   10.8631    0.0000 [ 0.2616; 0.3603 ] 
#>   eta2 <~ y22      0.4473      0.0242   18.5111    0.0000 [ 0.3996; 0.4838 ] 
#>   eta2 <~ y23      0.4735      0.0229   20.6769    0.0000 [ 0.4191; 0.5075 ] 
#>   eta3 <~ y31      0.4400      0.0178   24.6859    0.0000 [ 0.4141; 0.4755 ] 
#>   eta3 <~ y32      0.3521      0.0187   18.8501    0.0000 [ 0.3173; 0.3833 ] 
#>   eta3 <~ y33      0.3999      0.0218   18.3326    0.0000 [ 0.3581; 0.4282 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0436   15.3927    0.0000 [ 0.5948; 0.7409 ] 
#>   eta3 ~ eta1       0.6634      0.0416   15.9638    0.0000 [ 0.5698; 0.7286 ] 
#>   eta3 ~ eta2       0.3052      0.0656    4.6528    0.0000 [ 0.1834; 0.4446 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0451    4.5405    0.0000 [ 0.1273; 0.3065 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04207234 15.760233  5.841679e-56
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04165732 15.586169  9.039512e-55
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03270047 23.282413 6.682551e-120
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05171600  9.986363  1.748800e-23
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03702719 20.400891  1.641964e-92
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04674145 17.108236  1.288437e-65
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03332484 24.674606 2.003941e-134
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03954223 16.642178  3.448447e-62
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04104631 18.209290  4.355557e-74
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5859616          0.7282800
#> 2          0.5662852          0.7164262
#> 3          0.7097260          0.8156122
#> 4          0.4349102          0.6139228
#> 5          0.6904659          0.8235882
#> 6          0.7076437          0.8808689
#> 7          0.7660536          0.8781219
#> 8          0.5817474          0.7185974
#> 9          0.6510524          0.7984855

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1215101660
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0436   15.3927    0.0000 [ 0.5590; 0.7846 ] 
#>   eta3 ~ eta1      0.4585      0.0680    6.7429    0.0000 [ 0.2916; 0.6432 ] 
#>   eta3 ~ eta2      0.3052      0.0656    4.6528    0.0000 [ 0.1343; 0.4734 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0421   15.7602    0.0000 [ 0.5531; 0.7707 ] 
#>   eta1 =~ y12      0.6493      0.0417   15.5862    0.0000 [ 0.5466; 0.7620 ] 
#>   eta1 =~ y13      0.7613      0.0327   23.2824    0.0000 [ 0.6773; 0.8464 ] 
#>   eta2 =~ y21      0.5165      0.0517    9.9864    0.0000 [ 0.3691; 0.6365 ] 
#>   eta2 =~ y22      0.7554      0.0370   20.4009    0.0000 [ 0.6575; 0.8490 ] 
#>   eta2 =~ y23      0.7997      0.0467   17.1082    0.0000 [ 0.6875; 0.9292 ] 
#>   eta3 =~ y31      0.8223      0.0333   24.6746    0.0000 [ 0.7362; 0.9085 ] 
#>   eta3 =~ y32      0.6581      0.0395   16.6422    0.0000 [ 0.5593; 0.7638 ] 
#>   eta3 =~ y33      0.7474      0.0410   18.2093    0.0000 [ 0.6416; 0.8539 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0203   19.4521    0.0000 [ 0.3414; 0.4465 ] 
#>   eta1 <~ y12      0.3873      0.0183   21.1929    0.0000 [ 0.3423; 0.4368 ] 
#>   eta1 <~ y13      0.4542      0.0230   19.7772    0.0000 [ 0.3936; 0.5124 ] 
#>   eta2 <~ y21      0.3058      0.0282   10.8631    0.0000 [ 0.2265; 0.3721 ] 
#>   eta2 <~ y22      0.4473      0.0242   18.5111    0.0000 [ 0.3854; 0.5103 ] 
#>   eta2 <~ y23      0.4735      0.0229   20.6769    0.0000 [ 0.4217; 0.5402 ] 
#>   eta3 <~ y31      0.4400      0.0178   24.6859    0.0000 [ 0.3930; 0.4852 ] 
#>   eta3 <~ y32      0.3521      0.0187   18.8501    0.0000 [ 0.3051; 0.4017 ] 
#>   eta3 <~ y33      0.3999      0.0218   18.3326    0.0000 [ 0.3429; 0.4557 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0436   15.3927    0.0000 [ 0.5590; 0.7846 ] 
#>   eta3 ~ eta1       0.6634      0.0416   15.9638    0.0000 [ 0.5644; 0.7793 ] 
#>   eta3 ~ eta2       0.3052      0.0656    4.6528    0.0000 [ 0.1343; 0.4734 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0451    4.5405    0.0000 [ 0.0878; 0.3211 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04361372 15.392711 1.831932e-53
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.06799837  6.742908 1.552478e-11
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.06558506  4.652753 3.275320e-06
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5590207          0.7845662          0.5861043          0.7574826
#> 2          0.2915767          0.6432259          0.3338029          0.6009997
#> 3          0.1342547          0.4734237          0.1749823          0.4326961
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5898119          0.7596163          0.5948339          0.7409147
#> 2          0.2940000          0.5806710          0.3058845          0.5706751
#> 3          0.1660900          0.4451520          0.1833584          0.4446361