Summarize model

Usage

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -602767520
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0519   12.9309    0.0000 [ 0.5692; 0.7558 ] 
#>   eta3 ~ eta1      0.4585      0.0829    5.5295    0.0000 [ 0.3415; 0.6295 ] 
#>   eta3 ~ eta2      0.3052      0.0857    3.5625    0.0004 [ 0.1345; 0.4206 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0290   22.8893    0.0000 [ 0.6162; 0.7301 ] 
#>   eta1 =~ y12      0.6493      0.0381   17.0268    0.0000 [ 0.5720; 0.7063 ] 
#>   eta1 =~ y13      0.7613      0.0326   23.3871    0.0000 [ 0.7021; 0.8089 ] 
#>   eta2 =~ y21      0.5165      0.0581    8.8920    0.0000 [ 0.3735; 0.5861 ] 
#>   eta2 =~ y22      0.7554      0.0361   20.9261    0.0000 [ 0.6927; 0.8202 ] 
#>   eta2 =~ y23      0.7997      0.0346   23.0986    0.0000 [ 0.7373; 0.8618 ] 
#>   eta3 =~ y31      0.8223      0.0315   26.0956    0.0000 [ 0.7520; 0.8691 ] 
#>   eta3 =~ y32      0.6581      0.0489   13.4516    0.0000 [ 0.5585; 0.7339 ] 
#>   eta3 =~ y33      0.7474      0.0477   15.6739    0.0000 [ 0.6440; 0.8236 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0197   20.0653    0.0000 [ 0.3676; 0.4369 ] 
#>   eta1 <~ y12      0.3873      0.0147   26.3275    0.0000 [ 0.3599; 0.4157 ] 
#>   eta1 <~ y13      0.4542      0.0187   24.2284    0.0000 [ 0.4100; 0.4860 ] 
#>   eta2 <~ y21      0.3058      0.0305   10.0327    0.0000 [ 0.2255; 0.3406 ] 
#>   eta2 <~ y22      0.4473      0.0189   23.7081    0.0000 [ 0.4273; 0.4803 ] 
#>   eta2 <~ y23      0.4735      0.0239   19.8352    0.0000 [ 0.4326; 0.5192 ] 
#>   eta3 <~ y31      0.4400      0.0233   18.9095    0.0000 [ 0.3864; 0.4785 ] 
#>   eta3 <~ y32      0.3521      0.0196   18.0106    0.0000 [ 0.3122; 0.3959 ] 
#>   eta3 <~ y33      0.3999      0.0233   17.1858    0.0000 [ 0.3505; 0.4487 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0519   12.9309    0.0000 [ 0.5692; 0.7558 ] 
#>   eta3 ~ eta1       0.6634      0.0419   15.8453    0.0000 [ 0.5886; 0.7339 ] 
#>   eta3 ~ eta2       0.3052      0.0857    3.5625    0.0004 [ 0.1345; 0.4206 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0574    3.5718    0.0004 [ 0.0916; 0.3061 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.02896850 22.889339 5.933408e-116
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03813279 17.026763  5.200264e-65
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03255407 23.387119 5.779608e-121
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05808096  8.891982  6.002843e-19
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03609784 20.926121  3.096995e-97
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03461955 23.098619 4.780379e-118
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03151014 26.095640 4.085870e-150
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04892138 13.451561  3.014392e-41
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04768603 15.673858  2.282909e-55
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6161619          0.7301045
#> 2          0.5719685          0.7063172
#> 3          0.7021251          0.8088675
#> 4          0.3735259          0.5861137
#> 5          0.6927195          0.8201709
#> 6          0.7372719          0.8617874
#> 7          0.7519506          0.8690608
#> 8          0.5584594          0.7338764
#> 9          0.6439522          0.8235972

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -602767520
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0519   12.9309    0.0000 [ 0.5425; 0.8110 ] 
#>   eta3 ~ eta1      0.4585      0.0829    5.5295    0.0000 [ 0.2340; 0.6628 ] 
#>   eta3 ~ eta2      0.3052      0.0857    3.5625    0.0004 [ 0.0965; 0.5395 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0290   22.8893    0.0000 [ 0.5833; 0.7331 ] 
#>   eta1 =~ y12      0.6493      0.0381   17.0268    0.0000 [ 0.5576; 0.7548 ] 
#>   eta1 =~ y13      0.7613      0.0326   23.3871    0.0000 [ 0.6793; 0.8476 ] 
#>   eta2 =~ y21      0.5165      0.0581    8.8920    0.0000 [ 0.3839; 0.6842 ] 
#>   eta2 =~ y22      0.7554      0.0361   20.9261    0.0000 [ 0.6554; 0.8421 ] 
#>   eta2 =~ y23      0.7997      0.0346   23.0986    0.0000 [ 0.7163; 0.8953 ] 
#>   eta3 =~ y31      0.8223      0.0315   26.0956    0.0000 [ 0.7531; 0.9160 ] 
#>   eta3 =~ y32      0.6581      0.0489   13.4516    0.0000 [ 0.5268; 0.7798 ] 
#>   eta3 =~ y33      0.7474      0.0477   15.6739    0.0000 [ 0.6292; 0.8758 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0197   20.0653    0.0000 [ 0.3404; 0.4424 ] 
#>   eta1 <~ y12      0.3873      0.0147   26.3275    0.0000 [ 0.3527; 0.4288 ] 
#>   eta1 <~ y13      0.4542      0.0187   24.2284    0.0000 [ 0.4057; 0.5027 ] 
#>   eta2 <~ y21      0.3058      0.0305   10.0327    0.0000 [ 0.2351; 0.3927 ] 
#>   eta2 <~ y22      0.4473      0.0189   23.7081    0.0000 [ 0.3904; 0.4880 ] 
#>   eta2 <~ y23      0.4735      0.0239   19.8352    0.0000 [ 0.4108; 0.5342 ] 
#>   eta3 <~ y31      0.4400      0.0233   18.9095    0.0000 [ 0.3823; 0.5027 ] 
#>   eta3 <~ y32      0.3521      0.0196   18.0106    0.0000 [ 0.2964; 0.3975 ] 
#>   eta3 <~ y33      0.3999      0.0233   17.1858    0.0000 [ 0.3392; 0.4595 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0519   12.9309    0.0000 [ 0.5425; 0.8110 ] 
#>   eta3 ~ eta1       0.6634      0.0419   15.8453    0.0000 [ 0.5562; 0.7727 ] 
#>   eta3 ~ eta2       0.3052      0.0857    3.5625    0.0004 [ 0.0965; 0.5395 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0574    3.5718    0.0004 [ 0.0677; 0.3643 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05191709 12.930876 3.013671e-38
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08292029  5.529488 3.211659e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08565626  3.562508 3.673286e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54254865          0.8110345          0.5747886          0.7787946
#> 2         0.23402727          0.6628442          0.2855198          0.6113516
#> 3         0.09649013          0.5394559          0.1496817          0.4862644
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5628779          0.7725017          0.5691582          0.7558416
#> 2          0.3394982          0.6298116          0.3414924          0.6295352
#> 3          0.1263031          0.4370055          0.1344842          0.4205521