Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 227164290
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0464   14.4624    0.0000 [ 0.5971; 0.7610 ] 
#>   eta3 ~ eta1      0.4585      0.0749    6.1249    0.0000 [ 0.3075; 0.5660 ] 
#>   eta3 ~ eta2      0.3052      0.0838    3.6431    0.0003 [ 0.1739; 0.4371 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0357   18.5852    0.0000 [ 0.5978; 0.7216 ] 
#>   eta1 =~ y12      0.6493      0.0399   16.2783    0.0000 [ 0.5817; 0.7143 ] 
#>   eta1 =~ y13      0.7613      0.0344   22.1301    0.0000 [ 0.7151; 0.8379 ] 
#>   eta2 =~ y21      0.5165      0.0614    8.4111    0.0000 [ 0.4241; 0.6121 ] 
#>   eta2 =~ y22      0.7554      0.0330   22.8667    0.0000 [ 0.6929; 0.7955 ] 
#>   eta2 =~ y23      0.7997      0.0365   21.9245    0.0000 [ 0.7354; 0.8687 ] 
#>   eta3 =~ y31      0.8223      0.0332   24.7483    0.0000 [ 0.7648; 0.8809 ] 
#>   eta3 =~ y32      0.6581      0.0379   17.3573    0.0000 [ 0.5617; 0.7020 ] 
#>   eta3 =~ y33      0.7474      0.0373   20.0553    0.0000 [ 0.6713; 0.8028 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0186   21.3071    0.0000 [ 0.3638; 0.4271 ] 
#>   eta1 <~ y12      0.3873      0.0195   19.8968    0.0000 [ 0.3525; 0.4234 ] 
#>   eta1 <~ y13      0.4542      0.0214   21.1906    0.0000 [ 0.4260; 0.5002 ] 
#>   eta2 <~ y21      0.3058      0.0311    9.8233    0.0000 [ 0.2573; 0.3621 ] 
#>   eta2 <~ y22      0.4473      0.0235   19.0283    0.0000 [ 0.4038; 0.4856 ] 
#>   eta2 <~ y23      0.4735      0.0195   24.2575    0.0000 [ 0.4410; 0.5080 ] 
#>   eta3 <~ y31      0.4400      0.0213   20.6918    0.0000 [ 0.4126; 0.4819 ] 
#>   eta3 <~ y32      0.3521      0.0166   21.1786    0.0000 [ 0.3203; 0.3762 ] 
#>   eta3 <~ y33      0.3999      0.0161   24.8774    0.0000 [ 0.3755; 0.4368 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0464   14.4624    0.0000 [ 0.5971; 0.7610 ] 
#>   eta3 ~ eta1       0.6634      0.0316   20.9747    0.0000 [ 0.5923; 0.7028 ] 
#>   eta3 ~ eta2       0.3052      0.0838    3.6431    0.0003 [ 0.1739; 0.4371 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0595    3.4441    0.0006 [ 0.1050; 0.3061 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03567733 18.585187  4.235214e-77
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03988606 16.278316  1.406883e-59
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03440323 22.130068 1.623308e-108
#> 4 eta2 =~ y21  Common factor 0.5164548 0.06140123  8.411148  4.060169e-17
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03303437 22.866721 9.964457e-116
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03647352 21.924503 1.516848e-106
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03322560 24.748311 3.232882e-135
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03791317 17.357264  1.738116e-67
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03726825 20.055253  1.816161e-89
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5977523          0.7216131
#> 2          0.5817431          0.7143279
#> 3          0.7150505          0.8379334
#> 4          0.4241293          0.6120885
#> 5          0.6928532          0.7954862
#> 6          0.7353641          0.8687297
#> 7          0.7648169          0.8809010
#> 8          0.5616739          0.7019609
#> 9          0.6712961          0.8027842

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 227164290
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0464   14.4624    0.0000 [ 0.5424; 0.7824 ] 
#>   eta3 ~ eta1      0.4585      0.0749    6.1249    0.0000 [ 0.2720; 0.6592 ] 
#>   eta3 ~ eta2      0.3052      0.0838    3.6431    0.0003 [ 0.0798; 0.5129 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0357   18.5852    0.0000 [ 0.5762; 0.7607 ] 
#>   eta1 =~ y12      0.6493      0.0399   16.2783    0.0000 [ 0.5384; 0.7447 ] 
#>   eta1 =~ y13      0.7613      0.0344   22.1301    0.0000 [ 0.6620; 0.8400 ] 
#>   eta2 =~ y21      0.5165      0.0614    8.4111    0.0000 [ 0.3556; 0.6732 ] 
#>   eta2 =~ y22      0.7554      0.0330   22.8667    0.0000 [ 0.6767; 0.8476 ] 
#>   eta2 =~ y23      0.7997      0.0365   21.9245    0.0000 [ 0.7020; 0.8907 ] 
#>   eta3 =~ y31      0.8223      0.0332   24.7483    0.0000 [ 0.7282; 0.9001 ] 
#>   eta3 =~ y32      0.6581      0.0379   17.3573    0.0000 [ 0.5724; 0.7685 ] 
#>   eta3 =~ y33      0.7474      0.0373   20.0553    0.0000 [ 0.6520; 0.8447 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0186   21.3071    0.0000 [ 0.3546; 0.4506 ] 
#>   eta1 <~ y12      0.3873      0.0195   19.8968    0.0000 [ 0.3363; 0.4370 ] 
#>   eta1 <~ y13      0.4542      0.0214   21.1906    0.0000 [ 0.3969; 0.5077 ] 
#>   eta2 <~ y21      0.3058      0.0311    9.8233    0.0000 [ 0.2246; 0.3856 ] 
#>   eta2 <~ y22      0.4473      0.0235   19.0283    0.0000 [ 0.3902; 0.5118 ] 
#>   eta2 <~ y23      0.4735      0.0195   24.2575    0.0000 [ 0.4210; 0.5220 ] 
#>   eta3 <~ y31      0.4400      0.0213   20.6918    0.0000 [ 0.3793; 0.4892 ] 
#>   eta3 <~ y32      0.3521      0.0166   21.1786    0.0000 [ 0.3150; 0.4010 ] 
#>   eta3 <~ y33      0.3999      0.0161   24.8774    0.0000 [ 0.3579; 0.4410 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0464   14.4624    0.0000 [ 0.5424; 0.7824 ] 
#>   eta3 ~ eta1       0.6634      0.0316   20.9747    0.0000 [ 0.5799; 0.7434 ] 
#>   eta3 ~ eta2       0.3052      0.0838    3.6431    0.0003 [ 0.0798; 0.5129 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0595    3.4441    0.0006 [ 0.0423; 0.3499 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04641934 14.462365 2.094761e-47
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07485901  6.124938 9.071907e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08376147  3.643097 2.693776e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54235530          0.7824099          0.5711812          0.7535841
#> 2         0.27202213          0.6591507          0.3185087          0.6126641
#> 3         0.07976016          0.5129272          0.1317751          0.4609123
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5740227          0.7667169          0.5970795          0.7610221
#> 2          0.3002564          0.6036183          0.3074693          0.5660488
#> 3          0.1337393          0.4544559          0.1738661          0.4371037