Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1233608938
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0382   17.5599    0.0000 [ 0.6031; 0.7341 ] 
#>   eta3 ~ eta1      0.4585      0.0730    6.2817    0.0000 [ 0.3452; 0.5893 ] 
#>   eta3 ~ eta2      0.3052      0.0849    3.5958    0.0003 [ 0.1604; 0.4128 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0430   15.4216    0.0000 [ 0.5790; 0.7238 ] 
#>   eta1 =~ y12      0.6493      0.0403   16.1149    0.0000 [ 0.5772; 0.7368 ] 
#>   eta1 =~ y13      0.7613      0.0335   22.7162    0.0000 [ 0.7088; 0.8365 ] 
#>   eta2 =~ y21      0.5165      0.0523    9.8747    0.0000 [ 0.4320; 0.6004 ] 
#>   eta2 =~ y22      0.7554      0.0395   19.1110    0.0000 [ 0.6920; 0.8276 ] 
#>   eta2 =~ y23      0.7997      0.0260   30.7936    0.0000 [ 0.7592; 0.8527 ] 
#>   eta3 =~ y31      0.8223      0.0341   24.0918    0.0000 [ 0.7650; 0.8767 ] 
#>   eta3 =~ y32      0.6581      0.0439   14.9762    0.0000 [ 0.5624; 0.7244 ] 
#>   eta3 =~ y33      0.7474      0.0429   17.4190    0.0000 [ 0.6695; 0.8114 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0224   17.6606    0.0000 [ 0.3548; 0.4341 ] 
#>   eta1 <~ y12      0.3873      0.0174   22.3062    0.0000 [ 0.3498; 0.4126 ] 
#>   eta1 <~ y13      0.4542      0.0242   18.7739    0.0000 [ 0.4256; 0.5076 ] 
#>   eta2 <~ y21      0.3058      0.0270   11.3196    0.0000 [ 0.2642; 0.3447 ] 
#>   eta2 <~ y22      0.4473      0.0178   25.0638    0.0000 [ 0.4152; 0.4816 ] 
#>   eta2 <~ y23      0.4735      0.0218   21.7528    0.0000 [ 0.4425; 0.5166 ] 
#>   eta3 <~ y31      0.4400      0.0216   20.3957    0.0000 [ 0.4061; 0.4823 ] 
#>   eta3 <~ y32      0.3521      0.0188   18.7768    0.0000 [ 0.3175; 0.3801 ] 
#>   eta3 <~ y33      0.3999      0.0214   18.6576    0.0000 [ 0.3532; 0.4299 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0382   17.5599    0.0000 [ 0.6031; 0.7341 ] 
#>   eta3 ~ eta1       0.6634      0.0372   17.8324    0.0000 [ 0.5820; 0.7232 ] 
#>   eta3 ~ eta2       0.3052      0.0849    3.5958    0.0003 [ 0.1604; 0.4128 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0562    3.6472    0.0003 [ 0.1082; 0.2880 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04299618 15.421601  1.171637e-53
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04029053 16.114902  2.004801e-58
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03351548 22.716243 3.095815e-114
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05230102  9.874661  5.361381e-23
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03952624 19.111045  2.043362e-81
#> 6 eta2 =~ y23  Common factor 0.7996637 0.02596854 30.793558 3.196161e-208
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03413095 24.091839 3.043984e-128
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04394094 14.976212  1.050354e-50
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04290867 17.418953  5.925174e-68
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5790328          0.7237570
#> 2          0.5772185          0.7367687
#> 3          0.7088077          0.8365293
#> 4          0.4320112          0.6004363
#> 5          0.6919888          0.8276393
#> 6          0.7591763          0.8527473
#> 7          0.7650144          0.8766789
#> 8          0.5623933          0.7243813
#> 9          0.6695269          0.8113609

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1233608938
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0382   17.5599    0.0000 [ 0.5704; 0.7681 ] 
#>   eta3 ~ eta1      0.4585      0.0730    6.2817    0.0000 [ 0.2594; 0.6369 ] 
#>   eta3 ~ eta2      0.3052      0.0849    3.5958    0.0003 [ 0.1012; 0.5401 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0430   15.4216    0.0000 [ 0.5497; 0.7720 ] 
#>   eta1 =~ y12      0.6493      0.0403   16.1149    0.0000 [ 0.5422; 0.7505 ] 
#>   eta1 =~ y13      0.7613      0.0335   22.7162    0.0000 [ 0.6754; 0.8487 ] 
#>   eta2 =~ y21      0.5165      0.0523    9.8747    0.0000 [ 0.3794; 0.6498 ] 
#>   eta2 =~ y22      0.7554      0.0395   19.1110    0.0000 [ 0.6552; 0.8596 ] 
#>   eta2 =~ y23      0.7997      0.0260   30.7936    0.0000 [ 0.7315; 0.8658 ] 
#>   eta3 =~ y31      0.8223      0.0341   24.0918    0.0000 [ 0.7323; 0.9088 ] 
#>   eta3 =~ y32      0.6581      0.0439   14.9762    0.0000 [ 0.5396; 0.7668 ] 
#>   eta3 =~ y33      0.7474      0.0429   17.4190    0.0000 [ 0.6359; 0.8578 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0224   17.6606    0.0000 [ 0.3379; 0.4537 ] 
#>   eta1 <~ y12      0.3873      0.0174   22.3062    0.0000 [ 0.3423; 0.4321 ] 
#>   eta1 <~ y13      0.4542      0.0242   18.7739    0.0000 [ 0.3934; 0.5185 ] 
#>   eta2 <~ y21      0.3058      0.0270   11.3196    0.0000 [ 0.2356; 0.3753 ] 
#>   eta2 <~ y22      0.4473      0.0178   25.0638    0.0000 [ 0.4031; 0.4954 ] 
#>   eta2 <~ y23      0.4735      0.0218   21.7528    0.0000 [ 0.4169; 0.5295 ] 
#>   eta3 <~ y31      0.4400      0.0216   20.3957    0.0000 [ 0.3854; 0.4970 ] 
#>   eta3 <~ y32      0.3521      0.0188   18.7768    0.0000 [ 0.3031; 0.4001 ] 
#>   eta3 <~ y33      0.3999      0.0214   18.6576    0.0000 [ 0.3463; 0.4572 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0382   17.5599    0.0000 [ 0.5704; 0.7681 ] 
#>   eta3 ~ eta1       0.6634      0.0372   17.8324    0.0000 [ 0.5673; 0.7597 ] 
#>   eta3 ~ eta2       0.3052      0.0849    3.5958    0.0003 [ 0.1012; 0.5401 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0562    3.6472    0.0003 [ 0.0701; 0.3605 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03823095 17.559946 4.992288e-69
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07299132  6.281661 3.349742e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08486431  3.595753 3.234541e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5703871          0.7680960          0.5941281          0.7443550
#> 2          0.2594497          0.6369196          0.3047765          0.5915929
#> 3          0.1012377          0.5401080          0.1539374          0.4874082
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5914433          0.7601067          0.6031451          0.7340992
#> 2          0.3256196          0.6197328          0.3452080          0.5893223
#> 3          0.1580022          0.4315795          0.1604442          0.4127805