Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -362897310
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0429   15.6479    0.0000 [ 0.5878; 0.7367 ] 
#>   eta3 ~ eta1      0.4585      0.0764    6.0005    0.0000 [ 0.2698; 0.5590 ] 
#>   eta3 ~ eta2      0.3052      0.0890    3.4279    0.0006 [ 0.1778; 0.5085 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0388   17.0887    0.0000 [ 0.5805; 0.7233 ] 
#>   eta1 =~ y12      0.6493      0.0356   18.2500    0.0000 [ 0.6045; 0.7163 ] 
#>   eta1 =~ y13      0.7613      0.0311   24.4681    0.0000 [ 0.7125; 0.8216 ] 
#>   eta2 =~ y21      0.5165      0.0537    9.6260    0.0000 [ 0.4138; 0.6081 ] 
#>   eta2 =~ y22      0.7554      0.0361   20.9080    0.0000 [ 0.6906; 0.8172 ] 
#>   eta2 =~ y23      0.7997      0.0388   20.5969    0.0000 [ 0.7141; 0.8685 ] 
#>   eta3 =~ y31      0.8223      0.0326   25.2107    0.0000 [ 0.7622; 0.8660 ] 
#>   eta3 =~ y32      0.6581      0.0395   16.6792    0.0000 [ 0.5819; 0.7246 ] 
#>   eta3 =~ y33      0.7474      0.0428   17.4581    0.0000 [ 0.6872; 0.8210 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0192   20.6206    0.0000 [ 0.3599; 0.4255 ] 
#>   eta1 <~ y12      0.3873      0.0153   25.3100    0.0000 [ 0.3631; 0.4195 ] 
#>   eta1 <~ y13      0.4542      0.0208   21.8538    0.0000 [ 0.4161; 0.4862 ] 
#>   eta2 <~ y21      0.3058      0.0271   11.2700    0.0000 [ 0.2519; 0.3543 ] 
#>   eta2 <~ y22      0.4473      0.0211   21.2274    0.0000 [ 0.4133; 0.4870 ] 
#>   eta2 <~ y23      0.4735      0.0224   21.1129    0.0000 [ 0.4375; 0.5164 ] 
#>   eta3 <~ y31      0.4400      0.0165   26.6618    0.0000 [ 0.4099; 0.4683 ] 
#>   eta3 <~ y32      0.3521      0.0208   16.9144    0.0000 [ 0.3197; 0.3894 ] 
#>   eta3 <~ y33      0.3999      0.0210   19.0439    0.0000 [ 0.3608; 0.4303 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0429   15.6479    0.0000 [ 0.5878; 0.7367 ] 
#>   eta3 ~ eta1       0.6634      0.0318   20.8693    0.0000 [ 0.5945; 0.6982 ] 
#>   eta3 ~ eta2       0.3052      0.0890    3.4279    0.0006 [ 0.1778; 0.5085 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0598    3.4258    0.0006 [ 0.1211; 0.3257 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03880167 17.088692  1.801692e-65
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03557679 18.250043  2.067454e-74
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03111581 24.468131 3.227447e-132
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05365226  9.625966  6.212043e-22
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03612907 20.908031  4.525284e-97
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03882446 20.596906  2.925452e-94
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03261618 25.210721 3.055869e-140
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03945449 16.679188  1.857257e-62
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04281251 17.458078  2.988267e-68
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5805393          0.7233105
#> 2          0.6045112          0.7163244
#> 3          0.7125146          0.8216482
#> 4          0.4138135          0.6080604
#> 5          0.6906239          0.8171669
#> 6          0.7141183          0.8684606
#> 7          0.7621834          0.8659824
#> 8          0.5819294          0.7246336
#> 9          0.6871814          0.8209801

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -362897310
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0429   15.6479    0.0000 [ 0.5549; 0.7767 ] 
#>   eta3 ~ eta1      0.4585      0.0764    6.0005    0.0000 [ 0.2817; 0.6768 ] 
#>   eta3 ~ eta2      0.3052      0.0890    3.4279    0.0006 [ 0.0509; 0.5112 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0388   17.0887    0.0000 [ 0.5651; 0.7657 ] 
#>   eta1 =~ y12      0.6493      0.0356   18.2500    0.0000 [ 0.5517; 0.7357 ] 
#>   eta1 =~ y13      0.7613      0.0311   24.4681    0.0000 [ 0.6797; 0.8406 ] 
#>   eta2 =~ y21      0.5165      0.0537    9.6260    0.0000 [ 0.3871; 0.6645 ] 
#>   eta2 =~ y22      0.7554      0.0361   20.9080    0.0000 [ 0.6602; 0.8471 ] 
#>   eta2 =~ y23      0.7997      0.0388   20.5969    0.0000 [ 0.7017; 0.9024 ] 
#>   eta3 =~ y31      0.8223      0.0326   25.2107    0.0000 [ 0.7376; 0.9063 ] 
#>   eta3 =~ y32      0.6581      0.0395   16.6792    0.0000 [ 0.5456; 0.7496 ] 
#>   eta3 =~ y33      0.7474      0.0428   17.4581    0.0000 [ 0.6376; 0.8590 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0192   20.6206    0.0000 [ 0.3488; 0.4480 ] 
#>   eta1 <~ y12      0.3873      0.0153   25.3100    0.0000 [ 0.3460; 0.4251 ] 
#>   eta1 <~ y13      0.4542      0.0208   21.8538    0.0000 [ 0.4011; 0.5086 ] 
#>   eta2 <~ y21      0.3058      0.0271   11.2700    0.0000 [ 0.2400; 0.3803 ] 
#>   eta2 <~ y22      0.4473      0.0211   21.2274    0.0000 [ 0.3893; 0.4982 ] 
#>   eta2 <~ y23      0.4735      0.0224   21.1129    0.0000 [ 0.4143; 0.5303 ] 
#>   eta3 <~ y31      0.4400      0.0165   26.6618    0.0000 [ 0.4002; 0.4855 ] 
#>   eta3 <~ y32      0.3521      0.0208   16.9144    0.0000 [ 0.2952; 0.4028 ] 
#>   eta3 <~ y33      0.3999      0.0210   19.0439    0.0000 [ 0.3489; 0.4576 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0429   15.6479    0.0000 [ 0.5549; 0.7767 ] 
#>   eta3 ~ eta1       0.6634      0.0318   20.8693    0.0000 [ 0.5844; 0.7488 ] 
#>   eta3 ~ eta2       0.3052      0.0890    3.4279    0.0006 [ 0.0509; 0.5112 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0598    3.4258    0.0006 [ 0.0328; 0.3420 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04290246 15.647900 3.433636e-55
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07641190  6.000463 1.967554e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08902094  3.427858 6.083642e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55485561          0.7767229          0.5814975          0.7500810
#> 2         0.28165588          0.6768151          0.3291068          0.6293642
#> 3         0.05088253          0.5112486          0.1061635          0.4559676
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5755117          0.7626072          0.5878372          0.7366853
#> 2          0.2694848          0.5890290          0.2697617          0.5590108
#> 3          0.1311078          0.5326603          0.1777958          0.5085019