Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 2022163901
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0430   15.5980    0.0000 [ 0.5791; 0.7403 ] 
#>   eta3 ~ eta1      0.4585      0.0887    5.1663    0.0000 [ 0.2899; 0.6208 ] 
#>   eta3 ~ eta2      0.3052      0.0922    3.3084    0.0009 [ 0.1310; 0.5094 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0393   16.8528    0.0000 [ 0.5809; 0.7284 ] 
#>   eta1 =~ y12      0.6493      0.0328   19.8040    0.0000 [ 0.5852; 0.7026 ] 
#>   eta1 =~ y13      0.7613      0.0378   20.1244    0.0000 [ 0.6952; 0.8196 ] 
#>   eta2 =~ y21      0.5165      0.0572    9.0291    0.0000 [ 0.4016; 0.6068 ] 
#>   eta2 =~ y22      0.7554      0.0314   24.0426    0.0000 [ 0.7009; 0.7987 ] 
#>   eta2 =~ y23      0.7997      0.0304   26.3351    0.0000 [ 0.7543; 0.8555 ] 
#>   eta3 =~ y31      0.8223      0.0288   28.5288    0.0000 [ 0.7796; 0.8796 ] 
#>   eta3 =~ y32      0.6581      0.0377   17.4604    0.0000 [ 0.6050; 0.7348 ] 
#>   eta3 =~ y33      0.7474      0.0399   18.7278    0.0000 [ 0.6791; 0.8093 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0209   18.9437    0.0000 [ 0.3503; 0.4337 ] 
#>   eta1 <~ y12      0.3873      0.0142   27.3701    0.0000 [ 0.3541; 0.4071 ] 
#>   eta1 <~ y13      0.4542      0.0239   19.0026    0.0000 [ 0.4097; 0.5002 ] 
#>   eta2 <~ y21      0.3058      0.0278   10.9941    0.0000 [ 0.2556; 0.3443 ] 
#>   eta2 <~ y22      0.4473      0.0220   20.3383    0.0000 [ 0.4145; 0.4873 ] 
#>   eta2 <~ y23      0.4735      0.0207   22.9103    0.0000 [ 0.4460; 0.5151 ] 
#>   eta3 <~ y31      0.4400      0.0177   24.9232    0.0000 [ 0.4105; 0.4697 ] 
#>   eta3 <~ y32      0.3521      0.0189   18.6349    0.0000 [ 0.3234; 0.3893 ] 
#>   eta3 <~ y33      0.3999      0.0168   23.8493    0.0000 [ 0.3595; 0.4291 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0430   15.5980    0.0000 [ 0.5791; 0.7403 ] 
#>   eta3 ~ eta1       0.6634      0.0427   15.5515    0.0000 [ 0.5679; 0.7273 ] 
#>   eta3 ~ eta2       0.3052      0.0922    3.3084    0.0009 [ 0.1310; 0.5094 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0618    3.3129    0.0009 [ 0.0975; 0.3485 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03934486 16.852768  1.001175e-63
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03278518 19.804007  2.749300e-87
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03783200 20.124385  4.513373e-90
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05719920  9.029056  1.731592e-19
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03141875 24.042572 9.983238e-128
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03036493 26.335110 7.601721e-153
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02882269 28.528817 5.145383e-179
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03768915 17.460432  2.867535e-68
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03990994 18.727769  2.939813e-78
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5808909          0.7283725
#> 2          0.5852327          0.7026124
#> 3          0.6952252          0.8196145
#> 4          0.4015871          0.6067585
#> 5          0.7009020          0.7987431
#> 6          0.7543442          0.8555145
#> 7          0.7795806          0.8795841
#> 8          0.6049686          0.7348144
#> 9          0.6791301          0.8093105

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 2022163901
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0430   15.5980    0.0000 [ 0.5682; 0.7908 ] 
#>   eta3 ~ eta1      0.4585      0.0887    5.1663    0.0000 [ 0.2503; 0.7093 ] 
#>   eta3 ~ eta2      0.3052      0.0922    3.3084    0.0009 [ 0.0447; 0.5217 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0393   16.8528    0.0000 [ 0.5587; 0.7622 ] 
#>   eta1 =~ y12      0.6493      0.0328   19.8040    0.0000 [ 0.5654; 0.7350 ] 
#>   eta1 =~ y13      0.7613      0.0378   20.1244    0.0000 [ 0.6677; 0.8634 ] 
#>   eta2 =~ y21      0.5165      0.0572    9.0291    0.0000 [ 0.3702; 0.6660 ] 
#>   eta2 =~ y22      0.7554      0.0314   24.0426    0.0000 [ 0.6848; 0.8473 ] 
#>   eta2 =~ y23      0.7997      0.0304   26.3351    0.0000 [ 0.7174; 0.8744 ] 
#>   eta3 =~ y31      0.8223      0.0288   28.5288    0.0000 [ 0.7432; 0.8922 ] 
#>   eta3 =~ y32      0.6581      0.0377   17.4604    0.0000 [ 0.5533; 0.7482 ] 
#>   eta3 =~ y33      0.7474      0.0399   18.7278    0.0000 [ 0.6523; 0.8587 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0209   18.9437    0.0000 [ 0.3395; 0.4475 ] 
#>   eta1 <~ y12      0.3873      0.0142   27.3701    0.0000 [ 0.3509; 0.4240 ] 
#>   eta1 <~ y13      0.4542      0.0239   19.0026    0.0000 [ 0.3941; 0.5177 ] 
#>   eta2 <~ y21      0.3058      0.0278   10.9941    0.0000 [ 0.2339; 0.3778 ] 
#>   eta2 <~ y22      0.4473      0.0220   20.3383    0.0000 [ 0.3943; 0.5080 ] 
#>   eta2 <~ y23      0.4735      0.0207   22.9103    0.0000 [ 0.4152; 0.5221 ] 
#>   eta3 <~ y31      0.4400      0.0177   24.9232    0.0000 [ 0.3930; 0.4843 ] 
#>   eta3 <~ y32      0.3521      0.0189   18.6349    0.0000 [ 0.3004; 0.3981 ] 
#>   eta3 <~ y33      0.3999      0.0168   23.8493    0.0000 [ 0.3622; 0.4489 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0430   15.5980    0.0000 [ 0.5682; 0.7908 ] 
#>   eta3 ~ eta1       0.6634      0.0427   15.5515    0.0000 [ 0.5630; 0.7836 ] 
#>   eta3 ~ eta2       0.3052      0.0922    3.3084    0.0009 [ 0.0447; 0.5217 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0618    3.3129    0.0009 [ 0.0336; 0.3534 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04303965 15.598021 7.508597e-55
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08874938  5.166309 2.387613e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09223487  3.308414 9.382606e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56821967          0.7907965          0.5949468          0.7640693
#> 2         0.25030907          0.7092708          0.3054214          0.6541584
#> 3         0.04474827          0.5217349          0.1020250          0.4644582
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5783687          0.7450574          0.5791493          0.7402942
#> 2          0.2538107          0.6459515          0.2898620          0.6207670
#> 3          0.1283409          0.5435355          0.1309649          0.5093966