Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 485658682
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0461   14.5710    0.0000 [ 0.6023; 0.7758 ] 
#>   eta3 ~ eta1      0.4585      0.0746    6.1426    0.0000 [ 0.3212; 0.5770 ] 
#>   eta3 ~ eta2      0.3052      0.0832    3.6681    0.0002 [ 0.1530; 0.4589 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0394   16.8240    0.0000 [ 0.6002; 0.7317 ] 
#>   eta1 =~ y12      0.6493      0.0394   16.4875    0.0000 [ 0.5543; 0.7156 ] 
#>   eta1 =~ y13      0.7613      0.0349   21.8026    0.0000 [ 0.6903; 0.8197 ] 
#>   eta2 =~ y21      0.5165      0.0504   10.2405    0.0000 [ 0.4333; 0.6128 ] 
#>   eta2 =~ y22      0.7554      0.0351   21.4955    0.0000 [ 0.6871; 0.8186 ] 
#>   eta2 =~ y23      0.7997      0.0348   22.9976    0.0000 [ 0.7261; 0.8673 ] 
#>   eta3 =~ y31      0.8223      0.0362   22.7101    0.0000 [ 0.7626; 0.8811 ] 
#>   eta3 =~ y32      0.6581      0.0434   15.1508    0.0000 [ 0.5769; 0.7315 ] 
#>   eta3 =~ y33      0.7474      0.0379   19.7314    0.0000 [ 0.6757; 0.8096 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0227   17.4244    0.0000 [ 0.3552; 0.4349 ] 
#>   eta1 <~ y12      0.3873      0.0175   22.1862    0.0000 [ 0.3609; 0.4162 ] 
#>   eta1 <~ y13      0.4542      0.0190   23.8569    0.0000 [ 0.4229; 0.4834 ] 
#>   eta2 <~ y21      0.3058      0.0237   12.8829    0.0000 [ 0.2739; 0.3505 ] 
#>   eta2 <~ y22      0.4473      0.0213   21.0291    0.0000 [ 0.4115; 0.4820 ] 
#>   eta2 <~ y23      0.4735      0.0195   24.2757    0.0000 [ 0.4351; 0.5043 ] 
#>   eta3 <~ y31      0.4400      0.0180   24.4491    0.0000 [ 0.4044; 0.4773 ] 
#>   eta3 <~ y32      0.3521      0.0202   17.4516    0.0000 [ 0.3135; 0.3833 ] 
#>   eta3 <~ y33      0.3999      0.0215   18.5979    0.0000 [ 0.3646; 0.4403 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0461   14.5710    0.0000 [ 0.6023; 0.7758 ] 
#>   eta3 ~ eta1       0.6634      0.0345   19.2311    0.0000 [ 0.6067; 0.7199 ] 
#>   eta3 ~ eta2       0.3052      0.0832    3.6681    0.0002 [ 0.1530; 0.4589 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0536    3.8197    0.0001 [ 0.1090; 0.3138 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03941219 16.82398  1.628426e-63
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03938000 16.48750  4.512001e-61
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03491999 21.80258 2.193085e-105
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05043278 10.24046  1.306142e-24
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03514159 21.49555 1.713562e-102
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03477157 22.99763 4.922730e-117
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03620748 22.71015 3.556282e-114
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04343464 15.15078  7.486711e-52
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03787986 19.73144  1.158243e-86
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6001604          0.7316933
#> 2          0.5542691          0.7156234
#> 3          0.6902569          0.8196741
#> 4          0.4333155          0.6127926
#> 5          0.6870893          0.8186049
#> 6          0.7261113          0.8673293
#> 7          0.7626220          0.8811444
#> 8          0.5769138          0.7314609
#> 9          0.6757441          0.8095578

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 485658682
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0461   14.5710    0.0000 [ 0.5434; 0.7816 ] 
#>   eta3 ~ eta1      0.4585      0.0746    6.1426    0.0000 [ 0.2827; 0.6687 ] 
#>   eta3 ~ eta2      0.3052      0.0832    3.6681    0.0002 [ 0.0783; 0.5085 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0394   16.8240    0.0000 [ 0.5566; 0.7604 ] 
#>   eta1 =~ y12      0.6493      0.0394   16.4875    0.0000 [ 0.5479; 0.7516 ] 
#>   eta1 =~ y13      0.7613      0.0349   21.8026    0.0000 [ 0.6763; 0.8569 ] 
#>   eta2 =~ y21      0.5165      0.0504   10.2405    0.0000 [ 0.3737; 0.6345 ] 
#>   eta2 =~ y22      0.7554      0.0351   21.4955    0.0000 [ 0.6665; 0.8482 ] 
#>   eta2 =~ y23      0.7997      0.0348   22.9976    0.0000 [ 0.7188; 0.8986 ] 
#>   eta3 =~ y31      0.8223      0.0362   22.7101    0.0000 [ 0.7266; 0.9138 ] 
#>   eta3 =~ y32      0.6581      0.0434   15.1508    0.0000 [ 0.5534; 0.7780 ] 
#>   eta3 =~ y33      0.7474      0.0379   19.7314    0.0000 [ 0.6526; 0.8485 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0227   17.4244    0.0000 [ 0.3339; 0.4512 ] 
#>   eta1 <~ y12      0.3873      0.0175   22.1862    0.0000 [ 0.3425; 0.4327 ] 
#>   eta1 <~ y13      0.4542      0.0190   23.8569    0.0000 [ 0.4078; 0.5062 ] 
#>   eta2 <~ y21      0.3058      0.0237   12.8829    0.0000 [ 0.2376; 0.3604 ] 
#>   eta2 <~ y22      0.4473      0.0213   21.0291    0.0000 [ 0.3934; 0.5034 ] 
#>   eta2 <~ y23      0.4735      0.0195   24.2757    0.0000 [ 0.4284; 0.5293 ] 
#>   eta3 <~ y31      0.4400      0.0180   24.4491    0.0000 [ 0.3901; 0.4831 ] 
#>   eta3 <~ y32      0.3521      0.0202   17.4516    0.0000 [ 0.3024; 0.4067 ] 
#>   eta3 <~ y33      0.3999      0.0215   18.5979    0.0000 [ 0.3438; 0.4550 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0461   14.5710    0.0000 [ 0.5434; 0.7816 ] 
#>   eta3 ~ eta1       0.6634      0.0345   19.2311    0.0000 [ 0.5819; 0.7603 ] 
#>   eta3 ~ eta2       0.3052      0.0832    3.6681    0.0002 [ 0.0783; 0.5085 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0536    3.8197    0.0001 [ 0.0568; 0.3341 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04607325 14.571003 4.295349e-48
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07464346  6.142625 8.116893e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08318962  3.668140 2.443218e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54337048          0.7816353          0.5719814          0.7530243
#> 2         0.28265994          0.6686738          0.3290127          0.6223211
#> 3         0.07829045          0.5085002          0.1299502          0.4568404
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5603963          0.7997425          0.6023390          0.7758342
#> 2          0.3161584          0.6140927          0.3211551          0.5770285
#> 3          0.1403991          0.4979348          0.1529513          0.4589234