Summarize model

stable

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 969472733
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0346   19.4187    0.0000 [ 0.6116; 0.7391 ] 
#>   eta3 ~ eta1      0.4585      0.0900    5.0937    0.0000 [ 0.3368; 0.6917 ] 
#>   eta3 ~ eta2      0.3052      0.0862    3.5415    0.0004 [ 0.0960; 0.4296 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0315   21.0829    0.0000 [ 0.5859; 0.7058 ] 
#>   eta1 =~ y12      0.6493      0.0325   19.9545    0.0000 [ 0.5977; 0.7170 ] 
#>   eta1 =~ y13      0.7613      0.0272   27.9672    0.0000 [ 0.7281; 0.8043 ] 
#>   eta2 =~ y21      0.5165      0.0526    9.8250    0.0000 [ 0.4204; 0.6032 ] 
#>   eta2 =~ y22      0.7554      0.0370   20.3925    0.0000 [ 0.6626; 0.8134 ] 
#>   eta2 =~ y23      0.7997      0.0395   20.2598    0.0000 [ 0.7302; 0.8756 ] 
#>   eta3 =~ y31      0.8223      0.0354   23.2552    0.0000 [ 0.7481; 0.9000 ] 
#>   eta3 =~ y32      0.6581      0.0355   18.5322    0.0000 [ 0.6148; 0.7491 ] 
#>   eta3 =~ y33      0.7474      0.0327   22.8406    0.0000 [ 0.6827; 0.7959 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0134   29.4110    0.0000 [ 0.3636; 0.4114 ] 
#>   eta1 <~ y12      0.3873      0.0172   22.5039    0.0000 [ 0.3581; 0.4231 ] 
#>   eta1 <~ y13      0.4542      0.0177   25.6885    0.0000 [ 0.4248; 0.4859 ] 
#>   eta2 <~ y21      0.3058      0.0278   11.0136    0.0000 [ 0.2559; 0.3537 ] 
#>   eta2 <~ y22      0.4473      0.0217   20.5969    0.0000 [ 0.4089; 0.4852 ] 
#>   eta2 <~ y23      0.4735      0.0229   20.6416    0.0000 [ 0.4312; 0.5187 ] 
#>   eta3 <~ y31      0.4400      0.0155   28.3477    0.0000 [ 0.4096; 0.4613 ] 
#>   eta3 <~ y32      0.3521      0.0181   19.4683    0.0000 [ 0.3332; 0.3912 ] 
#>   eta3 <~ y33      0.3999      0.0172   23.2063    0.0000 [ 0.3644; 0.4253 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0346   19.4187    0.0000 [ 0.6116; 0.7391 ] 
#>   eta3 ~ eta1       0.6634      0.0474   14.0093    0.0000 [ 0.5975; 0.7731 ] 
#>   eta3 ~ eta2       0.3052      0.0862    3.5415    0.0004 [ 0.0960; 0.4296 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0561    3.6534    0.0003 [ 0.0691; 0.2846 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03145065 21.082867  1.142477e-98
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03253793 19.954495  1.369996e-88
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02722277 27.967239 4.068692e-172
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05256535  9.825005  8.786926e-23
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03704242 20.392504  1.949089e-92
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03947038 20.259844  2.908806e-91
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03535884 23.255216 1.259771e-119
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03550955 18.532167  1.136186e-76
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03272353 22.840571 1.813391e-115
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5858862          0.7057580
#> 2          0.5976538          0.7170312
#> 3          0.7280780          0.8043376
#> 4          0.4203857          0.6031658
#> 5          0.6626045          0.8133634
#> 6          0.7301878          0.8756215
#> 7          0.7480647          0.8999861
#> 8          0.6147893          0.7490938
#> 9          0.6826840          0.7959479

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 969472733
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0346   19.4187    0.0000 [ 0.5842; 0.7630 ] 
#>   eta3 ~ eta1      0.4585      0.0900    5.0937    0.0000 [ 0.2021; 0.6676 ] 
#>   eta3 ~ eta2      0.3052      0.0862    3.5415    0.0004 [ 0.0975; 0.5431 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0315   21.0829    0.0000 [ 0.5886; 0.7512 ] 
#>   eta1 =~ y12      0.6493      0.0325   19.9545    0.0000 [ 0.5623; 0.7306 ] 
#>   eta1 =~ y13      0.7613      0.0272   27.9672    0.0000 [ 0.6858; 0.8266 ] 
#>   eta2 =~ y21      0.5165      0.0526    9.8250    0.0000 [ 0.3813; 0.6531 ] 
#>   eta2 =~ y22      0.7554      0.0370   20.3925    0.0000 [ 0.6616; 0.8532 ] 
#>   eta2 =~ y23      0.7997      0.0395   20.2598    0.0000 [ 0.6923; 0.8965 ] 
#>   eta3 =~ y31      0.8223      0.0354   23.2552    0.0000 [ 0.7356; 0.9185 ] 
#>   eta3 =~ y32      0.6581      0.0355   18.5322    0.0000 [ 0.5501; 0.7337 ] 
#>   eta3 =~ y33      0.7474      0.0327   22.8406    0.0000 [ 0.6744; 0.8436 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0134   29.4110    0.0000 [ 0.3655; 0.4350 ] 
#>   eta1 <~ y12      0.3873      0.0172   22.5039    0.0000 [ 0.3417; 0.4307 ] 
#>   eta1 <~ y13      0.4542      0.0177   25.6885    0.0000 [ 0.4058; 0.4972 ] 
#>   eta2 <~ y21      0.3058      0.0278   11.0136    0.0000 [ 0.2356; 0.3792 ] 
#>   eta2 <~ y22      0.4473      0.0217   20.5969    0.0000 [ 0.3936; 0.5059 ] 
#>   eta2 <~ y23      0.4735      0.0229   20.6416    0.0000 [ 0.4124; 0.5310 ] 
#>   eta3 <~ y31      0.4400      0.0155   28.3477    0.0000 [ 0.4023; 0.4826 ] 
#>   eta3 <~ y32      0.3521      0.0181   19.4683    0.0000 [ 0.2966; 0.3901 ] 
#>   eta3 <~ y33      0.3999      0.0172   23.2063    0.0000 [ 0.3614; 0.4505 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0346   19.4187    0.0000 [ 0.5842; 0.7630 ] 
#>   eta3 ~ eta1       0.6634      0.0474   14.0093    0.0000 [ 0.5291; 0.7739 ] 
#>   eta3 ~ eta2       0.3052      0.0862    3.5415    0.0004 [ 0.0975; 0.5431 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0561    3.6534    0.0003 [ 0.0716; 0.3616 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03457154 19.418673 5.365877e-84
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09001386  5.093735 3.510775e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08616492  3.541477 3.978929e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.58418301          0.7629675          0.6056515          0.7414990
#> 2         0.20214155          0.6676424          0.2580391          0.6117449
#> 3         0.09752564          0.5431220          0.1510330          0.4896146
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.58586131          0.7770572         0.61157961          0.7390979
#> 2         0.33283438          0.7010966         0.33680246          0.6916523
#> 3         0.08708989          0.4416822         0.09597682          0.4295653