Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -543359892
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0516   13.0097    0.0000 [ 0.5573; 0.7455 ] 
#>   eta3 ~ eta1      0.4585      0.0743    6.1741    0.0000 [ 0.2784; 0.5522 ] 
#>   eta3 ~ eta2      0.3052      0.0706    4.3238    0.0000 [ 0.1965; 0.4567 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0436   15.1912    0.0000 [ 0.5957; 0.7329 ] 
#>   eta1 =~ y12      0.6493      0.0400   16.2185    0.0000 [ 0.5731; 0.7125 ] 
#>   eta1 =~ y13      0.7613      0.0320   23.8273    0.0000 [ 0.7046; 0.8385 ] 
#>   eta2 =~ y21      0.5165      0.0519    9.9562    0.0000 [ 0.4290; 0.6245 ] 
#>   eta2 =~ y22      0.7554      0.0304   24.8244    0.0000 [ 0.6999; 0.7943 ] 
#>   eta2 =~ y23      0.7997      0.0421   18.9847    0.0000 [ 0.7226; 0.8675 ] 
#>   eta3 =~ y31      0.8223      0.0364   22.6080    0.0000 [ 0.7237; 0.8659 ] 
#>   eta3 =~ y32      0.6581      0.0424   15.5025    0.0000 [ 0.5907; 0.7219 ] 
#>   eta3 =~ y33      0.7474      0.0355   21.0640    0.0000 [ 0.6991; 0.8291 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0215   18.3980    0.0000 [ 0.3563; 0.4289 ] 
#>   eta1 <~ y12      0.3873      0.0214   18.1403    0.0000 [ 0.3491; 0.4296 ] 
#>   eta1 <~ y13      0.4542      0.0218   20.8088    0.0000 [ 0.4129; 0.4832 ] 
#>   eta2 <~ y21      0.3058      0.0275   11.1151    0.0000 [ 0.2613; 0.3611 ] 
#>   eta2 <~ y22      0.4473      0.0222   20.1363    0.0000 [ 0.4161; 0.4930 ] 
#>   eta2 <~ y23      0.4735      0.0195   24.3266    0.0000 [ 0.4360; 0.5102 ] 
#>   eta3 <~ y31      0.4400      0.0213   20.7045    0.0000 [ 0.4009; 0.4684 ] 
#>   eta3 <~ y32      0.3521      0.0191   18.4341    0.0000 [ 0.3127; 0.3852 ] 
#>   eta3 <~ y33      0.3999      0.0162   24.6364    0.0000 [ 0.3804; 0.4394 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0516   13.0097    0.0000 [ 0.5573; 0.7455 ] 
#>   eta3 ~ eta1       0.6634      0.0380   17.4604    0.0000 [ 0.6005; 0.7197 ] 
#>   eta3 ~ eta2       0.3052      0.0706    4.3238    0.0000 [ 0.1965; 0.4567 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0492    4.1640    0.0000 [ 0.1176; 0.3115 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04364834 15.191181  4.045660e-52
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04003308 16.218537  3.729722e-59
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03195272 23.827262 1.742886e-125
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05187285  9.956168  2.370221e-23
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03042921 24.824430 4.885042e-136
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04212152 18.984683  2.283097e-80
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03637105 22.608019 3.613981e-113
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04244934 15.502452  3.339018e-54
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03548345 21.064019  1.701113e-98
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5957014          0.7328741
#> 2          0.5730543          0.7124862
#> 3          0.7046223          0.8384537
#> 4          0.4290192          0.6245005
#> 5          0.6998832          0.7942783
#> 6          0.7225658          0.8674572
#> 7          0.7236521          0.8659327
#> 8          0.5907178          0.7219387
#> 9          0.6991287          0.8290691

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -543359892
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0516   13.0097    0.0000 [ 0.5404; 0.8073 ] 
#>   eta3 ~ eta1      0.4585      0.0743    6.1741    0.0000 [ 0.2862; 0.6703 ] 
#>   eta3 ~ eta2      0.3052      0.0706    4.3238    0.0000 [ 0.1035; 0.4685 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0436   15.1912    0.0000 [ 0.5491; 0.7748 ] 
#>   eta1 =~ y12      0.6493      0.0400   16.2185    0.0000 [ 0.5442; 0.7512 ] 
#>   eta1 =~ y13      0.7613      0.0320   23.8273    0.0000 [ 0.6819; 0.8471 ] 
#>   eta2 =~ y21      0.5165      0.0519    9.9562    0.0000 [ 0.3849; 0.6532 ] 
#>   eta2 =~ y22      0.7554      0.0304   24.8244    0.0000 [ 0.6867; 0.8441 ] 
#>   eta2 =~ y23      0.7997      0.0421   18.9847    0.0000 [ 0.6921; 0.9100 ] 
#>   eta3 =~ y31      0.8223      0.0364   22.6080    0.0000 [ 0.7363; 0.9244 ] 
#>   eta3 =~ y32      0.6581      0.0424   15.5025    0.0000 [ 0.5430; 0.7625 ] 
#>   eta3 =~ y33      0.7474      0.0355   21.0640    0.0000 [ 0.6437; 0.8272 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0215   18.3980    0.0000 [ 0.3396; 0.4507 ] 
#>   eta1 <~ y12      0.3873      0.0214   18.1403    0.0000 [ 0.3313; 0.4417 ] 
#>   eta1 <~ y13      0.4542      0.0218   20.8088    0.0000 [ 0.3995; 0.5124 ] 
#>   eta2 <~ y21      0.3058      0.0275   11.1151    0.0000 [ 0.2336; 0.3759 ] 
#>   eta2 <~ y22      0.4473      0.0222   20.1363    0.0000 [ 0.3914; 0.5063 ] 
#>   eta2 <~ y23      0.4735      0.0195   24.3266    0.0000 [ 0.4198; 0.5204 ] 
#>   eta3 <~ y31      0.4400      0.0213   20.7045    0.0000 [ 0.3920; 0.5019 ] 
#>   eta3 <~ y32      0.3521      0.0191   18.4341    0.0000 [ 0.3023; 0.4011 ] 
#>   eta3 <~ y33      0.3999      0.0162   24.6364    0.0000 [ 0.3542; 0.4382 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0516   13.0097    0.0000 [ 0.5404; 0.8073 ] 
#>   eta3 ~ eta1       0.6634      0.0380   17.4604    0.0000 [ 0.5728; 0.7693 ] 
#>   eta3 ~ eta2       0.3052      0.0706    4.3238    0.0000 [ 0.1035; 0.4685 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0492    4.1640    0.0000 [ 0.0656; 0.3200 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05160269 13.009661 1.078199e-38
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07426346  6.174056 6.655978e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07057404  4.323844 1.533338e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5404242          0.8072842          0.5724689          0.7752395
#> 2          0.2862282          0.6702769          0.3323449          0.6241601
#> 3          0.1035214          0.4684905          0.1473471          0.4246649
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5175512          0.8083990          0.5573069          0.7454975
#> 2          0.2776951          0.5827413          0.2783634          0.5521787
#> 3          0.1922231          0.5146225          0.1965208          0.4566867