Summarize model

stable

Usage

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 397658049
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0468   14.3447    0.0000 [ 0.5717; 0.7616 ] 
#>   eta3 ~ eta1      0.4585      0.0878    5.2214    0.0000 [ 0.2839; 0.5585 ] 
#>   eta3 ~ eta2      0.3052      0.0924    3.3011    0.0010 [ 0.1670; 0.4823 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0440   15.0858    0.0000 [ 0.5978; 0.7535 ] 
#>   eta1 =~ y12      0.6493      0.0354   18.3259    0.0000 [ 0.5833; 0.6981 ] 
#>   eta1 =~ y13      0.7613      0.0411   18.5323    0.0000 [ 0.6834; 0.8194 ] 
#>   eta2 =~ y21      0.5165      0.0422   12.2258    0.0000 [ 0.4580; 0.6130 ] 
#>   eta2 =~ y22      0.7554      0.0291   25.9477    0.0000 [ 0.6724; 0.7898 ] 
#>   eta2 =~ y23      0.7997      0.0367   21.7679    0.0000 [ 0.7347; 0.8605 ] 
#>   eta3 =~ y31      0.8223      0.0333   24.7060    0.0000 [ 0.7548; 0.8805 ] 
#>   eta3 =~ y32      0.6581      0.0453   14.5296    0.0000 [ 0.5527; 0.7363 ] 
#>   eta3 =~ y33      0.7474      0.0401   18.6346    0.0000 [ 0.6766; 0.8217 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0264   14.9987    0.0000 [ 0.3592; 0.4494 ] 
#>   eta1 <~ y12      0.3873      0.0181   21.3752    0.0000 [ 0.3505; 0.4185 ] 
#>   eta1 <~ y13      0.4542      0.0209   21.7098    0.0000 [ 0.4099; 0.4828 ] 
#>   eta2 <~ y21      0.3058      0.0223   13.7439    0.0000 [ 0.2772; 0.3704 ] 
#>   eta2 <~ y22      0.4473      0.0202   22.1906    0.0000 [ 0.4159; 0.4873 ] 
#>   eta2 <~ y23      0.4735      0.0182   25.9943    0.0000 [ 0.4387; 0.4993 ] 
#>   eta3 <~ y31      0.4400      0.0196   22.4588    0.0000 [ 0.4039; 0.4717 ] 
#>   eta3 <~ y32      0.3521      0.0214   16.4759    0.0000 [ 0.3018; 0.3956 ] 
#>   eta3 <~ y33      0.3999      0.0200   20.0095    0.0000 [ 0.3668; 0.4426 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0468   14.3447    0.0000 [ 0.5717; 0.7616 ] 
#>   eta3 ~ eta1       0.6634      0.0365   18.1712    0.0000 [ 0.5847; 0.7041 ] 
#>   eta3 ~ eta2       0.3052      0.0924    3.3011    0.0010 [ 0.1670; 0.4823 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0614    3.3387    0.0008 [ 0.1176; 0.3250 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04395315 15.08583  2.007242e-51
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03542948 18.32593  5.139947e-75
#> 3 eta1 =~ y13  Common factor 0.7613458 0.04108204 18.53233  1.132793e-76
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04224290 12.22584  2.262417e-34
#> 5 eta2 =~ y22  Common factor 0.7553877 0.02911197 25.94767 1.931865e-148
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03673593 21.76789 4.676380e-105
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03328255 24.70596 9.228959e-135
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04529153 14.52962  7.865053e-48
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04010953 18.63458  1.684835e-77
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5977550          0.7534560
#> 2          0.5833276          0.6981269
#> 3          0.6833655          0.8193574
#> 4          0.4579580          0.6130144
#> 5          0.6724143          0.7897722
#> 6          0.7347326          0.8604703
#> 7          0.7547625          0.8804972
#> 8          0.5527013          0.7362800
#> 9          0.6765880          0.8217103

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 397658049
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0468   14.3447    0.0000 [ 0.5468; 0.7888 ] 
#>   eta3 ~ eta1      0.4585      0.0878    5.2214    0.0000 [ 0.2534; 0.7075 ] 
#>   eta3 ~ eta2      0.3052      0.0924    3.3011    0.0010 [ 0.0435; 0.5216 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0440   15.0858    0.0000 [ 0.5407; 0.7680 ] 
#>   eta1 =~ y12      0.6493      0.0354   18.3259    0.0000 [ 0.5626; 0.7459 ] 
#>   eta1 =~ y13      0.7613      0.0411   18.5323    0.0000 [ 0.6560; 0.8685 ] 
#>   eta2 =~ y21      0.5165      0.0422   12.2258    0.0000 [ 0.4002; 0.6187 ] 
#>   eta2 =~ y22      0.7554      0.0291   25.9477    0.0000 [ 0.6852; 0.8357 ] 
#>   eta2 =~ y23      0.7997      0.0367   21.7679    0.0000 [ 0.7198; 0.9098 ] 
#>   eta3 =~ y31      0.8223      0.0333   24.7060    0.0000 [ 0.7405; 0.9126 ] 
#>   eta3 =~ y32      0.6581      0.0453   14.5296    0.0000 [ 0.5416; 0.7759 ] 
#>   eta3 =~ y33      0.7474      0.0401   18.6346    0.0000 [ 0.6366; 0.8440 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0264   14.9987    0.0000 [ 0.3231; 0.4595 ] 
#>   eta1 <~ y12      0.3873      0.0181   21.3752    0.0000 [ 0.3445; 0.4382 ] 
#>   eta1 <~ y13      0.4542      0.0209   21.7098    0.0000 [ 0.4019; 0.5101 ] 
#>   eta2 <~ y21      0.3058      0.0223   13.7439    0.0000 [ 0.2411; 0.3562 ] 
#>   eta2 <~ y22      0.4473      0.0202   22.1906    0.0000 [ 0.3935; 0.4977 ] 
#>   eta2 <~ y23      0.4735      0.0182   25.9943    0.0000 [ 0.4307; 0.5249 ] 
#>   eta3 <~ y31      0.4400      0.0196   22.4588    0.0000 [ 0.3924; 0.4937 ] 
#>   eta3 <~ y32      0.3521      0.0214   16.4759    0.0000 [ 0.2981; 0.4086 ] 
#>   eta3 <~ y33      0.3999      0.0200   20.0095    0.0000 [ 0.3453; 0.4487 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0468   14.3447    0.0000 [ 0.5468; 0.7888 ] 
#>   eta3 ~ eta1       0.6634      0.0365   18.1712    0.0000 [ 0.5755; 0.7643 ] 
#>   eta3 ~ eta2       0.3052      0.0924    3.3011    0.0010 [ 0.0435; 0.5216 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0614    3.3387    0.0008 [ 0.0308; 0.3481 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04680008 14.344706 1.149867e-46
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08781266  5.221420 1.775566e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09243853  3.301125 9.629803e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54677081          0.7887944          0.5758331          0.7597321
#> 2         0.25337746          0.7074950          0.3079081          0.6529643
#> 3         0.04351151          0.5215514          0.1009148          0.4641481
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5654905          0.7772779          0.5717030          0.7616300
#> 2          0.2643837          0.5729385          0.2838706          0.5585133
#> 3          0.1584966          0.4888904          0.1670366          0.4823349