Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1631270560
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0384   17.4735    0.0000 [ 0.5847; 0.7370 ] 
#>   eta3 ~ eta1      0.4585      0.0696    6.5859    0.0000 [ 0.3387; 0.6087 ] 
#>   eta3 ~ eta2      0.3052      0.0748    4.0797    0.0000 [ 0.1731; 0.4251 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0345   19.2126    0.0000 [ 0.6152; 0.7307 ] 
#>   eta1 =~ y12      0.6493      0.0388   16.7144    0.0000 [ 0.5435; 0.6948 ] 
#>   eta1 =~ y13      0.7613      0.0293   25.9494    0.0000 [ 0.6986; 0.8016 ] 
#>   eta2 =~ y21      0.5165      0.0522    9.8989    0.0000 [ 0.4450; 0.6188 ] 
#>   eta2 =~ y22      0.7554      0.0333   22.6810    0.0000 [ 0.6684; 0.7946 ] 
#>   eta2 =~ y23      0.7997      0.0356   22.4783    0.0000 [ 0.7446; 0.8787 ] 
#>   eta3 =~ y31      0.8223      0.0334   24.6392    0.0000 [ 0.7548; 0.8850 ] 
#>   eta3 =~ y32      0.6581      0.0358   18.3801    0.0000 [ 0.5898; 0.7293 ] 
#>   eta3 =~ y33      0.7474      0.0426   17.5600    0.0000 [ 0.6650; 0.8176 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0182   21.7399    0.0000 [ 0.3726; 0.4417 ] 
#>   eta1 <~ y12      0.3873      0.0188   20.6399    0.0000 [ 0.3454; 0.4084 ] 
#>   eta1 <~ y13      0.4542      0.0174   26.0911    0.0000 [ 0.4230; 0.4972 ] 
#>   eta2 <~ y21      0.3058      0.0273   11.2101    0.0000 [ 0.2693; 0.3573 ] 
#>   eta2 <~ y22      0.4473      0.0165   27.0920    0.0000 [ 0.4064; 0.4622 ] 
#>   eta2 <~ y23      0.4735      0.0244   19.3994    0.0000 [ 0.4317; 0.5120 ] 
#>   eta3 <~ y31      0.4400      0.0215   20.4643    0.0000 [ 0.4013; 0.4790 ] 
#>   eta3 <~ y32      0.3521      0.0150   23.4356    0.0000 [ 0.3265; 0.3810 ] 
#>   eta3 <~ y33      0.3999      0.0201   19.9334    0.0000 [ 0.3655; 0.4312 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0384   17.4735    0.0000 [ 0.5847; 0.7370 ] 
#>   eta3 ~ eta1       0.6634      0.0361   18.3874    0.0000 [ 0.6179; 0.7319 ] 
#>   eta3 ~ eta2       0.3052      0.0748    4.0797    0.0000 [ 0.1731; 0.4251 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0501    4.0904    0.0000 [ 0.1150; 0.2937 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03451228 19.212577  2.904823e-82
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03884532 16.714443  1.028766e-62
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02933964 25.949393 1.847230e-148
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05217297  9.898896  4.208946e-23
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03330491 22.680973 6.904563e-114
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03557500 22.478251 6.775644e-112
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03337270 24.639224 4.801662e-134
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03580324 18.380149  1.894469e-75
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04256410 17.559964  4.990647e-69
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6151591          0.7306709
#> 2          0.5434636          0.6948482
#> 3          0.6986142          0.8015924
#> 4          0.4449726          0.6187763
#> 5          0.6683627          0.7946024
#> 6          0.7445849          0.8787050
#> 7          0.7548057          0.8849917
#> 8          0.5898191          0.7292985
#> 9          0.6650461          0.8175525

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1631270560
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0384   17.4735    0.0000 [ 0.5806; 0.7793 ] 
#>   eta3 ~ eta1      0.4585      0.0696    6.5859    0.0000 [ 0.2757; 0.6357 ] 
#>   eta3 ~ eta2      0.3052      0.0748    4.0797    0.0000 [ 0.1093; 0.4961 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0345   19.2126    0.0000 [ 0.5690; 0.7475 ] 
#>   eta1 =~ y12      0.6493      0.0388   16.7144    0.0000 [ 0.5541; 0.7550 ] 
#>   eta1 =~ y13      0.7613      0.0293   25.9494    0.0000 [ 0.6840; 0.8358 ] 
#>   eta2 =~ y21      0.5165      0.0522    9.8989    0.0000 [ 0.3715; 0.6413 ] 
#>   eta2 =~ y22      0.7554      0.0333   22.6810    0.0000 [ 0.6779; 0.8501 ] 
#>   eta2 =~ y23      0.7997      0.0356   22.4783    0.0000 [ 0.7017; 0.8856 ] 
#>   eta3 =~ y31      0.8223      0.0334   24.6392    0.0000 [ 0.7359; 0.9085 ] 
#>   eta3 =~ y32      0.6581      0.0358   18.3801    0.0000 [ 0.5697; 0.7549 ] 
#>   eta3 =~ y33      0.7474      0.0426   17.5600    0.0000 [ 0.6368; 0.8569 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0182   21.7399    0.0000 [ 0.3461; 0.4402 ] 
#>   eta1 <~ y12      0.3873      0.0188   20.6399    0.0000 [ 0.3425; 0.4396 ] 
#>   eta1 <~ y13      0.4542      0.0174   26.0911    0.0000 [ 0.4087; 0.4987 ] 
#>   eta2 <~ y21      0.3058      0.0273   11.2101    0.0000 [ 0.2312; 0.3723 ] 
#>   eta2 <~ y22      0.4473      0.0165   27.0920    0.0000 [ 0.4121; 0.4975 ] 
#>   eta2 <~ y23      0.4735      0.0244   19.3994    0.0000 [ 0.4091; 0.5354 ] 
#>   eta3 <~ y31      0.4400      0.0215   20.4643    0.0000 [ 0.3833; 0.4945 ] 
#>   eta3 <~ y32      0.3521      0.0150   23.4356    0.0000 [ 0.3150; 0.3927 ] 
#>   eta3 <~ y33      0.3999      0.0201   19.9334    0.0000 [ 0.3471; 0.4508 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0384   17.4735    0.0000 [ 0.5806; 0.7793 ] 
#>   eta3 ~ eta1       0.6634      0.0361   18.3874    0.0000 [ 0.5685; 0.7551 ] 
#>   eta3 ~ eta2       0.3052      0.0748    4.0797    0.0000 [ 0.1093; 0.4961 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0501    4.0904    0.0000 [ 0.0766; 0.3356 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03842001 17.473537 2.279164e-68
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.06961980  6.585867 4.522388e-11
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07479793  4.079673 4.509904e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5805848          0.7792714          0.6044432          0.7554130
#> 2          0.2756758          0.6357101          0.3189089          0.5924770
#> 3          0.1093303          0.4961429          0.1557789          0.4496943
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5658598          0.7374739          0.5846891          0.7369892
#> 2          0.3377800          0.6393003          0.3387231          0.6086621
#> 3          0.1688751          0.4546780          0.1731433          0.4251325