Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -558195079
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0410   16.3695    0.0000 [ 0.5927; 0.7366 ] 
#>   eta3 ~ eta1      0.4585      0.0733    6.2558    0.0000 [ 0.3552; 0.6342 ] 
#>   eta3 ~ eta2      0.3052      0.0767    3.9766    0.0001 [ 0.1484; 0.4213 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0403   16.4725    0.0000 [ 0.5884; 0.7330 ] 
#>   eta1 =~ y12      0.6493      0.0407   15.9419    0.0000 [ 0.5676; 0.7241 ] 
#>   eta1 =~ y13      0.7613      0.0286   26.6159    0.0000 [ 0.7065; 0.8167 ] 
#>   eta2 =~ y21      0.5165      0.0545    9.4806    0.0000 [ 0.3943; 0.6170 ] 
#>   eta2 =~ y22      0.7554      0.0273   27.6630    0.0000 [ 0.7006; 0.7899 ] 
#>   eta2 =~ y23      0.7997      0.0322   24.8130    0.0000 [ 0.7314; 0.8451 ] 
#>   eta3 =~ y31      0.8223      0.0374   22.0119    0.0000 [ 0.7499; 0.8849 ] 
#>   eta3 =~ y32      0.6581      0.0384   17.1484    0.0000 [ 0.5964; 0.7162 ] 
#>   eta3 =~ y33      0.7474      0.0360   20.7872    0.0000 [ 0.6838; 0.8042 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0216   18.3453    0.0000 [ 0.3492; 0.4309 ] 
#>   eta1 <~ y12      0.3873      0.0184   21.0415    0.0000 [ 0.3512; 0.4229 ] 
#>   eta1 <~ y13      0.4542      0.0163   27.7905    0.0000 [ 0.4271; 0.4848 ] 
#>   eta2 <~ y21      0.3058      0.0281   10.8850    0.0000 [ 0.2508; 0.3558 ] 
#>   eta2 <~ y22      0.4473      0.0193   23.1887    0.0000 [ 0.4083; 0.4737 ] 
#>   eta2 <~ y23      0.4735      0.0193   24.5675    0.0000 [ 0.4319; 0.5064 ] 
#>   eta3 <~ y31      0.4400      0.0191   22.9954    0.0000 [ 0.4056; 0.4757 ] 
#>   eta3 <~ y32      0.3521      0.0183   19.2614    0.0000 [ 0.3228; 0.3865 ] 
#>   eta3 <~ y33      0.3999      0.0187   21.3924    0.0000 [ 0.3667; 0.4402 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0410   16.3695    0.0000 [ 0.5927; 0.7366 ] 
#>   eta3 ~ eta1       0.6634      0.0391   16.9829    0.0000 [ 0.6073; 0.7496 ] 
#>   eta3 ~ eta2       0.3052      0.0767    3.9766    0.0001 [ 0.1484; 0.4213 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0511    4.0128    0.0001 [ 0.1001; 0.2725 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04025314 16.472501  5.782893e-61
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04072780 15.941885  3.244171e-57
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02860492 26.615900 4.443776e-156
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05447481  9.480617  2.527848e-21
#> 5 eta2 =~ y22  Common factor 0.7553877 0.02730675 27.663034 1.945241e-168
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03222755 24.813045 6.483085e-136
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03735607 22.011883 2.215928e-107
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03837485 17.148441  6.456142e-66
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03595602 20.787174  5.654382e-96
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5884194          0.7329541
#> 2          0.5675513          0.7241018
#> 3          0.7064633          0.8166599
#> 4          0.3943234          0.6169772
#> 5          0.7006250          0.7898886
#> 6          0.7314200          0.8450935
#> 7          0.7499043          0.8849036
#> 8          0.5964364          0.7161620
#> 9          0.6837883          0.8042473

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -558195079
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0410   16.3695    0.0000 [ 0.5662; 0.7783 ] 
#>   eta3 ~ eta1      0.4585      0.0733    6.2558    0.0000 [ 0.2571; 0.6361 ] 
#>   eta3 ~ eta2      0.3052      0.0767    3.9766    0.0001 [ 0.1143; 0.5111 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0403   16.4725    0.0000 [ 0.5658; 0.7739 ] 
#>   eta1 =~ y12      0.6493      0.0407   15.9419    0.0000 [ 0.5398; 0.7504 ] 
#>   eta1 =~ y13      0.7613      0.0286   26.6159    0.0000 [ 0.6788; 0.8268 ] 
#>   eta2 =~ y21      0.5165      0.0545    9.4806    0.0000 [ 0.3761; 0.6578 ] 
#>   eta2 =~ y22      0.7554      0.0273   27.6630    0.0000 [ 0.6864; 0.8276 ] 
#>   eta2 =~ y23      0.7997      0.0322   24.8130    0.0000 [ 0.7169; 0.8836 ] 
#>   eta3 =~ y31      0.8223      0.0374   22.0119    0.0000 [ 0.7239; 0.9171 ] 
#>   eta3 =~ y32      0.6581      0.0384   17.1484    0.0000 [ 0.5629; 0.7614 ] 
#>   eta3 =~ y33      0.7474      0.0360   20.7872    0.0000 [ 0.6573; 0.8432 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0216   18.3453    0.0000 [ 0.3457; 0.4572 ] 
#>   eta1 <~ y12      0.3873      0.0184   21.0415    0.0000 [ 0.3393; 0.4345 ] 
#>   eta1 <~ y13      0.4542      0.0163   27.7905    0.0000 [ 0.4089; 0.4934 ] 
#>   eta2 <~ y21      0.3058      0.0281   10.8850    0.0000 [ 0.2336; 0.3789 ] 
#>   eta2 <~ y22      0.4473      0.0193   23.1887    0.0000 [ 0.3978; 0.4975 ] 
#>   eta2 <~ y23      0.4735      0.0193   24.5675    0.0000 [ 0.4235; 0.5232 ] 
#>   eta3 <~ y31      0.4400      0.0191   22.9954    0.0000 [ 0.3882; 0.4872 ] 
#>   eta3 <~ y32      0.3521      0.0183   19.2614    0.0000 [ 0.3061; 0.4006 ] 
#>   eta3 <~ y33      0.3999      0.0187   21.3924    0.0000 [ 0.3519; 0.4486 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0410   16.3695    0.0000 [ 0.5662; 0.7783 ] 
#>   eta3 ~ eta1       0.6634      0.0391   16.9829    0.0000 [ 0.5564; 0.7584 ] 
#>   eta3 ~ eta2       0.3052      0.0767    3.9766    0.0001 [ 0.1143; 0.5111 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0511    4.0128    0.0001 [ 0.0788; 0.3428 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04101114 16.369537 3.156076e-60
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07329314  6.255794 3.954995e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07673741  3.976563 6.991861e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5661866          0.7782731          0.5916540          0.7528056
#> 2          0.2571046          0.6361353          0.3026188          0.5906211
#> 3          0.1142520          0.5110946          0.1619050          0.4634415
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5914787          0.7438465          0.5926634          0.7366162
#> 2          0.3522830          0.6343465          0.3551566          0.6342166
#> 3          0.1437393          0.4306219          0.1483531          0.4212881