Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -408141552
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0422   15.9216    0.0000 [ 0.5925; 0.7620 ] 
#>   eta3 ~ eta1      0.4585      0.1025    4.4721    0.0000 [ 0.2990; 0.6269 ] 
#>   eta3 ~ eta2      0.3052      0.1038    2.9406    0.0033 [ 0.1249; 0.4407 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0414   15.9984    0.0000 [ 0.5945; 0.7401 ] 
#>   eta1 =~ y12      0.6493      0.0400   16.2272    0.0000 [ 0.5932; 0.7153 ] 
#>   eta1 =~ y13      0.7613      0.0295   25.7889    0.0000 [ 0.7093; 0.8022 ] 
#>   eta2 =~ y21      0.5165      0.0491   10.5236    0.0000 [ 0.4388; 0.6396 ] 
#>   eta2 =~ y22      0.7554      0.0392   19.2808    0.0000 [ 0.6714; 0.8179 ] 
#>   eta2 =~ y23      0.7997      0.0386   20.7207    0.0000 [ 0.7454; 0.8912 ] 
#>   eta3 =~ y31      0.8223      0.0309   26.5811    0.0000 [ 0.7590; 0.8872 ] 
#>   eta3 =~ y32      0.6581      0.0377   17.4536    0.0000 [ 0.5926; 0.7236 ] 
#>   eta3 =~ y33      0.7474      0.0353   21.1982    0.0000 [ 0.6858; 0.8116 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0197   20.0641    0.0000 [ 0.3577; 0.4299 ] 
#>   eta1 <~ y12      0.3873      0.0186   20.8406    0.0000 [ 0.3529; 0.4185 ] 
#>   eta1 <~ y13      0.4542      0.0205   22.1736    0.0000 [ 0.4212; 0.4937 ] 
#>   eta2 <~ y21      0.3058      0.0259   11.7880    0.0000 [ 0.2689; 0.3490 ] 
#>   eta2 <~ y22      0.4473      0.0197   22.7178    0.0000 [ 0.4034; 0.4711 ] 
#>   eta2 <~ y23      0.4735      0.0238   19.9144    0.0000 [ 0.4369; 0.5156 ] 
#>   eta3 <~ y31      0.4400      0.0160   27.5066    0.0000 [ 0.4054; 0.4735 ] 
#>   eta3 <~ y32      0.3521      0.0171   20.5738    0.0000 [ 0.3257; 0.3831 ] 
#>   eta3 <~ y33      0.3999      0.0179   22.3333    0.0000 [ 0.3755; 0.4337 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0422   15.9216    0.0000 [ 0.5925; 0.7620 ] 
#>   eta3 ~ eta1       0.6634      0.0424   15.6526    0.0000 [ 0.5836; 0.7325 ] 
#>   eta3 ~ eta2       0.3052      0.1038    2.9406    0.0033 [ 0.1249; 0.4407 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0719    2.8501    0.0044 [ 0.0835; 0.3095 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04144597 15.99842  1.310694e-57
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04001163 16.22723  3.237335e-59
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02952221 25.78892 1.180592e-146
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04907604 10.52356  6.727975e-26
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03917826 19.28079  7.788211e-83
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03859244 20.72073  2.252333e-95
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03093464 26.58112 1.122250e-155
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03770385 17.45363  3.230588e-68
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03525879 21.19823 9.915010e-100
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5944570          0.7400891
#> 2          0.5932339          0.7152945
#> 3          0.7093134          0.8022016
#> 4          0.4388284          0.6396163
#> 5          0.6713893          0.8179306
#> 6          0.7453596          0.8911541
#> 7          0.7589895          0.8872316
#> 8          0.5925686          0.7235804
#> 9          0.6857709          0.8115770

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -408141552
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0422   15.9216    0.0000 [ 0.5653; 0.7833 ] 
#>   eta3 ~ eta1      0.4585      0.1025    4.4721    0.0000 [ 0.1969; 0.7271 ] 
#>   eta3 ~ eta2      0.3052      0.1038    2.9406    0.0033 [ 0.0242; 0.5608 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0414   15.9984    0.0000 [ 0.5544; 0.7687 ] 
#>   eta1 =~ y12      0.6493      0.0400   16.2272    0.0000 [ 0.5431; 0.7500 ] 
#>   eta1 =~ y13      0.7613      0.0295   25.7889    0.0000 [ 0.6881; 0.8407 ] 
#>   eta2 =~ y21      0.5165      0.0491   10.5236    0.0000 [ 0.3800; 0.6338 ] 
#>   eta2 =~ y22      0.7554      0.0392   19.2808    0.0000 [ 0.6576; 0.8602 ] 
#>   eta2 =~ y23      0.7997      0.0386   20.7207    0.0000 [ 0.6998; 0.8994 ] 
#>   eta3 =~ y31      0.8223      0.0309   26.5811    0.0000 [ 0.7499; 0.9099 ] 
#>   eta3 =~ y32      0.6581      0.0377   17.4536    0.0000 [ 0.5600; 0.7550 ] 
#>   eta3 =~ y33      0.7474      0.0353   21.1982    0.0000 [ 0.6496; 0.8320 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0197   20.0641    0.0000 [ 0.3443; 0.4462 ] 
#>   eta1 <~ y12      0.3873      0.0186   20.8406    0.0000 [ 0.3382; 0.4344 ] 
#>   eta1 <~ y13      0.4542      0.0205   22.1736    0.0000 [ 0.4033; 0.5092 ] 
#>   eta2 <~ y21      0.3058      0.0259   11.7880    0.0000 [ 0.2346; 0.3687 ] 
#>   eta2 <~ y22      0.4473      0.0197   22.7178    0.0000 [ 0.4004; 0.5022 ] 
#>   eta2 <~ y23      0.4735      0.0238   19.9144    0.0000 [ 0.4138; 0.5368 ] 
#>   eta3 <~ y31      0.4400      0.0160   27.5066    0.0000 [ 0.4026; 0.4853 ] 
#>   eta3 <~ y32      0.3521      0.0171   20.5738    0.0000 [ 0.3076; 0.3962 ] 
#>   eta3 <~ y33      0.3999      0.0179   22.3333    0.0000 [ 0.3500; 0.4426 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0422   15.9216    0.0000 [ 0.5653; 0.7833 ] 
#>   eta3 ~ eta1       0.6634      0.0424   15.6526    0.0000 [ 0.5494; 0.7686 ] 
#>   eta3 ~ eta2       0.3052      0.1038    2.9406    0.0033 [ 0.0242; 0.5608 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0719    2.8501    0.0044 [ 0.0111; 0.3828 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04216484 15.921642 4.484489e-57
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.10252656  4.472078 7.746322e-06
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.10377030  2.940640 3.275346e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56526781          0.7833205         0.59145168          0.7571367
#> 2         0.19693878          0.7271483         0.26060658          0.6634805
#> 3         0.02417961          0.5608210         0.08861975          0.4963809
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5811742          0.7794115          0.5925457          0.7619708
#> 2          0.2882087          0.6517622          0.2989597          0.6269034
#> 3          0.1231929          0.4439661          0.1249237          0.4406988