Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -363085665
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0506   13.2592    0.0000 [ 0.5945; 0.7653 ] 
#>   eta3 ~ eta1      0.4585      0.0798    5.7483    0.0000 [ 0.3000; 0.5985 ] 
#>   eta3 ~ eta2      0.3052      0.0803    3.7989    0.0001 [ 0.1507; 0.4500 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0343   19.3111    0.0000 [ 0.5961; 0.7239 ] 
#>   eta1 =~ y12      0.6493      0.0324   20.0299    0.0000 [ 0.5886; 0.6993 ] 
#>   eta1 =~ y13      0.7613      0.0299   25.4912    0.0000 [ 0.7157; 0.8239 ] 
#>   eta2 =~ y21      0.5165      0.0519    9.9603    0.0000 [ 0.4298; 0.6163 ] 
#>   eta2 =~ y22      0.7554      0.0408   18.4956    0.0000 [ 0.6936; 0.8152 ] 
#>   eta2 =~ y23      0.7997      0.0448   17.8614    0.0000 [ 0.7385; 0.8794 ] 
#>   eta3 =~ y31      0.8223      0.0321   25.6031    0.0000 [ 0.7699; 0.8732 ] 
#>   eta3 =~ y32      0.6581      0.0416   15.8172    0.0000 [ 0.5760; 0.7205 ] 
#>   eta3 =~ y33      0.7474      0.0400   18.6691    0.0000 [ 0.6689; 0.8196 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0190   20.8527    0.0000 [ 0.3579; 0.4280 ] 
#>   eta1 <~ y12      0.3873      0.0170   22.8374    0.0000 [ 0.3524; 0.4117 ] 
#>   eta1 <~ y13      0.4542      0.0174   26.0891    0.0000 [ 0.4289; 0.4928 ] 
#>   eta2 <~ y21      0.3058      0.0320    9.5496    0.0000 [ 0.2480; 0.3607 ] 
#>   eta2 <~ y22      0.4473      0.0232   19.2963    0.0000 [ 0.4098; 0.4880 ] 
#>   eta2 <~ y23      0.4735      0.0162   29.2616    0.0000 [ 0.4359; 0.5029 ] 
#>   eta3 <~ y31      0.4400      0.0181   24.3568    0.0000 [ 0.4094; 0.4756 ] 
#>   eta3 <~ y32      0.3521      0.0195   18.0427    0.0000 [ 0.3047; 0.3831 ] 
#>   eta3 <~ y33      0.3999      0.0209   19.1304    0.0000 [ 0.3700; 0.4382 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0506   13.2592    0.0000 [ 0.5945; 0.7653 ] 
#>   eta3 ~ eta1       0.6634      0.0435   15.2405    0.0000 [ 0.5836; 0.7406 ] 
#>   eta3 ~ eta2       0.3052      0.0803    3.7989    0.0001 [ 0.1507; 0.4500 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0549    3.7290    0.0002 [ 0.1046; 0.2957 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03433627 19.311060  4.335808e-83
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03241548 20.029875  3.024173e-89
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02986695 25.491248 2.464904e-143
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05185119  9.960328  2.273115e-23
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04084142 18.495629  2.239144e-76
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04477049 17.861400  2.356547e-71
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03211635 25.603080 1.409711e-144
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04160468 15.817184  2.368517e-56
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04003532 18.669118  8.829885e-78
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5961014          0.7239419
#> 2          0.5886340          0.6993063
#> 3          0.7157143          0.8239162
#> 4          0.4297954          0.6162529
#> 5          0.6936141          0.8152142
#> 6          0.7385011          0.8793704
#> 7          0.7699229          0.8731853
#> 8          0.5760157          0.7205500
#> 9          0.6689421          0.8195729

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -363085665
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0506   13.2592    0.0000 [ 0.5388; 0.8006 ] 
#>   eta3 ~ eta1      0.4585      0.0798    5.7483    0.0000 [ 0.2594; 0.6719 ] 
#>   eta3 ~ eta2      0.3052      0.0803    3.7989    0.0001 [ 0.0926; 0.5080 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0343   19.3111    0.0000 [ 0.5830; 0.7606 ] 
#>   eta1 =~ y12      0.6493      0.0324   20.0299    0.0000 [ 0.5705; 0.7381 ] 
#>   eta1 =~ y13      0.7613      0.0299   25.4912    0.0000 [ 0.6787; 0.8332 ] 
#>   eta2 =~ y21      0.5165      0.0519    9.9603    0.0000 [ 0.3921; 0.6602 ] 
#>   eta2 =~ y22      0.7554      0.0408   18.4956    0.0000 [ 0.6496; 0.8608 ] 
#>   eta2 =~ y23      0.7997      0.0448   17.8614    0.0000 [ 0.6790; 0.9105 ] 
#>   eta3 =~ y31      0.8223      0.0321   25.6031    0.0000 [ 0.7396; 0.9056 ] 
#>   eta3 =~ y32      0.6581      0.0416   15.8172    0.0000 [ 0.5552; 0.7703 ] 
#>   eta3 =~ y33      0.7474      0.0400   18.6691    0.0000 [ 0.6421; 0.8491 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0190   20.8527    0.0000 [ 0.3497; 0.4478 ] 
#>   eta1 <~ y12      0.3873      0.0170   22.8374    0.0000 [ 0.3445; 0.4322 ] 
#>   eta1 <~ y13      0.4542      0.0174   26.0891    0.0000 [ 0.4035; 0.4935 ] 
#>   eta2 <~ y21      0.3058      0.0320    9.5496    0.0000 [ 0.2282; 0.3938 ] 
#>   eta2 <~ y22      0.4473      0.0232   19.2963    0.0000 [ 0.3867; 0.5066 ] 
#>   eta2 <~ y23      0.4735      0.0162   29.2616    0.0000 [ 0.4286; 0.5122 ] 
#>   eta3 <~ y31      0.4400      0.0181   24.3568    0.0000 [ 0.3927; 0.4861 ] 
#>   eta3 <~ y32      0.3521      0.0195   18.0427    0.0000 [ 0.3038; 0.4047 ] 
#>   eta3 <~ y33      0.3999      0.0209   19.1304    0.0000 [ 0.3443; 0.4524 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0506   13.2592    0.0000 [ 0.5388; 0.8006 ] 
#>   eta3 ~ eta1       0.6634      0.0435   15.2405    0.0000 [ 0.5546; 0.7797 ] 
#>   eta3 ~ eta2       0.3052      0.0803    3.7989    0.0001 [ 0.0926; 0.5080 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0549    3.7290    0.0002 [ 0.0595; 0.3436 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05063146 13.259215 3.990501e-40
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07976325  5.748346 9.012063e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08032550  3.798932 1.453209e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.53875006          0.8005874          0.5701916          0.7691458
#> 2         0.25943753          0.6719280          0.3089696          0.6223960
#> 3         0.09255202          0.5079502          0.1424332          0.4580690
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5894108          0.7763237          0.5945062          0.7652862
#> 2          0.2605909          0.6116423          0.3000221          0.5984511
#> 3          0.1455275          0.4806176          0.1507089          0.4499737