Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 761970059
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0508   13.2258    0.0000 [ 0.5824; 0.7789 ] 
#>   eta3 ~ eta1      0.4585      0.0790    5.8023    0.0000 [ 0.3319; 0.5751 ] 
#>   eta3 ~ eta2      0.3052      0.0864    3.5307    0.0004 [ 0.1408; 0.4715 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0342   19.4073    0.0000 [ 0.5911; 0.7221 ] 
#>   eta1 =~ y12      0.6493      0.0386   16.8143    0.0000 [ 0.5666; 0.7144 ] 
#>   eta1 =~ y13      0.7613      0.0263   28.9692    0.0000 [ 0.7170; 0.8063 ] 
#>   eta2 =~ y21      0.5165      0.0506   10.2087    0.0000 [ 0.4157; 0.5857 ] 
#>   eta2 =~ y22      0.7554      0.0410   18.4115    0.0000 [ 0.6597; 0.8100 ] 
#>   eta2 =~ y23      0.7997      0.0425   18.8351    0.0000 [ 0.7267; 0.8690 ] 
#>   eta3 =~ y31      0.8223      0.0335   24.5540    0.0000 [ 0.7644; 0.8623 ] 
#>   eta3 =~ y32      0.6581      0.0431   15.2774    0.0000 [ 0.5833; 0.7234 ] 
#>   eta3 =~ y33      0.7474      0.0405   18.4512    0.0000 [ 0.6839; 0.8207 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0211   18.7219    0.0000 [ 0.3626; 0.4285 ] 
#>   eta1 <~ y12      0.3873      0.0180   21.5254    0.0000 [ 0.3594; 0.4183 ] 
#>   eta1 <~ y13      0.4542      0.0148   30.7026    0.0000 [ 0.4262; 0.4785 ] 
#>   eta2 <~ y21      0.3058      0.0290   10.5557    0.0000 [ 0.2583; 0.3476 ] 
#>   eta2 <~ y22      0.4473      0.0233   19.2256    0.0000 [ 0.4121; 0.4808 ] 
#>   eta2 <~ y23      0.4735      0.0231   20.5376    0.0000 [ 0.4455; 0.4985 ] 
#>   eta3 <~ y31      0.4400      0.0206   21.3867    0.0000 [ 0.4071; 0.4756 ] 
#>   eta3 <~ y32      0.3521      0.0173   20.3238    0.0000 [ 0.3293; 0.3923 ] 
#>   eta3 <~ y33      0.3999      0.0212   18.8299    0.0000 [ 0.3673; 0.4400 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0508   13.2258    0.0000 [ 0.5824; 0.7789 ] 
#>   eta3 ~ eta1       0.6634      0.0413   16.0765    0.0000 [ 0.5943; 0.7591 ] 
#>   eta3 ~ eta2       0.3052      0.0864    3.5307    0.0004 [ 0.1408; 0.4715 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0588    3.4861    0.0005 [ 0.1008; 0.3159 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03416599 19.40731  6.694438e-84
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03861469 16.81427  1.918387e-63
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02628123 28.96919 1.608888e-184
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05058981 10.20867  1.813293e-24
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04102801 18.41151  1.062168e-75
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04245611 18.83507  3.896153e-79
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03348859 24.55395 3.924239e-133
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04307471 15.27738  1.081986e-52
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04050818 18.45119  5.101027e-76
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5910767          0.7220704
#> 2          0.5666280          0.7143571
#> 3          0.7170233          0.8063422
#> 4          0.4157146          0.5856617
#> 5          0.6597079          0.8099965
#> 6          0.7266909          0.8689951
#> 7          0.7644451          0.8622602
#> 8          0.5833285          0.7234062
#> 9          0.6838650          0.8206847

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 761970059
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0508   13.2258    0.0000 [ 0.5315; 0.7940 ] 
#>   eta3 ~ eta1      0.4585      0.0790    5.8023    0.0000 [ 0.2572; 0.6658 ] 
#>   eta3 ~ eta2      0.3052      0.0864    3.5307    0.0004 [ 0.0827; 0.5297 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0342   19.4073    0.0000 [ 0.5784; 0.7551 ] 
#>   eta1 =~ y12      0.6493      0.0386   16.8143    0.0000 [ 0.5470; 0.7467 ] 
#>   eta1 =~ y13      0.7613      0.0263   28.9692    0.0000 [ 0.6919; 0.8278 ] 
#>   eta2 =~ y21      0.5165      0.0506   10.2087    0.0000 [ 0.3831; 0.6447 ] 
#>   eta2 =~ y22      0.7554      0.0410   18.4115    0.0000 [ 0.6570; 0.8692 ] 
#>   eta2 =~ y23      0.7997      0.0425   18.8351    0.0000 [ 0.6965; 0.9161 ] 
#>   eta3 =~ y31      0.8223      0.0335   24.5540    0.0000 [ 0.7418; 0.9150 ] 
#>   eta3 =~ y32      0.6581      0.0431   15.2774    0.0000 [ 0.5491; 0.7718 ] 
#>   eta3 =~ y33      0.7474      0.0405   18.4512    0.0000 [ 0.6433; 0.8528 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0211   18.7219    0.0000 [ 0.3433; 0.4525 ] 
#>   eta1 <~ y12      0.3873      0.0180   21.5254    0.0000 [ 0.3398; 0.4328 ] 
#>   eta1 <~ y13      0.4542      0.0148   30.7026    0.0000 [ 0.4153; 0.4918 ] 
#>   eta2 <~ y21      0.3058      0.0290   10.5557    0.0000 [ 0.2269; 0.3768 ] 
#>   eta2 <~ y22      0.4473      0.0233   19.2256    0.0000 [ 0.3881; 0.5084 ] 
#>   eta2 <~ y23      0.4735      0.0231   20.5376    0.0000 [ 0.4140; 0.5332 ] 
#>   eta3 <~ y31      0.4400      0.0206   21.3867    0.0000 [ 0.3873; 0.4937 ] 
#>   eta3 <~ y32      0.3521      0.0173   20.3238    0.0000 [ 0.3068; 0.3964 ] 
#>   eta3 <~ y33      0.3999      0.0212   18.8299    0.0000 [ 0.3430; 0.4528 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0508   13.2258    0.0000 [ 0.5315; 0.7940 ] 
#>   eta3 ~ eta1       0.6634      0.0413   16.0765    0.0000 [ 0.5588; 0.7722 ] 
#>   eta3 ~ eta2       0.3052      0.0864    3.5307    0.0004 [ 0.0827; 0.5297 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0588    3.4861    0.0005 [ 0.0521; 0.3560 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05075937 13.225803 6.226893e-40
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07902107  5.802336 6.539743e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08642853  3.530676 4.144995e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.53147233          0.7939711          0.5629933          0.7624502
#> 2         0.25716257          0.6658149          0.3062337          0.6167438
#> 3         0.08271344          0.5296730          0.1363845          0.4760019
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5773390          0.7955691          0.5824457          0.7788577
#> 2          0.2728413          0.5897162          0.3319174          0.5750904
#> 3          0.1307555          0.5218563          0.1408478          0.4714577