Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 455560321
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0465   14.4363    0.0000 [ 0.6135; 0.7947 ] 
#>   eta3 ~ eta1      0.4585      0.0799    5.7410    0.0000 [ 0.3093; 0.5676 ] 
#>   eta3 ~ eta2      0.3052      0.0854    3.5745    0.0004 [ 0.1756; 0.4728 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0415   15.9686    0.0000 [ 0.5899; 0.7255 ] 
#>   eta1 =~ y12      0.6493      0.0380   17.0680    0.0000 [ 0.5909; 0.7153 ] 
#>   eta1 =~ y13      0.7613      0.0305   24.9730    0.0000 [ 0.6926; 0.8063 ] 
#>   eta2 =~ y21      0.5165      0.0633    8.1621    0.0000 [ 0.4069; 0.6331 ] 
#>   eta2 =~ y22      0.7554      0.0341   22.1410    0.0000 [ 0.7040; 0.8223 ] 
#>   eta2 =~ y23      0.7997      0.0331   24.1242    0.0000 [ 0.7477; 0.8522 ] 
#>   eta3 =~ y31      0.8223      0.0251   32.7938    0.0000 [ 0.7845; 0.8664 ] 
#>   eta3 =~ y32      0.6581      0.0376   17.5074    0.0000 [ 0.5751; 0.7209 ] 
#>   eta3 =~ y33      0.7474      0.0410   18.2090    0.0000 [ 0.6721; 0.8179 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0217   18.2310    0.0000 [ 0.3391; 0.4308 ] 
#>   eta1 <~ y12      0.3873      0.0197   19.6412    0.0000 [ 0.3580; 0.4336 ] 
#>   eta1 <~ y13      0.4542      0.0185   24.5251    0.0000 [ 0.4172; 0.4844 ] 
#>   eta2 <~ y21      0.3058      0.0321    9.5254    0.0000 [ 0.2429; 0.3631 ] 
#>   eta2 <~ y22      0.4473      0.0227   19.7100    0.0000 [ 0.3986; 0.4810 ] 
#>   eta2 <~ y23      0.4735      0.0207   22.8500    0.0000 [ 0.4353; 0.5146 ] 
#>   eta3 <~ y31      0.4400      0.0166   26.4844    0.0000 [ 0.4143; 0.4757 ] 
#>   eta3 <~ y32      0.3521      0.0194   18.1268    0.0000 [ 0.3130; 0.3868 ] 
#>   eta3 <~ y33      0.3999      0.0182   21.9708    0.0000 [ 0.3752; 0.4324 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0465   14.4363    0.0000 [ 0.6135; 0.7947 ] 
#>   eta3 ~ eta1       0.6634      0.0394   16.8543    0.0000 [ 0.5947; 0.7169 ] 
#>   eta3 ~ eta2       0.3052      0.0854    3.5745    0.0004 [ 0.1756; 0.4728 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0596    3.4358    0.0006 [ 0.1226; 0.3478 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04152335 15.968601  2.114765e-57
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03804063 17.068014  2.567888e-65
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03048672 24.973028 1.200696e-137
#> 4 eta2 =~ y21  Common factor 0.5164548 0.06327438  8.162147  3.291211e-16
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03411719 22.140973 1.274567e-108
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03314782 24.124173 1.394179e-128
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02507416 32.793811 7.214678e-236
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03758809 17.507377  1.258613e-68
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04104700 18.208985  4.379927e-74
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5898962          0.7255030
#> 2          0.5908942          0.7152926
#> 3          0.6925759          0.8062977
#> 4          0.4069181          0.6330732
#> 5          0.7040314          0.8223142
#> 6          0.7476948          0.8521757
#> 7          0.7844927          0.8664102
#> 8          0.5750583          0.7209420
#> 9          0.6720552          0.8178531

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 455560321
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0465   14.4363    0.0000 [ 0.5371; 0.7776 ] 
#>   eta3 ~ eta1      0.4585      0.0799    5.7410    0.0000 [ 0.2538; 0.6668 ] 
#>   eta3 ~ eta2      0.3052      0.0854    3.5745    0.0004 [ 0.0802; 0.5217 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0415   15.9686    0.0000 [ 0.5518; 0.7666 ] 
#>   eta1 =~ y12      0.6493      0.0380   17.0680    0.0000 [ 0.5408; 0.7376 ] 
#>   eta1 =~ y13      0.7613      0.0305   24.9730    0.0000 [ 0.6835; 0.8411 ] 
#>   eta2 =~ y21      0.5165      0.0633    8.1621    0.0000 [ 0.3458; 0.6730 ] 
#>   eta2 =~ y22      0.7554      0.0341   22.1410    0.0000 [ 0.6649; 0.8414 ] 
#>   eta2 =~ y23      0.7997      0.0331   24.1242    0.0000 [ 0.7147; 0.8861 ] 
#>   eta3 =~ y31      0.8223      0.0251   32.7938    0.0000 [ 0.7578; 0.8875 ] 
#>   eta3 =~ y32      0.6581      0.0376   17.5074    0.0000 [ 0.5616; 0.7560 ] 
#>   eta3 =~ y33      0.7474      0.0410   18.2090    0.0000 [ 0.6443; 0.8566 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0217   18.2310    0.0000 [ 0.3410; 0.4532 ] 
#>   eta1 <~ y12      0.3873      0.0197   19.6412    0.0000 [ 0.3342; 0.4361 ] 
#>   eta1 <~ y13      0.4542      0.0185   24.5251    0.0000 [ 0.4111; 0.5068 ] 
#>   eta2 <~ y21      0.3058      0.0321    9.5254    0.0000 [ 0.2213; 0.3873 ] 
#>   eta2 <~ y22      0.4473      0.0227   19.7100    0.0000 [ 0.3901; 0.5075 ] 
#>   eta2 <~ y23      0.4735      0.0207   22.8500    0.0000 [ 0.4234; 0.5306 ] 
#>   eta3 <~ y31      0.4400      0.0166   26.4844    0.0000 [ 0.3961; 0.4820 ] 
#>   eta3 <~ y32      0.3521      0.0194   18.1268    0.0000 [ 0.3015; 0.4019 ] 
#>   eta3 <~ y33      0.3999      0.0182   21.9708    0.0000 [ 0.3537; 0.4478 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0465   14.4363    0.0000 [ 0.5371; 0.7776 ] 
#>   eta3 ~ eta1       0.6634      0.0394   16.8543    0.0000 [ 0.5569; 0.7605 ] 
#>   eta3 ~ eta2       0.3052      0.0854    3.5745    0.0004 [ 0.0802; 0.5217 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0596    3.4358    0.0006 [ 0.0442; 0.3525 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04650323 14.436274 3.059397e-47
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07986600  5.740950 9.414670e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08536933  3.574482 3.509218e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.53707872          0.7775672          0.5659567          0.7486892
#> 2         0.25382424          0.6668461          0.3034201          0.6172503
#> 3         0.08022227          0.5217042          0.1332356          0.4686909
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.6071166          0.7984506          0.6135173          0.7947150
#> 2          0.2793208          0.6183301          0.3093456          0.5676434
#> 3          0.1124882          0.4993920          0.1756333          0.4727874