Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 25020659
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0400   16.8010    0.0000 [ 0.6197; 0.7415 ] 
#>   eta3 ~ eta1      0.4585      0.0820    5.5902    0.0000 [ 0.2786; 0.5884 ] 
#>   eta3 ~ eta2      0.3052      0.0987    3.0930    0.0020 [ 0.1333; 0.5323 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0404   16.4067    0.0000 [ 0.6133; 0.7352 ] 
#>   eta1 =~ y12      0.6493      0.0403   16.0963    0.0000 [ 0.5583; 0.7111 ] 
#>   eta1 =~ y13      0.7613      0.0284   26.7802    0.0000 [ 0.7226; 0.8239 ] 
#>   eta2 =~ y21      0.5165      0.0463   11.1657    0.0000 [ 0.4404; 0.6136 ] 
#>   eta2 =~ y22      0.7554      0.0280   26.9341    0.0000 [ 0.7098; 0.8079 ] 
#>   eta2 =~ y23      0.7997      0.0343   23.3006    0.0000 [ 0.7468; 0.8680 ] 
#>   eta3 =~ y31      0.8223      0.0310   26.4864    0.0000 [ 0.7765; 0.8856 ] 
#>   eta3 =~ y32      0.6581      0.0407   16.1844    0.0000 [ 0.5742; 0.7338 ] 
#>   eta3 =~ y33      0.7474      0.0406   18.4243    0.0000 [ 0.6519; 0.8046 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0203   19.5069    0.0000 [ 0.3612; 0.4398 ] 
#>   eta1 <~ y12      0.3873      0.0211   18.3455    0.0000 [ 0.3349; 0.4196 ] 
#>   eta1 <~ y13      0.4542      0.0184   24.7107    0.0000 [ 0.4154; 0.4825 ] 
#>   eta2 <~ y21      0.3058      0.0237   12.8888    0.0000 [ 0.2684; 0.3438 ] 
#>   eta2 <~ y22      0.4473      0.0180   24.8020    0.0000 [ 0.4093; 0.4759 ] 
#>   eta2 <~ y23      0.4735      0.0155   30.4737    0.0000 [ 0.4510; 0.5097 ] 
#>   eta3 <~ y31      0.4400      0.0172   25.5605    0.0000 [ 0.4131; 0.4696 ] 
#>   eta3 <~ y32      0.3521      0.0194   18.1933    0.0000 [ 0.3246; 0.3892 ] 
#>   eta3 <~ y33      0.3999      0.0193   20.7356    0.0000 [ 0.3606; 0.4274 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0400   16.8010    0.0000 [ 0.6197; 0.7415 ] 
#>   eta3 ~ eta1       0.6634      0.0350   18.9348    0.0000 [ 0.6063; 0.7393 ] 
#>   eta3 ~ eta2       0.3052      0.0987    3.0930    0.0020 [ 0.1333; 0.5323 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0633    3.2342    0.0012 [ 0.0980; 0.3376 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04041447 16.40674  1.711425e-60
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04033714 16.09628  2.709142e-58
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02842943 26.78020 5.495632e-158
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04625371 11.16569  6.002039e-29
#> 5 eta2 =~ y22  Common factor 0.7553877 0.02804573 26.93415 8.749145e-160
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03431941 23.30063 4.368995e-120
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03104528 26.48639 1.390836e-154
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04066080 16.18436  6.502720e-59
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04056736 18.42427  8.391268e-76
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6132704          0.7351645
#> 2          0.5582818          0.7111010
#> 3          0.7226318          0.8239484
#> 4          0.4404402          0.6136113
#> 5          0.7098375          0.8079070
#> 6          0.7467972          0.8679962
#> 7          0.7765484          0.8856227
#> 8          0.5741947          0.7338049
#> 9          0.6518956          0.8045852

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 25020659
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0400   16.8010    0.0000 [ 0.5602; 0.7668 ] 
#>   eta3 ~ eta1      0.4585      0.0820    5.5902    0.0000 [ 0.2492; 0.6734 ] 
#>   eta3 ~ eta2      0.3052      0.0987    3.0930    0.0020 [ 0.0443; 0.5545 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0404   16.4067    0.0000 [ 0.5454; 0.7544 ] 
#>   eta1 =~ y12      0.6493      0.0403   16.0963    0.0000 [ 0.5508; 0.7594 ] 
#>   eta1 =~ y13      0.7613      0.0284   26.7802    0.0000 [ 0.6841; 0.8311 ] 
#>   eta2 =~ y21      0.5165      0.0463   11.1657    0.0000 [ 0.3951; 0.6343 ] 
#>   eta2 =~ y22      0.7554      0.0280   26.9341    0.0000 [ 0.6832; 0.8283 ] 
#>   eta2 =~ y23      0.7997      0.0343   23.3006    0.0000 [ 0.7044; 0.8818 ] 
#>   eta3 =~ y31      0.8223      0.0310   26.4864    0.0000 [ 0.7376; 0.8982 ] 
#>   eta3 =~ y32      0.6581      0.0407   16.1844    0.0000 [ 0.5464; 0.7566 ] 
#>   eta3 =~ y33      0.7474      0.0406   18.4243    0.0000 [ 0.6510; 0.8608 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0203   19.5069    0.0000 [ 0.3388; 0.4437 ] 
#>   eta1 <~ y12      0.3873      0.0211   18.3455    0.0000 [ 0.3396; 0.4488 ] 
#>   eta1 <~ y13      0.4542      0.0184   24.7107    0.0000 [ 0.4082; 0.5032 ] 
#>   eta2 <~ y21      0.3058      0.0237   12.8888    0.0000 [ 0.2456; 0.3683 ] 
#>   eta2 <~ y22      0.4473      0.0180   24.8020    0.0000 [ 0.4035; 0.4968 ] 
#>   eta2 <~ y23      0.4735      0.0155   30.4737    0.0000 [ 0.4325; 0.5129 ] 
#>   eta3 <~ y31      0.4400      0.0172   25.5605    0.0000 [ 0.3939; 0.4829 ] 
#>   eta3 <~ y32      0.3521      0.0194   18.1933    0.0000 [ 0.2993; 0.3994 ] 
#>   eta3 <~ y33      0.3999      0.0193   20.7356    0.0000 [ 0.3555; 0.4552 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0400   16.8010    0.0000 [ 0.5602; 0.7668 ] 
#>   eta3 ~ eta1       0.6634      0.0350   18.9348    0.0000 [ 0.5706; 0.7518 ] 
#>   eta3 ~ eta2       0.3052      0.0987    3.0930    0.0020 [ 0.0443; 0.5545 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0633    3.2342    0.0012 [ 0.0361; 0.3637 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03995805 16.800956 2.401422e-63
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08202008  5.590177 2.268386e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09865906  3.092986 1.981533e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56016044          0.7668009          0.5849739          0.7419874
#> 2         0.24924405          0.6734056          0.3001776          0.6224721
#> 3         0.04433831          0.5545473          0.1056044          0.4932812
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5828922          0.7645056          0.6196891          0.7414617
#> 2          0.2505020          0.6003654          0.2785777          0.5883742
#> 3          0.1246416          0.5360955          0.1333345          0.5323237