Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2123189044
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0418   16.0650    0.0000 [ 0.6039; 0.7316 ] 
#>   eta3 ~ eta1      0.4585      0.0629    7.2932    0.0000 [ 0.3350; 0.5405 ] 
#>   eta3 ~ eta2      0.3052      0.0668    4.5663    0.0000 [ 0.2078; 0.4286 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0388   17.0728    0.0000 [ 0.6066; 0.7152 ] 
#>   eta1 =~ y12      0.6493      0.0428   15.1546    0.0000 [ 0.5790; 0.7217 ] 
#>   eta1 =~ y13      0.7613      0.0431   17.6668    0.0000 [ 0.6898; 0.8406 ] 
#>   eta2 =~ y21      0.5165      0.0613    8.4247    0.0000 [ 0.4083; 0.6002 ] 
#>   eta2 =~ y22      0.7554      0.0300   25.1790    0.0000 [ 0.7118; 0.8292 ] 
#>   eta2 =~ y23      0.7997      0.0384   20.8298    0.0000 [ 0.7344; 0.8686 ] 
#>   eta3 =~ y31      0.8223      0.0368   22.3347    0.0000 [ 0.7613; 0.8925 ] 
#>   eta3 =~ y32      0.6581      0.0427   15.4283    0.0000 [ 0.5775; 0.7201 ] 
#>   eta3 =~ y33      0.7474      0.0444   16.8170    0.0000 [ 0.6546; 0.8114 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0217   18.2504    0.0000 [ 0.3545; 0.4289 ] 
#>   eta1 <~ y12      0.3873      0.0227   17.0579    0.0000 [ 0.3440; 0.4249 ] 
#>   eta1 <~ y13      0.4542      0.0231   19.6837    0.0000 [ 0.4108; 0.4923 ] 
#>   eta2 <~ y21      0.3058      0.0315    9.7078    0.0000 [ 0.2453; 0.3534 ] 
#>   eta2 <~ y22      0.4473      0.0231   19.3274    0.0000 [ 0.4156; 0.4988 ] 
#>   eta2 <~ y23      0.4735      0.0213   22.2626    0.0000 [ 0.4331; 0.5127 ] 
#>   eta3 <~ y31      0.4400      0.0231   19.0684    0.0000 [ 0.3999; 0.4831 ] 
#>   eta3 <~ y32      0.3521      0.0170   20.6742    0.0000 [ 0.3218; 0.3760 ] 
#>   eta3 <~ y33      0.3999      0.0222   18.0020    0.0000 [ 0.3512; 0.4308 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0418   16.0650    0.0000 [ 0.6039; 0.7316 ] 
#>   eta3 ~ eta1       0.6634      0.0388   17.0904    0.0000 [ 0.5837; 0.7221 ] 
#>   eta3 ~ eta2       0.3052      0.0668    4.5663    0.0000 [ 0.2078; 0.4286 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0461    4.4477    0.0000 [ 0.1377; 0.2903 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03883789 17.072759  2.367417e-65
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04284368 15.154579  7.066654e-52
#> 3 eta1 =~ y13  Common factor 0.7613458 0.04309482 17.666761  7.561640e-70
#> 4 eta2 =~ y21  Common factor 0.5164548 0.06130242  8.424705  3.616590e-17
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03000074 25.178965 6.810144e-140
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03839033 20.829820  2.323334e-96
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03681608 22.334733 1.699423e-110
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04265347 15.428263  1.056771e-53
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04444461 16.816981  1.832639e-63
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6066208          0.7151778
#> 2          0.5789596          0.7216536
#> 3          0.6898149          0.8405695
#> 4          0.4083396          0.6001536
#> 5          0.7118262          0.8292004
#> 6          0.7344064          0.8686177
#> 7          0.7613298          0.8924927
#> 8          0.5774652          0.7201035
#> 9          0.6545630          0.8114288

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2123189044
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0418   16.0650    0.0000 [ 0.5652; 0.7813 ] 
#>   eta3 ~ eta1      0.4585      0.0629    7.2932    0.0000 [ 0.3101; 0.6352 ] 
#>   eta3 ~ eta2      0.3052      0.0668    4.5663    0.0000 [ 0.1186; 0.4642 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0388   17.0728    0.0000 [ 0.5617; 0.7626 ] 
#>   eta1 =~ y12      0.6493      0.0428   15.1546    0.0000 [ 0.5444; 0.7659 ] 
#>   eta1 =~ y13      0.7613      0.0431   17.6668    0.0000 [ 0.6466; 0.8695 ] 
#>   eta2 =~ y21      0.5165      0.0613    8.4247    0.0000 [ 0.3677; 0.6847 ] 
#>   eta2 =~ y22      0.7554      0.0300   25.1790    0.0000 [ 0.6770; 0.8321 ] 
#>   eta2 =~ y23      0.7997      0.0384   20.8298    0.0000 [ 0.7023; 0.9008 ] 
#>   eta3 =~ y31      0.8223      0.0368   22.3347    0.0000 [ 0.7208; 0.9112 ] 
#>   eta3 =~ y32      0.6581      0.0427   15.4283    0.0000 [ 0.5516; 0.7722 ] 
#>   eta3 =~ y33      0.7474      0.0444   16.8170    0.0000 [ 0.6378; 0.8676 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0217   18.2504    0.0000 [ 0.3388; 0.4508 ] 
#>   eta1 <~ y12      0.3873      0.0227   17.0579    0.0000 [ 0.3320; 0.4494 ] 
#>   eta1 <~ y13      0.4542      0.0231   19.6837    0.0000 [ 0.3924; 0.5117 ] 
#>   eta2 <~ y21      0.3058      0.0315    9.7078    0.0000 [ 0.2289; 0.3919 ] 
#>   eta2 <~ y22      0.4473      0.0231   19.3274    0.0000 [ 0.3842; 0.5039 ] 
#>   eta2 <~ y23      0.4735      0.0213   22.2626    0.0000 [ 0.4170; 0.5270 ] 
#>   eta3 <~ y31      0.4400      0.0231   19.0684    0.0000 [ 0.3762; 0.4955 ] 
#>   eta3 <~ y32      0.3521      0.0170   20.6742    0.0000 [ 0.3100; 0.3981 ] 
#>   eta3 <~ y33      0.3999      0.0222   18.0020    0.0000 [ 0.3448; 0.4597 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0418   16.0650    0.0000 [ 0.5652; 0.7813 ] 
#>   eta3 ~ eta1       0.6634      0.0388   17.0904    0.0000 [ 0.5687; 0.7695 ] 
#>   eta3 ~ eta2       0.3052      0.0668    4.5663    0.0000 [ 0.1186; 0.4642 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0461    4.4477    0.0000 [ 0.0774; 0.3156 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04178849 16.065032 4.486296e-58
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.06286762  7.293210 3.026562e-13
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.06682739  4.566258 4.965073e-06
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5651505          0.7812570          0.5911007          0.7553068
#> 2          0.3100666          0.6351825          0.3491067          0.5961424
#> 3          0.1186424          0.4642359          0.1601414          0.4227369
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5767180          0.7601469          0.6039187          0.7315737
#> 2          0.3223132          0.5805790          0.3349930          0.5405236
#> 3          0.1603978          0.4736347          0.2078090          0.4286040