Summarize model

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -2062964741
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0373   18.0171    0.0000 [ 0.6110; 0.7428 ] 
#>   eta3 ~ eta1      0.4585      0.1041    4.4037    0.0000 [ 0.3102; 0.6718 ] 
#>   eta3 ~ eta2      0.3052      0.1071    2.8494    0.0044 [ 0.1330; 0.5143 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0414   16.0165    0.0000 [ 0.5965; 0.7332 ] 
#>   eta1 =~ y12      0.6493      0.0359   18.0760    0.0000 [ 0.5711; 0.6980 ] 
#>   eta1 =~ y13      0.7613      0.0404   18.8623    0.0000 [ 0.6981; 0.8296 ] 
#>   eta2 =~ y21      0.5165      0.0539    9.5893    0.0000 [ 0.4024; 0.6091 ] 
#>   eta2 =~ y22      0.7554      0.0300   25.1387    0.0000 [ 0.7056; 0.7985 ] 
#>   eta2 =~ y23      0.7997      0.0370   21.6213    0.0000 [ 0.7519; 0.8747 ] 
#>   eta3 =~ y31      0.8223      0.0331   24.8389    0.0000 [ 0.7604; 0.8808 ] 
#>   eta3 =~ y32      0.6581      0.0361   18.2175    0.0000 [ 0.5977; 0.7201 ] 
#>   eta3 =~ y33      0.7474      0.0404   18.5018    0.0000 [ 0.6900; 0.8210 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0238   16.6079    0.0000 [ 0.3565; 0.4393 ] 
#>   eta1 <~ y12      0.3873      0.0206   18.8162    0.0000 [ 0.3499; 0.4251 ] 
#>   eta1 <~ y13      0.4542      0.0191   23.7770    0.0000 [ 0.4252; 0.4837 ] 
#>   eta2 <~ y21      0.3058      0.0246   12.4099    0.0000 [ 0.2484; 0.3492 ] 
#>   eta2 <~ y22      0.4473      0.0243   18.3766    0.0000 [ 0.4046; 0.4924 ] 
#>   eta2 <~ y23      0.4735      0.0186   25.4739    0.0000 [ 0.4452; 0.5083 ] 
#>   eta3 <~ y31      0.4400      0.0180   24.5079    0.0000 [ 0.4057; 0.4735 ] 
#>   eta3 <~ y32      0.3521      0.0183   19.1925    0.0000 [ 0.3137; 0.3805 ] 
#>   eta3 <~ y33      0.3999      0.0187   21.4393    0.0000 [ 0.3708; 0.4334 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0373   18.0171    0.0000 [ 0.6110; 0.7428 ] 
#>   eta3 ~ eta1       0.6634      0.0480   13.8313    0.0000 [ 0.5897; 0.7562 ] 
#>   eta3 ~ eta2       0.3052      0.1071    2.8494    0.0044 [ 0.1330; 0.5143 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0695    2.9459    0.0032 [ 0.0991; 0.3120 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04139906 16.016545  9.794285e-58
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03591937 18.075985  4.926798e-73
#> 3 eta1 =~ y13  Common factor 0.7613458 0.04036328 18.862337  2.327043e-79
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05385750  9.589283  8.869947e-22
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03004884 25.138660 1.880313e-139
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03698492 21.621347 1.131259e-103
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03310448 24.838854 3.412412e-136
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03612285 18.217523  3.747384e-74
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04039738 18.501797  1.997036e-76
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5965337          0.7332222
#> 2          0.5711445          0.6980311
#> 3          0.6980582          0.8296241
#> 4          0.4024223          0.6090986
#> 5          0.7056107          0.7984711
#> 6          0.7518777          0.8746658
#> 7          0.7603820          0.8808326
#> 8          0.5976598          0.7200974
#> 9          0.6900378          0.8210277

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -2062964741
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0373   18.0171    0.0000 [ 0.5665; 0.7591 ] 
#>   eta3 ~ eta1      0.4585      0.1041    4.4037    0.0000 [ 0.1952; 0.7337 ] 
#>   eta3 ~ eta2      0.3052      0.1071    2.8494    0.0044 [ 0.0244; 0.5782 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0414   16.0165    0.0000 [ 0.5574; 0.7715 ] 
#>   eta1 =~ y12      0.6493      0.0359   18.0760    0.0000 [ 0.5562; 0.7420 ] 
#>   eta1 =~ y13      0.7613      0.0404   18.8623    0.0000 [ 0.6564; 0.8651 ] 
#>   eta2 =~ y21      0.5165      0.0539    9.5893    0.0000 [ 0.3721; 0.6506 ] 
#>   eta2 =~ y22      0.7554      0.0300   25.1387    0.0000 [ 0.6743; 0.8297 ] 
#>   eta2 =~ y23      0.7997      0.0370   21.6213    0.0000 [ 0.6944; 0.8856 ] 
#>   eta3 =~ y31      0.8223      0.0331   24.8389    0.0000 [ 0.7342; 0.9054 ] 
#>   eta3 =~ y32      0.6581      0.0361   18.2175    0.0000 [ 0.5665; 0.7533 ] 
#>   eta3 =~ y33      0.7474      0.0404   18.5018    0.0000 [ 0.6383; 0.8472 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0238   16.6079    0.0000 [ 0.3348; 0.4580 ] 
#>   eta1 <~ y12      0.3873      0.0206   18.8162    0.0000 [ 0.3340; 0.4404 ] 
#>   eta1 <~ y13      0.4542      0.0191   23.7770    0.0000 [ 0.4046; 0.5034 ] 
#>   eta2 <~ y21      0.3058      0.0246   12.4099    0.0000 [ 0.2439; 0.3713 ] 
#>   eta2 <~ y22      0.4473      0.0243   18.3766    0.0000 [ 0.3881; 0.5139 ] 
#>   eta2 <~ y23      0.4735      0.0186   25.4739    0.0000 [ 0.4263; 0.5224 ] 
#>   eta3 <~ y31      0.4400      0.0180   24.5079    0.0000 [ 0.3940; 0.4868 ] 
#>   eta3 <~ y32      0.3521      0.0183   19.1925    0.0000 [ 0.3071; 0.4020 ] 
#>   eta3 <~ y33      0.3999      0.0187   21.4393    0.0000 [ 0.3510; 0.4474 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0373   18.0171    0.0000 [ 0.5665; 0.7591 ] 
#>   eta3 ~ eta1       0.6634      0.0480   13.8313    0.0000 [ 0.5416; 0.7896 ] 
#>   eta3 ~ eta2       0.3052      0.1071    2.8494    0.0044 [ 0.0244; 0.5782 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0695    2.9459    0.0032 [ 0.0213; 0.3809 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03726095 18.017076 1.431090e-72
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.10411813  4.403717 1.064117e-05
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.10709468  2.849358 4.380752e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56645400          0.7591466          0.5895926          0.7360080
#> 2         0.19524459          0.7336848          0.2599007          0.6690286
#> 3         0.02438456          0.5782178          0.0908891          0.5117132
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.57867947          0.7447184          0.6109575          0.7428268
#> 2         0.24771819          0.6898931          0.3102051          0.6717589
#> 3         0.05890778          0.5390904          0.1330245          0.5142584