Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -511296746
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0296   22.6983    0.0000 [ 0.6283; 0.7373 ] 
#>   eta3 ~ eta1      0.4585      0.0810    5.6628    0.0000 [ 0.2782; 0.5827 ] 
#>   eta3 ~ eta2      0.3052      0.0931    3.2767    0.0011 [ 0.1186; 0.4867 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0387   17.1361    0.0000 [ 0.5911; 0.7255 ] 
#>   eta1 =~ y12      0.6493      0.0360   18.0448    0.0000 [ 0.5903; 0.7091 ] 
#>   eta1 =~ y13      0.7613      0.0296   25.6953    0.0000 [ 0.7015; 0.8051 ] 
#>   eta2 =~ y21      0.5165      0.0486   10.6238    0.0000 [ 0.4380; 0.6119 ] 
#>   eta2 =~ y22      0.7554      0.0420   17.9804    0.0000 [ 0.6709; 0.8491 ] 
#>   eta2 =~ y23      0.7997      0.0390   20.5067    0.0000 [ 0.7256; 0.8750 ] 
#>   eta3 =~ y31      0.8223      0.0278   29.5910    0.0000 [ 0.7672; 0.8714 ] 
#>   eta3 =~ y32      0.6581      0.0475   13.8641    0.0000 [ 0.5729; 0.7376 ] 
#>   eta3 =~ y33      0.7474      0.0332   22.5205    0.0000 [ 0.6961; 0.8112 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0172   23.0175    0.0000 [ 0.3668; 0.4204 ] 
#>   eta1 <~ y12      0.3873      0.0200   19.3551    0.0000 [ 0.3529; 0.4284 ] 
#>   eta1 <~ y13      0.4542      0.0195   23.3063    0.0000 [ 0.4169; 0.4883 ] 
#>   eta2 <~ y21      0.3058      0.0235   13.0168    0.0000 [ 0.2655; 0.3449 ] 
#>   eta2 <~ y22      0.4473      0.0248   18.0405    0.0000 [ 0.4050; 0.4822 ] 
#>   eta2 <~ y23      0.4735      0.0203   23.3322    0.0000 [ 0.4357; 0.5101 ] 
#>   eta3 <~ y31      0.4400      0.0185   23.7838    0.0000 [ 0.4092; 0.4747 ] 
#>   eta3 <~ y32      0.3521      0.0198   17.7526    0.0000 [ 0.3155; 0.3808 ] 
#>   eta3 <~ y33      0.3999      0.0188   21.2278    0.0000 [ 0.3687; 0.4389 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0296   22.6983    0.0000 [ 0.6283; 0.7373 ] 
#>   eta3 ~ eta1       0.6634      0.0312   21.2880    0.0000 [ 0.6123; 0.7281 ] 
#>   eta3 ~ eta2       0.3052      0.0931    3.2767    0.0011 [ 0.1186; 0.4867 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0666    3.0772    0.0021 [ 0.0808; 0.3443 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03869438 17.13608  7.985976e-66
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03598147 18.04478  8.670218e-73
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02962978 25.69529 1.319541e-145
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04861300 10.62380  2.309633e-26
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04201176 17.98039  2.775556e-72
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03899525 20.50670  1.876157e-93
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02778813 29.59096 1.953249e-192
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04746570 13.86409  1.045375e-43
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03318861 22.52050 2.613951e-112
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5910639          0.7254639
#> 2          0.5903198          0.7090941
#> 3          0.7014712          0.8050817
#> 4          0.4379537          0.6118843
#> 5          0.6708561          0.8490893
#> 6          0.7256395          0.8749704
#> 7          0.7672109          0.8714358
#> 8          0.5729476          0.7376427
#> 9          0.6961116          0.8111652

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -511296746
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0296   22.6983    0.0000 [ 0.5871; 0.7400 ] 
#>   eta3 ~ eta1      0.4585      0.0810    5.6628    0.0000 [ 0.2445; 0.6632 ] 
#>   eta3 ~ eta2      0.3052      0.0931    3.2767    0.0011 [ 0.0639; 0.5455 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0387   17.1361    0.0000 [ 0.5698; 0.7699 ] 
#>   eta1 =~ y12      0.6493      0.0360   18.0448    0.0000 [ 0.5525; 0.7386 ] 
#>   eta1 =~ y13      0.7613      0.0296   25.6953    0.0000 [ 0.6881; 0.8414 ] 
#>   eta2 =~ y21      0.5165      0.0486   10.6238    0.0000 [ 0.3821; 0.6335 ] 
#>   eta2 =~ y22      0.7554      0.0420   17.9804    0.0000 [ 0.6437; 0.8610 ] 
#>   eta2 =~ y23      0.7997      0.0390   20.5067    0.0000 [ 0.7009; 0.9025 ] 
#>   eta3 =~ y31      0.8223      0.0278   29.5910    0.0000 [ 0.7558; 0.8995 ] 
#>   eta3 =~ y32      0.6581      0.0475   13.8641    0.0000 [ 0.5348; 0.7802 ] 
#>   eta3 =~ y33      0.7474      0.0332   22.5205    0.0000 [ 0.6534; 0.8250 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0172   23.0175    0.0000 [ 0.3537; 0.4426 ] 
#>   eta1 <~ y12      0.3873      0.0200   19.3551    0.0000 [ 0.3317; 0.4352 ] 
#>   eta1 <~ y13      0.4542      0.0195   23.3063    0.0000 [ 0.4037; 0.5045 ] 
#>   eta2 <~ y21      0.3058      0.0235   13.0168    0.0000 [ 0.2424; 0.3639 ] 
#>   eta2 <~ y22      0.4473      0.0248   18.0405    0.0000 [ 0.3843; 0.5125 ] 
#>   eta2 <~ y23      0.4735      0.0203   23.3322    0.0000 [ 0.4254; 0.5304 ] 
#>   eta3 <~ y31      0.4400      0.0185   23.7838    0.0000 [ 0.3961; 0.4917 ] 
#>   eta3 <~ y32      0.3521      0.0198   17.7526    0.0000 [ 0.3018; 0.4044 ] 
#>   eta3 <~ y33      0.3999      0.0188   21.2278    0.0000 [ 0.3478; 0.4453 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0296   22.6983    0.0000 [ 0.5871; 0.7400 ] 
#>   eta3 ~ eta1       0.6634      0.0312   21.2880    0.0000 [ 0.5751; 0.7362 ] 
#>   eta3 ~ eta2       0.3052      0.0931    3.2767    0.0011 [ 0.0639; 0.5455 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0666    3.0772    0.0021 [ 0.0296; 0.3739 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.02957644 22.698250 4.661807e-114
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08096770  5.662835  1.488921e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09312720  3.276713  1.050229e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.58708015          0.7400328          0.6054468          0.7216662
#> 2         0.24449584          0.6632151          0.2947758          0.6129351
#> 3         0.06389431          0.5454956          0.1217252          0.4876647
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.6158066          0.7456232          0.6283119          0.7372940
#> 2          0.2569542          0.6575603          0.2781696          0.5827334
#> 3          0.1065670          0.5347747          0.1185744          0.4866767