Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1390507919
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0384   17.4733    0.0000 [ 0.6020; 0.7378 ] 
#>   eta3 ~ eta1      0.4585      0.0829    5.5288    0.0000 [ 0.3498; 0.6310 ] 
#>   eta3 ~ eta2      0.3052      0.0836    3.6507    0.0003 [ 0.1417; 0.4104 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0377   17.5824    0.0000 [ 0.5940; 0.7210 ] 
#>   eta1 =~ y12      0.6493      0.0272   23.9104    0.0000 [ 0.5895; 0.7005 ] 
#>   eta1 =~ y13      0.7613      0.0389   19.5857    0.0000 [ 0.6690; 0.8182 ] 
#>   eta2 =~ y21      0.5165      0.0544    9.4894    0.0000 [ 0.4289; 0.6125 ] 
#>   eta2 =~ y22      0.7554      0.0290   26.0182    0.0000 [ 0.7051; 0.8091 ] 
#>   eta2 =~ y23      0.7997      0.0436   18.3456    0.0000 [ 0.7246; 0.8712 ] 
#>   eta3 =~ y31      0.8223      0.0323   25.4582    0.0000 [ 0.7572; 0.8857 ] 
#>   eta3 =~ y32      0.6581      0.0363   18.1415    0.0000 [ 0.5650; 0.7151 ] 
#>   eta3 =~ y33      0.7474      0.0401   18.6200    0.0000 [ 0.6776; 0.8134 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0191   20.6730    0.0000 [ 0.3658; 0.4361 ] 
#>   eta1 <~ y12      0.3873      0.0167   23.1569    0.0000 [ 0.3611; 0.4252 ] 
#>   eta1 <~ y13      0.4542      0.0205   22.1136    0.0000 [ 0.4158; 0.4868 ] 
#>   eta2 <~ y21      0.3058      0.0302   10.1180    0.0000 [ 0.2593; 0.3671 ] 
#>   eta2 <~ y22      0.4473      0.0208   21.4890    0.0000 [ 0.4078; 0.4840 ] 
#>   eta2 <~ y23      0.4735      0.0209   22.6396    0.0000 [ 0.4389; 0.5116 ] 
#>   eta3 <~ y31      0.4400      0.0184   23.8596    0.0000 [ 0.4101; 0.4698 ] 
#>   eta3 <~ y32      0.3521      0.0171   20.5437    0.0000 [ 0.3150; 0.3766 ] 
#>   eta3 <~ y33      0.3999      0.0192   20.8473    0.0000 [ 0.3735; 0.4378 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0384   17.4733    0.0000 [ 0.6020; 0.7378 ] 
#>   eta3 ~ eta1       0.6634      0.0399   16.6202    0.0000 [ 0.5839; 0.7466 ] 
#>   eta3 ~ eta2       0.3052      0.0836    3.6507    0.0003 [ 0.1417; 0.4104 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0529    3.8691    0.0001 [ 0.0974; 0.2818 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03771206 17.58243  3.358389e-69
#> 2 eta1 =~ y12  Common factor 0.6492779 0.02715467 23.91036 2.390041e-126
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03887251 19.58571  2.047289e-85
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05442439  9.48940  2.323685e-21
#> 5 eta2 =~ y22  Common factor 0.7553877 0.02903301 26.01824 3.079492e-149
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04358883 18.34561  3.578874e-75
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03229910 25.45821 5.725743e-143
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03627414 18.14154  1.497689e-73
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04014103 18.61996  2.213978e-77
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5940474          0.7209576
#> 2          0.5895451          0.7005117
#> 3          0.6689866          0.8182483
#> 4          0.4288733          0.6124548
#> 5          0.7051423          0.8091363
#> 6          0.7245746          0.8711986
#> 7          0.7571512          0.8856958
#> 8          0.5650257          0.7151429
#> 9          0.6775956          0.8134258

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1390507919
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0384   17.4733    0.0000 [ 0.5723; 0.7710 ] 
#>   eta3 ~ eta1      0.4585      0.0829    5.5288    0.0000 [ 0.2241; 0.6530 ] 
#>   eta3 ~ eta2      0.3052      0.0836    3.6507    0.0003 [ 0.1111; 0.5434 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0377   17.5824    0.0000 [ 0.5667; 0.7617 ] 
#>   eta1 =~ y12      0.6493      0.0272   23.9104    0.0000 [ 0.5890; 0.7294 ] 
#>   eta1 =~ y13      0.7613      0.0389   19.5857    0.0000 [ 0.6656; 0.8666 ] 
#>   eta2 =~ y21      0.5165      0.0544    9.4894    0.0000 [ 0.3832; 0.6646 ] 
#>   eta2 =~ y22      0.7554      0.0290   26.0182    0.0000 [ 0.6886; 0.8387 ] 
#>   eta2 =~ y23      0.7997      0.0436   18.3456    0.0000 [ 0.6833; 0.9087 ] 
#>   eta3 =~ y31      0.8223      0.0323   25.4582    0.0000 [ 0.7415; 0.9086 ] 
#>   eta3 =~ y32      0.6581      0.0363   18.1415    0.0000 [ 0.5670; 0.7546 ] 
#>   eta3 =~ y33      0.7474      0.0401   18.6200    0.0000 [ 0.6431; 0.8507 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0191   20.6730    0.0000 [ 0.3425; 0.4415 ] 
#>   eta1 <~ y12      0.3873      0.0167   23.1569    0.0000 [ 0.3457; 0.4322 ] 
#>   eta1 <~ y13      0.4542      0.0205   22.1136    0.0000 [ 0.3990; 0.5052 ] 
#>   eta2 <~ y21      0.3058      0.0302   10.1180    0.0000 [ 0.2301; 0.3864 ] 
#>   eta2 <~ y22      0.4473      0.0208   21.4890    0.0000 [ 0.3951; 0.5027 ] 
#>   eta2 <~ y23      0.4735      0.0209   22.6396    0.0000 [ 0.4141; 0.5223 ] 
#>   eta3 <~ y31      0.4400      0.0184   23.8596    0.0000 [ 0.3924; 0.4877 ] 
#>   eta3 <~ y32      0.3521      0.0171   20.5437    0.0000 [ 0.3083; 0.3969 ] 
#>   eta3 <~ y33      0.3999      0.0192   20.8473    0.0000 [ 0.3489; 0.4481 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0384   17.4733    0.0000 [ 0.5723; 0.7710 ] 
#>   eta3 ~ eta1       0.6634      0.0399   16.6202    0.0000 [ 0.5565; 0.7629 ] 
#>   eta3 ~ eta2       0.3052      0.0836    3.6507    0.0003 [ 0.1111; 0.5434 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0529    3.8691    0.0001 [ 0.0843; 0.3581 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03842053 17.473300 2.288676e-68
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08293064  5.528798 3.224319e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08358805  3.650655 2.615724e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5722938          0.7709831          0.5961525          0.7471244
#> 2          0.2241180          0.6529884          0.2756169          0.6014895
#> 3          0.1111028          0.5433729          0.1630100          0.4914657
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5797978          0.7512464          0.6019692          0.7378090
#> 2          0.2349409          0.6499513          0.3498350          0.6309712
#> 3          0.1086709          0.4992041          0.1416760          0.4104499