Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -460013068
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0486   13.8094    0.0000 [ 0.5652; 0.7583 ] 
#>   eta3 ~ eta1      0.4585      0.0888    5.1638    0.0000 [ 0.3195; 0.6418 ] 
#>   eta3 ~ eta2      0.3052      0.0992    3.0754    0.0021 [ 0.0974; 0.4845 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0396   16.7642    0.0000 [ 0.5908; 0.7212 ] 
#>   eta1 =~ y12      0.6493      0.0366   17.7214    0.0000 [ 0.5839; 0.6985 ] 
#>   eta1 =~ y13      0.7613      0.0410   18.5550    0.0000 [ 0.6772; 0.8331 ] 
#>   eta2 =~ y21      0.5165      0.0510   10.1211    0.0000 [ 0.4153; 0.6035 ] 
#>   eta2 =~ y22      0.7554      0.0355   21.2831    0.0000 [ 0.6973; 0.8089 ] 
#>   eta2 =~ y23      0.7997      0.0398   20.1090    0.0000 [ 0.7169; 0.8531 ] 
#>   eta3 =~ y31      0.8223      0.0296   27.7472    0.0000 [ 0.7683; 0.8696 ] 
#>   eta3 =~ y32      0.6581      0.0347   18.9799    0.0000 [ 0.5846; 0.7137 ] 
#>   eta3 =~ y33      0.7474      0.0388   19.2846    0.0000 [ 0.6676; 0.8109 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0223   17.7135    0.0000 [ 0.3579; 0.4316 ] 
#>   eta1 <~ y12      0.3873      0.0224   17.2585    0.0000 [ 0.3457; 0.4221 ] 
#>   eta1 <~ y13      0.4542      0.0205   22.1612    0.0000 [ 0.4126; 0.4823 ] 
#>   eta2 <~ y21      0.3058      0.0291   10.5081    0.0000 [ 0.2572; 0.3602 ] 
#>   eta2 <~ y22      0.4473      0.0167   26.8310    0.0000 [ 0.4218; 0.4741 ] 
#>   eta2 <~ y23      0.4735      0.0215   22.0471    0.0000 [ 0.4343; 0.5077 ] 
#>   eta3 <~ y31      0.4400      0.0143   30.8471    0.0000 [ 0.4122; 0.4647 ] 
#>   eta3 <~ y32      0.3521      0.0195   18.0607    0.0000 [ 0.3184; 0.3840 ] 
#>   eta3 <~ y33      0.3999      0.0155   25.7940    0.0000 [ 0.3729; 0.4264 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0486   13.8094    0.0000 [ 0.5652; 0.7583 ] 
#>   eta3 ~ eta1       0.6634      0.0381   17.4315    0.0000 [ 0.5924; 0.7211 ] 
#>   eta3 ~ eta2       0.3052      0.0992    3.0754    0.0021 [ 0.0974; 0.4845 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0680    3.0123    0.0026 [ 0.0726; 0.3241 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03955281 16.76416  4.462474e-63
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03663803 17.72142  2.865891e-70
#> 3 eta1 =~ y13  Common factor 0.7613458 0.04103190 18.55497  7.434483e-77
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05102774 10.12106  4.455627e-24
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03549236 21.28311 1.627869e-100
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03976653 20.10896  6.159822e-90
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02963458 27.74723 1.882133e-169
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03467189 18.97989  2.500970e-80
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03875764 19.28456  7.240095e-83
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5907916          0.7212186
#> 2          0.5839421          0.6984610
#> 3          0.6771772          0.8331329
#> 4          0.4152628          0.6035469
#> 5          0.6972870          0.8089241
#> 6          0.7169126          0.8531316
#> 7          0.7682677          0.8695824
#> 8          0.5846079          0.7137417
#> 9          0.6676310          0.8109469

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -460013068
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0486   13.8094    0.0000 [ 0.5390; 0.7904 ] 
#>   eta3 ~ eta1      0.4585      0.0888    5.1638    0.0000 [ 0.2241; 0.6833 ] 
#>   eta3 ~ eta2      0.3052      0.0992    3.0754    0.0021 [ 0.0560; 0.5691 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0396   16.7642    0.0000 [ 0.5607; 0.7653 ] 
#>   eta1 =~ y12      0.6493      0.0366   17.7214    0.0000 [ 0.5537; 0.7432 ] 
#>   eta1 =~ y13      0.7613      0.0410   18.5550    0.0000 [ 0.6606; 0.8728 ] 
#>   eta2 =~ y21      0.5165      0.0510   10.1211    0.0000 [ 0.3894; 0.6533 ] 
#>   eta2 =~ y22      0.7554      0.0355   21.2831    0.0000 [ 0.6639; 0.8475 ] 
#>   eta2 =~ y23      0.7997      0.0398   20.1090    0.0000 [ 0.7049; 0.9106 ] 
#>   eta3 =~ y31      0.8223      0.0296   27.7472    0.0000 [ 0.7468; 0.9000 ] 
#>   eta3 =~ y32      0.6581      0.0347   18.9799    0.0000 [ 0.5699; 0.7492 ] 
#>   eta3 =~ y33      0.7474      0.0388   19.2846    0.0000 [ 0.6464; 0.8468 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0223   17.7135    0.0000 [ 0.3367; 0.4522 ] 
#>   eta1 <~ y12      0.3873      0.0224   17.2585    0.0000 [ 0.3277; 0.4437 ] 
#>   eta1 <~ y13      0.4542      0.0205   22.1612    0.0000 [ 0.4033; 0.5093 ] 
#>   eta2 <~ y21      0.3058      0.0291   10.5081    0.0000 [ 0.2309; 0.3814 ] 
#>   eta2 <~ y22      0.4473      0.0167   26.8310    0.0000 [ 0.4007; 0.4869 ] 
#>   eta2 <~ y23      0.4735      0.0215   22.0471    0.0000 [ 0.4188; 0.5299 ] 
#>   eta3 <~ y31      0.4400      0.0143   30.8471    0.0000 [ 0.4033; 0.4770 ] 
#>   eta3 <~ y32      0.3521      0.0195   18.0607    0.0000 [ 0.3021; 0.4029 ] 
#>   eta3 <~ y33      0.3999      0.0155   25.7940    0.0000 [ 0.3592; 0.4394 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0486   13.8094    0.0000 [ 0.5390; 0.7904 ] 
#>   eta3 ~ eta1       0.6634      0.0381   17.4315    0.0000 [ 0.5639; 0.7607 ] 
#>   eta3 ~ eta2       0.3052      0.0992    3.0754    0.0021 [ 0.0560; 0.5691 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0680    3.0123    0.0026 [ 0.0328; 0.3845 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04861423 13.809403 2.236781e-43
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08879271  5.163789 2.420009e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09922345  3.075393 2.102253e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.53898096          0.7903863          0.5691698          0.7601974
#> 2         0.22413472          0.6833205          0.2792740          0.6281812
#> 3         0.05599104          0.5691187          0.1176076          0.5075021
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.55952466          0.7672399         0.56519279          0.7583387
#> 2         0.29571757          0.6464225         0.31953321          0.6418287
#> 3         0.06398087          0.4889678         0.09735634          0.4845128