Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1110179390
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0406   16.5423    0.0000 [ 0.5960; 0.7343 ] 
#>   eta3 ~ eta1      0.4585      0.0802    5.7200    0.0000 [ 0.2999; 0.6306 ] 
#>   eta3 ~ eta2      0.3052      0.0835    3.6552    0.0003 [ 0.1560; 0.4267 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0401   16.5307    0.0000 [ 0.5767; 0.7192 ] 
#>   eta1 =~ y12      0.6493      0.0370   17.5648    0.0000 [ 0.5848; 0.7171 ] 
#>   eta1 =~ y13      0.7613      0.0286   26.5987    0.0000 [ 0.7058; 0.8030 ] 
#>   eta2 =~ y21      0.5165      0.0436   11.8553    0.0000 [ 0.4359; 0.5977 ] 
#>   eta2 =~ y22      0.7554      0.0363   20.8129    0.0000 [ 0.6839; 0.8178 ] 
#>   eta2 =~ y23      0.7997      0.0448   17.8673    0.0000 [ 0.7027; 0.8849 ] 
#>   eta3 =~ y31      0.8223      0.0270   30.4039    0.0000 [ 0.7786; 0.8731 ] 
#>   eta3 =~ y32      0.6581      0.0406   16.2121    0.0000 [ 0.5645; 0.7249 ] 
#>   eta3 =~ y33      0.7474      0.0414   18.0340    0.0000 [ 0.6636; 0.8091 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0191   20.7101    0.0000 [ 0.3646; 0.4277 ] 
#>   eta1 <~ y12      0.3873      0.0212   18.2542    0.0000 [ 0.3556; 0.4306 ] 
#>   eta1 <~ y13      0.4542      0.0181   25.0680    0.0000 [ 0.4256; 0.4903 ] 
#>   eta2 <~ y21      0.3058      0.0205   14.9027    0.0000 [ 0.2706; 0.3391 ] 
#>   eta2 <~ y22      0.4473      0.0241   18.5714    0.0000 [ 0.4128; 0.4930 ] 
#>   eta2 <~ y23      0.4735      0.0248   19.1226    0.0000 [ 0.4119; 0.5090 ] 
#>   eta3 <~ y31      0.4400      0.0151   29.0487    0.0000 [ 0.4138; 0.4688 ] 
#>   eta3 <~ y32      0.3521      0.0212   16.6188    0.0000 [ 0.3079; 0.3951 ] 
#>   eta3 <~ y33      0.3999      0.0202   19.7653    0.0000 [ 0.3635; 0.4360 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0406   16.5423    0.0000 [ 0.5960; 0.7343 ] 
#>   eta3 ~ eta1       0.6634      0.0383   17.3142    0.0000 [ 0.5948; 0.7099 ] 
#>   eta3 ~ eta2       0.3052      0.0835    3.6552    0.0003 [ 0.1560; 0.4267 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0565    3.6252    0.0003 [ 0.0991; 0.2867 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04011132 16.53074  2.204204e-61
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03696473 17.56480  4.583397e-69
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02862339 26.59873 7.021298e-156
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04356303 11.85535  2.018880e-32
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03629413 20.81294  3.304156e-96
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04475569 17.86731  2.119844e-71
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02704517 30.40386 4.884501e-203
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04059125 16.21209  4.142481e-59
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04144519 18.03404  1.053107e-72
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5766522          0.7192275
#> 2          0.5848377          0.7170773
#> 3          0.7058036          0.8030363
#> 4          0.4359071          0.5977414
#> 5          0.6838862          0.8178341
#> 6          0.7027033          0.8849326
#> 7          0.7785614          0.8731485
#> 8          0.5645126          0.7248573
#> 9          0.6635912          0.8091060

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1110179390
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0406   16.5423    0.0000 [ 0.5648; 0.7747 ] 
#>   eta3 ~ eta1      0.4585      0.0802    5.7200    0.0000 [ 0.2531; 0.6676 ] 
#>   eta3 ~ eta2      0.3052      0.0835    3.6552    0.0003 [ 0.0982; 0.5300 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0401   16.5307    0.0000 [ 0.5671; 0.7745 ] 
#>   eta1 =~ y12      0.6493      0.0370   17.5648    0.0000 [ 0.5528; 0.7440 ] 
#>   eta1 =~ y13      0.7613      0.0286   26.5987    0.0000 [ 0.6975; 0.8455 ] 
#>   eta2 =~ y21      0.5165      0.0436   11.8553    0.0000 [ 0.4039; 0.6292 ] 
#>   eta2 =~ y22      0.7554      0.0363   20.8129    0.0000 [ 0.6602; 0.8479 ] 
#>   eta2 =~ y23      0.7997      0.0448   17.8673    0.0000 [ 0.6860; 0.9175 ] 
#>   eta3 =~ y31      0.8223      0.0270   30.4039    0.0000 [ 0.7468; 0.8867 ] 
#>   eta3 =~ y32      0.6581      0.0406   16.2121    0.0000 [ 0.5623; 0.7723 ] 
#>   eta3 =~ y33      0.7474      0.0414   18.0340    0.0000 [ 0.6448; 0.8591 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0191   20.7101    0.0000 [ 0.3463; 0.4450 ] 
#>   eta1 <~ y12      0.3873      0.0212   18.2542    0.0000 [ 0.3273; 0.4370 ] 
#>   eta1 <~ y13      0.4542      0.0181   25.0680    0.0000 [ 0.4079; 0.5016 ] 
#>   eta2 <~ y21      0.3058      0.0205   14.9027    0.0000 [ 0.2531; 0.3593 ] 
#>   eta2 <~ y22      0.4473      0.0241   18.5714    0.0000 [ 0.3840; 0.5086 ] 
#>   eta2 <~ y23      0.4735      0.0248   19.1226    0.0000 [ 0.4107; 0.5388 ] 
#>   eta3 <~ y31      0.4400      0.0151   29.0487    0.0000 [ 0.3957; 0.4740 ] 
#>   eta3 <~ y32      0.3521      0.0212   16.6188    0.0000 [ 0.3006; 0.4102 ] 
#>   eta3 <~ y33      0.3999      0.0202   19.7653    0.0000 [ 0.3482; 0.4529 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0406   16.5423    0.0000 [ 0.5648; 0.7747 ] 
#>   eta3 ~ eta1       0.6634      0.0383   17.3142    0.0000 [ 0.5724; 0.7705 ] 
#>   eta3 ~ eta2       0.3052      0.0835    3.6552    0.0003 [ 0.0982; 0.5300 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0565    3.6252    0.0003 [ 0.0649; 0.3572 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04058288 16.542283 1.819917e-61
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08015876  5.719983 1.065344e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08348428  3.655193 2.569888e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56482900          0.7747007          0.5900305          0.7494992
#> 2         0.25310485          0.6676407          0.3028825          0.6178630
#> 3         0.09822121          0.5299547          0.1500640          0.4781120
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.59443979          0.7703586          0.5960239          0.7342633
#> 2         0.29757136          0.6667657          0.2998503          0.6306386
#> 3         0.06550995          0.4509828          0.1559745          0.4267428