Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1526245725
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0388   17.2975    0.0000 [ 0.6056; 0.7344 ] 
#>   eta3 ~ eta1      0.4585      0.0904    5.0731    0.0000 [ 0.3724; 0.6676 ] 
#>   eta3 ~ eta2      0.3052      0.1009    3.0246    0.0025 [ 0.0279; 0.4030 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0320   20.6953    0.0000 [ 0.6034; 0.7176 ] 
#>   eta1 =~ y12      0.6493      0.0364   17.8565    0.0000 [ 0.5810; 0.7114 ] 
#>   eta1 =~ y13      0.7613      0.0271   28.1146    0.0000 [ 0.7260; 0.8088 ] 
#>   eta2 =~ y21      0.5165      0.0646    7.9943    0.0000 [ 0.3980; 0.6095 ] 
#>   eta2 =~ y22      0.7554      0.0374   20.1906    0.0000 [ 0.6864; 0.8095 ] 
#>   eta2 =~ y23      0.7997      0.0396   20.1689    0.0000 [ 0.7322; 0.8728 ] 
#>   eta3 =~ y31      0.8223      0.0299   27.5260    0.0000 [ 0.7668; 0.8727 ] 
#>   eta3 =~ y32      0.6581      0.0467   14.1050    0.0000 [ 0.5539; 0.7157 ] 
#>   eta3 =~ y33      0.7474      0.0405   18.4448    0.0000 [ 0.6879; 0.8345 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0188   20.9908    0.0000 [ 0.3604; 0.4256 ] 
#>   eta1 <~ y12      0.3873      0.0173   22.3678    0.0000 [ 0.3557; 0.4210 ] 
#>   eta1 <~ y13      0.4542      0.0155   29.3569    0.0000 [ 0.4245; 0.4790 ] 
#>   eta2 <~ y21      0.3058      0.0330    9.2540    0.0000 [ 0.2417; 0.3588 ] 
#>   eta2 <~ y22      0.4473      0.0236   18.9488    0.0000 [ 0.4185; 0.4985 ] 
#>   eta2 <~ y23      0.4735      0.0238   19.8858    0.0000 [ 0.4398; 0.5278 ] 
#>   eta3 <~ y31      0.4400      0.0169   26.0472    0.0000 [ 0.4051; 0.4709 ] 
#>   eta3 <~ y32      0.3521      0.0235   14.9954    0.0000 [ 0.3013; 0.3819 ] 
#>   eta3 <~ y33      0.3999      0.0200   19.9732    0.0000 [ 0.3704; 0.4359 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0388   17.2975    0.0000 [ 0.6056; 0.7344 ] 
#>   eta3 ~ eta1       0.6634      0.0376   17.6397    0.0000 [ 0.6126; 0.7365 ] 
#>   eta3 ~ eta2       0.3052      0.1009    3.0246    0.0025 [ 0.0279; 0.4030 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0671    3.0539    0.0023 [ 0.0197; 0.2769 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03203959 20.695331  3.816018e-95
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03636093 17.856470  2.574136e-71
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02708005 28.114641 6.487857e-174
#> 4 eta2 =~ y21  Common factor 0.5164548 0.06460300  7.994286  1.303274e-15
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03741293 20.190551  1.185344e-90
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03964830 20.168929  1.835704e-90
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02987271 27.526042 8.567491e-167
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04665514 14.104960  3.539875e-45
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04052226 18.444779  5.743332e-76
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6033785          0.7175863
#> 2          0.5810458          0.7114252
#> 3          0.7260092          0.8087776
#> 4          0.3980113          0.6094516
#> 5          0.6864483          0.8095002
#> 6          0.7322187          0.8727726
#> 7          0.7668412          0.8726811
#> 8          0.5538747          0.7156549
#> 9          0.6878516          0.8345187

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1526245725
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0388   17.2975    0.0000 [ 0.5626; 0.7634 ] 
#>   eta3 ~ eta1      0.4585      0.0904    5.0731    0.0000 [ 0.1859; 0.6533 ] 
#>   eta3 ~ eta2      0.3052      0.1009    3.0246    0.0025 [ 0.0780; 0.5997 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0320   20.6953    0.0000 [ 0.5786; 0.7443 ] 
#>   eta1 =~ y12      0.6493      0.0364   17.8565    0.0000 [ 0.5549; 0.7430 ] 
#>   eta1 =~ y13      0.7613      0.0271   28.1146    0.0000 [ 0.6906; 0.8306 ] 
#>   eta2 =~ y21      0.5165      0.0646    7.9943    0.0000 [ 0.3617; 0.6958 ] 
#>   eta2 =~ y22      0.7554      0.0374   20.1906    0.0000 [ 0.6627; 0.8562 ] 
#>   eta2 =~ y23      0.7997      0.0396   20.1689    0.0000 [ 0.6878; 0.8928 ] 
#>   eta3 =~ y31      0.8223      0.0299   27.5260    0.0000 [ 0.7428; 0.8973 ] 
#>   eta3 =~ y32      0.6581      0.0467   14.1050    0.0000 [ 0.5538; 0.7951 ] 
#>   eta3 =~ y33      0.7474      0.0405   18.4448    0.0000 [ 0.6416; 0.8512 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0188   20.9908    0.0000 [ 0.3467; 0.4442 ] 
#>   eta1 <~ y12      0.3873      0.0173   22.3678    0.0000 [ 0.3434; 0.4329 ] 
#>   eta1 <~ y13      0.4542      0.0155   29.3569    0.0000 [ 0.4147; 0.4947 ] 
#>   eta2 <~ y21      0.3058      0.0330    9.2540    0.0000 [ 0.2275; 0.3984 ] 
#>   eta2 <~ y22      0.4473      0.0236   18.9488    0.0000 [ 0.3875; 0.5096 ] 
#>   eta2 <~ y23      0.4735      0.0238   19.8858    0.0000 [ 0.4052; 0.5284 ] 
#>   eta3 <~ y31      0.4400      0.0169   26.0472    0.0000 [ 0.3916; 0.4790 ] 
#>   eta3 <~ y32      0.3521      0.0235   14.9954    0.0000 [ 0.2976; 0.4190 ] 
#>   eta3 <~ y33      0.3999      0.0200   19.9732    0.0000 [ 0.3445; 0.4481 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0388   17.2975    0.0000 [ 0.5626; 0.7634 ] 
#>   eta3 ~ eta1       0.6634      0.0376   17.6397    0.0000 [ 0.5486; 0.7431 ] 
#>   eta3 ~ eta2       0.3052      0.1009    3.0246    0.0025 [ 0.0780; 0.5997 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0671    3.0539    0.0023 [ 0.0527; 0.3997 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03881090 17.297547 4.908504e-67
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09037988  5.073107 3.913730e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.10088809  3.024649 2.489214e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5626476          0.7633557          0.5867487          0.7392546
#> 2          0.1859205          0.6533142          0.2420454          0.5971894
#> 3          0.0779768          0.5997131          0.1406271          0.5370627
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.59709039          0.7458174         0.60556895          0.7343557
#> 2         0.28481330          0.7064137         0.37240651          0.6675707
#> 3         0.02712046          0.4701792         0.02789709          0.4029813