Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1946697912
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0377   17.7936    0.0000 [ 0.6133; 0.7304 ] 
#>   eta3 ~ eta1      0.4585      0.0532    8.6242    0.0000 [ 0.3936; 0.5641 ] 
#>   eta3 ~ eta2      0.3052      0.0507    6.0167    0.0000 [ 0.2139; 0.3977 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0421   15.7633    0.0000 [ 0.5607; 0.7605 ] 
#>   eta1 =~ y12      0.6493      0.0384   16.9160    0.0000 [ 0.5849; 0.7364 ] 
#>   eta1 =~ y13      0.7613      0.0275   27.6540    0.0000 [ 0.7153; 0.7998 ] 
#>   eta2 =~ y21      0.5165      0.0629    8.2073    0.0000 [ 0.3541; 0.5781 ] 
#>   eta2 =~ y22      0.7554      0.0372   20.3086    0.0000 [ 0.6753; 0.8219 ] 
#>   eta2 =~ y23      0.7997      0.0370   21.6288    0.0000 [ 0.7361; 0.8614 ] 
#>   eta3 =~ y31      0.8223      0.0384   21.4238    0.0000 [ 0.7548; 0.8796 ] 
#>   eta3 =~ y32      0.6581      0.0377   17.4705    0.0000 [ 0.6103; 0.7368 ] 
#>   eta3 =~ y33      0.7474      0.0354   21.0989    0.0000 [ 0.6765; 0.8056 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0210   18.8779    0.0000 [ 0.3501; 0.4334 ] 
#>   eta1 <~ y12      0.3873      0.0227   17.0531    0.0000 [ 0.3452; 0.4245 ] 
#>   eta1 <~ y13      0.4542      0.0150   30.2865    0.0000 [ 0.4276; 0.4793 ] 
#>   eta2 <~ y21      0.3058      0.0328    9.3320    0.0000 [ 0.2356; 0.3592 ] 
#>   eta2 <~ y22      0.4473      0.0230   19.4839    0.0000 [ 0.4148; 0.5010 ] 
#>   eta2 <~ y23      0.4735      0.0245   19.3510    0.0000 [ 0.4443; 0.5318 ] 
#>   eta3 <~ y31      0.4400      0.0154   28.4833    0.0000 [ 0.4089; 0.4641 ] 
#>   eta3 <~ y32      0.3521      0.0203   17.3704    0.0000 [ 0.3222; 0.3981 ] 
#>   eta3 <~ y33      0.3999      0.0204   19.5610    0.0000 [ 0.3671; 0.4392 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0377   17.7936    0.0000 [ 0.6133; 0.7304 ] 
#>   eta3 ~ eta1       0.6634      0.0318   20.8833    0.0000 [ 0.6206; 0.7310 ] 
#>   eta3 ~ eta2       0.3052      0.0507    6.0167    0.0000 [ 0.2139; 0.3977 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0368    5.5734    0.0000 [ 0.1366; 0.2587 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04206425 15.763262  5.568314e-56
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03838237 16.916047  3.426722e-64
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02753111 27.654014 2.497219e-168
#> 4 eta2 =~ y21  Common factor 0.5164548 0.06292614  8.207317  2.261857e-16
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03719548 20.308590  1.079604e-91
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03697224 21.628760 9.633860e-104
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03838152 21.423785 8.019914e-102
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03766739 17.470522  2.402857e-68
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03542486 21.098858  8.147972e-99
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5607196          0.7605174
#> 2          0.5849230          0.7364132
#> 3          0.7152822          0.7998156
#> 4          0.3541331          0.5780502
#> 5          0.6753176          0.8219275
#> 6          0.7360728          0.8614128
#> 7          0.7547776          0.8796007
#> 8          0.6102913          0.7367623
#> 9          0.6764994          0.8056033

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1946697912
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0377   17.7936    0.0000 [ 0.5779; 0.7730 ] 
#>   eta3 ~ eta1      0.4585      0.0532    8.6242    0.0000 [ 0.3118; 0.5867 ] 
#>   eta3 ~ eta2      0.3052      0.0507    6.0167    0.0000 [ 0.1738; 0.4360 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0421   15.7633    0.0000 [ 0.5569; 0.7745 ] 
#>   eta1 =~ y12      0.6493      0.0384   16.9160    0.0000 [ 0.5419; 0.7404 ] 
#>   eta1 =~ y13      0.7613      0.0275   27.6540    0.0000 [ 0.6919; 0.8342 ] 
#>   eta2 =~ y21      0.5165      0.0629    8.2073    0.0000 [ 0.3676; 0.6930 ] 
#>   eta2 =~ y22      0.7554      0.0372   20.3086    0.0000 [ 0.6597; 0.8521 ] 
#>   eta2 =~ y23      0.7997      0.0370   21.6288    0.0000 [ 0.7066; 0.8978 ] 
#>   eta3 =~ y31      0.8223      0.0384   21.4238    0.0000 [ 0.7306; 0.9291 ] 
#>   eta3 =~ y32      0.6581      0.0377   17.4705    0.0000 [ 0.5619; 0.7567 ] 
#>   eta3 =~ y33      0.7474      0.0354   21.0989    0.0000 [ 0.6460; 0.8292 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0210   18.8779    0.0000 [ 0.3443; 0.4527 ] 
#>   eta1 <~ y12      0.3873      0.0227   17.0531    0.0000 [ 0.3249; 0.4423 ] 
#>   eta1 <~ y13      0.4542      0.0150   30.2865    0.0000 [ 0.4178; 0.4954 ] 
#>   eta2 <~ y21      0.3058      0.0328    9.3320    0.0000 [ 0.2271; 0.3965 ] 
#>   eta2 <~ y22      0.4473      0.0230   19.4839    0.0000 [ 0.3839; 0.5026 ] 
#>   eta2 <~ y23      0.4735      0.0245   19.3510    0.0000 [ 0.4071; 0.5336 ] 
#>   eta3 <~ y31      0.4400      0.0154   28.4833    0.0000 [ 0.4048; 0.4846 ] 
#>   eta3 <~ y32      0.3521      0.0203   17.3704    0.0000 [ 0.3007; 0.4056 ] 
#>   eta3 <~ y33      0.3999      0.0204   19.5610    0.0000 [ 0.3422; 0.4479 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0377   17.7936    0.0000 [ 0.5779; 0.7730 ] 
#>   eta3 ~ eta1       0.6634      0.0318   20.8833    0.0000 [ 0.5728; 0.7371 ] 
#>   eta3 ~ eta2       0.3052      0.0507    6.0167    0.0000 [ 0.1738; 0.4360 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0368    5.5734    0.0000 [ 0.1107; 0.3007 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03772885 17.793637 7.917090e-71
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.05316488  8.624241 6.451885e-18
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.05071713  6.016727 1.779793e-09
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5778685          0.7729808          0.6012976          0.7495516
#> 2          0.3117551          0.5866939          0.3447699          0.5536791
#> 3          0.1737638          0.4360442          0.2052585          0.4045494
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5823429          0.7455877          0.6133298          0.7304030
#> 2          0.3813425          0.5674400          0.3936320          0.5640709
#> 3          0.1937396          0.3993976          0.2138550          0.3977181