Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -630203846
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0336   19.9685    0.0000 [ 0.6086; 0.7319 ] 
#>   eta3 ~ eta1      0.4585      0.0777    5.9015    0.0000 [ 0.3018; 0.5594 ] 
#>   eta3 ~ eta2      0.3052      0.0748    4.0805    0.0000 [ 0.1813; 0.4477 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0419   15.8271    0.0000 [ 0.5915; 0.7345 ] 
#>   eta1 =~ y12      0.6493      0.0346   18.7914    0.0000 [ 0.5833; 0.6990 ] 
#>   eta1 =~ y13      0.7613      0.0308   24.7475    0.0000 [ 0.7051; 0.8026 ] 
#>   eta2 =~ y21      0.5165      0.0450   11.4874    0.0000 [ 0.4604; 0.5878 ] 
#>   eta2 =~ y22      0.7554      0.0354   21.3173    0.0000 [ 0.6898; 0.8299 ] 
#>   eta2 =~ y23      0.7997      0.0402   19.9009    0.0000 [ 0.7249; 0.8599 ] 
#>   eta3 =~ y31      0.8223      0.0326   25.2478    0.0000 [ 0.7663; 0.8917 ] 
#>   eta3 =~ y32      0.6581      0.0393   16.7494    0.0000 [ 0.5728; 0.7177 ] 
#>   eta3 =~ y33      0.7474      0.0351   21.2910    0.0000 [ 0.6998; 0.8057 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0179   22.1374    0.0000 [ 0.3671; 0.4312 ] 
#>   eta1 <~ y12      0.3873      0.0185   20.8924    0.0000 [ 0.3546; 0.4121 ] 
#>   eta1 <~ y13      0.4542      0.0221   20.5487    0.0000 [ 0.4069; 0.4902 ] 
#>   eta2 <~ y21      0.3058      0.0228   13.4357    0.0000 [ 0.2787; 0.3482 ] 
#>   eta2 <~ y22      0.4473      0.0209   21.4439    0.0000 [ 0.4094; 0.4844 ] 
#>   eta2 <~ y23      0.4735      0.0230   20.5799    0.0000 [ 0.4358; 0.5100 ] 
#>   eta3 <~ y31      0.4400      0.0161   27.2869    0.0000 [ 0.4151; 0.4742 ] 
#>   eta3 <~ y32      0.3521      0.0180   19.5599    0.0000 [ 0.3163; 0.3756 ] 
#>   eta3 <~ y33      0.3999      0.0185   21.6340    0.0000 [ 0.3688; 0.4397 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0336   19.9685    0.0000 [ 0.6086; 0.7319 ] 
#>   eta3 ~ eta1       0.6634      0.0417   15.9055    0.0000 [ 0.5679; 0.7196 ] 
#>   eta3 ~ eta2       0.3052      0.0748    4.0805    0.0000 [ 0.1813; 0.4477 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0515    3.9755    0.0001 [ 0.1202; 0.3128 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04189456 15.82711  2.022915e-56
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03455183 18.79142  8.877485e-79
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03076460 24.74747 3.301249e-135
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04495845 11.48738  1.526705e-30
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03543538 21.31733 7.840459e-101
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04018231 19.90089  3.997635e-88
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03256834 25.24775 1.198789e-140
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03928911 16.74940  5.720246e-63
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03510520 21.29098 1.376113e-100
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5915219          0.7345370
#> 2          0.5833496          0.6990337
#> 3          0.7050519          0.8026164
#> 4          0.4603535          0.5878419
#> 5          0.6898314          0.8299366
#> 6          0.7248647          0.8598604
#> 7          0.7663318          0.8917377
#> 8          0.5727659          0.7177460
#> 9          0.6998081          0.8056785

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -630203846
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0336   19.9685    0.0000 [ 0.5914; 0.7653 ] 
#>   eta3 ~ eta1      0.4585      0.0777    5.9015    0.0000 [ 0.2838; 0.6856 ] 
#>   eta3 ~ eta2      0.3052      0.0748    4.0805    0.0000 [ 0.0881; 0.4748 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0419   15.8271    0.0000 [ 0.5501; 0.7667 ] 
#>   eta1 =~ y12      0.6493      0.0346   18.7914    0.0000 [ 0.5677; 0.7464 ] 
#>   eta1 =~ y13      0.7613      0.0308   24.7475    0.0000 [ 0.6838; 0.8429 ] 
#>   eta2 =~ y21      0.5165      0.0450   11.4874    0.0000 [ 0.3857; 0.6182 ] 
#>   eta2 =~ y22      0.7554      0.0354   21.3173    0.0000 [ 0.6636; 0.8469 ] 
#>   eta2 =~ y23      0.7997      0.0402   19.9009    0.0000 [ 0.7045; 0.9123 ] 
#>   eta3 =~ y31      0.8223      0.0326   25.2478    0.0000 [ 0.7394; 0.9079 ] 
#>   eta3 =~ y32      0.6581      0.0393   16.7494    0.0000 [ 0.5581; 0.7613 ] 
#>   eta3 =~ y33      0.7474      0.0351   21.2910    0.0000 [ 0.6487; 0.8302 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0179   22.1374    0.0000 [ 0.3456; 0.4380 ] 
#>   eta1 <~ y12      0.3873      0.0185   20.8924    0.0000 [ 0.3428; 0.4387 ] 
#>   eta1 <~ y13      0.4542      0.0221   20.5487    0.0000 [ 0.3965; 0.5108 ] 
#>   eta2 <~ y21      0.3058      0.0228   13.4357    0.0000 [ 0.2396; 0.3574 ] 
#>   eta2 <~ y22      0.4473      0.0209   21.4439    0.0000 [ 0.3946; 0.5025 ] 
#>   eta2 <~ y23      0.4735      0.0230   20.5799    0.0000 [ 0.4207; 0.5397 ] 
#>   eta3 <~ y31      0.4400      0.0161   27.2869    0.0000 [ 0.4007; 0.4841 ] 
#>   eta3 <~ y32      0.3521      0.0180   19.5599    0.0000 [ 0.3079; 0.4010 ] 
#>   eta3 <~ y33      0.3999      0.0185   21.6340    0.0000 [ 0.3494; 0.4450 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0336   19.9685    0.0000 [ 0.5914; 0.7653 ] 
#>   eta3 ~ eta1       0.6634      0.0417   15.9055    0.0000 [ 0.5680; 0.7837 ] 
#>   eta3 ~ eta2       0.3052      0.0748    4.0805    0.0000 [ 0.0881; 0.4748 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0515    3.9755    0.0001 [ 0.0579; 0.3244 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03361968 19.968464 1.035903e-88
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07769289  5.901528 3.601501e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07478270  4.080504 4.493808e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.59139936          0.7652614          0.6122768          0.7443839
#> 2         0.28380377          0.6855875          0.3320501          0.6373411
#> 3         0.08809221          0.4748261          0.1345314          0.4283869
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5913886          0.7389201          0.6085580          0.7318669
#> 2          0.2577936          0.5907062          0.3017692          0.5593890
#> 3          0.1737077          0.4827048          0.1812951          0.4477429