Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1471703724
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0551   12.1909    0.0000 [ 0.5583; 0.7487 ] 
#>   eta3 ~ eta1      0.4585      0.0793    5.7841    0.0000 [ 0.3423; 0.6297 ] 
#>   eta3 ~ eta2      0.3052      0.0888    3.4369    0.0006 [ 0.1244; 0.4319 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0433   15.3245    0.0000 [ 0.6062; 0.7500 ] 
#>   eta1 =~ y12      0.6493      0.0429   15.1256    0.0000 [ 0.5713; 0.7211 ] 
#>   eta1 =~ y13      0.7613      0.0304   25.0729    0.0000 [ 0.6828; 0.7966 ] 
#>   eta2 =~ y21      0.5165      0.0738    6.9940    0.0000 [ 0.3742; 0.6431 ] 
#>   eta2 =~ y22      0.7554      0.0385   19.6316    0.0000 [ 0.6891; 0.8282 ] 
#>   eta2 =~ y23      0.7997      0.0379   21.0977    0.0000 [ 0.7121; 0.8447 ] 
#>   eta3 =~ y31      0.8223      0.0332   24.7747    0.0000 [ 0.7553; 0.8706 ] 
#>   eta3 =~ y32      0.6581      0.0490   13.4358    0.0000 [ 0.5486; 0.7396 ] 
#>   eta3 =~ y33      0.7474      0.0358   20.8673    0.0000 [ 0.6935; 0.8388 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0226   17.4667    0.0000 [ 0.3657; 0.4395 ] 
#>   eta1 <~ y12      0.3873      0.0208   18.6555    0.0000 [ 0.3560; 0.4291 ] 
#>   eta1 <~ y13      0.4542      0.0185   24.4901    0.0000 [ 0.4240; 0.4853 ] 
#>   eta2 <~ y21      0.3058      0.0389    7.8573    0.0000 [ 0.2342; 0.3693 ] 
#>   eta2 <~ y22      0.4473      0.0256   17.4433    0.0000 [ 0.4080; 0.4859 ] 
#>   eta2 <~ y23      0.4735      0.0224   21.1557    0.0000 [ 0.4364; 0.5087 ] 
#>   eta3 <~ y31      0.4400      0.0199   22.1222    0.0000 [ 0.4006; 0.4638 ] 
#>   eta3 <~ y32      0.3521      0.0220   16.0024    0.0000 [ 0.3060; 0.3942 ] 
#>   eta3 <~ y33      0.3999      0.0183   21.9108    0.0000 [ 0.3720; 0.4386 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0551   12.1909    0.0000 [ 0.5583; 0.7487 ] 
#>   eta3 ~ eta1       0.6634      0.0397   16.7148    0.0000 [ 0.5929; 0.7317 ] 
#>   eta3 ~ eta2       0.3052      0.0888    3.4369    0.0006 [ 0.1244; 0.4319 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0581    3.5263    0.0004 [ 0.0855; 0.2828 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04326866 15.324484  5.246928e-53
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04292586 15.125565  1.098536e-51
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03036534 25.072856 9.836635e-139
#> 4 eta2 =~ y21  Common factor 0.5164548 0.07384244  6.994011  2.671365e-12
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03847810 19.631626  8.301455e-86
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03790280 21.097746  8.341808e-99
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03319025 24.774663 1.681685e-135
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04897877 13.435799  3.730187e-41
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03581792 20.867325  1.061066e-96
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6062297          0.7499976
#> 2          0.5713384          0.7210524
#> 3          0.6828367          0.7966343
#> 4          0.3742485          0.6431213
#> 5          0.6891273          0.8281872
#> 6          0.7121261          0.8446793
#> 7          0.7553163          0.8706100
#> 8          0.5486368          0.7396070
#> 9          0.6935161          0.8388498

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1471703724
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0551   12.1909    0.0000 [ 0.5260; 0.8108 ] 
#>   eta3 ~ eta1      0.4585      0.0793    5.7841    0.0000 [ 0.2452; 0.6551 ] 
#>   eta3 ~ eta2      0.3052      0.0888    3.4369    0.0006 [ 0.0838; 0.5429 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0433   15.3245    0.0000 [ 0.5469; 0.7706 ] 
#>   eta1 =~ y12      0.6493      0.0429   15.1256    0.0000 [ 0.5402; 0.7622 ] 
#>   eta1 =~ y13      0.7613      0.0304   25.0729    0.0000 [ 0.6907; 0.8477 ] 
#>   eta2 =~ y21      0.5165      0.0738    6.9940    0.0000 [ 0.3224; 0.7043 ] 
#>   eta2 =~ y22      0.7554      0.0385   19.6316    0.0000 [ 0.6589; 0.8579 ] 
#>   eta2 =~ y23      0.7997      0.0379   21.0977    0.0000 [ 0.7057; 0.9017 ] 
#>   eta3 =~ y31      0.8223      0.0332   24.7747    0.0000 [ 0.7431; 0.9147 ] 
#>   eta3 =~ y32      0.6581      0.0490   13.4358    0.0000 [ 0.5252; 0.7785 ] 
#>   eta3 =~ y33      0.7474      0.0358   20.8673    0.0000 [ 0.6541; 0.8393 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0226   17.4667    0.0000 [ 0.3330; 0.4501 ] 
#>   eta1 <~ y12      0.3873      0.0208   18.6555    0.0000 [ 0.3336; 0.4410 ] 
#>   eta1 <~ y13      0.4542      0.0185   24.4901    0.0000 [ 0.4091; 0.5050 ] 
#>   eta2 <~ y21      0.3058      0.0389    7.8573    0.0000 [ 0.2035; 0.4048 ] 
#>   eta2 <~ y22      0.4473      0.0256   17.4433    0.0000 [ 0.3818; 0.5145 ] 
#>   eta2 <~ y23      0.4735      0.0224   21.1557    0.0000 [ 0.4172; 0.5330 ] 
#>   eta3 <~ y31      0.4400      0.0199   22.1222    0.0000 [ 0.3921; 0.4950 ] 
#>   eta3 <~ y32      0.3521      0.0220   16.0024    0.0000 [ 0.2923; 0.4061 ] 
#>   eta3 <~ y33      0.3999      0.0183   21.9108    0.0000 [ 0.3525; 0.4469 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0551   12.1909    0.0000 [ 0.5260; 0.8108 ] 
#>   eta3 ~ eta1       0.6634      0.0397   16.7148    0.0000 [ 0.5585; 0.7637 ] 
#>   eta3 ~ eta2       0.3052      0.0888    3.4369    0.0006 [ 0.0838; 0.5429 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0581    3.5263    0.0004 [ 0.0607; 0.3612 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05506844 12.190893 3.475910e-34
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07926976  5.784132 7.288769e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08878744  3.436873 5.884723e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.52599161          0.8107745          0.5601885          0.7765776
#> 2         0.24516161          0.6551001          0.2943872          0.6058745
#> 3         0.08376554          0.5429241          0.1389015          0.4877881
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5060371          0.7805741          0.5582873          0.7486773
#> 2          0.3314429          0.6397281          0.3422506          0.6296933
#> 3          0.1094755          0.4854821          0.1244149          0.4318684