Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1040488065
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0514   13.0586    0.0000 [ 0.5924; 0.7360 ] 
#>   eta3 ~ eta1      0.4585      0.0914    5.0167    0.0000 [ 0.2851; 0.6367 ] 
#>   eta3 ~ eta2      0.3052      0.0969    3.1493    0.0016 [ 0.1280; 0.4974 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0319   20.8148    0.0000 [ 0.6028; 0.7100 ] 
#>   eta1 =~ y12      0.6493      0.0405   16.0121    0.0000 [ 0.5983; 0.7376 ] 
#>   eta1 =~ y13      0.7613      0.0292   26.0936    0.0000 [ 0.7127; 0.8024 ] 
#>   eta2 =~ y21      0.5165      0.0521    9.9065    0.0000 [ 0.4036; 0.5924 ] 
#>   eta2 =~ y22      0.7554      0.0395   19.1118    0.0000 [ 0.6974; 0.8281 ] 
#>   eta2 =~ y23      0.7997      0.0387   20.6619    0.0000 [ 0.7203; 0.8664 ] 
#>   eta3 =~ y31      0.8223      0.0375   21.9371    0.0000 [ 0.7603; 0.8820 ] 
#>   eta3 =~ y32      0.6581      0.0505   13.0275    0.0000 [ 0.5770; 0.7460 ] 
#>   eta3 =~ y33      0.7474      0.0372   20.0716    0.0000 [ 0.6827; 0.8227 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0175   22.5622    0.0000 [ 0.3654; 0.4268 ] 
#>   eta1 <~ y12      0.3873      0.0194   19.9510    0.0000 [ 0.3580; 0.4221 ] 
#>   eta1 <~ y13      0.4542      0.0180   25.1736    0.0000 [ 0.4187; 0.4778 ] 
#>   eta2 <~ y21      0.3058      0.0278   10.9954    0.0000 [ 0.2439; 0.3546 ] 
#>   eta2 <~ y22      0.4473      0.0220   20.3395    0.0000 [ 0.4188; 0.4893 ] 
#>   eta2 <~ y23      0.4735      0.0219   21.6370    0.0000 [ 0.4398; 0.5093 ] 
#>   eta3 <~ y31      0.4400      0.0218   20.1528    0.0000 [ 0.4042; 0.4735 ] 
#>   eta3 <~ y32      0.3521      0.0204   17.2298    0.0000 [ 0.3121; 0.3808 ] 
#>   eta3 <~ y33      0.3999      0.0208   19.2383    0.0000 [ 0.3691; 0.4388 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0514   13.0586    0.0000 [ 0.5924; 0.7360 ] 
#>   eta3 ~ eta1       0.6634      0.0400   16.5797    0.0000 [ 0.5978; 0.7312 ] 
#>   eta3 ~ eta2       0.3052      0.0969    3.1493    0.0016 [ 0.1280; 0.4974 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0675    3.0361    0.0024 [ 0.0831; 0.3327 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03185563 20.814844  3.175770e-96
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04054927 16.012075  1.052400e-57
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02917746 26.093629 4.306253e-150
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05213307  9.906473  3.901775e-23
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03952477 19.111754  2.015792e-81
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03870228 20.661927  7.625887e-95
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03748346 21.937071 1.150780e-106
#> 8 eta3 =~ y32  Common factor 0.6580689 0.05051397 13.027464  8.539895e-39
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03723781 20.071644  1.306130e-89
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6028361          0.7099638
#> 2          0.5982593          0.7375823
#> 3          0.7127151          0.8024476
#> 4          0.4036081          0.5923555
#> 5          0.6973599          0.8281455
#> 6          0.7202848          0.8664270
#> 7          0.7603097          0.8820019
#> 8          0.5769686          0.7460221
#> 9          0.6826769          0.8226731

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1040488065
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0514   13.0586    0.0000 [ 0.5437; 0.8096 ] 
#>   eta3 ~ eta1      0.4585      0.0914    5.0167    0.0000 [ 0.2244; 0.6970 ] 
#>   eta3 ~ eta2      0.3052      0.0969    3.1493    0.0016 [ 0.0524; 0.5535 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0319   20.8148    0.0000 [ 0.5905; 0.7552 ] 
#>   eta1 =~ y12      0.6493      0.0405   16.0121    0.0000 [ 0.5352; 0.7449 ] 
#>   eta1 =~ y13      0.7613      0.0292   26.0936    0.0000 [ 0.6896; 0.8405 ] 
#>   eta2 =~ y21      0.5165      0.0521    9.9065    0.0000 [ 0.3998; 0.6694 ] 
#>   eta2 =~ y22      0.7554      0.0395   19.1118    0.0000 [ 0.6493; 0.8537 ] 
#>   eta2 =~ y23      0.7997      0.0387   20.6619    0.0000 [ 0.7094; 0.9095 ] 
#>   eta3 =~ y31      0.8223      0.0375   21.9371    0.0000 [ 0.7273; 0.9212 ] 
#>   eta3 =~ y32      0.6581      0.0505   13.0275    0.0000 [ 0.5289; 0.7902 ] 
#>   eta3 =~ y33      0.7474      0.0372   20.0716    0.0000 [ 0.6458; 0.8383 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0175   22.5622    0.0000 [ 0.3550; 0.4456 ] 
#>   eta1 <~ y12      0.3873      0.0194   19.9510    0.0000 [ 0.3307; 0.4311 ] 
#>   eta1 <~ y13      0.4542      0.0180   25.1736    0.0000 [ 0.4084; 0.5017 ] 
#>   eta2 <~ y21      0.3058      0.0278   10.9954    0.0000 [ 0.2406; 0.3844 ] 
#>   eta2 <~ y22      0.4473      0.0220   20.3395    0.0000 [ 0.3814; 0.4951 ] 
#>   eta2 <~ y23      0.4735      0.0219   21.6370    0.0000 [ 0.4157; 0.5289 ] 
#>   eta3 <~ y31      0.4400      0.0218   20.1528    0.0000 [ 0.3852; 0.4982 ] 
#>   eta3 <~ y32      0.3521      0.0204   17.2298    0.0000 [ 0.3011; 0.4068 ] 
#>   eta3 <~ y33      0.3999      0.0208   19.2383    0.0000 [ 0.3439; 0.4514 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0514   13.0586    0.0000 [ 0.5437; 0.8096 ] 
#>   eta3 ~ eta1       0.6634      0.0400   16.5797    0.0000 [ 0.5620; 0.7689 ] 
#>   eta3 ~ eta2       0.3052      0.0969    3.1493    0.0016 [ 0.0524; 0.5535 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0675    3.0361    0.0024 [ 0.0302; 0.3792 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05140944 13.058564 5.678884e-39
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09139627  5.016690 5.256923e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09689471  3.149306 1.636585e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54374567          0.8096063          0.5756703          0.7776816
#> 2         0.22439219          0.6970421          0.2811482          0.6402861
#> 3         0.05241746          0.5535022          0.1125879          0.4933317
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5755405          0.7703151          0.5924432          0.7360363
#> 2          0.2697679          0.6427429          0.2850871          0.6366500
#> 3          0.1103872          0.5025674          0.1279996          0.4974020