Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1281271222
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0433   15.5041    0.0000 [ 0.5904; 0.7541 ] 
#>   eta3 ~ eta1      0.4585      0.0917    4.9987    0.0000 [ 0.2876; 0.6431 ] 
#>   eta3 ~ eta2      0.3052      0.0944    3.2319    0.0012 [ 0.1050; 0.4597 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0469   14.1233    0.0000 [ 0.5771; 0.7262 ] 
#>   eta1 =~ y12      0.6493      0.0432   15.0224    0.0000 [ 0.5677; 0.7302 ] 
#>   eta1 =~ y13      0.7613      0.0306   24.8511    0.0000 [ 0.7139; 0.8229 ] 
#>   eta2 =~ y21      0.5165      0.0570    9.0569    0.0000 [ 0.3908; 0.6095 ] 
#>   eta2 =~ y22      0.7554      0.0359   21.0471    0.0000 [ 0.6806; 0.8080 ] 
#>   eta2 =~ y23      0.7997      0.0430   18.6083    0.0000 [ 0.7159; 0.8531 ] 
#>   eta3 =~ y31      0.8223      0.0308   26.6921    0.0000 [ 0.7652; 0.8743 ] 
#>   eta3 =~ y32      0.6581      0.0407   16.1650    0.0000 [ 0.5887; 0.7184 ] 
#>   eta3 =~ y33      0.7474      0.0399   18.7466    0.0000 [ 0.6847; 0.8303 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0218   18.1418    0.0000 [ 0.3543; 0.4332 ] 
#>   eta1 <~ y12      0.3873      0.0234   16.5315    0.0000 [ 0.3379; 0.4279 ] 
#>   eta1 <~ y13      0.4542      0.0222   20.4541    0.0000 [ 0.4233; 0.4976 ] 
#>   eta2 <~ y21      0.3058      0.0316    9.6627    0.0000 [ 0.2463; 0.3595 ] 
#>   eta2 <~ y22      0.4473      0.0230   19.4513    0.0000 [ 0.4096; 0.4868 ] 
#>   eta2 <~ y23      0.4735      0.0222   21.3042    0.0000 [ 0.4354; 0.5199 ] 
#>   eta3 <~ y31      0.4400      0.0203   21.6255    0.0000 [ 0.4046; 0.4759 ] 
#>   eta3 <~ y32      0.3521      0.0189   18.6390    0.0000 [ 0.3216; 0.3847 ] 
#>   eta3 <~ y33      0.3999      0.0168   23.8283    0.0000 [ 0.3720; 0.4303 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0433   15.5041    0.0000 [ 0.5904; 0.7541 ] 
#>   eta3 ~ eta1       0.6634      0.0398   16.6652    0.0000 [ 0.5851; 0.7303 ] 
#>   eta3 ~ eta2       0.3052      0.0944    3.2319    0.0012 [ 0.1050; 0.4597 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0659    3.1084    0.0019 [ 0.0773; 0.3133 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04694852 14.123340  2.727533e-45
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04322059 15.022422  5.235908e-51
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03063630 24.851098 2.516127e-136
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05702305  9.056948  1.341517e-19
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03589027 21.047146  2.428704e-98
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04297347 18.608313  2.751473e-77
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03080602 26.692103 5.813206e-157
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04070948 16.165002  8.903438e-59
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03986975 18.746648  2.061836e-78
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5771388          0.7261716
#> 2          0.5677086          0.7301627
#> 3          0.7139463          0.8229155
#> 4          0.3907710          0.6095135
#> 5          0.6805863          0.8080212
#> 6          0.7159392          0.8530869
#> 7          0.7651878          0.8742518
#> 8          0.5886639          0.7184258
#> 9          0.6847403          0.8302609

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1281271222
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0433   15.5041    0.0000 [ 0.5585; 0.7825 ] 
#>   eta3 ~ eta1      0.4585      0.0917    4.9987    0.0000 [ 0.2101; 0.6844 ] 
#>   eta3 ~ eta2      0.3052      0.0944    3.2319    0.0012 [ 0.0756; 0.5639 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0469   14.1233    0.0000 [ 0.5458; 0.7886 ] 
#>   eta1 =~ y12      0.6493      0.0432   15.0224    0.0000 [ 0.5434; 0.7669 ] 
#>   eta1 =~ y13      0.7613      0.0306   24.8511    0.0000 [ 0.6868; 0.8453 ] 
#>   eta2 =~ y21      0.5165      0.0570    9.0569    0.0000 [ 0.3747; 0.6696 ] 
#>   eta2 =~ y22      0.7554      0.0359   21.0471    0.0000 [ 0.6608; 0.8464 ] 
#>   eta2 =~ y23      0.7997      0.0430   18.6083    0.0000 [ 0.6948; 0.9170 ] 
#>   eta3 =~ y31      0.8223      0.0308   26.6921    0.0000 [ 0.7472; 0.9065 ] 
#>   eta3 =~ y32      0.6581      0.0407   16.1650    0.0000 [ 0.5532; 0.7637 ] 
#>   eta3 =~ y33      0.7474      0.0399   18.7466    0.0000 [ 0.6431; 0.8493 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0218   18.1418    0.0000 [ 0.3380; 0.4508 ] 
#>   eta1 <~ y12      0.3873      0.0234   16.5315    0.0000 [ 0.3265; 0.4477 ] 
#>   eta1 <~ y13      0.4542      0.0222   20.4541    0.0000 [ 0.3948; 0.5096 ] 
#>   eta2 <~ y21      0.3058      0.0316    9.6627    0.0000 [ 0.2259; 0.3896 ] 
#>   eta2 <~ y22      0.4473      0.0230   19.4513    0.0000 [ 0.3840; 0.5030 ] 
#>   eta2 <~ y23      0.4735      0.0222   21.3042    0.0000 [ 0.4171; 0.5321 ] 
#>   eta3 <~ y31      0.4400      0.0203   21.6255    0.0000 [ 0.3886; 0.4939 ] 
#>   eta3 <~ y32      0.3521      0.0189   18.6390    0.0000 [ 0.3028; 0.4005 ] 
#>   eta3 <~ y33      0.3999      0.0168   23.8283    0.0000 [ 0.3551; 0.4419 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0433   15.5041    0.0000 [ 0.5585; 0.7825 ] 
#>   eta3 ~ eta1       0.6634      0.0398   16.6652    0.0000 [ 0.5587; 0.7646 ] 
#>   eta3 ~ eta2       0.3052      0.0944    3.2319    0.0012 [ 0.0756; 0.5639 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0659    3.1084    0.0019 [ 0.0440; 0.3848 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04330041 15.504089 3.255028e-54
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09172512  4.998705 5.771672e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09441888  3.231887 1.229758e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55854484          0.7824701          0.5854339          0.7555811
#> 2         0.21005690          0.6844074          0.2670171          0.6274472
#> 3         0.07561939          0.5639005          0.1342524          0.5052675
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5770777          0.7585847          0.5903810          0.7540713
#> 2          0.2338423          0.6520932          0.2875715          0.6431483
#> 3          0.0777744          0.4715593          0.1050049          0.4597076