Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1370802402
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0557   12.0542    0.0000 [ 0.5857; 0.7479 ] 
#>   eta3 ~ eta1      0.4585      0.0819    5.6014    0.0000 [ 0.3364; 0.5760 ] 
#>   eta3 ~ eta2      0.3052      0.0828    3.6865    0.0002 [ 0.1598; 0.4442 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0412   16.0828    0.0000 [ 0.5883; 0.7252 ] 
#>   eta1 =~ y12      0.6493      0.0324   20.0368    0.0000 [ 0.5834; 0.6901 ] 
#>   eta1 =~ y13      0.7613      0.0396   19.2280    0.0000 [ 0.6951; 0.8316 ] 
#>   eta2 =~ y21      0.5165      0.0607    8.5061    0.0000 [ 0.3885; 0.6012 ] 
#>   eta2 =~ y22      0.7554      0.0344   21.9800    0.0000 [ 0.6864; 0.7999 ] 
#>   eta2 =~ y23      0.7997      0.0360   22.2381    0.0000 [ 0.7318; 0.8610 ] 
#>   eta3 =~ y31      0.8223      0.0309   26.6443    0.0000 [ 0.7637; 0.8707 ] 
#>   eta3 =~ y32      0.6581      0.0385   17.0922    0.0000 [ 0.6094; 0.7337 ] 
#>   eta3 =~ y33      0.7474      0.0369   20.2525    0.0000 [ 0.6710; 0.7988 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0237   16.6942    0.0000 [ 0.3460; 0.4311 ] 
#>   eta1 <~ y12      0.3873      0.0186   20.8705    0.0000 [ 0.3508; 0.4171 ] 
#>   eta1 <~ y13      0.4542      0.0215   21.1446    0.0000 [ 0.4126; 0.4949 ] 
#>   eta2 <~ y21      0.3058      0.0329    9.2826    0.0000 [ 0.2401; 0.3491 ] 
#>   eta2 <~ y22      0.4473      0.0175   25.4968    0.0000 [ 0.4169; 0.4877 ] 
#>   eta2 <~ y23      0.4735      0.0237   19.9664    0.0000 [ 0.4363; 0.5193 ] 
#>   eta3 <~ y31      0.4400      0.0190   23.1725    0.0000 [ 0.4001; 0.4770 ] 
#>   eta3 <~ y32      0.3521      0.0173   20.3171    0.0000 [ 0.3241; 0.3917 ] 
#>   eta3 <~ y33      0.3999      0.0162   24.7395    0.0000 [ 0.3661; 0.4236 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0557   12.0542    0.0000 [ 0.5857; 0.7479 ] 
#>   eta3 ~ eta1       0.6634      0.0432   15.3393    0.0000 [ 0.5726; 0.7336 ] 
#>   eta3 ~ eta2       0.3052      0.0828    3.6865    0.0002 [ 0.1598; 0.4442 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0529    3.8757    0.0001 [ 0.1160; 0.2971 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04122841 16.082840  3.365877e-58
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03240422 20.036832  2.629838e-89
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03959564 19.228019  2.157119e-82
#> 4 eta2 =~ y21  Common factor 0.5164548 0.06071605  8.506067  1.799332e-17
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03436702 21.980014 4.473203e-107
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03595910 22.238145 1.469066e-109
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03086131 26.644281 2.084765e-156
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03850102 17.092245  1.695209e-65
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03690525 20.252515  3.375569e-91
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5883125          0.7251576
#> 2          0.5833581          0.6901481
#> 3          0.6950705          0.8316074
#> 4          0.3885203          0.6011831
#> 5          0.6864085          0.7999354
#> 6          0.7317770          0.8609666
#> 7          0.7637307          0.8707467
#> 8          0.6094265          0.7337428
#> 9          0.6710070          0.7988118

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1370802402
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0557   12.0542    0.0000 [ 0.5210; 0.8090 ] 
#>   eta3 ~ eta1      0.4585      0.0819    5.6014    0.0000 [ 0.2503; 0.6736 ] 
#>   eta3 ~ eta2      0.3052      0.0828    3.6865    0.0002 [ 0.0969; 0.5250 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0412   16.0828    0.0000 [ 0.5562; 0.7694 ] 
#>   eta1 =~ y12      0.6493      0.0324   20.0368    0.0000 [ 0.5738; 0.7413 ] 
#>   eta1 =~ y13      0.7613      0.0396   19.2280    0.0000 [ 0.6569; 0.8616 ] 
#>   eta2 =~ y21      0.5165      0.0607    8.5061    0.0000 [ 0.3655; 0.6795 ] 
#>   eta2 =~ y22      0.7554      0.0344   21.9800    0.0000 [ 0.6731; 0.8508 ] 
#>   eta2 =~ y23      0.7997      0.0360   22.2381    0.0000 [ 0.7005; 0.8865 ] 
#>   eta3 =~ y31      0.8223      0.0309   26.6443    0.0000 [ 0.7392; 0.8988 ] 
#>   eta3 =~ y32      0.6581      0.0385   17.0922    0.0000 [ 0.5477; 0.7469 ] 
#>   eta3 =~ y33      0.7474      0.0369   20.2525    0.0000 [ 0.6573; 0.8482 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0237   16.6942    0.0000 [ 0.3327; 0.4553 ] 
#>   eta1 <~ y12      0.3873      0.0186   20.8705    0.0000 [ 0.3429; 0.4389 ] 
#>   eta1 <~ y13      0.4542      0.0215   21.1446    0.0000 [ 0.3959; 0.5069 ] 
#>   eta2 <~ y21      0.3058      0.0329    9.2826    0.0000 [ 0.2237; 0.3941 ] 
#>   eta2 <~ y22      0.4473      0.0175   25.4968    0.0000 [ 0.4047; 0.4954 ] 
#>   eta2 <~ y23      0.4735      0.0237   19.9664    0.0000 [ 0.4070; 0.5297 ] 
#>   eta3 <~ y31      0.4400      0.0190   23.1725    0.0000 [ 0.3916; 0.4898 ] 
#>   eta3 <~ y32      0.3521      0.0173   20.3171    0.0000 [ 0.3038; 0.3934 ] 
#>   eta3 <~ y33      0.3999      0.0162   24.7395    0.0000 [ 0.3635; 0.4471 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0557   12.0542    0.0000 [ 0.5210; 0.8090 ] 
#>   eta3 ~ eta1       0.6634      0.0432   15.3393    0.0000 [ 0.5586; 0.7822 ] 
#>   eta3 ~ eta2       0.3052      0.0828    3.6865    0.0002 [ 0.0969; 0.5250 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0529    3.8757    0.0001 [ 0.0718; 0.3451 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05569306 12.054168 1.843856e-33
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08185544  5.601421 2.126015e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08277477  3.686523 2.273387e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.52103473          0.8090478          0.5556195          0.7744630
#> 2         0.25026689          0.6735770          0.3010982          0.6227457
#> 3         0.09694581          0.5250102          0.1483480          0.4736080
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5035462          0.7774594          0.5856580          0.7479330
#> 2          0.2687652          0.5927657          0.3363718          0.5759888
#> 3          0.1540742          0.4827462          0.1597654          0.4442437