Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 474716227
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0474   14.1504    0.0000 [ 0.5903; 0.7419 ] 
#>   eta3 ~ eta1      0.4585      0.0742    6.1763    0.0000 [ 0.2962; 0.5602 ] 
#>   eta3 ~ eta2      0.3052      0.0753    4.0525    0.0001 [ 0.1830; 0.4796 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0375   17.6656    0.0000 [ 0.5883; 0.7255 ] 
#>   eta1 =~ y12      0.6493      0.0399   16.2740    0.0000 [ 0.6003; 0.7288 ] 
#>   eta1 =~ y13      0.7613      0.0383   19.8586    0.0000 [ 0.6927; 0.8258 ] 
#>   eta2 =~ y21      0.5165      0.0434   11.8884    0.0000 [ 0.4472; 0.5959 ] 
#>   eta2 =~ y22      0.7554      0.0305   24.7840    0.0000 [ 0.6874; 0.8044 ] 
#>   eta2 =~ y23      0.7997      0.0310   25.7708    0.0000 [ 0.7337; 0.8466 ] 
#>   eta3 =~ y31      0.8223      0.0350   23.5021    0.0000 [ 0.7485; 0.8701 ] 
#>   eta3 =~ y32      0.6581      0.0478   13.7621    0.0000 [ 0.5893; 0.7281 ] 
#>   eta3 =~ y33      0.7474      0.0398   18.7851    0.0000 [ 0.6878; 0.8219 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0208   18.9748    0.0000 [ 0.3564; 0.4325 ] 
#>   eta1 <~ y12      0.3873      0.0172   22.4578    0.0000 [ 0.3627; 0.4204 ] 
#>   eta1 <~ y13      0.4542      0.0216   21.0028    0.0000 [ 0.4127; 0.4934 ] 
#>   eta2 <~ y21      0.3058      0.0250   12.2283    0.0000 [ 0.2724; 0.3613 ] 
#>   eta2 <~ y22      0.4473      0.0186   24.0404    0.0000 [ 0.4146; 0.4765 ] 
#>   eta2 <~ y23      0.4735      0.0148   32.0572    0.0000 [ 0.4454; 0.5019 ] 
#>   eta3 <~ y31      0.4400      0.0187   23.5181    0.0000 [ 0.3970; 0.4649 ] 
#>   eta3 <~ y32      0.3521      0.0198   17.7426    0.0000 [ 0.3215; 0.3893 ] 
#>   eta3 <~ y33      0.3999      0.0237   16.8504    0.0000 [ 0.3616; 0.4465 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0474   14.1504    0.0000 [ 0.5903; 0.7419 ] 
#>   eta3 ~ eta1       0.6634      0.0398   16.6878    0.0000 [ 0.5896; 0.7250 ] 
#>   eta3 ~ eta2       0.3052      0.0753    4.0525    0.0001 [ 0.1830; 0.4796 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0481    4.2556    0.0000 [ 0.1401; 0.3227 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03753462 17.66555  7.725050e-70
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03989675 16.27396  1.510707e-59
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03833834 19.85860  9.285679e-88
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04344198 11.88838  1.360192e-32
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03047887 24.78398 1.334430e-135
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03102983 25.77080 1.884728e-146
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03498744 23.50207 3.884594e-122
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04781759 13.76207  4.310231e-43
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03978823 18.78506  1.000774e-78
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5882579          0.7255334
#> 2          0.6002738          0.7287532
#> 3          0.6926949          0.8257639
#> 4          0.4472496          0.5958676
#> 5          0.6874221          0.8044441
#> 6          0.7337382          0.8465830
#> 7          0.7485302          0.8700818
#> 8          0.5893415          0.7280784
#> 9          0.6877715          0.8218549

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 474716227
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0474   14.1504    0.0000 [ 0.5428; 0.7881 ] 
#>   eta3 ~ eta1      0.4585      0.0742    6.1763    0.0000 [ 0.2686; 0.6525 ] 
#>   eta3 ~ eta2      0.3052      0.0753    4.0525    0.0001 [ 0.1070; 0.4964 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0375   17.6656    0.0000 [ 0.5683; 0.7624 ] 
#>   eta1 =~ y12      0.6493      0.0399   16.2740    0.0000 [ 0.5395; 0.7458 ] 
#>   eta1 =~ y13      0.7613      0.0383   19.8586    0.0000 [ 0.6591; 0.8573 ] 
#>   eta2 =~ y21      0.5165      0.0434   11.8884    0.0000 [ 0.4043; 0.6290 ] 
#>   eta2 =~ y22      0.7554      0.0305   24.7840    0.0000 [ 0.6765; 0.8341 ] 
#>   eta2 =~ y23      0.7997      0.0310   25.7708    0.0000 [ 0.7330; 0.8935 ] 
#>   eta3 =~ y31      0.8223      0.0350   23.5021    0.0000 [ 0.7389; 0.9198 ] 
#>   eta3 =~ y32      0.6581      0.0478   13.7621    0.0000 [ 0.5381; 0.7853 ] 
#>   eta3 =~ y33      0.7474      0.0398   18.7851    0.0000 [ 0.6481; 0.8538 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0208   18.9748    0.0000 [ 0.3452; 0.4530 ] 
#>   eta1 <~ y12      0.3873      0.0172   22.4578    0.0000 [ 0.3412; 0.4304 ] 
#>   eta1 <~ y13      0.4542      0.0216   21.0028    0.0000 [ 0.3989; 0.5107 ] 
#>   eta2 <~ y21      0.3058      0.0250   12.2283    0.0000 [ 0.2384; 0.3677 ] 
#>   eta2 <~ y22      0.4473      0.0186   24.0404    0.0000 [ 0.3948; 0.4910 ] 
#>   eta2 <~ y23      0.4735      0.0148   32.0572    0.0000 [ 0.4390; 0.5154 ] 
#>   eta3 <~ y31      0.4400      0.0187   23.5181    0.0000 [ 0.3913; 0.4880 ] 
#>   eta3 <~ y32      0.3521      0.0198   17.7426    0.0000 [ 0.2998; 0.4024 ] 
#>   eta3 <~ y33      0.3999      0.0237   16.8504    0.0000 [ 0.3366; 0.4593 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0474   14.1504    0.0000 [ 0.5428; 0.7881 ] 
#>   eta3 ~ eta1       0.6634      0.0398   16.6878    0.0000 [ 0.5598; 0.7654 ] 
#>   eta3 ~ eta2       0.3052      0.0753    4.0525    0.0001 [ 0.1070; 0.4964 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0481    4.2556    0.0000 [ 0.0775; 0.3265 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04744261 14.150431 1.856169e-45
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07423627  6.176317 6.561409e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07530005  4.052469 5.067997e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5427606          0.7881070          0.5722219          0.7586457
#> 2          0.2686337          0.6525418          0.3147336          0.6064419
#> 3          0.1070397          0.4964490          0.1538001          0.4496886
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5894862          0.7609941          0.5903020          0.7419238
#> 2          0.2656715          0.5678146          0.2961501          0.5602094
#> 3          0.1722514          0.4899966          0.1830398          0.4796017