Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2123039818
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0370   18.1349    0.0000 [ 0.6075; 0.7300 ] 
#>   eta3 ~ eta1      0.4585      0.0897    5.1093    0.0000 [ 0.3151; 0.6900 ] 
#>   eta3 ~ eta2      0.3052      0.1025    2.9784    0.0029 [ 0.0122; 0.4441 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0424   15.6222    0.0000 [ 0.5888; 0.7439 ] 
#>   eta1 =~ y12      0.6493      0.0383   16.9402    0.0000 [ 0.5767; 0.7196 ] 
#>   eta1 =~ y13      0.7613      0.0404   18.8456    0.0000 [ 0.6922; 0.8237 ] 
#>   eta2 =~ y21      0.5165      0.0434   11.8869    0.0000 [ 0.4341; 0.5677 ] 
#>   eta2 =~ y22      0.7554      0.0397   19.0356    0.0000 [ 0.6995; 0.8341 ] 
#>   eta2 =~ y23      0.7997      0.0370   21.6024    0.0000 [ 0.7282; 0.8578 ] 
#>   eta3 =~ y31      0.8223      0.0359   22.8888    0.0000 [ 0.7696; 0.8881 ] 
#>   eta3 =~ y32      0.6581      0.0376   17.5092    0.0000 [ 0.6035; 0.7311 ] 
#>   eta3 =~ y33      0.7474      0.0357   20.9203    0.0000 [ 0.6832; 0.8143 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0271   14.5957    0.0000 [ 0.3487; 0.4423 ] 
#>   eta1 <~ y12      0.3873      0.0172   22.4902    0.0000 [ 0.3577; 0.4218 ] 
#>   eta1 <~ y13      0.4542      0.0215   21.1110    0.0000 [ 0.4171; 0.4863 ] 
#>   eta2 <~ y21      0.3058      0.0248   12.3154    0.0000 [ 0.2564; 0.3469 ] 
#>   eta2 <~ y22      0.4473      0.0174   25.7568    0.0000 [ 0.4197; 0.4756 ] 
#>   eta2 <~ y23      0.4735      0.0227   20.8754    0.0000 [ 0.4332; 0.5086 ] 
#>   eta3 <~ y31      0.4400      0.0194   22.6568    0.0000 [ 0.4065; 0.4775 ] 
#>   eta3 <~ y32      0.3521      0.0162   21.7633    0.0000 [ 0.3164; 0.3740 ] 
#>   eta3 <~ y33      0.3999      0.0176   22.7082    0.0000 [ 0.3691; 0.4315 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0370   18.1349    0.0000 [ 0.6075; 0.7300 ] 
#>   eta3 ~ eta1       0.6634      0.0377   17.5747    0.0000 [ 0.5791; 0.7234 ] 
#>   eta3 ~ eta2       0.3052      0.1025    2.9784    0.0029 [ 0.0122; 0.4441 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0663    3.0920    0.0020 [ 0.0088; 0.2753 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04244399 15.62223  5.137646e-55
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03832763 16.94021  2.273297e-64
#> 3 eta1 =~ y13  Common factor 0.7613458 0.04039904 18.84564  3.190882e-79
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04344756 11.88685  1.385278e-32
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03968286 19.03562  8.646670e-81
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03701738 21.60239 1.705738e-103
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03592487 22.88881 6.006250e-116
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03758425 17.50917  1.219670e-68
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03572716 20.92033  3.496571e-97
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5888294          0.7439325
#> 2          0.5767155          0.7195945
#> 3          0.6921533          0.8237298
#> 4          0.4340580          0.5676842
#> 5          0.6995096          0.8341309
#> 6          0.7282427          0.8577522
#> 7          0.7695848          0.8880584
#> 8          0.6035475          0.7311084
#> 9          0.6832385          0.8142500

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2123039818
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0370   18.1349    0.0000 [ 0.5758; 0.7672 ] 
#>   eta3 ~ eta1      0.4585      0.0897    5.1093    0.0000 [ 0.2134; 0.6774 ] 
#>   eta3 ~ eta2      0.3052      0.1025    2.9784    0.0029 [ 0.0531; 0.5830 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0424   15.6222    0.0000 [ 0.5569; 0.7764 ] 
#>   eta1 =~ y12      0.6493      0.0383   16.9402    0.0000 [ 0.5469; 0.7451 ] 
#>   eta1 =~ y13      0.7613      0.0404   18.8456    0.0000 [ 0.6560; 0.8649 ] 
#>   eta2 =~ y21      0.5165      0.0434   11.8869    0.0000 [ 0.4021; 0.6267 ] 
#>   eta2 =~ y22      0.7554      0.0397   19.0356    0.0000 [ 0.6515; 0.8567 ] 
#>   eta2 =~ y23      0.7997      0.0370   21.6024    0.0000 [ 0.7047; 0.8961 ] 
#>   eta3 =~ y31      0.8223      0.0359   22.8888    0.0000 [ 0.7304; 0.9162 ] 
#>   eta3 =~ y32      0.6581      0.0376   17.5092    0.0000 [ 0.5608; 0.7552 ] 
#>   eta3 =~ y33      0.7474      0.0357   20.9203    0.0000 [ 0.6521; 0.8368 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0271   14.5957    0.0000 [ 0.3278; 0.4680 ] 
#>   eta1 <~ y12      0.3873      0.0172   22.4902    0.0000 [ 0.3415; 0.4305 ] 
#>   eta1 <~ y13      0.4542      0.0215   21.1110    0.0000 [ 0.3985; 0.5098 ] 
#>   eta2 <~ y21      0.3058      0.0248   12.3154    0.0000 [ 0.2412; 0.3696 ] 
#>   eta2 <~ y22      0.4473      0.0174   25.7568    0.0000 [ 0.4029; 0.4927 ] 
#>   eta2 <~ y23      0.4735      0.0227   20.8754    0.0000 [ 0.4163; 0.5336 ] 
#>   eta3 <~ y31      0.4400      0.0194   22.6568    0.0000 [ 0.3910; 0.4914 ] 
#>   eta3 <~ y32      0.3521      0.0162   21.7633    0.0000 [ 0.3110; 0.3946 ] 
#>   eta3 <~ y33      0.3999      0.0176   22.7082    0.0000 [ 0.3535; 0.4446 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0370   18.1349    0.0000 [ 0.5758; 0.7672 ] 
#>   eta3 ~ eta1       0.6634      0.0377   17.5747    0.0000 [ 0.5628; 0.7580 ] 
#>   eta3 ~ eta2       0.3052      0.1025    2.9784    0.0029 [ 0.0531; 0.5830 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0663    3.0920    0.0020 [ 0.0437; 0.3863 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03701877 18.134947 1.688632e-73
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08973980  5.109291 3.233694e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.10245592  2.978365 2.897908e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.57575391          0.7671941          0.5987421          0.7442059
#> 2         0.21335805          0.6774416          0.2690854          0.6217142
#> 3         0.05314371          0.5829879          0.1167676          0.5193640
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.60260547          0.7409092         0.60751811          0.7300017
#> 2         0.30367058          0.7020209         0.31505560          0.6900274
#> 3        -0.01429945          0.4506002         0.01222989          0.4440712