Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -447304973
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0326   20.6125    0.0000 [ 0.6287; 0.7461 ] 
#>   eta3 ~ eta1      0.4585      0.0944    4.8577    0.0000 [ 0.2536; 0.6435 ] 
#>   eta3 ~ eta2      0.3052      0.0956    3.1908    0.0014 [ 0.1388; 0.5358 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0367   18.0733    0.0000 [ 0.6011; 0.7289 ] 
#>   eta1 =~ y12      0.6493      0.0381   17.0446    0.0000 [ 0.5840; 0.7175 ] 
#>   eta1 =~ y13      0.7613      0.0348   21.8987    0.0000 [ 0.7017; 0.8243 ] 
#>   eta2 =~ y21      0.5165      0.0510   10.1364    0.0000 [ 0.4199; 0.6074 ] 
#>   eta2 =~ y22      0.7554      0.0421   17.9275    0.0000 [ 0.6969; 0.8198 ] 
#>   eta2 =~ y23      0.7997      0.0366   21.8779    0.0000 [ 0.7285; 0.8601 ] 
#>   eta3 =~ y31      0.8223      0.0333   24.6763    0.0000 [ 0.7603; 0.8807 ] 
#>   eta3 =~ y32      0.6581      0.0440   14.9511    0.0000 [ 0.5898; 0.7306 ] 
#>   eta3 =~ y33      0.7474      0.0464   16.1181    0.0000 [ 0.6656; 0.8537 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0203   19.4822    0.0000 [ 0.3588; 0.4311 ] 
#>   eta1 <~ y12      0.3873      0.0208   18.5797    0.0000 [ 0.3467; 0.4227 ] 
#>   eta1 <~ y13      0.4542      0.0195   23.2748    0.0000 [ 0.4221; 0.4857 ] 
#>   eta2 <~ y21      0.3058      0.0278   10.9839    0.0000 [ 0.2451; 0.3628 ] 
#>   eta2 <~ y22      0.4473      0.0223   20.0452    0.0000 [ 0.4051; 0.4821 ] 
#>   eta2 <~ y23      0.4735      0.0202   23.4815    0.0000 [ 0.4337; 0.5007 ] 
#>   eta3 <~ y31      0.4400      0.0222   19.8607    0.0000 [ 0.4099; 0.4842 ] 
#>   eta3 <~ y32      0.3521      0.0215   16.3417    0.0000 [ 0.3140; 0.3996 ] 
#>   eta3 <~ y33      0.3999      0.0197   20.3138    0.0000 [ 0.3742; 0.4398 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0326   20.6125    0.0000 [ 0.6287; 0.7461 ] 
#>   eta3 ~ eta1       0.6634      0.0419   15.8197    0.0000 [ 0.5824; 0.7411 ] 
#>   eta3 ~ eta2       0.3052      0.0956    3.1908    0.0014 [ 0.1388; 0.5358 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0623    3.2887    0.0010 [ 0.0959; 0.3288 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03668779 18.07331  5.171325e-73
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03809293 17.04458  3.835063e-65
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03476664 21.89874 2.670387e-106
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05095031 10.13644  3.807197e-24
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04213574 17.92748  7.196643e-72
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03655126 21.87787 4.220782e-106
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03332256 24.67630 1.921947e-134
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04401467 14.95113  1.531368e-50
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04637181 16.11807  1.904579e-58
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6010661          0.7289301
#> 2          0.5840174          0.7175284
#> 3          0.7016870          0.8242942
#> 4          0.4198884          0.6073893
#> 5          0.6969013          0.8197906
#> 6          0.7285359          0.8601319
#> 7          0.7602993          0.8806937
#> 8          0.5897585          0.7305505
#> 9          0.6656028          0.8537485

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -447304973
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0326   20.6125    0.0000 [ 0.5801; 0.7485 ] 
#>   eta3 ~ eta1      0.4585      0.0944    4.8577    0.0000 [ 0.2155; 0.7036 ] 
#>   eta3 ~ eta2      0.3052      0.0956    3.1908    0.0014 [ 0.0478; 0.5424 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0367   18.0733    0.0000 [ 0.5606; 0.7503 ] 
#>   eta1 =~ y12      0.6493      0.0381   17.0446    0.0000 [ 0.5485; 0.7455 ] 
#>   eta1 =~ y13      0.7613      0.0348   21.8987    0.0000 [ 0.6699; 0.8496 ] 
#>   eta2 =~ y21      0.5165      0.0510   10.1364    0.0000 [ 0.3760; 0.6395 ] 
#>   eta2 =~ y22      0.7554      0.0421   17.9275    0.0000 [ 0.6436; 0.8615 ] 
#>   eta2 =~ y23      0.7997      0.0366   21.8779    0.0000 [ 0.7061; 0.8951 ] 
#>   eta3 =~ y31      0.8223      0.0333   24.6763    0.0000 [ 0.7427; 0.9151 ] 
#>   eta3 =~ y32      0.6581      0.0440   14.9511    0.0000 [ 0.5459; 0.7735 ] 
#>   eta3 =~ y33      0.7474      0.0464   16.1181    0.0000 [ 0.6225; 0.8623 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0203   19.4822    0.0000 [ 0.3420; 0.4470 ] 
#>   eta1 <~ y12      0.3873      0.0208   18.5797    0.0000 [ 0.3355; 0.4433 ] 
#>   eta1 <~ y13      0.4542      0.0195   23.2748    0.0000 [ 0.4067; 0.5076 ] 
#>   eta2 <~ y21      0.3058      0.0278   10.9839    0.0000 [ 0.2312; 0.3752 ] 
#>   eta2 <~ y22      0.4473      0.0223   20.0452    0.0000 [ 0.3914; 0.5068 ] 
#>   eta2 <~ y23      0.4735      0.0202   23.4815    0.0000 [ 0.4255; 0.5298 ] 
#>   eta3 <~ y31      0.4400      0.0222   19.8607    0.0000 [ 0.3852; 0.4998 ] 
#>   eta3 <~ y32      0.3521      0.0215   16.3417    0.0000 [ 0.2968; 0.4082 ] 
#>   eta3 <~ y33      0.3999      0.0197   20.3138    0.0000 [ 0.3459; 0.4477 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0326   20.6125    0.0000 [ 0.5801; 0.7485 ] 
#>   eta3 ~ eta1       0.6634      0.0419   15.8197    0.0000 [ 0.5479; 0.7647 ] 
#>   eta3 ~ eta2       0.3052      0.0956    3.1908    0.0014 [ 0.0478; 0.5424 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0623    3.2887    0.0010 [ 0.0357; 0.3579 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03256930 20.612461 2.121668e-94
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09438765  4.857699 1.187580e-06
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09563384  3.190828 1.418657e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.58007009          0.7485001          0.6002952          0.7282750
#> 2         0.21545220          0.7035719          0.2740658          0.6449582
#> 3         0.04779757          0.5423618          0.1071851          0.4829743
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.61589445          0.7461426          0.6287408          0.7460798
#> 2         0.23158104          0.6465502          0.2535874          0.6435207
#> 3         0.09585989          0.5392398          0.1387932          0.5357517