Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -309643377
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0400   16.7700    0.0000 [ 0.5997; 0.7442 ] 
#>   eta3 ~ eta1      0.4585      0.0874    5.2475    0.0000 [ 0.2843; 0.5834 ] 
#>   eta3 ~ eta2      0.3052      0.0839    3.6370    0.0003 [ 0.1636; 0.4584 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0430   15.4125    0.0000 [ 0.5798; 0.7332 ] 
#>   eta1 =~ y12      0.6493      0.0350   18.5388    0.0000 [ 0.5800; 0.7013 ] 
#>   eta1 =~ y13      0.7613      0.0311   24.4720    0.0000 [ 0.7003; 0.8227 ] 
#>   eta2 =~ y21      0.5165      0.0561    9.2008    0.0000 [ 0.4139; 0.6160 ] 
#>   eta2 =~ y22      0.7554      0.0367   20.6098    0.0000 [ 0.6726; 0.8105 ] 
#>   eta2 =~ y23      0.7997      0.0340   23.5031    0.0000 [ 0.7411; 0.8607 ] 
#>   eta3 =~ y31      0.8223      0.0384   21.4348    0.0000 [ 0.7714; 0.8990 ] 
#>   eta3 =~ y32      0.6581      0.0381   17.2776    0.0000 [ 0.5751; 0.7202 ] 
#>   eta3 =~ y33      0.7474      0.0468   15.9762    0.0000 [ 0.6723; 0.8193 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0220   18.0019    0.0000 [ 0.3529; 0.4400 ] 
#>   eta1 <~ y12      0.3873      0.0167   23.1434    0.0000 [ 0.3528; 0.4112 ] 
#>   eta1 <~ y13      0.4542      0.0209   21.7751    0.0000 [ 0.4197; 0.4925 ] 
#>   eta2 <~ y21      0.3058      0.0277   11.0421    0.0000 [ 0.2597; 0.3526 ] 
#>   eta2 <~ y22      0.4473      0.0240   18.6677    0.0000 [ 0.4069; 0.4885 ] 
#>   eta2 <~ y23      0.4735      0.0215   22.0529    0.0000 [ 0.4409; 0.5219 ] 
#>   eta3 <~ y31      0.4400      0.0186   23.6759    0.0000 [ 0.4052; 0.4812 ] 
#>   eta3 <~ y32      0.3521      0.0179   19.6213    0.0000 [ 0.3080; 0.3788 ] 
#>   eta3 <~ y33      0.3999      0.0253   15.8367    0.0000 [ 0.3585; 0.4537 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0400   16.7700    0.0000 [ 0.5997; 0.7442 ] 
#>   eta3 ~ eta1       0.6634      0.0422   15.7311    0.0000 [ 0.5820; 0.7271 ] 
#>   eta3 ~ eta2       0.3052      0.0839    3.6370    0.0003 [ 0.1636; 0.4584 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0584    3.5101    0.0004 [ 0.1150; 0.3258 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04302151 15.412518  1.348523e-53
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03502263 18.538814  1.004116e-76
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03111095 24.471957 2.938500e-132
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05613180  9.200752  3.554522e-20
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03665181 20.609831  2.240103e-94
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03402376 23.503098 3.791817e-122
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03836182 21.434784 6.332654e-102
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03808799 17.277596  6.938046e-67
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04678373 15.976154  1.873535e-57
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5797857          0.7332204
#> 2          0.5799994          0.7012716
#> 3          0.7003378          0.8226716
#> 4          0.4139230          0.6159950
#> 5          0.6725599          0.8104515
#> 6          0.7410522          0.8607039
#> 7          0.7713640          0.8990160
#> 8          0.5751030          0.7202453
#> 9          0.6723128          0.8192991

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -309643377
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0400   16.7700    0.0000 [ 0.5653; 0.7723 ] 
#>   eta3 ~ eta1      0.4585      0.0874    5.2475    0.0000 [ 0.2512; 0.7030 ] 
#>   eta3 ~ eta2      0.3052      0.0839    3.6370    0.0003 [ 0.0763; 0.5102 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0430   15.4125    0.0000 [ 0.5599; 0.7824 ] 
#>   eta1 =~ y12      0.6493      0.0350   18.5388    0.0000 [ 0.5633; 0.7444 ] 
#>   eta1 =~ y13      0.7613      0.0311   24.4720    0.0000 [ 0.6799; 0.8408 ] 
#>   eta2 =~ y21      0.5165      0.0561    9.2008    0.0000 [ 0.3779; 0.6682 ] 
#>   eta2 =~ y22      0.7554      0.0367   20.6098    0.0000 [ 0.6740; 0.8635 ] 
#>   eta2 =~ y23      0.7997      0.0340   23.5031    0.0000 [ 0.7174; 0.8933 ] 
#>   eta3 =~ y31      0.8223      0.0384   21.4348    0.0000 [ 0.7202; 0.9186 ] 
#>   eta3 =~ y32      0.6581      0.0381   17.2776    0.0000 [ 0.5658; 0.7628 ] 
#>   eta3 =~ y33      0.7474      0.0468   15.9762    0.0000 [ 0.6204; 0.8623 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0220   18.0019    0.0000 [ 0.3407; 0.4543 ] 
#>   eta1 <~ y12      0.3873      0.0167   23.1434    0.0000 [ 0.3439; 0.4305 ] 
#>   eta1 <~ y13      0.4542      0.0209   21.7751    0.0000 [ 0.3959; 0.5038 ] 
#>   eta2 <~ y21      0.3058      0.0277   11.0421    0.0000 [ 0.2333; 0.3766 ] 
#>   eta2 <~ y22      0.4473      0.0240   18.6677    0.0000 [ 0.3853; 0.5093 ] 
#>   eta2 <~ y23      0.4735      0.0215   22.0529    0.0000 [ 0.4130; 0.5240 ] 
#>   eta3 <~ y31      0.4400      0.0186   23.6759    0.0000 [ 0.3916; 0.4877 ] 
#>   eta3 <~ y32      0.3521      0.0179   19.6213    0.0000 [ 0.3101; 0.4029 ] 
#>   eta3 <~ y33      0.3999      0.0253   15.8367    0.0000 [ 0.3324; 0.4630 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0400   16.7700    0.0000 [ 0.5653; 0.7723 ] 
#>   eta3 ~ eta1       0.6634      0.0422   15.7311    0.0000 [ 0.5642; 0.7823 ] 
#>   eta3 ~ eta2       0.3052      0.0839    3.6370    0.0003 [ 0.0763; 0.5102 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0584    3.5101    0.0004 [ 0.0453; 0.3471 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04003173 16.770035 4.042723e-63
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08737549  5.247545 1.541397e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08390183  3.637002 2.758297e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5653277          0.7723491          0.5901869          0.7474899
#> 2          0.2511517          0.7030084          0.3054109          0.6487492
#> 3          0.0762989          0.5101918          0.1284010          0.4580897
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5750353          0.7520284          0.5996712          0.7442223
#> 2          0.2754693          0.5920758          0.2842938          0.5833718
#> 3          0.1621361          0.4706504          0.1635746          0.4583934