Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1976093989
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0460   14.5961    0.0000 [ 0.5945; 0.7332 ] 
#>   eta3 ~ eta1      0.4585      0.0745    6.1535    0.0000 [ 0.3182; 0.5729 ] 
#>   eta3 ~ eta2      0.3052      0.0786    3.8822    0.0001 [ 0.1914; 0.4933 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0342   19.3792    0.0000 [ 0.6054; 0.7184 ] 
#>   eta1 =~ y12      0.6493      0.0555   11.6934    0.0000 [ 0.5415; 0.7186 ] 
#>   eta1 =~ y13      0.7613      0.0377   20.2056    0.0000 [ 0.6951; 0.8410 ] 
#>   eta2 =~ y21      0.5165      0.0600    8.6090    0.0000 [ 0.3683; 0.6020 ] 
#>   eta2 =~ y22      0.7554      0.0377   20.0427    0.0000 [ 0.6788; 0.8133 ] 
#>   eta2 =~ y23      0.7997      0.0335   23.8994    0.0000 [ 0.7392; 0.8674 ] 
#>   eta3 =~ y31      0.8223      0.0333   24.6798    0.0000 [ 0.7676; 0.8670 ] 
#>   eta3 =~ y32      0.6581      0.0478   13.7692    0.0000 [ 0.5577; 0.7496 ] 
#>   eta3 =~ y33      0.7474      0.0420   17.7986    0.0000 [ 0.6842; 0.8343 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0239   16.5304    0.0000 [ 0.3579; 0.4557 ] 
#>   eta1 <~ y12      0.3873      0.0228   16.9728    0.0000 [ 0.3314; 0.4168 ] 
#>   eta1 <~ y13      0.4542      0.0246   18.4741    0.0000 [ 0.4170; 0.5092 ] 
#>   eta2 <~ y21      0.3058      0.0274   11.1777    0.0000 [ 0.2332; 0.3443 ] 
#>   eta2 <~ y22      0.4473      0.0264   16.9336    0.0000 [ 0.4084; 0.5075 ] 
#>   eta2 <~ y23      0.4735      0.0201   23.5785    0.0000 [ 0.4464; 0.5196 ] 
#>   eta3 <~ y31      0.4400      0.0189   23.2206    0.0000 [ 0.4038; 0.4719 ] 
#>   eta3 <~ y32      0.3521      0.0230   15.3134    0.0000 [ 0.3093; 0.3925 ] 
#>   eta3 <~ y33      0.3999      0.0225   17.7868    0.0000 [ 0.3692; 0.4557 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0460   14.5961    0.0000 [ 0.5945; 0.7332 ] 
#>   eta3 ~ eta1       0.6634      0.0428   15.5022    0.0000 [ 0.5795; 0.7249 ] 
#>   eta3 ~ eta2       0.3052      0.0786    3.8822    0.0001 [ 0.1914; 0.4933 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0518    3.9548    0.0001 [ 0.1230; 0.3399 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03421551 19.379220  1.155852e-83
#> 2 eta1 =~ y12  Common factor 0.6492779 0.05552507 11.693420  1.377262e-31
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03768002 20.205559  8.747359e-91
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05999007  8.609005  7.369741e-18
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03768894 20.042688  2.337946e-89
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03345961 23.899376 3.108820e-126
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03331786 24.679774 1.763644e-134
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04779291 13.769176  3.906456e-43
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04199351 17.798562  7.250717e-71
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6054450          0.7183736
#> 2          0.5414539          0.7186088
#> 3          0.6951465          0.8409638
#> 4          0.3682793          0.6020048
#> 5          0.6787884          0.8132982
#> 6          0.7392446          0.8673676
#> 7          0.7675819          0.8670328
#> 8          0.5576615          0.7496420
#> 9          0.6842441          0.8343183

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1976093989
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0460   14.5961    0.0000 [ 0.5600; 0.7978 ] 
#>   eta3 ~ eta1      0.4585      0.0745    6.1535    0.0000 [ 0.2706; 0.6559 ] 
#>   eta3 ~ eta2      0.3052      0.0786    3.8822    0.0001 [ 0.0905; 0.4970 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0342   19.3792    0.0000 [ 0.5694; 0.7463 ] 
#>   eta1 =~ y12      0.6493      0.0555   11.6934    0.0000 [ 0.5195; 0.8067 ] 
#>   eta1 =~ y13      0.7613      0.0377   20.2056    0.0000 [ 0.6618; 0.8566 ] 
#>   eta2 =~ y21      0.5165      0.0600    8.6090    0.0000 [ 0.3755; 0.6857 ] 
#>   eta2 =~ y22      0.7554      0.0377   20.0427    0.0000 [ 0.6594; 0.8543 ] 
#>   eta2 =~ y23      0.7997      0.0335   23.8994    0.0000 [ 0.7102; 0.8832 ] 
#>   eta3 =~ y31      0.8223      0.0333   24.6798    0.0000 [ 0.7417; 0.9140 ] 
#>   eta3 =~ y32      0.6581      0.0478   13.7692    0.0000 [ 0.5352; 0.7823 ] 
#>   eta3 =~ y33      0.7474      0.0420   17.7986    0.0000 [ 0.6362; 0.8534 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0239   16.5304    0.0000 [ 0.3288; 0.4525 ] 
#>   eta1 <~ y12      0.3873      0.0228   16.9728    0.0000 [ 0.3358; 0.4538 ] 
#>   eta1 <~ y13      0.4542      0.0246   18.4741    0.0000 [ 0.3874; 0.5145 ] 
#>   eta2 <~ y21      0.3058      0.0274   11.1777    0.0000 [ 0.2423; 0.3838 ] 
#>   eta2 <~ y22      0.4473      0.0264   16.9336    0.0000 [ 0.3766; 0.5132 ] 
#>   eta2 <~ y23      0.4735      0.0201   23.5785    0.0000 [ 0.4165; 0.5204 ] 
#>   eta3 <~ y31      0.4400      0.0189   23.2206    0.0000 [ 0.3931; 0.4910 ] 
#>   eta3 <~ y32      0.3521      0.0230   15.3134    0.0000 [ 0.2925; 0.4114 ] 
#>   eta3 <~ y33      0.3999      0.0225   17.7868    0.0000 [ 0.3396; 0.4559 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0460   14.5961    0.0000 [ 0.5600; 0.7978 ] 
#>   eta3 ~ eta1       0.6634      0.0428   15.5022    0.0000 [ 0.5528; 0.7741 ] 
#>   eta3 ~ eta2       0.3052      0.0786    3.8822    0.0001 [ 0.0905; 0.4970 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0518    3.9548    0.0001 [ 0.0663; 0.3342 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04599387 14.596149 2.971619e-48
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07451170  6.153487 7.579749e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07860186  3.882238 1.034995e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55999266          0.7978470          0.5885543          0.7692853
#> 2         0.27055432          0.6558867          0.3168252          0.6096158
#> 3         0.09052744          0.4970119          0.1393383          0.4482010
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5730807          0.7584102          0.5945273          0.7331738
#> 2          0.2733614          0.6194402          0.3182172          0.5729311
#> 3          0.1716702          0.5210631          0.1914270          0.4932897