Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 2018600287
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0351   19.1399    0.0000 [ 0.5919; 0.7298 ] 
#>   eta3 ~ eta1      0.4585      0.0810    5.6619    0.0000 [ 0.2898; 0.5986 ] 
#>   eta3 ~ eta2      0.3052      0.0773    3.9453    0.0001 [ 0.1791; 0.4635 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0387   17.1229    0.0000 [ 0.5859; 0.7256 ] 
#>   eta1 =~ y12      0.6493      0.0431   15.0615    0.0000 [ 0.5526; 0.7151 ] 
#>   eta1 =~ y13      0.7613      0.0413   18.4564    0.0000 [ 0.6821; 0.8279 ] 
#>   eta2 =~ y21      0.5165      0.0519    9.9540    0.0000 [ 0.4120; 0.5974 ] 
#>   eta2 =~ y22      0.7554      0.0360   21.0058    0.0000 [ 0.6775; 0.8090 ] 
#>   eta2 =~ y23      0.7997      0.0368   21.7179    0.0000 [ 0.7325; 0.8597 ] 
#>   eta3 =~ y31      0.8223      0.0265   31.0543    0.0000 [ 0.7662; 0.8630 ] 
#>   eta3 =~ y32      0.6581      0.0362   18.1875    0.0000 [ 0.5843; 0.7120 ] 
#>   eta3 =~ y33      0.7474      0.0314   23.8168    0.0000 [ 0.6979; 0.8017 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0194   20.3631    0.0000 [ 0.3481; 0.4257 ] 
#>   eta1 <~ y12      0.3873      0.0248   15.5958    0.0000 [ 0.3456; 0.4358 ] 
#>   eta1 <~ y13      0.4542      0.0245   18.5005    0.0000 [ 0.4190; 0.5046 ] 
#>   eta2 <~ y21      0.3058      0.0259   11.7981    0.0000 [ 0.2555; 0.3468 ] 
#>   eta2 <~ y22      0.4473      0.0213   20.9710    0.0000 [ 0.4054; 0.4865 ] 
#>   eta2 <~ y23      0.4735      0.0221   21.4121    0.0000 [ 0.4370; 0.5228 ] 
#>   eta3 <~ y31      0.4400      0.0143   30.7081    0.0000 [ 0.4164; 0.4665 ] 
#>   eta3 <~ y32      0.3521      0.0145   24.2381    0.0000 [ 0.3194; 0.3714 ] 
#>   eta3 <~ y33      0.3999      0.0173   23.1080    0.0000 [ 0.3783; 0.4343 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0351   19.1399    0.0000 [ 0.5919; 0.7298 ] 
#>   eta3 ~ eta1       0.6634      0.0430   15.4102    0.0000 [ 0.5758; 0.7240 ] 
#>   eta3 ~ eta2       0.3052      0.0773    3.9453    0.0001 [ 0.1791; 0.4635 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0545    3.7602    0.0002 [ 0.1265; 0.3222 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03872407 17.122940  1.000899e-65
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04310838 15.061525  2.900203e-51
#> 3 eta1 =~ y13  Common factor 0.7613458 0.04125097 18.456435  4.629023e-76
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05188423  9.953985  2.422821e-23
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03596094 21.005784  5.806740e-98
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03682055 21.717864 1.390976e-104
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02647871 31.054278 9.988084e-212
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03618247 18.187509  6.482039e-74
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03138219 23.816824 2.235907e-125
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5858810          0.7255897
#> 2          0.5525923          0.7150516
#> 3          0.6820942          0.8278756
#> 4          0.4120421          0.5973852
#> 5          0.6774797          0.8089651
#> 6          0.7325218          0.8596773
#> 7          0.7662184          0.8630136
#> 8          0.5842650          0.7120004
#> 9          0.6979248          0.8017020

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 2018600287
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0351   19.1399    0.0000 [ 0.5797; 0.7611 ] 
#>   eta3 ~ eta1      0.4585      0.0810    5.6619    0.0000 [ 0.2738; 0.6926 ] 
#>   eta3 ~ eta2      0.3052      0.0773    3.9453    0.0001 [ 0.0800; 0.4800 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0387   17.1229    0.0000 [ 0.5551; 0.7554 ] 
#>   eta1 =~ y12      0.6493      0.0431   15.0615    0.0000 [ 0.5415; 0.7644 ] 
#>   eta1 =~ y13      0.7613      0.0413   18.4564    0.0000 [ 0.6506; 0.8639 ] 
#>   eta2 =~ y21      0.5165      0.0519    9.9540    0.0000 [ 0.3904; 0.6587 ] 
#>   eta2 =~ y22      0.7554      0.0360   21.0058    0.0000 [ 0.6677; 0.8537 ] 
#>   eta2 =~ y23      0.7997      0.0368   21.7179    0.0000 [ 0.7076; 0.8980 ] 
#>   eta3 =~ y31      0.8223      0.0265   31.0543    0.0000 [ 0.7549; 0.8918 ] 
#>   eta3 =~ y32      0.6581      0.0362   18.1875    0.0000 [ 0.5694; 0.7565 ] 
#>   eta3 =~ y33      0.7474      0.0314   23.8168    0.0000 [ 0.6590; 0.8213 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0194   20.3631    0.0000 [ 0.3434; 0.4439 ] 
#>   eta1 <~ y12      0.3873      0.0248   15.5958    0.0000 [ 0.3279; 0.4563 ] 
#>   eta1 <~ y13      0.4542      0.0245   18.5005    0.0000 [ 0.3912; 0.5182 ] 
#>   eta2 <~ y21      0.3058      0.0259   11.7981    0.0000 [ 0.2408; 0.3748 ] 
#>   eta2 <~ y22      0.4473      0.0213   20.9710    0.0000 [ 0.3905; 0.5008 ] 
#>   eta2 <~ y23      0.4735      0.0221   21.4121    0.0000 [ 0.4131; 0.5274 ] 
#>   eta3 <~ y31      0.4400      0.0143   30.7081    0.0000 [ 0.4040; 0.4781 ] 
#>   eta3 <~ y32      0.3521      0.0145   24.2381    0.0000 [ 0.3178; 0.3930 ] 
#>   eta3 <~ y33      0.3999      0.0173   23.1080    0.0000 [ 0.3518; 0.4413 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0351   19.1399    0.0000 [ 0.5797; 0.7611 ] 
#>   eta3 ~ eta1       0.6634      0.0430   15.4102    0.0000 [ 0.5595; 0.7821 ] 
#>   eta3 ~ eta2       0.3052      0.0773    3.9453    0.0001 [ 0.0800; 0.4800 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0545    3.7602    0.0002 [ 0.0467; 0.3285 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03507510 19.139888 1.175217e-81
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08098167  5.661859 1.497423e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07734619  3.945264 7.971219e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.57972062          0.7611092          0.6015018          0.7393280
#> 2         0.27380917          0.6926007          0.3240978          0.6423120
#> 3         0.08004913          0.4800400          0.1280802          0.4320089
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5695641          0.7523350          0.5918608          0.7297557
#> 2          0.2617197          0.6224074          0.2897708          0.5986239
#> 3          0.1560042          0.4714005          0.1791074          0.4634989