Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -256519694
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0393   17.0687    0.0000 [ 0.5972; 0.7490 ] 
#>   eta3 ~ eta1      0.4585      0.0768    5.9733    0.0000 [ 0.2960; 0.5923 ] 
#>   eta3 ~ eta2      0.3052      0.0805    3.7914    0.0001 [ 0.1505; 0.4744 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0438   15.1224    0.0000 [ 0.5571; 0.7169 ] 
#>   eta1 =~ y12      0.6493      0.0339   19.1448    0.0000 [ 0.6038; 0.7186 ] 
#>   eta1 =~ y13      0.7613      0.0314   24.2511    0.0000 [ 0.6929; 0.8112 ] 
#>   eta2 =~ y21      0.5165      0.0426   12.1305    0.0000 [ 0.4283; 0.5995 ] 
#>   eta2 =~ y22      0.7554      0.0383   19.7355    0.0000 [ 0.6816; 0.8088 ] 
#>   eta2 =~ y23      0.7997      0.0407   19.6244    0.0000 [ 0.7296; 0.8747 ] 
#>   eta3 =~ y31      0.8223      0.0305   26.9950    0.0000 [ 0.7713; 0.8786 ] 
#>   eta3 =~ y32      0.6581      0.0299   22.0282    0.0000 [ 0.5898; 0.7046 ] 
#>   eta3 =~ y33      0.7474      0.0377   19.8004    0.0000 [ 0.6756; 0.8116 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0221   17.8881    0.0000 [ 0.3373; 0.4276 ] 
#>   eta1 <~ y12      0.3873      0.0211   18.3927    0.0000 [ 0.3568; 0.4278 ] 
#>   eta1 <~ y13      0.4542      0.0179   25.3098    0.0000 [ 0.4234; 0.4867 ] 
#>   eta2 <~ y21      0.3058      0.0257   11.8785    0.0000 [ 0.2625; 0.3619 ] 
#>   eta2 <~ y22      0.4473      0.0200   22.4018    0.0000 [ 0.4189; 0.4917 ] 
#>   eta2 <~ y23      0.4735      0.0201   23.5375    0.0000 [ 0.4505; 0.5158 ] 
#>   eta3 <~ y31      0.4400      0.0154   28.6406    0.0000 [ 0.4163; 0.4669 ] 
#>   eta3 <~ y32      0.3521      0.0155   22.7480    0.0000 [ 0.3243; 0.3786 ] 
#>   eta3 <~ y33      0.3999      0.0147   27.2468    0.0000 [ 0.3807; 0.4258 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0393   17.0687    0.0000 [ 0.5972; 0.7490 ] 
#>   eta3 ~ eta1       0.6634      0.0398   16.6513    0.0000 [ 0.5941; 0.7429 ] 
#>   eta3 ~ eta2       0.3052      0.0805    3.7914    0.0001 [ 0.1505; 0.4744 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0572    3.5842    0.0003 [ 0.1020; 0.3341 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04384699 15.12236  1.153397e-51
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03391404 19.14481  1.069210e-81
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03139427 24.25111 6.436158e-130
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04257472 12.13055  7.275925e-34
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03827550 19.73554  1.067981e-86
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04074846 19.62439  9.571601e-86
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03046037 26.99499 1.692196e-160
#> 8 eta3 =~ y32  Common factor 0.6580689 0.02987398 22.02816 1.547229e-107
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03774791 19.80041  2.952817e-87
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5571457          0.7168613
#> 2          0.6038483          0.7185882
#> 3          0.6928821          0.8112413
#> 4          0.4282706          0.5995352
#> 5          0.6815593          0.8088025
#> 6          0.7295778          0.8747338
#> 7          0.7713381          0.8785752
#> 8          0.5898116          0.7045620
#> 9          0.6756326          0.8115776

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -256519694
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0393   17.0687    0.0000 [ 0.5671; 0.7705 ] 
#>   eta3 ~ eta1      0.4585      0.0768    5.9733    0.0000 [ 0.2653; 0.6622 ] 
#>   eta3 ~ eta2      0.3052      0.0805    3.7914    0.0001 [ 0.0871; 0.5033 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0438   15.1224    0.0000 [ 0.5560; 0.7828 ] 
#>   eta1 =~ y12      0.6493      0.0339   19.1448    0.0000 [ 0.5564; 0.7318 ] 
#>   eta1 =~ y13      0.7613      0.0314   24.2511    0.0000 [ 0.6836; 0.8460 ] 
#>   eta2 =~ y21      0.5165      0.0426   12.1305    0.0000 [ 0.4087; 0.6289 ] 
#>   eta2 =~ y22      0.7554      0.0383   19.7355    0.0000 [ 0.6592; 0.8572 ] 
#>   eta2 =~ y23      0.7997      0.0407   19.6244    0.0000 [ 0.7018; 0.9125 ] 
#>   eta3 =~ y31      0.8223      0.0305   26.9950    0.0000 [ 0.7426; 0.9001 ] 
#>   eta3 =~ y32      0.6581      0.0299   22.0282    0.0000 [ 0.5890; 0.7435 ] 
#>   eta3 =~ y33      0.7474      0.0377   19.8004    0.0000 [ 0.6488; 0.8440 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0221   17.8881    0.0000 [ 0.3412; 0.4556 ] 
#>   eta1 <~ y12      0.3873      0.0211   18.3927    0.0000 [ 0.3286; 0.4375 ] 
#>   eta1 <~ y13      0.4542      0.0179   25.3098    0.0000 [ 0.4085; 0.5013 ] 
#>   eta2 <~ y21      0.3058      0.0257   11.8785    0.0000 [ 0.2379; 0.3711 ] 
#>   eta2 <~ y22      0.4473      0.0200   22.4018    0.0000 [ 0.3936; 0.4968 ] 
#>   eta2 <~ y23      0.4735      0.0201   23.5375    0.0000 [ 0.4220; 0.5261 ] 
#>   eta3 <~ y31      0.4400      0.0154   28.6406    0.0000 [ 0.3980; 0.4775 ] 
#>   eta3 <~ y32      0.3521      0.0155   22.7480    0.0000 [ 0.3151; 0.3951 ] 
#>   eta3 <~ y33      0.3999      0.0147   27.2468    0.0000 [ 0.3600; 0.4359 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0393   17.0687    0.0000 [ 0.5671; 0.7705 ] 
#>   eta3 ~ eta1       0.6634      0.0398   16.6513    0.0000 [ 0.5579; 0.7640 ] 
#>   eta3 ~ eta2       0.3052      0.0805    3.7914    0.0001 [ 0.0871; 0.5033 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0572    3.5842    0.0003 [ 0.0494; 0.3450 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03933135 17.068658 2.539728e-65
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07675923  5.973311 2.324862e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08048497  3.791405 1.497974e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56713636          0.7705359          0.5915607          0.7461116
#> 2         0.26526244          0.6622179          0.3129290          0.6145513
#> 3         0.08709656          0.5033194          0.1370768          0.4533392
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5854241          0.7654462          0.5971662          0.7489752
#> 2          0.2555701          0.6304774          0.2960160          0.5922648
#> 3          0.1329763          0.5083466          0.1505170          0.4743517