Summarize model

Usage

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 969185924
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0426   15.7421    0.0000 [ 0.6065; 0.7620 ] 
#>   eta3 ~ eta1      0.4585      0.0746    6.1448    0.0000 [ 0.3285; 0.6058 ] 
#>   eta3 ~ eta2      0.3052      0.0816    3.7414    0.0002 [ 0.0901; 0.4024 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0347   19.0947    0.0000 [ 0.5937; 0.7281 ] 
#>   eta1 =~ y12      0.6493      0.0362   17.9224    0.0000 [ 0.5834; 0.7050 ] 
#>   eta1 =~ y13      0.7613      0.0324   23.5047    0.0000 [ 0.7132; 0.8165 ] 
#>   eta2 =~ y21      0.5165      0.0493   10.4833    0.0000 [ 0.4292; 0.6201 ] 
#>   eta2 =~ y22      0.7554      0.0358   21.0961    0.0000 [ 0.6995; 0.8197 ] 
#>   eta2 =~ y23      0.7997      0.0399   20.0348    0.0000 [ 0.7256; 0.8806 ] 
#>   eta3 =~ y31      0.8223      0.0257   31.9652    0.0000 [ 0.7798; 0.8767 ] 
#>   eta3 =~ y32      0.6581      0.0471   13.9712    0.0000 [ 0.5565; 0.7288 ] 
#>   eta3 =~ y33      0.7474      0.0350   21.3809    0.0000 [ 0.6731; 0.7893 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0178   22.1997    0.0000 [ 0.3650; 0.4261 ] 
#>   eta1 <~ y12      0.3873      0.0175   22.1805    0.0000 [ 0.3518; 0.4108 ] 
#>   eta1 <~ y13      0.4542      0.0218   20.7870    0.0000 [ 0.4121; 0.4947 ] 
#>   eta2 <~ y21      0.3058      0.0237   12.8984    0.0000 [ 0.2641; 0.3428 ] 
#>   eta2 <~ y22      0.4473      0.0200   22.3514    0.0000 [ 0.4181; 0.4890 ] 
#>   eta2 <~ y23      0.4735      0.0233   20.3175    0.0000 [ 0.4259; 0.5089 ] 
#>   eta3 <~ y31      0.4400      0.0211   20.8057    0.0000 [ 0.4085; 0.5050 ] 
#>   eta3 <~ y32      0.3521      0.0195   18.0186    0.0000 [ 0.3072; 0.3803 ] 
#>   eta3 <~ y33      0.3999      0.0144   27.7469    0.0000 [ 0.3706; 0.4216 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0426   15.7421    0.0000 [ 0.6065; 0.7620 ] 
#>   eta3 ~ eta1       0.6634      0.0371   17.8866    0.0000 [ 0.5902; 0.7294 ] 
#>   eta3 ~ eta2       0.3052      0.0816    3.7414    0.0002 [ 0.0901; 0.4024 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0535    3.8320    0.0001 [ 0.0641; 0.2720 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03472532 19.09471  2.794017e-81
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03622723 17.92237  7.888716e-72
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03239122 23.50470 3.651869e-122
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04926461 10.48328  1.031023e-25
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03580701 21.09609  8.639849e-99
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03991364 20.03485  2.736875e-89
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02572414 31.96520 3.322491e-224
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04710193 13.97117  2.337678e-44
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03495749 21.38094 2.010220e-101
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5937394          0.7280749
#> 2          0.5834330          0.7049566
#> 3          0.7131918          0.8165489
#> 4          0.4291765          0.6200509
#> 5          0.6995173          0.8197132
#> 6          0.7256394          0.8805678
#> 7          0.7798276          0.8767384
#> 8          0.5565178          0.7287502
#> 9          0.6730698          0.7893209

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 969185924
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0426   15.7421    0.0000 [ 0.5488; 0.7693 ] 
#>   eta3 ~ eta1      0.4585      0.0746    6.1448    0.0000 [ 0.2497; 0.6355 ] 
#>   eta3 ~ eta2      0.3052      0.0816    3.7414    0.0002 [ 0.1123; 0.5341 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0347   19.0947    0.0000 [ 0.5671; 0.7467 ] 
#>   eta1 =~ y12      0.6493      0.0362   17.9224    0.0000 [ 0.5587; 0.7460 ] 
#>   eta1 =~ y13      0.7613      0.0324   23.5047    0.0000 [ 0.6715; 0.8390 ] 
#>   eta2 =~ y21      0.5165      0.0493   10.4833    0.0000 [ 0.3902; 0.6450 ] 
#>   eta2 =~ y22      0.7554      0.0358   21.0961    0.0000 [ 0.6571; 0.8422 ] 
#>   eta2 =~ y23      0.7997      0.0399   20.0348    0.0000 [ 0.6945; 0.9009 ] 
#>   eta3 =~ y31      0.8223      0.0257   31.9652    0.0000 [ 0.7502; 0.8833 ] 
#>   eta3 =~ y32      0.6581      0.0471   13.9712    0.0000 [ 0.5388; 0.7824 ] 
#>   eta3 =~ y33      0.7474      0.0350   21.3809    0.0000 [ 0.6657; 0.8465 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0178   22.1997    0.0000 [ 0.3487; 0.4408 ] 
#>   eta1 <~ y12      0.3873      0.0175   22.1805    0.0000 [ 0.3469; 0.4372 ] 
#>   eta1 <~ y13      0.4542      0.0218   20.7870    0.0000 [ 0.3971; 0.5101 ] 
#>   eta2 <~ y21      0.3058      0.0237   12.8984    0.0000 [ 0.2473; 0.3699 ] 
#>   eta2 <~ y22      0.4473      0.0200   22.3514    0.0000 [ 0.3947; 0.4981 ] 
#>   eta2 <~ y23      0.4735      0.0233   20.3175    0.0000 [ 0.4147; 0.5352 ] 
#>   eta3 <~ y31      0.4400      0.0211   20.8057    0.0000 [ 0.3805; 0.4899 ] 
#>   eta3 <~ y32      0.3521      0.0195   18.0186    0.0000 [ 0.3021; 0.4032 ] 
#>   eta3 <~ y33      0.3999      0.0144   27.7469    0.0000 [ 0.3661; 0.4406 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0426   15.7421    0.0000 [ 0.5488; 0.7693 ] 
#>   eta3 ~ eta1       0.6634      0.0371   17.8866    0.0000 [ 0.5613; 0.7531 ] 
#>   eta3 ~ eta2       0.3052      0.0816    3.7414    0.0002 [ 0.1123; 0.5341 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0535    3.8320    0.0001 [ 0.0763; 0.3528 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04264564 15.742135 7.777460e-56
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07461679  6.144820 8.005411e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08155974  3.741443 1.829666e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5487867          0.7693258          0.5752691          0.7428434
#> 2          0.2496564          0.6355323          0.2959926          0.5891962
#> 3          0.1122976          0.5340786          0.1629453          0.4834309
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.60049426          0.7821808         0.60649355          0.7620329
#> 2         0.30143851          0.6683270         0.32846508          0.6058266
#> 3         0.06937343          0.4171427         0.09008635          0.4024178