Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1696280240
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0518   12.9710    0.0000 [ 0.5522; 0.7682 ] 
#>   eta3 ~ eta1      0.4585      0.0771    5.9435    0.0000 [ 0.3212; 0.5970 ] 
#>   eta3 ~ eta2      0.3052      0.0756    4.0345    0.0001 [ 0.1551; 0.4300 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0453   14.6427    0.0000 [ 0.5792; 0.7576 ] 
#>   eta1 =~ y12      0.6493      0.0358   18.1251    0.0000 [ 0.5864; 0.7300 ] 
#>   eta1 =~ y13      0.7613      0.0325   23.4162    0.0000 [ 0.6989; 0.8261 ] 
#>   eta2 =~ y21      0.5165      0.0483   10.6887    0.0000 [ 0.4360; 0.6249 ] 
#>   eta2 =~ y22      0.7554      0.0405   18.6421    0.0000 [ 0.6760; 0.8103 ] 
#>   eta2 =~ y23      0.7997      0.0409   19.5603    0.0000 [ 0.7340; 0.8800 ] 
#>   eta3 =~ y31      0.8223      0.0308   26.6844    0.0000 [ 0.7674; 0.8782 ] 
#>   eta3 =~ y32      0.6581      0.0413   15.9229    0.0000 [ 0.5796; 0.7220 ] 
#>   eta3 =~ y33      0.7474      0.0374   20.0012    0.0000 [ 0.6843; 0.8177 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0229   17.2653    0.0000 [ 0.3396; 0.4217 ] 
#>   eta1 <~ y12      0.3873      0.0185   20.9582    0.0000 [ 0.3610; 0.4265 ] 
#>   eta1 <~ y13      0.4542      0.0196   23.2105    0.0000 [ 0.4158; 0.4900 ] 
#>   eta2 <~ y21      0.3058      0.0269   11.3787    0.0000 [ 0.2646; 0.3556 ] 
#>   eta2 <~ y22      0.4473      0.0195   22.9827    0.0000 [ 0.4127; 0.4758 ] 
#>   eta2 <~ y23      0.4735      0.0202   23.4078    0.0000 [ 0.4364; 0.5060 ] 
#>   eta3 <~ y31      0.4400      0.0188   23.4546    0.0000 [ 0.4076; 0.4710 ] 
#>   eta3 <~ y32      0.3521      0.0177   19.9377    0.0000 [ 0.3231; 0.3747 ] 
#>   eta3 <~ y33      0.3999      0.0193   20.7199    0.0000 [ 0.3687; 0.4383 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0518   12.9710    0.0000 [ 0.5522; 0.7682 ] 
#>   eta3 ~ eta1       0.6634      0.0441   15.0537    0.0000 [ 0.5416; 0.7361 ] 
#>   eta3 ~ eta2       0.3052      0.0756    4.0345    0.0001 [ 0.1551; 0.4300 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0560    3.6560    0.0003 [ 0.0966; 0.3239 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04528328 14.64271  1.499736e-48
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03582211 18.12506  2.021227e-73
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03251369 23.41616 2.925649e-121
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04831795 10.68867  1.150038e-26
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04052058 18.64208  1.464475e-77
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04088195 19.56031  3.369997e-85
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03081489 26.68442 7.138430e-157
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04132842 15.92292  4.394113e-57
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03736894 20.00121  5.375171e-89
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5791904          0.7576036
#> 2          0.5864045          0.7300317
#> 3          0.6988946          0.8260647
#> 4          0.4359830          0.6248952
#> 5          0.6759816          0.8102979
#> 6          0.7339532          0.8800123
#> 7          0.7673773          0.8782356
#> 8          0.5796150          0.7220283
#> 9          0.6843234          0.8176915

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1696280240
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0518   12.9710    0.0000 [ 0.5408; 0.8085 ] 
#>   eta3 ~ eta1      0.4585      0.0771    5.9435    0.0000 [ 0.2647; 0.6637 ] 
#>   eta3 ~ eta2      0.3052      0.0756    4.0345    0.0001 [ 0.1060; 0.4972 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0453   14.6427    0.0000 [ 0.5547; 0.7889 ] 
#>   eta1 =~ y12      0.6493      0.0358   18.1251    0.0000 [ 0.5504; 0.7357 ] 
#>   eta1 =~ y13      0.7613      0.0325   23.4162    0.0000 [ 0.6723; 0.8404 ] 
#>   eta2 =~ y21      0.5165      0.0483   10.6887    0.0000 [ 0.3844; 0.6343 ] 
#>   eta2 =~ y22      0.7554      0.0405   18.6421    0.0000 [ 0.6516; 0.8611 ] 
#>   eta2 =~ y23      0.7997      0.0409   19.5603    0.0000 [ 0.6988; 0.9102 ] 
#>   eta3 =~ y31      0.8223      0.0308   26.6844    0.0000 [ 0.7421; 0.9014 ] 
#>   eta3 =~ y32      0.6581      0.0413   15.9229    0.0000 [ 0.5577; 0.7714 ] 
#>   eta3 =~ y33      0.7474      0.0374   20.0012    0.0000 [ 0.6454; 0.8386 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0229   17.2653    0.0000 [ 0.3426; 0.4611 ] 
#>   eta1 <~ y12      0.3873      0.0185   20.9582    0.0000 [ 0.3368; 0.4324 ] 
#>   eta1 <~ y13      0.4542      0.0196   23.2105    0.0000 [ 0.4016; 0.5028 ] 
#>   eta2 <~ y21      0.3058      0.0269   11.3787    0.0000 [ 0.2324; 0.3714 ] 
#>   eta2 <~ y22      0.4473      0.0195   22.9827    0.0000 [ 0.3980; 0.4986 ] 
#>   eta2 <~ y23      0.4735      0.0202   23.4078    0.0000 [ 0.4244; 0.5291 ] 
#>   eta3 <~ y31      0.4400      0.0188   23.4546    0.0000 [ 0.3912; 0.4882 ] 
#>   eta3 <~ y32      0.3521      0.0177   19.9377    0.0000 [ 0.3102; 0.4016 ] 
#>   eta3 <~ y33      0.3999      0.0193   20.7199    0.0000 [ 0.3472; 0.4470 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0518   12.9710    0.0000 [ 0.5408; 0.8085 ] 
#>   eta3 ~ eta1       0.6634      0.0441   15.0537    0.0000 [ 0.5530; 0.7809 ] 
#>   eta3 ~ eta2       0.3052      0.0756    4.0345    0.0001 [ 0.1060; 0.4972 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0560    3.6560    0.0003 [ 0.0578; 0.3476 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05175652 12.970991 1.787040e-38
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07714482  5.943455 2.790758e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07563454  4.034547 5.470766e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5407968          0.8084523          0.5729369          0.7763121
#> 2          0.2647494          0.6636988          0.3126554          0.6157928
#> 3          0.1060127          0.4971518          0.1529808          0.4501836
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5518602          0.7765629          0.5522482          0.7681542
#> 2          0.3191616          0.6177634          0.3212135          0.5969646
#> 3          0.1496851          0.4851307          0.1551022          0.4299720