Summarize model

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -884384831
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0479   14.0217    0.0000 [ 0.5501; 0.7277 ] 
#>   eta3 ~ eta1      0.4585      0.0658    6.9721    0.0000 [ 0.3160; 0.5584 ] 
#>   eta3 ~ eta2      0.3052      0.0522    5.8419    0.0000 [ 0.2552; 0.4404 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0395   16.7751    0.0000 [ 0.5988; 0.7355 ] 
#>   eta1 =~ y12      0.6493      0.0382   17.0019    0.0000 [ 0.5849; 0.7239 ] 
#>   eta1 =~ y13      0.7613      0.0349   21.8221    0.0000 [ 0.6948; 0.8135 ] 
#>   eta2 =~ y21      0.5165      0.0657    7.8631    0.0000 [ 0.3815; 0.6100 ] 
#>   eta2 =~ y22      0.7554      0.0423   17.8713    0.0000 [ 0.6565; 0.8083 ] 
#>   eta2 =~ y23      0.7997      0.0397   20.1259    0.0000 [ 0.7146; 0.8604 ] 
#>   eta3 =~ y31      0.8223      0.0397   20.7014    0.0000 [ 0.7578; 0.8773 ] 
#>   eta3 =~ y32      0.6581      0.0394   16.6905    0.0000 [ 0.5909; 0.7270 ] 
#>   eta3 =~ y33      0.7474      0.0405   18.4357    0.0000 [ 0.6745; 0.8123 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0233   16.9509    0.0000 [ 0.3466; 0.4340 ] 
#>   eta1 <~ y12      0.3873      0.0154   25.1365    0.0000 [ 0.3727; 0.4210 ] 
#>   eta1 <~ y13      0.4542      0.0218   20.8692    0.0000 [ 0.4085; 0.4860 ] 
#>   eta2 <~ y21      0.3058      0.0367    8.3356    0.0000 [ 0.2415; 0.3788 ] 
#>   eta2 <~ y22      0.4473      0.0233   19.2094    0.0000 [ 0.4093; 0.4894 ] 
#>   eta2 <~ y23      0.4735      0.0214   22.1150    0.0000 [ 0.4353; 0.5090 ] 
#>   eta3 <~ y31      0.4400      0.0188   23.3867    0.0000 [ 0.4143; 0.4773 ] 
#>   eta3 <~ y32      0.3521      0.0193   18.2644    0.0000 [ 0.3228; 0.3918 ] 
#>   eta3 <~ y33      0.3999      0.0205   19.5113    0.0000 [ 0.3672; 0.4356 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0479   14.0217    0.0000 [ 0.5501; 0.7277 ] 
#>   eta3 ~ eta1       0.6634      0.0424   15.6426    0.0000 [ 0.5971; 0.7248 ] 
#>   eta3 ~ eta2       0.3052      0.0522    5.8419    0.0000 [ 0.2552; 0.4404 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0377    5.4405    0.0000 [ 0.1617; 0.2933 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03952700 16.77511  3.711564e-63
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03818866 17.00185  7.956421e-65
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03488876 21.82209 1.431568e-105
#> 4 eta2 =~ y21  Common factor 0.5164548 0.06568040  7.86315  3.745927e-15
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04226816 17.87132  1.972768e-71
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03973305 20.12591  4.376767e-90
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03972084 20.70141  3.363956e-95
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03942767 16.69053  1.535923e-62
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04054229 18.43566  6.797806e-76
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5987998          0.7354585
#> 2          0.5848505          0.7239298
#> 3          0.6948289          0.8134699
#> 4          0.3814880          0.6099669
#> 5          0.6565299          0.8083392
#> 6          0.7146355          0.8603865
#> 7          0.7577574          0.8772728
#> 8          0.5908643          0.7270258
#> 9          0.6744530          0.8122545

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -884384831
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0479   14.0217    0.0000 [ 0.5562; 0.8038 ] 
#>   eta3 ~ eta1      0.4585      0.0658    6.9721    0.0000 [ 0.2979; 0.6380 ] 
#>   eta3 ~ eta2      0.3052      0.0522    5.8419    0.0000 [ 0.1457; 0.4159 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0395   16.7751    0.0000 [ 0.5686; 0.7730 ] 
#>   eta1 =~ y12      0.6493      0.0382   17.0019    0.0000 [ 0.5333; 0.7308 ] 
#>   eta1 =~ y13      0.7613      0.0349   21.8221    0.0000 [ 0.6769; 0.8573 ] 
#>   eta2 =~ y21      0.5165      0.0657    7.8631    0.0000 [ 0.3426; 0.6823 ] 
#>   eta2 =~ y22      0.7554      0.0423   17.8713    0.0000 [ 0.6488; 0.8674 ] 
#>   eta2 =~ y23      0.7997      0.0397   20.1259    0.0000 [ 0.7053; 0.9107 ] 
#>   eta3 =~ y31      0.8223      0.0397   20.7014    0.0000 [ 0.7259; 0.9313 ] 
#>   eta3 =~ y32      0.6581      0.0394   16.6905    0.0000 [ 0.5501; 0.7540 ] 
#>   eta3 =~ y33      0.7474      0.0405   18.4357    0.0000 [ 0.6489; 0.8585 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0233   16.9509    0.0000 [ 0.3409; 0.4616 ] 
#>   eta1 <~ y12      0.3873      0.0154   25.1365    0.0000 [ 0.3386; 0.4183 ] 
#>   eta1 <~ y13      0.4542      0.0218   20.8692    0.0000 [ 0.4025; 0.5151 ] 
#>   eta2 <~ y21      0.3058      0.0367    8.3356    0.0000 [ 0.2075; 0.3972 ] 
#>   eta2 <~ y22      0.4473      0.0233   19.2094    0.0000 [ 0.3867; 0.5072 ] 
#>   eta2 <~ y23      0.4735      0.0214   22.1150    0.0000 [ 0.4210; 0.5318 ] 
#>   eta3 <~ y31      0.4400      0.0188   23.3867    0.0000 [ 0.3928; 0.4901 ] 
#>   eta3 <~ y32      0.3521      0.0193   18.2644    0.0000 [ 0.2975; 0.3972 ] 
#>   eta3 <~ y33      0.3999      0.0205   19.5113    0.0000 [ 0.3485; 0.4545 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0479   14.0217    0.0000 [ 0.5562; 0.8038 ] 
#>   eta3 ~ eta1       0.6634      0.0424   15.6426    0.0000 [ 0.5498; 0.7691 ] 
#>   eta3 ~ eta2       0.3052      0.0522    5.8419    0.0000 [ 0.1457; 0.4159 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0377    5.4405    0.0000 [ 0.0942; 0.2889 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04787831 14.021660 1.148957e-44
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.06576322  6.972085 3.122778e-12
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.05223464  5.841930 5.159934e-09
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5562474          0.8038470          0.5859793          0.7741152
#> 2          0.2978649          0.6379552          0.3387031          0.5971170
#> 3          0.1457495          0.4158775          0.1781866          0.3834404
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5419918          0.7287841          0.5501410          0.7277264
#> 2          0.3106747          0.5746044          0.3160069          0.5583931
#> 3          0.2299067          0.4566258          0.2551815          0.4404335