Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 324549252
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0420   15.9987    0.0000 [ 0.5851; 0.7386 ] 
#>   eta3 ~ eta1      0.4585      0.0644    7.1224    0.0000 [ 0.2944; 0.5572 ] 
#>   eta3 ~ eta2      0.3052      0.0715    4.2683    0.0000 [ 0.1931; 0.4808 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0332   19.9841    0.0000 [ 0.5960; 0.7116 ] 
#>   eta1 =~ y12      0.6493      0.0354   18.3234    0.0000 [ 0.5686; 0.7014 ] 
#>   eta1 =~ y13      0.7613      0.0315   24.1655    0.0000 [ 0.7235; 0.8439 ] 
#>   eta2 =~ y21      0.5165      0.0468   11.0333    0.0000 [ 0.4441; 0.6219 ] 
#>   eta2 =~ y22      0.7554      0.0383   19.7050    0.0000 [ 0.6738; 0.8272 ] 
#>   eta2 =~ y23      0.7997      0.0332   24.1055    0.0000 [ 0.7307; 0.8675 ] 
#>   eta3 =~ y31      0.8223      0.0316   26.0527    0.0000 [ 0.7619; 0.8724 ] 
#>   eta3 =~ y32      0.6581      0.0422   15.5882    0.0000 [ 0.5733; 0.7221 ] 
#>   eta3 =~ y33      0.7474      0.0394   18.9604    0.0000 [ 0.6749; 0.8084 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0158   25.1129    0.0000 [ 0.3634; 0.4147 ] 
#>   eta1 <~ y12      0.3873      0.0176   21.9727    0.0000 [ 0.3483; 0.4121 ] 
#>   eta1 <~ y13      0.4542      0.0204   22.3024    0.0000 [ 0.4210; 0.4885 ] 
#>   eta2 <~ y21      0.3058      0.0267   11.4688    0.0000 [ 0.2554; 0.3654 ] 
#>   eta2 <~ y22      0.4473      0.0177   25.3002    0.0000 [ 0.4088; 0.4744 ] 
#>   eta2 <~ y23      0.4735      0.0211   22.4598    0.0000 [ 0.4379; 0.5103 ] 
#>   eta3 <~ y31      0.4400      0.0207   21.2832    0.0000 [ 0.4106; 0.4814 ] 
#>   eta3 <~ y32      0.3521      0.0193   18.2647    0.0000 [ 0.3123; 0.3718 ] 
#>   eta3 <~ y33      0.3999      0.0179   22.3373    0.0000 [ 0.3737; 0.4311 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0420   15.9987    0.0000 [ 0.5851; 0.7386 ] 
#>   eta3 ~ eta1       0.6634      0.0379   17.5242    0.0000 [ 0.5653; 0.7325 ] 
#>   eta3 ~ eta2       0.3052      0.0715    4.2683    0.0000 [ 0.1931; 0.4808 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0508    4.0333    0.0001 [ 0.1298; 0.3098 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03317985 19.98411  7.572503e-89
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03543431 18.32343  5.381432e-75
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03150544 24.16554 5.126880e-129
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04680888 11.03327  2.640954e-28
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03833491 19.70495  1.955235e-86
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03317356 24.10546 2.191228e-128
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03156210 26.05268 1.254386e-149
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04221582 15.58821  8.755814e-55
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03942037 18.96035  3.627132e-80
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5959502          0.7115852
#> 2          0.5685999          0.7013599
#> 3          0.7234608          0.8438511
#> 4          0.4440604          0.6219037
#> 5          0.6738082          0.8271651
#> 6          0.7307444          0.8675220
#> 7          0.7619244          0.8723906
#> 8          0.5732903          0.7221137
#> 9          0.6749063          0.8084436

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 324549252
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0420   15.9987    0.0000 [ 0.5749; 0.7919 ] 
#>   eta3 ~ eta1      0.4585      0.0644    7.1224    0.0000 [ 0.3096; 0.6426 ] 
#>   eta3 ~ eta2      0.3052      0.0715    4.2683    0.0000 [ 0.0985; 0.4682 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0332   19.9841    0.0000 [ 0.5805; 0.7521 ] 
#>   eta1 =~ y12      0.6493      0.0354   18.3234    0.0000 [ 0.5604; 0.7437 ] 
#>   eta1 =~ y13      0.7613      0.0315   24.1655    0.0000 [ 0.6699; 0.8328 ] 
#>   eta2 =~ y21      0.5165      0.0468   11.0333    0.0000 [ 0.3875; 0.6295 ] 
#>   eta2 =~ y22      0.7554      0.0383   19.7050    0.0000 [ 0.6656; 0.8639 ] 
#>   eta2 =~ y23      0.7997      0.0332   24.1055    0.0000 [ 0.7145; 0.8860 ] 
#>   eta3 =~ y31      0.8223      0.0316   26.0527    0.0000 [ 0.7401; 0.9034 ] 
#>   eta3 =~ y32      0.6581      0.0422   15.5882    0.0000 [ 0.5587; 0.7770 ] 
#>   eta3 =~ y33      0.7474      0.0394   18.9604    0.0000 [ 0.6479; 0.8517 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0158   25.1129    0.0000 [ 0.3582; 0.4397 ] 
#>   eta1 <~ y12      0.3873      0.0176   21.9727    0.0000 [ 0.3449; 0.4360 ] 
#>   eta1 <~ y13      0.4542      0.0204   22.3024    0.0000 [ 0.3970; 0.5023 ] 
#>   eta2 <~ y21      0.3058      0.0267   11.4688    0.0000 [ 0.2317; 0.3696 ] 
#>   eta2 <~ y22      0.4473      0.0177   25.3002    0.0000 [ 0.4066; 0.4980 ] 
#>   eta2 <~ y23      0.4735      0.0211   22.4598    0.0000 [ 0.4185; 0.5275 ] 
#>   eta3 <~ y31      0.4400      0.0207   21.2832    0.0000 [ 0.3829; 0.4898 ] 
#>   eta3 <~ y32      0.3521      0.0193   18.2647    0.0000 [ 0.3052; 0.4049 ] 
#>   eta3 <~ y33      0.3999      0.0179   22.3373    0.0000 [ 0.3521; 0.4447 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0420   15.9987    0.0000 [ 0.5749; 0.7919 ] 
#>   eta3 ~ eta1       0.6634      0.0379   17.5242    0.0000 [ 0.5724; 0.7682 ] 
#>   eta3 ~ eta2       0.3052      0.0715    4.2683    0.0000 [ 0.0985; 0.4682 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0508    4.0333    0.0001 [ 0.0629; 0.3255 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04196165 15.998736 1.303957e-57
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.06437504  7.122431 1.060401e-12
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07149199  4.268326 1.969455e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.57486934          0.7918713          0.6009270          0.7658136
#> 2         0.30964769          0.6425591          0.3496238          0.6025829
#> 3         0.09851744          0.4682337          0.1429131          0.4238380
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5794891          0.7417271          0.5850841          0.7386326
#> 2          0.2852559          0.5668740          0.2943893          0.5571532
#> 3          0.1870206          0.5147762          0.1930804          0.4807527