Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1737328448
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0348   19.2662    0.0000 [ 0.6114; 0.7344 ] 
#>   eta3 ~ eta1      0.4585      0.0748    6.1296    0.0000 [ 0.3144; 0.5915 ] 
#>   eta3 ~ eta2      0.3052      0.0823    3.7078    0.0002 [ 0.1500; 0.4536 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0429   15.4463    0.0000 [ 0.6090; 0.7503 ] 
#>   eta1 =~ y12      0.6493      0.0352   18.4231    0.0000 [ 0.5952; 0.7111 ] 
#>   eta1 =~ y13      0.7613      0.0297   25.6455    0.0000 [ 0.6988; 0.8030 ] 
#>   eta2 =~ y21      0.5165      0.0508   10.1576    0.0000 [ 0.4087; 0.6199 ] 
#>   eta2 =~ y22      0.7554      0.0356   21.1945    0.0000 [ 0.6773; 0.8107 ] 
#>   eta2 =~ y23      0.7997      0.0306   26.1278    0.0000 [ 0.7446; 0.8528 ] 
#>   eta3 =~ y31      0.8223      0.0359   22.9240    0.0000 [ 0.7617; 0.8970 ] 
#>   eta3 =~ y32      0.6581      0.0406   16.1987    0.0000 [ 0.6065; 0.7513 ] 
#>   eta3 =~ y33      0.7474      0.0473   15.8150    0.0000 [ 0.6727; 0.8359 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0235   16.8449    0.0000 [ 0.3610; 0.4369 ] 
#>   eta1 <~ y12      0.3873      0.0200   19.3597    0.0000 [ 0.3525; 0.4143 ] 
#>   eta1 <~ y13      0.4542      0.0146   31.1812    0.0000 [ 0.4236; 0.4699 ] 
#>   eta2 <~ y21      0.3058      0.0265   11.5451    0.0000 [ 0.2560; 0.3476 ] 
#>   eta2 <~ y22      0.4473      0.0174   25.7661    0.0000 [ 0.4079; 0.4673 ] 
#>   eta2 <~ y23      0.4735      0.0214   22.1239    0.0000 [ 0.4441; 0.5169 ] 
#>   eta3 <~ y31      0.4400      0.0210   20.9344    0.0000 [ 0.3988; 0.4735 ] 
#>   eta3 <~ y32      0.3521      0.0201   17.4908    0.0000 [ 0.3174; 0.3892 ] 
#>   eta3 <~ y33      0.3999      0.0206   19.4054    0.0000 [ 0.3625; 0.4401 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0348   19.2662    0.0000 [ 0.6114; 0.7344 ] 
#>   eta3 ~ eta1       0.6634      0.0368   18.0422    0.0000 [ 0.6112; 0.7297 ] 
#>   eta3 ~ eta2       0.3052      0.0823    3.7078    0.0002 [ 0.1500; 0.4536 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0549    3.7282    0.0002 [ 0.0975; 0.3032 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04292747 15.44628  7.992232e-54
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03524263 18.42309  8.577022e-76
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02968730 25.64551 4.745110e-145
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05084416 10.15760  3.065161e-24
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03564081 21.19446  1.074232e-99
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03060591 26.12775 1.764310e-150
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03586977 22.92396 2.680387e-116
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04062487 16.19867  5.152863e-59
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04726044 15.81500  2.451929e-56
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6090046          0.7502835
#> 2          0.5951636          0.7110897
#> 3          0.6988440          0.8030423
#> 4          0.4087228          0.6198998
#> 5          0.6772850          0.8106705
#> 6          0.7446229          0.8527527
#> 7          0.7616804          0.8969528
#> 8          0.6064929          0.7512864
#> 9          0.6726608          0.8359393

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1737328448
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0348   19.2662    0.0000 [ 0.5853; 0.7655 ] 
#>   eta3 ~ eta1      0.4585      0.0748    6.1296    0.0000 [ 0.2746; 0.6615 ] 
#>   eta3 ~ eta2      0.3052      0.0823    3.7078    0.0002 [ 0.0779; 0.5035 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0429   15.4463    0.0000 [ 0.5444; 0.7664 ] 
#>   eta1 =~ y12      0.6493      0.0352   18.4231    0.0000 [ 0.5562; 0.7384 ] 
#>   eta1 =~ y13      0.7613      0.0297   25.6455    0.0000 [ 0.6816; 0.8351 ] 
#>   eta2 =~ y21      0.5165      0.0508   10.1576    0.0000 [ 0.3855; 0.6485 ] 
#>   eta2 =~ y22      0.7554      0.0356   21.1945    0.0000 [ 0.6806; 0.8649 ] 
#>   eta2 =~ y23      0.7997      0.0306   26.1278    0.0000 [ 0.7133; 0.8715 ] 
#>   eta3 =~ y31      0.8223      0.0359   22.9240    0.0000 [ 0.7278; 0.9133 ] 
#>   eta3 =~ y32      0.6581      0.0406   16.1987    0.0000 [ 0.5459; 0.7560 ] 
#>   eta3 =~ y33      0.7474      0.0473   15.8150    0.0000 [ 0.6154; 0.8598 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0235   16.8449    0.0000 [ 0.3341; 0.4556 ] 
#>   eta1 <~ y12      0.3873      0.0200   19.3597    0.0000 [ 0.3381; 0.4415 ] 
#>   eta1 <~ y13      0.4542      0.0146   31.1812    0.0000 [ 0.4191; 0.4944 ] 
#>   eta2 <~ y21      0.3058      0.0265   11.5451    0.0000 [ 0.2358; 0.3728 ] 
#>   eta2 <~ y22      0.4473      0.0174   25.7661    0.0000 [ 0.4097; 0.4995 ] 
#>   eta2 <~ y23      0.4735      0.0214   22.1239    0.0000 [ 0.4103; 0.5210 ] 
#>   eta3 <~ y31      0.4400      0.0210   20.9344    0.0000 [ 0.3903; 0.4990 ] 
#>   eta3 <~ y32      0.3521      0.0201   17.4908    0.0000 [ 0.3009; 0.4050 ] 
#>   eta3 <~ y33      0.3999      0.0206   19.4054    0.0000 [ 0.3468; 0.4534 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0348   19.2662    0.0000 [ 0.5853; 0.7655 ] 
#>   eta3 ~ eta1       0.6634      0.0368   18.0422    0.0000 [ 0.5698; 0.7599 ] 
#>   eta3 ~ eta2       0.3052      0.0823    3.7078    0.0002 [ 0.0779; 0.5035 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0549    3.7282    0.0002 [ 0.0547; 0.3389 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03484518 19.266178 1.032884e-82
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07480180  6.129622 8.808805e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08229918  3.707827 2.090455e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5852987          0.7654983          0.6069371          0.7438598
#> 2          0.2746285          0.6614612          0.3210795          0.6150101
#> 3          0.0779001          0.5035050          0.1290069          0.4523982
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.60958105          0.7638098          0.6113729          0.7343826
#> 2         0.31353494          0.6147531          0.3143687          0.5915399
#> 3         0.08487418          0.4699102          0.1500246          0.4535769