Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -358231022
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0501   13.3941    0.0000 [ 0.5782; 0.7458 ] 
#>   eta3 ~ eta1      0.4585      0.0780    5.8767    0.0000 [ 0.3028; 0.5860 ] 
#>   eta3 ~ eta2      0.3052      0.0772    3.9540    0.0001 [ 0.1737; 0.4870 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0435   15.2263    0.0000 [ 0.5805; 0.7353 ] 
#>   eta1 =~ y12      0.6493      0.0458   14.1758    0.0000 [ 0.5564; 0.7430 ] 
#>   eta1 =~ y13      0.7613      0.0339   22.4324    0.0000 [ 0.6923; 0.8319 ] 
#>   eta2 =~ y21      0.5165      0.0458   11.2716    0.0000 [ 0.4398; 0.5977 ] 
#>   eta2 =~ y22      0.7554      0.0329   22.9404    0.0000 [ 0.7048; 0.8185 ] 
#>   eta2 =~ y23      0.7997      0.0439   18.1967    0.0000 [ 0.7292; 0.8731 ] 
#>   eta3 =~ y31      0.8223      0.0285   28.8161    0.0000 [ 0.7748; 0.8748 ] 
#>   eta3 =~ y32      0.6581      0.0462   14.2433    0.0000 [ 0.5681; 0.7413 ] 
#>   eta3 =~ y33      0.7474      0.0395   18.9069    0.0000 [ 0.6601; 0.8201 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0242   16.3201    0.0000 [ 0.3444; 0.4320 ] 
#>   eta1 <~ y12      0.3873      0.0230   16.8174    0.0000 [ 0.3486; 0.4206 ] 
#>   eta1 <~ y13      0.4542      0.0205   22.1074    0.0000 [ 0.4203; 0.4840 ] 
#>   eta2 <~ y21      0.3058      0.0244   12.5452    0.0000 [ 0.2665; 0.3515 ] 
#>   eta2 <~ y22      0.4473      0.0178   25.0663    0.0000 [ 0.4242; 0.4887 ] 
#>   eta2 <~ y23      0.4735      0.0249   19.0457    0.0000 [ 0.4288; 0.5305 ] 
#>   eta3 <~ y31      0.4400      0.0185   23.7408    0.0000 [ 0.4124; 0.4799 ] 
#>   eta3 <~ y32      0.3521      0.0215   16.3939    0.0000 [ 0.3169; 0.3891 ] 
#>   eta3 <~ y33      0.3999      0.0177   22.6028    0.0000 [ 0.3690; 0.4298 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0501   13.3941    0.0000 [ 0.5782; 0.7458 ] 
#>   eta3 ~ eta1       0.6634      0.0417   15.9221    0.0000 [ 0.5955; 0.7210 ] 
#>   eta3 ~ eta2       0.3052      0.0772    3.9540    0.0001 [ 0.1737; 0.4870 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0508    4.0365    0.0001 [ 0.1134; 0.3053 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04354776 15.22627  2.367229e-52
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04580198 14.17576  1.294329e-45
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03393949 22.43245 1.899045e-111
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04581895 11.27164  1.811578e-29
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03292824 22.94042 1.836526e-116
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04394553 18.19670  5.480851e-74
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02853531 28.81614 1.346727e-182
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04620189 14.24333  4.931745e-46
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03953191 18.90686  1.001489e-79
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5805323          0.7352723
#> 2          0.5564172          0.7429510
#> 3          0.6923254          0.8318730
#> 4          0.4397935          0.5976708
#> 5          0.7047993          0.8185316
#> 6          0.7291683          0.8731115
#> 7          0.7747572          0.8747543
#> 8          0.5680767          0.7413286
#> 9          0.6600960          0.8200960

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -358231022
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0501   13.3941    0.0000 [ 0.5432; 0.8024 ] 
#>   eta3 ~ eta1      0.4585      0.0780    5.8767    0.0000 [ 0.2715; 0.6750 ] 
#>   eta3 ~ eta2      0.3052      0.0772    3.9540    0.0001 [ 0.0834; 0.4825 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0435   15.2263    0.0000 [ 0.5551; 0.7803 ] 
#>   eta1 =~ y12      0.6493      0.0458   14.1758    0.0000 [ 0.5228; 0.7597 ] 
#>   eta1 =~ y13      0.7613      0.0339   22.4324    0.0000 [ 0.6704; 0.8459 ] 
#>   eta2 =~ y21      0.5165      0.0458   11.2716    0.0000 [ 0.3976; 0.6345 ] 
#>   eta2 =~ y22      0.7554      0.0329   22.9404    0.0000 [ 0.6650; 0.8353 ] 
#>   eta2 =~ y23      0.7997      0.0439   18.1967    0.0000 [ 0.6869; 0.9142 ] 
#>   eta3 =~ y31      0.8223      0.0285   28.8161    0.0000 [ 0.7500; 0.8976 ] 
#>   eta3 =~ y32      0.6581      0.0462   14.2433    0.0000 [ 0.5480; 0.7869 ] 
#>   eta3 =~ y33      0.7474      0.0395   18.9069    0.0000 [ 0.6431; 0.8476 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0242   16.3201    0.0000 [ 0.3378; 0.4631 ] 
#>   eta1 <~ y12      0.3873      0.0230   16.8174    0.0000 [ 0.3253; 0.4444 ] 
#>   eta1 <~ y13      0.4542      0.0205   22.1074    0.0000 [ 0.4015; 0.5077 ] 
#>   eta2 <~ y21      0.3058      0.0244   12.5452    0.0000 [ 0.2440; 0.3701 ] 
#>   eta2 <~ y22      0.4473      0.0178   25.0663    0.0000 [ 0.3999; 0.4922 ] 
#>   eta2 <~ y23      0.4735      0.0249   19.0457    0.0000 [ 0.4117; 0.5403 ] 
#>   eta3 <~ y31      0.4400      0.0185   23.7408    0.0000 [ 0.3904; 0.4862 ] 
#>   eta3 <~ y32      0.3521      0.0215   16.3939    0.0000 [ 0.2999; 0.4110 ] 
#>   eta3 <~ y33      0.3999      0.0177   22.6028    0.0000 [ 0.3511; 0.4426 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0501   13.3941    0.0000 [ 0.5432; 0.8024 ] 
#>   eta3 ~ eta1       0.6634      0.0417   15.9221    0.0000 [ 0.5569; 0.7723 ] 
#>   eta3 ~ eta2       0.3052      0.0772    3.9540    0.0001 [ 0.0834; 0.4825 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0508    4.0365    0.0001 [ 0.0601; 0.3225 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05012144 13.394136 6.543272e-41
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07802122  5.876693 4.185424e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07717609  3.953960 7.686843e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54318841          0.8023882          0.5743132          0.7712634
#> 2         0.27154877          0.6750304          0.3199990          0.6265802
#> 3         0.08335972          0.4824709          0.1312852          0.4345454
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5633891          0.7574139          0.5781719          0.7457960
#> 2          0.2519560          0.6150288          0.3027856          0.5859705
#> 3          0.1187590          0.5188483          0.1737023          0.4869700