Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 40678459
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0453   14.8122    0.0000 [ 0.5862; 0.7413 ] 
#>   eta3 ~ eta1      0.4585      0.0843    5.4396    0.0000 [ 0.3039; 0.6032 ] 
#>   eta3 ~ eta2      0.3052      0.0861    3.5445    0.0004 [ 0.1497; 0.4547 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0411   16.1256    0.0000 [ 0.5902; 0.7171 ] 
#>   eta1 =~ y12      0.6493      0.0322   20.1378    0.0000 [ 0.5905; 0.7113 ] 
#>   eta1 =~ y13      0.7613      0.0359   21.2130    0.0000 [ 0.6927; 0.8175 ] 
#>   eta2 =~ y21      0.5165      0.0477   10.8353    0.0000 [ 0.4464; 0.6169 ] 
#>   eta2 =~ y22      0.7554      0.0407   18.5568    0.0000 [ 0.6769; 0.8194 ] 
#>   eta2 =~ y23      0.7997      0.0307   26.0065    0.0000 [ 0.7251; 0.8432 ] 
#>   eta3 =~ y31      0.8223      0.0344   23.8993    0.0000 [ 0.7535; 0.8756 ] 
#>   eta3 =~ y32      0.6581      0.0402   16.3763    0.0000 [ 0.5750; 0.7283 ] 
#>   eta3 =~ y33      0.7474      0.0437   17.1111    0.0000 [ 0.6797; 0.8249 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0191   20.7406    0.0000 [ 0.3600; 0.4298 ] 
#>   eta1 <~ y12      0.3873      0.0186   20.8219    0.0000 [ 0.3600; 0.4268 ] 
#>   eta1 <~ y13      0.4542      0.0211   21.5129    0.0000 [ 0.4163; 0.4892 ] 
#>   eta2 <~ y21      0.3058      0.0234   13.0559    0.0000 [ 0.2745; 0.3538 ] 
#>   eta2 <~ y22      0.4473      0.0241   18.5894    0.0000 [ 0.4011; 0.4938 ] 
#>   eta2 <~ y23      0.4735      0.0179   26.5100    0.0000 [ 0.4327; 0.5056 ] 
#>   eta3 <~ y31      0.4400      0.0218   20.1679    0.0000 [ 0.4071; 0.4845 ] 
#>   eta3 <~ y32      0.3521      0.0201   17.5191    0.0000 [ 0.3059; 0.3878 ] 
#>   eta3 <~ y33      0.3999      0.0195   20.4623    0.0000 [ 0.3679; 0.4362 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0453   14.8122    0.0000 [ 0.5862; 0.7413 ] 
#>   eta3 ~ eta1       0.6634      0.0390   17.0061    0.0000 [ 0.5891; 0.7143 ] 
#>   eta3 ~ eta2       0.3052      0.0861    3.5445    0.0004 [ 0.1497; 0.4547 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0578    3.5414    0.0004 [ 0.0952; 0.3047 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04111905 16.12561  1.685815e-58
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03224182 20.13776  3.445784e-90
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03589050 21.21302 7.241447e-100
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04766395 10.83533  2.341091e-27
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04070685 18.55677  7.189729e-77
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03074856 26.00654 4.176580e-149
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03440597 23.89926 3.117403e-126
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04018424 16.37629  2.824316e-60
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04368057 17.11113  1.225889e-65
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5901893          0.7171133
#> 2          0.5905402          0.7113233
#> 3          0.6927494          0.8175485
#> 4          0.4464124          0.6169012
#> 5          0.6768689          0.8193533
#> 6          0.7251262          0.8431981
#> 7          0.7534543          0.8756118
#> 8          0.5749633          0.7283130
#> 9          0.6797205          0.8249296

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 40678459
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0453   14.8122    0.0000 [ 0.5609; 0.7952 ] 
#>   eta3 ~ eta1      0.4585      0.0843    5.4396    0.0000 [ 0.2542; 0.6901 ] 
#>   eta3 ~ eta2      0.3052      0.0861    3.5445    0.0004 [ 0.0754; 0.5206 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0411   16.1256    0.0000 [ 0.5658; 0.7785 ] 
#>   eta1 =~ y12      0.6493      0.0322   20.1378    0.0000 [ 0.5709; 0.7376 ] 
#>   eta1 =~ y13      0.7613      0.0359   21.2130    0.0000 [ 0.6675; 0.8531 ] 
#>   eta2 =~ y21      0.5165      0.0477   10.8353    0.0000 [ 0.3899; 0.6364 ] 
#>   eta2 =~ y22      0.7554      0.0407   18.5568    0.0000 [ 0.6560; 0.8665 ] 
#>   eta2 =~ y23      0.7997      0.0307   26.0065    0.0000 [ 0.7193; 0.8783 ] 
#>   eta3 =~ y31      0.8223      0.0344   23.8993    0.0000 [ 0.7356; 0.9136 ] 
#>   eta3 =~ y32      0.6581      0.0402   16.3763    0.0000 [ 0.5584; 0.7662 ] 
#>   eta3 =~ y33      0.7474      0.0437   17.1111    0.0000 [ 0.6424; 0.8683 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0191   20.7406    0.0000 [ 0.3484; 0.4471 ] 
#>   eta1 <~ y12      0.3873      0.0186   20.8219    0.0000 [ 0.3387; 0.4349 ] 
#>   eta1 <~ y13      0.4542      0.0211   21.5129    0.0000 [ 0.3949; 0.5040 ] 
#>   eta2 <~ y21      0.3058      0.0234   13.0559    0.0000 [ 0.2433; 0.3644 ] 
#>   eta2 <~ y22      0.4473      0.0241   18.5894    0.0000 [ 0.3880; 0.5125 ] 
#>   eta2 <~ y23      0.4735      0.0179   26.5100    0.0000 [ 0.4263; 0.5187 ] 
#>   eta3 <~ y31      0.4400      0.0218   20.1679    0.0000 [ 0.3805; 0.4933 ] 
#>   eta3 <~ y32      0.3521      0.0201   17.5191    0.0000 [ 0.2992; 0.4031 ] 
#>   eta3 <~ y33      0.3999      0.0195   20.4623    0.0000 [ 0.3501; 0.4512 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0453   14.8122    0.0000 [ 0.5609; 0.7952 ] 
#>   eta3 ~ eta1       0.6634      0.0390   17.0061    0.0000 [ 0.5739; 0.7756 ] 
#>   eta3 ~ eta2       0.3052      0.0861    3.5445    0.0004 [ 0.0754; 0.5206 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0578    3.5414    0.0004 [ 0.0530; 0.3521 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04532290 14.812235 1.221120e-49
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08429086  5.439579 5.340675e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08609157  3.544495 3.933660e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56086339          0.7952478          0.5890084          0.7671028
#> 2         0.25423967          0.6901444          0.3065833          0.6378008
#> 3         0.07539402          0.5206110          0.1288559          0.4671491
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5712548          0.7723084          0.5861987          0.7413319
#> 2          0.2904120          0.6529708          0.3039212          0.6032117
#> 3          0.1051705          0.4712411          0.1497420          0.4547262