Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 596543445
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0426   15.7605    0.0000 [ 0.5958; 0.7447 ] 
#>   eta3 ~ eta1      0.4585      0.0894    5.1272    0.0000 [ 0.3051; 0.6334 ] 
#>   eta3 ~ eta2      0.3052      0.0945    3.2277    0.0012 [ 0.1381; 0.4615 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0411   16.1290    0.0000 [ 0.5974; 0.7185 ] 
#>   eta1 =~ y12      0.6493      0.0402   16.1373    0.0000 [ 0.5482; 0.7208 ] 
#>   eta1 =~ y13      0.7613      0.0338   22.5444    0.0000 [ 0.7017; 0.8221 ] 
#>   eta2 =~ y21      0.5165      0.0475   10.8801    0.0000 [ 0.4168; 0.5995 ] 
#>   eta2 =~ y22      0.7554      0.0353   21.4244    0.0000 [ 0.6759; 0.8067 ] 
#>   eta2 =~ y23      0.7997      0.0372   21.4922    0.0000 [ 0.7351; 0.8620 ] 
#>   eta3 =~ y31      0.8223      0.0295   27.8315    0.0000 [ 0.7846; 0.8885 ] 
#>   eta3 =~ y32      0.6581      0.0368   17.9051    0.0000 [ 0.5704; 0.7161 ] 
#>   eta3 =~ y33      0.7474      0.0443   16.8635    0.0000 [ 0.6561; 0.8249 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0230   17.1862    0.0000 [ 0.3506; 0.4348 ] 
#>   eta1 <~ y12      0.3873      0.0190   20.3917    0.0000 [ 0.3533; 0.4231 ] 
#>   eta1 <~ y13      0.4542      0.0197   23.0731    0.0000 [ 0.4221; 0.4901 ] 
#>   eta2 <~ y21      0.3058      0.0248   12.3169    0.0000 [ 0.2513; 0.3396 ] 
#>   eta2 <~ y22      0.4473      0.0199   22.4837    0.0000 [ 0.3982; 0.4784 ] 
#>   eta2 <~ y23      0.4735      0.0185   25.6463    0.0000 [ 0.4415; 0.5017 ] 
#>   eta3 <~ y31      0.4400      0.0193   22.7800    0.0000 [ 0.4125; 0.4749 ] 
#>   eta3 <~ y32      0.3521      0.0175   20.0821    0.0000 [ 0.3159; 0.3777 ] 
#>   eta3 <~ y33      0.3999      0.0206   19.4567    0.0000 [ 0.3558; 0.4330 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0426   15.7605    0.0000 [ 0.5958; 0.7447 ] 
#>   eta3 ~ eta1       0.6634      0.0382   17.3517    0.0000 [ 0.6068; 0.7478 ] 
#>   eta3 ~ eta2       0.3052      0.0945    3.2277    0.0012 [ 0.1381; 0.4615 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0663    3.0909    0.0020 [ 0.0995; 0.3182 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04111052 16.12896  1.596958e-58
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04023454 16.13733  1.394473e-58
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03377089 22.54444 1.522624e-112
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04746767 10.88014  1.433471e-27
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03525821 21.42445 7.906460e-102
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03720719 21.49218 1.842436e-102
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02954488 27.83147 1.805611e-170
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03675320 17.90508  1.076445e-71
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04432211 16.86346  8.355163e-64
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5973880          0.7185490
#> 2          0.5482409          0.7207803
#> 3          0.7016906          0.8220766
#> 4          0.4168349          0.5995029
#> 5          0.6759460          0.8066866
#> 6          0.7351454          0.8620196
#> 7          0.7845621          0.8884895
#> 8          0.5704338          0.7161023
#> 9          0.6560620          0.8248621

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 596543445
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0426   15.7605    0.0000 [ 0.5554; 0.7757 ] 
#>   eta3 ~ eta1      0.4585      0.0894    5.1272    0.0000 [ 0.2122; 0.6747 ] 
#>   eta3 ~ eta2      0.3052      0.0945    3.2277    0.0012 [ 0.0738; 0.5627 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0411   16.1290    0.0000 [ 0.5600; 0.7726 ] 
#>   eta1 =~ y12      0.6493      0.0402   16.1373    0.0000 [ 0.5466; 0.7547 ] 
#>   eta1 =~ y13      0.7613      0.0338   22.5444    0.0000 [ 0.6762; 0.8509 ] 
#>   eta2 =~ y21      0.5165      0.0475   10.8801    0.0000 [ 0.3957; 0.6412 ] 
#>   eta2 =~ y22      0.7554      0.0353   21.4244    0.0000 [ 0.6674; 0.8498 ] 
#>   eta2 =~ y23      0.7997      0.0372   21.4922    0.0000 [ 0.7019; 0.8943 ] 
#>   eta3 =~ y31      0.8223      0.0295   27.8315    0.0000 [ 0.7436; 0.8964 ] 
#>   eta3 =~ y32      0.6581      0.0368   17.9051    0.0000 [ 0.5623; 0.7524 ] 
#>   eta3 =~ y33      0.7474      0.0443   16.8635    0.0000 [ 0.6400; 0.8692 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0230   17.1862    0.0000 [ 0.3362; 0.4552 ] 
#>   eta1 <~ y12      0.3873      0.0190   20.3917    0.0000 [ 0.3375; 0.4358 ] 
#>   eta1 <~ y13      0.4542      0.0197   23.0731    0.0000 [ 0.4025; 0.5043 ] 
#>   eta2 <~ y21      0.3058      0.0248   12.3169    0.0000 [ 0.2424; 0.3708 ] 
#>   eta2 <~ y22      0.4473      0.0199   22.4837    0.0000 [ 0.3969; 0.4998 ] 
#>   eta2 <~ y23      0.4735      0.0185   25.6463    0.0000 [ 0.4240; 0.5195 ] 
#>   eta3 <~ y31      0.4400      0.0193   22.7800    0.0000 [ 0.3876; 0.4875 ] 
#>   eta3 <~ y32      0.3521      0.0175   20.0821    0.0000 [ 0.3056; 0.3963 ] 
#>   eta3 <~ y33      0.3999      0.0206   19.4567    0.0000 [ 0.3498; 0.4561 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0426   15.7605    0.0000 [ 0.5554; 0.7757 ] 
#>   eta3 ~ eta1       0.6634      0.0382   17.3517    0.0000 [ 0.5569; 0.7546 ] 
#>   eta3 ~ eta2       0.3052      0.0945    3.2277    0.0012 [ 0.0738; 0.5627 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0663    3.0909    0.0020 [ 0.0409; 0.3837 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04259607 15.760456 5.821122e-56
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08942607  5.127216 2.940587e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09454037  3.227734 1.247751e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55544247          0.7757253          0.5818941          0.7492736
#> 2         0.21220384          0.6746650          0.2677364          0.6191325
#> 3         0.07376237          0.5626718          0.1324708          0.5039633
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.59418076          0.7538057          0.5958287          0.7446863
#> 2         0.29348207          0.6746908          0.3050901          0.6333596
#> 3         0.08333889          0.5079009          0.1381018          0.4615483