Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2078801711
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0380   17.6626    0.0000 [ 0.6052; 0.7409 ] 
#>   eta3 ~ eta1      0.4585      0.0628    7.2955    0.0000 [ 0.3705; 0.5900 ] 
#>   eta3 ~ eta2      0.3052      0.0659    4.6316    0.0000 [ 0.1684; 0.4013 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0402   16.4991    0.0000 [ 0.5923; 0.7410 ] 
#>   eta1 =~ y12      0.6493      0.0361   17.9851    0.0000 [ 0.5972; 0.7219 ] 
#>   eta1 =~ y13      0.7613      0.0347   21.9457    0.0000 [ 0.7065; 0.8320 ] 
#>   eta2 =~ y21      0.5165      0.0533    9.6925    0.0000 [ 0.4260; 0.5975 ] 
#>   eta2 =~ y22      0.7554      0.0308   24.4912    0.0000 [ 0.7102; 0.8040 ] 
#>   eta2 =~ y23      0.7997      0.0398   20.0993    0.0000 [ 0.7330; 0.8688 ] 
#>   eta3 =~ y31      0.8223      0.0357   23.0084    0.0000 [ 0.7410; 0.8724 ] 
#>   eta3 =~ y32      0.6581      0.0371   17.7376    0.0000 [ 0.6086; 0.7481 ] 
#>   eta3 =~ y33      0.7474      0.0346   21.6101    0.0000 [ 0.6819; 0.8037 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0218   18.1088    0.0000 [ 0.3467; 0.4307 ] 
#>   eta1 <~ y12      0.3873      0.0200   19.3740    0.0000 [ 0.3540; 0.4225 ] 
#>   eta1 <~ y13      0.4542      0.0174   26.0642    0.0000 [ 0.4350; 0.4905 ] 
#>   eta2 <~ y21      0.3058      0.0265   11.5611    0.0000 [ 0.2545; 0.3533 ] 
#>   eta2 <~ y22      0.4473      0.0201   22.2903    0.0000 [ 0.4106; 0.4769 ] 
#>   eta2 <~ y23      0.4735      0.0230   20.5683    0.0000 [ 0.4326; 0.5121 ] 
#>   eta3 <~ y31      0.4400      0.0160   27.5134    0.0000 [ 0.4078; 0.4558 ] 
#>   eta3 <~ y32      0.3521      0.0171   20.6417    0.0000 [ 0.3311; 0.3904 ] 
#>   eta3 <~ y33      0.3999      0.0183   21.8420    0.0000 [ 0.3628; 0.4347 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0380   17.6626    0.0000 [ 0.6052; 0.7409 ] 
#>   eta3 ~ eta1       0.6634      0.0346   19.1972    0.0000 [ 0.6129; 0.7312 ] 
#>   eta3 ~ eta2       0.3052      0.0659    4.6316    0.0000 [ 0.1684; 0.4013 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0458    4.4718    0.0000 [ 0.1090; 0.2804 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04018823 16.499106  3.723672e-61
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03610094 17.985070  2.550649e-72
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03469221 21.945725 9.513871e-107
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05328370  9.692548  3.243341e-22
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03084317 24.491243 1.831189e-132
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03978559 20.099329  7.479862e-90
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03573813 23.008404 3.840451e-117
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03710014 17.737640  2.147669e-70
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03458674 21.610134 1.442297e-103
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5923397          0.7410291
#> 2          0.5971961          0.7218537
#> 3          0.7064520          0.8319993
#> 4          0.4259769          0.5974982
#> 5          0.7102256          0.8040401
#> 6          0.7330463          0.8687717
#> 7          0.7410414          0.8724205
#> 8          0.6086447          0.7480828
#> 9          0.6819449          0.8036836

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2078801711
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0380   17.6626    0.0000 [ 0.5763; 0.7728 ] 
#>   eta3 ~ eta1      0.4585      0.0628    7.2955    0.0000 [ 0.2827; 0.6077 ] 
#>   eta3 ~ eta2      0.3052      0.0659    4.6316    0.0000 [ 0.1408; 0.4815 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0402   16.4991    0.0000 [ 0.5670; 0.7748 ] 
#>   eta1 =~ y12      0.6493      0.0361   17.9851    0.0000 [ 0.5545; 0.7412 ] 
#>   eta1 =~ y13      0.7613      0.0347   21.9457    0.0000 [ 0.6724; 0.8519 ] 
#>   eta2 =~ y21      0.5165      0.0533    9.6925    0.0000 [ 0.3862; 0.6617 ] 
#>   eta2 =~ y22      0.7554      0.0308   24.4912    0.0000 [ 0.6691; 0.8286 ] 
#>   eta2 =~ y23      0.7997      0.0398   20.0993    0.0000 [ 0.6972; 0.9029 ] 
#>   eta3 =~ y31      0.8223      0.0357   23.0084    0.0000 [ 0.7377; 0.9226 ] 
#>   eta3 =~ y32      0.6581      0.0371   17.7376    0.0000 [ 0.5541; 0.7460 ] 
#>   eta3 =~ y33      0.7474      0.0346   21.6101    0.0000 [ 0.6586; 0.8375 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0218   18.1088    0.0000 [ 0.3419; 0.4549 ] 
#>   eta1 <~ y12      0.3873      0.0200   19.3740    0.0000 [ 0.3330; 0.4363 ] 
#>   eta1 <~ y13      0.4542      0.0174   26.0642    0.0000 [ 0.4075; 0.4976 ] 
#>   eta2 <~ y21      0.3058      0.0265   11.5611    0.0000 [ 0.2425; 0.3793 ] 
#>   eta2 <~ y22      0.4473      0.0201   22.2903    0.0000 [ 0.3917; 0.4955 ] 
#>   eta2 <~ y23      0.4735      0.0230   20.5683    0.0000 [ 0.4146; 0.5336 ] 
#>   eta3 <~ y31      0.4400      0.0160   27.5134    0.0000 [ 0.4027; 0.4854 ] 
#>   eta3 <~ y32      0.3521      0.0171   20.6417    0.0000 [ 0.3036; 0.3919 ] 
#>   eta3 <~ y33      0.3999      0.0183   21.8420    0.0000 [ 0.3527; 0.4474 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0380   17.6626    0.0000 [ 0.5763; 0.7728 ] 
#>   eta3 ~ eta1       0.6634      0.0346   19.1972    0.0000 [ 0.5659; 0.7446 ] 
#>   eta3 ~ eta2       0.3052      0.0659    4.6316    0.0000 [ 0.1408; 0.4815 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0458    4.4718    0.0000 [ 0.0916; 0.3285 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03800881 17.662572 8.144244e-70
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.06284800  7.295487 2.975800e-13
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.06588400  4.631642 3.627767e-06
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5762663          0.7728265          0.5998694          0.7492234
#> 2          0.2826766          0.6076910          0.3217045          0.5686631
#> 3          0.1407765          0.4814914          0.1816897          0.4405782
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5747023          0.7473496          0.6051963          0.7408998
#> 2          0.3569381          0.6425736          0.3705190          0.5899636
#> 3          0.1052463          0.4095299          0.1684154          0.4012679