Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 433802086
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0431   15.5668    0.0000 [ 0.6023; 0.7596 ] 
#>   eta3 ~ eta1      0.4585      0.0825    5.5574    0.0000 [ 0.3383; 0.6221 ] 
#>   eta3 ~ eta2      0.3052      0.0878    3.4760    0.0005 [ 0.1332; 0.4632 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0467   14.1924    0.0000 [ 0.5820; 0.7544 ] 
#>   eta1 =~ y12      0.6493      0.0400   16.2179    0.0000 [ 0.5884; 0.7173 ] 
#>   eta1 =~ y13      0.7613      0.0249   30.6250    0.0000 [ 0.7245; 0.8112 ] 
#>   eta2 =~ y21      0.5165      0.0440   11.7243    0.0000 [ 0.4314; 0.5895 ] 
#>   eta2 =~ y22      0.7554      0.0331   22.8413    0.0000 [ 0.6804; 0.8227 ] 
#>   eta2 =~ y23      0.7997      0.0342   23.3756    0.0000 [ 0.7414; 0.8526 ] 
#>   eta3 =~ y31      0.8223      0.0405   20.3250    0.0000 [ 0.7475; 0.8835 ] 
#>   eta3 =~ y32      0.6581      0.0361   18.2233    0.0000 [ 0.5959; 0.7084 ] 
#>   eta3 =~ y33      0.7474      0.0388   19.2743    0.0000 [ 0.6865; 0.8207 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0205   19.2937    0.0000 [ 0.3484; 0.4248 ] 
#>   eta1 <~ y12      0.3873      0.0175   22.0868    0.0000 [ 0.3521; 0.4156 ] 
#>   eta1 <~ y13      0.4542      0.0246   18.4705    0.0000 [ 0.4162; 0.4989 ] 
#>   eta2 <~ y21      0.3058      0.0198   15.4248    0.0000 [ 0.2769; 0.3466 ] 
#>   eta2 <~ y22      0.4473      0.0196   22.8273    0.0000 [ 0.4179; 0.4866 ] 
#>   eta2 <~ y23      0.4735      0.0191   24.7290    0.0000 [ 0.4372; 0.5142 ] 
#>   eta3 <~ y31      0.4400      0.0178   24.7642    0.0000 [ 0.3946; 0.4712 ] 
#>   eta3 <~ y32      0.3521      0.0167   21.0260    0.0000 [ 0.3238; 0.3781 ] 
#>   eta3 <~ y33      0.3999      0.0219   18.2992    0.0000 [ 0.3674; 0.4334 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0431   15.5668    0.0000 [ 0.6023; 0.7596 ] 
#>   eta3 ~ eta1       0.6634      0.0416   15.9297    0.0000 [ 0.6048; 0.7399 ] 
#>   eta3 ~ eta2       0.3052      0.0878    3.4760    0.0005 [ 0.1332; 0.4632 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0565    3.6282    0.0003 [ 0.0953; 0.2935 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04672005 14.19241  1.020943e-45
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04003464 16.21790  3.768411e-59
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02486026 30.62501 5.686477e-206
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04404992 11.72431  9.568126e-32
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03307120 22.84125 1.785427e-115
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03420930 23.37562 7.565480e-121
#> 7 eta3 =~ y31  Common factor 0.8222773 0.04045641 20.32502  7.725784e-92
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03611135 18.22333  3.370353e-74
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03877819 19.27434  8.821034e-83
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5819957          0.7544330
#> 2          0.5884349          0.7172803
#> 3          0.7244643          0.8111745
#> 4          0.4313685          0.5894972
#> 5          0.6803864          0.8226536
#> 6          0.7414415          0.8525928
#> 7          0.7475226          0.8834718
#> 8          0.5959335          0.7083730
#> 9          0.6864731          0.8206503

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 433802086
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0431   15.5668    0.0000 [ 0.5563; 0.7793 ] 
#>   eta3 ~ eta1      0.4585      0.0825    5.5574    0.0000 [ 0.2283; 0.6550 ] 
#>   eta3 ~ eta2      0.3052      0.0878    3.4760    0.0005 [ 0.0922; 0.5462 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0467   14.1924    0.0000 [ 0.5468; 0.7885 ] 
#>   eta1 =~ y12      0.6493      0.0400   16.2179    0.0000 [ 0.5439; 0.7509 ] 
#>   eta1 =~ y13      0.7613      0.0249   30.6250    0.0000 [ 0.6937; 0.8223 ] 
#>   eta2 =~ y21      0.5165      0.0440   11.7243    0.0000 [ 0.4027; 0.6305 ] 
#>   eta2 =~ y22      0.7554      0.0331   22.8413    0.0000 [ 0.6722; 0.8432 ] 
#>   eta2 =~ y23      0.7997      0.0342   23.3756    0.0000 [ 0.7118; 0.8887 ] 
#>   eta3 =~ y31      0.8223      0.0405   20.3250    0.0000 [ 0.7237; 0.9329 ] 
#>   eta3 =~ y32      0.6581      0.0361   18.2233    0.0000 [ 0.5662; 0.7529 ] 
#>   eta3 =~ y33      0.7474      0.0388   19.2743    0.0000 [ 0.6389; 0.8394 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0205   19.2937    0.0000 [ 0.3460; 0.4521 ] 
#>   eta1 <~ y12      0.3873      0.0175   22.0868    0.0000 [ 0.3415; 0.4322 ] 
#>   eta1 <~ y13      0.4542      0.0246   18.4705    0.0000 [ 0.3886; 0.5157 ] 
#>   eta2 <~ y21      0.3058      0.0198   15.4248    0.0000 [ 0.2545; 0.3570 ] 
#>   eta2 <~ y22      0.4473      0.0196   22.8273    0.0000 [ 0.3970; 0.4984 ] 
#>   eta2 <~ y23      0.4735      0.0191   24.7290    0.0000 [ 0.4234; 0.5224 ] 
#>   eta3 <~ y31      0.4400      0.0178   24.7642    0.0000 [ 0.3977; 0.4896 ] 
#>   eta3 <~ y32      0.3521      0.0167   21.0260    0.0000 [ 0.3100; 0.3966 ] 
#>   eta3 <~ y33      0.3999      0.0219   18.2992    0.0000 [ 0.3392; 0.4522 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0431   15.5668    0.0000 [ 0.5563; 0.7793 ] 
#>   eta3 ~ eta1       0.6634      0.0416   15.9297    0.0000 [ 0.5486; 0.7639 ] 
#>   eta3 ~ eta2       0.3052      0.0878    3.4760    0.0005 [ 0.0922; 0.5462 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0565    3.6282    0.0003 [ 0.0686; 0.3606 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04312590 15.566826 1.223286e-54
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08250355  5.557419 2.737934e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08778812  3.475996 5.089606e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55627787          0.7793007          0.5830586          0.7525200
#> 2         0.22828962          0.6549514          0.2795234          0.6037177
#> 3         0.09223731          0.5462279          0.1467527          0.4917125
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.6011945          0.7793516          0.6023193          0.7596391
#> 2          0.2758136          0.6543563          0.3382882          0.6220965
#> 3          0.1162497          0.5162823          0.1332261          0.4631688