Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -681903872
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0472   14.2273    0.0000 [ 0.5959; 0.7511 ] 
#>   eta3 ~ eta1      0.4585      0.0752    6.0955    0.0000 [ 0.3256; 0.5884 ] 
#>   eta3 ~ eta2      0.3052      0.0850    3.5885    0.0003 [ 0.1539; 0.4800 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0430   15.4140    0.0000 [ 0.5959; 0.7411 ] 
#>   eta1 =~ y12      0.6493      0.0401   16.2044    0.0000 [ 0.5623; 0.7029 ] 
#>   eta1 =~ y13      0.7613      0.0330   23.0774    0.0000 [ 0.7014; 0.8172 ] 
#>   eta2 =~ y21      0.5165      0.0422   12.2259    0.0000 [ 0.4700; 0.6213 ] 
#>   eta2 =~ y22      0.7554      0.0417   18.1265    0.0000 [ 0.6757; 0.8322 ] 
#>   eta2 =~ y23      0.7997      0.0380   21.0517    0.0000 [ 0.7060; 0.8438 ] 
#>   eta3 =~ y31      0.8223      0.0354   23.2439    0.0000 [ 0.7574; 0.8758 ] 
#>   eta3 =~ y32      0.6581      0.0337   19.5204    0.0000 [ 0.5797; 0.7072 ] 
#>   eta3 =~ y33      0.7474      0.0409   18.2858    0.0000 [ 0.6779; 0.8140 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0195   20.2449    0.0000 [ 0.3620; 0.4359 ] 
#>   eta1 <~ y12      0.3873      0.0186   20.8774    0.0000 [ 0.3545; 0.4196 ] 
#>   eta1 <~ y13      0.4542      0.0239   19.0346    0.0000 [ 0.4212; 0.5054 ] 
#>   eta2 <~ y21      0.3058      0.0244   12.5587    0.0000 [ 0.2821; 0.3595 ] 
#>   eta2 <~ y22      0.4473      0.0213   20.9758    0.0000 [ 0.4100; 0.4895 ] 
#>   eta2 <~ y23      0.4735      0.0197   24.0412    0.0000 [ 0.4351; 0.4973 ] 
#>   eta3 <~ y31      0.4400      0.0215   20.4348    0.0000 [ 0.4010; 0.4799 ] 
#>   eta3 <~ y32      0.3521      0.0160   22.0398    0.0000 [ 0.3258; 0.3798 ] 
#>   eta3 <~ y33      0.3999      0.0181   22.1240    0.0000 [ 0.3731; 0.4359 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0472   14.2273    0.0000 [ 0.5959; 0.7511 ] 
#>   eta3 ~ eta1       0.6634      0.0323   20.5505    0.0000 [ 0.6020; 0.7182 ] 
#>   eta3 ~ eta2       0.3052      0.0850    3.5885    0.0003 [ 0.1539; 0.4800 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0608    3.3685    0.0008 [ 0.1088; 0.3400 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04301736 15.41401  1.317808e-53
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04006799 16.20441  4.693967e-59
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03299098 23.07739 7.810928e-118
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04224271 12.22589  2.260885e-34
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04167304 18.12653  1.967839e-73
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03798575 21.05168  2.207335e-98
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03537611 23.24386 1.641162e-119
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03371191 19.52037  7.370873e-85
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04087460 18.28578  1.074096e-74
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5958747          0.7410832
#> 2          0.5622841          0.7029497
#> 3          0.7013577          0.8171681
#> 4          0.4699736          0.6212561
#> 5          0.6757326          0.8321686
#> 6          0.7060374          0.8438124
#> 7          0.7574299          0.8757628
#> 8          0.5796717          0.7071721
#> 9          0.6779427          0.8139920

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -681903872
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0472   14.2273    0.0000 [ 0.5464; 0.7904 ] 
#>   eta3 ~ eta1      0.4585      0.0752    6.0955    0.0000 [ 0.2542; 0.6432 ] 
#>   eta3 ~ eta2      0.3052      0.0850    3.5885    0.0003 [ 0.0958; 0.5356 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0430   15.4140    0.0000 [ 0.5519; 0.7744 ] 
#>   eta1 =~ y12      0.6493      0.0401   16.2044    0.0000 [ 0.5451; 0.7523 ] 
#>   eta1 =~ y13      0.7613      0.0330   23.0774    0.0000 [ 0.6822; 0.8528 ] 
#>   eta2 =~ y21      0.5165      0.0422   12.2259    0.0000 [ 0.3928; 0.6113 ] 
#>   eta2 =~ y22      0.7554      0.0417   18.1265    0.0000 [ 0.6553; 0.8709 ] 
#>   eta2 =~ y23      0.7997      0.0380   21.0517    0.0000 [ 0.7116; 0.9080 ] 
#>   eta3 =~ y31      0.8223      0.0354   23.2439    0.0000 [ 0.7325; 0.9155 ] 
#>   eta3 =~ y32      0.6581      0.0337   19.5204    0.0000 [ 0.5755; 0.7499 ] 
#>   eta3 =~ y33      0.7474      0.0409   18.2858    0.0000 [ 0.6438; 0.8552 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0195   20.2449    0.0000 [ 0.3438; 0.4448 ] 
#>   eta1 <~ y12      0.3873      0.0186   20.8774    0.0000 [ 0.3377; 0.4337 ] 
#>   eta1 <~ y13      0.4542      0.0239   19.0346    0.0000 [ 0.3941; 0.5175 ] 
#>   eta2 <~ y21      0.3058      0.0244   12.5587    0.0000 [ 0.2331; 0.3591 ] 
#>   eta2 <~ y22      0.4473      0.0213   20.9758    0.0000 [ 0.3951; 0.5054 ] 
#>   eta2 <~ y23      0.4735      0.0197   24.0412    0.0000 [ 0.4269; 0.5287 ] 
#>   eta3 <~ y31      0.4400      0.0215   20.4348    0.0000 [ 0.3827; 0.4941 ] 
#>   eta3 <~ y32      0.3521      0.0160   22.0398    0.0000 [ 0.3115; 0.3941 ] 
#>   eta3 <~ y33      0.3999      0.0181   22.1240    0.0000 [ 0.3523; 0.4458 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0472   14.2273    0.0000 [ 0.5464; 0.7904 ] 
#>   eta3 ~ eta1       0.6634      0.0323   20.5505    0.0000 [ 0.5759; 0.7428 ] 
#>   eta3 ~ eta2       0.3052      0.0850    3.5885    0.0003 [ 0.0958; 0.5356 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0608    3.3685    0.0008 [ 0.0534; 0.3679 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04718615 14.227342 6.199381e-46
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07522048  6.095504 1.090930e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08503691  3.588455 3.326432e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54637169          0.7903918          0.5756737          0.7610897
#> 2         0.25419971          0.6431976          0.3009108          0.5964865
#> 3         0.09580375          0.5355666          0.1486107          0.4827597
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5955249          0.7527421          0.5958891          0.7510752
#> 2          0.3127051          0.6382519          0.3255900          0.5883567
#> 3          0.0817661          0.5180379          0.1539089          0.4800459