Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1018387326
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0344   19.5174    0.0000 [ 0.5950; 0.7253 ] 
#>   eta3 ~ eta1      0.4585      0.0643    7.1361    0.0000 [ 0.3467; 0.5574 ] 
#>   eta3 ~ eta2      0.3052      0.0685    4.4545    0.0000 [ 0.1740; 0.4206 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0467   14.1987    0.0000 [ 0.5720; 0.7249 ] 
#>   eta1 =~ y12      0.6493      0.0360   18.0300    0.0000 [ 0.5554; 0.7130 ] 
#>   eta1 =~ y13      0.7613      0.0356   21.3822    0.0000 [ 0.7017; 0.8307 ] 
#>   eta2 =~ y21      0.5165      0.0506   10.2059    0.0000 [ 0.4185; 0.5905 ] 
#>   eta2 =~ y22      0.7554      0.0335   22.5565    0.0000 [ 0.6735; 0.8130 ] 
#>   eta2 =~ y23      0.7997      0.0267   29.9856    0.0000 [ 0.7618; 0.8492 ] 
#>   eta3 =~ y31      0.8223      0.0239   34.3820    0.0000 [ 0.7878; 0.8643 ] 
#>   eta3 =~ y32      0.6581      0.0435   15.1315    0.0000 [ 0.5861; 0.7471 ] 
#>   eta3 =~ y33      0.7474      0.0406   18.4124    0.0000 [ 0.6691; 0.8179 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0243   16.2650    0.0000 [ 0.3458; 0.4357 ] 
#>   eta1 <~ y12      0.3873      0.0191   20.2868    0.0000 [ 0.3584; 0.4214 ] 
#>   eta1 <~ y13      0.4542      0.0219   20.7856    0.0000 [ 0.4225; 0.4988 ] 
#>   eta2 <~ y21      0.3058      0.0241   12.6713    0.0000 [ 0.2574; 0.3430 ] 
#>   eta2 <~ y22      0.4473      0.0207   21.5601    0.0000 [ 0.4132; 0.4780 ] 
#>   eta2 <~ y23      0.4735      0.0183   25.8776    0.0000 [ 0.4406; 0.5175 ] 
#>   eta3 <~ y31      0.4400      0.0145   30.2895    0.0000 [ 0.4083; 0.4690 ] 
#>   eta3 <~ y32      0.3521      0.0221   15.9465    0.0000 [ 0.3032; 0.3953 ] 
#>   eta3 <~ y33      0.3999      0.0183   21.8429    0.0000 [ 0.3736; 0.4392 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0344   19.5174    0.0000 [ 0.5950; 0.7253 ] 
#>   eta3 ~ eta1       0.6634      0.0355   18.6707    0.0000 [ 0.5999; 0.7180 ] 
#>   eta3 ~ eta2       0.3052      0.0685    4.4545    0.0000 [ 0.1740; 0.4206 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0483    4.2390    0.0000 [ 0.1109; 0.2855 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04669941 14.19868  9.335737e-46
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03601100 18.02999  1.133175e-72
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03560652 21.38220 1.956745e-101
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05060356 10.20590  1.865857e-24
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03348870 22.55649 1.159611e-112
#> 6 eta2 =~ y23  Common factor 0.7996637 0.02666822 29.98564 1.510259e-197
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02391594 34.38198 4.688953e-259
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04349004 15.13149  1.004029e-51
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04059354 18.41239  1.045021e-75
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5719792          0.7248745
#> 2          0.5553520          0.7130211
#> 3          0.7017293          0.8307161
#> 4          0.4185211          0.5905465
#> 5          0.6735138          0.8130383
#> 6          0.7618492          0.8491653
#> 7          0.7878321          0.8643193
#> 8          0.5860586          0.7471041
#> 9          0.6691000          0.8179100

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1018387326
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0344   19.5174    0.0000 [ 0.5881; 0.7660 ] 
#>   eta3 ~ eta1      0.4585      0.0643    7.1361    0.0000 [ 0.2940; 0.6263 ] 
#>   eta3 ~ eta2      0.3052      0.0685    4.4545    0.0000 [ 0.1194; 0.4736 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0467   14.1987    0.0000 [ 0.5508; 0.7923 ] 
#>   eta1 =~ y12      0.6493      0.0360   18.0300    0.0000 [ 0.5592; 0.7454 ] 
#>   eta1 =~ y13      0.7613      0.0356   21.3822    0.0000 [ 0.6709; 0.8551 ] 
#>   eta2 =~ y21      0.5165      0.0506   10.2059    0.0000 [ 0.3991; 0.6608 ] 
#>   eta2 =~ y22      0.7554      0.0335   22.5565    0.0000 [ 0.6750; 0.8482 ] 
#>   eta2 =~ y23      0.7997      0.0267   29.9856    0.0000 [ 0.7259; 0.8638 ] 
#>   eta3 =~ y31      0.8223      0.0239   34.3820    0.0000 [ 0.7621; 0.8857 ] 
#>   eta3 =~ y32      0.6581      0.0435   15.1315    0.0000 [ 0.5451; 0.7700 ] 
#>   eta3 =~ y33      0.7474      0.0406   18.4124    0.0000 [ 0.6399; 0.8498 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0243   16.2650    0.0000 [ 0.3344; 0.4602 ] 
#>   eta1 <~ y12      0.3873      0.0191   20.2868    0.0000 [ 0.3364; 0.4351 ] 
#>   eta1 <~ y13      0.4542      0.0219   20.7856    0.0000 [ 0.3945; 0.5075 ] 
#>   eta2 <~ y21      0.3058      0.0241   12.6713    0.0000 [ 0.2493; 0.3741 ] 
#>   eta2 <~ y22      0.4473      0.0207   21.5601    0.0000 [ 0.3934; 0.5007 ] 
#>   eta2 <~ y23      0.4735      0.0183   25.8776    0.0000 [ 0.4191; 0.5137 ] 
#>   eta3 <~ y31      0.4400      0.0145   30.2895    0.0000 [ 0.4038; 0.4790 ] 
#>   eta3 <~ y32      0.3521      0.0221   15.9465    0.0000 [ 0.2953; 0.4095 ] 
#>   eta3 <~ y33      0.3999      0.0183   21.8429    0.0000 [ 0.3519; 0.4466 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0344   19.5174    0.0000 [ 0.5881; 0.7660 ] 
#>   eta3 ~ eta1       0.6634      0.0355   18.6707    0.0000 [ 0.5690; 0.7527 ] 
#>   eta3 ~ eta2       0.3052      0.0685    4.4545    0.0000 [ 0.1194; 0.4736 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0483    4.2390    0.0000 [ 0.0758; 0.3257 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03439673 19.517362 7.817110e-85
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.06425129  7.136149 9.598211e-13
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.06850433  4.454479 8.409698e-06
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5881296          0.7660100          0.6094895          0.7446501
#> 2          0.2940101          0.6262815          0.3339094          0.5863822
#> 3          0.1193708          0.4736365          0.1619112          0.4310961
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5915651          0.7300778          0.5949936          0.7253013
#> 2          0.3263146          0.5704495          0.3466676          0.5574494
#> 3          0.1571140          0.4473716          0.1739567          0.4206444