Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -1141257839
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0308   21.8049    0.0000 [ 0.6118; 0.7327 ] 
#>   eta3 ~ eta1      0.4585      0.0583    7.8645    0.0000 [ 0.3268; 0.5311 ] 
#>   eta3 ~ eta2      0.3052      0.0648    4.7071    0.0000 [ 0.2224; 0.4316 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0376   17.6454    0.0000 [ 0.5943; 0.7303 ] 
#>   eta1 =~ y12      0.6493      0.0389   16.7086    0.0000 [ 0.5788; 0.7155 ] 
#>   eta1 =~ y13      0.7613      0.0358   21.2393    0.0000 [ 0.6902; 0.8353 ] 
#>   eta2 =~ y21      0.5165      0.0388   13.3185    0.0000 [ 0.4523; 0.5698 ] 
#>   eta2 =~ y22      0.7554      0.0337   22.4107    0.0000 [ 0.6860; 0.8032 ] 
#>   eta2 =~ y23      0.7997      0.0379   21.1077    0.0000 [ 0.7397; 0.8697 ] 
#>   eta3 =~ y31      0.8223      0.0311   26.4160    0.0000 [ 0.7637; 0.8594 ] 
#>   eta3 =~ y32      0.6581      0.0399   16.4866    0.0000 [ 0.5924; 0.7276 ] 
#>   eta3 =~ y33      0.7474      0.0443   16.8878    0.0000 [ 0.6332; 0.8199 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0222   17.8152    0.0000 [ 0.3708; 0.4378 ] 
#>   eta1 <~ y12      0.3873      0.0181   21.3820    0.0000 [ 0.3545; 0.4161 ] 
#>   eta1 <~ y13      0.4542      0.0213   21.2950    0.0000 [ 0.4199; 0.4862 ] 
#>   eta2 <~ y21      0.3058      0.0203   15.0566    0.0000 [ 0.2613; 0.3400 ] 
#>   eta2 <~ y22      0.4473      0.0173   25.9250    0.0000 [ 0.4137; 0.4793 ] 
#>   eta2 <~ y23      0.4735      0.0201   23.5403    0.0000 [ 0.4398; 0.5066 ] 
#>   eta3 <~ y31      0.4400      0.0199   22.1094    0.0000 [ 0.4056; 0.4748 ] 
#>   eta3 <~ y32      0.3521      0.0193   18.2412    0.0000 [ 0.3276; 0.3809 ] 
#>   eta3 <~ y33      0.3999      0.0205   19.5351    0.0000 [ 0.3594; 0.4384 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0308   21.8049    0.0000 [ 0.6118; 0.7327 ] 
#>   eta3 ~ eta1       0.6634      0.0331   20.0667    0.0000 [ 0.5798; 0.7100 ] 
#>   eta3 ~ eta2       0.3052      0.0648    4.7071    0.0000 [ 0.2224; 0.4316 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0423    4.8430    0.0000 [ 0.1463; 0.2927 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03757746 17.64541  1.103617e-69
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03885883 16.70863  1.134045e-62
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03584605 21.23932 4.138409e-100
#> 4 eta2 =~ y21  Common factor 0.5164548 0.03877732 13.31848  1.807588e-40
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03370659 22.41068 3.096705e-111
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03788502 21.10765  6.765385e-99
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03112802 26.41598 8.978585e-154
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03991530 16.48663  4.577454e-61
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04425824 16.88780  5.533075e-64
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5943070          0.7303113
#> 2          0.5787707          0.7155029
#> 3          0.6901912          0.8352541
#> 4          0.4523375          0.5698192
#> 5          0.6860361          0.8032073
#> 6          0.7397297          0.8697017
#> 7          0.7636597          0.8594440
#> 8          0.5923627          0.7275997
#> 9          0.6331646          0.8199344

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -1141257839
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0308   21.8049    0.0000 [ 0.5913; 0.7505 ] 
#>   eta3 ~ eta1      0.4585      0.0583    7.8645    0.0000 [ 0.3233; 0.6248 ] 
#>   eta3 ~ eta2      0.3052      0.0648    4.7071    0.0000 [ 0.1187; 0.4540 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0376   17.6454    0.0000 [ 0.5651; 0.7595 ] 
#>   eta1 =~ y12      0.6493      0.0389   16.7086    0.0000 [ 0.5469; 0.7478 ] 
#>   eta1 =~ y13      0.7613      0.0358   21.2393    0.0000 [ 0.6719; 0.8573 ] 
#>   eta2 =~ y21      0.5165      0.0388   13.3185    0.0000 [ 0.4142; 0.6147 ] 
#>   eta2 =~ y22      0.7554      0.0337   22.4107    0.0000 [ 0.6735; 0.8478 ] 
#>   eta2 =~ y23      0.7997      0.0379   21.1077    0.0000 [ 0.7023; 0.8982 ] 
#>   eta3 =~ y31      0.8223      0.0311   26.4160    0.0000 [ 0.7511; 0.9121 ] 
#>   eta3 =~ y32      0.6581      0.0399   16.4866    0.0000 [ 0.5563; 0.7627 ] 
#>   eta3 =~ y33      0.7474      0.0443   16.8878    0.0000 [ 0.6238; 0.8526 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0222   17.8152    0.0000 [ 0.3376; 0.4525 ] 
#>   eta1 <~ y12      0.3873      0.0181   21.3820    0.0000 [ 0.3395; 0.4332 ] 
#>   eta1 <~ y13      0.4542      0.0213   21.2950    0.0000 [ 0.4009; 0.5112 ] 
#>   eta2 <~ y21      0.3058      0.0203   15.0566    0.0000 [ 0.2514; 0.3564 ] 
#>   eta2 <~ y22      0.4473      0.0173   25.9250    0.0000 [ 0.4046; 0.4938 ] 
#>   eta2 <~ y23      0.4735      0.0201   23.5403    0.0000 [ 0.4206; 0.5246 ] 
#>   eta3 <~ y31      0.4400      0.0199   22.1094    0.0000 [ 0.3931; 0.4960 ] 
#>   eta3 <~ y32      0.3521      0.0193   18.2412    0.0000 [ 0.3028; 0.4027 ] 
#>   eta3 <~ y33      0.3999      0.0205   19.5351    0.0000 [ 0.3419; 0.4478 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0308   21.8049    0.0000 [ 0.5913; 0.7505 ] 
#>   eta3 ~ eta1       0.6634      0.0331   20.0667    0.0000 [ 0.5811; 0.7520 ] 
#>   eta3 ~ eta2       0.3052      0.0648    4.7071    0.0000 [ 0.1187; 0.4540 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0423    4.8430    0.0000 [ 0.0831; 0.3019 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03078822 21.804879 2.085554e-105
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.05830110  7.864462  3.706863e-15
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.06482797  4.707091  2.512773e-06
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5912618          0.7504811          0.6103810          0.7313620
#> 2          0.3232965          0.6247968          0.3595008          0.5885925
#> 3          0.1187383          0.4539920          0.1589957          0.4137346
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.6070846          0.7329173          0.6118221          0.7327245
#> 2          0.3039240          0.5573435          0.3267749          0.5310645
#> 3          0.2172555          0.4755846          0.2224449          0.4315930