Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -151378221
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0448   15.0015    0.0000 [ 0.6006; 0.7595 ] 
#>   eta3 ~ eta1      0.4585      0.0976    4.6973    0.0000 [ 0.2520; 0.6173 ] 
#>   eta3 ~ eta2      0.3052      0.0991    3.0801    0.0021 [ 0.1527; 0.5391 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0430   15.4278    0.0000 [ 0.5827; 0.7300 ] 
#>   eta1 =~ y12      0.6493      0.0336   19.2953    0.0000 [ 0.6016; 0.7300 ] 
#>   eta1 =~ y13      0.7613      0.0356   21.3921    0.0000 [ 0.6888; 0.8076 ] 
#>   eta2 =~ y21      0.5165      0.0508   10.1597    0.0000 [ 0.4150; 0.6081 ] 
#>   eta2 =~ y22      0.7554      0.0385   19.6095    0.0000 [ 0.6839; 0.8282 ] 
#>   eta2 =~ y23      0.7997      0.0309   25.8918    0.0000 [ 0.7393; 0.8450 ] 
#>   eta3 =~ y31      0.8223      0.0391   21.0244    0.0000 [ 0.7566; 0.9112 ] 
#>   eta3 =~ y32      0.6581      0.0437   15.0571    0.0000 [ 0.5795; 0.7364 ] 
#>   eta3 =~ y33      0.7474      0.0401   18.6240    0.0000 [ 0.6600; 0.8208 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0211   18.7655    0.0000 [ 0.3557; 0.4229 ] 
#>   eta1 <~ y12      0.3873      0.0176   21.9684    0.0000 [ 0.3658; 0.4266 ] 
#>   eta1 <~ y13      0.4542      0.0214   21.2004    0.0000 [ 0.4058; 0.4888 ] 
#>   eta2 <~ y21      0.3058      0.0256   11.9659    0.0000 [ 0.2520; 0.3430 ] 
#>   eta2 <~ y22      0.4473      0.0190   23.4973    0.0000 [ 0.4139; 0.4881 ] 
#>   eta2 <~ y23      0.4735      0.0208   22.7802    0.0000 [ 0.4396; 0.5126 ] 
#>   eta3 <~ y31      0.4400      0.0204   21.5958    0.0000 [ 0.4039; 0.4769 ] 
#>   eta3 <~ y32      0.3521      0.0221   15.9550    0.0000 [ 0.3109; 0.3937 ] 
#>   eta3 <~ y33      0.3999      0.0196   20.4229    0.0000 [ 0.3536; 0.4281 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0448   15.0015    0.0000 [ 0.6006; 0.7595 ] 
#>   eta3 ~ eta1       0.6634      0.0410   16.1626    0.0000 [ 0.5908; 0.7321 ] 
#>   eta3 ~ eta2       0.3052      0.0991    3.0801    0.0021 [ 0.1527; 0.5391 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0689    2.9738    0.0029 [ 0.1130; 0.3907 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04297890 15.42780  1.064387e-53
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03364955 19.29529  5.882906e-83
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03558999 21.39213 1.581548e-101
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05083376 10.15968  3.000559e-24
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03852144 19.60954  1.281863e-85
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03088482 25.89180 8.237619e-148
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03911065 21.02439  3.924319e-98
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04370486 15.05711  3.100462e-51
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04013238 18.62397  2.054217e-77
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5827303          0.7300344
#> 2          0.6015816          0.7300239
#> 3          0.6888283          0.8076440
#> 4          0.4149918          0.6080960
#> 5          0.6839301          0.8281678
#> 6          0.7392676          0.8450495
#> 7          0.7565673          0.9111888
#> 8          0.5794858          0.7364466
#> 9          0.6600014          0.8208445

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -151378221
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0448   15.0015    0.0000 [ 0.5485; 0.7799 ] 
#>   eta3 ~ eta1      0.4585      0.0976    4.6973    0.0000 [ 0.2193; 0.7241 ] 
#>   eta3 ~ eta2      0.3052      0.0991    3.0801    0.0021 [ 0.0315; 0.5438 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0430   15.4278    0.0000 [ 0.5540; 0.7762 ] 
#>   eta1 =~ y12      0.6493      0.0336   19.2953    0.0000 [ 0.5534; 0.7274 ] 
#>   eta1 =~ y13      0.7613      0.0356   21.3921    0.0000 [ 0.6700; 0.8541 ] 
#>   eta2 =~ y21      0.5165      0.0508   10.1597    0.0000 [ 0.3769; 0.6398 ] 
#>   eta2 =~ y22      0.7554      0.0385   19.6095    0.0000 [ 0.6597; 0.8590 ] 
#>   eta2 =~ y23      0.7997      0.0309   25.8918    0.0000 [ 0.7205; 0.8802 ] 
#>   eta3 =~ y31      0.8223      0.0391   21.0244    0.0000 [ 0.7244; 0.9266 ] 
#>   eta3 =~ y32      0.6581      0.0437   15.0571    0.0000 [ 0.5430; 0.7691 ] 
#>   eta3 =~ y33      0.7474      0.0401   18.6240    0.0000 [ 0.6351; 0.8427 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0211   18.7655    0.0000 [ 0.3443; 0.4533 ] 
#>   eta1 <~ y12      0.3873      0.0176   21.9684    0.0000 [ 0.3383; 0.4295 ] 
#>   eta1 <~ y13      0.4542      0.0214   21.2004    0.0000 [ 0.4012; 0.5120 ] 
#>   eta2 <~ y21      0.3058      0.0256   11.9659    0.0000 [ 0.2360; 0.3681 ] 
#>   eta2 <~ y22      0.4473      0.0190   23.4973    0.0000 [ 0.4015; 0.4999 ] 
#>   eta2 <~ y23      0.4735      0.0208   22.7802    0.0000 [ 0.4211; 0.5285 ] 
#>   eta3 <~ y31      0.4400      0.0204   21.5958    0.0000 [ 0.3914; 0.4967 ] 
#>   eta3 <~ y32      0.3521      0.0221   15.9550    0.0000 [ 0.2960; 0.4101 ] 
#>   eta3 <~ y33      0.3999      0.0196   20.4229    0.0000 [ 0.3470; 0.4482 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0448   15.0015    0.0000 [ 0.5485; 0.7799 ] 
#>   eta3 ~ eta1       0.6634      0.0410   16.1626    0.0000 [ 0.5571; 0.7694 ] 
#>   eta3 ~ eta2       0.3052      0.0991    3.0801    0.0021 [ 0.0315; 0.5438 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0689    2.9738    0.0029 [ 0.0134; 0.3696 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04475110 15.001497 7.178215e-51
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09761062  4.697304 2.636178e-06
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09907093  3.080128 2.069119e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54850162          0.7799290         0.57629152          0.7521391
#> 2         0.21934090          0.7241279         0.27995595          0.6635129
#> 3         0.03146409          0.5438030         0.09298598          0.4822811
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5978294          0.7716266          0.6006167          0.7594750
#> 2          0.1856073          0.6291609          0.2520343          0.6173208
#> 3          0.1375062          0.5650288          0.1526978          0.5391115