Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 480202620
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0462   14.5223    0.0000 [ 0.6090; 0.7701 ] 
#>   eta3 ~ eta1      0.4585      0.0871    5.2666    0.0000 [ 0.3096; 0.6655 ] 
#>   eta3 ~ eta2      0.3052      0.0855    3.5681    0.0004 [ 0.1232; 0.4468 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0350   18.9421    0.0000 [ 0.5994; 0.7262 ] 
#>   eta1 =~ y12      0.6493      0.0434   14.9734    0.0000 [ 0.5410; 0.7066 ] 
#>   eta1 =~ y13      0.7613      0.0289   26.3674    0.0000 [ 0.7011; 0.8085 ] 
#>   eta2 =~ y21      0.5165      0.0596    8.6603    0.0000 [ 0.3588; 0.5928 ] 
#>   eta2 =~ y22      0.7554      0.0388   19.4875    0.0000 [ 0.6775; 0.8091 ] 
#>   eta2 =~ y23      0.7997      0.0393   20.3504    0.0000 [ 0.7444; 0.8740 ] 
#>   eta3 =~ y31      0.8223      0.0346   23.7368    0.0000 [ 0.7771; 0.8800 ] 
#>   eta3 =~ y32      0.6581      0.0367   17.9229    0.0000 [ 0.5975; 0.7239 ] 
#>   eta3 =~ y33      0.7474      0.0319   23.4188    0.0000 [ 0.6919; 0.8172 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0195   20.3270    0.0000 [ 0.3621; 0.4335 ] 
#>   eta1 <~ y12      0.3873      0.0205   18.9133    0.0000 [ 0.3415; 0.4157 ] 
#>   eta1 <~ y13      0.4542      0.0209   21.7040    0.0000 [ 0.4223; 0.5004 ] 
#>   eta2 <~ y21      0.3058      0.0281   10.8671    0.0000 [ 0.2423; 0.3483 ] 
#>   eta2 <~ y22      0.4473      0.0237   18.9021    0.0000 [ 0.4173; 0.5105 ] 
#>   eta2 <~ y23      0.4735      0.0252   18.7918    0.0000 [ 0.4348; 0.5226 ] 
#>   eta3 <~ y31      0.4400      0.0177   24.8070    0.0000 [ 0.4108; 0.4704 ] 
#>   eta3 <~ y32      0.3521      0.0179   19.6834    0.0000 [ 0.3209; 0.3785 ] 
#>   eta3 <~ y33      0.3999      0.0153   26.1396    0.0000 [ 0.3711; 0.4252 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0462   14.5223    0.0000 [ 0.6090; 0.7701 ] 
#>   eta3 ~ eta1       0.6634      0.0459   14.4523    0.0000 [ 0.5968; 0.7525 ] 
#>   eta3 ~ eta2       0.3052      0.0855    3.5681    0.0004 [ 0.1232; 0.4468 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0576    3.5543    0.0004 [ 0.0834; 0.2865 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03500505 18.942118  5.129172e-80
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04336196 14.973446  1.094971e-50
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02887447 26.367440 3.238891e-153
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05963452  8.660333  4.703897e-18
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03876259 19.487543  1.400470e-84
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03929480 20.350370  4.607869e-92
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03464141 23.736833 1.502830e-124
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03671670 17.922877  7.817852e-72
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03191549 23.418849 2.746715e-121
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5993841          0.7261978
#> 2          0.5410282          0.7066233
#> 3          0.7011089          0.8085344
#> 4          0.3588476          0.5928176
#> 5          0.6775096          0.8091281
#> 6          0.7444123          0.8740411
#> 7          0.7771468          0.8799801
#> 8          0.5975400          0.7238565
#> 9          0.6919075          0.8171674

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 480202620
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0462   14.5223    0.0000 [ 0.5406; 0.7797 ] 
#>   eta3 ~ eta1      0.4585      0.0871    5.2666    0.0000 [ 0.1921; 0.6423 ] 
#>   eta3 ~ eta2      0.3052      0.0855    3.5681    0.0004 [ 0.1269; 0.5692 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0350   18.9421    0.0000 [ 0.5675; 0.7485 ] 
#>   eta1 =~ y12      0.6493      0.0434   14.9734    0.0000 [ 0.5502; 0.7744 ] 
#>   eta1 =~ y13      0.7613      0.0289   26.3674    0.0000 [ 0.6925; 0.8418 ] 
#>   eta2 =~ y21      0.5165      0.0596    8.6603    0.0000 [ 0.3807; 0.6891 ] 
#>   eta2 =~ y22      0.7554      0.0388   19.4875    0.0000 [ 0.6478; 0.8483 ] 
#>   eta2 =~ y23      0.7997      0.0393   20.3504    0.0000 [ 0.7034; 0.9066 ] 
#>   eta3 =~ y31      0.8223      0.0346   23.7368    0.0000 [ 0.7324; 0.9116 ] 
#>   eta3 =~ y32      0.6581      0.0367   17.9229    0.0000 [ 0.5550; 0.7449 ] 
#>   eta3 =~ y33      0.7474      0.0319   23.4188    0.0000 [ 0.6654; 0.8305 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0195   20.3270    0.0000 [ 0.3385; 0.4391 ] 
#>   eta1 <~ y12      0.3873      0.0205   18.9133    0.0000 [ 0.3389; 0.4448 ] 
#>   eta1 <~ y13      0.4542      0.0209   21.7040    0.0000 [ 0.3991; 0.5073 ] 
#>   eta2 <~ y21      0.3058      0.0281   10.8671    0.0000 [ 0.2421; 0.3876 ] 
#>   eta2 <~ y22      0.4473      0.0237   18.9021    0.0000 [ 0.3776; 0.4999 ] 
#>   eta2 <~ y23      0.4735      0.0252   18.7918    0.0000 [ 0.4071; 0.5374 ] 
#>   eta3 <~ y31      0.4400      0.0177   24.8070    0.0000 [ 0.3964; 0.4881 ] 
#>   eta3 <~ y32      0.3521      0.0179   19.6834    0.0000 [ 0.3035; 0.3960 ] 
#>   eta3 <~ y33      0.3999      0.0153   26.1396    0.0000 [ 0.3629; 0.4420 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0462   14.5223    0.0000 [ 0.5406; 0.7797 ] 
#>   eta3 ~ eta1       0.6634      0.0459   14.4523    0.0000 [ 0.5299; 0.7672 ] 
#>   eta3 ~ eta2       0.3052      0.0855    3.5681    0.0004 [ 0.1269; 0.5692 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0576    3.5543    0.0004 [ 0.0823; 0.3804 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04622762 14.522345 8.746567e-48
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08705866  5.266642 1.389418e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08552290  3.568063 3.596297e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5406227          0.7796859          0.5693295          0.7509790
#> 2          0.1920959          0.6423141          0.2461583          0.5882517
#> 3          0.1268923          0.5691685          0.1800010          0.5160598
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.55930898          0.7798836          0.6089777          0.7700964
#> 2         0.29518710          0.6714690          0.3096364          0.6654923
#> 3         0.09895171          0.4526183          0.1232244          0.4468158