Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -963244388
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0480   13.9781    0.0000 [ 0.6201; 0.7821 ] 
#>   eta3 ~ eta1      0.4585      0.0806    5.6896    0.0000 [ 0.3249; 0.6034 ] 
#>   eta3 ~ eta2      0.3052      0.0894    3.4120    0.0006 [ 0.1290; 0.4676 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0401   16.5443    0.0000 [ 0.5806; 0.7325 ] 
#>   eta1 =~ y12      0.6493      0.0485   13.3813    0.0000 [ 0.5521; 0.7137 ] 
#>   eta1 =~ y13      0.7613      0.0326   23.3791    0.0000 [ 0.7101; 0.8254 ] 
#>   eta2 =~ y21      0.5165      0.0504   10.2468    0.0000 [ 0.4154; 0.6147 ] 
#>   eta2 =~ y22      0.7554      0.0378   19.9583    0.0000 [ 0.6647; 0.8124 ] 
#>   eta2 =~ y23      0.7997      0.0444   18.0265    0.0000 [ 0.7234; 0.8725 ] 
#>   eta3 =~ y31      0.8223      0.0265   31.0182    0.0000 [ 0.7687; 0.8585 ] 
#>   eta3 =~ y32      0.6581      0.0407   16.1858    0.0000 [ 0.5816; 0.7332 ] 
#>   eta3 =~ y33      0.7474      0.0314   23.7944    0.0000 [ 0.7039; 0.7982 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0237   16.7148    0.0000 [ 0.3539; 0.4293 ] 
#>   eta1 <~ y12      0.3873      0.0222   17.4779    0.0000 [ 0.3448; 0.4239 ] 
#>   eta1 <~ y13      0.4542      0.0212   21.4092    0.0000 [ 0.4205; 0.4920 ] 
#>   eta2 <~ y21      0.3058      0.0278   10.9958    0.0000 [ 0.2512; 0.3625 ] 
#>   eta2 <~ y22      0.4473      0.0224   19.9812    0.0000 [ 0.4125; 0.4867 ] 
#>   eta2 <~ y23      0.4735      0.0240   19.7465    0.0000 [ 0.4368; 0.5169 ] 
#>   eta3 <~ y31      0.4400      0.0143   30.6733    0.0000 [ 0.4158; 0.4677 ] 
#>   eta3 <~ y32      0.3521      0.0193   18.2066    0.0000 [ 0.3131; 0.3847 ] 
#>   eta3 <~ y33      0.3999      0.0176   22.6710    0.0000 [ 0.3736; 0.4339 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0480   13.9781    0.0000 [ 0.6201; 0.7821 ] 
#>   eta3 ~ eta1       0.6634      0.0356   18.6366    0.0000 [ 0.5942; 0.7316 ] 
#>   eta3 ~ eta2       0.3052      0.0894    3.4120    0.0006 [ 0.1290; 0.4676 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0577    3.5476    0.0004 [ 0.1006; 0.3033 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04007846 16.54429  1.760141e-61
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04852117 13.38133  7.774114e-41
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03256527 23.37908 6.977963e-121
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05040146 10.24682  1.222970e-24
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03784821 19.95835  1.268378e-88
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04436051 18.02648  1.207406e-72
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02650954 31.01817 3.066578e-211
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04065711 16.18583  6.349225e-59
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03141179 23.79438 3.818321e-125
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5806439          0.7324645
#> 2          0.5521385          0.7136951
#> 3          0.7100684          0.8253815
#> 4          0.4154436          0.6147025
#> 5          0.6647475          0.8124354
#> 6          0.7234291          0.8725220
#> 7          0.7687494          0.8585320
#> 8          0.5816478          0.7331745
#> 9          0.7039205          0.7981845

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -963244388
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0480   13.9781    0.0000 [ 0.5422; 0.7905 ] 
#>   eta3 ~ eta1      0.4585      0.0806    5.6896    0.0000 [ 0.2593; 0.6760 ] 
#>   eta3 ~ eta2      0.3052      0.0894    3.4120    0.0006 [ 0.0737; 0.5362 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0401   16.5443    0.0000 [ 0.5652; 0.7724 ] 
#>   eta1 =~ y12      0.6493      0.0485   13.3813    0.0000 [ 0.5266; 0.7775 ] 
#>   eta1 =~ y13      0.7613      0.0326   23.3791    0.0000 [ 0.6729; 0.8413 ] 
#>   eta2 =~ y21      0.5165      0.0504   10.2468    0.0000 [ 0.3925; 0.6531 ] 
#>   eta2 =~ y22      0.7554      0.0378   19.9583    0.0000 [ 0.6507; 0.8464 ] 
#>   eta2 =~ y23      0.7997      0.0444   18.0265    0.0000 [ 0.6922; 0.9216 ] 
#>   eta3 =~ y31      0.8223      0.0265   31.0182    0.0000 [ 0.7557; 0.8928 ] 
#>   eta3 =~ y32      0.6581      0.0407   16.1858    0.0000 [ 0.5579; 0.7682 ] 
#>   eta3 =~ y33      0.7474      0.0314   23.7944    0.0000 [ 0.6640; 0.8265 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0237   16.7148    0.0000 [ 0.3368; 0.4592 ] 
#>   eta1 <~ y12      0.3873      0.0222   17.4779    0.0000 [ 0.3312; 0.4458 ] 
#>   eta1 <~ y13      0.4542      0.0212   21.4092    0.0000 [ 0.3956; 0.5053 ] 
#>   eta2 <~ y21      0.3058      0.0278   10.9958    0.0000 [ 0.2368; 0.3806 ] 
#>   eta2 <~ y22      0.4473      0.0224   19.9812    0.0000 [ 0.3838; 0.4996 ] 
#>   eta2 <~ y23      0.4735      0.0240   19.7465    0.0000 [ 0.4143; 0.5383 ] 
#>   eta3 <~ y31      0.4400      0.0143   30.6733    0.0000 [ 0.4027; 0.4769 ] 
#>   eta3 <~ y32      0.3521      0.0193   18.2066    0.0000 [ 0.3039; 0.4040 ] 
#>   eta3 <~ y33      0.3999      0.0176   22.6710    0.0000 [ 0.3520; 0.4432 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0480   13.9781    0.0000 [ 0.5422; 0.7905 ] 
#>   eta3 ~ eta1       0.6634      0.0356   18.6366    0.0000 [ 0.5800; 0.7641 ] 
#>   eta3 ~ eta2       0.3052      0.0894    3.4120    0.0006 [ 0.0737; 0.5362 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0577    3.5476    0.0004 [ 0.0550; 0.3537 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04802739 13.978137 2.119634e-44
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08058694  5.689592 1.273435e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08943475  3.411997 6.448875e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54217873          0.7905493          0.5720032          0.7607248
#> 2         0.25929276          0.6760429          0.3093363          0.6259994
#> 3         0.07365962          0.5361656          0.1291975          0.4806277
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.6028619          0.7865213          0.6201147          0.7820979
#> 2          0.2848079          0.6337312          0.3249361          0.6034262
#> 3          0.1012910          0.4713215          0.1289540          0.4675990