Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2023184093
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0354   18.9572    0.0000 [ 0.6228; 0.7341 ] 
#>   eta3 ~ eta1      0.4585      0.0756    6.0652    0.0000 [ 0.3192; 0.5964 ] 
#>   eta3 ~ eta2      0.3052      0.0774    3.9410    0.0001 [ 0.1478; 0.4215 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0504   13.1656    0.0000 [ 0.5584; 0.7467 ] 
#>   eta1 =~ y12      0.6493      0.0421   15.4246    0.0000 [ 0.5776; 0.7313 ] 
#>   eta1 =~ y13      0.7613      0.0335   22.7424    0.0000 [ 0.7109; 0.8280 ] 
#>   eta2 =~ y21      0.5165      0.0450   11.4880    0.0000 [ 0.4096; 0.5807 ] 
#>   eta2 =~ y22      0.7554      0.0316   23.9078    0.0000 [ 0.7023; 0.8124 ] 
#>   eta2 =~ y23      0.7997      0.0412   19.4077    0.0000 [ 0.7061; 0.8713 ] 
#>   eta3 =~ y31      0.8223      0.0334   24.6376    0.0000 [ 0.7681; 0.8834 ] 
#>   eta3 =~ y32      0.6581      0.0352   18.6858    0.0000 [ 0.5892; 0.7086 ] 
#>   eta3 =~ y33      0.7474      0.0410   18.2233    0.0000 [ 0.6620; 0.7986 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0267   14.8161    0.0000 [ 0.3421; 0.4425 ] 
#>   eta1 <~ y12      0.3873      0.0211   18.3591    0.0000 [ 0.3488; 0.4288 ] 
#>   eta1 <~ y13      0.4542      0.0239   19.0133    0.0000 [ 0.4150; 0.4932 ] 
#>   eta2 <~ y21      0.3058      0.0254   12.0340    0.0000 [ 0.2603; 0.3485 ] 
#>   eta2 <~ y22      0.4473      0.0206   21.6740    0.0000 [ 0.4166; 0.4893 ] 
#>   eta2 <~ y23      0.4735      0.0176   26.9418    0.0000 [ 0.4406; 0.5036 ] 
#>   eta3 <~ y31      0.4400      0.0189   23.3285    0.0000 [ 0.4095; 0.4751 ] 
#>   eta3 <~ y32      0.3521      0.0195   18.0518    0.0000 [ 0.3187; 0.3896 ] 
#>   eta3 <~ y33      0.3999      0.0176   22.7513    0.0000 [ 0.3584; 0.4279 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0354   18.9572    0.0000 [ 0.6228; 0.7341 ] 
#>   eta3 ~ eta1       0.6634      0.0351   18.9009    0.0000 [ 0.5964; 0.7331 ] 
#>   eta3 ~ eta2       0.3052      0.0774    3.9410    0.0001 [ 0.1478; 0.4215 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0538    3.8067    0.0001 [ 0.1044; 0.3058 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.05036390 13.16558  1.384731e-39
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04209356 15.42464  1.117781e-53
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03347688 22.74244 1.704941e-114
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04495610 11.48798  1.516161e-30
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03159593 23.90775 2.543933e-126
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04120342 19.40770  6.642955e-84
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03337495 24.63756 5.003171e-134
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03521765 18.68577  6.464068e-78
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04101470 18.22332  3.370594e-74
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5584468          0.7467220
#> 2          0.5775846          0.7313437
#> 3          0.7109033          0.8279545
#> 4          0.4096127          0.5807003
#> 5          0.7023268          0.8124224
#> 6          0.7060775          0.8712552
#> 7          0.7680792          0.8834317
#> 8          0.5892349          0.7086418
#> 9          0.6619576          0.7985872

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2023184093
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0354   18.9572    0.0000 [ 0.5744; 0.7576 ] 
#>   eta3 ~ eta1      0.4585      0.0756    6.0652    0.0000 [ 0.2542; 0.6451 ] 
#>   eta3 ~ eta2      0.3052      0.0774    3.9410    0.0001 [ 0.1163; 0.5168 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0504   13.1656    0.0000 [ 0.5268; 0.7873 ] 
#>   eta1 =~ y12      0.6493      0.0421   15.4246    0.0000 [ 0.5495; 0.7672 ] 
#>   eta1 =~ y13      0.7613      0.0335   22.7424    0.0000 [ 0.6786; 0.8517 ] 
#>   eta2 =~ y21      0.5165      0.0450   11.4880    0.0000 [ 0.4149; 0.6474 ] 
#>   eta2 =~ y22      0.7554      0.0316   23.9078    0.0000 [ 0.6685; 0.8319 ] 
#>   eta2 =~ y23      0.7997      0.0412   19.4077    0.0000 [ 0.6959; 0.9089 ] 
#>   eta3 =~ y31      0.8223      0.0334   24.6376    0.0000 [ 0.7343; 0.9069 ] 
#>   eta3 =~ y32      0.6581      0.0352   18.6858    0.0000 [ 0.5707; 0.7528 ] 
#>   eta3 =~ y33      0.7474      0.0410   18.2233    0.0000 [ 0.6529; 0.8650 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0267   14.8161    0.0000 [ 0.3215; 0.4595 ] 
#>   eta1 <~ y12      0.3873      0.0211   18.3591    0.0000 [ 0.3368; 0.4459 ] 
#>   eta1 <~ y13      0.4542      0.0239   19.0133    0.0000 [ 0.3925; 0.5161 ] 
#>   eta2 <~ y21      0.3058      0.0254   12.0340    0.0000 [ 0.2470; 0.3784 ] 
#>   eta2 <~ y22      0.4473      0.0206   21.6740    0.0000 [ 0.3877; 0.4944 ] 
#>   eta2 <~ y23      0.4735      0.0176   26.9418    0.0000 [ 0.4268; 0.5177 ] 
#>   eta3 <~ y31      0.4400      0.0189   23.3285    0.0000 [ 0.3864; 0.4839 ] 
#>   eta3 <~ y32      0.3521      0.0195   18.0518    0.0000 [ 0.3005; 0.4014 ] 
#>   eta3 <~ y33      0.3999      0.0176   22.7513    0.0000 [ 0.3574; 0.4483 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0354   18.9572    0.0000 [ 0.5744; 0.7576 ] 
#>   eta3 ~ eta1       0.6634      0.0351   18.9009    0.0000 [ 0.5701; 0.7516 ] 
#>   eta3 ~ eta2       0.3052      0.0774    3.9410    0.0001 [ 0.1163; 0.5168 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0538    3.8067    0.0001 [ 0.0721; 0.3504 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03541319 18.957155 3.854357e-80
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07559662  6.065175 1.318097e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07742957  3.941016 8.113736e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5744279          0.7575649          0.5964191          0.7355737
#> 2          0.2541709          0.6451139          0.3011155          0.5981693
#> 3          0.1163307          0.5167527          0.1644136          0.4686699
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.6056470          0.7495979          0.6227824          0.7340874
#> 2          0.2945215          0.5973613          0.3191596          0.5963701
#> 3          0.1355162          0.4944044          0.1477725          0.4214911