Summarize model

Usage

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1948102236
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0414   16.2033    0.0000 [ 0.5876; 0.7398 ] 
#>   eta3 ~ eta1      0.4585      0.0762    6.0153    0.0000 [ 0.3206; 0.5837 ] 
#>   eta3 ~ eta2      0.3052      0.0855    3.5678    0.0004 [ 0.1615; 0.4825 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0388   17.1027    0.0000 [ 0.6046; 0.7170 ] 
#>   eta1 =~ y12      0.6493      0.0359   18.0867    0.0000 [ 0.5720; 0.6994 ] 
#>   eta1 =~ y13      0.7613      0.0255   29.8528    0.0000 [ 0.7193; 0.7979 ] 
#>   eta2 =~ y21      0.5165      0.0532    9.7152    0.0000 [ 0.3920; 0.5929 ] 
#>   eta2 =~ y22      0.7554      0.0350   21.5957    0.0000 [ 0.6855; 0.8155 ] 
#>   eta2 =~ y23      0.7997      0.0436   18.3418    0.0000 [ 0.7237; 0.8869 ] 
#>   eta3 =~ y31      0.8223      0.0333   24.6951    0.0000 [ 0.7799; 0.8775 ] 
#>   eta3 =~ y32      0.6581      0.0459   14.3276    0.0000 [ 0.5685; 0.7523 ] 
#>   eta3 =~ y33      0.7474      0.0394   18.9540    0.0000 [ 0.6900; 0.8259 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0154   25.7004    0.0000 [ 0.3634; 0.4197 ] 
#>   eta1 <~ y12      0.3873      0.0181   21.4287    0.0000 [ 0.3580; 0.4205 ] 
#>   eta1 <~ y13      0.4542      0.0195   23.3205    0.0000 [ 0.4311; 0.4974 ] 
#>   eta2 <~ y21      0.3058      0.0283   10.7964    0.0000 [ 0.2431; 0.3464 ] 
#>   eta2 <~ y22      0.4473      0.0247   18.1170    0.0000 [ 0.3959; 0.4915 ] 
#>   eta2 <~ y23      0.4735      0.0181   26.1505    0.0000 [ 0.4480; 0.5099 ] 
#>   eta3 <~ y31      0.4400      0.0195   22.6074    0.0000 [ 0.4022; 0.4695 ] 
#>   eta3 <~ y32      0.3521      0.0198   17.8004    0.0000 [ 0.3224; 0.3885 ] 
#>   eta3 <~ y33      0.3999      0.0210   19.0496    0.0000 [ 0.3668; 0.4361 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0414   16.2033    0.0000 [ 0.5876; 0.7398 ] 
#>   eta3 ~ eta1       0.6634      0.0315   21.0686    0.0000 [ 0.6079; 0.7076 ] 
#>   eta3 ~ eta2       0.3052      0.0855    3.5678    0.0004 [ 0.1615; 0.4825 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0563    3.6417    0.0003 [ 0.1011; 0.2996 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03877000 17.102652  1.418011e-65
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03589814 18.086671  4.058746e-73
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02550330 29.852835 8.065693e-196
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05315923  9.715242  2.596309e-22
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03497860 21.595711 1.970876e-103
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04359781 18.341831  3.836564e-75
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03329721 24.695083 1.207826e-134
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04593030 14.327555  1.472125e-46
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03943354 18.954021  4.090942e-80
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6045799          0.7169662
#> 2          0.5719597          0.6994101
#> 3          0.7192849          0.7979229
#> 4          0.3919848          0.5929274
#> 5          0.6854700          0.8154850
#> 6          0.7236669          0.8868651
#> 7          0.7799401          0.8775042
#> 8          0.5684943          0.7523076
#> 9          0.6899754          0.8259006

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1948102236
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0414   16.2033    0.0000 [ 0.5649; 0.7791 ] 
#>   eta3 ~ eta1      0.4585      0.0762    6.0153    0.0000 [ 0.2411; 0.6352 ] 
#>   eta3 ~ eta2      0.3052      0.0855    3.5678    0.0004 [ 0.1095; 0.5518 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0388   17.1027    0.0000 [ 0.5669; 0.7674 ] 
#>   eta1 =~ y12      0.6493      0.0359   18.0867    0.0000 [ 0.5606; 0.7462 ] 
#>   eta1 =~ y13      0.7613      0.0255   29.8528    0.0000 [ 0.6975; 0.8294 ] 
#>   eta2 =~ y21      0.5165      0.0532    9.7152    0.0000 [ 0.3830; 0.6579 ] 
#>   eta2 =~ y22      0.7554      0.0350   21.5957    0.0000 [ 0.6628; 0.8437 ] 
#>   eta2 =~ y23      0.7997      0.0436   18.3418    0.0000 [ 0.6840; 0.9095 ] 
#>   eta3 =~ y31      0.8223      0.0333   24.6951    0.0000 [ 0.7357; 0.9079 ] 
#>   eta3 =~ y32      0.6581      0.0459   14.3276    0.0000 [ 0.5401; 0.7777 ] 
#>   eta3 =~ y33      0.7474      0.0394   18.9540    0.0000 [ 0.6405; 0.8444 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0154   25.7004    0.0000 [ 0.3557; 0.4353 ] 
#>   eta1 <~ y12      0.3873      0.0181   21.4287    0.0000 [ 0.3404; 0.4339 ] 
#>   eta1 <~ y13      0.4542      0.0195   23.3205    0.0000 [ 0.4016; 0.5023 ] 
#>   eta2 <~ y21      0.3058      0.0283   10.7964    0.0000 [ 0.2359; 0.3823 ] 
#>   eta2 <~ y22      0.4473      0.0247   18.1170    0.0000 [ 0.3829; 0.5106 ] 
#>   eta2 <~ y23      0.4735      0.0181   26.1505    0.0000 [ 0.4263; 0.5199 ] 
#>   eta3 <~ y31      0.4400      0.0195   22.6074    0.0000 [ 0.3909; 0.4915 ] 
#>   eta3 <~ y32      0.3521      0.0198   17.8004    0.0000 [ 0.3030; 0.4053 ] 
#>   eta3 <~ y33      0.3999      0.0210   19.0496    0.0000 [ 0.3444; 0.4530 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0414   16.2033    0.0000 [ 0.5649; 0.7791 ] 
#>   eta3 ~ eta1       0.6634      0.0315   21.0686    0.0000 [ 0.5794; 0.7422 ] 
#>   eta3 ~ eta2       0.3052      0.0855    3.5678    0.0004 [ 0.1095; 0.5518 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0563    3.6417    0.0003 [ 0.0772; 0.3681 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04143189 16.203302 4.779018e-59
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07622321  6.015317 1.795348e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08552816  3.567844 3.599306e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5648592          0.7791215          0.5905879          0.7533928
#> 2          0.2410646          0.6352480          0.2883983          0.5879143
#> 3          0.1095442          0.5518476          0.1626562          0.4987356
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5826799          0.7471177          0.5876488          0.7397598
#> 2          0.2828942          0.6062007          0.3206293          0.5836945
#> 3          0.1422372          0.4873446          0.1615414          0.4825210