Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 467113898
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0437   15.3620    0.0000 [ 0.6033; 0.7624 ] 
#>   eta3 ~ eta1      0.4585      0.0776    5.9115    0.0000 [ 0.2984; 0.6001 ] 
#>   eta3 ~ eta2      0.3052      0.0835    3.6553    0.0003 [ 0.1801; 0.4828 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0416   15.9533    0.0000 [ 0.5893; 0.7397 ] 
#>   eta1 =~ y12      0.6493      0.0420   15.4606    0.0000 [ 0.5784; 0.7338 ] 
#>   eta1 =~ y13      0.7613      0.0403   18.8817    0.0000 [ 0.6844; 0.8346 ] 
#>   eta2 =~ y21      0.5165      0.0534    9.6684    0.0000 [ 0.4158; 0.6210 ] 
#>   eta2 =~ y22      0.7554      0.0431   17.5222    0.0000 [ 0.6893; 0.8350 ] 
#>   eta2 =~ y23      0.7997      0.0352   22.7105    0.0000 [ 0.7312; 0.8749 ] 
#>   eta3 =~ y31      0.8223      0.0314   26.1905    0.0000 [ 0.7659; 0.8743 ] 
#>   eta3 =~ y32      0.6581      0.0583   11.2951    0.0000 [ 0.5376; 0.7559 ] 
#>   eta3 =~ y33      0.7474      0.0406   18.4019    0.0000 [ 0.6816; 0.8099 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0256   15.4387    0.0000 [ 0.3541; 0.4363 ] 
#>   eta1 <~ y12      0.3873      0.0205   18.9149    0.0000 [ 0.3528; 0.4267 ] 
#>   eta1 <~ y13      0.4542      0.0223   20.3835    0.0000 [ 0.4158; 0.4979 ] 
#>   eta2 <~ y21      0.3058      0.0288   10.6084    0.0000 [ 0.2461; 0.3735 ] 
#>   eta2 <~ y22      0.4473      0.0223   20.0838    0.0000 [ 0.4073; 0.4928 ] 
#>   eta2 <~ y23      0.4735      0.0208   22.7304    0.0000 [ 0.4372; 0.5044 ] 
#>   eta3 <~ y31      0.4400      0.0214   20.5989    0.0000 [ 0.3985; 0.4776 ] 
#>   eta3 <~ y32      0.3521      0.0247   14.2373    0.0000 [ 0.3042; 0.3939 ] 
#>   eta3 <~ y33      0.3999      0.0231   17.3230    0.0000 [ 0.3662; 0.4484 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0437   15.3620    0.0000 [ 0.6033; 0.7624 ] 
#>   eta3 ~ eta1       0.6634      0.0321   20.6695    0.0000 [ 0.5985; 0.7206 ] 
#>   eta3 ~ eta2       0.3052      0.0835    3.6553    0.0003 [ 0.1801; 0.4828 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0586    3.4971    0.0005 [ 0.1192; 0.3467 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04156313 15.953320  2.701454e-57
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04199558 15.460624  6.397531e-54
#> 3 eta1 =~ y13  Common factor 0.7613458 0.04032198 18.881659  1.614361e-79
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05341655  9.668442  4.105796e-22
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04311037 17.522180  9.703497e-69
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03521119 22.710499 3.528176e-114
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03139596 26.190548 3.405209e-151
#> 8 eta3 =~ y32  Common factor 0.6580689 0.05826147 11.295095  1.387534e-29
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04061661 18.401931  1.267657e-75
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5892509          0.7397399
#> 2          0.5784002          0.7338392
#> 3          0.6844388          0.8346433
#> 4          0.4158460          0.6209955
#> 5          0.6893191          0.8349906
#> 6          0.7312461          0.8749012
#> 7          0.7659083          0.8743297
#> 8          0.5376229          0.7558844
#> 9          0.6816304          0.8099143

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 467113898
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0437   15.3620    0.0000 [ 0.5495; 0.7755 ] 
#>   eta3 ~ eta1      0.4585      0.0776    5.9115    0.0000 [ 0.2636; 0.6647 ] 
#>   eta3 ~ eta2      0.3052      0.0835    3.6553    0.0003 [ 0.0899; 0.5216 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0416   15.9533    0.0000 [ 0.5512; 0.7661 ] 
#>   eta1 =~ y12      0.6493      0.0420   15.4606    0.0000 [ 0.5417; 0.7589 ] 
#>   eta1 =~ y13      0.7613      0.0403   18.8817    0.0000 [ 0.6630; 0.8716 ] 
#>   eta2 =~ y21      0.5165      0.0534    9.6684    0.0000 [ 0.3882; 0.6644 ] 
#>   eta2 =~ y22      0.7554      0.0431   17.5222    0.0000 [ 0.6458; 0.8688 ] 
#>   eta2 =~ y23      0.7997      0.0352   22.7105    0.0000 [ 0.7115; 0.8936 ] 
#>   eta3 =~ y31      0.8223      0.0314   26.1905    0.0000 [ 0.7367; 0.8990 ] 
#>   eta3 =~ y32      0.6581      0.0583   11.2951    0.0000 [ 0.5126; 0.8139 ] 
#>   eta3 =~ y33      0.7474      0.0406   18.4019    0.0000 [ 0.6505; 0.8606 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0256   15.4387    0.0000 [ 0.3260; 0.4585 ] 
#>   eta1 <~ y12      0.3873      0.0205   18.9149    0.0000 [ 0.3347; 0.4406 ] 
#>   eta1 <~ y13      0.4542      0.0223   20.3835    0.0000 [ 0.3995; 0.5147 ] 
#>   eta2 <~ y21      0.3058      0.0288   10.6084    0.0000 [ 0.2347; 0.3837 ] 
#>   eta2 <~ y22      0.4473      0.0223   20.0838    0.0000 [ 0.3870; 0.5022 ] 
#>   eta2 <~ y23      0.4735      0.0208   22.7304    0.0000 [ 0.4171; 0.5248 ] 
#>   eta3 <~ y31      0.4400      0.0214   20.5989    0.0000 [ 0.3799; 0.4903 ] 
#>   eta3 <~ y32      0.3521      0.0247   14.2373    0.0000 [ 0.2896; 0.4175 ] 
#>   eta3 <~ y33      0.3999      0.0231   17.3230    0.0000 [ 0.3424; 0.4618 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0437   15.3620    0.0000 [ 0.5495; 0.7755 ] 
#>   eta3 ~ eta1       0.6634      0.0321   20.6695    0.0000 [ 0.5839; 0.7499 ] 
#>   eta3 ~ eta2       0.3052      0.0835    3.6553    0.0003 [ 0.0899; 0.5216 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0586    3.4971    0.0005 [ 0.0513; 0.3542 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04370100 15.361970 2.944896e-53
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07756201  5.911486 3.390347e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08348263  3.655265 2.569165e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54950910          0.7755060          0.5766469          0.7483682
#> 2         0.26356450          0.6646714          0.3117296          0.6165063
#> 3         0.08988233          0.5216073          0.1417241          0.4697656
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5958958          0.7743556          0.6032551          0.7623956
#> 2          0.2313174          0.6080570          0.2983927          0.6001028
#> 3          0.1751025          0.4881429          0.1800637          0.4828245