Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 503268875
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0343   19.5795    0.0000 [ 0.6259; 0.7403 ] 
#>   eta3 ~ eta1      0.4585      0.0755    6.0746    0.0000 [ 0.3299; 0.5849 ] 
#>   eta3 ~ eta2      0.3052      0.0903    3.3777    0.0007 [ 0.1412; 0.4773 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0454   14.5913    0.0000 [ 0.5788; 0.7312 ] 
#>   eta1 =~ y12      0.6493      0.0364   17.8422    0.0000 [ 0.5733; 0.7049 ] 
#>   eta1 =~ y13      0.7613      0.0378   20.1345    0.0000 [ 0.6757; 0.8258 ] 
#>   eta2 =~ y21      0.5165      0.0567    9.1097    0.0000 [ 0.4404; 0.6406 ] 
#>   eta2 =~ y22      0.7554      0.0363   20.8099    0.0000 [ 0.6963; 0.8104 ] 
#>   eta2 =~ y23      0.7997      0.0388   20.6352    0.0000 [ 0.7357; 0.8632 ] 
#>   eta3 =~ y31      0.8223      0.0285   28.8236    0.0000 [ 0.7833; 0.8795 ] 
#>   eta3 =~ y32      0.6581      0.0335   19.6550    0.0000 [ 0.5980; 0.7221 ] 
#>   eta3 =~ y33      0.7474      0.0432   17.3010    0.0000 [ 0.6664; 0.8275 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0242   16.3377    0.0000 [ 0.3462; 0.4406 ] 
#>   eta1 <~ y12      0.3873      0.0191   20.2302    0.0000 [ 0.3439; 0.4258 ] 
#>   eta1 <~ y13      0.4542      0.0217   20.9018    0.0000 [ 0.4168; 0.4930 ] 
#>   eta2 <~ y21      0.3058      0.0286   10.6764    0.0000 [ 0.2634; 0.3543 ] 
#>   eta2 <~ y22      0.4473      0.0204   21.8985    0.0000 [ 0.4159; 0.4755 ] 
#>   eta2 <~ y23      0.4735      0.0238   19.8710    0.0000 [ 0.4312; 0.5109 ] 
#>   eta3 <~ y31      0.4400      0.0198   22.2538    0.0000 [ 0.4105; 0.4997 ] 
#>   eta3 <~ y32      0.3521      0.0166   21.1746    0.0000 [ 0.3189; 0.3831 ] 
#>   eta3 <~ y33      0.3999      0.0173   23.0656    0.0000 [ 0.3646; 0.4286 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0343   19.5795    0.0000 [ 0.6259; 0.7403 ] 
#>   eta3 ~ eta1       0.6634      0.0315   21.0338    0.0000 [ 0.6110; 0.7274 ] 
#>   eta3 ~ eta2       0.3052      0.0903    3.3777    0.0007 [ 0.1412; 0.4773 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0586    3.4963    0.0005 [ 0.1000; 0.3215 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04544269 14.591342  3.188633e-48
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03639004 17.842186  3.324349e-71
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03781306 20.134466  3.682607e-90
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05669263  9.109735  8.258380e-20
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03629937 20.809937  3.518039e-96
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03875246 20.635173  1.326687e-94
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02852794 28.823581 1.086417e-182
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03348106 19.654961  5.242947e-86
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04320117 17.301014  4.621790e-67
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5787850          0.7312478
#> 2          0.5733043          0.7049261
#> 3          0.6756523          0.8257775
#> 4          0.4404336          0.6405858
#> 5          0.6962576          0.8104485
#> 6          0.7357152          0.8631939
#> 7          0.7833173          0.8794616
#> 8          0.5979700          0.7221217
#> 9          0.6663529          0.8274825

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 503268875
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0343   19.5795    0.0000 [ 0.5717; 0.7490 ] 
#>   eta3 ~ eta1      0.4585      0.0755    6.0746    0.0000 [ 0.2669; 0.6572 ] 
#>   eta3 ~ eta2      0.3052      0.0903    3.3777    0.0007 [ 0.0602; 0.5274 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0454   14.5913    0.0000 [ 0.5431; 0.7781 ] 
#>   eta1 =~ y12      0.6493      0.0364   17.8422    0.0000 [ 0.5565; 0.7447 ] 
#>   eta1 =~ y13      0.7613      0.0378   20.1345    0.0000 [ 0.6697; 0.8652 ] 
#>   eta2 =~ y21      0.5165      0.0567    9.1097    0.0000 [ 0.3542; 0.6474 ] 
#>   eta2 =~ y22      0.7554      0.0363   20.8099    0.0000 [ 0.6621; 0.8498 ] 
#>   eta2 =~ y23      0.7997      0.0388   20.6352    0.0000 [ 0.7038; 0.9042 ] 
#>   eta3 =~ y31      0.8223      0.0285   28.8236    0.0000 [ 0.7432; 0.8907 ] 
#>   eta3 =~ y32      0.6581      0.0335   19.6550    0.0000 [ 0.5714; 0.7446 ] 
#>   eta3 =~ y33      0.7474      0.0432   17.3010    0.0000 [ 0.6393; 0.8627 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0242   16.3377    0.0000 [ 0.3304; 0.4556 ] 
#>   eta1 <~ y12      0.3873      0.0191   20.2302    0.0000 [ 0.3374; 0.4364 ] 
#>   eta1 <~ y13      0.4542      0.0217   20.9018    0.0000 [ 0.4001; 0.5125 ] 
#>   eta2 <~ y21      0.3058      0.0286   10.6764    0.0000 [ 0.2251; 0.3732 ] 
#>   eta2 <~ y22      0.4473      0.0204   21.8985    0.0000 [ 0.3979; 0.5035 ] 
#>   eta2 <~ y23      0.4735      0.0238   19.8710    0.0000 [ 0.4176; 0.5409 ] 
#>   eta3 <~ y31      0.4400      0.0198   22.2538    0.0000 [ 0.3866; 0.4888 ] 
#>   eta3 <~ y32      0.3521      0.0166   21.1746    0.0000 [ 0.3097; 0.3957 ] 
#>   eta3 <~ y33      0.3999      0.0173   23.0656    0.0000 [ 0.3579; 0.4476 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0343   19.5795    0.0000 [ 0.5717; 0.7490 ] 
#>   eta3 ~ eta1       0.6634      0.0315   21.0338    0.0000 [ 0.5754; 0.7385 ] 
#>   eta3 ~ eta2       0.3052      0.0903    3.3777    0.0007 [ 0.0602; 0.5274 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0586    3.4963    0.0005 [ 0.0434; 0.3464 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03428760 19.579479 2.313760e-85
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07547944  6.074592 1.243037e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09034173  3.377743 7.308337e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.57167556          0.7489917          0.5929678          0.7276995
#> 2         0.26687890          0.6572159          0.3137507          0.6103441
#> 3         0.06020537          0.5274018          0.1163065          0.4713006
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.6232991          0.7516324          0.6259480          0.7403127
#> 2          0.2965814          0.5989126          0.3298754          0.5849209
#> 3          0.1284922          0.4874158          0.1411701          0.4773247