Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 490140016
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0360   18.6719    0.0000 [ 0.6070; 0.7273 ] 
#>   eta3 ~ eta1      0.4585      0.0672    6.8234    0.0000 [ 0.3585; 0.5846 ] 
#>   eta3 ~ eta2      0.3052      0.0678    4.5009    0.0000 [ 0.1817; 0.4254 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0426   15.5784    0.0000 [ 0.5712; 0.7327 ] 
#>   eta1 =~ y12      0.6493      0.0396   16.3759    0.0000 [ 0.6058; 0.7342 ] 
#>   eta1 =~ y13      0.7613      0.0342   22.2347    0.0000 [ 0.6878; 0.8175 ] 
#>   eta2 =~ y21      0.5165      0.0698    7.3944    0.0000 [ 0.3833; 0.6482 ] 
#>   eta2 =~ y22      0.7554      0.0396   19.0994    0.0000 [ 0.6790; 0.8170 ] 
#>   eta2 =~ y23      0.7997      0.0356   22.4376    0.0000 [ 0.7319; 0.8539 ] 
#>   eta3 =~ y31      0.8223      0.0346   23.7476    0.0000 [ 0.7429; 0.8769 ] 
#>   eta3 =~ y32      0.6581      0.0342   19.2323    0.0000 [ 0.5825; 0.7054 ] 
#>   eta3 =~ y33      0.7474      0.0367   20.3673    0.0000 [ 0.6914; 0.8277 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0236   16.7613    0.0000 [ 0.3474; 0.4338 ] 
#>   eta1 <~ y12      0.3873      0.0196   19.7836    0.0000 [ 0.3650; 0.4230 ] 
#>   eta1 <~ y13      0.4542      0.0196   23.1240    0.0000 [ 0.4151; 0.4874 ] 
#>   eta2 <~ y21      0.3058      0.0365    8.3798    0.0000 [ 0.2329; 0.3766 ] 
#>   eta2 <~ y22      0.4473      0.0244   18.3034    0.0000 [ 0.3991; 0.4879 ] 
#>   eta2 <~ y23      0.4735      0.0226   20.9541    0.0000 [ 0.4438; 0.5189 ] 
#>   eta3 <~ y31      0.4400      0.0187   23.5638    0.0000 [ 0.3999; 0.4753 ] 
#>   eta3 <~ y32      0.3521      0.0152   23.1219    0.0000 [ 0.3159; 0.3747 ] 
#>   eta3 <~ y33      0.3999      0.0170   23.5475    0.0000 [ 0.3801; 0.4401 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0360   18.6719    0.0000 [ 0.6070; 0.7273 ] 
#>   eta3 ~ eta1       0.6634      0.0366   18.1384    0.0000 [ 0.6127; 0.7288 ] 
#>   eta3 ~ eta2       0.3052      0.0678    4.5009    0.0000 [ 0.1817; 0.4254 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0476    4.3065    0.0000 [ 0.1199; 0.2868 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04256331 15.578439  1.020154e-54
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03964830 16.375932  2.841215e-60
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03424137 22.234678 1.587043e-109
#> 4 eta2 =~ y21  Common factor 0.5164548 0.06984406  7.394398  1.420499e-13
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03955043 19.099352  2.556411e-81
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03563941 22.437626 1.690288e-111
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03462564 23.747643 1.162120e-124
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03421679 19.232341  1.984657e-82
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03669732 20.367266  3.263956e-92
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5712438          0.7327402
#> 2          0.6057803          0.7341822
#> 3          0.6878341          0.8175168
#> 4          0.3833034          0.6481713
#> 5          0.6789590          0.8170255
#> 6          0.7318763          0.8539243
#> 7          0.7428809          0.8769108
#> 8          0.5824978          0.7053709
#> 9          0.6913949          0.8276636

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 490140016
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0360   18.6719    0.0000 [ 0.5826; 0.7685 ] 
#>   eta3 ~ eta1      0.4585      0.0672    6.8234    0.0000 [ 0.2665; 0.6140 ] 
#>   eta3 ~ eta2      0.3052      0.0678    4.5009    0.0000 [ 0.1451; 0.4957 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0426   15.5784    0.0000 [ 0.5626; 0.7827 ] 
#>   eta1 =~ y12      0.6493      0.0396   16.3759    0.0000 [ 0.5342; 0.7392 ] 
#>   eta1 =~ y13      0.7613      0.0342   22.2347    0.0000 [ 0.6672; 0.8443 ] 
#>   eta2 =~ y21      0.5165      0.0698    7.3944    0.0000 [ 0.3427; 0.7038 ] 
#>   eta2 =~ y22      0.7554      0.0396   19.0994    0.0000 [ 0.6589; 0.8634 ] 
#>   eta2 =~ y23      0.7997      0.0356   22.4376    0.0000 [ 0.7074; 0.8917 ] 
#>   eta3 =~ y31      0.8223      0.0346   23.7476    0.0000 [ 0.7340; 0.9131 ] 
#>   eta3 =~ y32      0.6581      0.0342   19.2323    0.0000 [ 0.5781; 0.7551 ] 
#>   eta3 =~ y33      0.7474      0.0367   20.3673    0.0000 [ 0.6415; 0.8312 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0236   16.7613    0.0000 [ 0.3429; 0.4649 ] 
#>   eta1 <~ y12      0.3873      0.0196   19.7836    0.0000 [ 0.3320; 0.4332 ] 
#>   eta1 <~ y13      0.4542      0.0196   23.1240    0.0000 [ 0.4031; 0.5046 ] 
#>   eta2 <~ y21      0.3058      0.0365    8.3798    0.0000 [ 0.2140; 0.4027 ] 
#>   eta2 <~ y22      0.4473      0.0244   18.3034    0.0000 [ 0.3843; 0.5107 ] 
#>   eta2 <~ y23      0.4735      0.0226   20.9541    0.0000 [ 0.4116; 0.5285 ] 
#>   eta3 <~ y31      0.4400      0.0187   23.5638    0.0000 [ 0.3929; 0.4895 ] 
#>   eta3 <~ y32      0.3521      0.0152   23.1219    0.0000 [ 0.3179; 0.3967 ] 
#>   eta3 <~ y33      0.3999      0.0170   23.5475    0.0000 [ 0.3507; 0.4386 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0360   18.6719    0.0000 [ 0.5826; 0.7685 ] 
#>   eta3 ~ eta1       0.6634      0.0366   18.1384    0.0000 [ 0.5619; 0.7510 ] 
#>   eta3 ~ eta2       0.3052      0.0678    4.5009    0.0000 [ 0.1451; 0.4957 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0476    4.3065    0.0000 [ 0.0932; 0.3392 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03595420 18.671907 8.380596e-78
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.06719664  6.823358 8.893661e-12
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.06779753  4.500918 6.766065e-06
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5825910          0.7685258          0.6049181          0.7461986
#> 2          0.2665057          0.6140088          0.3082340          0.5722805
#> 3          0.1451152          0.4957258          0.1872167          0.4536243
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5970276          0.7439521          0.6070132          0.7272987
#> 2          0.3134441          0.5846414          0.3585172          0.5845918
#> 3          0.1660474          0.4344629          0.1816745          0.4254217