Summarize model

Usage

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -251302023
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0530   12.6720    0.0000 [ 0.5970; 0.7504 ] 
#>   eta3 ~ eta1      0.4585      0.0659    6.9585    0.0000 [ 0.3053; 0.5475 ] 
#>   eta3 ~ eta2      0.3052      0.0700    4.3570    0.0000 [ 0.2080; 0.4141 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0448   14.7910    0.0000 [ 0.5938; 0.7526 ] 
#>   eta1 =~ y12      0.6493      0.0352   18.4510    0.0000 [ 0.5870; 0.7204 ] 
#>   eta1 =~ y13      0.7613      0.0333   22.8475    0.0000 [ 0.7119; 0.8364 ] 
#>   eta2 =~ y21      0.5165      0.0435   11.8661    0.0000 [ 0.4351; 0.5845 ] 
#>   eta2 =~ y22      0.7554      0.0328   23.0068    0.0000 [ 0.6912; 0.8022 ] 
#>   eta2 =~ y23      0.7997      0.0293   27.3059    0.0000 [ 0.7374; 0.8420 ] 
#>   eta3 =~ y31      0.8223      0.0329   25.0107    0.0000 [ 0.7739; 0.8939 ] 
#>   eta3 =~ y32      0.6581      0.0352   18.7177    0.0000 [ 0.6097; 0.7364 ] 
#>   eta3 =~ y33      0.7474      0.0376   19.8669    0.0000 [ 0.6657; 0.8079 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0221   17.9275    0.0000 [ 0.3515; 0.4331 ] 
#>   eta1 <~ y12      0.3873      0.0194   19.9230    0.0000 [ 0.3429; 0.4111 ] 
#>   eta1 <~ y13      0.4542      0.0184   24.6937    0.0000 [ 0.4211; 0.4833 ] 
#>   eta2 <~ y21      0.3058      0.0229   13.3273    0.0000 [ 0.2655; 0.3583 ] 
#>   eta2 <~ y22      0.4473      0.0178   25.0944    0.0000 [ 0.4124; 0.4782 ] 
#>   eta2 <~ y23      0.4735      0.0184   25.6732    0.0000 [ 0.4391; 0.5135 ] 
#>   eta3 <~ y31      0.4400      0.0167   26.3282    0.0000 [ 0.4173; 0.4762 ] 
#>   eta3 <~ y32      0.3521      0.0174   20.2905    0.0000 [ 0.3270; 0.3966 ] 
#>   eta3 <~ y33      0.3999      0.0193   20.7490    0.0000 [ 0.3605; 0.4263 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0530   12.6720    0.0000 [ 0.5970; 0.7504 ] 
#>   eta3 ~ eta1       0.6634      0.0406   16.3195    0.0000 [ 0.5531; 0.7253 ] 
#>   eta3 ~ eta2       0.3052      0.0700    4.3570    0.0000 [ 0.2080; 0.4141 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0468    4.3771    0.0000 [ 0.1416; 0.2970 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04482937 14.79097  1.675220e-49
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03518937 18.45097  5.121757e-76
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03332290 22.84753 1.546563e-115
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04352346 11.86612  1.775004e-32
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03283318 23.00684 3.981907e-117
#> 6 eta2 =~ y23  Common factor 0.7996637 0.02928533 27.30594 3.605370e-164
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03287699 25.01072 4.673760e-138
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03515756 18.71771  3.551187e-78
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03762152 19.86693  7.866649e-88
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5938290          0.7526255
#> 2          0.5870409          0.7203639
#> 3          0.7119260          0.8363562
#> 4          0.4350577          0.5844500
#> 5          0.6912168          0.8022268
#> 6          0.7373805          0.8420439
#> 7          0.7738567          0.8939144
#> 8          0.6096616          0.7363908
#> 9          0.6656887          0.8078984

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -251302023
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0530   12.6720    0.0000 [ 0.5324; 0.8064 ] 
#>   eta3 ~ eta1      0.4585      0.0659    6.9585    0.0000 [ 0.3042; 0.6449 ] 
#>   eta3 ~ eta2      0.3052      0.0700    4.3570    0.0000 [ 0.1141; 0.4763 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0448   14.7910    0.0000 [ 0.5366; 0.7684 ] 
#>   eta1 =~ y12      0.6493      0.0352   18.4510    0.0000 [ 0.5591; 0.7411 ] 
#>   eta1 =~ y13      0.7613      0.0333   22.8475    0.0000 [ 0.6651; 0.8375 ] 
#>   eta2 =~ y21      0.5165      0.0435   11.8661    0.0000 [ 0.3953; 0.6204 ] 
#>   eta2 =~ y22      0.7554      0.0328   23.0068    0.0000 [ 0.6774; 0.8472 ] 
#>   eta2 =~ y23      0.7997      0.0293   27.3059    0.0000 [ 0.7283; 0.8797 ] 
#>   eta3 =~ y31      0.8223      0.0329   25.0107    0.0000 [ 0.7277; 0.8978 ] 
#>   eta3 =~ y32      0.6581      0.0352   18.7177    0.0000 [ 0.5652; 0.7471 ] 
#>   eta3 =~ y33      0.7474      0.0376   19.8669    0.0000 [ 0.6570; 0.8515 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0221   17.9275    0.0000 [ 0.3382; 0.4523 ] 
#>   eta1 <~ y12      0.3873      0.0194   19.9230    0.0000 [ 0.3432; 0.4437 ] 
#>   eta1 <~ y13      0.4542      0.0184   24.6937    0.0000 [ 0.4073; 0.5024 ] 
#>   eta2 <~ y21      0.3058      0.0229   13.3273    0.0000 [ 0.2408; 0.3595 ] 
#>   eta2 <~ y22      0.4473      0.0178   25.0944    0.0000 [ 0.4042; 0.4964 ] 
#>   eta2 <~ y23      0.4735      0.0184   25.6732    0.0000 [ 0.4272; 0.5225 ] 
#>   eta3 <~ y31      0.4400      0.0167   26.3282    0.0000 [ 0.3933; 0.4797 ] 
#>   eta3 <~ y32      0.3521      0.0174   20.2905    0.0000 [ 0.3075; 0.3973 ] 
#>   eta3 <~ y33      0.3999      0.0193   20.7490    0.0000 [ 0.3552; 0.4549 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0530   12.6720    0.0000 [ 0.5324; 0.8064 ] 
#>   eta3 ~ eta1       0.6634      0.0406   16.3195    0.0000 [ 0.5678; 0.7780 ] 
#>   eta3 ~ eta2       0.3052      0.0700    4.3570    0.0000 [ 0.1141; 0.4763 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0468    4.3771    0.0000 [ 0.0773; 0.3194 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05297786 12.671963 8.457163e-37
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.06589133  6.958530 3.438420e-12
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07003771  4.356955 1.318846e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5324171          0.8063887          0.5653158          0.7734901
#> 2          0.3041539          0.6449066          0.3450716          0.6039889
#> 3          0.1141335          0.4763290          0.1576261          0.4328364
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5797710          0.7846085          0.5969749          0.7504049
#> 2          0.2962135          0.5612812          0.3052631          0.5475468
#> 3          0.1965514          0.4336397          0.2080114          0.4140518