Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1354729602
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0479   14.0214    0.0000 [ 0.5961; 0.7508 ] 
#>   eta3 ~ eta1      0.4585      0.0795    5.7672    0.0000 [ 0.3087; 0.6135 ] 
#>   eta3 ~ eta2      0.3052      0.0850    3.5896    0.0003 [ 0.1320; 0.4454 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0399   16.6384    0.0000 [ 0.5831; 0.7214 ] 
#>   eta1 =~ y12      0.6493      0.0374   17.3789    0.0000 [ 0.5735; 0.7069 ] 
#>   eta1 =~ y13      0.7613      0.0300   25.3951    0.0000 [ 0.7100; 0.8230 ] 
#>   eta2 =~ y21      0.5165      0.0522    9.8996    0.0000 [ 0.4387; 0.6250 ] 
#>   eta2 =~ y22      0.7554      0.0347   21.7386    0.0000 [ 0.6887; 0.8079 ] 
#>   eta2 =~ y23      0.7997      0.0333   24.0044    0.0000 [ 0.7272; 0.8666 ] 
#>   eta3 =~ y31      0.8223      0.0309   26.6475    0.0000 [ 0.7722; 0.8859 ] 
#>   eta3 =~ y32      0.6581      0.0415   15.8403    0.0000 [ 0.5876; 0.7142 ] 
#>   eta3 =~ y33      0.7474      0.0333   22.4718    0.0000 [ 0.7023; 0.8056 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0201   19.7061    0.0000 [ 0.3660; 0.4469 ] 
#>   eta1 <~ y12      0.3873      0.0191   20.2596    0.0000 [ 0.3536; 0.4202 ] 
#>   eta1 <~ y13      0.4542      0.0212   21.4493    0.0000 [ 0.4304; 0.4911 ] 
#>   eta2 <~ y21      0.3058      0.0247   12.4039    0.0000 [ 0.2532; 0.3650 ] 
#>   eta2 <~ y22      0.4473      0.0226   19.7954    0.0000 [ 0.4025; 0.4879 ] 
#>   eta2 <~ y23      0.4735      0.0184   25.7021    0.0000 [ 0.4424; 0.5035 ] 
#>   eta3 <~ y31      0.4400      0.0158   27.8771    0.0000 [ 0.4144; 0.4734 ] 
#>   eta3 <~ y32      0.3521      0.0166   21.2143    0.0000 [ 0.3182; 0.3781 ] 
#>   eta3 <~ y33      0.3999      0.0210   19.0789    0.0000 [ 0.3673; 0.4431 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0479   14.0214    0.0000 [ 0.5961; 0.7508 ] 
#>   eta3 ~ eta1       0.6634      0.0338   19.6437    0.0000 [ 0.6042; 0.7271 ] 
#>   eta3 ~ eta2       0.3052      0.0850    3.5896    0.0003 [ 0.1320; 0.4454 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0609    3.3621    0.0008 [ 0.0946; 0.3245 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03985171 16.638429  3.671245e-62
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03736021 17.378864  1.192938e-67
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02997997 25.395149 2.853079e-142
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05216941  9.899572  4.180604e-23
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03474868 21.738602 8.855564e-105
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03331314 24.004450 2.498628e-127
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03085757 26.647506 1.912829e-156
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04154406 15.840263  1.641334e-56
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03326051 22.471818 7.831881e-112
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5831469          0.7214186
#> 2          0.5735330          0.7069102
#> 3          0.7100245          0.8230290
#> 4          0.4387149          0.6250193
#> 5          0.6887283          0.8079057
#> 6          0.7271955          0.8666095
#> 7          0.7721633          0.8858917
#> 8          0.5876046          0.7141899
#> 9          0.7023445          0.8056448

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1354729602
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0479   14.0214    0.0000 [ 0.5431; 0.7907 ] 
#>   eta3 ~ eta1      0.4585      0.0795    5.7672    0.0000 [ 0.2397; 0.6509 ] 
#>   eta3 ~ eta2      0.3052      0.0850    3.5896    0.0003 [ 0.0974; 0.5370 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0399   16.6384    0.0000 [ 0.5673; 0.7733 ] 
#>   eta1 =~ y12      0.6493      0.0374   17.3789    0.0000 [ 0.5555; 0.7487 ] 
#>   eta1 =~ y13      0.7613      0.0300   25.3951    0.0000 [ 0.6868; 0.8419 ] 
#>   eta2 =~ y21      0.5165      0.0522    9.8996    0.0000 [ 0.3671; 0.6369 ] 
#>   eta2 =~ y22      0.7554      0.0347   21.7386    0.0000 [ 0.6708; 0.8505 ] 
#>   eta2 =~ y23      0.7997      0.0333   24.0044    0.0000 [ 0.7182; 0.8904 ] 
#>   eta3 =~ y31      0.8223      0.0309   26.6475    0.0000 [ 0.7357; 0.8952 ] 
#>   eta3 =~ y32      0.6581      0.0415   15.8403    0.0000 [ 0.5526; 0.7674 ] 
#>   eta3 =~ y33      0.7474      0.0333   22.4718    0.0000 [ 0.6571; 0.8291 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0201   19.7061    0.0000 [ 0.3446; 0.4484 ] 
#>   eta1 <~ y12      0.3873      0.0191   20.2596    0.0000 [ 0.3363; 0.4351 ] 
#>   eta1 <~ y13      0.4542      0.0212   21.4493    0.0000 [ 0.3970; 0.5065 ] 
#>   eta2 <~ y21      0.3058      0.0247   12.4039    0.0000 [ 0.2347; 0.3622 ] 
#>   eta2 <~ y22      0.4473      0.0226   19.7954    0.0000 [ 0.3928; 0.5097 ] 
#>   eta2 <~ y23      0.4735      0.0184   25.7021    0.0000 [ 0.4297; 0.5249 ] 
#>   eta3 <~ y31      0.4400      0.0158   27.8771    0.0000 [ 0.3986; 0.4802 ] 
#>   eta3 <~ y32      0.3521      0.0166   21.2143    0.0000 [ 0.3129; 0.3987 ] 
#>   eta3 <~ y33      0.3999      0.0210   19.0789    0.0000 [ 0.3460; 0.4544 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0479   14.0214    0.0000 [ 0.5431; 0.7907 ] 
#>   eta3 ~ eta1       0.6634      0.0338   19.6437    0.0000 [ 0.5695; 0.7441 ] 
#>   eta3 ~ eta2       0.3052      0.0850    3.5896    0.0003 [ 0.0974; 0.5370 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0609    3.3621    0.0008 [ 0.0539; 0.3690 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04787919 14.021403 1.153119e-44
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07950296  5.767166 8.061575e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08501080  3.589557 3.312400e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54308771          0.7906918          0.5728201          0.7609594
#> 2         0.23971489          0.6508593          0.2890853          0.6014889
#> 3         0.09737475          0.5370026          0.1501655          0.4842119
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5785051          0.7691695          0.5961176          0.7507980
#> 2          0.2816508          0.6426050          0.3086916          0.6134692
#> 3          0.1081550          0.4590696          0.1319905          0.4453863