Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1187556123
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0430   15.6121    0.0000 [ 0.6052; 0.7468 ] 
#>   eta3 ~ eta1      0.4585      0.0789    5.8104    0.0000 [ 0.3290; 0.6162 ] 
#>   eta3 ~ eta2      0.3052      0.0849    3.5963    0.0003 [ 0.1286; 0.4344 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0332   19.9859    0.0000 [ 0.5971; 0.7121 ] 
#>   eta1 =~ y12      0.6493      0.0377   17.2258    0.0000 [ 0.5736; 0.7033 ] 
#>   eta1 =~ y13      0.7613      0.0275   27.6468    0.0000 [ 0.7080; 0.8076 ] 
#>   eta2 =~ y21      0.5165      0.0418   12.3568    0.0000 [ 0.4284; 0.5823 ] 
#>   eta2 =~ y22      0.7554      0.0341   22.1451    0.0000 [ 0.6923; 0.8041 ] 
#>   eta2 =~ y23      0.7997      0.0285   28.0298    0.0000 [ 0.7460; 0.8556 ] 
#>   eta3 =~ y31      0.8223      0.0276   29.7433    0.0000 [ 0.7597; 0.8821 ] 
#>   eta3 =~ y32      0.6581      0.0451   14.5935    0.0000 [ 0.5652; 0.7305 ] 
#>   eta3 =~ y33      0.7474      0.0443   16.8598    0.0000 [ 0.6569; 0.8181 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0191   20.6748    0.0000 [ 0.3553; 0.4262 ] 
#>   eta1 <~ y12      0.3873      0.0181   21.4046    0.0000 [ 0.3503; 0.4211 ] 
#>   eta1 <~ y13      0.4542      0.0166   27.3005    0.0000 [ 0.4267; 0.4918 ] 
#>   eta2 <~ y21      0.3058      0.0216   14.1429    0.0000 [ 0.2628; 0.3365 ] 
#>   eta2 <~ y22      0.4473      0.0202   22.1307    0.0000 [ 0.4130; 0.4839 ] 
#>   eta2 <~ y23      0.4735      0.0155   30.5311    0.0000 [ 0.4478; 0.5076 ] 
#>   eta3 <~ y31      0.4400      0.0196   22.4224    0.0000 [ 0.4042; 0.4784 ] 
#>   eta3 <~ y32      0.3521      0.0208   16.9516    0.0000 [ 0.3122; 0.3924 ] 
#>   eta3 <~ y33      0.3999      0.0197   20.2826    0.0000 [ 0.3647; 0.4369 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0430   15.6121    0.0000 [ 0.6052; 0.7468 ] 
#>   eta3 ~ eta1       0.6634      0.0401   16.5379    0.0000 [ 0.5859; 0.7165 ] 
#>   eta3 ~ eta2       0.3052      0.0849    3.5963    0.0003 [ 0.1286; 0.4344 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0573    3.5766    0.0003 [ 0.0845; 0.2824 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03317691 19.98589  7.307913e-89
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03769223 17.22578  1.701284e-66
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02753829 27.64681 3.048625e-168
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04179515 12.35681  4.475619e-35
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03411085 22.14508 1.163417e-108
#> 6 eta2 =~ y23  Common factor 0.7996637 0.02852905 28.02981 7.041652e-173
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02764578 29.74332 2.115964e-194
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04509318 14.59354  3.087688e-48
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04433177 16.85979  8.890934e-64
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5970541          0.7120916
#> 2          0.5735678          0.7032826
#> 3          0.7079742          0.8075753
#> 4          0.4283968          0.5823325
#> 5          0.6923026          0.8040613
#> 6          0.7459963          0.8555942
#> 7          0.7597324          0.8820705
#> 8          0.5651665          0.7304964
#> 9          0.6569480          0.8181459

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1187556123
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0430   15.6121    0.0000 [ 0.5551; 0.7775 ] 
#>   eta3 ~ eta1      0.4585      0.0789    5.8104    0.0000 [ 0.2519; 0.6600 ] 
#>   eta3 ~ eta2      0.3052      0.0849    3.5963    0.0003 [ 0.0918; 0.5306 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0332   19.9859    0.0000 [ 0.5862; 0.7578 ] 
#>   eta1 =~ y12      0.6493      0.0377   17.2258    0.0000 [ 0.5525; 0.7475 ] 
#>   eta1 =~ y13      0.7613      0.0275   27.6468    0.0000 [ 0.6833; 0.8257 ] 
#>   eta2 =~ y21      0.5165      0.0418   12.3568    0.0000 [ 0.4168; 0.6330 ] 
#>   eta2 =~ y22      0.7554      0.0341   22.1451    0.0000 [ 0.6715; 0.8479 ] 
#>   eta2 =~ y23      0.7997      0.0285   28.0298    0.0000 [ 0.7230; 0.8705 ] 
#>   eta3 =~ y31      0.8223      0.0276   29.7433    0.0000 [ 0.7544; 0.8974 ] 
#>   eta3 =~ y32      0.6581      0.0451   14.5935    0.0000 [ 0.5449; 0.7781 ] 
#>   eta3 =~ y33      0.7474      0.0443   16.8598    0.0000 [ 0.6286; 0.8578 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0191   20.6748    0.0000 [ 0.3508; 0.4498 ] 
#>   eta1 <~ y12      0.3873      0.0181   21.4046    0.0000 [ 0.3406; 0.4342 ] 
#>   eta1 <~ y13      0.4542      0.0166   27.3005    0.0000 [ 0.4064; 0.4924 ] 
#>   eta2 <~ y21      0.3058      0.0216   14.1429    0.0000 [ 0.2534; 0.3652 ] 
#>   eta2 <~ y22      0.4473      0.0202   22.1307    0.0000 [ 0.3950; 0.4995 ] 
#>   eta2 <~ y23      0.4735      0.0155   30.5311    0.0000 [ 0.4290; 0.5092 ] 
#>   eta3 <~ y31      0.4400      0.0196   22.4224    0.0000 [ 0.3903; 0.4918 ] 
#>   eta3 <~ y32      0.3521      0.0208   16.9516    0.0000 [ 0.2999; 0.4074 ] 
#>   eta3 <~ y33      0.3999      0.0197   20.2826    0.0000 [ 0.3463; 0.4483 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0430   15.6121    0.0000 [ 0.5551; 0.7775 ] 
#>   eta3 ~ eta1       0.6634      0.0401   16.5379    0.0000 [ 0.5602; 0.7676 ] 
#>   eta3 ~ eta2       0.3052      0.0849    3.5963    0.0003 [ 0.0918; 0.5306 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0573    3.5766    0.0003 [ 0.0599; 0.3561 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04300084 15.612102 6.021989e-55
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07891180  5.810370 6.233495e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08485160  3.596292 3.227854e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55512596          0.7775020          0.5818290          0.7507990
#> 2         0.25186427          0.6599516          0.3008676          0.6109483
#> 3         0.09180936          0.5306139          0.1445012          0.4779221
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5820710          0.7497187          0.6051508          0.7468251
#> 2          0.3083410          0.6284138          0.3290044          0.6161759
#> 3          0.1137864          0.4440239          0.1286010          0.4343775