Summarize model

stable

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1448516543
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0450   14.9256    0.0000 [ 0.5797; 0.7422 ] 
#>   eta3 ~ eta1      0.4585      0.0816    5.6166    0.0000 [ 0.2951; 0.6608 ] 
#>   eta3 ~ eta2      0.3052      0.0808    3.7746    0.0002 [ 0.1142; 0.4427 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0430   15.4167    0.0000 [ 0.5844; 0.7694 ] 
#>   eta1 =~ y12      0.6493      0.0435   14.9359    0.0000 [ 0.5361; 0.6913 ] 
#>   eta1 =~ y13      0.7613      0.0376   20.2732    0.0000 [ 0.6801; 0.8230 ] 
#>   eta2 =~ y21      0.5165      0.0585    8.8221    0.0000 [ 0.3864; 0.5877 ] 
#>   eta2 =~ y22      0.7554      0.0342   22.0990    0.0000 [ 0.7037; 0.8197 ] 
#>   eta2 =~ y23      0.7997      0.0355   22.5237    0.0000 [ 0.7343; 0.8617 ] 
#>   eta3 =~ y31      0.8223      0.0353   23.2870    0.0000 [ 0.7547; 0.8870 ] 
#>   eta3 =~ y32      0.6581      0.0379   17.3646    0.0000 [ 0.5919; 0.7234 ] 
#>   eta3 =~ y33      0.7474      0.0382   19.5765    0.0000 [ 0.6681; 0.8074 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0280   14.1044    0.0000 [ 0.3641; 0.4693 ] 
#>   eta1 <~ y12      0.3873      0.0203   19.1000    0.0000 [ 0.3402; 0.4088 ] 
#>   eta1 <~ y13      0.4542      0.0186   24.3816    0.0000 [ 0.4256; 0.4897 ] 
#>   eta2 <~ y21      0.3058      0.0294   10.4070    0.0000 [ 0.2391; 0.3532 ] 
#>   eta2 <~ y22      0.4473      0.0207   21.5898    0.0000 [ 0.4144; 0.4852 ] 
#>   eta2 <~ y23      0.4735      0.0219   21.6550    0.0000 [ 0.4322; 0.5136 ] 
#>   eta3 <~ y31      0.4400      0.0149   29.5423    0.0000 [ 0.4199; 0.4645 ] 
#>   eta3 <~ y32      0.3521      0.0181   19.4721    0.0000 [ 0.3189; 0.3896 ] 
#>   eta3 <~ y33      0.3999      0.0208   19.2066    0.0000 [ 0.3557; 0.4283 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0450   14.9256    0.0000 [ 0.5797; 0.7422 ] 
#>   eta3 ~ eta1       0.6634      0.0428   15.5039    0.0000 [ 0.5834; 0.7399 ] 
#>   eta3 ~ eta2       0.3052      0.0808    3.7746    0.0002 [ 0.1142; 0.4427 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0516    3.9693    0.0001 [ 0.0791; 0.2892 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04300998 15.416652  1.264943e-53
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04347106 14.935867  1.925539e-50
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03755424 20.273230  2.216221e-91
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05854107  8.822094  1.123393e-18
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03418199 22.098998 3.231609e-108
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03550327 22.523662 2.433985e-112
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03531062 23.286972 6.008335e-120
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03789720 17.364577  1.530230e-67
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03817959 19.576536  2.451340e-85
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5844162          0.7693603
#> 2          0.5360643          0.6912905
#> 3          0.6801222          0.8229997
#> 4          0.3863727          0.5877064
#> 5          0.7036584          0.8196760
#> 6          0.7343090          0.8617070
#> 7          0.7547130          0.8870358
#> 8          0.5919415          0.7234484
#> 9          0.6680650          0.8074060

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1448516543
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0450   14.9256    0.0000 [ 0.5512; 0.7838 ] 
#>   eta3 ~ eta1      0.4585      0.0816    5.6166    0.0000 [ 0.2286; 0.6508 ] 
#>   eta3 ~ eta2      0.3052      0.0808    3.7746    0.0002 [ 0.1139; 0.5320 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0430   15.4167    0.0000 [ 0.5435; 0.7659 ] 
#>   eta1 =~ y12      0.6493      0.0435   14.9359    0.0000 [ 0.5549; 0.7797 ] 
#>   eta1 =~ y13      0.7613      0.0376   20.2732    0.0000 [ 0.6646; 0.8588 ] 
#>   eta2 =~ y21      0.5165      0.0585    8.8221    0.0000 [ 0.3745; 0.6772 ] 
#>   eta2 =~ y22      0.7554      0.0342   22.0990    0.0000 [ 0.6569; 0.8336 ] 
#>   eta2 =~ y23      0.7997      0.0355   22.5237    0.0000 [ 0.7132; 0.8968 ] 
#>   eta3 =~ y31      0.8223      0.0353   23.2870    0.0000 [ 0.7301; 0.9127 ] 
#>   eta3 =~ y32      0.6581      0.0379   17.3646    0.0000 [ 0.5528; 0.7487 ] 
#>   eta3 =~ y33      0.7474      0.0382   19.5765    0.0000 [ 0.6565; 0.8539 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0280   14.1044    0.0000 [ 0.3153; 0.4603 ] 
#>   eta1 <~ y12      0.3873      0.0203   19.1000    0.0000 [ 0.3435; 0.4484 ] 
#>   eta1 <~ y13      0.4542      0.0186   24.3816    0.0000 [ 0.4034; 0.4998 ] 
#>   eta2 <~ y21      0.3058      0.0294   10.4070    0.0000 [ 0.2355; 0.3875 ] 
#>   eta2 <~ y22      0.4473      0.0207   21.5898    0.0000 [ 0.3871; 0.4942 ] 
#>   eta2 <~ y23      0.4735      0.0219   21.6550    0.0000 [ 0.4195; 0.5325 ] 
#>   eta3 <~ y31      0.4400      0.0149   29.5423    0.0000 [ 0.4013; 0.4783 ] 
#>   eta3 <~ y32      0.3521      0.0181   19.4721    0.0000 [ 0.3017; 0.3952 ] 
#>   eta3 <~ y33      0.3999      0.0208   19.2066    0.0000 [ 0.3503; 0.4580 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0450   14.9256    0.0000 [ 0.5512; 0.7838 ] 
#>   eta3 ~ eta1       0.6634      0.0428   15.5039    0.0000 [ 0.5462; 0.7675 ] 
#>   eta3 ~ eta2       0.3052      0.0808    3.7746    0.0002 [ 0.1139; 0.5320 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0516    3.9693    0.0001 [ 0.0837; 0.3506 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04497877 14.925563 2.247316e-50
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08163369  5.616636 1.947107e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08084418  3.774559 1.602911e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5511944          0.7837992          0.5791257          0.7558679
#> 2          0.2286396          0.6508030          0.2793332          0.6001095
#> 3          0.1138769          0.5319574          0.1640802          0.4817541
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5638533          0.7602275          0.5796621          0.7421987
#> 2          0.2925006          0.6696826          0.2950961          0.6608253
#> 3          0.1085743          0.4604986          0.1141781          0.4427171