Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 902047534
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0408   16.4463    0.0000 [ 0.6207; 0.7646 ] 
#>   eta3 ~ eta1      0.4585      0.0811    5.6548    0.0000 [ 0.3033; 0.6042 ] 
#>   eta3 ~ eta2      0.3052      0.0855    3.5690    0.0004 [ 0.1358; 0.4804 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0459   14.4378    0.0000 [ 0.5990; 0.7605 ] 
#>   eta1 =~ y12      0.6493      0.0280   23.1720    0.0000 [ 0.5847; 0.6881 ] 
#>   eta1 =~ y13      0.7613      0.0288   26.4356    0.0000 [ 0.7009; 0.8044 ] 
#>   eta2 =~ y21      0.5165      0.0571    9.0393    0.0000 [ 0.3929; 0.6105 ] 
#>   eta2 =~ y22      0.7554      0.0377   20.0625    0.0000 [ 0.6747; 0.8068 ] 
#>   eta2 =~ y23      0.7997      0.0291   27.4986    0.0000 [ 0.7476; 0.8470 ] 
#>   eta3 =~ y31      0.8223      0.0292   28.2039    0.0000 [ 0.7761; 0.8749 ] 
#>   eta3 =~ y32      0.6581      0.0375   17.5640    0.0000 [ 0.5852; 0.7171 ] 
#>   eta3 =~ y33      0.7474      0.0459   16.2954    0.0000 [ 0.6506; 0.8223 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0197   20.0307    0.0000 [ 0.3693; 0.4353 ] 
#>   eta1 <~ y12      0.3873      0.0184   21.0342    0.0000 [ 0.3501; 0.4208 ] 
#>   eta1 <~ y13      0.4542      0.0179   25.3053    0.0000 [ 0.4197; 0.4869 ] 
#>   eta2 <~ y21      0.3058      0.0296   10.3275    0.0000 [ 0.2505; 0.3527 ] 
#>   eta2 <~ y22      0.4473      0.0248   18.0219    0.0000 [ 0.3950; 0.4952 ] 
#>   eta2 <~ y23      0.4735      0.0167   28.2966    0.0000 [ 0.4577; 0.5162 ] 
#>   eta3 <~ y31      0.4400      0.0157   28.0953    0.0000 [ 0.4135; 0.4688 ] 
#>   eta3 <~ y32      0.3521      0.0187   18.8580    0.0000 [ 0.3189; 0.3916 ] 
#>   eta3 <~ y33      0.3999      0.0216   18.5482    0.0000 [ 0.3604; 0.4373 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0408   16.4463    0.0000 [ 0.6207; 0.7646 ] 
#>   eta3 ~ eta1       0.6634      0.0345   19.2128    0.0000 [ 0.6105; 0.7143 ] 
#>   eta3 ~ eta2       0.3052      0.0855    3.5690    0.0004 [ 0.1358; 0.4804 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0643    3.1852    0.0014 [ 0.0844; 0.3343 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04592589 14.437823  2.991423e-47
#> 2 eta1 =~ y12  Common factor 0.6492779 0.02801991 23.172020 8.721443e-119
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02880002 26.435600 5.342600e-154
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05713421  9.039326  1.576402e-19
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03765171 20.062505  1.569727e-89
#> 6 eta2 =~ y23  Common factor 0.7996637 0.02908017 27.498588 1.825243e-166
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02915470 28.203942 5.231530e-175
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03746689 17.564013  4.647048e-69
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04586731 16.295355  1.064835e-59
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5989829          0.7605081
#> 2          0.5847013          0.6881197
#> 3          0.7009433          0.8044036
#> 4          0.3928653          0.6105252
#> 5          0.6747008          0.8067540
#> 6          0.7476152          0.8470078
#> 7          0.7761225          0.8749105
#> 8          0.5852167          0.7171177
#> 9          0.6506162          0.8223481

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 902047534
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0408   16.4463    0.0000 [ 0.5601; 0.7712 ] 
#>   eta3 ~ eta1      0.4585      0.0811    5.6548    0.0000 [ 0.2679; 0.6872 ] 
#>   eta3 ~ eta2      0.3052      0.0855    3.5690    0.0004 [ 0.0586; 0.5008 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0459   14.4378    0.0000 [ 0.5366; 0.7741 ] 
#>   eta1 =~ y12      0.6493      0.0280   23.1720    0.0000 [ 0.5836; 0.7285 ] 
#>   eta1 =~ y13      0.7613      0.0288   26.4356    0.0000 [ 0.6930; 0.8419 ] 
#>   eta2 =~ y21      0.5165      0.0571    9.0393    0.0000 [ 0.3721; 0.6676 ] 
#>   eta2 =~ y22      0.7554      0.0377   20.0625    0.0000 [ 0.6669; 0.8616 ] 
#>   eta2 =~ y23      0.7997      0.0291   27.4986    0.0000 [ 0.7225; 0.8729 ] 
#>   eta3 =~ y31      0.8223      0.0292   28.2039    0.0000 [ 0.7526; 0.9034 ] 
#>   eta3 =~ y32      0.6581      0.0375   17.5640    0.0000 [ 0.5499; 0.7437 ] 
#>   eta3 =~ y33      0.7474      0.0459   16.2954    0.0000 [ 0.6324; 0.8696 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0197   20.0307    0.0000 [ 0.3389; 0.4410 ] 
#>   eta1 <~ y12      0.3873      0.0184   21.0342    0.0000 [ 0.3423; 0.4375 ] 
#>   eta1 <~ y13      0.4542      0.0179   25.3053    0.0000 [ 0.4097; 0.5026 ] 
#>   eta2 <~ y21      0.3058      0.0296   10.3275    0.0000 [ 0.2298; 0.3829 ] 
#>   eta2 <~ y22      0.4473      0.0248   18.0219    0.0000 [ 0.3855; 0.5138 ] 
#>   eta2 <~ y23      0.4735      0.0167   28.2966    0.0000 [ 0.4261; 0.5127 ] 
#>   eta3 <~ y31      0.4400      0.0157   28.0953    0.0000 [ 0.4030; 0.4840 ] 
#>   eta3 <~ y32      0.3521      0.0187   18.8580    0.0000 [ 0.2983; 0.3949 ] 
#>   eta3 <~ y33      0.3999      0.0216   18.5482    0.0000 [ 0.3467; 0.4582 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0408   16.4463    0.0000 [ 0.5601; 0.7712 ] 
#>   eta3 ~ eta1       0.6634      0.0345   19.2128    0.0000 [ 0.5730; 0.7515 ] 
#>   eta3 ~ eta2       0.3052      0.0855    3.5690    0.0004 [ 0.0586; 0.5008 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0643    3.1852    0.0014 [ 0.0184; 0.3510 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04081962 16.446343 8.908682e-61
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08108215  5.654842 1.559896e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08549984  3.569026 3.583115e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56009934          0.7711953          0.5854478          0.7458468
#> 2         0.26786978          0.6871809          0.3182208          0.6368298
#> 3         0.05864051          0.5007974          0.1117349          0.4477030
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.61712840          0.7916731          0.6206839          0.7645779
#> 2         0.26697928          0.6524550          0.3032528          0.6042416
#> 3         0.08381442          0.5012789          0.1357910          0.4804181