Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1408148628
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0443   15.1456    0.0000 [ 0.5834; 0.7726 ] 
#>   eta3 ~ eta1      0.4585      0.0757    6.0588    0.0000 [ 0.3267; 0.5760 ] 
#>   eta3 ~ eta2      0.3052      0.0804    3.7933    0.0001 [ 0.1446; 0.4153 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0369   17.9807    0.0000 [ 0.6199; 0.7283 ] 
#>   eta1 =~ y12      0.6493      0.0351   18.4905    0.0000 [ 0.5472; 0.6851 ] 
#>   eta1 =~ y13      0.7613      0.0324   23.4883    0.0000 [ 0.6961; 0.8132 ] 
#>   eta2 =~ y21      0.5165      0.0501   10.3142    0.0000 [ 0.4022; 0.5799 ] 
#>   eta2 =~ y22      0.7554      0.0366   20.6325    0.0000 [ 0.6931; 0.8137 ] 
#>   eta2 =~ y23      0.7997      0.0395   20.2587    0.0000 [ 0.7460; 0.8790 ] 
#>   eta3 =~ y31      0.8223      0.0300   27.3849    0.0000 [ 0.7722; 0.8836 ] 
#>   eta3 =~ y32      0.6581      0.0397   16.5645    0.0000 [ 0.5946; 0.7580 ] 
#>   eta3 =~ y33      0.7474      0.0409   18.2738    0.0000 [ 0.6751; 0.8119 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0209   18.9489    0.0000 [ 0.3678; 0.4433 ] 
#>   eta1 <~ y12      0.3873      0.0161   24.0200    0.0000 [ 0.3438; 0.4037 ] 
#>   eta1 <~ y13      0.4542      0.0189   24.0844    0.0000 [ 0.4176; 0.4930 ] 
#>   eta2 <~ y21      0.3058      0.0273   11.2130    0.0000 [ 0.2517; 0.3370 ] 
#>   eta2 <~ y22      0.4473      0.0221   20.2454    0.0000 [ 0.4111; 0.5018 ] 
#>   eta2 <~ y23      0.4735      0.0188   25.2491    0.0000 [ 0.4367; 0.5113 ] 
#>   eta3 <~ y31      0.4400      0.0198   22.2772    0.0000 [ 0.4208; 0.4886 ] 
#>   eta3 <~ y32      0.3521      0.0178   19.8071    0.0000 [ 0.3230; 0.3927 ] 
#>   eta3 <~ y33      0.3999      0.0187   21.4058    0.0000 [ 0.3598; 0.4227 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0443   15.1456    0.0000 [ 0.5834; 0.7726 ] 
#>   eta3 ~ eta1       0.6634      0.0349   18.9882    0.0000 [ 0.5816; 0.7034 ] 
#>   eta3 ~ eta2       0.3052      0.0804    3.7933    0.0001 [ 0.1446; 0.4153 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0549    3.7332    0.0002 [ 0.0961; 0.2864 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03687670 17.98073  2.758528e-72
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03511417 18.49048  2.463473e-76
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03241386 23.48828 5.374229e-122
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05007218 10.31421  6.078062e-25
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03661158 20.63248  1.402629e-94
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03947257 20.25872  2.976080e-91
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03002663 27.38493 4.145820e-165
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03972778 16.56445  1.259254e-61
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04090147 18.27377  1.338723e-74
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6198705          0.7282862
#> 2          0.5472491          0.6851049
#> 3          0.6961079          0.8131768
#> 4          0.4022209          0.5799353
#> 5          0.6931292          0.8136861
#> 6          0.7459564          0.8789711
#> 7          0.7722430          0.8836231
#> 8          0.5946236          0.7579853
#> 9          0.6750761          0.8118655

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1408148628
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0443   15.1456    0.0000 [ 0.5554; 0.7846 ] 
#>   eta3 ~ eta1      0.4585      0.0757    6.0588    0.0000 [ 0.2559; 0.6472 ] 
#>   eta3 ~ eta2      0.3052      0.0804    3.7933    0.0001 [ 0.1245; 0.5405 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0369   17.9807    0.0000 [ 0.5563; 0.7470 ] 
#>   eta1 =~ y12      0.6493      0.0351   18.4905    0.0000 [ 0.5773; 0.7589 ] 
#>   eta1 =~ y13      0.7613      0.0324   23.4883    0.0000 [ 0.6812; 0.8489 ] 
#>   eta2 =~ y21      0.5165      0.0501   10.3142    0.0000 [ 0.3961; 0.6551 ] 
#>   eta2 =~ y22      0.7554      0.0366   20.6325    0.0000 [ 0.6644; 0.8538 ] 
#>   eta2 =~ y23      0.7997      0.0395   20.2587    0.0000 [ 0.6875; 0.8917 ] 
#>   eta3 =~ y31      0.8223      0.0300   27.3849    0.0000 [ 0.7303; 0.8856 ] 
#>   eta3 =~ y32      0.6581      0.0397   16.5645    0.0000 [ 0.5535; 0.7589 ] 
#>   eta3 =~ y33      0.7474      0.0409   18.2738    0.0000 [ 0.6507; 0.8622 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0209   18.9489    0.0000 [ 0.3318; 0.4397 ] 
#>   eta1 <~ y12      0.3873      0.0161   24.0200    0.0000 [ 0.3543; 0.4377 ] 
#>   eta1 <~ y13      0.4542      0.0189   24.0844    0.0000 [ 0.4042; 0.5017 ] 
#>   eta2 <~ y21      0.3058      0.0273   11.2130    0.0000 [ 0.2409; 0.3819 ] 
#>   eta2 <~ y22      0.4473      0.0221   20.2454    0.0000 [ 0.3923; 0.5065 ] 
#>   eta2 <~ y23      0.4735      0.0188   25.2491    0.0000 [ 0.4192; 0.5162 ] 
#>   eta3 <~ y31      0.4400      0.0198   22.2772    0.0000 [ 0.3835; 0.4856 ] 
#>   eta3 <~ y32      0.3521      0.0178   19.8071    0.0000 [ 0.3072; 0.3992 ] 
#>   eta3 <~ y33      0.3999      0.0187   21.4058    0.0000 [ 0.3587; 0.4553 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0443   15.1456    0.0000 [ 0.5554; 0.7846 ] 
#>   eta3 ~ eta1       0.6634      0.0349   18.9882    0.0000 [ 0.5842; 0.7649 ] 
#>   eta3 ~ eta2       0.3052      0.0804    3.7933    0.0001 [ 0.1245; 0.5405 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0549    3.7332    0.0002 [ 0.0811; 0.3649 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04432543 15.145561 8.106012e-52
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07567672  6.058756 1.371782e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08044516  3.793281 1.486696e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5553769          0.7846030          0.5829025          0.7570774
#> 2          0.2558608          0.6472180          0.3028552          0.6002237
#> 3          0.1245157          0.5405326          0.1744712          0.4905771
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5452088          0.7763419          0.5834172          0.7726239
#> 2          0.3126080          0.6270324          0.3266929          0.5760324
#> 3          0.0802686          0.4225205          0.1445685          0.4153309