Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1202602573
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0403   16.6633    0.0000 [ 0.5962; 0.7418 ] 
#>   eta3 ~ eta1      0.4585      0.0756    6.0632    0.0000 [ 0.3286; 0.5911 ] 
#>   eta3 ~ eta2      0.3052      0.0774    3.9446    0.0001 [ 0.1652; 0.4459 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0442   15.0013    0.0000 [ 0.5751; 0.7257 ] 
#>   eta1 =~ y12      0.6493      0.0383   16.9419    0.0000 [ 0.5824; 0.7033 ] 
#>   eta1 =~ y13      0.7613      0.0327   23.3040    0.0000 [ 0.7118; 0.8222 ] 
#>   eta2 =~ y21      0.5165      0.0545    9.4691    0.0000 [ 0.4293; 0.6089 ] 
#>   eta2 =~ y22      0.7554      0.0348   21.7230    0.0000 [ 0.7136; 0.8155 ] 
#>   eta2 =~ y23      0.7997      0.0301   26.5573    0.0000 [ 0.7434; 0.8672 ] 
#>   eta3 =~ y31      0.8223      0.0356   23.0922    0.0000 [ 0.7613; 0.8765 ] 
#>   eta3 =~ y32      0.6581      0.0313   20.9999    0.0000 [ 0.5920; 0.7040 ] 
#>   eta3 =~ y33      0.7474      0.0427   17.5244    0.0000 [ 0.6726; 0.8056 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0231   17.1132    0.0000 [ 0.3458; 0.4354 ] 
#>   eta1 <~ y12      0.3873      0.0200   19.3386    0.0000 [ 0.3445; 0.4209 ] 
#>   eta1 <~ y13      0.4542      0.0201   22.6376    0.0000 [ 0.4298; 0.4867 ] 
#>   eta2 <~ y21      0.3058      0.0258   11.8370    0.0000 [ 0.2649; 0.3475 ] 
#>   eta2 <~ y22      0.4473      0.0216   20.6843    0.0000 [ 0.4065; 0.4839 ] 
#>   eta2 <~ y23      0.4735      0.0203   23.2758    0.0000 [ 0.4306; 0.4990 ] 
#>   eta3 <~ y31      0.4400      0.0206   21.3351    0.0000 [ 0.4053; 0.4808 ] 
#>   eta3 <~ y32      0.3521      0.0155   22.7393    0.0000 [ 0.3223; 0.3774 ] 
#>   eta3 <~ y33      0.3999      0.0177   22.6129    0.0000 [ 0.3702; 0.4343 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0403   16.6633    0.0000 [ 0.5962; 0.7418 ] 
#>   eta3 ~ eta1       0.6634      0.0398   16.6670    0.0000 [ 0.6058; 0.7448 ] 
#>   eta3 ~ eta2       0.3052      0.0774    3.9446    0.0001 [ 0.1652; 0.4459 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0556    3.6846    0.0002 [ 0.1134; 0.3143 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04420073 15.001332  7.196090e-51
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03832380 16.941901  2.208719e-64
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03267019 23.303993 4.038640e-120
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05454123  9.469071  2.823434e-21
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03477370 21.722956 1.245034e-104
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03011093 26.557260 2.117312e-155
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03560844 23.092201 5.545680e-118
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03133682 20.999862  6.577599e-98
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04265059 17.524355  9.339462e-69
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5751111          0.7257370
#> 2          0.5824196          0.7033381
#> 3          0.7118458          0.8221537
#> 4          0.4292555          0.6088831
#> 5          0.7136185          0.8154891
#> 6          0.7433904          0.8672077
#> 7          0.7612874          0.8765274
#> 8          0.5920302          0.7039700
#> 9          0.6726234          0.8056446

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1202602573
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0403   16.6633    0.0000 [ 0.5689; 0.7772 ] 
#>   eta3 ~ eta1      0.4585      0.0756    6.0632    0.0000 [ 0.2694; 0.6605 ] 
#>   eta3 ~ eta2      0.3052      0.0774    3.9446    0.0001 [ 0.0960; 0.4961 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0442   15.0013    0.0000 [ 0.5477; 0.7763 ] 
#>   eta1 =~ y12      0.6493      0.0383   16.9419    0.0000 [ 0.5540; 0.7522 ] 
#>   eta1 =~ y13      0.7613      0.0327   23.3040    0.0000 [ 0.6698; 0.8387 ] 
#>   eta2 =~ y21      0.5165      0.0545    9.4691    0.0000 [ 0.3656; 0.6476 ] 
#>   eta2 =~ y22      0.7554      0.0348   21.7230    0.0000 [ 0.6617; 0.8415 ] 
#>   eta2 =~ y23      0.7997      0.0301   26.5573    0.0000 [ 0.7228; 0.8785 ] 
#>   eta3 =~ y31      0.8223      0.0356   23.0922    0.0000 [ 0.7312; 0.9153 ] 
#>   eta3 =~ y32      0.6581      0.0313   20.9999    0.0000 [ 0.5758; 0.7378 ] 
#>   eta3 =~ y33      0.7474      0.0427   17.5244    0.0000 [ 0.6420; 0.8626 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0231   17.1132    0.0000 [ 0.3368; 0.4563 ] 
#>   eta1 <~ y12      0.3873      0.0200   19.3386    0.0000 [ 0.3394; 0.4430 ] 
#>   eta1 <~ y13      0.4542      0.0201   22.6376    0.0000 [ 0.3997; 0.5035 ] 
#>   eta2 <~ y21      0.3058      0.0258   11.8370    0.0000 [ 0.2365; 0.3701 ] 
#>   eta2 <~ y22      0.4473      0.0216   20.6843    0.0000 [ 0.3931; 0.5049 ] 
#>   eta2 <~ y23      0.4735      0.0203   23.2758    0.0000 [ 0.4256; 0.5308 ] 
#>   eta3 <~ y31      0.4400      0.0206   21.3351    0.0000 [ 0.3857; 0.4924 ] 
#>   eta3 <~ y32      0.3521      0.0155   22.7393    0.0000 [ 0.3104; 0.3904 ] 
#>   eta3 <~ y33      0.3999      0.0177   22.6129    0.0000 [ 0.3558; 0.4473 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0403   16.6633    0.0000 [ 0.5689; 0.7772 ] 
#>   eta3 ~ eta1       0.6634      0.0398   16.6670    0.0000 [ 0.5608; 0.7666 ] 
#>   eta3 ~ eta2       0.3052      0.0774    3.9446    0.0001 [ 0.0960; 0.4961 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0556    3.6846    0.0002 [ 0.0550; 0.3425 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04028804 16.663341 2.421121e-62
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07562074  6.063241 1.334053e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07735922  3.944600 7.993345e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56885182          0.7771988          0.5938702          0.7521804
#> 2         0.26944023          0.6605080          0.3163998          0.6135484
#> 3         0.09601687          0.4960750          0.1440560          0.4480359
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5779314          0.7459122          0.5962237          0.7417888
#> 2          0.3235253          0.6171645          0.3285602          0.5910975
#> 3          0.1333548          0.4757748          0.1651857          0.4459415