Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -1968378111
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0417   16.1046    0.0000 [ 0.5954; 0.7374 ] 
#>   eta3 ~ eta1      0.4585      0.0753    6.0918    0.0000 [ 0.3022; 0.6048 ] 
#>   eta3 ~ eta2      0.3052      0.0870    3.5058    0.0005 [ 0.1580; 0.5057 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0358   18.5403    0.0000 [ 0.6089; 0.7259 ] 
#>   eta1 =~ y12      0.6493      0.0409   15.8592    0.0000 [ 0.5661; 0.7068 ] 
#>   eta1 =~ y13      0.7613      0.0359   21.2300    0.0000 [ 0.7124; 0.8649 ] 
#>   eta2 =~ y21      0.5165      0.0472   10.9458    0.0000 [ 0.4229; 0.5940 ] 
#>   eta2 =~ y22      0.7554      0.0374   20.2147    0.0000 [ 0.6921; 0.8216 ] 
#>   eta2 =~ y23      0.7997      0.0444   18.0179    0.0000 [ 0.7102; 0.8617 ] 
#>   eta3 =~ y31      0.8223      0.0326   25.1862    0.0000 [ 0.7809; 0.8889 ] 
#>   eta3 =~ y32      0.6581      0.0422   15.5796    0.0000 [ 0.5932; 0.7204 ] 
#>   eta3 =~ y33      0.7474      0.0412   18.1260    0.0000 [ 0.6771; 0.8300 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0228   17.3795    0.0000 [ 0.3568; 0.4450 ] 
#>   eta1 <~ y12      0.3873      0.0192   20.1716    0.0000 [ 0.3463; 0.4152 ] 
#>   eta1 <~ y13      0.4542      0.0204   22.3120    0.0000 [ 0.4275; 0.5040 ] 
#>   eta2 <~ y21      0.3058      0.0248   12.3314    0.0000 [ 0.2693; 0.3611 ] 
#>   eta2 <~ y22      0.4473      0.0202   22.1590    0.0000 [ 0.4132; 0.4793 ] 
#>   eta2 <~ y23      0.4735      0.0242   19.5671    0.0000 [ 0.4325; 0.5299 ] 
#>   eta3 <~ y31      0.4400      0.0209   21.0771    0.0000 [ 0.4077; 0.4931 ] 
#>   eta3 <~ y32      0.3521      0.0211   16.6674    0.0000 [ 0.3175; 0.3831 ] 
#>   eta3 <~ y33      0.3999      0.0179   22.3285    0.0000 [ 0.3716; 0.4323 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0417   16.1046    0.0000 [ 0.5954; 0.7374 ] 
#>   eta3 ~ eta1       0.6634      0.0363   18.2997    0.0000 [ 0.5799; 0.7188 ] 
#>   eta3 ~ eta2       0.3052      0.0870    3.5058    0.0005 [ 0.1580; 0.5057 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0547    3.7469    0.0002 [ 0.1100; 0.3227 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03576377 18.54027  9.773299e-77
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04094003 15.85924  1.213476e-56
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03586176 21.23002 5.044264e-100
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04718282 10.94582  6.958310e-28
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03736815 20.21474  7.262049e-91
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04438169 18.01787  1.410596e-72
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03264791 25.18622 5.671950e-140
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04223906 15.57963  1.001347e-54
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04123500 18.12596  1.988279e-73
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6089333          0.7259135
#> 2          0.5660747          0.7068460
#> 3          0.7124133          0.8649392
#> 4          0.4228890          0.5940093
#> 5          0.6920538          0.8216357
#> 6          0.7102451          0.8616750
#> 7          0.7808646          0.8889461
#> 8          0.5931815          0.7203808
#> 9          0.6770643          0.8299624

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -1968378111
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0417   16.1046    0.0000 [ 0.5600; 0.7756 ] 
#>   eta3 ~ eta1      0.4585      0.0753    6.0918    0.0000 [ 0.2891; 0.6783 ] 
#>   eta3 ~ eta2      0.3052      0.0870    3.5058    0.0005 [ 0.0610; 0.5111 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0358   18.5403    0.0000 [ 0.5645; 0.7495 ] 
#>   eta1 =~ y12      0.6493      0.0409   15.8592    0.0000 [ 0.5555; 0.7672 ] 
#>   eta1 =~ y13      0.7613      0.0359   21.2300    0.0000 [ 0.6674; 0.8529 ] 
#>   eta2 =~ y21      0.5165      0.0472   10.9458    0.0000 [ 0.3928; 0.6368 ] 
#>   eta2 =~ y22      0.7554      0.0374   20.2147    0.0000 [ 0.6640; 0.8572 ] 
#>   eta2 =~ y23      0.7997      0.0444   18.0179    0.0000 [ 0.6909; 0.9205 ] 
#>   eta3 =~ y31      0.8223      0.0326   25.1862    0.0000 [ 0.7349; 0.9037 ] 
#>   eta3 =~ y32      0.6581      0.0422   15.5796    0.0000 [ 0.5469; 0.7654 ] 
#>   eta3 =~ y33      0.7474      0.0412   18.1260    0.0000 [ 0.6419; 0.8551 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0228   17.3795    0.0000 [ 0.3319; 0.4496 ] 
#>   eta1 <~ y12      0.3873      0.0192   20.1716    0.0000 [ 0.3441; 0.4434 ] 
#>   eta1 <~ y13      0.4542      0.0204   22.3120    0.0000 [ 0.3995; 0.5048 ] 
#>   eta2 <~ y21      0.3058      0.0248   12.3314    0.0000 [ 0.2388; 0.3671 ] 
#>   eta2 <~ y22      0.4473      0.0202   22.1590    0.0000 [ 0.3951; 0.4995 ] 
#>   eta2 <~ y23      0.4735      0.0242   19.5671    0.0000 [ 0.4113; 0.5365 ] 
#>   eta3 <~ y31      0.4400      0.0209   21.0771    0.0000 [ 0.3856; 0.4936 ] 
#>   eta3 <~ y32      0.3521      0.0211   16.6674    0.0000 [ 0.2976; 0.4069 ] 
#>   eta3 <~ y33      0.3999      0.0179   22.3285    0.0000 [ 0.3556; 0.4482 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0417   16.1046    0.0000 [ 0.5600; 0.7756 ] 
#>   eta3 ~ eta1       0.6634      0.0363   18.2997    0.0000 [ 0.5824; 0.7699 ] 
#>   eta3 ~ eta2       0.3052      0.0870    3.5058    0.0005 [ 0.0610; 0.5111 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0547    3.7469    0.0002 [ 0.0511; 0.3338 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04168579 16.104611 2.367806e-58
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07526610  6.091810 1.116413e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08704182  3.505799 4.552391e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55998247          0.7755578          0.5858689          0.7496714
#> 2         0.28906279          0.6782966          0.3358022          0.6315572
#> 3         0.06101045          0.5111416          0.1150624          0.4570897
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5843015          0.7376280          0.5954070          0.7373940
#> 2          0.2748216          0.6080722          0.3021763          0.6048204
#> 3          0.1560908          0.5234196          0.1580494          0.5056553