Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -123142337
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0415   16.1885    0.0000 [ 0.5918; 0.7404 ] 
#>   eta3 ~ eta1      0.4585      0.0766    5.9857    0.0000 [ 0.3410; 0.6071 ] 
#>   eta3 ~ eta2      0.3052      0.0806    3.7864    0.0002 [ 0.1115; 0.4554 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0418   15.8803    0.0000 [ 0.5945; 0.7394 ] 
#>   eta1 =~ y12      0.6493      0.0318   20.4154    0.0000 [ 0.5929; 0.7058 ] 
#>   eta1 =~ y13      0.7613      0.0355   21.4641    0.0000 [ 0.6834; 0.8210 ] 
#>   eta2 =~ y21      0.5165      0.0572    9.0291    0.0000 [ 0.3720; 0.5884 ] 
#>   eta2 =~ y22      0.7554      0.0392   19.2572    0.0000 [ 0.6922; 0.8378 ] 
#>   eta2 =~ y23      0.7997      0.0478   16.7135    0.0000 [ 0.7040; 0.8769 ] 
#>   eta3 =~ y31      0.8223      0.0266   30.9029    0.0000 [ 0.7653; 0.8597 ] 
#>   eta3 =~ y32      0.6581      0.0344   19.1327    0.0000 [ 0.5964; 0.7244 ] 
#>   eta3 =~ y33      0.7474      0.0359   20.7980    0.0000 [ 0.6690; 0.8131 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0231   17.1324    0.0000 [ 0.3538; 0.4336 ] 
#>   eta1 <~ y12      0.3873      0.0183   21.2050    0.0000 [ 0.3545; 0.4211 ] 
#>   eta1 <~ y13      0.4542      0.0187   24.2628    0.0000 [ 0.4169; 0.4859 ] 
#>   eta2 <~ y21      0.3058      0.0302   10.1303    0.0000 [ 0.2400; 0.3484 ] 
#>   eta2 <~ y22      0.4473      0.0251   17.8184    0.0000 [ 0.4109; 0.4967 ] 
#>   eta2 <~ y23      0.4735      0.0238   19.9167    0.0000 [ 0.4351; 0.5127 ] 
#>   eta3 <~ y31      0.4400      0.0152   28.9598    0.0000 [ 0.4110; 0.4663 ] 
#>   eta3 <~ y32      0.3521      0.0164   21.4777    0.0000 [ 0.3271; 0.3872 ] 
#>   eta3 <~ y33      0.3999      0.0176   22.7727    0.0000 [ 0.3668; 0.4364 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0415   16.1885    0.0000 [ 0.5918; 0.7404 ] 
#>   eta3 ~ eta1       0.6634      0.0392   16.9336    0.0000 [ 0.5824; 0.7404 ] 
#>   eta3 ~ eta2       0.3052      0.0806    3.7864    0.0002 [ 0.1115; 0.4554 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0555    3.6935    0.0002 [ 0.0866; 0.3354 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04175433 15.88027  8.681077e-57
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03180339 20.41537  1.221128e-92
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03547060 21.46413 3.369674e-102
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05719874  9.02913  1.730422e-19
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03922619 19.25723  1.227826e-82
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04784545 16.71347  1.045646e-62
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02660840 30.90292 1.091189e-209
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03439493 19.13273  1.348240e-81
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03593730 20.79800  4.512065e-96
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5944547          0.7394428
#> 2          0.5929334          0.7057875
#> 3          0.6834065          0.8209614
#> 4          0.3719595          0.5884202
#> 5          0.6921825          0.8378305
#> 6          0.7040059          0.8769090
#> 7          0.7652877          0.8597358
#> 8          0.5964029          0.7244480
#> 9          0.6690243          0.8131062

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -123142337
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0415   16.1885    0.0000 [ 0.5638; 0.7783 ] 
#>   eta3 ~ eta1      0.4585      0.0766    5.9857    0.0000 [ 0.2653; 0.6615 ] 
#>   eta3 ~ eta2      0.3052      0.0806    3.7864    0.0002 [ 0.0823; 0.4991 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0418   15.8803    0.0000 [ 0.5570; 0.7729 ] 
#>   eta1 =~ y12      0.6493      0.0318   20.4154    0.0000 [ 0.5610; 0.7255 ] 
#>   eta1 =~ y13      0.7613      0.0355   21.4641    0.0000 [ 0.6697; 0.8531 ] 
#>   eta2 =~ y21      0.5165      0.0572    9.0291    0.0000 [ 0.3844; 0.6802 ] 
#>   eta2 =~ y22      0.7554      0.0392   19.2572    0.0000 [ 0.6503; 0.8532 ] 
#>   eta2 =~ y23      0.7997      0.0478   16.7135    0.0000 [ 0.6859; 0.9333 ] 
#>   eta3 =~ y31      0.8223      0.0266   30.9029    0.0000 [ 0.7627; 0.9003 ] 
#>   eta3 =~ y32      0.6581      0.0344   19.1327    0.0000 [ 0.5637; 0.7415 ] 
#>   eta3 =~ y33      0.7474      0.0359   20.7980    0.0000 [ 0.6495; 0.8354 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0231   17.1324    0.0000 [ 0.3383; 0.4577 ] 
#>   eta1 <~ y12      0.3873      0.0183   21.2050    0.0000 [ 0.3378; 0.4322 ] 
#>   eta1 <~ y13      0.4542      0.0187   24.2628    0.0000 [ 0.4074; 0.5042 ] 
#>   eta2 <~ y21      0.3058      0.0302   10.1303    0.0000 [ 0.2335; 0.3896 ] 
#>   eta2 <~ y22      0.4473      0.0251   17.8184    0.0000 [ 0.3739; 0.5037 ] 
#>   eta2 <~ y23      0.4735      0.0238   19.9167    0.0000 [ 0.4117; 0.5346 ] 
#>   eta3 <~ y31      0.4400      0.0152   28.9598    0.0000 [ 0.4059; 0.4845 ] 
#>   eta3 <~ y32      0.3521      0.0164   21.4777    0.0000 [ 0.3072; 0.3920 ] 
#>   eta3 <~ y33      0.3999      0.0176   22.7727    0.0000 [ 0.3522; 0.4431 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0415   16.1885    0.0000 [ 0.5638; 0.7783 ] 
#>   eta3 ~ eta1       0.6634      0.0392   16.9336    0.0000 [ 0.5578; 0.7604 ] 
#>   eta3 ~ eta2       0.3052      0.0806    3.7864    0.0002 [ 0.0823; 0.4991 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0555    3.6935    0.0002 [ 0.0523; 0.3391 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err   t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04146973 16.18852 6.077392e-59
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07660023  5.98571 2.154482e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08059092  3.78642 1.528330e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56384274          0.7783007          0.5895950          0.7525485
#> 2         0.26533475          0.6614679          0.3129026          0.6139001
#> 3         0.08231188          0.4990826          0.1323579          0.4490366
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5846397          0.7713069          0.5917872          0.7404326
#> 2          0.3366898          0.6319015          0.3410098          0.6070514
#> 3          0.1083859          0.4728570          0.1114556          0.4553622