Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 982697944
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0480   13.9856    0.0000 [ 0.5920; 0.7568 ] 
#>   eta3 ~ eta1      0.4585      0.0839    5.4651    0.0000 [ 0.3321; 0.6130 ] 
#>   eta3 ~ eta2      0.3052      0.0859    3.5505    0.0004 [ 0.1602; 0.4825 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0397   16.7145    0.0000 [ 0.5788; 0.7244 ] 
#>   eta1 =~ y12      0.6493      0.0306   21.2146    0.0000 [ 0.5862; 0.7063 ] 
#>   eta1 =~ y13      0.7613      0.0324   23.5130    0.0000 [ 0.6944; 0.8140 ] 
#>   eta2 =~ y21      0.5165      0.0495   10.4263    0.0000 [ 0.4160; 0.5891 ] 
#>   eta2 =~ y22      0.7554      0.0335   22.5224    0.0000 [ 0.6816; 0.8080 ] 
#>   eta2 =~ y23      0.7997      0.0401   19.9222    0.0000 [ 0.7314; 0.8668 ] 
#>   eta3 =~ y31      0.8223      0.0283   29.0528    0.0000 [ 0.7736; 0.8676 ] 
#>   eta3 =~ y32      0.6581      0.0404   16.2881    0.0000 [ 0.5440; 0.7182 ] 
#>   eta3 =~ y33      0.7474      0.0359   20.8437    0.0000 [ 0.6906; 0.7995 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0208   19.0595    0.0000 [ 0.3521; 0.4294 ] 
#>   eta1 <~ y12      0.3873      0.0172   22.5050    0.0000 [ 0.3540; 0.4219 ] 
#>   eta1 <~ y13      0.4542      0.0175   25.9598    0.0000 [ 0.4294; 0.4962 ] 
#>   eta2 <~ y21      0.3058      0.0243   12.5900    0.0000 [ 0.2607; 0.3426 ] 
#>   eta2 <~ y22      0.4473      0.0232   19.2997    0.0000 [ 0.4134; 0.4951 ] 
#>   eta2 <~ y23      0.4735      0.0178   26.5468    0.0000 [ 0.4375; 0.5052 ] 
#>   eta3 <~ y31      0.4400      0.0172   25.5145    0.0000 [ 0.4119; 0.4707 ] 
#>   eta3 <~ y32      0.3521      0.0171   20.6100    0.0000 [ 0.3200; 0.3867 ] 
#>   eta3 <~ y33      0.3999      0.0178   22.5246    0.0000 [ 0.3771; 0.4423 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0480   13.9856    0.0000 [ 0.5920; 0.7568 ] 
#>   eta3 ~ eta1       0.6634      0.0421   15.7592    0.0000 [ 0.5874; 0.7361 ] 
#>   eta3 ~ eta2       0.3052      0.0859    3.5505    0.0004 [ 0.1602; 0.4825 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0561    3.6513    0.0003 [ 0.1066; 0.3116 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03967042 16.71446  1.028402e-62
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03060521 21.21462 6.998930e-100
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03237981 23.51298 3.004738e-122
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04953379 10.42631  1.880447e-25
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03353945 22.52236 2.506497e-112
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04013927 19.92223  2.611182e-88
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02830290 29.05276 1.420183e-185
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04040183 16.28810  1.199062e-59
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03585856 20.84367  1.739644e-96
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5788332          0.7244382
#> 2          0.5861557          0.7063199
#> 3          0.6943959          0.8140384
#> 4          0.4160475          0.5891218
#> 5          0.6816260          0.8080035
#> 6          0.7313820          0.8667531
#> 7          0.7735501          0.8675560
#> 8          0.5439695          0.7182325
#> 9          0.6906136          0.7994581

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 982697944
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0480   13.9856    0.0000 [ 0.5368; 0.7850 ] 
#>   eta3 ~ eta1      0.4585      0.0839    5.4651    0.0000 [ 0.2240; 0.6579 ] 
#>   eta3 ~ eta2      0.3052      0.0859    3.5505    0.0004 [ 0.0956; 0.5401 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0397   16.7145    0.0000 [ 0.5734; 0.7786 ] 
#>   eta1 =~ y12      0.6493      0.0306   21.2146    0.0000 [ 0.5719; 0.7302 ] 
#>   eta1 =~ y13      0.7613      0.0324   23.5130    0.0000 [ 0.6774; 0.8449 ] 
#>   eta2 =~ y21      0.5165      0.0495   10.4263    0.0000 [ 0.3914; 0.6475 ] 
#>   eta2 =~ y22      0.7554      0.0335   22.5224    0.0000 [ 0.6671; 0.8405 ] 
#>   eta2 =~ y23      0.7997      0.0401   19.9222    0.0000 [ 0.6975; 0.9051 ] 
#>   eta3 =~ y31      0.8223      0.0283   29.0528    0.0000 [ 0.7518; 0.8981 ] 
#>   eta3 =~ y32      0.6581      0.0404   16.2881    0.0000 [ 0.5522; 0.7611 ] 
#>   eta3 =~ y33      0.7474      0.0359   20.8437    0.0000 [ 0.6572; 0.8426 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0208   19.0595    0.0000 [ 0.3458; 0.4532 ] 
#>   eta1 <~ y12      0.3873      0.0172   22.5050    0.0000 [ 0.3400; 0.4290 ] 
#>   eta1 <~ y13      0.4542      0.0175   25.9598    0.0000 [ 0.4043; 0.4948 ] 
#>   eta2 <~ y21      0.3058      0.0243   12.5900    0.0000 [ 0.2447; 0.3704 ] 
#>   eta2 <~ y22      0.4473      0.0232   19.2997    0.0000 [ 0.3856; 0.5054 ] 
#>   eta2 <~ y23      0.4735      0.0178   26.5468    0.0000 [ 0.4279; 0.5201 ] 
#>   eta3 <~ y31      0.4400      0.0172   25.5145    0.0000 [ 0.3957; 0.4848 ] 
#>   eta3 <~ y32      0.3521      0.0171   20.6100    0.0000 [ 0.3066; 0.3949 ] 
#>   eta3 <~ y33      0.3999      0.0178   22.5246    0.0000 [ 0.3544; 0.4462 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0480   13.9856    0.0000 [ 0.5368; 0.7850 ] 
#>   eta3 ~ eta1       0.6634      0.0421   15.7592    0.0000 [ 0.5438; 0.7615 ] 
#>   eta3 ~ eta2       0.3052      0.0859    3.5505    0.0004 [ 0.0956; 0.5401 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0561    3.6513    0.0003 [ 0.0666; 0.3567 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04800166 13.985630 1.907801e-44
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08389775  5.465066 4.627338e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08594695  3.550459 3.845598e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.53680294          0.7850404          0.5666114          0.7552319
#> 2         0.22404930          0.6579211          0.2761488          0.6058216
#> 3         0.09562787          0.5400970          0.1489999          0.4867249
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.54394267          0.7622848          0.5920042          0.7568146
#> 2         0.32936511          0.6677424          0.3320952          0.6130253
#> 3         0.09213591          0.4826857          0.1601795          0.4825153