Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1121475708
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0473   14.2080    0.0000 [ 0.6017; 0.7600 ] 
#>   eta3 ~ eta1      0.4585      0.0827    5.5447    0.0000 [ 0.2940; 0.6329 ] 
#>   eta3 ~ eta2      0.3052      0.0934    3.2681    0.0011 [ 0.1259; 0.4694 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0431   15.3795    0.0000 [ 0.5789; 0.7288 ] 
#>   eta1 =~ y12      0.6493      0.0443   14.6584    0.0000 [ 0.5381; 0.7082 ] 
#>   eta1 =~ y13      0.7613      0.0328   23.2042    0.0000 [ 0.7036; 0.8122 ] 
#>   eta2 =~ y21      0.5165      0.0488   10.5746    0.0000 [ 0.4153; 0.5947 ] 
#>   eta2 =~ y22      0.7554      0.0440   17.1835    0.0000 [ 0.6443; 0.8107 ] 
#>   eta2 =~ y23      0.7997      0.0299   26.7200    0.0000 [ 0.7348; 0.8433 ] 
#>   eta3 =~ y31      0.8223      0.0301   27.3379    0.0000 [ 0.7746; 0.8775 ] 
#>   eta3 =~ y32      0.6581      0.0370   17.7890    0.0000 [ 0.5719; 0.7384 ] 
#>   eta3 =~ y33      0.7474      0.0375   19.9191    0.0000 [ 0.6760; 0.8058 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0234   16.9328    0.0000 [ 0.3527; 0.4290 ] 
#>   eta1 <~ y12      0.3873      0.0217   17.8576    0.0000 [ 0.3443; 0.4193 ] 
#>   eta1 <~ y13      0.4542      0.0217   20.9237    0.0000 [ 0.4281; 0.5043 ] 
#>   eta2 <~ y21      0.3058      0.0252   12.1317    0.0000 [ 0.2578; 0.3432 ] 
#>   eta2 <~ y22      0.4473      0.0225   19.8928    0.0000 [ 0.4043; 0.4909 ] 
#>   eta2 <~ y23      0.4735      0.0212   22.3031    0.0000 [ 0.4358; 0.5140 ] 
#>   eta3 <~ y31      0.4400      0.0158   27.9204    0.0000 [ 0.4142; 0.4721 ] 
#>   eta3 <~ y32      0.3521      0.0171   20.5580    0.0000 [ 0.3169; 0.3835 ] 
#>   eta3 <~ y33      0.3999      0.0189   21.1091    0.0000 [ 0.3588; 0.4315 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0473   14.2080    0.0000 [ 0.6017; 0.7600 ] 
#>   eta3 ~ eta1       0.6634      0.0362   18.3137    0.0000 [ 0.5973; 0.7226 ] 
#>   eta3 ~ eta2       0.3052      0.0934    3.2681    0.0011 [ 0.1259; 0.4694 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0641    3.1952    0.0014 [ 0.0890; 0.3292 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04311390 15.37949  2.247069e-53
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04429396 14.65839  1.190680e-48
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03281069 23.20420 4.129404e-119
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04883911 10.57462  3.907769e-26
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04396018 17.18345  3.532446e-66
#> 6 eta2 =~ y23  Common factor 0.7996637 0.02992753 26.72000 2.756709e-157
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03007832 27.33788 1.505060e-164
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03699307 17.78898  8.603424e-71
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03752301 19.91909  2.780079e-88
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5788521          0.7287523
#> 2          0.5381320          0.7082335
#> 3          0.7035911          0.8121973
#> 4          0.4153276          0.5946502
#> 5          0.6443124          0.8106843
#> 6          0.7348025          0.8432797
#> 7          0.7745722          0.8774605
#> 8          0.5718632          0.7383814
#> 9          0.6760339          0.8058069

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1121475708
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0473   14.2080    0.0000 [ 0.5488; 0.7932 ] 
#>   eta3 ~ eta1      0.4585      0.0827    5.5447    0.0000 [ 0.2397; 0.6674 ] 
#>   eta3 ~ eta2      0.3052      0.0934    3.2681    0.0011 [ 0.0734; 0.5562 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0431   15.3795    0.0000 [ 0.5558; 0.7788 ] 
#>   eta1 =~ y12      0.6493      0.0443   14.6584    0.0000 [ 0.5384; 0.7675 ] 
#>   eta1 =~ y13      0.7613      0.0328   23.2042    0.0000 [ 0.6813; 0.8510 ] 
#>   eta2 =~ y21      0.5165      0.0488   10.5746    0.0000 [ 0.3986; 0.6512 ] 
#>   eta2 =~ y22      0.7554      0.0440   17.1835    0.0000 [ 0.6486; 0.8759 ] 
#>   eta2 =~ y23      0.7997      0.0299   26.7200    0.0000 [ 0.7180; 0.8727 ] 
#>   eta3 =~ y31      0.8223      0.0301   27.3379    0.0000 [ 0.7429; 0.8985 ] 
#>   eta3 =~ y32      0.6581      0.0370   17.7890    0.0000 [ 0.5606; 0.7519 ] 
#>   eta3 =~ y33      0.7474      0.0375   19.9191    0.0000 [ 0.6586; 0.8526 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0234   16.9328    0.0000 [ 0.3344; 0.4552 ] 
#>   eta1 <~ y12      0.3873      0.0217   17.8576    0.0000 [ 0.3303; 0.4425 ] 
#>   eta1 <~ y13      0.4542      0.0217   20.9237    0.0000 [ 0.3969; 0.5091 ] 
#>   eta2 <~ y21      0.3058      0.0252   12.1317    0.0000 [ 0.2440; 0.3744 ] 
#>   eta2 <~ y22      0.4473      0.0225   19.8928    0.0000 [ 0.3905; 0.5068 ] 
#>   eta2 <~ y23      0.4735      0.0212   22.3031    0.0000 [ 0.4128; 0.5226 ] 
#>   eta3 <~ y31      0.4400      0.0158   27.9204    0.0000 [ 0.3970; 0.4785 ] 
#>   eta3 <~ y32      0.3521      0.0171   20.5580    0.0000 [ 0.3059; 0.3945 ] 
#>   eta3 <~ y33      0.3999      0.0189   21.1091    0.0000 [ 0.3541; 0.4521 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0473   14.2080    0.0000 [ 0.5488; 0.7932 ] 
#>   eta3 ~ eta1       0.6634      0.0362   18.3137    0.0000 [ 0.5716; 0.7589 ] 
#>   eta3 ~ eta2       0.3052      0.0934    3.2681    0.0011 [ 0.0734; 0.5562 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0641    3.1952    0.0014 [ 0.0459; 0.3775 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04725053 14.207956 8.177755e-46
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08269327  5.544668 2.945117e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09337215  3.268117 1.082655e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54883897          0.7931920          0.5781810          0.7638500
#> 2         0.23974270          0.6673856          0.2910943          0.6160341
#> 3         0.07337675          0.5562448          0.1313598          0.4982618
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.55089713          0.7715395          0.6017466          0.7599670
#> 2         0.28266786          0.6636984          0.2939748          0.6328548
#> 3         0.07442207          0.5054881          0.1259216          0.4693935