Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2039653835
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0460   14.5961    0.0000 [ 0.5914; 0.7419 ] 
#>   eta3 ~ eta1      0.4585      0.0805    5.6948    0.0000 [ 0.3048; 0.5898 ] 
#>   eta3 ~ eta2      0.3052      0.0776    3.9319    0.0001 [ 0.1377; 0.4488 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0415   15.9822    0.0000 [ 0.5992; 0.7350 ] 
#>   eta1 =~ y12      0.6493      0.0341   19.0394    0.0000 [ 0.5784; 0.6981 ] 
#>   eta1 =~ y13      0.7613      0.0284   26.7981    0.0000 [ 0.7231; 0.8224 ] 
#>   eta2 =~ y21      0.5165      0.0572    9.0331    0.0000 [ 0.3973; 0.5932 ] 
#>   eta2 =~ y22      0.7554      0.0362   20.8908    0.0000 [ 0.6866; 0.8301 ] 
#>   eta2 =~ y23      0.7997      0.0389   20.5800    0.0000 [ 0.7398; 0.8701 ] 
#>   eta3 =~ y31      0.8223      0.0314   26.2148    0.0000 [ 0.7791; 0.8825 ] 
#>   eta3 =~ y32      0.6581      0.0346   19.0260    0.0000 [ 0.5907; 0.7111 ] 
#>   eta3 =~ y33      0.7474      0.0377   19.8009    0.0000 [ 0.6771; 0.8054 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0211   18.7816    0.0000 [ 0.3598; 0.4283 ] 
#>   eta1 <~ y12      0.3873      0.0201   19.2498    0.0000 [ 0.3447; 0.4163 ] 
#>   eta1 <~ y13      0.4542      0.0153   29.7659    0.0000 [ 0.4348; 0.4889 ] 
#>   eta2 <~ y21      0.3058      0.0286   10.7002    0.0000 [ 0.2464; 0.3437 ] 
#>   eta2 <~ y22      0.4473      0.0228   19.5772    0.0000 [ 0.4253; 0.5026 ] 
#>   eta2 <~ y23      0.4735      0.0236   20.0617    0.0000 [ 0.4451; 0.5180 ] 
#>   eta3 <~ y31      0.4400      0.0182   24.1363    0.0000 [ 0.4115; 0.4755 ] 
#>   eta3 <~ y32      0.3521      0.0155   22.6826    0.0000 [ 0.3225; 0.3817 ] 
#>   eta3 <~ y33      0.3999      0.0174   23.0135    0.0000 [ 0.3661; 0.4314 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0460   14.5961    0.0000 [ 0.5914; 0.7419 ] 
#>   eta3 ~ eta1       0.6634      0.0492   13.4863    0.0000 [ 0.5624; 0.7181 ] 
#>   eta3 ~ eta2       0.3052      0.0776    3.9319    0.0001 [ 0.1377; 0.4488 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0507    4.0407    0.0001 [ 0.0904; 0.2795 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04148811 15.982168  1.701254e-57
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03410173 19.039442  8.037948e-81
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02841042 26.798119 3.398505e-158
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05717387  9.033057  1.669412e-19
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03615886 20.890807  6.491374e-97
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03885640 20.579973  4.149126e-94
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03136690 26.214808 1.801634e-151
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03458784 19.026019  1.038495e-80
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03774701 19.800883  2.925212e-87
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5991725          0.7350277
#> 2          0.5783714          0.6980917
#> 3          0.7231107          0.8224246
#> 4          0.3973068          0.5931990
#> 5          0.6866297          0.8300572
#> 6          0.7398153          0.8700509
#> 7          0.7791064          0.8824844
#> 8          0.5907035          0.7111356
#> 9          0.6770539          0.8054092

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2039653835
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0460   14.5961    0.0000 [ 0.5487; 0.7865 ] 
#>   eta3 ~ eta1      0.4585      0.0805    5.6948    0.0000 [ 0.2366; 0.6529 ] 
#>   eta3 ~ eta2      0.3052      0.0776    3.9319    0.0001 [ 0.1214; 0.5227 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0415   15.9822    0.0000 [ 0.5510; 0.7656 ] 
#>   eta1 =~ y12      0.6493      0.0341   19.0394    0.0000 [ 0.5653; 0.7417 ] 
#>   eta1 =~ y13      0.7613      0.0284   26.7981    0.0000 [ 0.6857; 0.8326 ] 
#>   eta2 =~ y21      0.5165      0.0572    9.0331    0.0000 [ 0.3918; 0.6874 ] 
#>   eta2 =~ y22      0.7554      0.0362   20.8908    0.0000 [ 0.6598; 0.8468 ] 
#>   eta2 =~ y23      0.7997      0.0389   20.5800    0.0000 [ 0.6905; 0.8914 ] 
#>   eta3 =~ y31      0.8223      0.0314   26.2148    0.0000 [ 0.7351; 0.8973 ] 
#>   eta3 =~ y32      0.6581      0.0346   19.0260    0.0000 [ 0.5696; 0.7485 ] 
#>   eta3 =~ y33      0.7474      0.0377   19.8009    0.0000 [ 0.6569; 0.8521 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0211   18.7816    0.0000 [ 0.3394; 0.4483 ] 
#>   eta1 <~ y12      0.3873      0.0201   19.2498    0.0000 [ 0.3387; 0.4428 ] 
#>   eta1 <~ y13      0.4542      0.0153   29.7659    0.0000 [ 0.4146; 0.4935 ] 
#>   eta2 <~ y21      0.3058      0.0286   10.7002    0.0000 [ 0.2447; 0.3925 ] 
#>   eta2 <~ y22      0.4473      0.0228   19.5772    0.0000 [ 0.3846; 0.5027 ] 
#>   eta2 <~ y23      0.4735      0.0236   20.0617    0.0000 [ 0.4048; 0.5269 ] 
#>   eta3 <~ y31      0.4400      0.0182   24.1363    0.0000 [ 0.3891; 0.4833 ] 
#>   eta3 <~ y32      0.3521      0.0155   22.6826    0.0000 [ 0.3122; 0.3925 ] 
#>   eta3 <~ y33      0.3999      0.0174   23.0135    0.0000 [ 0.3585; 0.4484 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0460   14.5961    0.0000 [ 0.5487; 0.7865 ] 
#>   eta3 ~ eta1       0.6634      0.0492   13.4863    0.0000 [ 0.5334; 0.7877 ] 
#>   eta3 ~ eta2       0.3052      0.0776    3.9319    0.0001 [ 0.1214; 0.5227 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0507    4.0407    0.0001 [ 0.0847; 0.3469 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04599416 14.596059 2.975560e-48
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08051326  5.694799 1.235177e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07760936  3.931885 8.428226e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5486678          0.7865236          0.5772296          0.7579618
#> 2          0.2365646          0.6529337          0.2865624          0.6029359
#> 3          0.1213633          0.5227151          0.1695578          0.4745206
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5908162          0.7782014          0.5913550          0.7418610
#> 2          0.2352779          0.6329111          0.3047579          0.5897558
#> 3          0.1186011          0.4557818          0.1376822          0.4488028