Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 2113075901
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0378   17.7708    0.0000 [ 0.6044; 0.7335 ] 
#>   eta3 ~ eta1      0.4585      0.0847    5.4139    0.0000 [ 0.3429; 0.6435 ] 
#>   eta3 ~ eta2      0.3052      0.0864    3.5324    0.0004 [ 0.1182; 0.3980 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0407   16.3104    0.0000 [ 0.5947; 0.7539 ] 
#>   eta1 =~ y12      0.6493      0.0333   19.4850    0.0000 [ 0.6042; 0.7101 ] 
#>   eta1 =~ y13      0.7613      0.0331   23.0310    0.0000 [ 0.7110; 0.8229 ] 
#>   eta2 =~ y21      0.5165      0.0516   10.0124    0.0000 [ 0.4093; 0.5921 ] 
#>   eta2 =~ y22      0.7554      0.0304   24.8860    0.0000 [ 0.7038; 0.8178 ] 
#>   eta2 =~ y23      0.7997      0.0399   20.0613    0.0000 [ 0.7105; 0.8525 ] 
#>   eta3 =~ y31      0.8223      0.0273   30.0668    0.0000 [ 0.7804; 0.8719 ] 
#>   eta3 =~ y32      0.6581      0.0395   16.6597    0.0000 [ 0.5960; 0.7269 ] 
#>   eta3 =~ y33      0.7474      0.0364   20.5324    0.0000 [ 0.6718; 0.8107 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0205   19.2765    0.0000 [ 0.3605; 0.4272 ] 
#>   eta1 <~ y12      0.3873      0.0194   19.9354    0.0000 [ 0.3442; 0.4194 ] 
#>   eta1 <~ y13      0.4542      0.0176   25.7853    0.0000 [ 0.4210; 0.4826 ] 
#>   eta2 <~ y21      0.3058      0.0303   10.0883    0.0000 [ 0.2545; 0.3507 ] 
#>   eta2 <~ y22      0.4473      0.0133   33.5725    0.0000 [ 0.4217; 0.4740 ] 
#>   eta2 <~ y23      0.4735      0.0226   20.9242    0.0000 [ 0.4432; 0.5144 ] 
#>   eta3 <~ y31      0.4400      0.0176   24.9447    0.0000 [ 0.4127; 0.4797 ] 
#>   eta3 <~ y32      0.3521      0.0174   20.1789    0.0000 [ 0.3193; 0.3896 ] 
#>   eta3 <~ y33      0.3999      0.0166   24.0482    0.0000 [ 0.3666; 0.4304 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0378   17.7708    0.0000 [ 0.6044; 0.7335 ] 
#>   eta3 ~ eta1       0.6634      0.0388   17.0856    0.0000 [ 0.6043; 0.7412 ] 
#>   eta3 ~ eta2       0.3052      0.0864    3.5324    0.0004 [ 0.1182; 0.3980 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0592    3.4580    0.0005 [ 0.0776; 0.2761 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04065328 16.31037  8.329239e-60
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03332190 19.48502  1.471113e-84
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03305740 23.03102 2.279516e-117
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05158128 10.01245  1.343876e-23
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03035396 24.88597 1.055619e-136
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03986106 20.06127  1.609111e-89
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02734832 30.06683 1.315574e-198
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03950072 16.65967  2.574510e-62
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03640215 20.53242  1.105359e-93
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5947219          0.7538509
#> 2          0.6042495          0.7100657
#> 3          0.7110301          0.8228850
#> 4          0.4093194          0.5920847
#> 5          0.7037831          0.8177748
#> 6          0.7105029          0.8525113
#> 7          0.7804355          0.8719139
#> 8          0.5959721          0.7269053
#> 9          0.6718176          0.8107024

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 2113075901
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0378   17.7708    0.0000 [ 0.5736; 0.7689 ] 
#>   eta3 ~ eta1      0.4585      0.0847    5.4139    0.0000 [ 0.2234; 0.6613 ] 
#>   eta3 ~ eta2      0.3052      0.0864    3.5324    0.0004 [ 0.0961; 0.5429 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0407   16.3104    0.0000 [ 0.5580; 0.7682 ] 
#>   eta1 =~ y12      0.6493      0.0333   19.4850    0.0000 [ 0.5570; 0.7293 ] 
#>   eta1 =~ y13      0.7613      0.0331   23.0310    0.0000 [ 0.6785; 0.8495 ] 
#>   eta2 =~ y21      0.5165      0.0516   10.0124    0.0000 [ 0.3852; 0.6520 ] 
#>   eta2 =~ y22      0.7554      0.0304   24.8860    0.0000 [ 0.6776; 0.8346 ] 
#>   eta2 =~ y23      0.7997      0.0399   20.0613    0.0000 [ 0.7015; 0.9077 ] 
#>   eta3 =~ y31      0.8223      0.0273   30.0668    0.0000 [ 0.7435; 0.8849 ] 
#>   eta3 =~ y32      0.6581      0.0395   16.6597    0.0000 [ 0.5514; 0.7557 ] 
#>   eta3 =~ y33      0.7474      0.0364   20.5324    0.0000 [ 0.6559; 0.8442 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0205   19.2765    0.0000 [ 0.3437; 0.4498 ] 
#>   eta1 <~ y12      0.3873      0.0194   19.9354    0.0000 [ 0.3344; 0.4349 ] 
#>   eta1 <~ y13      0.4542      0.0176   25.7853    0.0000 [ 0.4115; 0.5026 ] 
#>   eta2 <~ y21      0.3058      0.0303   10.0883    0.0000 [ 0.2273; 0.3841 ] 
#>   eta2 <~ y22      0.4473      0.0133   33.5725    0.0000 [ 0.4113; 0.4802 ] 
#>   eta2 <~ y23      0.4735      0.0226   20.9242    0.0000 [ 0.4158; 0.5328 ] 
#>   eta3 <~ y31      0.4400      0.0176   24.9447    0.0000 [ 0.3931; 0.4843 ] 
#>   eta3 <~ y32      0.3521      0.0174   20.1789    0.0000 [ 0.3073; 0.3975 ] 
#>   eta3 <~ y33      0.3999      0.0166   24.0482    0.0000 [ 0.3613; 0.4473 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0378   17.7708    0.0000 [ 0.5736; 0.7689 ] 
#>   eta3 ~ eta1       0.6634      0.0388   17.0856    0.0000 [ 0.5563; 0.7571 ] 
#>   eta3 ~ eta2       0.3052      0.0864    3.5324    0.0004 [ 0.0961; 0.5429 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0592    3.4580    0.0005 [ 0.0612; 0.3675 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03777742 17.770757 1.190733e-70
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08469111  5.413871 6.167653e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08638697  3.532374 4.118460e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.57358629          0.7689498          0.5970456          0.7454905
#> 2         0.22335685          0.6613315          0.2759490          0.6087393
#> 3         0.09611072          0.5428554          0.1497560          0.4892101
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.59501580          0.7470123          0.6044321          0.7334946
#> 2         0.33189759          0.6963230          0.3429162          0.6435343
#> 3         0.08331859          0.4560665          0.1182285          0.3979927