Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 2096234684
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0438   15.3138    0.0000 [ 0.5961; 0.7536 ] 
#>   eta3 ~ eta1      0.4585      0.0678    6.7592    0.0000 [ 0.3736; 0.5889 ] 
#>   eta3 ~ eta2      0.3052      0.0731    4.1716    0.0000 [ 0.1420; 0.4124 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0425   15.5840    0.0000 [ 0.5809; 0.7401 ] 
#>   eta1 =~ y12      0.6493      0.0326   19.9443    0.0000 [ 0.6054; 0.7273 ] 
#>   eta1 =~ y13      0.7613      0.0348   21.8881    0.0000 [ 0.6979; 0.8082 ] 
#>   eta2 =~ y21      0.5165      0.0468   11.0238    0.0000 [ 0.4104; 0.5887 ] 
#>   eta2 =~ y22      0.7554      0.0322   23.4883    0.0000 [ 0.7004; 0.8175 ] 
#>   eta2 =~ y23      0.7997      0.0359   22.2847    0.0000 [ 0.7451; 0.8740 ] 
#>   eta3 =~ y31      0.8223      0.0335   24.5592    0.0000 [ 0.7641; 0.8797 ] 
#>   eta3 =~ y32      0.6581      0.0397   16.5882    0.0000 [ 0.6089; 0.7345 ] 
#>   eta3 =~ y33      0.7474      0.0439   17.0122    0.0000 [ 0.6582; 0.8348 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0227   17.4459    0.0000 [ 0.3491; 0.4331 ] 
#>   eta1 <~ y12      0.3873      0.0194   19.9670    0.0000 [ 0.3541; 0.4258 ] 
#>   eta1 <~ y13      0.4542      0.0188   24.1654    0.0000 [ 0.4164; 0.4836 ] 
#>   eta2 <~ y21      0.3058      0.0249   12.2968    0.0000 [ 0.2577; 0.3385 ] 
#>   eta2 <~ y22      0.4473      0.0202   22.1817    0.0000 [ 0.4089; 0.4941 ] 
#>   eta2 <~ y23      0.4735      0.0166   28.5770    0.0000 [ 0.4495; 0.5086 ] 
#>   eta3 <~ y31      0.4400      0.0215   20.4563    0.0000 [ 0.3974; 0.4767 ] 
#>   eta3 <~ y32      0.3521      0.0191   18.4238    0.0000 [ 0.3222; 0.3903 ] 
#>   eta3 <~ y33      0.3999      0.0193   20.7273    0.0000 [ 0.3722; 0.4345 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0438   15.3138    0.0000 [ 0.5961; 0.7536 ] 
#>   eta3 ~ eta1       0.6634      0.0360   18.4379    0.0000 [ 0.6059; 0.7336 ] 
#>   eta3 ~ eta2       0.3052      0.0731    4.1716    0.0000 [ 0.1420; 0.4124 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0477    4.2981    0.0000 [ 0.0935; 0.2701 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04254825 15.58395  9.358496e-55
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03255463 19.94426  1.681333e-88
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03478348 21.88814 3.369793e-106
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04684896 11.02383  2.933185e-28
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03216012 23.48833 5.367523e-122
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03588396 22.28471 5.198811e-110
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03348147 24.55917 3.451369e-133
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03967082 16.58823  8.477751e-62
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04393461 17.01219  6.669289e-65
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5808510          0.7401136
#> 2          0.6054335          0.7272643
#> 3          0.6978579          0.8082327
#> 4          0.4103696          0.5886970
#> 5          0.7003979          0.8174612
#> 6          0.7450550          0.8739778
#> 7          0.7640789          0.8797057
#> 8          0.6089498          0.7344905
#> 9          0.6582274          0.8348255

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 2096234684
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0438   15.3138    0.0000 [ 0.5596; 0.7863 ] 
#>   eta3 ~ eta1      0.4585      0.0678    6.7592    0.0000 [ 0.2668; 0.6176 ] 
#>   eta3 ~ eta2      0.3052      0.0731    4.1716    0.0000 [ 0.1362; 0.5145 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0425   15.5840    0.0000 [ 0.5494; 0.7695 ] 
#>   eta1 =~ y12      0.6493      0.0326   19.9443    0.0000 [ 0.5626; 0.7309 ] 
#>   eta1 =~ y13      0.7613      0.0348   21.8881    0.0000 [ 0.6808; 0.8607 ] 
#>   eta2 =~ y21      0.5165      0.0468   11.0238    0.0000 [ 0.4119; 0.6542 ] 
#>   eta2 =~ y22      0.7554      0.0322   23.4883    0.0000 [ 0.6643; 0.8306 ] 
#>   eta2 =~ y23      0.7997      0.0359   22.2847    0.0000 [ 0.7013; 0.8868 ] 
#>   eta3 =~ y31      0.8223      0.0335   24.5592    0.0000 [ 0.7386; 0.9117 ] 
#>   eta3 =~ y32      0.6581      0.0397   16.5882    0.0000 [ 0.5529; 0.7581 ] 
#>   eta3 =~ y33      0.7474      0.0439   17.0122    0.0000 [ 0.6390; 0.8662 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0227   17.4459    0.0000 [ 0.3341; 0.4513 ] 
#>   eta1 <~ y12      0.3873      0.0194   19.9670    0.0000 [ 0.3348; 0.4351 ] 
#>   eta1 <~ y13      0.4542      0.0188   24.1654    0.0000 [ 0.4103; 0.5075 ] 
#>   eta2 <~ y21      0.3058      0.0249   12.2968    0.0000 [ 0.2517; 0.3803 ] 
#>   eta2 <~ y22      0.4473      0.0202   22.1817    0.0000 [ 0.3906; 0.4949 ] 
#>   eta2 <~ y23      0.4735      0.0166   28.5770    0.0000 [ 0.4278; 0.5135 ] 
#>   eta3 <~ y31      0.4400      0.0215   20.4563    0.0000 [ 0.3842; 0.4954 ] 
#>   eta3 <~ y32      0.3521      0.0191   18.4238    0.0000 [ 0.3002; 0.3990 ] 
#>   eta3 <~ y33      0.3999      0.0193   20.7273    0.0000 [ 0.3516; 0.4514 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0438   15.3138    0.0000 [ 0.5596; 0.7863 ] 
#>   eta3 ~ eta1       0.6634      0.0360   18.4379    0.0000 [ 0.5685; 0.7546 ] 
#>   eta3 ~ eta2       0.3052      0.0731    4.1716    0.0000 [ 0.1362; 0.5145 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0477    4.2981    0.0000 [ 0.0961; 0.3426 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04383841 15.313816 6.182736e-53
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.06783472  6.759175 1.387798e-11
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07314883  4.171647 3.024055e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5595778          0.7862853          0.5868009          0.7590622
#> 2          0.2668010          0.6176039          0.3089256          0.5754793
#> 3          0.1362072          0.5144917          0.1816318          0.4690671
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5686543          0.7559036          0.5960766          0.7536291
#> 2          0.2987210          0.6358716          0.3735785          0.5888970
#> 3          0.1382144          0.4537723          0.1420117          0.4124130