Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1558913949
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0420   15.9824    0.0000 [ 0.5996; 0.7591 ] 
#>   eta3 ~ eta1      0.4585      0.0714    6.4228    0.0000 [ 0.3904; 0.6302 ] 
#>   eta3 ~ eta2      0.3052      0.0766    3.9835    0.0001 [ 0.1349; 0.4129 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0378   17.5641    0.0000 [ 0.5739; 0.7088 ] 
#>   eta1 =~ y12      0.6493      0.0327   19.8386    0.0000 [ 0.5762; 0.6996 ] 
#>   eta1 =~ y13      0.7613      0.0390   19.5321    0.0000 [ 0.6732; 0.8073 ] 
#>   eta2 =~ y21      0.5165      0.0487   10.6155    0.0000 [ 0.4200; 0.5819 ] 
#>   eta2 =~ y22      0.7554      0.0361   20.8988    0.0000 [ 0.6940; 0.8224 ] 
#>   eta2 =~ y23      0.7997      0.0381   20.9749    0.0000 [ 0.7293; 0.8626 ] 
#>   eta3 =~ y31      0.8223      0.0356   23.0967    0.0000 [ 0.7612; 0.8804 ] 
#>   eta3 =~ y32      0.6581      0.0430   15.3209    0.0000 [ 0.5665; 0.7312 ] 
#>   eta3 =~ y33      0.7474      0.0415   18.0175    0.0000 [ 0.6376; 0.8059 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0218   18.1133    0.0000 [ 0.3578; 0.4399 ] 
#>   eta1 <~ y12      0.3873      0.0175   22.1880    0.0000 [ 0.3612; 0.4212 ] 
#>   eta1 <~ y13      0.4542      0.0196   23.1442    0.0000 [ 0.4171; 0.4986 ] 
#>   eta2 <~ y21      0.3058      0.0246   12.4220    0.0000 [ 0.2632; 0.3337 ] 
#>   eta2 <~ y22      0.4473      0.0194   23.0788    0.0000 [ 0.4249; 0.4930 ] 
#>   eta2 <~ y23      0.4735      0.0224   21.1358    0.0000 [ 0.4352; 0.5156 ] 
#>   eta3 <~ y31      0.4400      0.0214   20.6061    0.0000 [ 0.4084; 0.4673 ] 
#>   eta3 <~ y32      0.3521      0.0199   17.7379    0.0000 [ 0.3171; 0.3856 ] 
#>   eta3 <~ y33      0.3999      0.0206   19.4599    0.0000 [ 0.3539; 0.4366 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0420   15.9824    0.0000 [ 0.5996; 0.7591 ] 
#>   eta3 ~ eta1       0.6634      0.0346   19.1753    0.0000 [ 0.6225; 0.7551 ] 
#>   eta3 ~ eta2       0.3052      0.0766    3.9835    0.0001 [ 0.1349; 0.4129 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0518    3.9539    0.0001 [ 0.0970; 0.2884 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03775137 17.56413  4.637561e-69
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03272802 19.83860  1.382679e-87
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03897919 19.53211  5.857087e-85
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04865101 10.61550  2.524356e-26
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03614501 20.89881  5.489650e-97
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03812477 20.97491  1.111679e-97
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03560146 23.09673 4.993889e-118
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04295226 15.32094  5.541074e-53
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04148319 18.01752  1.419756e-72
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5739460          0.7088470
#> 2          0.5762082          0.6995656
#> 3          0.6732169          0.8072947
#> 4          0.4199506          0.5819238
#> 5          0.6940279          0.8224170
#> 6          0.7293376          0.8625929
#> 7          0.7611561          0.8804261
#> 8          0.5665019          0.7311584
#> 9          0.6375781          0.8058971

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1558913949
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0420   15.9824    0.0000 [ 0.5550; 0.7723 ] 
#>   eta3 ~ eta1      0.4585      0.0714    6.4228    0.0000 [ 0.2451; 0.6143 ] 
#>   eta3 ~ eta2      0.3052      0.0766    3.9835    0.0001 [ 0.1299; 0.5260 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0378   17.5641    0.0000 [ 0.5729; 0.7681 ] 
#>   eta1 =~ y12      0.6493      0.0327   19.8386    0.0000 [ 0.5682; 0.7374 ] 
#>   eta1 =~ y13      0.7613      0.0390   19.5321    0.0000 [ 0.6713; 0.8729 ] 
#>   eta2 =~ y21      0.5165      0.0487   10.6155    0.0000 [ 0.4090; 0.6606 ] 
#>   eta2 =~ y22      0.7554      0.0361   20.8988    0.0000 [ 0.6577; 0.8446 ] 
#>   eta2 =~ y23      0.7997      0.0381   20.9749    0.0000 [ 0.6994; 0.8966 ] 
#>   eta3 =~ y31      0.8223      0.0356   23.0967    0.0000 [ 0.7361; 0.9202 ] 
#>   eta3 =~ y32      0.6581      0.0430   15.3209    0.0000 [ 0.5424; 0.7646 ] 
#>   eta3 =~ y33      0.7474      0.0415   18.0175    0.0000 [ 0.6380; 0.8525 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0218   18.1133    0.0000 [ 0.3375; 0.4504 ] 
#>   eta1 <~ y12      0.3873      0.0175   22.1880    0.0000 [ 0.3384; 0.4286 ] 
#>   eta1 <~ y13      0.4542      0.0196   23.1442    0.0000 [ 0.4030; 0.5045 ] 
#>   eta2 <~ y21      0.3058      0.0246   12.4220    0.0000 [ 0.2515; 0.3788 ] 
#>   eta2 <~ y22      0.4473      0.0194   23.0788    0.0000 [ 0.3919; 0.4921 ] 
#>   eta2 <~ y23      0.4735      0.0224   21.1358    0.0000 [ 0.4116; 0.5274 ] 
#>   eta3 <~ y31      0.4400      0.0214   20.6061    0.0000 [ 0.3882; 0.4986 ] 
#>   eta3 <~ y32      0.3521      0.0199   17.7379    0.0000 [ 0.2988; 0.4015 ] 
#>   eta3 <~ y33      0.3999      0.0206   19.4599    0.0000 [ 0.3460; 0.4523 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0420   15.9824    0.0000 [ 0.5550; 0.7723 ] 
#>   eta3 ~ eta1       0.6634      0.0346   19.1753    0.0000 [ 0.5590; 0.7379 ] 
#>   eta3 ~ eta2       0.3052      0.0766    3.9835    0.0001 [ 0.1299; 0.5260 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0518    3.9539    0.0001 [ 0.0848; 0.3528 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04200459 15.982384 1.695364e-57
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07138726  6.422809 1.337819e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07660426  3.983475 6.791494e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5550337          0.7722577          0.5811180          0.7461733
#> 2          0.2450940          0.6142686          0.2894247          0.5699380
#> 3          0.1298563          0.5260103          0.1774266          0.4784399
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.57834491          0.7776994          0.5995641          0.7590821
#> 2         0.33216780          0.6649981          0.3903544          0.6301623
#> 3         0.08755759          0.4196083          0.1349204          0.4128628