Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -163318726
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0434   15.4572    0.0000 [ 0.5818; 0.7594 ] 
#>   eta3 ~ eta1      0.4585      0.0837    5.4762    0.0000 [ 0.2938; 0.6387 ] 
#>   eta3 ~ eta2      0.3052      0.0956    3.1931    0.0014 [ 0.1118; 0.4647 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0393   16.8649    0.0000 [ 0.6029; 0.7496 ] 
#>   eta1 =~ y12      0.6493      0.0435   14.9236    0.0000 [ 0.5496; 0.7512 ] 
#>   eta1 =~ y13      0.7613      0.0287   26.5016    0.0000 [ 0.7191; 0.8214 ] 
#>   eta2 =~ y21      0.5165      0.0548    9.4223    0.0000 [ 0.4007; 0.5867 ] 
#>   eta2 =~ y22      0.7554      0.0477   15.8267    0.0000 [ 0.6678; 0.8333 ] 
#>   eta2 =~ y23      0.7997      0.0396   20.2056    0.0000 [ 0.7402; 0.8684 ] 
#>   eta3 =~ y31      0.8223      0.0306   26.8916    0.0000 [ 0.7708; 0.8727 ] 
#>   eta3 =~ y32      0.6581      0.0428   15.3582    0.0000 [ 0.6007; 0.7653 ] 
#>   eta3 =~ y33      0.7474      0.0395   18.9041    0.0000 [ 0.6616; 0.8120 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0217   18.2565    0.0000 [ 0.3569; 0.4317 ] 
#>   eta1 <~ y12      0.3873      0.0208   18.6495    0.0000 [ 0.3499; 0.4286 ] 
#>   eta1 <~ y13      0.4542      0.0183   24.8569    0.0000 [ 0.4255; 0.4924 ] 
#>   eta2 <~ y21      0.3058      0.0315    9.6955    0.0000 [ 0.2524; 0.3548 ] 
#>   eta2 <~ y22      0.4473      0.0233   19.1917    0.0000 [ 0.4012; 0.4999 ] 
#>   eta2 <~ y23      0.4735      0.0256   18.4830    0.0000 [ 0.4391; 0.5226 ] 
#>   eta3 <~ y31      0.4400      0.0169   26.0971    0.0000 [ 0.4126; 0.4701 ] 
#>   eta3 <~ y32      0.3521      0.0191   18.4745    0.0000 [ 0.3252; 0.3933 ] 
#>   eta3 <~ y33      0.3999      0.0215   18.5715    0.0000 [ 0.3646; 0.4378 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0434   15.4572    0.0000 [ 0.5818; 0.7594 ] 
#>   eta3 ~ eta1       0.6634      0.0364   18.2231    0.0000 [ 0.5912; 0.7193 ] 
#>   eta3 ~ eta2       0.3052      0.0956    3.1931    0.0014 [ 0.1118; 0.4647 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0647    3.1642    0.0016 [ 0.0770; 0.3105 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03931667 16.864851  8.160702e-64
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04350673 14.923620  2.313721e-50
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02872833 26.501567 9.297510e-155
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05481193  9.422307  4.412825e-21
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04772861 15.826727  2.035379e-56
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03957626 20.205639  8.733248e-91
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03057750 26.891581 2.755211e-159
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04284804 15.358205  3.120987e-53
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03953777 18.904052  1.056216e-79
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6029276          0.7495732
#> 2          0.5496184          0.7512135
#> 3          0.7190604          0.8214393
#> 4          0.4006593          0.5866534
#> 5          0.6677790          0.8333303
#> 6          0.7402262          0.8683851
#> 7          0.7708171          0.8727062
#> 8          0.6007016          0.7652716
#> 9          0.6616082          0.8120235

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -163318726
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0434   15.4572    0.0000 [ 0.5545; 0.7791 ] 
#>   eta3 ~ eta1      0.4585      0.0837    5.4762    0.0000 [ 0.2306; 0.6636 ] 
#>   eta3 ~ eta2      0.3052      0.0956    3.1931    0.0014 [ 0.0723; 0.5665 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0393   16.8649    0.0000 [ 0.5624; 0.7657 ] 
#>   eta1 =~ y12      0.6493      0.0435   14.9236    0.0000 [ 0.5351; 0.7600 ] 
#>   eta1 =~ y13      0.7613      0.0287   26.5016    0.0000 [ 0.6796; 0.8282 ] 
#>   eta2 =~ y21      0.5165      0.0548    9.4223    0.0000 [ 0.3835; 0.6670 ] 
#>   eta2 =~ y22      0.7554      0.0477   15.8267    0.0000 [ 0.6318; 0.8786 ] 
#>   eta2 =~ y23      0.7997      0.0396   20.2056    0.0000 [ 0.6881; 0.8927 ] 
#>   eta3 =~ y31      0.8223      0.0306   26.8916    0.0000 [ 0.7456; 0.9037 ] 
#>   eta3 =~ y32      0.6581      0.0428   15.3582    0.0000 [ 0.5426; 0.7642 ] 
#>   eta3 =~ y33      0.7474      0.0395   18.9041    0.0000 [ 0.6496; 0.8540 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0217   18.2565    0.0000 [ 0.3427; 0.4547 ] 
#>   eta1 <~ y12      0.3873      0.0208   18.6495    0.0000 [ 0.3353; 0.4427 ] 
#>   eta1 <~ y13      0.4542      0.0183   24.8569    0.0000 [ 0.4053; 0.4998 ] 
#>   eta2 <~ y21      0.3058      0.0315    9.6955    0.0000 [ 0.2304; 0.3935 ] 
#>   eta2 <~ y22      0.4473      0.0233   19.1917    0.0000 [ 0.3884; 0.5090 ] 
#>   eta2 <~ y23      0.4735      0.0256   18.4830    0.0000 [ 0.4030; 0.5355 ] 
#>   eta3 <~ y31      0.4400      0.0169   26.0971    0.0000 [ 0.3971; 0.4843 ] 
#>   eta3 <~ y32      0.3521      0.0191   18.4745    0.0000 [ 0.3001; 0.3987 ] 
#>   eta3 <~ y33      0.3999      0.0215   18.5715    0.0000 [ 0.3461; 0.4575 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0434   15.4572    0.0000 [ 0.5545; 0.7791 ] 
#>   eta3 ~ eta1       0.6634      0.0364   18.2231    0.0000 [ 0.5667; 0.7550 ] 
#>   eta3 ~ eta2       0.3052      0.0956    3.1931    0.0014 [ 0.0723; 0.5665 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0647    3.1642    0.0016 [ 0.0464; 0.3812 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04343184 15.457172 6.749788e-54
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08372730  5.476192 4.345761e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09556619  3.193087 1.407607e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55454247          0.7791474          0.5815131          0.7521768
#> 2         0.23056791          0.6635582          0.2825616          0.6115646
#> 3         0.07230399          0.5665183          0.1316495          0.5071729
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.57369914          0.7676834          0.5817787          0.7593661
#> 2         0.27513373          0.6407962          0.2937731          0.6387171
#> 3         0.07537043          0.4933273          0.1117724          0.4647211