Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 209800595
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0530   12.6618    0.0000 [ 0.5775; 0.7651 ] 
#>   eta3 ~ eta1      0.4585      0.0665    6.8961    0.0000 [ 0.3327; 0.5649 ] 
#>   eta3 ~ eta2      0.3052      0.0824    3.7019    0.0002 [ 0.1345; 0.4230 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0410   16.1591    0.0000 [ 0.5742; 0.7453 ] 
#>   eta1 =~ y12      0.6493      0.0403   16.0980    0.0000 [ 0.5909; 0.7473 ] 
#>   eta1 =~ y13      0.7613      0.0277   27.5097    0.0000 [ 0.7109; 0.8120 ] 
#>   eta2 =~ y21      0.5165      0.0587    8.7994    0.0000 [ 0.4232; 0.6142 ] 
#>   eta2 =~ y22      0.7554      0.0437   17.2983    0.0000 [ 0.6894; 0.8235 ] 
#>   eta2 =~ y23      0.7997      0.0360   22.2398    0.0000 [ 0.7319; 0.8616 ] 
#>   eta3 =~ y31      0.8223      0.0360   22.8598    0.0000 [ 0.7465; 0.8779 ] 
#>   eta3 =~ y32      0.6581      0.0407   16.1550    0.0000 [ 0.5901; 0.7334 ] 
#>   eta3 =~ y33      0.7474      0.0441   16.9525    0.0000 [ 0.6613; 0.8134 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0234   16.9057    0.0000 [ 0.3480; 0.4390 ] 
#>   eta1 <~ y12      0.3873      0.0192   20.1897    0.0000 [ 0.3595; 0.4305 ] 
#>   eta1 <~ y13      0.4542      0.0169   26.9523    0.0000 [ 0.4255; 0.4861 ] 
#>   eta2 <~ y21      0.3058      0.0324    9.4297    0.0000 [ 0.2445; 0.3623 ] 
#>   eta2 <~ y22      0.4473      0.0219   20.3984    0.0000 [ 0.4141; 0.4881 ] 
#>   eta2 <~ y23      0.4735      0.0238   19.8936    0.0000 [ 0.4356; 0.5150 ] 
#>   eta3 <~ y31      0.4400      0.0204   21.6105    0.0000 [ 0.4063; 0.4693 ] 
#>   eta3 <~ y32      0.3521      0.0194   18.1051    0.0000 [ 0.3201; 0.3845 ] 
#>   eta3 <~ y33      0.3999      0.0197   20.3128    0.0000 [ 0.3606; 0.4276 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0530   12.6618    0.0000 [ 0.5775; 0.7651 ] 
#>   eta3 ~ eta1       0.6634      0.0362   18.3119    0.0000 [ 0.6059; 0.7410 ] 
#>   eta3 ~ eta2       0.3052      0.0824    3.7019    0.0002 [ 0.1345; 0.4230 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0522    3.9269    0.0001 [ 0.0938; 0.3023 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04103372 16.15915  9.790724e-59
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04033275 16.09803  2.633463e-58
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02767553 27.50971 1.343700e-166
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05869231  8.79936  1.375983e-18
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04366830 17.29831  4.844120e-67
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03595637 22.23984 1.414707e-109
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03597038 22.85985 1.166425e-115
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04073469 16.15500  1.047208e-58
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04408926 16.95252  1.843754e-64
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5741978          0.7452571
#> 2          0.5908858          0.7473112
#> 3          0.7109066          0.8119844
#> 4          0.4231562          0.6142290
#> 5          0.6893675          0.8234546
#> 6          0.7319459          0.8615735
#> 7          0.7465371          0.8779081
#> 8          0.5901449          0.7333782
#> 9          0.6612823          0.8134440

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 209800595
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0530   12.6618    0.0000 [ 0.5341; 0.8083 ] 
#>   eta3 ~ eta1      0.4585      0.0665    6.8961    0.0000 [ 0.2965; 0.6404 ] 
#>   eta3 ~ eta2      0.3052      0.0824    3.7019    0.0002 [ 0.0823; 0.5085 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0410   16.1591    0.0000 [ 0.5553; 0.7675 ] 
#>   eta1 =~ y12      0.6493      0.0403   16.0980    0.0000 [ 0.5373; 0.7459 ] 
#>   eta1 =~ y13      0.7613      0.0277   27.5097    0.0000 [ 0.6893; 0.8324 ] 
#>   eta2 =~ y21      0.5165      0.0587    8.7994    0.0000 [ 0.3734; 0.6769 ] 
#>   eta2 =~ y22      0.7554      0.0437   17.2983    0.0000 [ 0.6446; 0.8704 ] 
#>   eta2 =~ y23      0.7997      0.0360   22.2398    0.0000 [ 0.7090; 0.8950 ] 
#>   eta3 =~ y31      0.8223      0.0360   22.8598    0.0000 [ 0.7283; 0.9143 ] 
#>   eta3 =~ y32      0.6581      0.0407   16.1550    0.0000 [ 0.5556; 0.7662 ] 
#>   eta3 =~ y33      0.7474      0.0441   16.9525    0.0000 [ 0.6394; 0.8674 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0234   16.9057    0.0000 [ 0.3370; 0.4580 ] 
#>   eta1 <~ y12      0.3873      0.0192   20.1897    0.0000 [ 0.3362; 0.4354 ] 
#>   eta1 <~ y13      0.4542      0.0169   26.9523    0.0000 [ 0.4136; 0.5007 ] 
#>   eta2 <~ y21      0.3058      0.0324    9.4297    0.0000 [ 0.2251; 0.3928 ] 
#>   eta2 <~ y22      0.4473      0.0219   20.3984    0.0000 [ 0.3887; 0.5021 ] 
#>   eta2 <~ y23      0.4735      0.0238   19.8936    0.0000 [ 0.4096; 0.5327 ] 
#>   eta3 <~ y31      0.4400      0.0204   21.6105    0.0000 [ 0.3845; 0.4898 ] 
#>   eta3 <~ y32      0.3521      0.0194   18.1051    0.0000 [ 0.3017; 0.4023 ] 
#>   eta3 <~ y33      0.3999      0.0197   20.3128    0.0000 [ 0.3504; 0.4523 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0530   12.6618    0.0000 [ 0.5341; 0.8083 ] 
#>   eta3 ~ eta1       0.6634      0.0362   18.3119    0.0000 [ 0.5748; 0.7621 ] 
#>   eta3 ~ eta2       0.3052      0.0824    3.7019    0.0002 [ 0.0823; 0.5085 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0522    3.9269    0.0001 [ 0.0651; 0.3349 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05302041 12.661793 9.627543e-37
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.06648810  6.896073 5.345975e-12
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08243106  3.701895 2.139951e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.53406796          0.8082596          0.5669930          0.7753345
#> 2         0.29653701          0.6403759          0.3378253          0.5990876
#> 3         0.08226104          0.5085479          0.1334498          0.4573592
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5686911          0.7668703          0.5774719          0.7651220
#> 2          0.3249256          0.5839943          0.3327390          0.5648629
#> 3          0.1269329          0.4514842          0.1344700          0.4230007