Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2065158711
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0512   13.1224    0.0000 [ 0.5834; 0.7637 ] 
#>   eta3 ~ eta1      0.4585      0.0874    5.2439    0.0000 [ 0.2967; 0.6115 ] 
#>   eta3 ~ eta2      0.3052      0.0937    3.2558    0.0011 [ 0.1866; 0.4661 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0350   18.9273    0.0000 [ 0.6030; 0.7282 ] 
#>   eta1 =~ y12      0.6493      0.0265   24.5243    0.0000 [ 0.6016; 0.6915 ] 
#>   eta1 =~ y13      0.7613      0.0383   19.8976    0.0000 [ 0.6914; 0.8144 ] 
#>   eta2 =~ y21      0.5165      0.0511   10.1136    0.0000 [ 0.4152; 0.5915 ] 
#>   eta2 =~ y22      0.7554      0.0361   20.9266    0.0000 [ 0.7034; 0.8314 ] 
#>   eta2 =~ y23      0.7997      0.0347   23.0212    0.0000 [ 0.7423; 0.8582 ] 
#>   eta3 =~ y31      0.8223      0.0311   26.4751    0.0000 [ 0.7733; 0.8747 ] 
#>   eta3 =~ y32      0.6581      0.0468   14.0588    0.0000 [ 0.5878; 0.7335 ] 
#>   eta3 =~ y33      0.7474      0.0300   24.8935    0.0000 [ 0.6983; 0.7928 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0205   19.2647    0.0000 [ 0.3629; 0.4320 ] 
#>   eta1 <~ y12      0.3873      0.0137   28.3181    0.0000 [ 0.3669; 0.4205 ] 
#>   eta1 <~ y13      0.4542      0.0200   22.7052    0.0000 [ 0.4231; 0.4868 ] 
#>   eta2 <~ y21      0.3058      0.0281   10.8801    0.0000 [ 0.2476; 0.3377 ] 
#>   eta2 <~ y22      0.4473      0.0195   22.9597    0.0000 [ 0.4119; 0.4799 ] 
#>   eta2 <~ y23      0.4735      0.0174   27.2731    0.0000 [ 0.4382; 0.4978 ] 
#>   eta3 <~ y31      0.4400      0.0186   23.6430    0.0000 [ 0.4045; 0.4710 ] 
#>   eta3 <~ y32      0.3521      0.0181   19.5049    0.0000 [ 0.3216; 0.3826 ] 
#>   eta3 <~ y33      0.3999      0.0193   20.7708    0.0000 [ 0.3724; 0.4303 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0512   13.1224    0.0000 [ 0.5834; 0.7637 ] 
#>   eta3 ~ eta1       0.6634      0.0390   17.0270    0.0000 [ 0.5924; 0.7327 ] 
#>   eta3 ~ eta2       0.3052      0.0937    3.2558    0.0011 [ 0.1866; 0.4661 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0640    3.2007    0.0014 [ 0.1302; 0.3240 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03503241 18.92733  6.791985e-80
#> 2 eta1 =~ y12  Common factor 0.6492779 0.02647483 24.52434 8.125393e-133
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03826316 19.89762  4.267033e-88
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05106520 10.11364  4.806648e-24
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03609693 20.92665  3.063116e-97
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03473590 23.02124 2.856280e-117
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03105850 26.47512 1.875359e-154
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04680839 14.05878  6.804716e-45
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03002486 24.89351 8.747621e-137
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6029725          0.7282437
#> 2          0.6016101          0.6915091
#> 3          0.6913766          0.8144376
#> 4          0.4152374          0.5915309
#> 5          0.7034082          0.8313840
#> 6          0.7422622          0.8582128
#> 7          0.7733309          0.8746902
#> 8          0.5878452          0.7334586
#> 9          0.6982756          0.7927912

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2065158711
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0512   13.1224    0.0000 [ 0.5445; 0.8091 ] 
#>   eta3 ~ eta1      0.4585      0.0874    5.2439    0.0000 [ 0.2493; 0.7015 ] 
#>   eta3 ~ eta2      0.3052      0.0937    3.2558    0.0011 [ 0.0443; 0.5289 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0350   18.9273    0.0000 [ 0.5731; 0.7543 ] 
#>   eta1 =~ y12      0.6493      0.0265   24.5243    0.0000 [ 0.5816; 0.7185 ] 
#>   eta1 =~ y13      0.7613      0.0383   19.8976    0.0000 [ 0.6638; 0.8617 ] 
#>   eta2 =~ y21      0.5165      0.0511   10.1136    0.0000 [ 0.3888; 0.6528 ] 
#>   eta2 =~ y22      0.7554      0.0361   20.9266    0.0000 [ 0.6519; 0.8386 ] 
#>   eta2 =~ y23      0.7997      0.0347   23.0212    0.0000 [ 0.7092; 0.8889 ] 
#>   eta3 =~ y31      0.8223      0.0311   26.4751    0.0000 [ 0.7443; 0.9050 ] 
#>   eta3 =~ y32      0.6581      0.0468   14.0588    0.0000 [ 0.5353; 0.7773 ] 
#>   eta3 =~ y33      0.7474      0.0300   24.8935    0.0000 [ 0.6800; 0.8353 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0205   19.2647    0.0000 [ 0.3421; 0.4483 ] 
#>   eta1 <~ y12      0.3873      0.0137   28.3181    0.0000 [ 0.3518; 0.4225 ] 
#>   eta1 <~ y13      0.4542      0.0200   22.7052    0.0000 [ 0.4026; 0.5060 ] 
#>   eta2 <~ y21      0.3058      0.0281   10.8801    0.0000 [ 0.2377; 0.3830 ] 
#>   eta2 <~ y22      0.4473      0.0195   22.9597    0.0000 [ 0.3937; 0.4945 ] 
#>   eta2 <~ y23      0.4735      0.0174   27.2731    0.0000 [ 0.4312; 0.5210 ] 
#>   eta3 <~ y31      0.4400      0.0186   23.6430    0.0000 [ 0.3899; 0.4861 ] 
#>   eta3 <~ y32      0.3521      0.0181   19.5049    0.0000 [ 0.3024; 0.3957 ] 
#>   eta3 <~ y33      0.3999      0.0193   20.7708    0.0000 [ 0.3526; 0.4522 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0512   13.1224    0.0000 [ 0.5445; 0.8091 ] 
#>   eta3 ~ eta1       0.6634      0.0390   17.0270    0.0000 [ 0.5697; 0.7711 ] 
#>   eta3 ~ eta2       0.3052      0.0937    3.2558    0.0011 [ 0.0443; 0.5289 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0640    3.2007    0.0014 [ 0.0295; 0.3605 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05115921 13.122435 2.449370e-39
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08743542  5.243948 1.571766e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09372477  3.255822 1.130648e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54449267          0.8090592          0.5762619          0.7772900
#> 2         0.24928495          0.7014516          0.3035813          0.6471552
#> 3         0.04425305          0.5289446          0.1024550          0.4707426
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5777321          0.7659364          0.5833914          0.7636636
#> 2          0.2487075          0.6285068          0.2967108          0.6114983
#> 3          0.0812571          0.4919814          0.1865621          0.4660937