Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1396638097
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0463   14.4928    0.0000 [ 0.5828; 0.7400 ] 
#>   eta3 ~ eta1      0.4585      0.0856    5.3590    0.0000 [ 0.2897; 0.6168 ] 
#>   eta3 ~ eta2      0.3052      0.0895    3.4106    0.0006 [ 0.1372; 0.4506 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0494   13.4111    0.0000 [ 0.5484; 0.7319 ] 
#>   eta1 =~ y12      0.6493      0.0432   15.0353    0.0000 [ 0.5790; 0.7342 ] 
#>   eta1 =~ y13      0.7613      0.0333   22.8350    0.0000 [ 0.7146; 0.8434 ] 
#>   eta2 =~ y21      0.5165      0.0589    8.7752    0.0000 [ 0.4015; 0.6337 ] 
#>   eta2 =~ y22      0.7554      0.0321   23.4976    0.0000 [ 0.6972; 0.8050 ] 
#>   eta2 =~ y23      0.7997      0.0327   24.4799    0.0000 [ 0.7278; 0.8392 ] 
#>   eta3 =~ y31      0.8223      0.0347   23.7295    0.0000 [ 0.7730; 0.9079 ] 
#>   eta3 =~ y32      0.6581      0.0368   17.8708    0.0000 [ 0.5916; 0.7186 ] 
#>   eta3 =~ y33      0.7474      0.0368   20.3253    0.0000 [ 0.7007; 0.8207 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0265   14.9077    0.0000 [ 0.3496; 0.4294 ] 
#>   eta1 <~ y12      0.3873      0.0214   18.0638    0.0000 [ 0.3413; 0.4259 ] 
#>   eta1 <~ y13      0.4542      0.0227   20.0055    0.0000 [ 0.4250; 0.5034 ] 
#>   eta2 <~ y21      0.3058      0.0285   10.7458    0.0000 [ 0.2445; 0.3533 ] 
#>   eta2 <~ y22      0.4473      0.0246   18.1576    0.0000 [ 0.4075; 0.4888 ] 
#>   eta2 <~ y23      0.4735      0.0177   26.8185    0.0000 [ 0.4460; 0.5069 ] 
#>   eta3 <~ y31      0.4400      0.0176   25.0410    0.0000 [ 0.4092; 0.4792 ] 
#>   eta3 <~ y32      0.3521      0.0168   20.9160    0.0000 [ 0.3140; 0.3727 ] 
#>   eta3 <~ y33      0.3999      0.0197   20.3126    0.0000 [ 0.3689; 0.4349 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0463   14.4928    0.0000 [ 0.5828; 0.7400 ] 
#>   eta3 ~ eta1       0.6634      0.0387   17.1384    0.0000 [ 0.5728; 0.7310 ] 
#>   eta3 ~ eta2       0.3052      0.0895    3.4106    0.0006 [ 0.1372; 0.4506 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0571    3.5862    0.0003 [ 0.0964; 0.2937 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04944201 13.411061  5.208808e-41
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04318360 15.035288  4.311681e-51
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03334118 22.835001 2.059890e-115
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05885393  8.775197  1.706056e-18
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03214737 23.497649 4.310840e-122
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03266613 24.479905 2.418268e-132
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03465208 23.729525 1.788010e-124
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03682369 17.870801  1.991142e-71
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03677305 20.325325  7.677945e-92
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5483647          0.7319250
#> 2          0.5789681          0.7342460
#> 3          0.7145966          0.8434452
#> 4          0.4014953          0.6336929
#> 5          0.6972198          0.8050354
#> 6          0.7278125          0.8391782
#> 7          0.7729882          0.9078905
#> 8          0.5915734          0.7186267
#> 9          0.7006648          0.8207027

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1396638097
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0463   14.4928    0.0000 [ 0.5574; 0.7969 ] 
#>   eta3 ~ eta1      0.4585      0.0856    5.3590    0.0000 [ 0.2344; 0.6768 ] 
#>   eta3 ~ eta2      0.3052      0.0895    3.4106    0.0006 [ 0.0750; 0.5377 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0494   13.4111    0.0000 [ 0.5374; 0.7931 ] 
#>   eta1 =~ y12      0.6493      0.0432   15.0353    0.0000 [ 0.5350; 0.7583 ] 
#>   eta1 =~ y13      0.7613      0.0333   22.8350    0.0000 [ 0.6662; 0.8387 ] 
#>   eta2 =~ y21      0.5165      0.0589    8.7752    0.0000 [ 0.3583; 0.6627 ] 
#>   eta2 =~ y22      0.7554      0.0321   23.4976    0.0000 [ 0.6801; 0.8464 ] 
#>   eta2 =~ y23      0.7997      0.0327   24.4799    0.0000 [ 0.7233; 0.8922 ] 
#>   eta3 =~ y31      0.8223      0.0347   23.7295    0.0000 [ 0.7323; 0.9115 ] 
#>   eta3 =~ y32      0.6581      0.0368   17.8708    0.0000 [ 0.5702; 0.7607 ] 
#>   eta3 =~ y33      0.7474      0.0368   20.3253    0.0000 [ 0.6426; 0.8328 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0265   14.9077    0.0000 [ 0.3314; 0.4686 ] 
#>   eta1 <~ y12      0.3873      0.0214   18.0638    0.0000 [ 0.3334; 0.4443 ] 
#>   eta1 <~ y13      0.4542      0.0227   20.0055    0.0000 [ 0.3934; 0.5108 ] 
#>   eta2 <~ y21      0.3058      0.0285   10.7458    0.0000 [ 0.2271; 0.3743 ] 
#>   eta2 <~ y22      0.4473      0.0246   18.1576    0.0000 [ 0.3847; 0.5121 ] 
#>   eta2 <~ y23      0.4735      0.0177   26.8185    0.0000 [ 0.4292; 0.5205 ] 
#>   eta3 <~ y31      0.4400      0.0176   25.0410    0.0000 [ 0.3954; 0.4863 ] 
#>   eta3 <~ y32      0.3521      0.0168   20.9160    0.0000 [ 0.3135; 0.4006 ] 
#>   eta3 <~ y33      0.3999      0.0197   20.3126    0.0000 [ 0.3448; 0.4466 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0463   14.4928    0.0000 [ 0.5574; 0.7969 ] 
#>   eta3 ~ eta1       0.6634      0.0387   17.1384    0.0000 [ 0.5639; 0.7641 ] 
#>   eta3 ~ eta2       0.3052      0.0895    3.4106    0.0006 [ 0.0750; 0.5377 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0571    3.5862    0.0003 [ 0.0607; 0.3561 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04632197 14.492766 1.346107e-47
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08555861  5.358979 8.369377e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08947224  3.410567 6.482785e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55739740          0.7969485          0.5861628          0.7681831
#> 2         0.23435074          0.6768116          0.2874816          0.6236807
#> 3         0.07495477          0.5376547          0.1305160          0.4820935
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5569998          0.7700913          0.5828290          0.7399553
#> 2          0.2814997          0.6302755          0.2897069          0.6167819
#> 3          0.1121845          0.4652642          0.1371799          0.4505996