Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -49670491
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0369   18.1905    0.0000 [ 0.5954; 0.7470 ] 
#>   eta3 ~ eta1      0.4585      0.0929    4.9360    0.0000 [ 0.2417; 0.5882 ] 
#>   eta3 ~ eta2      0.3052      0.0915    3.3342    0.0009 [ 0.1923; 0.5197 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0398   16.6530    0.0000 [ 0.5864; 0.7194 ] 
#>   eta1 =~ y12      0.6493      0.0354   18.3315    0.0000 [ 0.5751; 0.7035 ] 
#>   eta1 =~ y13      0.7613      0.0344   22.1149    0.0000 [ 0.6944; 0.8268 ] 
#>   eta2 =~ y21      0.5165      0.0409   12.6254    0.0000 [ 0.4609; 0.5887 ] 
#>   eta2 =~ y22      0.7554      0.0366   20.6527    0.0000 [ 0.6823; 0.8123 ] 
#>   eta2 =~ y23      0.7997      0.0344   23.2690    0.0000 [ 0.7375; 0.8455 ] 
#>   eta3 =~ y31      0.8223      0.0317   25.9131    0.0000 [ 0.7606; 0.8711 ] 
#>   eta3 =~ y32      0.6581      0.0384   17.1483    0.0000 [ 0.5662; 0.7184 ] 
#>   eta3 =~ y33      0.7474      0.0332   22.5264    0.0000 [ 0.6863; 0.8005 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0189   20.9305    0.0000 [ 0.3533; 0.4210 ] 
#>   eta1 <~ y12      0.3873      0.0205   18.9248    0.0000 [ 0.3522; 0.4262 ] 
#>   eta1 <~ y13      0.4542      0.0209   21.7396    0.0000 [ 0.4208; 0.4919 ] 
#>   eta2 <~ y21      0.3058      0.0242   12.6563    0.0000 [ 0.2646; 0.3547 ] 
#>   eta2 <~ y22      0.4473      0.0208   21.4773    0.0000 [ 0.4111; 0.4914 ] 
#>   eta2 <~ y23      0.4735      0.0147   32.2021    0.0000 [ 0.4505; 0.5020 ] 
#>   eta3 <~ y31      0.4400      0.0192   22.8850    0.0000 [ 0.4030; 0.4658 ] 
#>   eta3 <~ y32      0.3521      0.0181   19.4068    0.0000 [ 0.3101; 0.3834 ] 
#>   eta3 <~ y33      0.3999      0.0133   30.0538    0.0000 [ 0.3766; 0.4238 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0369   18.1905    0.0000 [ 0.5954; 0.7470 ] 
#>   eta3 ~ eta1       0.6634      0.0404   16.4197    0.0000 [ 0.5895; 0.7427 ] 
#>   eta3 ~ eta2       0.3052      0.0915    3.3342    0.0009 [ 0.1923; 0.5197 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0618    3.3131    0.0009 [ 0.1247; 0.3488 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03981695 16.65296  2.880162e-62
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03541868 18.33151  4.638427e-75
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03442689 22.11486 2.274108e-108
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04090595 12.62542  1.529252e-36
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03657578 20.65268  9.235874e-95
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03436607 23.26899 9.139469e-120
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03173209 25.91311 4.739027e-148
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03837513 17.14832  6.470041e-66
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03317997 22.52637 2.289813e-112
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5864440          0.7193544
#> 2          0.5750803          0.7035011
#> 3          0.6944382          0.8267522
#> 4          0.4608938          0.5886590
#> 5          0.6822644          0.8122926
#> 6          0.7374889          0.8454647
#> 7          0.7606054          0.8711299
#> 8          0.5661540          0.7184166
#> 9          0.6862843          0.8004821

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -49670491
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0369   18.1905    0.0000 [ 0.5741; 0.7649 ] 
#>   eta3 ~ eta1      0.4585      0.0929    4.9360    0.0000 [ 0.2262; 0.7066 ] 
#>   eta3 ~ eta2      0.3052      0.0915    3.3342    0.0009 [ 0.0577; 0.5310 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0398   16.6530    0.0000 [ 0.5665; 0.7725 ] 
#>   eta1 =~ y12      0.6493      0.0354   18.3315    0.0000 [ 0.5456; 0.7288 ] 
#>   eta1 =~ y13      0.7613      0.0344   22.1149    0.0000 [ 0.6687; 0.8468 ] 
#>   eta2 =~ y21      0.5165      0.0409   12.6254    0.0000 [ 0.4015; 0.6130 ] 
#>   eta2 =~ y22      0.7554      0.0366   20.6527    0.0000 [ 0.6663; 0.8554 ] 
#>   eta2 =~ y23      0.7997      0.0344   23.2690    0.0000 [ 0.7135; 0.8912 ] 
#>   eta3 =~ y31      0.8223      0.0317   25.9131    0.0000 [ 0.7439; 0.9080 ] 
#>   eta3 =~ y32      0.6581      0.0384   17.1483    0.0000 [ 0.5599; 0.7584 ] 
#>   eta3 =~ y33      0.7474      0.0332   22.5264    0.0000 [ 0.6603; 0.8319 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0189   20.9305    0.0000 [ 0.3535; 0.4512 ] 
#>   eta1 <~ y12      0.3873      0.0205   18.9248    0.0000 [ 0.3300; 0.4358 ] 
#>   eta1 <~ y13      0.4542      0.0209   21.7396    0.0000 [ 0.4012; 0.5092 ] 
#>   eta2 <~ y21      0.3058      0.0242   12.6563    0.0000 [ 0.2379; 0.3629 ] 
#>   eta2 <~ y22      0.4473      0.0208   21.4773    0.0000 [ 0.3968; 0.5045 ] 
#>   eta2 <~ y23      0.4735      0.0147   32.2021    0.0000 [ 0.4374; 0.5134 ] 
#>   eta3 <~ y31      0.4400      0.0192   22.8850    0.0000 [ 0.3912; 0.4906 ] 
#>   eta3 <~ y32      0.3521      0.0181   19.4068    0.0000 [ 0.3051; 0.3990 ] 
#>   eta3 <~ y33      0.3999      0.0133   30.0538    0.0000 [ 0.3641; 0.4329 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0369   18.1905    0.0000 [ 0.5741; 0.7649 ] 
#>   eta3 ~ eta1       0.6634      0.0404   16.4197    0.0000 [ 0.5594; 0.7683 ] 
#>   eta3 ~ eta2       0.3052      0.0915    3.3342    0.0009 [ 0.0577; 0.5310 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0618    3.3131    0.0009 [ 0.0376; 0.3574 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03690580 18.190456 6.142655e-74
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09288961  4.936039 7.972498e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09152223  3.334175 8.555284e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.57406982          0.7649258          0.5969879          0.7420077
#> 2         0.22620069          0.7065733          0.2838840          0.6488899
#> 3         0.05774219          0.5310435          0.1145764          0.4742093
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5886518          0.7552853          0.5954317          0.7469517
#> 2          0.2001178          0.6382196          0.2416877          0.5881604
#> 3          0.1367686          0.5769735          0.1923120          0.5197136