Summarize model

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1027215127
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0460   14.5903    0.0000 [ 0.5670; 0.7358 ] 
#>   eta3 ~ eta1      0.4585      0.0779    5.8885    0.0000 [ 0.2963; 0.5870 ] 
#>   eta3 ~ eta2      0.3052      0.0893    3.4172    0.0006 [ 0.1599; 0.5005 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0417   15.9006    0.0000 [ 0.5905; 0.7340 ] 
#>   eta1 =~ y12      0.6493      0.0382   16.9959    0.0000 [ 0.5912; 0.7364 ] 
#>   eta1 =~ y13      0.7613      0.0258   29.4740    0.0000 [ 0.7190; 0.8026 ] 
#>   eta2 =~ y21      0.5165      0.0586    8.8195    0.0000 [ 0.4251; 0.6291 ] 
#>   eta2 =~ y22      0.7554      0.0339   22.2973    0.0000 [ 0.6932; 0.8188 ] 
#>   eta2 =~ y23      0.7997      0.0371   21.5732    0.0000 [ 0.7154; 0.8412 ] 
#>   eta3 =~ y31      0.8223      0.0300   27.4370    0.0000 [ 0.7594; 0.8683 ] 
#>   eta3 =~ y32      0.6581      0.0472   13.9360    0.0000 [ 0.5966; 0.7625 ] 
#>   eta3 =~ y33      0.7474      0.0314   23.7878    0.0000 [ 0.6945; 0.7940 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0204   19.3447    0.0000 [ 0.3613; 0.4275 ] 
#>   eta1 <~ y12      0.3873      0.0170   22.8204    0.0000 [ 0.3531; 0.4163 ] 
#>   eta1 <~ y13      0.4542      0.0211   21.5111    0.0000 [ 0.4125; 0.4882 ] 
#>   eta2 <~ y21      0.3058      0.0295   10.3571    0.0000 [ 0.2581; 0.3689 ] 
#>   eta2 <~ y22      0.4473      0.0226   19.7931    0.0000 [ 0.4112; 0.4986 ] 
#>   eta2 <~ y23      0.4735      0.0235   20.1540    0.0000 [ 0.4285; 0.5134 ] 
#>   eta3 <~ y31      0.4400      0.0158   27.7635    0.0000 [ 0.4074; 0.4622 ] 
#>   eta3 <~ y32      0.3521      0.0195   18.0480    0.0000 [ 0.3210; 0.3968 ] 
#>   eta3 <~ y33      0.3999      0.0206   19.4111    0.0000 [ 0.3585; 0.4334 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0460   14.5903    0.0000 [ 0.5670; 0.7358 ] 
#>   eta3 ~ eta1       0.6634      0.0297   22.3553    0.0000 [ 0.5998; 0.7127 ] 
#>   eta3 ~ eta2       0.3052      0.0893    3.4172    0.0006 [ 0.1599; 0.5005 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0576    3.5579    0.0004 [ 0.0907; 0.3331 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04170081 15.900646  6.271427e-57
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03820214 16.995853  8.813872e-65
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02583106 29.474039 6.195477e-191
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05855825  8.819505  1.149663e-18
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03387797 22.297311 3.923795e-110
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03706743 21.573215 3.206209e-103
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02996967 27.436985 9.934904e-166
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04722087 13.935975  3.829324e-44
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03142049 23.787791 4.467862e-125
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5905294          0.7339879
#> 2          0.5911674          0.7363921
#> 3          0.7190173          0.8026473
#> 4          0.4251447          0.6291193
#> 5          0.6931892          0.8188329
#> 6          0.7153586          0.8411877
#> 7          0.7594216          0.8682631
#> 8          0.5965836          0.7625429
#> 9          0.6945488          0.7940212

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1027215127
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0460   14.5903    0.0000 [ 0.5685; 0.8064 ] 
#>   eta3 ~ eta1      0.4585      0.0779    5.8885    0.0000 [ 0.2527; 0.6553 ] 
#>   eta3 ~ eta2      0.3052      0.0893    3.4172    0.0006 [ 0.0740; 0.5358 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0417   15.9006    0.0000 [ 0.5532; 0.7689 ] 
#>   eta1 =~ y12      0.6493      0.0382   16.9959    0.0000 [ 0.5425; 0.7401 ] 
#>   eta1 =~ y13      0.7613      0.0258   29.4740    0.0000 [ 0.6926; 0.8262 ] 
#>   eta2 =~ y21      0.5165      0.0586    8.8195    0.0000 [ 0.3712; 0.6741 ] 
#>   eta2 =~ y22      0.7554      0.0339   22.2973    0.0000 [ 0.6653; 0.8405 ] 
#>   eta2 =~ y23      0.7997      0.0371   21.5732    0.0000 [ 0.7128; 0.9045 ] 
#>   eta3 =~ y31      0.8223      0.0300   27.4370    0.0000 [ 0.7468; 0.9018 ] 
#>   eta3 =~ y32      0.6581      0.0472   13.9360    0.0000 [ 0.5323; 0.7765 ] 
#>   eta3 =~ y33      0.7474      0.0314   23.7878    0.0000 [ 0.6649; 0.8274 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0204   19.3447    0.0000 [ 0.3452; 0.4509 ] 
#>   eta1 <~ y12      0.3873      0.0170   22.8204    0.0000 [ 0.3423; 0.4301 ] 
#>   eta1 <~ y13      0.4542      0.0211   21.5111    0.0000 [ 0.4022; 0.5114 ] 
#>   eta2 <~ y21      0.3058      0.0295   10.3571    0.0000 [ 0.2313; 0.3840 ] 
#>   eta2 <~ y22      0.4473      0.0226   19.7931    0.0000 [ 0.3838; 0.5006 ] 
#>   eta2 <~ y23      0.4735      0.0235   20.1540    0.0000 [ 0.4144; 0.5359 ] 
#>   eta3 <~ y31      0.4400      0.0158   27.7635    0.0000 [ 0.4011; 0.4830 ] 
#>   eta3 <~ y32      0.3521      0.0195   18.0480    0.0000 [ 0.3008; 0.4017 ] 
#>   eta3 <~ y33      0.3999      0.0206   19.4111    0.0000 [ 0.3467; 0.4533 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0460   14.5903    0.0000 [ 0.5685; 0.8064 ] 
#>   eta3 ~ eta1       0.6634      0.0297   22.3553    0.0000 [ 0.5874; 0.7409 ] 
#>   eta3 ~ eta2       0.3052      0.0893    3.4172    0.0006 [ 0.0740; 0.5358 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0576    3.5579    0.0004 [ 0.0613; 0.3590 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04601236 14.590285 3.238401e-48
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07786519  5.888469 3.897888e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08929747  3.417243 6.325889e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56847843          0.8064284          0.5970516          0.7778552
#> 2         0.25266655          0.6553413          0.3010199          0.6069880
#> 3         0.07403237          0.5358285          0.1294851          0.4803758
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5655997          0.7446207          0.5669524          0.7357788
#> 2          0.2860030          0.5900631          0.2963045          0.5870056
#> 3          0.1370764          0.5182873          0.1599076          0.5004834