Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 179919035
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0457   14.6917    0.0000 [ 0.5863; 0.7423 ] 
#>   eta3 ~ eta1      0.4585      0.1015    4.5192    0.0000 [ 0.3243; 0.6856 ] 
#>   eta3 ~ eta2      0.3052      0.1004    3.0398    0.0024 [ 0.0860; 0.4330 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0383   17.2978    0.0000 [ 0.5698; 0.7114 ] 
#>   eta1 =~ y12      0.6493      0.0373   17.4181    0.0000 [ 0.5800; 0.7130 ] 
#>   eta1 =~ y13      0.7613      0.0350   21.7834    0.0000 [ 0.6932; 0.8194 ] 
#>   eta2 =~ y21      0.5165      0.0458   11.2740    0.0000 [ 0.4104; 0.5908 ] 
#>   eta2 =~ y22      0.7554      0.0372   20.2841    0.0000 [ 0.6922; 0.8113 ] 
#>   eta2 =~ y23      0.7997      0.0403   19.8659    0.0000 [ 0.7387; 0.8734 ] 
#>   eta3 =~ y31      0.8223      0.0297   27.6989    0.0000 [ 0.7656; 0.8732 ] 
#>   eta3 =~ y32      0.6581      0.0366   17.9918    0.0000 [ 0.5797; 0.7129 ] 
#>   eta3 =~ y33      0.7474      0.0368   20.2907    0.0000 [ 0.6930; 0.8109 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0202   19.6158    0.0000 [ 0.3630; 0.4304 ] 
#>   eta1 <~ y12      0.3873      0.0156   24.7620    0.0000 [ 0.3659; 0.4178 ] 
#>   eta1 <~ y13      0.4542      0.0246   18.4460    0.0000 [ 0.4220; 0.5088 ] 
#>   eta2 <~ y21      0.3058      0.0246   12.4514    0.0000 [ 0.2581; 0.3462 ] 
#>   eta2 <~ y22      0.4473      0.0203   22.0101    0.0000 [ 0.4181; 0.4867 ] 
#>   eta2 <~ y23      0.4735      0.0206   22.9716    0.0000 [ 0.4407; 0.5034 ] 
#>   eta3 <~ y31      0.4400      0.0172   25.5788    0.0000 [ 0.4043; 0.4676 ] 
#>   eta3 <~ y32      0.3521      0.0172   20.4529    0.0000 [ 0.3171; 0.3775 ] 
#>   eta3 <~ y33      0.3999      0.0163   24.5198    0.0000 [ 0.3693; 0.4305 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0457   14.6917    0.0000 [ 0.5863; 0.7423 ] 
#>   eta3 ~ eta1       0.6634      0.0461   14.3791    0.0000 [ 0.5780; 0.7454 ] 
#>   eta3 ~ eta2       0.3052      0.1004    3.0398    0.0024 [ 0.0860; 0.4330 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0707    2.8987    0.0037 [ 0.0560; 0.3065 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03833264 17.29779  4.888115e-67
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03727595 17.41815  6.009105e-68
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03495078 21.78337 3.336064e-105
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04580944 11.27398  1.764048e-29
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03724037 20.28411  1.776569e-91
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04025315 19.86587  8.035474e-88
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02968624 27.69894 7.190591e-169
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03657600 17.99182  2.258105e-72
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03683587 20.29066  1.554968e-91
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5697515          0.7113939
#> 2          0.5799929          0.7130222
#> 3          0.6931766          0.8194481
#> 4          0.4103730          0.5907984
#> 5          0.6921502          0.8112508
#> 6          0.7386988          0.8734344
#> 7          0.7655706          0.8731716
#> 8          0.5796820          0.7129030
#> 9          0.6929651          0.8109040

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 179919035
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0457   14.6917    0.0000 [ 0.5521; 0.7884 ] 
#>   eta3 ~ eta1      0.4585      0.1015    4.5192    0.0000 [ 0.1825; 0.7071 ] 
#>   eta3 ~ eta2      0.3052      0.1004    3.0398    0.0024 [ 0.0613; 0.5804 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0383   17.2978    0.0000 [ 0.5756; 0.7738 ] 
#>   eta1 =~ y12      0.6493      0.0373   17.4181    0.0000 [ 0.5547; 0.7474 ] 
#>   eta1 =~ y13      0.7613      0.0350   21.7834    0.0000 [ 0.6805; 0.8613 ] 
#>   eta2 =~ y21      0.5165      0.0458   11.2740    0.0000 [ 0.3983; 0.6352 ] 
#>   eta2 =~ y22      0.7554      0.0372   20.2841    0.0000 [ 0.6625; 0.8551 ] 
#>   eta2 =~ y23      0.7997      0.0403   19.8659    0.0000 [ 0.6962; 0.9043 ] 
#>   eta3 =~ y31      0.8223      0.0297   27.6989    0.0000 [ 0.7514; 0.9049 ] 
#>   eta3 =~ y32      0.6581      0.0366   17.9918    0.0000 [ 0.5627; 0.7519 ] 
#>   eta3 =~ y33      0.7474      0.0368   20.2907    0.0000 [ 0.6459; 0.8364 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0202   19.6158    0.0000 [ 0.3441; 0.4484 ] 
#>   eta1 <~ y12      0.3873      0.0156   24.7620    0.0000 [ 0.3419; 0.4228 ] 
#>   eta1 <~ y13      0.4542      0.0246   18.4460    0.0000 [ 0.3887; 0.5160 ] 
#>   eta2 <~ y21      0.3058      0.0246   12.4514    0.0000 [ 0.2418; 0.3689 ] 
#>   eta2 <~ y22      0.4473      0.0203   22.0101    0.0000 [ 0.3956; 0.5007 ] 
#>   eta2 <~ y23      0.4735      0.0206   22.9716    0.0000 [ 0.4194; 0.5260 ] 
#>   eta3 <~ y31      0.4400      0.0172   25.5788    0.0000 [ 0.3990; 0.4879 ] 
#>   eta3 <~ y32      0.3521      0.0172   20.4529    0.0000 [ 0.3076; 0.3967 ] 
#>   eta3 <~ y33      0.3999      0.0163   24.5198    0.0000 [ 0.3549; 0.4392 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0457   14.6917    0.0000 [ 0.5521; 0.7884 ] 
#>   eta3 ~ eta1       0.6634      0.0461   14.3791    0.0000 [ 0.5407; 0.7793 ] 
#>   eta3 ~ eta2       0.3052      0.1004    3.0398    0.0024 [ 0.0613; 0.5804 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0707    2.8987    0.0037 [ 0.0324; 0.3979 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04569461 14.691743 7.281621e-49
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.10145773  4.519190 6.207670e-06
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.10038642  3.039765 2.367629e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55211713          0.7884238          0.5804929          0.7600480
#> 2         0.18246495          0.7071471          0.2454690          0.6441430
#> 3         0.06125024          0.5803921          0.1235890          0.5180533
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1        0.537469643          0.7478709         0.58630033          0.7423441
#> 2        0.284080060          0.7419712         0.32434552          0.6855643
#> 3        0.005351293          0.5000415         0.08599507          0.4330133