Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1846702017
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0449   14.9511    0.0000 [ 0.5786; 0.7276 ] 
#>   eta3 ~ eta1      0.4585      0.0729    6.2859    0.0000 [ 0.3267; 0.5910 ] 
#>   eta3 ~ eta2      0.3052      0.0819    3.7262    0.0002 [ 0.1533; 0.4487 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0426   15.5583    0.0000 [ 0.5797; 0.7420 ] 
#>   eta1 =~ y12      0.6493      0.0332   19.5278    0.0000 [ 0.5819; 0.7133 ] 
#>   eta1 =~ y13      0.7613      0.0392   19.4125    0.0000 [ 0.7183; 0.8713 ] 
#>   eta2 =~ y21      0.5165      0.0498   10.3656    0.0000 [ 0.4414; 0.6242 ] 
#>   eta2 =~ y22      0.7554      0.0383   19.7148    0.0000 [ 0.6837; 0.8105 ] 
#>   eta2 =~ y23      0.7997      0.0382   20.9300    0.0000 [ 0.7044; 0.8587 ] 
#>   eta3 =~ y31      0.8223      0.0321   25.6134    0.0000 [ 0.7663; 0.8817 ] 
#>   eta3 =~ y32      0.6581      0.0406   16.1976    0.0000 [ 0.5694; 0.7106 ] 
#>   eta3 =~ y33      0.7474      0.0310   24.0880    0.0000 [ 0.6801; 0.7853 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0215   18.3767    0.0000 [ 0.3505; 0.4272 ] 
#>   eta1 <~ y12      0.3873      0.0187   20.6647    0.0000 [ 0.3550; 0.4201 ] 
#>   eta1 <~ y13      0.4542      0.0204   22.2204    0.0000 [ 0.4159; 0.4914 ] 
#>   eta2 <~ y21      0.3058      0.0265   11.5557    0.0000 [ 0.2665; 0.3745 ] 
#>   eta2 <~ y22      0.4473      0.0253   17.6932    0.0000 [ 0.3967; 0.4919 ] 
#>   eta2 <~ y23      0.4735      0.0191   24.7630    0.0000 [ 0.4322; 0.5011 ] 
#>   eta3 <~ y31      0.4400      0.0163   26.9150    0.0000 [ 0.4191; 0.4763 ] 
#>   eta3 <~ y32      0.3521      0.0195   18.0588    0.0000 [ 0.3142; 0.3859 ] 
#>   eta3 <~ y33      0.3999      0.0157   25.5338    0.0000 [ 0.3702; 0.4222 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0449   14.9511    0.0000 [ 0.5786; 0.7276 ] 
#>   eta3 ~ eta1       0.6634      0.0400   16.5941    0.0000 [ 0.5806; 0.7188 ] 
#>   eta3 ~ eta2       0.3052      0.0819    3.7262    0.0002 [ 0.1533; 0.4487 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0537    3.8173    0.0001 [ 0.1058; 0.2997 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04261853 15.55825  1.398670e-54
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03324889 19.52781  6.371823e-85
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03921936 19.41250  6.050985e-84
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04982383 10.36562  3.554536e-25
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03831572 19.71482  1.608738e-86
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03820650 20.93005  2.852269e-97
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03210335 25.61344 1.080686e-144
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04062745 16.19764  5.239870e-59
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03102896 24.08795 3.343155e-128
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5797141          0.7420491
#> 2          0.5819274          0.7133141
#> 3          0.7183180          0.8712546
#> 4          0.4413859          0.6242026
#> 5          0.6837339          0.8104718
#> 6          0.7043738          0.8586637
#> 7          0.7662635          0.8817486
#> 8          0.5694410          0.7105858
#> 9          0.6801128          0.7853285

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1846702017
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0449   14.9511    0.0000 [ 0.5655; 0.7977 ] 
#>   eta3 ~ eta1      0.4585      0.0729    6.2859    0.0000 [ 0.2798; 0.6571 ] 
#>   eta3 ~ eta2      0.3052      0.0819    3.7262    0.0002 [ 0.0795; 0.5030 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0426   15.5583    0.0000 [ 0.5527; 0.7731 ] 
#>   eta1 =~ y12      0.6493      0.0332   19.5278    0.0000 [ 0.5641; 0.7361 ] 
#>   eta1 =~ y13      0.7613      0.0392   19.4125    0.0000 [ 0.6487; 0.8515 ] 
#>   eta2 =~ y21      0.5165      0.0498   10.3656    0.0000 [ 0.3910; 0.6487 ] 
#>   eta2 =~ y22      0.7554      0.0383   19.7148    0.0000 [ 0.6629; 0.8611 ] 
#>   eta2 =~ y23      0.7997      0.0382   20.9300    0.0000 [ 0.7143; 0.9119 ] 
#>   eta3 =~ y31      0.8223      0.0321   25.6134    0.0000 [ 0.7329; 0.8989 ] 
#>   eta3 =~ y32      0.6581      0.0406   16.1976    0.0000 [ 0.5589; 0.7690 ] 
#>   eta3 =~ y33      0.7474      0.0310   24.0880    0.0000 [ 0.6783; 0.8387 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0215   18.3767    0.0000 [ 0.3432; 0.4545 ] 
#>   eta1 <~ y12      0.3873      0.0187   20.6647    0.0000 [ 0.3425; 0.4394 ] 
#>   eta1 <~ y13      0.4542      0.0204   22.2204    0.0000 [ 0.3984; 0.5041 ] 
#>   eta2 <~ y21      0.3058      0.0265   11.5557    0.0000 [ 0.2347; 0.3715 ] 
#>   eta2 <~ y22      0.4473      0.0253   17.6932    0.0000 [ 0.3783; 0.5091 ] 
#>   eta2 <~ y23      0.4735      0.0191   24.7630    0.0000 [ 0.4245; 0.5233 ] 
#>   eta3 <~ y31      0.4400      0.0163   26.9150    0.0000 [ 0.3913; 0.4759 ] 
#>   eta3 <~ y32      0.3521      0.0195   18.0588    0.0000 [ 0.3026; 0.4035 ] 
#>   eta3 <~ y33      0.3999      0.0157   25.5338    0.0000 [ 0.3627; 0.4437 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0449   14.9511    0.0000 [ 0.5655; 0.7977 ] 
#>   eta3 ~ eta1       0.6634      0.0400   16.5941    0.0000 [ 0.5648; 0.7716 ] 
#>   eta3 ~ eta2       0.3052      0.0819    3.7262    0.0002 [ 0.0795; 0.5030 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0537    3.8173    0.0001 [ 0.0610; 0.3385 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04490209 14.951052 1.533090e-50
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07294207  6.285903 3.259544e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08189280  3.726226 1.943678e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56546446          0.7976727          0.5933481          0.7697890
#> 2         0.27984189          0.6570571          0.3251381          0.6117609
#> 3         0.07950251          0.5030058          0.1303570          0.4521514
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5782945          0.7290554          0.5785613          0.7275821
#> 2          0.3146616          0.6349028          0.3266843          0.5909817
#> 3          0.1122323          0.4681523          0.1532930          0.4486518