Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 259160724
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0463   14.5083    0.0000 [ 0.5882; 0.7537 ] 
#>   eta3 ~ eta1      0.4585      0.0736    6.2292    0.0000 [ 0.3154; 0.6103 ] 
#>   eta3 ~ eta2      0.3052      0.0811    3.7614    0.0002 [ 0.1531; 0.4467 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0444   14.9459    0.0000 [ 0.5910; 0.7326 ] 
#>   eta1 =~ y12      0.6493      0.0438   14.8297    0.0000 [ 0.5750; 0.7210 ] 
#>   eta1 =~ y13      0.7613      0.0296   25.7565    0.0000 [ 0.7128; 0.8224 ] 
#>   eta2 =~ y21      0.5165      0.0563    9.1766    0.0000 [ 0.4150; 0.5939 ] 
#>   eta2 =~ y22      0.7554      0.0277   27.2537    0.0000 [ 0.7035; 0.8091 ] 
#>   eta2 =~ y23      0.7997      0.0322   24.8188    0.0000 [ 0.7439; 0.8582 ] 
#>   eta3 =~ y31      0.8223      0.0374   21.9754    0.0000 [ 0.7544; 0.8770 ] 
#>   eta3 =~ y32      0.6581      0.0338   19.4694    0.0000 [ 0.6127; 0.7289 ] 
#>   eta3 =~ y33      0.7474      0.0423   17.6817    0.0000 [ 0.6676; 0.8155 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0203   19.5237    0.0000 [ 0.3548; 0.4289 ] 
#>   eta1 <~ y12      0.3873      0.0225   17.2238    0.0000 [ 0.3428; 0.4187 ] 
#>   eta1 <~ y13      0.4542      0.0209   21.7169    0.0000 [ 0.4236; 0.4927 ] 
#>   eta2 <~ y21      0.3058      0.0250   12.2432    0.0000 [ 0.2548; 0.3389 ] 
#>   eta2 <~ y22      0.4473      0.0188   23.7959    0.0000 [ 0.4221; 0.4821 ] 
#>   eta2 <~ y23      0.4735      0.0214   22.0963    0.0000 [ 0.4372; 0.5031 ] 
#>   eta3 <~ y31      0.4400      0.0218   20.1487    0.0000 [ 0.4003; 0.4796 ] 
#>   eta3 <~ y32      0.3521      0.0168   20.9339    0.0000 [ 0.3288; 0.3926 ] 
#>   eta3 <~ y33      0.3999      0.0194   20.6455    0.0000 [ 0.3695; 0.4374 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0463   14.5083    0.0000 [ 0.5882; 0.7537 ] 
#>   eta3 ~ eta1       0.6634      0.0394   16.8488    0.0000 [ 0.5995; 0.7222 ] 
#>   eta3 ~ eta2       0.3052      0.0811    3.7614    0.0002 [ 0.1531; 0.4467 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0557    3.6810    0.0002 [ 0.1083; 0.3057 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04436465 14.945905  1.656262e-50
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04378224 14.829711  9.413723e-50
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02955938 25.756491 2.726524e-146
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05627950  9.176606  4.448933e-20
#> 5 eta2 =~ y22  Common factor 0.7553877 0.02771684 27.253747 1.500319e-163
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03222007 24.818811 5.617448e-136
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03741812 21.975379 4.954014e-107
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03380018 19.469390  1.996331e-84
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04227104 17.681706  5.801362e-70
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5909799          0.7325927
#> 2          0.5750009          0.7209566
#> 3          0.7128065          0.8224459
#> 4          0.4149719          0.5939181
#> 5          0.7034592          0.8090896
#> 6          0.7438654          0.8582423
#> 7          0.7544286          0.8769774
#> 8          0.6126529          0.7288573
#> 9          0.6675918          0.8154980

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 259160724
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0463   14.5083    0.0000 [ 0.5514; 0.7907 ] 
#>   eta3 ~ eta1      0.4585      0.0736    6.2292    0.0000 [ 0.2803; 0.6610 ] 
#>   eta3 ~ eta2      0.3052      0.0811    3.7614    0.0002 [ 0.0830; 0.5025 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0444   14.9459    0.0000 [ 0.5571; 0.7866 ] 
#>   eta1 =~ y12      0.6493      0.0438   14.8297    0.0000 [ 0.5368; 0.7632 ] 
#>   eta1 =~ y13      0.7613      0.0296   25.7565    0.0000 [ 0.6755; 0.8283 ] 
#>   eta2 =~ y21      0.5165      0.0563    9.1766    0.0000 [ 0.3779; 0.6689 ] 
#>   eta2 =~ y22      0.7554      0.0277   27.2537    0.0000 [ 0.6791; 0.8225 ] 
#>   eta2 =~ y23      0.7997      0.0322   24.8188    0.0000 [ 0.7146; 0.8812 ] 
#>   eta3 =~ y31      0.8223      0.0374   21.9754    0.0000 [ 0.7291; 0.9226 ] 
#>   eta3 =~ y32      0.6581      0.0338   19.4694    0.0000 [ 0.5699; 0.7447 ] 
#>   eta3 =~ y33      0.7474      0.0423   17.6817    0.0000 [ 0.6446; 0.8632 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0203   19.5237    0.0000 [ 0.3489; 0.4537 ] 
#>   eta1 <~ y12      0.3873      0.0225   17.2238    0.0000 [ 0.3301; 0.4464 ] 
#>   eta1 <~ y13      0.4542      0.0209   21.7169    0.0000 [ 0.3946; 0.5027 ] 
#>   eta2 <~ y21      0.3058      0.0250   12.2432    0.0000 [ 0.2464; 0.3756 ] 
#>   eta2 <~ y22      0.4473      0.0188   23.7959    0.0000 [ 0.3963; 0.4935 ] 
#>   eta2 <~ y23      0.4735      0.0214   22.0963    0.0000 [ 0.4174; 0.5282 ] 
#>   eta3 <~ y31      0.4400      0.0218   20.1487    0.0000 [ 0.3826; 0.4956 ] 
#>   eta3 <~ y32      0.3521      0.0168   20.9339    0.0000 [ 0.3061; 0.3931 ] 
#>   eta3 <~ y33      0.3999      0.0194   20.6455    0.0000 [ 0.3511; 0.4512 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0463   14.5083    0.0000 [ 0.5514; 0.7907 ] 
#>   eta3 ~ eta1       0.6634      0.0394   16.8488    0.0000 [ 0.5655; 0.7691 ] 
#>   eta3 ~ eta2       0.3052      0.0811    3.7614    0.0002 [ 0.0830; 0.5025 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0557    3.6810    0.0002 [ 0.0528; 0.3406 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04627244 14.508277 1.073839e-47
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07360556  6.229241 4.686999e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08112630  3.761433 1.689428e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55142948          0.7907244          0.5801641          0.7619898
#> 2         0.28032402          0.6609704          0.3260322          0.6152622
#> 3         0.08299085          0.5025303          0.1333693          0.4521518
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5755383          0.7614787          0.5881776          0.7536601
#> 2          0.2921866          0.6123291          0.3153658          0.6103472
#> 3          0.1391006          0.4646052          0.1531159          0.4467326