Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 2008648436
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0519   12.9423    0.0000 [ 0.5938; 0.7538 ] 
#>   eta3 ~ eta1      0.4585      0.0843    5.4361    0.0000 [ 0.3571; 0.6184 ] 
#>   eta3 ~ eta2      0.3052      0.0938    3.2543    0.0011 [ 0.1511; 0.4471 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0345   19.2261    0.0000 [ 0.5967; 0.7247 ] 
#>   eta1 =~ y12      0.6493      0.0395   16.4505    0.0000 [ 0.5769; 0.7105 ] 
#>   eta1 =~ y13      0.7613      0.0282   26.9554    0.0000 [ 0.7143; 0.8167 ] 
#>   eta2 =~ y21      0.5165      0.0491   10.5273    0.0000 [ 0.4324; 0.6075 ] 
#>   eta2 =~ y22      0.7554      0.0331   22.8408    0.0000 [ 0.6577; 0.7892 ] 
#>   eta2 =~ y23      0.7997      0.0388   20.6306    0.0000 [ 0.7346; 0.8611 ] 
#>   eta3 =~ y31      0.8223      0.0374   22.0053    0.0000 [ 0.7470; 0.8954 ] 
#>   eta3 =~ y32      0.6581      0.0428   15.3860    0.0000 [ 0.5631; 0.7195 ] 
#>   eta3 =~ y33      0.7474      0.0429   17.4401    0.0000 [ 0.6790; 0.8404 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0196   20.1925    0.0000 [ 0.3518; 0.4275 ] 
#>   eta1 <~ y12      0.3873      0.0178   21.8099    0.0000 [ 0.3557; 0.4179 ] 
#>   eta1 <~ y13      0.4542      0.0176   25.7680    0.0000 [ 0.4216; 0.4845 ] 
#>   eta2 <~ y21      0.3058      0.0247   12.3568    0.0000 [ 0.2608; 0.3430 ] 
#>   eta2 <~ y22      0.4473      0.0222   20.1043    0.0000 [ 0.4005; 0.4902 ] 
#>   eta2 <~ y23      0.4735      0.0214   22.1694    0.0000 [ 0.4431; 0.5085 ] 
#>   eta3 <~ y31      0.4400      0.0160   27.5563    0.0000 [ 0.4221; 0.4763 ] 
#>   eta3 <~ y32      0.3521      0.0193   18.2659    0.0000 [ 0.3090; 0.3845 ] 
#>   eta3 <~ y33      0.3999      0.0263   15.1877    0.0000 [ 0.3660; 0.4609 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0519   12.9423    0.0000 [ 0.5938; 0.7538 ] 
#>   eta3 ~ eta1       0.6634      0.0321   20.6473    0.0000 [ 0.6125; 0.7389 ] 
#>   eta3 ~ eta2       0.3052      0.0938    3.2543    0.0011 [ 0.1511; 0.4471 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0634    3.2296    0.0012 [ 0.0952; 0.2916 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03448799 19.22611  2.237906e-82
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03946851 16.45053  8.313843e-61
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02824466 26.95539 4.932200e-160
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04905852 10.52732  6.464725e-26
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03307191 22.84076 1.805539e-115
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03876106 20.63059  1.458469e-94
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03736727 22.00528 2.563064e-107
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04277062 15.38600  2.031980e-53
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04285653 17.44015  4.090292e-68
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5967348          0.7247381
#> 2          0.5768528          0.7105116
#> 3          0.7142894          0.8166734
#> 4          0.4324165          0.6075006
#> 5          0.6576503          0.7891675
#> 6          0.7346450          0.8611379
#> 7          0.7469848          0.8954490
#> 8          0.5631122          0.7194563
#> 9          0.6789978          0.8404039

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 2008648436
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0519   12.9423    0.0000 [ 0.5524; 0.8206 ] 
#>   eta3 ~ eta1      0.4585      0.0843    5.4361    0.0000 [ 0.2383; 0.6745 ] 
#>   eta3 ~ eta2      0.3052      0.0938    3.2543    0.0011 [ 0.0583; 0.5432 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0345   19.2261    0.0000 [ 0.5748; 0.7532 ] 
#>   eta1 =~ y12      0.6493      0.0395   16.4505    0.0000 [ 0.5488; 0.7529 ] 
#>   eta1 =~ y13      0.7613      0.0282   26.9554    0.0000 [ 0.6792; 0.8253 ] 
#>   eta2 =~ y21      0.5165      0.0491   10.5273    0.0000 [ 0.4050; 0.6587 ] 
#>   eta2 =~ y22      0.7554      0.0331   22.8408    0.0000 [ 0.6791; 0.8501 ] 
#>   eta2 =~ y23      0.7997      0.0388   20.6306    0.0000 [ 0.6996; 0.9001 ] 
#>   eta3 =~ y31      0.8223      0.0374   22.0053    0.0000 [ 0.7279; 0.9211 ] 
#>   eta3 =~ y32      0.6581      0.0428   15.3860    0.0000 [ 0.5543; 0.7755 ] 
#>   eta3 =~ y33      0.7474      0.0429   17.4401    0.0000 [ 0.6437; 0.8653 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0196   20.1925    0.0000 [ 0.3475; 0.4488 ] 
#>   eta1 <~ y12      0.3873      0.0178   21.8099    0.0000 [ 0.3446; 0.4365 ] 
#>   eta1 <~ y13      0.4542      0.0176   25.7680    0.0000 [ 0.4055; 0.4967 ] 
#>   eta2 <~ y21      0.3058      0.0247   12.3568    0.0000 [ 0.2465; 0.3745 ] 
#>   eta2 <~ y22      0.4473      0.0222   20.1043    0.0000 [ 0.3880; 0.5031 ] 
#>   eta2 <~ y23      0.4735      0.0214   22.1694    0.0000 [ 0.4109; 0.5213 ] 
#>   eta3 <~ y31      0.4400      0.0160   27.5563    0.0000 [ 0.3954; 0.4780 ] 
#>   eta3 <~ y32      0.3521      0.0193   18.2659    0.0000 [ 0.3025; 0.4021 ] 
#>   eta3 <~ y33      0.3999      0.0263   15.1877    0.0000 [ 0.3313; 0.4674 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0519   12.9423    0.0000 [ 0.5524; 0.8206 ] 
#>   eta3 ~ eta1       0.6634      0.0321   20.6473    0.0000 [ 0.5811; 0.7472 ] 
#>   eta3 ~ eta2       0.3052      0.0938    3.2543    0.0011 [ 0.0583; 0.5432 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0634    3.2296    0.0012 [ 0.0438; 0.3718 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05187138 12.942270 2.598376e-38
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08434545  5.436058 5.447220e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09376818  3.254314 1.136664e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55235748          0.8206070          0.5845690          0.7883955
#> 2         0.23827493          0.6744620          0.2906525          0.6220844
#> 3         0.05830143          0.5432175          0.1165304          0.4849886
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1        0.529962009          0.7988315          0.5938243          0.7537671
#> 2        0.341411789          0.7374150          0.3570567          0.6183918
#> 3        0.002061306          0.4570896          0.1510886          0.4471256