Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1047763364
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0394   17.0571    0.0000 [ 0.5878; 0.7287 ] 
#>   eta3 ~ eta1      0.4585      0.0786    5.8298    0.0000 [ 0.3367; 0.5929 ] 
#>   eta3 ~ eta2      0.3052      0.0954    3.1992    0.0014 [ 0.1754; 0.4823 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0422   15.7293    0.0000 [ 0.5915; 0.7397 ] 
#>   eta1 =~ y12      0.6493      0.0460   14.1123    0.0000 [ 0.5494; 0.7112 ] 
#>   eta1 =~ y13      0.7613      0.0307   24.8242    0.0000 [ 0.7080; 0.8221 ] 
#>   eta2 =~ y21      0.5165      0.0431   11.9733    0.0000 [ 0.4106; 0.5932 ] 
#>   eta2 =~ y22      0.7554      0.0414   18.2621    0.0000 [ 0.6748; 0.8283 ] 
#>   eta2 =~ y23      0.7997      0.0348   22.9762    0.0000 [ 0.7274; 0.8486 ] 
#>   eta3 =~ y31      0.8223      0.0354   23.2084    0.0000 [ 0.7600; 0.8834 ] 
#>   eta3 =~ y32      0.6581      0.0339   19.4213    0.0000 [ 0.6003; 0.7143 ] 
#>   eta3 =~ y33      0.7474      0.0419   17.8459    0.0000 [ 0.6640; 0.8140 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0216   18.3334    0.0000 [ 0.3544; 0.4364 ] 
#>   eta1 <~ y12      0.3873      0.0200   19.3215    0.0000 [ 0.3531; 0.4210 ] 
#>   eta1 <~ y13      0.4542      0.0227   20.0286    0.0000 [ 0.4177; 0.4955 ] 
#>   eta2 <~ y21      0.3058      0.0212   14.4321    0.0000 [ 0.2710; 0.3490 ] 
#>   eta2 <~ y22      0.4473      0.0236   18.9698    0.0000 [ 0.4028; 0.5005 ] 
#>   eta2 <~ y23      0.4735      0.0205   23.0737    0.0000 [ 0.4364; 0.5218 ] 
#>   eta3 <~ y31      0.4400      0.0205   21.4609    0.0000 [ 0.4108; 0.4750 ] 
#>   eta3 <~ y32      0.3521      0.0140   25.1626    0.0000 [ 0.3240; 0.3839 ] 
#>   eta3 <~ y33      0.3999      0.0196   20.4290    0.0000 [ 0.3694; 0.4324 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0394   17.0571    0.0000 [ 0.5878; 0.7287 ] 
#>   eta3 ~ eta1       0.6634      0.0299   22.2124    0.0000 [ 0.6183; 0.7248 ] 
#>   eta3 ~ eta2       0.3052      0.0954    3.1992    0.0014 [ 0.1754; 0.4823 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0622    3.2931    0.0010 [ 0.1159; 0.3023 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04215521 15.72925  9.533257e-56
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04600781 14.11234  3.188082e-45
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03066944 24.82425 4.907291e-136
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04313393 11.97329  4.904737e-33
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04136372 18.26208  1.658463e-74
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03480396 22.97623 8.059497e-117
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03543017 23.20839 3.746015e-119
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03388395 19.42126  5.102415e-84
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04188206 17.84593  3.109106e-71
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5915409          0.7396902
#> 2          0.5494282          0.7111659
#> 3          0.7080335          0.8221132
#> 4          0.4105978          0.5932017
#> 5          0.6748471          0.8282907
#> 6          0.7273542          0.8485860
#> 7          0.7599670          0.8833511
#> 8          0.6002837          0.7142927
#> 9          0.6640382          0.8140030

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1047763364
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0394   17.0571    0.0000 [ 0.5749; 0.7784 ] 
#>   eta3 ~ eta1      0.4585      0.0786    5.8298    0.0000 [ 0.2426; 0.6493 ] 
#>   eta3 ~ eta2      0.3052      0.0954    3.1992    0.0014 [ 0.0594; 0.5527 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0422   15.7293    0.0000 [ 0.5518; 0.7698 ] 
#>   eta1 =~ y12      0.6493      0.0460   14.1123    0.0000 [ 0.5373; 0.7752 ] 
#>   eta1 =~ y13      0.7613      0.0307   24.8242    0.0000 [ 0.6833; 0.8419 ] 
#>   eta2 =~ y21      0.5165      0.0431   11.9733    0.0000 [ 0.4055; 0.6286 ] 
#>   eta2 =~ y22      0.7554      0.0414   18.2621    0.0000 [ 0.6529; 0.8668 ] 
#>   eta2 =~ y23      0.7997      0.0348   22.9762    0.0000 [ 0.7178; 0.8978 ] 
#>   eta3 =~ y31      0.8223      0.0354   23.2084    0.0000 [ 0.7338; 0.9170 ] 
#>   eta3 =~ y32      0.6581      0.0339   19.4213    0.0000 [ 0.5668; 0.7421 ] 
#>   eta3 =~ y33      0.7474      0.0419   17.8459    0.0000 [ 0.6380; 0.8546 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0216   18.3334    0.0000 [ 0.3370; 0.4486 ] 
#>   eta1 <~ y12      0.3873      0.0200   19.3215    0.0000 [ 0.3386; 0.4423 ] 
#>   eta1 <~ y13      0.4542      0.0227   20.0286    0.0000 [ 0.3943; 0.5115 ] 
#>   eta2 <~ y21      0.3058      0.0212   14.4321    0.0000 [ 0.2488; 0.3584 ] 
#>   eta2 <~ y22      0.4473      0.0236   18.9698    0.0000 [ 0.3848; 0.5067 ] 
#>   eta2 <~ y23      0.4735      0.0205   23.0737    0.0000 [ 0.4208; 0.5269 ] 
#>   eta3 <~ y31      0.4400      0.0205   21.4609    0.0000 [ 0.3891; 0.4951 ] 
#>   eta3 <~ y32      0.3521      0.0140   25.1626    0.0000 [ 0.3146; 0.3869 ] 
#>   eta3 <~ y33      0.3999      0.0196   20.4290    0.0000 [ 0.3493; 0.4505 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0394   17.0571    0.0000 [ 0.5749; 0.7784 ] 
#>   eta3 ~ eta1       0.6634      0.0299   22.2124    0.0000 [ 0.5764; 0.7309 ] 
#>   eta3 ~ eta2       0.3052      0.0954    3.1992    0.0014 [ 0.0594; 0.5527 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0622    3.2931    0.0010 [ 0.0468; 0.3685 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03935801 17.057098 3.095551e-65
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07864875  5.829803 5.549269e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09538492  3.199155 1.378311e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.57489382          0.7784312          0.5993347          0.7539904
#> 2         0.24261336          0.6493403          0.2914533          0.6005003
#> 3         0.05943581          0.5527127          0.1186687          0.4934798
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5725823          0.7548289          0.5878384          0.7287255
#> 2          0.3365534          0.6349638          0.3367308          0.5928836
#> 3          0.1166958          0.4969563          0.1753602          0.4823073