Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1985714266
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0450   14.9201    0.0000 [ 0.5820; 0.7465 ] 
#>   eta3 ~ eta1      0.4585      0.0898    5.1033    0.0000 [ 0.3186; 0.6152 ] 
#>   eta3 ~ eta2      0.3052      0.0900    3.3900    0.0007 [ 0.1460; 0.4535 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0421   15.7496    0.0000 [ 0.5623; 0.7408 ] 
#>   eta1 =~ y12      0.6493      0.0431   15.0555    0.0000 [ 0.5592; 0.7164 ] 
#>   eta1 =~ y13      0.7613      0.0320   23.8132    0.0000 [ 0.7046; 0.8235 ] 
#>   eta2 =~ y21      0.5165      0.0505   10.2251    0.0000 [ 0.4159; 0.5824 ] 
#>   eta2 =~ y22      0.7554      0.0370   20.3911    0.0000 [ 0.7101; 0.8220 ] 
#>   eta2 =~ y23      0.7997      0.0361   22.1697    0.0000 [ 0.7277; 0.8643 ] 
#>   eta3 =~ y31      0.8223      0.0309   26.6158    0.0000 [ 0.7689; 0.8732 ] 
#>   eta3 =~ y32      0.6581      0.0408   16.1272    0.0000 [ 0.5671; 0.7243 ] 
#>   eta3 =~ y33      0.7474      0.0397   18.8111    0.0000 [ 0.6760; 0.8182 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0234   16.9138    0.0000 [ 0.3464; 0.4293 ] 
#>   eta1 <~ y12      0.3873      0.0202   19.1682    0.0000 [ 0.3513; 0.4202 ] 
#>   eta1 <~ y13      0.4542      0.0219   20.7398    0.0000 [ 0.4236; 0.5005 ] 
#>   eta2 <~ y21      0.3058      0.0272   11.2228    0.0000 [ 0.2517; 0.3455 ] 
#>   eta2 <~ y22      0.4473      0.0211   21.1557    0.0000 [ 0.4205; 0.4960 ] 
#>   eta2 <~ y23      0.4735      0.0191   24.8181    0.0000 [ 0.4405; 0.5035 ] 
#>   eta3 <~ y31      0.4400      0.0175   25.2028    0.0000 [ 0.4103; 0.4654 ] 
#>   eta3 <~ y32      0.3521      0.0174   20.2278    0.0000 [ 0.3156; 0.3793 ] 
#>   eta3 <~ y33      0.3999      0.0208   19.2139    0.0000 [ 0.3677; 0.4394 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0450   14.9201    0.0000 [ 0.5820; 0.7465 ] 
#>   eta3 ~ eta1       0.6634      0.0408   16.2684    0.0000 [ 0.5936; 0.7414 ] 
#>   eta3 ~ eta2       0.3052      0.0900    3.3900    0.0007 [ 0.1460; 0.4535 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0607    3.3748    0.0007 [ 0.0996; 0.3015 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04210073 15.74961  6.911000e-56
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04312549 15.05555  3.174505e-51
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03197152 23.81325 2.434919e-125
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05050849 10.22511  1.530552e-24
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03704489 20.39114  2.004128e-92
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03607008 22.16972 6.732568e-109
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03089433 26.61580 4.455140e-156
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04080496 16.12718  1.643553e-58
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03973317 18.81109  6.126587e-79
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5622690          0.7408205
#> 2          0.5592246          0.7164263
#> 3          0.7046335          0.8234823
#> 4          0.4159416          0.5823811
#> 5          0.7101179          0.8220339
#> 6          0.7277388          0.8642588
#> 7          0.7688788          0.8731570
#> 8          0.5671088          0.7243014
#> 9          0.6760395          0.8181766

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1985714266
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0450   14.9201    0.0000 [ 0.5562; 0.7889 ] 
#>   eta3 ~ eta1      0.4585      0.0898    5.1033    0.0000 [ 0.2146; 0.6792 ] 
#>   eta3 ~ eta2      0.3052      0.0900    3.3900    0.0007 [ 0.0834; 0.5490 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0421   15.7496    0.0000 [ 0.5653; 0.7830 ] 
#>   eta1 =~ y12      0.6493      0.0431   15.0555    0.0000 [ 0.5377; 0.7607 ] 
#>   eta1 =~ y13      0.7613      0.0320   23.8132    0.0000 [ 0.6774; 0.8428 ] 
#>   eta2 =~ y21      0.5165      0.0505   10.2251    0.0000 [ 0.3917; 0.6529 ] 
#>   eta2 =~ y22      0.7554      0.0370   20.3911    0.0000 [ 0.6473; 0.8389 ] 
#>   eta2 =~ y23      0.7997      0.0361   22.1697    0.0000 [ 0.7154; 0.9019 ] 
#>   eta3 =~ y31      0.8223      0.0309   26.6158    0.0000 [ 0.7438; 0.9036 ] 
#>   eta3 =~ y32      0.6581      0.0408   16.1272    0.0000 [ 0.5654; 0.7764 ] 
#>   eta3 =~ y33      0.7474      0.0397   18.8111    0.0000 [ 0.6310; 0.8365 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0234   16.9138    0.0000 [ 0.3393; 0.4602 ] 
#>   eta1 <~ y12      0.3873      0.0202   19.1682    0.0000 [ 0.3329; 0.4374 ] 
#>   eta1 <~ y13      0.4542      0.0219   20.7398    0.0000 [ 0.3937; 0.5070 ] 
#>   eta2 <~ y21      0.3058      0.0272   11.2228    0.0000 [ 0.2389; 0.3798 ] 
#>   eta2 <~ y22      0.4473      0.0211   21.1557    0.0000 [ 0.3852; 0.4945 ] 
#>   eta2 <~ y23      0.4735      0.0191   24.8181    0.0000 [ 0.4294; 0.5280 ] 
#>   eta3 <~ y31      0.4400      0.0175   25.2028    0.0000 [ 0.3957; 0.4860 ] 
#>   eta3 <~ y32      0.3521      0.0174   20.2278    0.0000 [ 0.3144; 0.4044 ] 
#>   eta3 <~ y33      0.3999      0.0208   19.2139    0.0000 [ 0.3390; 0.4466 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0450   14.9201    0.0000 [ 0.5562; 0.7889 ] 
#>   eta3 ~ eta1       0.6634      0.0408   16.2684    0.0000 [ 0.5545; 0.7654 ] 
#>   eta3 ~ eta2       0.3052      0.0900    3.3900    0.0007 [ 0.0834; 0.5490 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0607    3.3748    0.0007 [ 0.0561; 0.3700 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04499531 14.920077 2.439926e-50
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08984531  5.103291 3.337967e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09001457  3.390019 6.988767e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5561843          0.7888746          0.5841259          0.7609331
#> 2          0.2145864          0.6792156          0.2703793          0.6234228
#> 3          0.0834490          0.5489535          0.1393470          0.4930555
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5818296          0.7575581          0.5820217          0.7465381
#> 2          0.2525662          0.6708147          0.3186160          0.6152076
#> 3          0.1081353          0.5153909          0.1459675          0.4535404