Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1319421890
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0459   14.6350    0.0000 [ 0.5985; 0.7710 ] 
#>   eta3 ~ eta1      0.4585      0.0911    5.0344    0.0000 [ 0.2894; 0.6484 ] 
#>   eta3 ~ eta2      0.3052      0.0978    3.1191    0.0018 [ 0.1293; 0.4827 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0478   13.8849    0.0000 [ 0.5877; 0.7522 ] 
#>   eta1 =~ y12      0.6493      0.0343   18.9437    0.0000 [ 0.5839; 0.7131 ] 
#>   eta1 =~ y13      0.7613      0.0317   23.9833    0.0000 [ 0.6993; 0.8234 ] 
#>   eta2 =~ y21      0.5165      0.0615    8.3980    0.0000 [ 0.4100; 0.6099 ] 
#>   eta2 =~ y22      0.7554      0.0378   19.9970    0.0000 [ 0.6868; 0.8196 ] 
#>   eta2 =~ y23      0.7997      0.0321   24.9049    0.0000 [ 0.7448; 0.8558 ] 
#>   eta3 =~ y31      0.8223      0.0337   24.4222    0.0000 [ 0.7762; 0.8956 ] 
#>   eta3 =~ y32      0.6581      0.0473   13.9244    0.0000 [ 0.5741; 0.7389 ] 
#>   eta3 =~ y33      0.7474      0.0393   19.0093    0.0000 [ 0.6799; 0.8042 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0227   17.4133    0.0000 [ 0.3564; 0.4283 ] 
#>   eta1 <~ y12      0.3873      0.0190   20.3502    0.0000 [ 0.3609; 0.4270 ] 
#>   eta1 <~ y13      0.4542      0.0214   21.2728    0.0000 [ 0.4175; 0.4938 ] 
#>   eta2 <~ y21      0.3058      0.0338    9.0355    0.0000 [ 0.2384; 0.3486 ] 
#>   eta2 <~ y22      0.4473      0.0187   23.8791    0.0000 [ 0.4105; 0.4745 ] 
#>   eta2 <~ y23      0.4735      0.0235   20.1361    0.0000 [ 0.4344; 0.5133 ] 
#>   eta3 <~ y31      0.4400      0.0205   21.4588    0.0000 [ 0.4088; 0.4798 ] 
#>   eta3 <~ y32      0.3521      0.0208   16.9684    0.0000 [ 0.3098; 0.3889 ] 
#>   eta3 <~ y33      0.3999      0.0204   19.5918    0.0000 [ 0.3636; 0.4414 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0459   14.6350    0.0000 [ 0.5985; 0.7710 ] 
#>   eta3 ~ eta1       0.6634      0.0407   16.2803    0.0000 [ 0.5989; 0.7488 ] 
#>   eta3 ~ eta2       0.3052      0.0978    3.1191    0.0018 [ 0.1293; 0.4827 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0639    3.2074    0.0013 [ 0.0995; 0.3323 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04775465 13.884928  7.817929e-44
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03427405 18.943720  4.975433e-80
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03174485 23.983286 4.155509e-127
#> 4 eta2 =~ y21  Common factor 0.5164548 0.06149771  8.397953  4.543309e-17
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03777502 19.997016  5.846737e-89
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03210874 24.904858 6.591171e-137
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03366925 24.422209 9.935699e-132
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04726026 13.924360  4.505586e-44
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03931885 19.009310  1.428195e-80
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5877489          0.7521816
#> 2          0.5839178          0.7131193
#> 3          0.6993176          0.8233930
#> 4          0.4100155          0.6098849
#> 5          0.6867977          0.8195765
#> 6          0.7448134          0.8558282
#> 7          0.7762074          0.8956307
#> 8          0.5741306          0.7389197
#> 9          0.6798706          0.8041875

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1319421890
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0459   14.6350    0.0000 [ 0.5475; 0.7847 ] 
#>   eta3 ~ eta1      0.4585      0.0911    5.0344    0.0000 [ 0.2186; 0.6896 ] 
#>   eta3 ~ eta2      0.3052      0.0978    3.1191    0.0018 [ 0.0457; 0.5516 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0478   13.8849    0.0000 [ 0.5440; 0.7910 ] 
#>   eta1 =~ y12      0.6493      0.0343   18.9437    0.0000 [ 0.5608; 0.7381 ] 
#>   eta1 =~ y13      0.7613      0.0317   23.9833    0.0000 [ 0.6829; 0.8470 ] 
#>   eta2 =~ y21      0.5165      0.0615    8.3980    0.0000 [ 0.3642; 0.6823 ] 
#>   eta2 =~ y22      0.7554      0.0378   19.9970    0.0000 [ 0.6576; 0.8530 ] 
#>   eta2 =~ y23      0.7997      0.0321   24.9049    0.0000 [ 0.7144; 0.8805 ] 
#>   eta3 =~ y31      0.8223      0.0337   24.4222    0.0000 [ 0.7274; 0.9016 ] 
#>   eta3 =~ y32      0.6581      0.0473   13.9244    0.0000 [ 0.5398; 0.7842 ] 
#>   eta3 =~ y33      0.7474      0.0393   19.0093    0.0000 [ 0.6444; 0.8477 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0227   17.4133    0.0000 [ 0.3376; 0.4551 ] 
#>   eta1 <~ y12      0.3873      0.0190   20.3502    0.0000 [ 0.3361; 0.4345 ] 
#>   eta1 <~ y13      0.4542      0.0214   21.2728    0.0000 [ 0.3985; 0.5089 ] 
#>   eta2 <~ y21      0.3058      0.0338    9.0355    0.0000 [ 0.2223; 0.3973 ] 
#>   eta2 <~ y22      0.4473      0.0187   23.8791    0.0000 [ 0.3985; 0.4953 ] 
#>   eta2 <~ y23      0.4735      0.0235   20.1361    0.0000 [ 0.4107; 0.5323 ] 
#>   eta3 <~ y31      0.4400      0.0205   21.4588    0.0000 [ 0.3846; 0.4906 ] 
#>   eta3 <~ y32      0.3521      0.0208   16.9684    0.0000 [ 0.3023; 0.4097 ] 
#>   eta3 <~ y33      0.3999      0.0204   19.5918    0.0000 [ 0.3481; 0.4537 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0459   14.6350    0.0000 [ 0.5475; 0.7847 ] 
#>   eta3 ~ eta1       0.6634      0.0407   16.2803    0.0000 [ 0.5491; 0.7598 ] 
#>   eta3 ~ eta2       0.3052      0.0978    3.1191    0.0018 [ 0.0457; 0.5516 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0639    3.2074    0.0013 [ 0.0352; 0.3655 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04587170 14.635025 1.679146e-48
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09107389  5.034448 4.792278e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09783178  3.119141 1.813792e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5474779          0.7847004          0.5759637          0.7562146
#> 2          0.2185709          0.6895536          0.2751267          0.6329978
#> 3          0.0456827          0.5516134          0.1064351          0.4908611
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.59714002          0.7966445          0.5984535          0.7710242
#> 2         0.27254775          0.6763215          0.2894217          0.6483822
#> 3         0.06284222          0.4917700          0.1293277          0.4827118