Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1084552371
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0422   15.9046    0.0000 [ 0.5990; 0.7373 ] 
#>   eta3 ~ eta1      0.4585      0.0781    5.8696    0.0000 [ 0.3636; 0.5803 ] 
#>   eta3 ~ eta2      0.3052      0.0822    3.7120    0.0002 [ 0.1600; 0.4040 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0491   13.4956    0.0000 [ 0.5617; 0.7500 ] 
#>   eta1 =~ y12      0.6493      0.0338   19.1869    0.0000 [ 0.6053; 0.7151 ] 
#>   eta1 =~ y13      0.7613      0.0357   21.3015    0.0000 [ 0.7046; 0.8300 ] 
#>   eta2 =~ y21      0.5165      0.0472   10.9365    0.0000 [ 0.4223; 0.6017 ] 
#>   eta2 =~ y22      0.7554      0.0369   20.4763    0.0000 [ 0.6814; 0.8182 ] 
#>   eta2 =~ y23      0.7997      0.0370   21.6244    0.0000 [ 0.7330; 0.8660 ] 
#>   eta3 =~ y31      0.8223      0.0339   24.2870    0.0000 [ 0.7515; 0.8751 ] 
#>   eta3 =~ y32      0.6581      0.0409   16.0795    0.0000 [ 0.5905; 0.7292 ] 
#>   eta3 =~ y33      0.7474      0.0308   24.2349    0.0000 [ 0.6936; 0.8151 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0244   16.1849    0.0000 [ 0.3462; 0.4314 ] 
#>   eta1 <~ y12      0.3873      0.0194   20.0036    0.0000 [ 0.3586; 0.4275 ] 
#>   eta1 <~ y13      0.4542      0.0220   20.6036    0.0000 [ 0.4171; 0.4904 ] 
#>   eta2 <~ y21      0.3058      0.0251   12.1630    0.0000 [ 0.2538; 0.3590 ] 
#>   eta2 <~ y22      0.4473      0.0198   22.5645    0.0000 [ 0.4137; 0.4790 ] 
#>   eta2 <~ y23      0.4735      0.0208   22.7854    0.0000 [ 0.4294; 0.5151 ] 
#>   eta3 <~ y31      0.4400      0.0164   26.7762    0.0000 [ 0.4084; 0.4624 ] 
#>   eta3 <~ y32      0.3521      0.0188   18.7381    0.0000 [ 0.3154; 0.3876 ] 
#>   eta3 <~ y33      0.3999      0.0171   23.3938    0.0000 [ 0.3705; 0.4331 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0422   15.9046    0.0000 [ 0.5990; 0.7373 ] 
#>   eta3 ~ eta1       0.6634      0.0424   15.6410    0.0000 [ 0.5956; 0.7341 ] 
#>   eta3 ~ eta2       0.3052      0.0822    3.7120    0.0002 [ 0.1600; 0.4040 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0612    3.3452    0.0008 [ 0.1032; 0.2816 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04913244 13.49556  1.660837e-41
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03383956 19.18695  4.757573e-82
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03574142 21.30150 1.099425e-100
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04722317 10.93647  7.714503e-28
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03689082 20.47631  3.502402e-93
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03697967 21.62441 1.058561e-103
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03385671 24.28699 2.690705e-130
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04092600 16.07948  3.553285e-58
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03084078 24.23493 9.533629e-130
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5617194          0.7499783
#> 2          0.6052509          0.7150530
#> 3          0.7045543          0.8300136
#> 4          0.4222977          0.6017466
#> 5          0.6814140          0.8181653
#> 6          0.7330351          0.8659994
#> 7          0.7515284          0.8750538
#> 8          0.5904839          0.7291860
#> 9          0.6935861          0.8150671

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1084552371
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0422   15.9046    0.0000 [ 0.5520; 0.7703 ] 
#>   eta3 ~ eta1      0.4585      0.0781    5.8696    0.0000 [ 0.2403; 0.6443 ] 
#>   eta3 ~ eta2      0.3052      0.0822    3.7120    0.0002 [ 0.1130; 0.5381 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0491   13.4956    0.0000 [ 0.5392; 0.7933 ] 
#>   eta1 =~ y12      0.6493      0.0338   19.1869    0.0000 [ 0.5512; 0.7262 ] 
#>   eta1 =~ y13      0.7613      0.0357   21.3015    0.0000 [ 0.6738; 0.8586 ] 
#>   eta2 =~ y21      0.5165      0.0472   10.9365    0.0000 [ 0.3898; 0.6340 ] 
#>   eta2 =~ y22      0.7554      0.0369   20.4763    0.0000 [ 0.6586; 0.8494 ] 
#>   eta2 =~ y23      0.7997      0.0370   21.6244    0.0000 [ 0.6992; 0.8904 ] 
#>   eta3 =~ y31      0.8223      0.0339   24.2870    0.0000 [ 0.7368; 0.9119 ] 
#>   eta3 =~ y32      0.6581      0.0409   16.0795    0.0000 [ 0.5522; 0.7638 ] 
#>   eta3 =~ y33      0.7474      0.0308   24.2349    0.0000 [ 0.6665; 0.8260 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0244   16.1849    0.0000 [ 0.3354; 0.4617 ] 
#>   eta1 <~ y12      0.3873      0.0194   20.0036    0.0000 [ 0.3318; 0.4319 ] 
#>   eta1 <~ y13      0.4542      0.0220   20.6036    0.0000 [ 0.4009; 0.5149 ] 
#>   eta2 <~ y21      0.3058      0.0251   12.1630    0.0000 [ 0.2408; 0.3708 ] 
#>   eta2 <~ y22      0.4473      0.0198   22.5645    0.0000 [ 0.3989; 0.5014 ] 
#>   eta2 <~ y23      0.4735      0.0208   22.7854    0.0000 [ 0.4207; 0.5282 ] 
#>   eta3 <~ y31      0.4400      0.0164   26.7762    0.0000 [ 0.3985; 0.4835 ] 
#>   eta3 <~ y32      0.3521      0.0188   18.7381    0.0000 [ 0.3035; 0.4007 ] 
#>   eta3 <~ y33      0.3999      0.0171   23.3938    0.0000 [ 0.3549; 0.4433 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0422   15.9046    0.0000 [ 0.5520; 0.7703 ] 
#>   eta3 ~ eta1       0.6634      0.0424   15.6410    0.0000 [ 0.5475; 0.7668 ] 
#>   eta3 ~ eta2       0.3052      0.0822    3.7120    0.0002 [ 0.1130; 0.5381 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0612    3.3452    0.0008 [ 0.0565; 0.3732 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04220990 15.904643 5.883739e-57
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07811522  5.869622 4.367907e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08220576  3.712041 2.055950e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5520434          0.7703292          0.5782553          0.7441173
#> 2          0.2403350          0.6443028          0.2888437          0.5957942
#> 3          0.1129807          0.5381025          0.1640295          0.4870537
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5735371          0.7636257          0.5989781          0.7372582
#> 2          0.2533550          0.6057468          0.3635740          0.5802622
#> 3          0.1560705          0.4938957          0.1600058          0.4039854