Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 541098223
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0470   14.2970    0.0000 [ 0.5857; 0.7478 ] 
#>   eta3 ~ eta1      0.4585      0.0776    5.9073    0.0000 [ 0.2803; 0.6022 ] 
#>   eta3 ~ eta2      0.3052      0.0796    3.8337    0.0001 [ 0.1333; 0.4712 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0453   14.6329    0.0000 [ 0.5810; 0.7315 ] 
#>   eta1 =~ y12      0.6493      0.0373   17.4133    0.0000 [ 0.5731; 0.7264 ] 
#>   eta1 =~ y13      0.7613      0.0371   20.5232    0.0000 [ 0.6981; 0.8232 ] 
#>   eta2 =~ y21      0.5165      0.0494   10.4460    0.0000 [ 0.4275; 0.5985 ] 
#>   eta2 =~ y22      0.7554      0.0410   18.4092    0.0000 [ 0.7002; 0.8322 ] 
#>   eta2 =~ y23      0.7997      0.0308   25.9288    0.0000 [ 0.7420; 0.8454 ] 
#>   eta3 =~ y31      0.8223      0.0325   25.3270    0.0000 [ 0.7718; 0.8840 ] 
#>   eta3 =~ y32      0.6581      0.0410   16.0636    0.0000 [ 0.5830; 0.7285 ] 
#>   eta3 =~ y33      0.7474      0.0369   20.2619    0.0000 [ 0.6694; 0.7969 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0224   17.6326    0.0000 [ 0.3474; 0.4378 ] 
#>   eta1 <~ y12      0.3873      0.0189   20.5097    0.0000 [ 0.3470; 0.4158 ] 
#>   eta1 <~ y13      0.4542      0.0218   20.8780    0.0000 [ 0.4187; 0.5036 ] 
#>   eta2 <~ y21      0.3058      0.0261   11.7306    0.0000 [ 0.2469; 0.3479 ] 
#>   eta2 <~ y22      0.4473      0.0189   23.6491    0.0000 [ 0.4139; 0.4821 ] 
#>   eta2 <~ y23      0.4735      0.0212   22.3801    0.0000 [ 0.4376; 0.5039 ] 
#>   eta3 <~ y31      0.4400      0.0179   24.6448    0.0000 [ 0.4071; 0.4689 ] 
#>   eta3 <~ y32      0.3521      0.0205   17.1998    0.0000 [ 0.3205; 0.3923 ] 
#>   eta3 <~ y33      0.3999      0.0177   22.6181    0.0000 [ 0.3667; 0.4257 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0470   14.2970    0.0000 [ 0.5857; 0.7478 ] 
#>   eta3 ~ eta1       0.6634      0.0386   17.1876    0.0000 [ 0.5558; 0.7200 ] 
#>   eta3 ~ eta2       0.3052      0.0796    3.8337    0.0001 [ 0.1333; 0.4712 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0517    3.9644    0.0001 [ 0.1004; 0.2856 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04531350 14.63294  1.731288e-48
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03728639 17.41327  6.543857e-68
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03709680 20.52322  1.335616e-93
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04944033 10.44602  1.527986e-25
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04103320 18.40918  1.108791e-75
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03084077 25.92879 3.154827e-148
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03246648 25.32697 1.612325e-141
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04096634 16.06365  4.587662e-58
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03688820 20.26188  2.791095e-91
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5809700          0.7315198
#> 2          0.5730824          0.7264261
#> 3          0.6981332          0.8232144
#> 4          0.4275285          0.5984667
#> 5          0.7002154          0.8321626
#> 6          0.7419715          0.8454187
#> 7          0.7717535          0.8840284
#> 8          0.5830064          0.7285225
#> 9          0.6694408          0.7968797

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 541098223
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0470   14.2970    0.0000 [ 0.5500; 0.7928 ] 
#>   eta3 ~ eta1      0.4585      0.0776    5.9073    0.0000 [ 0.2728; 0.6742 ] 
#>   eta3 ~ eta2      0.3052      0.0796    3.8337    0.0001 [ 0.0896; 0.5012 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0453   14.6329    0.0000 [ 0.5536; 0.7880 ] 
#>   eta1 =~ y12      0.6493      0.0373   17.4133    0.0000 [ 0.5618; 0.7547 ] 
#>   eta1 =~ y13      0.7613      0.0371   20.5232    0.0000 [ 0.6684; 0.8603 ] 
#>   eta2 =~ y21      0.5165      0.0494   10.4460    0.0000 [ 0.3948; 0.6505 ] 
#>   eta2 =~ y22      0.7554      0.0410   18.4092    0.0000 [ 0.6492; 0.8614 ] 
#>   eta2 =~ y23      0.7997      0.0308   25.9288    0.0000 [ 0.7275; 0.8870 ] 
#>   eta3 =~ y31      0.8223      0.0325   25.3270    0.0000 [ 0.7372; 0.9051 ] 
#>   eta3 =~ y32      0.6581      0.0410   16.0636    0.0000 [ 0.5543; 0.7661 ] 
#>   eta3 =~ y33      0.7474      0.0369   20.2619    0.0000 [ 0.6546; 0.8454 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0224   17.6326    0.0000 [ 0.3370; 0.4530 ] 
#>   eta1 <~ y12      0.3873      0.0189   20.5097    0.0000 [ 0.3386; 0.4363 ] 
#>   eta1 <~ y13      0.4542      0.0218   20.8780    0.0000 [ 0.3934; 0.5059 ] 
#>   eta2 <~ y21      0.3058      0.0261   11.7306    0.0000 [ 0.2397; 0.3745 ] 
#>   eta2 <~ y22      0.4473      0.0189   23.6491    0.0000 [ 0.3946; 0.4924 ] 
#>   eta2 <~ y23      0.4735      0.0212   22.3801    0.0000 [ 0.4190; 0.5284 ] 
#>   eta3 <~ y31      0.4400      0.0179   24.6448    0.0000 [ 0.3923; 0.4846 ] 
#>   eta3 <~ y32      0.3521      0.0205   17.1998    0.0000 [ 0.2997; 0.4056 ] 
#>   eta3 <~ y33      0.3999      0.0177   22.6181    0.0000 [ 0.3549; 0.4463 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0470   14.2970    0.0000 [ 0.5500; 0.7928 ] 
#>   eta3 ~ eta1       0.6634      0.0386   17.1876    0.0000 [ 0.5729; 0.7725 ] 
#>   eta3 ~ eta2       0.3052      0.0796    3.8337    0.0001 [ 0.0896; 0.5012 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0517    3.9644    0.0001 [ 0.0656; 0.3328 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04695610 14.297044 2.283039e-46
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07761648  5.907338 3.476799e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07959770  3.833668 1.262468e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55000504          0.7928355          0.5791642          0.7636763
#> 2         0.27282587          0.6742145          0.3210248          0.6260155
#> 3         0.08960813          0.5012425          0.1390374          0.4518133
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5752805          0.7542663          0.5856533          0.7478168
#> 2          0.2765891          0.6468927          0.2802514          0.6021884
#> 3          0.1043602          0.5065311          0.1333025          0.4711936