Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 607265549
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0470   14.2904    0.0000 [ 0.5800; 0.7501 ] 
#>   eta3 ~ eta1      0.4585      0.0924    4.9645    0.0000 [ 0.2801; 0.5933 ] 
#>   eta3 ~ eta2      0.3052      0.0987    3.0923    0.0020 [ 0.1627; 0.4792 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0443   14.9623    0.0000 [ 0.5987; 0.7465 ] 
#>   eta1 =~ y12      0.6493      0.0471   13.7947    0.0000 [ 0.5337; 0.7088 ] 
#>   eta1 =~ y13      0.7613      0.0360   21.1523    0.0000 [ 0.6976; 0.8283 ] 
#>   eta2 =~ y21      0.5165      0.0510   10.1218    0.0000 [ 0.4265; 0.5954 ] 
#>   eta2 =~ y22      0.7554      0.0280   26.9738    0.0000 [ 0.7165; 0.8048 ] 
#>   eta2 =~ y23      0.7997      0.0413   19.3748    0.0000 [ 0.7277; 0.8854 ] 
#>   eta3 =~ y31      0.8223      0.0347   23.7202    0.0000 [ 0.7569; 0.8810 ] 
#>   eta3 =~ y32      0.6581      0.0298   22.0647    0.0000 [ 0.6152; 0.7222 ] 
#>   eta3 =~ y33      0.7474      0.0423   17.6521    0.0000 [ 0.6693; 0.8108 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0217   18.2070    0.0000 [ 0.3666; 0.4412 ] 
#>   eta1 <~ y12      0.3873      0.0239   16.2052    0.0000 [ 0.3331; 0.4140 ] 
#>   eta1 <~ y13      0.4542      0.0247   18.3989    0.0000 [ 0.4160; 0.5066 ] 
#>   eta2 <~ y21      0.3058      0.0256   11.9235    0.0000 [ 0.2597; 0.3515 ] 
#>   eta2 <~ y22      0.4473      0.0232   19.2408    0.0000 [ 0.4076; 0.4940 ] 
#>   eta2 <~ y23      0.4735      0.0182   26.0853    0.0000 [ 0.4427; 0.5037 ] 
#>   eta3 <~ y31      0.4400      0.0185   23.7844    0.0000 [ 0.4045; 0.4743 ] 
#>   eta3 <~ y32      0.3521      0.0151   23.3853    0.0000 [ 0.3256; 0.3771 ] 
#>   eta3 <~ y33      0.3999      0.0204   19.6333    0.0000 [ 0.3602; 0.4351 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0470   14.2904    0.0000 [ 0.5800; 0.7501 ] 
#>   eta3 ~ eta1       0.6634      0.0365   18.1789    0.0000 [ 0.5945; 0.7146 ] 
#>   eta3 ~ eta2       0.3052      0.0987    3.0923    0.0020 [ 0.1627; 0.4792 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0685    2.9892    0.0028 [ 0.1118; 0.3151 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04431609 14.96228  1.295064e-50
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04706705 13.79474  2.741256e-43
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03599357 21.15227  2.629792e-99
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05102412 10.12178  4.423057e-24
#> 5 eta2 =~ y22  Common factor 0.7553877 0.02800450 26.97380 3.000212e-160
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04127349 19.37475  1.260628e-83
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03466564 23.72024 2.229461e-124
#> 8 eta3 =~ y32  Common factor 0.6580689 0.02982444 22.06475 6.894932e-108
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04234197 17.65208  9.806830e-70
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5987273          0.7465381
#> 2          0.5336923          0.7088450
#> 3          0.6975996          0.8282697
#> 4          0.4265011          0.5954283
#> 5          0.7165152          0.8047990
#> 6          0.7277098          0.8853634
#> 7          0.7569387          0.8809589
#> 8          0.6151969          0.7222342
#> 9          0.6692531          0.8107744

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 607265549
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0470   14.2904    0.0000 [ 0.5531; 0.7961 ] 
#>   eta3 ~ eta1      0.4585      0.0924    4.9645    0.0000 [ 0.2297; 0.7073 ] 
#>   eta3 ~ eta2      0.3052      0.0987    3.0923    0.0020 [ 0.0399; 0.5502 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0443   14.9623    0.0000 [ 0.5533; 0.7825 ] 
#>   eta1 =~ y12      0.6493      0.0471   13.7947    0.0000 [ 0.5394; 0.7828 ] 
#>   eta1 =~ y13      0.7613      0.0360   21.1523    0.0000 [ 0.6681; 0.8542 ] 
#>   eta2 =~ y21      0.5165      0.0510   10.1218    0.0000 [ 0.3974; 0.6613 ] 
#>   eta2 =~ y22      0.7554      0.0280   26.9738    0.0000 [ 0.6760; 0.8208 ] 
#>   eta2 =~ y23      0.7997      0.0413   19.3748    0.0000 [ 0.6936; 0.9071 ] 
#>   eta3 =~ y31      0.8223      0.0347   23.7202    0.0000 [ 0.7266; 0.9059 ] 
#>   eta3 =~ y32      0.6581      0.0298   22.0647    0.0000 [ 0.5735; 0.7277 ] 
#>   eta3 =~ y33      0.7474      0.0423   17.6521    0.0000 [ 0.6412; 0.8602 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0217   18.2070    0.0000 [ 0.3381; 0.4505 ] 
#>   eta1 <~ y12      0.3873      0.0239   16.2052    0.0000 [ 0.3286; 0.4522 ] 
#>   eta1 <~ y13      0.4542      0.0247   18.3989    0.0000 [ 0.3850; 0.5126 ] 
#>   eta2 <~ y21      0.3058      0.0256   11.9235    0.0000 [ 0.2468; 0.3794 ] 
#>   eta2 <~ y22      0.4473      0.0232   19.2408    0.0000 [ 0.3816; 0.5018 ] 
#>   eta2 <~ y23      0.4735      0.0182   26.0853    0.0000 [ 0.4260; 0.5199 ] 
#>   eta3 <~ y31      0.4400      0.0185   23.7844    0.0000 [ 0.3921; 0.4877 ] 
#>   eta3 <~ y32      0.3521      0.0151   23.3853    0.0000 [ 0.3118; 0.3896 ] 
#>   eta3 <~ y33      0.3999      0.0204   19.6333    0.0000 [ 0.3519; 0.4573 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0470   14.2904    0.0000 [ 0.5531; 0.7961 ] 
#>   eta3 ~ eta1       0.6634      0.0365   18.1789    0.0000 [ 0.5735; 0.7623 ] 
#>   eta3 ~ eta2       0.3052      0.0987    3.0923    0.0020 [ 0.0399; 0.5502 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0685    2.9892    0.0028 [ 0.0222; 0.3766 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04697778 14.290447 2.509919e-46
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09235699  4.964505 6.887641e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09868073  3.092307 1.986073e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55313475          0.7960773          0.5823074          0.7669046
#> 2         0.22965483          0.7072730          0.2870074          0.6499204
#> 3         0.03986134          0.5501824          0.1011409          0.4889028
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.56899820          0.7521986          0.5800169          0.7500733
#> 2         0.25179116          0.6357560          0.2800864          0.5933266
#> 3         0.09186248          0.5270581          0.1627132          0.4792480