Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 2120513541
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0509   13.1875    0.0000 [ 0.5881; 0.7572 ] 
#>   eta3 ~ eta1      0.4585      0.0853    5.3731    0.0000 [ 0.3277; 0.6160 ] 
#>   eta3 ~ eta2      0.3052      0.0957    3.1884    0.0014 [ 0.1136; 0.4371 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0329   20.1334    0.0000 [ 0.6143; 0.7223 ] 
#>   eta1 =~ y12      0.6493      0.0374   17.3396    0.0000 [ 0.5817; 0.7128 ] 
#>   eta1 =~ y13      0.7613      0.0304   25.0387    0.0000 [ 0.6883; 0.8023 ] 
#>   eta2 =~ y21      0.5165      0.0541    9.5427    0.0000 [ 0.4165; 0.6073 ] 
#>   eta2 =~ y22      0.7554      0.0344   21.9785    0.0000 [ 0.6624; 0.7985 ] 
#>   eta2 =~ y23      0.7997      0.0390   20.4973    0.0000 [ 0.7332; 0.8655 ] 
#>   eta3 =~ y31      0.8223      0.0323   25.4786    0.0000 [ 0.7653; 0.8720 ] 
#>   eta3 =~ y32      0.6581      0.0370   17.8058    0.0000 [ 0.6035; 0.7219 ] 
#>   eta3 =~ y33      0.7474      0.0366   20.4015    0.0000 [ 0.6549; 0.8037 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0189   20.9285    0.0000 [ 0.3636; 0.4334 ] 
#>   eta1 <~ y12      0.3873      0.0195   19.8949    0.0000 [ 0.3444; 0.4355 ] 
#>   eta1 <~ y13      0.4542      0.0157   29.0028    0.0000 [ 0.4253; 0.4793 ] 
#>   eta2 <~ y21      0.3058      0.0258   11.8573    0.0000 [ 0.2590; 0.3460 ] 
#>   eta2 <~ y22      0.4473      0.0219   20.4546    0.0000 [ 0.4066; 0.4893 ] 
#>   eta2 <~ y23      0.4735      0.0228   20.7888    0.0000 [ 0.4330; 0.5213 ] 
#>   eta3 <~ y31      0.4400      0.0182   24.1407    0.0000 [ 0.4126; 0.4711 ] 
#>   eta3 <~ y32      0.3521      0.0160   22.0695    0.0000 [ 0.3247; 0.3849 ] 
#>   eta3 <~ y33      0.3999      0.0178   22.5191    0.0000 [ 0.3571; 0.4237 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0509   13.1875    0.0000 [ 0.5881; 0.7572 ] 
#>   eta3 ~ eta1       0.6634      0.0391   16.9863    0.0000 [ 0.5840; 0.7273 ] 
#>   eta3 ~ eta2       0.3052      0.0957    3.1884    0.0014 [ 0.1136; 0.4371 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0677    3.0276    0.0025 [ 0.0831; 0.3246 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03293378 20.133430  3.760427e-90
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03744481 17.339600  2.363777e-67
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03040678 25.038682 2.319031e-138
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05412022  9.542732  1.391176e-21
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03436941 21.978486 4.626292e-107
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03901320 20.497259  2.277680e-93
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03227320 25.478647 3.400006e-143
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03695820 17.805764  6.375660e-71
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03663576 20.401489  1.621999e-92
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6142954          0.7223147
#> 2          0.5816607          0.7127592
#> 3          0.6883155          0.8023172
#> 4          0.4165455          0.6072670
#> 5          0.6624450          0.7984717
#> 6          0.7332021          0.8655050
#> 7          0.7653164          0.8719521
#> 8          0.6034709          0.7218771
#> 9          0.6549050          0.8037477

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 2120513541
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0509   13.1875    0.0000 [ 0.5275; 0.7908 ] 
#>   eta3 ~ eta1      0.4585      0.0853    5.3731    0.0000 [ 0.2236; 0.6649 ] 
#>   eta3 ~ eta2      0.3052      0.0957    3.1884    0.0014 [ 0.0721; 0.5671 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0329   20.1334    0.0000 [ 0.5726; 0.7429 ] 
#>   eta1 =~ y12      0.6493      0.0374   17.3396    0.0000 [ 0.5518; 0.7455 ] 
#>   eta1 =~ y13      0.7613      0.0304   25.0387    0.0000 [ 0.6893; 0.8465 ] 
#>   eta2 =~ y21      0.5165      0.0541    9.5427    0.0000 [ 0.3893; 0.6692 ] 
#>   eta2 =~ y22      0.7554      0.0344   21.9785    0.0000 [ 0.6711; 0.8488 ] 
#>   eta2 =~ y23      0.7997      0.0390   20.4973    0.0000 [ 0.6971; 0.8989 ] 
#>   eta3 =~ y31      0.8223      0.0323   25.4786    0.0000 [ 0.7389; 0.9058 ] 
#>   eta3 =~ y32      0.6581      0.0370   17.8058    0.0000 [ 0.5576; 0.7487 ] 
#>   eta3 =~ y33      0.7474      0.0366   20.4015    0.0000 [ 0.6626; 0.8521 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0189   20.9285    0.0000 [ 0.3434; 0.4411 ] 
#>   eta1 <~ y12      0.3873      0.0195   19.8949    0.0000 [ 0.3366; 0.4373 ] 
#>   eta1 <~ y13      0.4542      0.0157   29.0028    0.0000 [ 0.4175; 0.4985 ] 
#>   eta2 <~ y21      0.3058      0.0258   11.8573    0.0000 [ 0.2445; 0.3779 ] 
#>   eta2 <~ y22      0.4473      0.0219   20.4546    0.0000 [ 0.3892; 0.5022 ] 
#>   eta2 <~ y23      0.4735      0.0228   20.7888    0.0000 [ 0.4091; 0.5269 ] 
#>   eta3 <~ y31      0.4400      0.0182   24.1407    0.0000 [ 0.3913; 0.4856 ] 
#>   eta3 <~ y32      0.3521      0.0160   22.0695    0.0000 [ 0.3072; 0.3897 ] 
#>   eta3 <~ y33      0.3999      0.0178   22.5191    0.0000 [ 0.3580; 0.4499 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0509   13.1875    0.0000 [ 0.5275; 0.7908 ] 
#>   eta3 ~ eta1       0.6634      0.0391   16.9863    0.0000 [ 0.5544; 0.7564 ] 
#>   eta3 ~ eta2       0.3052      0.0957    3.1884    0.0014 [ 0.0721; 0.5671 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0677    3.0276    0.0025 [ 0.0362; 0.3861 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05090688 13.187480 1.035911e-39
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08533435  5.373062 7.741052e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09570766  3.188367 1.430789e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5275469          0.7908085          0.5591594          0.7591959
#> 2          0.2235870          0.6648880          0.2765786          0.6118964
#> 3          0.0721075          0.5670535          0.1315408          0.5076201
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.58307863          0.7680757          0.5881310          0.7572139
#> 2         0.32104846          0.6546734          0.3277450          0.6160288
#> 3         0.06298187          0.4376504          0.1136028          0.4370903