Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -372093516
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0385   17.4180    0.0000 [ 0.6144; 0.7470 ] 
#>   eta3 ~ eta1      0.4585      0.0815    5.6290    0.0000 [ 0.3458; 0.6424 ] 
#>   eta3 ~ eta2      0.3052      0.0814    3.7497    0.0002 [ 0.1444; 0.4320 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0429   15.4538    0.0000 [ 0.5925; 0.7303 ] 
#>   eta1 =~ y12      0.6493      0.0413   15.7373    0.0000 [ 0.5705; 0.7291 ] 
#>   eta1 =~ y13      0.7613      0.0337   22.5813    0.0000 [ 0.6981; 0.8073 ] 
#>   eta2 =~ y21      0.5165      0.0584    8.8508    0.0000 [ 0.4051; 0.6130 ] 
#>   eta2 =~ y22      0.7554      0.0390   19.3820    0.0000 [ 0.6908; 0.8336 ] 
#>   eta2 =~ y23      0.7997      0.0406   19.6794    0.0000 [ 0.7242; 0.8655 ] 
#>   eta3 =~ y31      0.8223      0.0271   30.3549    0.0000 [ 0.7771; 0.8691 ] 
#>   eta3 =~ y32      0.6581      0.0420   15.6607    0.0000 [ 0.5796; 0.7117 ] 
#>   eta3 =~ y33      0.7474      0.0348   21.5064    0.0000 [ 0.6862; 0.8183 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0256   15.4782    0.0000 [ 0.3418; 0.4330 ] 
#>   eta1 <~ y12      0.3873      0.0192   20.1809    0.0000 [ 0.3497; 0.4168 ] 
#>   eta1 <~ y13      0.4542      0.0203   22.4222    0.0000 [ 0.4192; 0.4897 ] 
#>   eta2 <~ y21      0.3058      0.0283   10.7905    0.0000 [ 0.2496; 0.3479 ] 
#>   eta2 <~ y22      0.4473      0.0267   16.7346    0.0000 [ 0.4067; 0.5124 ] 
#>   eta2 <~ y23      0.4735      0.0208   22.8005    0.0000 [ 0.4320; 0.5072 ] 
#>   eta3 <~ y31      0.4400      0.0180   24.4938    0.0000 [ 0.4034; 0.4659 ] 
#>   eta3 <~ y32      0.3521      0.0180   19.5197    0.0000 [ 0.3199; 0.3799 ] 
#>   eta3 <~ y33      0.3999      0.0162   24.6897    0.0000 [ 0.3748; 0.4323 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0385   17.4180    0.0000 [ 0.6144; 0.7470 ] 
#>   eta3 ~ eta1       0.6634      0.0413   16.0632    0.0000 [ 0.6056; 0.7416 ] 
#>   eta3 ~ eta2       0.3052      0.0814    3.7497    0.0002 [ 0.1444; 0.4320 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0565    3.6252    0.0003 [ 0.1051; 0.2872 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04290651 15.453829  7.109172e-54
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04125736 15.737264  8.399794e-56
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03371570 22.581347 6.610355e-113
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05835109  8.850817  8.688064e-19
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03897366 19.382006  1.094944e-83
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04063452 19.679419  3.236907e-86
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02708877 30.354918 2.163715e-202
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04202047 15.660676  2.808945e-55
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03475349 21.506446 1.354961e-102
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5925084          0.7302651
#> 2          0.5705079          0.7291449
#> 3          0.6981465          0.8073415
#> 4          0.4050749          0.6130392
#> 5          0.6908133          0.8335574
#> 6          0.7242249          0.8655389
#> 7          0.7771138          0.8691007
#> 8          0.5796124          0.7116859
#> 9          0.6861512          0.8183437

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -372093516
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0385   17.4180    0.0000 [ 0.5669; 0.7662 ] 
#>   eta3 ~ eta1      0.4585      0.0815    5.6290    0.0000 [ 0.2192; 0.6405 ] 
#>   eta3 ~ eta2      0.3052      0.0814    3.7497    0.0002 [ 0.1229; 0.5437 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0429   15.4538    0.0000 [ 0.5584; 0.7803 ] 
#>   eta1 =~ y12      0.6493      0.0413   15.7373    0.0000 [ 0.5450; 0.7583 ] 
#>   eta1 =~ y13      0.7613      0.0337   22.5813    0.0000 [ 0.6803; 0.8547 ] 
#>   eta2 =~ y21      0.5165      0.0584    8.8508    0.0000 [ 0.3801; 0.6819 ] 
#>   eta2 =~ y22      0.7554      0.0390   19.3820    0.0000 [ 0.6450; 0.8466 ] 
#>   eta2 =~ y23      0.7997      0.0406   19.6794    0.0000 [ 0.7018; 0.9119 ] 
#>   eta3 =~ y31      0.8223      0.0271   30.3549    0.0000 [ 0.7491; 0.8892 ] 
#>   eta3 =~ y32      0.6581      0.0420   15.6607    0.0000 [ 0.5502; 0.7675 ] 
#>   eta3 =~ y33      0.7474      0.0348   21.5064    0.0000 [ 0.6608; 0.8405 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0256   15.4782    0.0000 [ 0.3292; 0.4613 ] 
#>   eta1 <~ y12      0.3873      0.0192   20.1809    0.0000 [ 0.3354; 0.4347 ] 
#>   eta1 <~ y13      0.4542      0.0203   22.4222    0.0000 [ 0.4008; 0.5056 ] 
#>   eta2 <~ y21      0.3058      0.0283   10.7905    0.0000 [ 0.2398; 0.3864 ] 
#>   eta2 <~ y22      0.4473      0.0267   16.7346    0.0000 [ 0.3694; 0.5076 ] 
#>   eta2 <~ y23      0.4735      0.0208   22.8005    0.0000 [ 0.4212; 0.5286 ] 
#>   eta3 <~ y31      0.4400      0.0180   24.4938    0.0000 [ 0.3916; 0.4845 ] 
#>   eta3 <~ y32      0.3521      0.0180   19.5197    0.0000 [ 0.3060; 0.3993 ] 
#>   eta3 <~ y33      0.3999      0.0162   24.6897    0.0000 [ 0.3597; 0.4435 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0385   17.4180    0.0000 [ 0.5669; 0.7662 ] 
#>   eta3 ~ eta1       0.6634      0.0413   16.0632    0.0000 [ 0.5457; 0.7593 ] 
#>   eta3 ~ eta2       0.3052      0.0814    3.7497    0.0002 [ 0.1229; 0.5437 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0565    3.6252    0.0003 [ 0.0765; 0.3688 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03854249 17.418007 6.023875e-68
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08145476  5.628974 1.812842e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08137924  3.749742 1.770168e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5668880          0.7662080          0.5908225          0.7422736
#> 2          0.2192175          0.6404555          0.2697999          0.5898731
#> 3          0.1228998          0.5437472          0.1734353          0.4932117
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5944685          0.7689764          0.6143878          0.7469758
#> 2          0.3250200          0.6495781          0.3457909          0.6423589
#> 3          0.1364963          0.4359963          0.1443881          0.4319605