Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2074147291
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0439   15.2921    0.0000 [ 0.5978; 0.7347 ] 
#>   eta3 ~ eta1      0.4585      0.0835    5.4899    0.0000 [ 0.3082; 0.6300 ] 
#>   eta3 ~ eta2      0.3052      0.0870    3.5078    0.0005 [ 0.1496; 0.4607 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0387   17.1498    0.0000 [ 0.5997; 0.7273 ] 
#>   eta1 =~ y12      0.6493      0.0363   17.8833    0.0000 [ 0.5962; 0.7276 ] 
#>   eta1 =~ y13      0.7613      0.0376   20.2302    0.0000 [ 0.6905; 0.8328 ] 
#>   eta2 =~ y21      0.5165      0.0502   10.2915    0.0000 [ 0.4084; 0.5862 ] 
#>   eta2 =~ y22      0.7554      0.0339   22.2540    0.0000 [ 0.7062; 0.8133 ] 
#>   eta2 =~ y23      0.7997      0.0382   20.9336    0.0000 [ 0.7125; 0.8575 ] 
#>   eta3 =~ y31      0.8223      0.0341   24.1091    0.0000 [ 0.7602; 0.8710 ] 
#>   eta3 =~ y32      0.6581      0.0427   15.4100    0.0000 [ 0.5966; 0.7451 ] 
#>   eta3 =~ y33      0.7474      0.0353   21.1728    0.0000 [ 0.6898; 0.8205 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0209   18.9421    0.0000 [ 0.3555; 0.4249 ] 
#>   eta1 <~ y12      0.3873      0.0195   19.8869    0.0000 [ 0.3596; 0.4344 ] 
#>   eta1 <~ y13      0.4542      0.0207   21.9036    0.0000 [ 0.4034; 0.4800 ] 
#>   eta2 <~ y21      0.3058      0.0248   12.3392    0.0000 [ 0.2542; 0.3433 ] 
#>   eta2 <~ y22      0.4473      0.0214   20.9354    0.0000 [ 0.4194; 0.4889 ] 
#>   eta2 <~ y23      0.4735      0.0226   20.9610    0.0000 [ 0.4240; 0.5035 ] 
#>   eta3 <~ y31      0.4400      0.0173   25.3733    0.0000 [ 0.4034; 0.4628 ] 
#>   eta3 <~ y32      0.3521      0.0211   16.6699    0.0000 [ 0.3244; 0.4028 ] 
#>   eta3 <~ y33      0.3999      0.0182   21.9308    0.0000 [ 0.3785; 0.4408 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0439   15.2921    0.0000 [ 0.5978; 0.7347 ] 
#>   eta3 ~ eta1       0.6634      0.0371   17.8720    0.0000 [ 0.5871; 0.7306 ] 
#>   eta3 ~ eta2       0.3052      0.0870    3.5078    0.0005 [ 0.1496; 0.4607 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0598    3.4270    0.0006 [ 0.0995; 0.3103 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03866335 17.14983  6.303781e-66
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03630633 17.88333  1.590521e-71
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03763414 20.23019  5.309645e-91
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05018244 10.29154  7.693236e-25
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03394395 22.25397 1.032327e-109
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03819993 20.93364  2.644811e-97
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03410648 24.10912 2.005436e-128
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04270394 15.41003  1.401487e-53
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03530115 21.17280  1.701580e-99
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5996827          0.7272768
#> 2          0.5962208          0.7275638
#> 3          0.6904521          0.8327540
#> 4          0.4084442          0.5861557
#> 5          0.7062456          0.8133397
#> 6          0.7125426          0.8575348
#> 7          0.7602064          0.8710397
#> 8          0.5966332          0.7450698
#> 9          0.6898116          0.8204928

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2074147291
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0439   15.2921    0.0000 [ 0.5592; 0.7862 ] 
#>   eta3 ~ eta1      0.4585      0.0835    5.4899    0.0000 [ 0.2560; 0.6879 ] 
#>   eta3 ~ eta2      0.3052      0.0870    3.5078    0.0005 [ 0.0590; 0.5089 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0387   17.1498    0.0000 [ 0.5637; 0.7637 ] 
#>   eta1 =~ y12      0.6493      0.0363   17.8833    0.0000 [ 0.5407; 0.7284 ] 
#>   eta1 =~ y13      0.7613      0.0376   20.2302    0.0000 [ 0.6658; 0.8604 ] 
#>   eta2 =~ y21      0.5165      0.0502   10.2915    0.0000 [ 0.3903; 0.6499 ] 
#>   eta2 =~ y22      0.7554      0.0339   22.2540    0.0000 [ 0.6639; 0.8394 ] 
#>   eta2 =~ y23      0.7997      0.0382   20.9336    0.0000 [ 0.7044; 0.9019 ] 
#>   eta3 =~ y31      0.8223      0.0341   24.1091    0.0000 [ 0.7418; 0.9182 ] 
#>   eta3 =~ y32      0.6581      0.0427   15.4100    0.0000 [ 0.5422; 0.7630 ] 
#>   eta3 =~ y33      0.7474      0.0353   21.1728    0.0000 [ 0.6511; 0.8336 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0209   18.9421    0.0000 [ 0.3455; 0.4535 ] 
#>   eta1 <~ y12      0.3873      0.0195   19.8869    0.0000 [ 0.3318; 0.4325 ] 
#>   eta1 <~ y13      0.4542      0.0207   21.9036    0.0000 [ 0.4057; 0.5129 ] 
#>   eta2 <~ y21      0.3058      0.0248   12.3392    0.0000 [ 0.2439; 0.3720 ] 
#>   eta2 <~ y22      0.4473      0.0214   20.9354    0.0000 [ 0.3891; 0.4996 ] 
#>   eta2 <~ y23      0.4735      0.0226   20.9610    0.0000 [ 0.4165; 0.5333 ] 
#>   eta3 <~ y31      0.4400      0.0173   25.3733    0.0000 [ 0.4002; 0.4899 ] 
#>   eta3 <~ y32      0.3521      0.0211   16.6699    0.0000 [ 0.2954; 0.4046 ] 
#>   eta3 <~ y33      0.3999      0.0182   21.9308    0.0000 [ 0.3509; 0.4452 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0439   15.2921    0.0000 [ 0.5592; 0.7862 ] 
#>   eta3 ~ eta1       0.6634      0.0371   17.8720    0.0000 [ 0.5674; 0.7593 ] 
#>   eta3 ~ eta2       0.3052      0.0870    3.5078    0.0005 [ 0.0590; 0.5089 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0598    3.4270    0.0006 [ 0.0368; 0.3459 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04390054 15.292146 8.625959e-53
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08351787  5.489924 4.021066e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08699211  3.507802 4.518248e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55915433          0.7861831          0.5864160          0.7589214
#> 2         0.25603517          0.6879424          0.3078988          0.6360788
#> 3         0.05903515          0.5089092          0.1130562          0.4548882
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5835241          0.7596817          0.5978002          0.7346507
#> 2          0.2812156          0.6342363          0.3082338          0.6299882
#> 3          0.1402262          0.5000428          0.1495713          0.4607300