Summarize model

Usage

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 638270906
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0375   17.9007    0.0000 [ 0.5949; 0.7310 ] 
#>   eta3 ~ eta1      0.4585      0.0821    5.5872    0.0000 [ 0.3267; 0.6183 ] 
#>   eta3 ~ eta2      0.3052      0.0917    3.3273    0.0009 [ 0.1176; 0.4657 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0372   17.8306    0.0000 [ 0.5740; 0.7155 ] 
#>   eta1 =~ y12      0.6493      0.0346   18.7384    0.0000 [ 0.5988; 0.7115 ] 
#>   eta1 =~ y13      0.7613      0.0328   23.2058    0.0000 [ 0.7023; 0.8161 ] 
#>   eta2 =~ y21      0.5165      0.0406   12.7297    0.0000 [ 0.4387; 0.5861 ] 
#>   eta2 =~ y22      0.7554      0.0426   17.7188    0.0000 [ 0.6843; 0.8297 ] 
#>   eta2 =~ y23      0.7997      0.0354   22.5832    0.0000 [ 0.7384; 0.8720 ] 
#>   eta3 =~ y31      0.8223      0.0351   23.3963    0.0000 [ 0.7543; 0.8850 ] 
#>   eta3 =~ y32      0.6581      0.0442   14.8804    0.0000 [ 0.5850; 0.7254 ] 
#>   eta3 =~ y33      0.7474      0.0444   16.8193    0.0000 [ 0.6528; 0.8178 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0180   21.9626    0.0000 [ 0.3588; 0.4303 ] 
#>   eta1 <~ y12      0.3873      0.0188   20.6277    0.0000 [ 0.3629; 0.4250 ] 
#>   eta1 <~ y13      0.4542      0.0222   20.4431    0.0000 [ 0.4222; 0.4915 ] 
#>   eta2 <~ y21      0.3058      0.0218   13.9987    0.0000 [ 0.2643; 0.3422 ] 
#>   eta2 <~ y22      0.4473      0.0220   20.3023    0.0000 [ 0.4075; 0.4911 ] 
#>   eta2 <~ y23      0.4735      0.0202   23.3869    0.0000 [ 0.4229; 0.5035 ] 
#>   eta3 <~ y31      0.4400      0.0221   19.8946    0.0000 [ 0.4127; 0.4775 ] 
#>   eta3 <~ y32      0.3521      0.0191   18.4827    0.0000 [ 0.3164; 0.3900 ] 
#>   eta3 <~ y33      0.3999      0.0222   18.0352    0.0000 [ 0.3653; 0.4400 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0375   17.9007    0.0000 [ 0.5949; 0.7310 ] 
#>   eta3 ~ eta1       0.6634      0.0356   18.6532    0.0000 [ 0.6106; 0.7310 ] 
#>   eta3 ~ eta2       0.3052      0.0917    3.3273    0.0009 [ 0.1176; 0.4657 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0594    3.4464    0.0006 [ 0.0810; 0.3005 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03718715 17.83062  4.088845e-71
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03464968 18.73835  2.409681e-78
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03280842 23.20580 3.978487e-119
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04057076 12.72973  4.042332e-37
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04263200 17.71880  3.002736e-70
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03540971 22.58318 6.342347e-113
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03514561 23.39630 4.660618e-121
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04422394 14.88038  4.419982e-50
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04443859 16.81926  1.763479e-63
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5739712          0.7155120
#> 2          0.5987679          0.7115309
#> 3          0.7023480          0.8160653
#> 4          0.4387337          0.5861350
#> 5          0.6843453          0.8297053
#> 6          0.7383946          0.8720212
#> 7          0.7543228          0.8849582
#> 8          0.5849667          0.7254244
#> 9          0.6528478          0.8178385

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 638270906
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0375   17.9007    0.0000 [ 0.5878; 0.7817 ] 
#>   eta3 ~ eta1      0.4585      0.0821    5.5872    0.0000 [ 0.2323; 0.6566 ] 
#>   eta3 ~ eta2      0.3052      0.0917    3.3273    0.0009 [ 0.0722; 0.5465 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0372   17.8306    0.0000 [ 0.5704; 0.7628 ] 
#>   eta1 =~ y12      0.6493      0.0346   18.7384    0.0000 [ 0.5478; 0.7270 ] 
#>   eta1 =~ y13      0.7613      0.0328   23.2058    0.0000 [ 0.6806; 0.8502 ] 
#>   eta2 =~ y21      0.5165      0.0406   12.7297    0.0000 [ 0.4108; 0.6206 ] 
#>   eta2 =~ y22      0.7554      0.0426   17.7188    0.0000 [ 0.6372; 0.8577 ] 
#>   eta2 =~ y23      0.7997      0.0354   22.5832    0.0000 [ 0.7083; 0.8914 ] 
#>   eta3 =~ y31      0.8223      0.0351   23.3963    0.0000 [ 0.7362; 0.9179 ] 
#>   eta3 =~ y32      0.6581      0.0442   14.8804    0.0000 [ 0.5478; 0.7765 ] 
#>   eta3 =~ y33      0.7474      0.0444   16.8193    0.0000 [ 0.6385; 0.8683 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0180   21.9626    0.0000 [ 0.3526; 0.4457 ] 
#>   eta1 <~ y12      0.3873      0.0188   20.6277    0.0000 [ 0.3330; 0.4301 ] 
#>   eta1 <~ y13      0.4542      0.0222   20.4431    0.0000 [ 0.4005; 0.5154 ] 
#>   eta2 <~ y21      0.3058      0.0218   13.9987    0.0000 [ 0.2510; 0.3640 ] 
#>   eta2 <~ y22      0.4473      0.0220   20.3023    0.0000 [ 0.3888; 0.5027 ] 
#>   eta2 <~ y23      0.4735      0.0202   23.3869    0.0000 [ 0.4244; 0.5291 ] 
#>   eta3 <~ y31      0.4400      0.0221   19.8946    0.0000 [ 0.3809; 0.4953 ] 
#>   eta3 <~ y32      0.3521      0.0191   18.4827    0.0000 [ 0.3019; 0.4004 ] 
#>   eta3 <~ y33      0.3999      0.0222   18.0352    0.0000 [ 0.3420; 0.4566 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0375   17.9007    0.0000 [ 0.5878; 0.7817 ] 
#>   eta3 ~ eta1       0.6634      0.0356   18.6532    0.0000 [ 0.5650; 0.7489 ] 
#>   eta3 ~ eta2       0.3052      0.0917    3.3273    0.0009 [ 0.0722; 0.5465 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0594    3.4464    0.0006 [ 0.0588; 0.3662 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03750324 17.900679 1.164937e-71
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08206395  5.587188 2.307754e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09171182  3.327283 8.769742e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5877567          0.7817023          0.6110458          0.7584132
#> 2          0.2322516          0.6566400          0.2832123          0.6056793
#> 3          0.0722036          0.5464853          0.1291556          0.4895334
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.59153701          0.7419593          0.5949315          0.7310473
#> 2         0.28298818          0.6548072          0.3267250          0.6182729
#> 3         0.09838448          0.5027810          0.1176087          0.4657433