Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 860834401
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0447   15.0223    0.0000 [ 0.5816; 0.7408 ] 
#>   eta3 ~ eta1      0.4585      0.0770    5.9543    0.0000 [ 0.2588; 0.5734 ] 
#>   eta3 ~ eta2      0.3052      0.0813    3.7535    0.0002 [ 0.1699; 0.4904 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0439   15.1190    0.0000 [ 0.5675; 0.7329 ] 
#>   eta1 =~ y12      0.6493      0.0342   18.9748    0.0000 [ 0.5874; 0.7096 ] 
#>   eta1 =~ y13      0.7613      0.0319   23.8455    0.0000 [ 0.7027; 0.8181 ] 
#>   eta2 =~ y21      0.5165      0.0504   10.2472    0.0000 [ 0.4091; 0.5958 ] 
#>   eta2 =~ y22      0.7554      0.0356   21.2253    0.0000 [ 0.6926; 0.8012 ] 
#>   eta2 =~ y23      0.7997      0.0403   19.8542    0.0000 [ 0.7175; 0.8597 ] 
#>   eta3 =~ y31      0.8223      0.0337   24.4282    0.0000 [ 0.7536; 0.8783 ] 
#>   eta3 =~ y32      0.6581      0.0370   17.7660    0.0000 [ 0.5798; 0.7365 ] 
#>   eta3 =~ y33      0.7474      0.0388   19.2552    0.0000 [ 0.6889; 0.8364 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0214   18.5149    0.0000 [ 0.3533; 0.4241 ] 
#>   eta1 <~ y12      0.3873      0.0197   19.6226    0.0000 [ 0.3542; 0.4241 ] 
#>   eta1 <~ y13      0.4542      0.0204   22.2745    0.0000 [ 0.4163; 0.4915 ] 
#>   eta2 <~ y21      0.3058      0.0250   12.2441    0.0000 [ 0.2536; 0.3521 ] 
#>   eta2 <~ y22      0.4473      0.0252   17.7724    0.0000 [ 0.4053; 0.4942 ] 
#>   eta2 <~ y23      0.4735      0.0194   24.3837    0.0000 [ 0.4315; 0.4991 ] 
#>   eta3 <~ y31      0.4400      0.0144   30.5371    0.0000 [ 0.4024; 0.4607 ] 
#>   eta3 <~ y32      0.3521      0.0196   17.9625    0.0000 [ 0.3191; 0.3888 ] 
#>   eta3 <~ y33      0.3999      0.0204   19.6043    0.0000 [ 0.3680; 0.4382 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0447   15.0223    0.0000 [ 0.5816; 0.7408 ] 
#>   eta3 ~ eta1       0.6634      0.0345   19.2479    0.0000 [ 0.5942; 0.7205 ] 
#>   eta3 ~ eta2       0.3052      0.0813    3.7535    0.0002 [ 0.1699; 0.4904 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0574    3.5711    0.0004 [ 0.1124; 0.3511 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04385671 15.11901  1.213599e-51
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03421786 18.97482  2.754353e-80
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03192831 23.84548 1.128167e-125
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05039961 10.24720  1.218242e-24
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03558910 21.22525 5.582420e-100
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04027688 19.85416  1.014356e-87
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03366096 24.42822 8.576530e-132
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03704100 17.76596  1.296992e-70
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03881679 19.25518  1.277400e-82
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5675496          0.7329010
#> 2          0.5874376          0.7096420
#> 3          0.7026576          0.8180795
#> 4          0.4091229          0.5957726
#> 5          0.6925875          0.8011908
#> 6          0.7175093          0.8597101
#> 7          0.7536445          0.8782840
#> 8          0.5798206          0.7365363
#> 9          0.6888828          0.8364123

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 860834401
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0447   15.0223    0.0000 [ 0.5512; 0.7823 ] 
#>   eta3 ~ eta1      0.4585      0.0770    5.9543    0.0000 [ 0.2628; 0.6610 ] 
#>   eta3 ~ eta2      0.3052      0.0813    3.7535    0.0002 [ 0.0927; 0.5131 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0439   15.1190    0.0000 [ 0.5525; 0.7793 ] 
#>   eta1 =~ y12      0.6493      0.0342   18.9748    0.0000 [ 0.5553; 0.7323 ] 
#>   eta1 =~ y13      0.7613      0.0319   23.8455    0.0000 [ 0.6830; 0.8481 ] 
#>   eta2 =~ y21      0.5165      0.0504   10.2472    0.0000 [ 0.3840; 0.6446 ] 
#>   eta2 =~ y22      0.7554      0.0356   21.2253    0.0000 [ 0.6596; 0.8436 ] 
#>   eta2 =~ y23      0.7997      0.0403   19.8542    0.0000 [ 0.6965; 0.9047 ] 
#>   eta3 =~ y31      0.8223      0.0337   24.4282    0.0000 [ 0.7508; 0.9249 ] 
#>   eta3 =~ y32      0.6581      0.0370   17.7660    0.0000 [ 0.5529; 0.7445 ] 
#>   eta3 =~ y33      0.7474      0.0388   19.2552    0.0000 [ 0.6346; 0.8353 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0214   18.5149    0.0000 [ 0.3419; 0.4524 ] 
#>   eta1 <~ y12      0.3873      0.0197   19.6226    0.0000 [ 0.3327; 0.4348 ] 
#>   eta1 <~ y13      0.4542      0.0204   22.2745    0.0000 [ 0.4035; 0.5089 ] 
#>   eta2 <~ y21      0.3058      0.0250   12.2441    0.0000 [ 0.2416; 0.3708 ] 
#>   eta2 <~ y22      0.4473      0.0252   17.7724    0.0000 [ 0.3816; 0.5117 ] 
#>   eta2 <~ y23      0.4735      0.0194   24.3837    0.0000 [ 0.4260; 0.5264 ] 
#>   eta3 <~ y31      0.4400      0.0144   30.5371    0.0000 [ 0.4130; 0.4875 ] 
#>   eta3 <~ y32      0.3521      0.0196   17.9625    0.0000 [ 0.2979; 0.3993 ] 
#>   eta3 <~ y33      0.3999      0.0204   19.6043    0.0000 [ 0.3422; 0.4477 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0447   15.0223    0.0000 [ 0.5512; 0.7823 ] 
#>   eta3 ~ eta1       0.6634      0.0345   19.2479    0.0000 [ 0.5744; 0.7527 ] 
#>   eta3 ~ eta2       0.3052      0.0813    3.7535    0.0002 [ 0.0927; 0.5131 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0574    3.5711    0.0004 [ 0.0533; 0.3500 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04468913 15.022297 5.245811e-51
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07700444  5.954290 2.612036e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08129690  3.753539 1.743551e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55122554          0.7823325          0.5789770          0.7545811
#> 2         0.26279380          0.6610173          0.3106127          0.6131984
#> 3         0.09265566          0.5130773          0.1431401          0.4625929
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5699809          0.7410359          0.5815656          0.7407608
#> 2          0.2421689          0.5862481          0.2587652          0.5734490
#> 3          0.1329510          0.5344433          0.1699136          0.4903864