Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1100778720
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0485   13.8401    0.0000 [ 0.5953; 0.7640 ] 
#>   eta3 ~ eta1      0.4585      0.0747    6.1407    0.0000 [ 0.2953; 0.5755 ] 
#>   eta3 ~ eta2      0.3052      0.0708    4.3128    0.0000 [ 0.1462; 0.4468 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0460   14.4182    0.0000 [ 0.5752; 0.7439 ] 
#>   eta1 =~ y12      0.6493      0.0485   13.3753    0.0000 [ 0.5573; 0.7433 ] 
#>   eta1 =~ y13      0.7613      0.0337   22.5846    0.0000 [ 0.7012; 0.8354 ] 
#>   eta2 =~ y21      0.5165      0.0557    9.2782    0.0000 [ 0.3961; 0.6078 ] 
#>   eta2 =~ y22      0.7554      0.0312   24.1945    0.0000 [ 0.7066; 0.8191 ] 
#>   eta2 =~ y23      0.7997      0.0393   20.3516    0.0000 [ 0.7191; 0.8628 ] 
#>   eta3 =~ y31      0.8223      0.0361   22.7514    0.0000 [ 0.7566; 0.8991 ] 
#>   eta3 =~ y32      0.6581      0.0401   16.4275    0.0000 [ 0.5938; 0.7418 ] 
#>   eta3 =~ y33      0.7474      0.0406   18.4092    0.0000 [ 0.6699; 0.8119 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0255   15.5188    0.0000 [ 0.3456; 0.4525 ] 
#>   eta1 <~ y12      0.3873      0.0265   14.6411    0.0000 [ 0.3402; 0.4363 ] 
#>   eta1 <~ y13      0.4542      0.0190   23.8448    0.0000 [ 0.4231; 0.4966 ] 
#>   eta2 <~ y21      0.3058      0.0270   11.3115    0.0000 [ 0.2414; 0.3506 ] 
#>   eta2 <~ y22      0.4473      0.0208   21.5025    0.0000 [ 0.4147; 0.4822 ] 
#>   eta2 <~ y23      0.4735      0.0223   21.2148    0.0000 [ 0.4311; 0.5195 ] 
#>   eta3 <~ y31      0.4400      0.0212   20.7121    0.0000 [ 0.4033; 0.4836 ] 
#>   eta3 <~ y32      0.3521      0.0167   21.0797    0.0000 [ 0.3319; 0.3934 ] 
#>   eta3 <~ y33      0.3999      0.0222   17.9926    0.0000 [ 0.3671; 0.4355 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0485   13.8401    0.0000 [ 0.5953; 0.7640 ] 
#>   eta3 ~ eta1       0.6634      0.0432   15.3414    0.0000 [ 0.5669; 0.7278 ] 
#>   eta3 ~ eta2       0.3052      0.0708    4.3128    0.0000 [ 0.1462; 0.4468 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0485    4.2263    0.0000 [ 0.1020; 0.2758 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04598839 14.418201  3.975722e-47
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04854295 13.375330  8.427913e-41
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03371083 22.584605 6.140595e-113
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05566300  9.278242  1.722979e-20
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03122146 24.194504 2.541722e-129
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03929240 20.351612  4.492566e-92
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03614187 22.751378 1.390625e-114
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04005904 16.427475  1.216178e-60
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04060054 18.409215  1.108178e-75
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5751625          0.7438797
#> 2          0.5573351          0.7432839
#> 3          0.7012164          0.8354110
#> 4          0.3961267          0.6078286
#> 5          0.7065961          0.8190636
#> 6          0.7191399          0.8627750
#> 7          0.7566237          0.8990505
#> 8          0.5938170          0.7418200
#> 9          0.6699419          0.8119453

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1100778720
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0485   13.8401    0.0000 [ 0.5450; 0.7959 ] 
#>   eta3 ~ eta1      0.4585      0.0747    6.1407    0.0000 [ 0.2589; 0.6451 ] 
#>   eta3 ~ eta2      0.3052      0.0708    4.3128    0.0000 [ 0.1340; 0.4999 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0460   14.4182    0.0000 [ 0.5361; 0.7739 ] 
#>   eta1 =~ y12      0.6493      0.0485   13.3753    0.0000 [ 0.5313; 0.7823 ] 
#>   eta1 =~ y13      0.7613      0.0337   22.5846    0.0000 [ 0.6783; 0.8526 ] 
#>   eta2 =~ y21      0.5165      0.0557    9.2782    0.0000 [ 0.3790; 0.6669 ] 
#>   eta2 =~ y22      0.7554      0.0312   24.1945    0.0000 [ 0.6734; 0.8349 ] 
#>   eta2 =~ y23      0.7997      0.0393   20.3516    0.0000 [ 0.7060; 0.9092 ] 
#>   eta3 =~ y31      0.8223      0.0361   22.7514    0.0000 [ 0.7273; 0.9142 ] 
#>   eta3 =~ y32      0.6581      0.0401   16.4275    0.0000 [ 0.5544; 0.7616 ] 
#>   eta3 =~ y33      0.7474      0.0406   18.4092    0.0000 [ 0.6477; 0.8576 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0255   15.5188    0.0000 [ 0.3241; 0.4559 ] 
#>   eta1 <~ y12      0.3873      0.0265   14.6411    0.0000 [ 0.3228; 0.4596 ] 
#>   eta1 <~ y13      0.4542      0.0190   23.8448    0.0000 [ 0.4065; 0.5050 ] 
#>   eta2 <~ y21      0.3058      0.0270   11.3115    0.0000 [ 0.2378; 0.3776 ] 
#>   eta2 <~ y22      0.4473      0.0208   21.5025    0.0000 [ 0.3889; 0.4965 ] 
#>   eta2 <~ y23      0.4735      0.0223   21.2148    0.0000 [ 0.4166; 0.5320 ] 
#>   eta3 <~ y31      0.4400      0.0212   20.7121    0.0000 [ 0.3832; 0.4931 ] 
#>   eta3 <~ y32      0.3521      0.0167   21.0797    0.0000 [ 0.3084; 0.3948 ] 
#>   eta3 <~ y33      0.3999      0.0222   17.9926    0.0000 [ 0.3444; 0.4593 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0485   13.8401    0.0000 [ 0.5450; 0.7959 ] 
#>   eta3 ~ eta1       0.6634      0.0432   15.3414    0.0000 [ 0.5531; 0.7767 ] 
#>   eta3 ~ eta2       0.3052      0.0708    4.3128    0.0000 [ 0.1340; 0.4999 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0485    4.2263    0.0000 [ 0.0875; 0.3382 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04850653 13.840063 1.460792e-43
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07466627  6.140748 8.213379e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07075467  4.312805 1.611962e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5450101          0.7958584          0.5751320          0.7657365
#> 2          0.2589428          0.6450746          0.3053097          0.5987077
#> 3          0.1340020          0.4999052          0.1779398          0.4559674
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5593022          0.7727300          0.5952539          0.7640450
#> 2          0.2402462          0.5879463          0.2953386          0.5755283
#> 3          0.1306540          0.4474538          0.1462475          0.4468228