Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1139732630
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0429   15.6653    0.0000 [ 0.5671; 0.7248 ] 
#>   eta3 ~ eta1      0.4585      0.0746    6.1445    0.0000 [ 0.3274; 0.5910 ] 
#>   eta3 ~ eta2      0.3052      0.0780    3.9138    0.0001 [ 0.1948; 0.4603 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0466   14.2215    0.0000 [ 0.6167; 0.7634 ] 
#>   eta1 =~ y12      0.6493      0.0373   17.3855    0.0000 [ 0.5984; 0.7164 ] 
#>   eta1 =~ y13      0.7613      0.0335   22.7191    0.0000 [ 0.7182; 0.8157 ] 
#>   eta2 =~ y21      0.5165      0.0480   10.7639    0.0000 [ 0.4331; 0.5926 ] 
#>   eta2 =~ y22      0.7554      0.0412   18.3313    0.0000 [ 0.6360; 0.8213 ] 
#>   eta2 =~ y23      0.7997      0.0349   22.9062    0.0000 [ 0.7399; 0.8629 ] 
#>   eta3 =~ y31      0.8223      0.0407   20.2227    0.0000 [ 0.7277; 0.8740 ] 
#>   eta3 =~ y32      0.6581      0.0332   19.8219    0.0000 [ 0.5943; 0.7096 ] 
#>   eta3 =~ y33      0.7474      0.0344   21.7107    0.0000 [ 0.6783; 0.8050 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0221   17.8998    0.0000 [ 0.3612; 0.4447 ] 
#>   eta1 <~ y12      0.3873      0.0210   18.4798    0.0000 [ 0.3469; 0.4231 ] 
#>   eta1 <~ y13      0.4542      0.0203   22.3504    0.0000 [ 0.4090; 0.4823 ] 
#>   eta2 <~ y21      0.3058      0.0267   11.4430    0.0000 [ 0.2478; 0.3623 ] 
#>   eta2 <~ y22      0.4473      0.0200   22.4140    0.0000 [ 0.3985; 0.4725 ] 
#>   eta2 <~ y23      0.4735      0.0212   22.3864    0.0000 [ 0.4377; 0.5113 ] 
#>   eta3 <~ y31      0.4400      0.0180   24.4374    0.0000 [ 0.4041; 0.4648 ] 
#>   eta3 <~ y32      0.3521      0.0164   21.5275    0.0000 [ 0.3302; 0.3864 ] 
#>   eta3 <~ y33      0.3999      0.0203   19.7188    0.0000 [ 0.3659; 0.4377 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0429   15.6653    0.0000 [ 0.5671; 0.7248 ] 
#>   eta3 ~ eta1       0.6634      0.0365   18.1697    0.0000 [ 0.5962; 0.7183 ] 
#>   eta3 ~ eta2       0.3052      0.0780    3.9138    0.0001 [ 0.1948; 0.4603 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0538    3.8072    0.0001 [ 0.1194; 0.3032 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04662453 14.22148  6.741162e-46
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03734589 17.38552  1.062108e-67
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03351124 22.71912 2.899761e-114
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04798024 10.76391  5.096258e-27
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04120764 18.33125  4.660463e-75
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03491038 22.90619 4.031244e-116
#> 7 eta3 =~ y31  Common factor 0.8222773 0.04066103 20.22274  6.175936e-91
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03319912 19.82188  1.927889e-87
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03442647 21.71074 1.624335e-104
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6166563          0.7633750
#> 2          0.5984012          0.7164425
#> 3          0.7182001          0.8156889
#> 4          0.4330744          0.5926399
#> 5          0.6360308          0.8212728
#> 6          0.7398697          0.8628810
#> 7          0.7277081          0.8739575
#> 8          0.5942781          0.7095533
#> 9          0.6783346          0.8049959

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1139732630
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0429   15.6653    0.0000 [ 0.5736; 0.7953 ] 
#>   eta3 ~ eta1      0.4585      0.0746    6.1445    0.0000 [ 0.2803; 0.6661 ] 
#>   eta3 ~ eta2      0.3052      0.0780    3.9138    0.0001 [ 0.0808; 0.4840 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0466   14.2215    0.0000 [ 0.5350; 0.7761 ] 
#>   eta1 =~ y12      0.6493      0.0373   17.3855    0.0000 [ 0.5455; 0.7386 ] 
#>   eta1 =~ y13      0.7613      0.0335   22.7191    0.0000 [ 0.6718; 0.8451 ] 
#>   eta2 =~ y21      0.5165      0.0480   10.7639    0.0000 [ 0.3746; 0.6228 ] 
#>   eta2 =~ y22      0.7554      0.0412   18.3313    0.0000 [ 0.6535; 0.8666 ] 
#>   eta2 =~ y23      0.7997      0.0349   22.9062    0.0000 [ 0.7100; 0.8905 ] 
#>   eta3 =~ y31      0.8223      0.0407   20.2227    0.0000 [ 0.7289; 0.9392 ] 
#>   eta3 =~ y32      0.6581      0.0332   19.8219    0.0000 [ 0.5710; 0.7427 ] 
#>   eta3 =~ y33      0.7474      0.0344   21.7107    0.0000 [ 0.6564; 0.8345 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0221   17.8998    0.0000 [ 0.3394; 0.4536 ] 
#>   eta1 <~ y12      0.3873      0.0210   18.4798    0.0000 [ 0.3338; 0.4422 ] 
#>   eta1 <~ y13      0.4542      0.0203   22.3504    0.0000 [ 0.4057; 0.5107 ] 
#>   eta2 <~ y21      0.3058      0.0267   11.4430    0.0000 [ 0.2288; 0.3670 ] 
#>   eta2 <~ y22      0.4473      0.0200   22.4140    0.0000 [ 0.4021; 0.5053 ] 
#>   eta2 <~ y23      0.4735      0.0212   22.3864    0.0000 [ 0.4228; 0.5322 ] 
#>   eta3 <~ y31      0.4400      0.0180   24.4374    0.0000 [ 0.3973; 0.4905 ] 
#>   eta3 <~ y32      0.3521      0.0164   21.5275    0.0000 [ 0.3072; 0.3918 ] 
#>   eta3 <~ y33      0.3999      0.0203   19.7188    0.0000 [ 0.3440; 0.4489 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0429   15.6653    0.0000 [ 0.5736; 0.7953 ] 
#>   eta3 ~ eta1       0.6634      0.0365   18.1697    0.0000 [ 0.5725; 0.7613 ] 
#>   eta3 ~ eta2       0.3052      0.0780    3.9138    0.0001 [ 0.0808; 0.4840 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0538    3.8072    0.0001 [ 0.0546; 0.3328 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04285487 15.665279 2.612799e-55
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07462071  6.144498 8.021693e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07796753  3.913823 9.084611e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.57364008          0.7952612          0.6002524          0.7686489
#> 2         0.28025067          0.6661468          0.3265893          0.6198083
#> 3         0.08075556          0.4839596          0.1291725          0.4355427
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5652256          0.7398604          0.5671120          0.7248001
#> 2          0.3202377          0.6184733          0.3274358          0.5910180
#> 3          0.1807908          0.4646645          0.1948099          0.4602830