Summarize model

stable

Usage

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1462191717
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0435   15.4159    0.0000 [ 0.5968; 0.7403 ] 
#>   eta3 ~ eta1      0.4585      0.0800    5.7346    0.0000 [ 0.3470; 0.6036 ] 
#>   eta3 ~ eta2      0.3052      0.0927    3.2920    0.0010 [ 0.1439; 0.4674 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0398   16.6449    0.0000 [ 0.5915; 0.7246 ] 
#>   eta1 =~ y12      0.6493      0.0425   15.2758    0.0000 [ 0.5727; 0.7161 ] 
#>   eta1 =~ y13      0.7613      0.0369   20.6414    0.0000 [ 0.6828; 0.8325 ] 
#>   eta2 =~ y21      0.5165      0.0465   11.0965    0.0000 [ 0.4277; 0.5780 ] 
#>   eta2 =~ y22      0.7554      0.0354   21.3139    0.0000 [ 0.6772; 0.8123 ] 
#>   eta2 =~ y23      0.7997      0.0353   22.6701    0.0000 [ 0.7256; 0.8392 ] 
#>   eta3 =~ y31      0.8223      0.0395   20.8372    0.0000 [ 0.7435; 0.8875 ] 
#>   eta3 =~ y32      0.6581      0.0421   15.6429    0.0000 [ 0.5716; 0.7473 ] 
#>   eta3 =~ y33      0.7474      0.0471   15.8722    0.0000 [ 0.6504; 0.8223 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0203   19.5049    0.0000 [ 0.3587; 0.4293 ] 
#>   eta1 <~ y12      0.3873      0.0224   17.2953    0.0000 [ 0.3476; 0.4230 ] 
#>   eta1 <~ y13      0.4542      0.0234   19.4441    0.0000 [ 0.4096; 0.4965 ] 
#>   eta2 <~ y21      0.3058      0.0244   12.5388    0.0000 [ 0.2576; 0.3432 ] 
#>   eta2 <~ y22      0.4473      0.0204   21.9389    0.0000 [ 0.4277; 0.4923 ] 
#>   eta2 <~ y23      0.4735      0.0207   22.8643    0.0000 [ 0.4396; 0.5137 ] 
#>   eta3 <~ y31      0.4400      0.0233   18.8737    0.0000 [ 0.3946; 0.4880 ] 
#>   eta3 <~ y32      0.3521      0.0201   17.5009    0.0000 [ 0.3241; 0.4026 ] 
#>   eta3 <~ y33      0.3999      0.0220   18.1701    0.0000 [ 0.3562; 0.4397 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0435   15.4159    0.0000 [ 0.5968; 0.7403 ] 
#>   eta3 ~ eta1       0.6634      0.0410   16.1799    0.0000 [ 0.5905; 0.7326 ] 
#>   eta3 ~ eta2       0.3052      0.0927    3.2920    0.0010 [ 0.1439; 0.4674 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0657    3.1162    0.0018 [ 0.0931; 0.3296 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03983613 16.64494  3.293183e-62
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04250371 15.27579  1.108683e-52
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03688440 20.64141  1.166204e-94
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04654205 11.09652  1.304235e-28
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03544103 21.31393 8.430800e-101
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03527393 22.67010 8.838762e-114
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03946199 20.83720  1.991573e-96
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04206824 15.64289  3.714789e-55
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04709016 15.87219  9.873164e-57
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5914768          0.7245785
#> 2          0.5727023          0.7161241
#> 3          0.6827650          0.8325099
#> 4          0.4277471          0.5779821
#> 5          0.6772389          0.8122941
#> 6          0.7255711          0.8391807
#> 7          0.7434549          0.8875284
#> 8          0.5716048          0.7472516
#> 9          0.6504197          0.8222683

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1462191717
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0435   15.4159    0.0000 [ 0.5587; 0.7839 ] 
#>   eta3 ~ eta1      0.4585      0.0800    5.7346    0.0000 [ 0.2512; 0.6646 ] 
#>   eta3 ~ eta2      0.3052      0.0927    3.2920    0.0010 [ 0.0749; 0.5543 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0398   16.6449    0.0000 [ 0.5663; 0.7723 ] 
#>   eta1 =~ y12      0.6493      0.0425   15.2758    0.0000 [ 0.5429; 0.7627 ] 
#>   eta1 =~ y13      0.7613      0.0369   20.6414    0.0000 [ 0.6638; 0.8545 ] 
#>   eta2 =~ y21      0.5165      0.0465   11.0965    0.0000 [ 0.4078; 0.6485 ] 
#>   eta2 =~ y22      0.7554      0.0354   21.3139    0.0000 [ 0.6600; 0.8433 ] 
#>   eta2 =~ y23      0.7997      0.0353   22.6701    0.0000 [ 0.7215; 0.9039 ] 
#>   eta3 =~ y31      0.8223      0.0395   20.8372    0.0000 [ 0.7185; 0.9226 ] 
#>   eta3 =~ y32      0.6581      0.0421   15.6429    0.0000 [ 0.5412; 0.7587 ] 
#>   eta3 =~ y33      0.7474      0.0471   15.8722    0.0000 [ 0.6341; 0.8776 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0203   19.5049    0.0000 [ 0.3451; 0.4500 ] 
#>   eta1 <~ y12      0.3873      0.0224   17.2953    0.0000 [ 0.3298; 0.4457 ] 
#>   eta1 <~ y13      0.4542      0.0234   19.4441    0.0000 [ 0.3902; 0.5110 ] 
#>   eta2 <~ y21      0.3058      0.0244   12.5388    0.0000 [ 0.2458; 0.3719 ] 
#>   eta2 <~ y22      0.4473      0.0204   21.9389    0.0000 [ 0.3862; 0.4916 ] 
#>   eta2 <~ y23      0.4735      0.0207   22.8643    0.0000 [ 0.4213; 0.5284 ] 
#>   eta3 <~ y31      0.4400      0.0233   18.8737    0.0000 [ 0.3792; 0.4997 ] 
#>   eta3 <~ y32      0.3521      0.0201   17.5009    0.0000 [ 0.2963; 0.4004 ] 
#>   eta3 <~ y33      0.3999      0.0220   18.1701    0.0000 [ 0.3482; 0.4620 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0435   15.4159    0.0000 [ 0.5587; 0.7839 ] 
#>   eta3 ~ eta1       0.6634      0.0410   16.1799    0.0000 [ 0.5625; 0.7745 ] 
#>   eta3 ~ eta2       0.3052      0.0927    3.2920    0.0010 [ 0.0749; 0.5543 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0657    3.1162    0.0018 [ 0.0406; 0.3806 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04354812 15.415901 1.279744e-53
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07995405  5.734629 9.772617e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09269524  3.291983 9.948369e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55866037          0.7838666          0.5857032          0.7568238
#> 2         0.25116603          0.6646432          0.3008166          0.6149927
#> 3         0.07494495          0.5543124          0.1325076          0.4967497
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5750647          0.7461736          0.5968386          0.7402552
#> 2          0.2791679          0.6224163          0.3469754          0.6036299
#> 3          0.1293563          0.4931034          0.1439083          0.4673877