Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1023848257
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0423   15.8894    0.0000 [ 0.5768; 0.7651 ] 
#>   eta3 ~ eta1      0.4585      0.0751    6.1081    0.0000 [ 0.3452; 0.5973 ] 
#>   eta3 ~ eta2      0.3052      0.0840    3.6348    0.0003 [ 0.1400; 0.4314 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0374   17.7195    0.0000 [ 0.5774; 0.7155 ] 
#>   eta1 =~ y12      0.6493      0.0367   17.6941    0.0000 [ 0.5701; 0.7016 ] 
#>   eta1 =~ y13      0.7613      0.0327   23.2693    0.0000 [ 0.6853; 0.8122 ] 
#>   eta2 =~ y21      0.5165      0.0617    8.3673    0.0000 [ 0.4009; 0.6159 ] 
#>   eta2 =~ y22      0.7554      0.0364   20.7270    0.0000 [ 0.7043; 0.8315 ] 
#>   eta2 =~ y23      0.7997      0.0337   23.7161    0.0000 [ 0.7338; 0.8445 ] 
#>   eta3 =~ y31      0.8223      0.0399   20.5970    0.0000 [ 0.7466; 0.8887 ] 
#>   eta3 =~ y32      0.6581      0.0395   16.6456    0.0000 [ 0.5937; 0.7147 ] 
#>   eta3 =~ y33      0.7474      0.0391   19.1163    0.0000 [ 0.6874; 0.8295 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0209   18.9558    0.0000 [ 0.3556; 0.4270 ] 
#>   eta1 <~ y12      0.3873      0.0184   21.0548    0.0000 [ 0.3590; 0.4207 ] 
#>   eta1 <~ y13      0.4542      0.0188   24.1484    0.0000 [ 0.4311; 0.4923 ] 
#>   eta2 <~ y21      0.3058      0.0313    9.7858    0.0000 [ 0.2518; 0.3567 ] 
#>   eta2 <~ y22      0.4473      0.0197   22.7595    0.0000 [ 0.4237; 0.4891 ] 
#>   eta2 <~ y23      0.4735      0.0245   19.3140    0.0000 [ 0.4393; 0.5155 ] 
#>   eta3 <~ y31      0.4400      0.0203   21.6352    0.0000 [ 0.4011; 0.4690 ] 
#>   eta3 <~ y32      0.3521      0.0201   17.5142    0.0000 [ 0.3204; 0.3933 ] 
#>   eta3 <~ y33      0.3999      0.0197   20.3213    0.0000 [ 0.3715; 0.4307 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0423   15.8894    0.0000 [ 0.5768; 0.7651 ] 
#>   eta3 ~ eta1       0.6634      0.0353   18.8124    0.0000 [ 0.5789; 0.7197 ] 
#>   eta3 ~ eta2       0.3052      0.0840    3.6348    0.0003 [ 0.1400; 0.4314 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0560    3.6605    0.0003 [ 0.0863; 0.2896 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03742035 17.71950  2.965462e-70
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03669454 17.69413  4.653819e-70
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03271890 23.26930 9.072991e-120
#> 4 eta2 =~ y21  Common factor 0.5164548 0.06172307  8.36729  5.895837e-17
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03644463 20.72699  1.977610e-95
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03371820 23.71608 2.460849e-124
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03992226 20.59697  2.921858e-94
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03953405 16.64562  3.255571e-62
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03909875 19.11632  1.847049e-81
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5773873          0.7154671
#> 2          0.5701387          0.7016160
#> 3          0.6853312          0.8121984
#> 4          0.4008759          0.6159223
#> 5          0.7042611          0.8315402
#> 6          0.7338151          0.8444661
#> 7          0.7465828          0.8887261
#> 8          0.5937442          0.7147315
#> 9          0.6874312          0.8295418

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1023848257
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0423   15.8894    0.0000 [ 0.5700; 0.7885 ] 
#>   eta3 ~ eta1      0.4585      0.0751    6.1081    0.0000 [ 0.2773; 0.6655 ] 
#>   eta3 ~ eta2      0.3052      0.0840    3.6348    0.0003 [ 0.0685; 0.5027 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0374   17.7195    0.0000 [ 0.5780; 0.7716 ] 
#>   eta1 =~ y12      0.6493      0.0367   17.6941    0.0000 [ 0.5572; 0.7470 ] 
#>   eta1 =~ y13      0.7613      0.0327   23.2693    0.0000 [ 0.6775; 0.8467 ] 
#>   eta2 =~ y21      0.5165      0.0617    8.3673    0.0000 [ 0.3662; 0.6854 ] 
#>   eta2 =~ y22      0.7554      0.0364   20.7270    0.0000 [ 0.6620; 0.8505 ] 
#>   eta2 =~ y23      0.7997      0.0337   23.7161    0.0000 [ 0.7267; 0.9010 ] 
#>   eta3 =~ y31      0.8223      0.0399   20.5970    0.0000 [ 0.7214; 0.9278 ] 
#>   eta3 =~ y32      0.6581      0.0395   16.6456    0.0000 [ 0.5526; 0.7570 ] 
#>   eta3 =~ y33      0.7474      0.0391   19.1163    0.0000 [ 0.6469; 0.8491 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0209   18.9558    0.0000 [ 0.3445; 0.4524 ] 
#>   eta1 <~ y12      0.3873      0.0184   21.0548    0.0000 [ 0.3375; 0.4326 ] 
#>   eta1 <~ y13      0.4542      0.0188   24.1484    0.0000 [ 0.4012; 0.4985 ] 
#>   eta2 <~ y21      0.3058      0.0313    9.7858    0.0000 [ 0.2263; 0.3879 ] 
#>   eta2 <~ y22      0.4473      0.0197   22.7595    0.0000 [ 0.3898; 0.4914 ] 
#>   eta2 <~ y23      0.4735      0.0245   19.3140    0.0000 [ 0.4107; 0.5375 ] 
#>   eta3 <~ y31      0.4400      0.0203   21.6352    0.0000 [ 0.3889; 0.4941 ] 
#>   eta3 <~ y32      0.3521      0.0201   17.5142    0.0000 [ 0.2986; 0.4026 ] 
#>   eta3 <~ y33      0.3999      0.0197   20.3213    0.0000 [ 0.3496; 0.4514 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0423   15.8894    0.0000 [ 0.5700; 0.7885 ] 
#>   eta3 ~ eta1       0.6634      0.0353   18.8124    0.0000 [ 0.5748; 0.7572 ] 
#>   eta3 ~ eta2       0.3052      0.0840    3.6348    0.0003 [ 0.0685; 0.5027 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0560    3.6605    0.0003 [ 0.0499; 0.3393 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04225038 15.889407 7.503433e-57
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07506505  6.108126 1.008080e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08395371  3.634755 2.782453e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56998941          0.7884845          0.5962264          0.7622475
#> 2         0.27732375          0.6655178          0.3239383          0.6189033
#> 3         0.06850645          0.5026676          0.1206407          0.4505333
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5701423          0.7691790          0.5767682          0.7651294
#> 2          0.3066631          0.5978756          0.3452339          0.5972606
#> 3          0.1276672          0.4692246          0.1400210          0.4314221