Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1951756845
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0501   13.4065    0.0000 [ 0.5616; 0.7580 ] 
#>   eta3 ~ eta1      0.4585      0.0878    5.2226    0.0000 [ 0.3112; 0.5865 ] 
#>   eta3 ~ eta2      0.3052      0.0911    3.3497    0.0008 [ 0.1466; 0.4466 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0396   16.7590    0.0000 [ 0.6039; 0.7536 ] 
#>   eta1 =~ y12      0.6493      0.0365   17.7804    0.0000 [ 0.5859; 0.7218 ] 
#>   eta1 =~ y13      0.7613      0.0295   25.7957    0.0000 [ 0.7209; 0.8225 ] 
#>   eta2 =~ y21      0.5165      0.0484   10.6721    0.0000 [ 0.4328; 0.5897 ] 
#>   eta2 =~ y22      0.7554      0.0372   20.3030    0.0000 [ 0.6926; 0.8302 ] 
#>   eta2 =~ y23      0.7997      0.0234   34.1616    0.0000 [ 0.7480; 0.8402 ] 
#>   eta3 =~ y31      0.8223      0.0247   33.3073    0.0000 [ 0.7861; 0.8768 ] 
#>   eta3 =~ y32      0.6581      0.0463   14.2107    0.0000 [ 0.5789; 0.7438 ] 
#>   eta3 =~ y33      0.7474      0.0419   17.8245    0.0000 [ 0.6636; 0.8234 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0194   20.3504    0.0000 [ 0.3529; 0.4308 ] 
#>   eta1 <~ y12      0.3873      0.0204   19.0071    0.0000 [ 0.3489; 0.4132 ] 
#>   eta1 <~ y13      0.4542      0.0172   26.3534    0.0000 [ 0.4307; 0.4973 ] 
#>   eta2 <~ y21      0.3058      0.0258   11.8506    0.0000 [ 0.2630; 0.3534 ] 
#>   eta2 <~ y22      0.4473      0.0190   23.5873    0.0000 [ 0.4196; 0.4939 ] 
#>   eta2 <~ y23      0.4735      0.0179   26.4391    0.0000 [ 0.4388; 0.4989 ] 
#>   eta3 <~ y31      0.4400      0.0153   28.7256    0.0000 [ 0.4197; 0.4662 ] 
#>   eta3 <~ y32      0.3521      0.0215   16.4122    0.0000 [ 0.3095; 0.3903 ] 
#>   eta3 <~ y33      0.3999      0.0218   18.3874    0.0000 [ 0.3596; 0.4401 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0501   13.4065    0.0000 [ 0.5616; 0.7580 ] 
#>   eta3 ~ eta1       0.6634      0.0411   16.1396    0.0000 [ 0.5998; 0.7310 ] 
#>   eta3 ~ eta2       0.3052      0.0911    3.3497    0.0008 [ 0.1466; 0.4466 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0619    3.3110    0.0009 [ 0.1062; 0.3098 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03956489 16.75905  4.863529e-63
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03651653 17.78038  1.002922e-70
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02951448 25.79567 9.916203e-147
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04839292 10.67211  1.374623e-26
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03720575 20.30298  1.210096e-91
#> 6 eta2 =~ y23  Common factor 0.7996637 0.02340823 34.16164 8.981034e-256
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02468762 33.30727 3.029734e-243
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04630795 14.21071  7.862211e-46
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04193243 17.82449  4.562480e-71
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6039016          0.7536282
#> 2          0.5859205          0.7218367
#> 3          0.7209197          0.8225334
#> 4          0.4327620          0.5896750
#> 5          0.6926440          0.8301744
#> 6          0.7480152          0.8401997
#> 7          0.7861476          0.8768379
#> 8          0.5789308          0.7437935
#> 9          0.6636369          0.8234134

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1951756845
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0501   13.4065    0.0000 [ 0.5374; 0.7964 ] 
#>   eta3 ~ eta1      0.4585      0.0878    5.2226    0.0000 [ 0.2308; 0.6848 ] 
#>   eta3 ~ eta2      0.3052      0.0911    3.3497    0.0008 [ 0.0758; 0.5469 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0396   16.7590    0.0000 [ 0.5578; 0.7624 ] 
#>   eta1 =~ y12      0.6493      0.0365   17.7804    0.0000 [ 0.5622; 0.7510 ] 
#>   eta1 =~ y13      0.7613      0.0295   25.7957    0.0000 [ 0.6810; 0.8336 ] 
#>   eta2 =~ y21      0.5165      0.0484   10.6721    0.0000 [ 0.3844; 0.6346 ] 
#>   eta2 =~ y22      0.7554      0.0372   20.3030    0.0000 [ 0.6549; 0.8473 ] 
#>   eta2 =~ y23      0.7997      0.0234   34.1616    0.0000 [ 0.7528; 0.8739 ] 
#>   eta3 =~ y31      0.8223      0.0247   33.3073    0.0000 [ 0.7541; 0.8818 ] 
#>   eta3 =~ y32      0.6581      0.0463   14.2107    0.0000 [ 0.5400; 0.7795 ] 
#>   eta3 =~ y33      0.7474      0.0419   17.8245    0.0000 [ 0.6415; 0.8583 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0194   20.3504    0.0000 [ 0.3438; 0.4443 ] 
#>   eta1 <~ y12      0.3873      0.0204   19.0071    0.0000 [ 0.3392; 0.4445 ] 
#>   eta1 <~ y13      0.4542      0.0172   26.3534    0.0000 [ 0.4073; 0.4965 ] 
#>   eta2 <~ y21      0.3058      0.0258   11.8506    0.0000 [ 0.2346; 0.3681 ] 
#>   eta2 <~ y22      0.4473      0.0190   23.5873    0.0000 [ 0.3949; 0.4930 ] 
#>   eta2 <~ y23      0.4735      0.0179   26.4391    0.0000 [ 0.4342; 0.5268 ] 
#>   eta3 <~ y31      0.4400      0.0153   28.7256    0.0000 [ 0.3983; 0.4775 ] 
#>   eta3 <~ y32      0.3521      0.0215   16.4122    0.0000 [ 0.2980; 0.4090 ] 
#>   eta3 <~ y33      0.3999      0.0218   18.3874    0.0000 [ 0.3453; 0.4578 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0501   13.4065    0.0000 [ 0.5374; 0.7964 ] 
#>   eta3 ~ eta1       0.6634      0.0411   16.1396    0.0000 [ 0.5598; 0.7723 ] 
#>   eta3 ~ eta2       0.3052      0.0911    3.3497    0.0008 [ 0.0758; 0.5469 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0619    3.3110    0.0009 [ 0.0483; 0.3682 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05007517 13.406514 5.538125e-41
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08779310  5.222583 1.764443e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09109805  3.349700 8.089919e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5374131          0.7963736          0.5685092          0.7652775
#> 2          0.2307632          0.6847795          0.2852817          0.6302611
#> 3          0.0757750          0.5468827          0.1323458          0.4903119
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5467854          0.7845949          0.5616256          0.7579914
#> 2          0.3047342          0.6528255          0.3112093          0.5864727
#> 3          0.0998088          0.4574618          0.1466000          0.4465720