Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 52703721
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0426   15.7580    0.0000 [ 0.5803; 0.7317 ] 
#>   eta3 ~ eta1      0.4585      0.0759    6.0398    0.0000 [ 0.3321; 0.5778 ] 
#>   eta3 ~ eta2      0.3052      0.0819    3.7245    0.0002 [ 0.1659; 0.4357 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0364   18.2383    0.0000 [ 0.6175; 0.7409 ] 
#>   eta1 =~ y12      0.6493      0.0384   16.8886    0.0000 [ 0.5856; 0.7077 ] 
#>   eta1 =~ y13      0.7613      0.0333   22.8799    0.0000 [ 0.7033; 0.8100 ] 
#>   eta2 =~ y21      0.5165      0.0685    7.5440    0.0000 [ 0.3483; 0.6018 ] 
#>   eta2 =~ y22      0.7554      0.0354   21.3331    0.0000 [ 0.7007; 0.8215 ] 
#>   eta2 =~ y23      0.7997      0.0302   26.5000    0.0000 [ 0.7521; 0.8429 ] 
#>   eta3 =~ y31      0.8223      0.0356   23.0986    0.0000 [ 0.7370; 0.8688 ] 
#>   eta3 =~ y32      0.6581      0.0312   21.0866    0.0000 [ 0.6137; 0.7155 ] 
#>   eta3 =~ y33      0.7474      0.0349   21.4229    0.0000 [ 0.6806; 0.8055 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0189   20.9168    0.0000 [ 0.3715; 0.4435 ] 
#>   eta1 <~ y12      0.3873      0.0176   21.9568    0.0000 [ 0.3573; 0.4143 ] 
#>   eta1 <~ y13      0.4542      0.0217   20.9316    0.0000 [ 0.4052; 0.4856 ] 
#>   eta2 <~ y21      0.3058      0.0352    8.6827    0.0000 [ 0.2173; 0.3486 ] 
#>   eta2 <~ y22      0.4473      0.0217   20.5763    0.0000 [ 0.4194; 0.4864 ] 
#>   eta2 <~ y23      0.4735      0.0229   20.6587    0.0000 [ 0.4451; 0.5244 ] 
#>   eta3 <~ y31      0.4400      0.0165   26.6922    0.0000 [ 0.4083; 0.4584 ] 
#>   eta3 <~ y32      0.3521      0.0170   20.6560    0.0000 [ 0.3284; 0.3876 ] 
#>   eta3 <~ y33      0.3999      0.0168   23.8750    0.0000 [ 0.3669; 0.4270 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0426   15.7580    0.0000 [ 0.5803; 0.7317 ] 
#>   eta3 ~ eta1       0.6634      0.0383   17.3178    0.0000 [ 0.5874; 0.7299 ] 
#>   eta3 ~ eta2       0.3052      0.0819    3.7245    0.0002 [ 0.1659; 0.4357 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0553    3.7046    0.0002 [ 0.1117; 0.2950 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03635589 18.238307  2.562740e-74
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03844481 16.888570  5.461052e-64
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03327575 22.879901 7.366859e-116
#> 4 eta2 =~ y21  Common factor 0.5164548 0.06845882  7.544021  4.556963e-14
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03540924 21.333063 5.601651e-101
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03017598 26.500008 9.690334e-155
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03559862 23.098572 4.785497e-118
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03120792 21.086599  1.055860e-98
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03488902 21.422901 8.173599e-102
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6175289          0.7408865
#> 2          0.5856419          0.7077436
#> 3          0.7032687          0.8100259
#> 4          0.3483448          0.6017865
#> 5          0.7007306          0.8215343
#> 6          0.7521318          0.8428507
#> 7          0.7369927          0.8687587
#> 8          0.6137267          0.7154823
#> 9          0.6806133          0.8054825

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 52703721
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0426   15.7580    0.0000 [ 0.5675; 0.7878 ] 
#>   eta3 ~ eta1      0.4585      0.0759    6.0398    0.0000 [ 0.2662; 0.6588 ] 
#>   eta3 ~ eta2      0.3052      0.0819    3.7245    0.0002 [ 0.0863; 0.5100 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0364   18.2383    0.0000 [ 0.5586; 0.7466 ] 
#>   eta1 =~ y12      0.6493      0.0384   16.8886    0.0000 [ 0.5488; 0.7476 ] 
#>   eta1 =~ y13      0.7613      0.0333   22.8799    0.0000 [ 0.6784; 0.8505 ] 
#>   eta2 =~ y21      0.5165      0.0685    7.5440    0.0000 [ 0.3605; 0.7146 ] 
#>   eta2 =~ y22      0.7554      0.0354   21.3331    0.0000 [ 0.6519; 0.8350 ] 
#>   eta2 =~ y23      0.7997      0.0302   26.5000    0.0000 [ 0.7187; 0.8747 ] 
#>   eta3 =~ y31      0.8223      0.0356   23.0986    0.0000 [ 0.7394; 0.9235 ] 
#>   eta3 =~ y32      0.6581      0.0312   21.0866    0.0000 [ 0.5633; 0.7246 ] 
#>   eta3 =~ y33      0.7474      0.0349   21.4229    0.0000 [ 0.6608; 0.8412 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0189   20.9168    0.0000 [ 0.3430; 0.4408 ] 
#>   eta1 <~ y12      0.3873      0.0176   21.9568    0.0000 [ 0.3437; 0.4349 ] 
#>   eta1 <~ y13      0.4542      0.0217   20.9316    0.0000 [ 0.4026; 0.5148 ] 
#>   eta2 <~ y21      0.3058      0.0352    8.6827    0.0000 [ 0.2279; 0.4101 ] 
#>   eta2 <~ y22      0.4473      0.0217   20.5763    0.0000 [ 0.3839; 0.4963 ] 
#>   eta2 <~ y23      0.4735      0.0229   20.6587    0.0000 [ 0.4122; 0.5307 ] 
#>   eta3 <~ y31      0.4400      0.0165   26.6922    0.0000 [ 0.4026; 0.4879 ] 
#>   eta3 <~ y32      0.3521      0.0170   20.6560    0.0000 [ 0.3007; 0.3888 ] 
#>   eta3 <~ y33      0.3999      0.0168   23.8750    0.0000 [ 0.3588; 0.4455 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0426   15.7580    0.0000 [ 0.5675; 0.7878 ] 
#>   eta3 ~ eta1       0.6634      0.0383   17.3178    0.0000 [ 0.5658; 0.7639 ] 
#>   eta3 ~ eta2       0.3052      0.0819    3.7245    0.0002 [ 0.0863; 0.5100 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0553    3.7046    0.0002 [ 0.0594; 0.3454 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04260264 15.758026 6.049325e-56
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07591404  6.039815 1.542906e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08192984  3.724542 1.956701e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56747562          0.7877924          0.5939314          0.7613367
#> 2         0.26620598          0.6587905          0.3133477          0.6116488
#> 3         0.08631959          0.5100144          0.1371971          0.4591370
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5750479          0.7513073          0.5802785          0.7316786
#> 2          0.3157411          0.5914050          0.3321259          0.5778082
#> 3          0.1615763          0.4394009          0.1659005          0.4356529