Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1858443328
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0419   16.0366    0.0000 [ 0.5895; 0.7490 ] 
#>   eta3 ~ eta1      0.4585      0.0917    5.0018    0.0000 [ 0.3280; 0.6528 ] 
#>   eta3 ~ eta2      0.3052      0.0948    3.2194    0.0013 [ 0.1099; 0.4383 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0321   20.6550    0.0000 [ 0.6103; 0.7049 ] 
#>   eta1 =~ y12      0.6493      0.0468   13.8599    0.0000 [ 0.5287; 0.7308 ] 
#>   eta1 =~ y13      0.7613      0.0373   20.3862    0.0000 [ 0.6958; 0.8165 ] 
#>   eta2 =~ y21      0.5165      0.0546    9.4653    0.0000 [ 0.3978; 0.5895 ] 
#>   eta2 =~ y22      0.7554      0.0384   19.6747    0.0000 [ 0.6986; 0.8356 ] 
#>   eta2 =~ y23      0.7997      0.0395   20.2486    0.0000 [ 0.7282; 0.8668 ] 
#>   eta3 =~ y31      0.8223      0.0366   22.4466    0.0000 [ 0.7398; 0.8906 ] 
#>   eta3 =~ y32      0.6581      0.0370   17.7824    0.0000 [ 0.5992; 0.7274 ] 
#>   eta3 =~ y33      0.7474      0.0397   18.8108    0.0000 [ 0.6540; 0.8012 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0198   19.9600    0.0000 [ 0.3596; 0.4319 ] 
#>   eta1 <~ y12      0.3873      0.0204   19.0326    0.0000 [ 0.3564; 0.4262 ] 
#>   eta1 <~ y13      0.4542      0.0209   21.6827    0.0000 [ 0.4188; 0.4987 ] 
#>   eta2 <~ y21      0.3058      0.0274   11.1578    0.0000 [ 0.2492; 0.3479 ] 
#>   eta2 <~ y22      0.4473      0.0268   16.6972    0.0000 [ 0.4083; 0.5046 ] 
#>   eta2 <~ y23      0.4735      0.0204   23.2665    0.0000 [ 0.4396; 0.5106 ] 
#>   eta3 <~ y31      0.4400      0.0194   22.6914    0.0000 [ 0.4021; 0.4713 ] 
#>   eta3 <~ y32      0.3521      0.0176   20.0401    0.0000 [ 0.3216; 0.3886 ] 
#>   eta3 <~ y33      0.3999      0.0187   21.3582    0.0000 [ 0.3605; 0.4281 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0419   16.0366    0.0000 [ 0.5895; 0.7490 ] 
#>   eta3 ~ eta1       0.6634      0.0428   15.5097    0.0000 [ 0.5811; 0.7365 ] 
#>   eta3 ~ eta2       0.3052      0.0948    3.2194    0.0013 [ 0.1099; 0.4383 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0601    3.4074    0.0007 [ 0.0762; 0.2842 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03210222 20.654957  8.810037e-95
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04684563 13.859946  1.107569e-43
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03734617 20.386179  2.218082e-92
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05456312  9.465272  2.927977e-21
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03839382 19.674722  3.551168e-86
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03949221 20.248643  3.651595e-91
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03663253 22.446644 1.380066e-111
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03700676 17.782397  9.675340e-71
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03973388 18.810752  6.165695e-79
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6103067          0.7048610
#> 2          0.5287461          0.7308284
#> 3          0.6957920          0.8165061
#> 4          0.3978205          0.5894962
#> 5          0.6986077          0.8356104
#> 6          0.7282102          0.8667786
#> 7          0.7398102          0.8906255
#> 8          0.5992200          0.7274412
#> 9          0.6539889          0.8012183

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1858443328
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0419   16.0366    0.0000 [ 0.5535; 0.7700 ] 
#>   eta3 ~ eta1      0.4585      0.0917    5.0018    0.0000 [ 0.2018; 0.6759 ] 
#>   eta3 ~ eta2      0.3052      0.0948    3.2194    0.0013 [ 0.0766; 0.5668 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0321   20.6550    0.0000 [ 0.5886; 0.7546 ] 
#>   eta1 =~ y12      0.6493      0.0468   13.8599    0.0000 [ 0.5229; 0.7652 ] 
#>   eta1 =~ y13      0.7613      0.0373   20.3862    0.0000 [ 0.6688; 0.8619 ] 
#>   eta2 =~ y21      0.5165      0.0546    9.4653    0.0000 [ 0.3815; 0.6637 ] 
#>   eta2 =~ y22      0.7554      0.0384   19.6747    0.0000 [ 0.6507; 0.8492 ] 
#>   eta2 =~ y23      0.7997      0.0395   20.2486    0.0000 [ 0.7006; 0.9048 ] 
#>   eta3 =~ y31      0.8223      0.0366   22.4466    0.0000 [ 0.7239; 0.9134 ] 
#>   eta3 =~ y32      0.6581      0.0370   17.7824    0.0000 [ 0.5596; 0.7510 ] 
#>   eta3 =~ y33      0.7474      0.0397   18.8108    0.0000 [ 0.6496; 0.8551 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0198   19.9600    0.0000 [ 0.3473; 0.4498 ] 
#>   eta1 <~ y12      0.3873      0.0204   19.0326    0.0000 [ 0.3300; 0.4353 ] 
#>   eta1 <~ y13      0.4542      0.0209   21.6827    0.0000 [ 0.4001; 0.5084 ] 
#>   eta2 <~ y21      0.3058      0.0274   11.1578    0.0000 [ 0.2387; 0.3805 ] 
#>   eta2 <~ y22      0.4473      0.0268   16.6972    0.0000 [ 0.3740; 0.5126 ] 
#>   eta2 <~ y23      0.4735      0.0204   23.2665    0.0000 [ 0.4222; 0.5275 ] 
#>   eta3 <~ y31      0.4400      0.0194   22.6914    0.0000 [ 0.3884; 0.4887 ] 
#>   eta3 <~ y32      0.3521      0.0176   20.0401    0.0000 [ 0.3057; 0.3966 ] 
#>   eta3 <~ y33      0.3999      0.0187   21.3582    0.0000 [ 0.3548; 0.4516 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0419   16.0366    0.0000 [ 0.5535; 0.7700 ] 
#>   eta3 ~ eta1       0.6634      0.0428   15.5097    0.0000 [ 0.5434; 0.7646 ] 
#>   eta3 ~ eta2       0.3052      0.0948    3.2194    0.0013 [ 0.0766; 0.5668 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0601    3.4074    0.0007 [ 0.0597; 0.3706 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04186260 16.036590 7.094431e-58
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09166796  5.001822 5.679115e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09478387  3.219441 1.284406e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55351906          0.7700088          0.5795152          0.7440126
#> 2         0.20180449          0.6758594          0.2587292          0.6189347
#> 3         0.07660509          0.5667738          0.1354648          0.5079141
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5765175          0.7552864          0.5895057          0.7490340
#> 2          0.2874961          0.6581238          0.3279990          0.6528055
#> 3          0.0688293          0.4616249          0.1099158          0.4383483