Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1986141089
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0445   15.0897    0.0000 [ 0.6085; 0.7613 ] 
#>   eta3 ~ eta1      0.4585      0.0958    4.7881    0.0000 [ 0.2518; 0.6339 ] 
#>   eta3 ~ eta2      0.3052      0.1049    2.9099    0.0036 [ 0.1078; 0.5033 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0420   15.7863    0.0000 [ 0.5625; 0.7320 ] 
#>   eta1 =~ y12      0.6493      0.0349   18.5981    0.0000 [ 0.5916; 0.7120 ] 
#>   eta1 =~ y13      0.7613      0.0296   25.7107    0.0000 [ 0.7208; 0.8164 ] 
#>   eta2 =~ y21      0.5165      0.0540    9.5667    0.0000 [ 0.3760; 0.5941 ] 
#>   eta2 =~ y22      0.7554      0.0328   23.0283    0.0000 [ 0.6868; 0.8121 ] 
#>   eta2 =~ y23      0.7997      0.0341   23.4575    0.0000 [ 0.7403; 0.8601 ] 
#>   eta3 =~ y31      0.8223      0.0311   26.4258    0.0000 [ 0.7641; 0.8671 ] 
#>   eta3 =~ y32      0.6581      0.0358   18.4066    0.0000 [ 0.6030; 0.7288 ] 
#>   eta3 =~ y33      0.7474      0.0328   22.7637    0.0000 [ 0.6848; 0.7960 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0207   19.1235    0.0000 [ 0.3526; 0.4269 ] 
#>   eta1 <~ y12      0.3873      0.0201   19.2954    0.0000 [ 0.3610; 0.4242 ] 
#>   eta1 <~ y13      0.4542      0.0177   25.6592    0.0000 [ 0.4259; 0.4890 ] 
#>   eta2 <~ y21      0.3058      0.0272   11.2622    0.0000 [ 0.2343; 0.3456 ] 
#>   eta2 <~ y22      0.4473      0.0240   18.6068    0.0000 [ 0.4065; 0.4960 ] 
#>   eta2 <~ y23      0.4735      0.0186   25.4414    0.0000 [ 0.4437; 0.5093 ] 
#>   eta3 <~ y31      0.4400      0.0177   24.8581    0.0000 [ 0.4093; 0.4734 ] 
#>   eta3 <~ y32      0.3521      0.0164   21.4878    0.0000 [ 0.3283; 0.3834 ] 
#>   eta3 <~ y33      0.3999      0.0161   24.8624    0.0000 [ 0.3664; 0.4259 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0445   15.0897    0.0000 [ 0.6085; 0.7613 ] 
#>   eta3 ~ eta1       0.6634      0.0448   14.7974    0.0000 [ 0.5948; 0.7607 ] 
#>   eta3 ~ eta2       0.3052      0.1049    2.9099    0.0036 [ 0.1078; 0.5033 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0671    3.0549    0.0023 [ 0.0818; 0.3594 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04200290 15.786289  3.866704e-56
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03491094 18.598121  3.327666e-77
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02961200 25.710715 8.870993e-146
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05398443  9.566737  1.103331e-21
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03280262 23.028276 2.428614e-117
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03408986 23.457522 1.107766e-121
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03111648 26.425781 6.928218e-154
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03575175 18.406619  1.162573e-75
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03283407 22.763674 1.050628e-114
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5625161          0.7319642
#> 2          0.5915687          0.7120039
#> 3          0.7208341          0.8163570
#> 4          0.3760159          0.5941054
#> 5          0.6868413          0.8121032
#> 6          0.7403213          0.8600539
#> 7          0.7640804          0.8670574
#> 8          0.6030466          0.7288495
#> 9          0.6847907          0.7960287

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1986141089
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0445   15.0897    0.0000 [ 0.5546; 0.7847 ] 
#>   eta3 ~ eta1      0.4585      0.0958    4.7881    0.0000 [ 0.2133; 0.7086 ] 
#>   eta3 ~ eta2      0.3052      0.1049    2.9099    0.0036 [ 0.0342; 0.5765 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0420   15.7863    0.0000 [ 0.5682; 0.7854 ] 
#>   eta1 =~ y12      0.6493      0.0349   18.5981    0.0000 [ 0.5515; 0.7320 ] 
#>   eta1 =~ y13      0.7613      0.0296   25.7107    0.0000 [ 0.6842; 0.8373 ] 
#>   eta2 =~ y21      0.5165      0.0540    9.5667    0.0000 [ 0.3918; 0.6710 ] 
#>   eta2 =~ y22      0.7554      0.0328   23.0283    0.0000 [ 0.6736; 0.8432 ] 
#>   eta2 =~ y23      0.7997      0.0341   23.4575    0.0000 [ 0.7103; 0.8866 ] 
#>   eta3 =~ y31      0.8223      0.0311   26.4258    0.0000 [ 0.7475; 0.9084 ] 
#>   eta3 =~ y32      0.6581      0.0358   18.4066    0.0000 [ 0.5565; 0.7414 ] 
#>   eta3 =~ y33      0.7474      0.0328   22.7637    0.0000 [ 0.6758; 0.8456 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0207   19.1235    0.0000 [ 0.3491; 0.4561 ] 
#>   eta1 <~ y12      0.3873      0.0201   19.2954    0.0000 [ 0.3295; 0.4333 ] 
#>   eta1 <~ y13      0.4542      0.0177   25.6592    0.0000 [ 0.4064; 0.4979 ] 
#>   eta2 <~ y21      0.3058      0.0272   11.2622    0.0000 [ 0.2421; 0.3825 ] 
#>   eta2 <~ y22      0.4473      0.0240   18.6068    0.0000 [ 0.3824; 0.5067 ] 
#>   eta2 <~ y23      0.4735      0.0186   25.4414    0.0000 [ 0.4201; 0.5163 ] 
#>   eta3 <~ y31      0.4400      0.0177   24.8581    0.0000 [ 0.3942; 0.4857 ] 
#>   eta3 <~ y32      0.3521      0.0164   21.4878    0.0000 [ 0.3025; 0.3873 ] 
#>   eta3 <~ y33      0.3999      0.0161   24.8624    0.0000 [ 0.3628; 0.4460 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0445   15.0897    0.0000 [ 0.5546; 0.7847 ] 
#>   eta3 ~ eta1       0.6634      0.0448   14.7974    0.0000 [ 0.5515; 0.7833 ] 
#>   eta3 ~ eta2       0.3052      0.1049    2.9099    0.0036 [ 0.0342; 0.5765 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0671    3.0549    0.0023 [ 0.0331; 0.3799 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04448939 15.089743 1.891773e-51
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09575954  4.788105 1.683633e-06
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.10486687  2.909891 3.615553e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55459181          0.7846658         0.58221920          0.7570384
#> 2         0.21334902          0.7085633         0.27281457          0.6490977
#> 3         0.03419795          0.5765102         0.09931905          0.5113891
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.59938741          0.7774667          0.6085326          0.7612785
#> 2         0.24321915          0.6684780          0.2517503          0.6339424
#> 3         0.09331947          0.5659510          0.1077846          0.5033347