Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -518562992
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0470   14.2686    0.0000 [ 0.6048; 0.7668 ] 
#>   eta3 ~ eta1      0.4585      0.0954    4.8056    0.0000 [ 0.2548; 0.6034 ] 
#>   eta3 ~ eta2      0.3052      0.1047    2.9139    0.0036 [ 0.1568; 0.5172 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0413   16.0511    0.0000 [ 0.5663; 0.7329 ] 
#>   eta1 =~ y12      0.6493      0.0360   18.0283    0.0000 [ 0.5856; 0.7022 ] 
#>   eta1 =~ y13      0.7613      0.0321   23.6930    0.0000 [ 0.7035; 0.8058 ] 
#>   eta2 =~ y21      0.5165      0.0461   11.1958    0.0000 [ 0.4457; 0.5985 ] 
#>   eta2 =~ y22      0.7554      0.0453   16.6667    0.0000 [ 0.6730; 0.8218 ] 
#>   eta2 =~ y23      0.7997      0.0363   22.0455    0.0000 [ 0.7405; 0.8640 ] 
#>   eta3 =~ y31      0.8223      0.0280   29.3949    0.0000 [ 0.7704; 0.8616 ] 
#>   eta3 =~ y32      0.6581      0.0318   20.6638    0.0000 [ 0.5924; 0.7031 ] 
#>   eta3 =~ y33      0.7474      0.0361   20.6949    0.0000 [ 0.6645; 0.8071 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0215   18.4273    0.0000 [ 0.3466; 0.4315 ] 
#>   eta1 <~ y12      0.3873      0.0187   20.6839    0.0000 [ 0.3480; 0.4156 ] 
#>   eta1 <~ y13      0.4542      0.0213   21.3560    0.0000 [ 0.4194; 0.5150 ] 
#>   eta2 <~ y21      0.3058      0.0258   11.8706    0.0000 [ 0.2692; 0.3551 ] 
#>   eta2 <~ y22      0.4473      0.0231   19.3598    0.0000 [ 0.4042; 0.4875 ] 
#>   eta2 <~ y23      0.4735      0.0181   26.1520    0.0000 [ 0.4374; 0.5009 ] 
#>   eta3 <~ y31      0.4400      0.0147   29.8669    0.0000 [ 0.4082; 0.4606 ] 
#>   eta3 <~ y32      0.3521      0.0172   20.4278    0.0000 [ 0.3181; 0.3817 ] 
#>   eta3 <~ y33      0.3999      0.0165   24.2874    0.0000 [ 0.3769; 0.4340 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0470   14.2686    0.0000 [ 0.6048; 0.7668 ] 
#>   eta3 ~ eta1       0.6634      0.0360   18.4062    0.0000 [ 0.5913; 0.7333 ] 
#>   eta3 ~ eta2       0.3052      0.1047    2.9139    0.0036 [ 0.1568; 0.5172 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0723    2.8331    0.0046 [ 0.1071; 0.3295 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04130981 16.05115  5.611708e-58
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03601441 18.02828  1.168678e-72
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03213384 23.69296 4.261410e-124
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04612930 11.19581  4.274901e-29
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04532318 16.66669  2.289071e-62
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03627332 22.04551 1.054942e-107
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02797343 29.39495 6.372347e-190
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03184648 20.66379  7.337724e-95
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03611638 20.69488  3.851711e-95
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5663044          0.7328868
#> 2          0.5856030          0.7022230
#> 3          0.7035480          0.8058477
#> 4          0.4456606          0.5985043
#> 5          0.6730278          0.8218229
#> 6          0.7404507          0.8639511
#> 7          0.7703994          0.8616430
#> 8          0.5923982          0.7031117
#> 9          0.6645303          0.8070601

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -518562992
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0470   14.2686    0.0000 [ 0.5462; 0.7895 ] 
#>   eta3 ~ eta1      0.4585      0.0954    4.8056    0.0000 [ 0.2185; 0.7119 ] 
#>   eta3 ~ eta2      0.3052      0.1047    2.9139    0.0036 [ 0.0200; 0.5616 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0413   16.0511    0.0000 [ 0.5670; 0.7806 ] 
#>   eta1 =~ y12      0.6493      0.0360   18.0283    0.0000 [ 0.5646; 0.7509 ] 
#>   eta1 =~ y13      0.7613      0.0321   23.6930    0.0000 [ 0.6784; 0.8446 ] 
#>   eta2 =~ y21      0.5165      0.0461   11.1958    0.0000 [ 0.3804; 0.6189 ] 
#>   eta2 =~ y22      0.7554      0.0453   16.6667    0.0000 [ 0.6363; 0.8707 ] 
#>   eta2 =~ y23      0.7997      0.0363   22.0455    0.0000 [ 0.7094; 0.8970 ] 
#>   eta3 =~ y31      0.8223      0.0280   29.3949    0.0000 [ 0.7541; 0.8988 ] 
#>   eta3 =~ y32      0.6581      0.0318   20.6638    0.0000 [ 0.5765; 0.7412 ] 
#>   eta3 =~ y33      0.7474      0.0361   20.6949    0.0000 [ 0.6541; 0.8408 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0215   18.4273    0.0000 [ 0.3414; 0.4524 ] 
#>   eta1 <~ y12      0.3873      0.0187   20.6839    0.0000 [ 0.3390; 0.4358 ] 
#>   eta1 <~ y13      0.4542      0.0213   21.3560    0.0000 [ 0.3931; 0.5031 ] 
#>   eta2 <~ y21      0.3058      0.0258   11.8706    0.0000 [ 0.2325; 0.3657 ] 
#>   eta2 <~ y22      0.4473      0.0231   19.3598    0.0000 [ 0.3909; 0.5104 ] 
#>   eta2 <~ y23      0.4735      0.0181   26.1520    0.0000 [ 0.4334; 0.5271 ] 
#>   eta3 <~ y31      0.4400      0.0147   29.8669    0.0000 [ 0.4027; 0.4789 ] 
#>   eta3 <~ y32      0.3521      0.0172   20.4278    0.0000 [ 0.3068; 0.3960 ] 
#>   eta3 <~ y33      0.3999      0.0165   24.2874    0.0000 [ 0.3562; 0.4414 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0470   14.2686    0.0000 [ 0.5462; 0.7895 ] 
#>   eta3 ~ eta1       0.6634      0.0360   18.4062    0.0000 [ 0.5661; 0.7525 ] 
#>   eta3 ~ eta2       0.3052      0.1047    2.9139    0.0036 [ 0.0200; 0.5616 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0723    2.8331    0.0046 [ 0.0071; 0.3810 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04704971 14.268598 3.434117e-46
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09541062  4.805616 1.542759e-06
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.10472410  2.913858 3.569926e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54623067          0.7895452         0.57544799          0.7603279
#> 2         0.21853602          0.7119459         0.27778489          0.6526970
#> 3         0.01997907          0.5615530         0.08501151          0.4965206
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5964603          0.7981440          0.6047762          0.7668133
#> 2          0.2392994          0.6071517          0.2548090          0.6033736
#> 3          0.1514281          0.5520649          0.1567856          0.5172215