Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1278327536
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0388   17.3150    0.0000 [ 0.6142; 0.7498 ] 
#>   eta3 ~ eta1      0.4585      0.0694    6.6059    0.0000 [ 0.3750; 0.6055 ] 
#>   eta3 ~ eta2      0.3052      0.0753    4.0547    0.0001 [ 0.1501; 0.4045 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0420   15.7971    0.0000 [ 0.5700; 0.7149 ] 
#>   eta1 =~ y12      0.6493      0.0342   18.9663    0.0000 [ 0.5934; 0.7137 ] 
#>   eta1 =~ y13      0.7613      0.0277   27.4395    0.0000 [ 0.6998; 0.8032 ] 
#>   eta2 =~ y21      0.5165      0.0545    9.4727    0.0000 [ 0.4390; 0.6178 ] 
#>   eta2 =~ y22      0.7554      0.0333   22.6963    0.0000 [ 0.6830; 0.7962 ] 
#>   eta2 =~ y23      0.7997      0.0309   25.9127    0.0000 [ 0.7565; 0.8620 ] 
#>   eta3 =~ y31      0.8223      0.0264   31.1824    0.0000 [ 0.7731; 0.8651 ] 
#>   eta3 =~ y32      0.6581      0.0316   20.8024    0.0000 [ 0.6049; 0.7146 ] 
#>   eta3 =~ y33      0.7474      0.0379   19.7381    0.0000 [ 0.6689; 0.8160 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0189   20.9068    0.0000 [ 0.3677; 0.4294 ] 
#>   eta1 <~ y12      0.3873      0.0175   22.1703    0.0000 [ 0.3601; 0.4161 ] 
#>   eta1 <~ y13      0.4542      0.0207   21.9179    0.0000 [ 0.4134; 0.4861 ] 
#>   eta2 <~ y21      0.3058      0.0253   12.1009    0.0000 [ 0.2641; 0.3471 ] 
#>   eta2 <~ y22      0.4473      0.0203   21.9875    0.0000 [ 0.4040; 0.4783 ] 
#>   eta2 <~ y23      0.4735      0.0216   21.9048    0.0000 [ 0.4339; 0.5196 ] 
#>   eta3 <~ y31      0.4400      0.0137   32.0744    0.0000 [ 0.4189; 0.4647 ] 
#>   eta3 <~ y32      0.3521      0.0176   20.0073    0.0000 [ 0.3194; 0.3845 ] 
#>   eta3 <~ y33      0.3999      0.0165   24.2038    0.0000 [ 0.3762; 0.4348 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0388   17.3150    0.0000 [ 0.6142; 0.7498 ] 
#>   eta3 ~ eta1       0.6634      0.0358   18.5143    0.0000 [ 0.5995; 0.7304 ] 
#>   eta3 ~ eta2       0.3052      0.0753    4.0547    0.0001 [ 0.1501; 0.4045 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0500    4.0944    0.0000 [ 0.1060; 0.2795 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04197416 15.797098  3.257734e-56
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03423318 18.966333  3.237144e-80
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02774636 27.439487 9.274762e-166
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05452052  9.472668  2.727844e-21
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03328240 22.696311 4.871966e-114
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03085994 25.912675 4.793264e-148
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02636994 31.182375 1.847272e-213
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03163421 20.802445  4.112957e-96
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03786700 19.738141  1.014415e-86
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5700361          0.7149235
#> 2          0.5934425          0.7136535
#> 3          0.6998324          0.8032489
#> 4          0.4389530          0.6177854
#> 5          0.6830027          0.7961820
#> 6          0.7564717          0.8620008
#> 7          0.7731185          0.8650667
#> 8          0.6049067          0.7145719
#> 9          0.6688660          0.8160071

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1278327536
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0388   17.3150    0.0000 [ 0.5668; 0.7673 ] 
#>   eta3 ~ eta1      0.4585      0.0694    6.6059    0.0000 [ 0.2578; 0.6168 ] 
#>   eta3 ~ eta2      0.3052      0.0753    4.0547    0.0001 [ 0.1325; 0.5217 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0420   15.7971    0.0000 [ 0.5555; 0.7726 ] 
#>   eta1 =~ y12      0.6493      0.0342   18.9663    0.0000 [ 0.5557; 0.7327 ] 
#>   eta1 =~ y13      0.7613      0.0277   27.4395    0.0000 [ 0.6995; 0.8430 ] 
#>   eta2 =~ y21      0.5165      0.0545    9.4727    0.0000 [ 0.3681; 0.6500 ] 
#>   eta2 =~ y22      0.7554      0.0333   22.6963    0.0000 [ 0.6796; 0.8517 ] 
#>   eta2 =~ y23      0.7997      0.0309   25.9127    0.0000 [ 0.7106; 0.8702 ] 
#>   eta3 =~ y31      0.8223      0.0264   31.1824    0.0000 [ 0.7499; 0.8863 ] 
#>   eta3 =~ y32      0.6581      0.0316   20.8024    0.0000 [ 0.5842; 0.7478 ] 
#>   eta3 =~ y33      0.7474      0.0379   19.7381    0.0000 [ 0.6380; 0.8338 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0189   20.9068    0.0000 [ 0.3459; 0.4437 ] 
#>   eta1 <~ y12      0.3873      0.0175   22.1703    0.0000 [ 0.3376; 0.4280 ] 
#>   eta1 <~ y13      0.4542      0.0207   21.9179    0.0000 [ 0.4044; 0.5116 ] 
#>   eta2 <~ y21      0.3058      0.0253   12.1009    0.0000 [ 0.2381; 0.3688 ] 
#>   eta2 <~ y22      0.4473      0.0203   21.9875    0.0000 [ 0.4027; 0.5079 ] 
#>   eta2 <~ y23      0.4735      0.0216   21.9048    0.0000 [ 0.4142; 0.5260 ] 
#>   eta3 <~ y31      0.4400      0.0137   32.0744    0.0000 [ 0.4049; 0.4758 ] 
#>   eta3 <~ y32      0.3521      0.0176   20.0073    0.0000 [ 0.3129; 0.4039 ] 
#>   eta3 <~ y33      0.3999      0.0165   24.2038    0.0000 [ 0.3536; 0.4391 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0388   17.3150    0.0000 [ 0.5668; 0.7673 ] 
#>   eta3 ~ eta1       0.6634      0.0358   18.5143    0.0000 [ 0.5637; 0.7490 ] 
#>   eta3 ~ eta2       0.3052      0.0753    4.0547    0.0001 [ 0.1325; 0.5217 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0500    4.0944    0.0000 [ 0.0897; 0.3484 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03877180 17.314991 3.625764e-67
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.06940816  6.605949 3.949793e-11
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07525881  4.054690 5.020090e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5667763          0.7672822          0.5908531          0.7432053
#> 2          0.2578246          0.6167644          0.3009262          0.5736627
#> 3          0.1325047          0.5217007          0.1792395          0.4749659
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.6047148          0.7591212          0.6142336          0.7497523
#> 2          0.3155500          0.6458178          0.3749602          0.6054562
#> 3          0.1100466          0.4151236          0.1501252          0.4045139