Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1776338981
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0384   17.4708    0.0000 [ 0.6102; 0.7487 ] 
#>   eta3 ~ eta1      0.4585      0.0746    6.1439    0.0000 [ 0.3254; 0.5934 ] 
#>   eta3 ~ eta2      0.3052      0.0874    3.4907    0.0005 [ 0.1554; 0.4409 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0380   17.4385    0.0000 [ 0.6023; 0.7378 ] 
#>   eta1 =~ y12      0.6493      0.0387   16.7619    0.0000 [ 0.5775; 0.7102 ] 
#>   eta1 =~ y13      0.7613      0.0308   24.7306    0.0000 [ 0.7039; 0.7986 ] 
#>   eta2 =~ y21      0.5165      0.0487   10.6078    0.0000 [ 0.4254; 0.5956 ] 
#>   eta2 =~ y22      0.7554      0.0328   23.0161    0.0000 [ 0.6993; 0.8141 ] 
#>   eta2 =~ y23      0.7997      0.0307   26.0763    0.0000 [ 0.7358; 0.8463 ] 
#>   eta3 =~ y31      0.8223      0.0344   23.8809    0.0000 [ 0.7637; 0.8987 ] 
#>   eta3 =~ y32      0.6581      0.0453   14.5294    0.0000 [ 0.5685; 0.7336 ] 
#>   eta3 =~ y33      0.7474      0.0377   19.8472    0.0000 [ 0.6930; 0.8189 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0214   18.4523    0.0000 [ 0.3636; 0.4338 ] 
#>   eta1 <~ y12      0.3873      0.0174   22.2759    0.0000 [ 0.3549; 0.4162 ] 
#>   eta1 <~ y13      0.4542      0.0189   24.0775    0.0000 [ 0.4167; 0.4842 ] 
#>   eta2 <~ y21      0.3058      0.0216   14.1472    0.0000 [ 0.2562; 0.3396 ] 
#>   eta2 <~ y22      0.4473      0.0195   22.9720    0.0000 [ 0.4204; 0.4824 ] 
#>   eta2 <~ y23      0.4735      0.0186   25.4223    0.0000 [ 0.4348; 0.5067 ] 
#>   eta3 <~ y31      0.4400      0.0195   22.5165    0.0000 [ 0.4067; 0.4758 ] 
#>   eta3 <~ y32      0.3521      0.0198   17.7475    0.0000 [ 0.3142; 0.3873 ] 
#>   eta3 <~ y33      0.3999      0.0200   19.9949    0.0000 [ 0.3600; 0.4290 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0384   17.4708    0.0000 [ 0.6102; 0.7487 ] 
#>   eta3 ~ eta1       0.6634      0.0284   23.3526    0.0000 [ 0.5984; 0.7124 ] 
#>   eta3 ~ eta2       0.3052      0.0874    3.4907    0.0005 [ 0.1554; 0.4409 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0590    3.4699    0.0005 [ 0.1041; 0.2958 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03802327 17.43853  4.207706e-68
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03873537 16.76189  4.636571e-63
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03078558 24.73059 5.014653e-135
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04868611 10.60785  2.739876e-26
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03281993 23.01613 3.213985e-117
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03066636 26.07625 6.780346e-150
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03443244 23.88089 4.838693e-126
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04529220 14.52941  7.889840e-48
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03765895 19.84719  1.165496e-87
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6022884          0.7378138
#> 2          0.5774791          0.7101592
#> 3          0.7039271          0.7986305
#> 4          0.4254050          0.5956454
#> 5          0.6993111          0.8140821
#> 6          0.7357742          0.8463367
#> 7          0.7636887          0.8986850
#> 8          0.5685106          0.7336134
#> 9          0.6930457          0.8189058

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1776338981
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0384   17.4708    0.0000 [ 0.5645; 0.7633 ] 
#>   eta3 ~ eta1      0.4585      0.0746    6.1439    0.0000 [ 0.2655; 0.6514 ] 
#>   eta3 ~ eta2      0.3052      0.0874    3.4907    0.0005 [ 0.0827; 0.5348 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0380   17.4385    0.0000 [ 0.5581; 0.7547 ] 
#>   eta1 =~ y12      0.6493      0.0387   16.7619    0.0000 [ 0.5441; 0.7444 ] 
#>   eta1 =~ y13      0.7613      0.0308   24.7306    0.0000 [ 0.6853; 0.8445 ] 
#>   eta2 =~ y21      0.5165      0.0487   10.6078    0.0000 [ 0.3837; 0.6354 ] 
#>   eta2 =~ y22      0.7554      0.0328   23.0161    0.0000 [ 0.6689; 0.8387 ] 
#>   eta2 =~ y23      0.7997      0.0307   26.0763    0.0000 [ 0.7210; 0.8796 ] 
#>   eta3 =~ y31      0.8223      0.0344   23.8809    0.0000 [ 0.7256; 0.9036 ] 
#>   eta3 =~ y32      0.6581      0.0453   14.5294    0.0000 [ 0.5383; 0.7725 ] 
#>   eta3 =~ y33      0.7474      0.0377   19.8472    0.0000 [ 0.6516; 0.8464 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0214   18.4523    0.0000 [ 0.3385; 0.4494 ] 
#>   eta1 <~ y12      0.3873      0.0174   22.2759    0.0000 [ 0.3419; 0.4319 ] 
#>   eta1 <~ y13      0.4542      0.0189   24.0775    0.0000 [ 0.4101; 0.5077 ] 
#>   eta2 <~ y21      0.3058      0.0216   14.1472    0.0000 [ 0.2480; 0.3598 ] 
#>   eta2 <~ y22      0.4473      0.0195   22.9720    0.0000 [ 0.3984; 0.4991 ] 
#>   eta2 <~ y23      0.4735      0.0186   25.4223    0.0000 [ 0.4282; 0.5245 ] 
#>   eta3 <~ y31      0.4400      0.0195   22.5165    0.0000 [ 0.3881; 0.4892 ] 
#>   eta3 <~ y32      0.3521      0.0198   17.7475    0.0000 [ 0.3019; 0.4045 ] 
#>   eta3 <~ y33      0.3999      0.0200   19.9949    0.0000 [ 0.3516; 0.4550 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0384   17.4708    0.0000 [ 0.5645; 0.7633 ] 
#>   eta3 ~ eta1       0.6634      0.0284   23.3526    0.0000 [ 0.5904; 0.7373 ] 
#>   eta3 ~ eta2       0.3052      0.0874    3.4907    0.0005 [ 0.0827; 0.5348 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0590    3.4699    0.0005 [ 0.0527; 0.3580 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03842609 17.470772 2.392335e-68
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07462822  6.143879 8.053026e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08741914  3.490667 4.818159e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56453489          0.7632529          0.5883970          0.7393908
#> 2         0.26548421          0.6514192          0.3118275          0.6050760
#> 3         0.08272898          0.5348114          0.1370152          0.4805251
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.6002039          0.7601214          0.6102338          0.7487278
#> 2          0.3197094          0.5975574          0.3253617          0.5933859
#> 3          0.1503352          0.4464484          0.1554274          0.4409217