Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 647694430
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0425   15.7863    0.0000 [ 0.6031; 0.7572 ] 
#>   eta3 ~ eta1      0.4585      0.0996    4.6031    0.0000 [ 0.2972; 0.6311 ] 
#>   eta3 ~ eta2      0.3052      0.1019    2.9941    0.0028 [ 0.1438; 0.4649 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0455   14.5649    0.0000 [ 0.5830; 0.7527 ] 
#>   eta1 =~ y12      0.6493      0.0408   15.8999    0.0000 [ 0.5784; 0.7145 ] 
#>   eta1 =~ y13      0.7613      0.0390   19.5319    0.0000 [ 0.6889; 0.8238 ] 
#>   eta2 =~ y21      0.5165      0.0527    9.7977    0.0000 [ 0.3961; 0.5971 ] 
#>   eta2 =~ y22      0.7554      0.0381   19.8199    0.0000 [ 0.6613; 0.8154 ] 
#>   eta2 =~ y23      0.7997      0.0379   21.0968    0.0000 [ 0.7230; 0.8576 ] 
#>   eta3 =~ y31      0.8223      0.0378   21.7264    0.0000 [ 0.7534; 0.8880 ] 
#>   eta3 =~ y32      0.6581      0.0437   15.0608    0.0000 [ 0.5781; 0.7378 ] 
#>   eta3 =~ y33      0.7474      0.0373   20.0366    0.0000 [ 0.6764; 0.8155 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0230   17.2269    0.0000 [ 0.3573; 0.4317 ] 
#>   eta1 <~ y12      0.3873      0.0239   16.2078    0.0000 [ 0.3539; 0.4340 ] 
#>   eta1 <~ y13      0.4542      0.0239   18.9695    0.0000 [ 0.4146; 0.4997 ] 
#>   eta2 <~ y21      0.3058      0.0280   10.9140    0.0000 [ 0.2433; 0.3536 ] 
#>   eta2 <~ y22      0.4473      0.0226   19.7869    0.0000 [ 0.4179; 0.4911 ] 
#>   eta2 <~ y23      0.4735      0.0218   21.6793    0.0000 [ 0.4445; 0.5179 ] 
#>   eta3 <~ y31      0.4400      0.0191   23.0345    0.0000 [ 0.3999; 0.4665 ] 
#>   eta3 <~ y32      0.3521      0.0205   17.1835    0.0000 [ 0.3108; 0.3793 ] 
#>   eta3 <~ y33      0.3999      0.0219   18.2998    0.0000 [ 0.3690; 0.4387 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0425   15.7863    0.0000 [ 0.6031; 0.7572 ] 
#>   eta3 ~ eta1       0.6634      0.0490   13.5481    0.0000 [ 0.5726; 0.7384 ] 
#>   eta3 ~ eta2       0.3052      0.1019    2.9941    0.0028 [ 0.1438; 0.4649 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0666    3.0773    0.0021 [ 0.0971; 0.3168 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04552508 14.564937  4.694169e-48
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04083544 15.899863  6.350390e-57
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03897955 19.531929  5.877684e-85
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05271188  9.797693  1.151858e-22
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03811250 19.819945  2.003259e-87
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03790446 21.096824  8.506091e-99
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03784692 21.726403 1.155034e-104
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04369405 15.060837  2.930529e-51
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03730297 20.036586  2.642837e-89
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5830059          0.7526625
#> 2          0.5783603          0.7145110
#> 3          0.6888559          0.8238474
#> 4          0.3960890          0.5971027
#> 5          0.6613111          0.8153750
#> 6          0.7229622          0.8575834
#> 7          0.7533534          0.8880447
#> 8          0.5780931          0.7378403
#> 9          0.6763671          0.8154879

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 647694430
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0425   15.7863    0.0000 [ 0.5605; 0.7804 ] 
#>   eta3 ~ eta1      0.4585      0.0996    4.6031    0.0000 [ 0.2068; 0.7219 ] 
#>   eta3 ~ eta2      0.3052      0.1019    2.9941    0.0028 [ 0.0344; 0.5615 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0455   14.5649    0.0000 [ 0.5480; 0.7835 ] 
#>   eta1 =~ y12      0.6493      0.0408   15.8999    0.0000 [ 0.5423; 0.7534 ] 
#>   eta1 =~ y13      0.7613      0.0390   19.5319    0.0000 [ 0.6654; 0.8670 ] 
#>   eta2 =~ y21      0.5165      0.0527    9.7977    0.0000 [ 0.3972; 0.6698 ] 
#>   eta2 =~ y22      0.7554      0.0381   19.8199    0.0000 [ 0.6593; 0.8564 ] 
#>   eta2 =~ y23      0.7997      0.0379   21.0968    0.0000 [ 0.7102; 0.9062 ] 
#>   eta3 =~ y31      0.8223      0.0378   21.7264    0.0000 [ 0.7269; 0.9227 ] 
#>   eta3 =~ y32      0.6581      0.0437   15.0608    0.0000 [ 0.5443; 0.7703 ] 
#>   eta3 =~ y33      0.7474      0.0373   20.0366    0.0000 [ 0.6443; 0.8372 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0230   17.2269    0.0000 [ 0.3365; 0.4552 ] 
#>   eta1 <~ y12      0.3873      0.0239   16.2078    0.0000 [ 0.3232; 0.4468 ] 
#>   eta1 <~ y13      0.4542      0.0239   18.9695    0.0000 [ 0.3934; 0.5172 ] 
#>   eta2 <~ y21      0.3058      0.0280   10.9140    0.0000 [ 0.2384; 0.3833 ] 
#>   eta2 <~ y22      0.4473      0.0226   19.7869    0.0000 [ 0.3822; 0.4991 ] 
#>   eta2 <~ y23      0.4735      0.0218   21.6793    0.0000 [ 0.4136; 0.5266 ] 
#>   eta3 <~ y31      0.4400      0.0191   23.0345    0.0000 [ 0.3937; 0.4924 ] 
#>   eta3 <~ y32      0.3521      0.0205   17.1835    0.0000 [ 0.3002; 0.4062 ] 
#>   eta3 <~ y33      0.3999      0.0219   18.2998    0.0000 [ 0.3413; 0.4543 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0425   15.7863    0.0000 [ 0.5605; 0.7804 ] 
#>   eta3 ~ eta1       0.6634      0.0490   13.5481    0.0000 [ 0.5384; 0.7917 ] 
#>   eta3 ~ eta2       0.3052      0.1019    2.9941    0.0028 [ 0.0344; 0.5615 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0666    3.0773    0.0021 [ 0.0286; 0.3728 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04252621 15.786345 3.863232e-56
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09960808  4.603108 4.162314e-06
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.10191659  2.994126 2.752322e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56051702          0.7804386          0.5869253          0.7540303
#> 2         0.20677844          0.7218952          0.2686339          0.6600397
#> 3         0.03441689          0.5614720          0.0977059          0.4981830
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5990099          0.7744968          0.6030826          0.7572330
#> 2          0.2191434          0.6475255          0.2971683          0.6311255
#> 3          0.1075352          0.5298317          0.1437743          0.4649476