Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1097535634
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0468   14.3555    0.0000 [ 0.5851; 0.7532 ] 
#>   eta3 ~ eta1      0.4585      0.0734    6.2449    0.0000 [ 0.3341; 0.6105 ] 
#>   eta3 ~ eta2      0.3052      0.0767    3.9780    0.0001 [ 0.1875; 0.4539 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0399   16.6346    0.0000 [ 0.6047; 0.7261 ] 
#>   eta1 =~ y12      0.6493      0.0357   18.1831    0.0000 [ 0.6069; 0.7114 ] 
#>   eta1 =~ y13      0.7613      0.0327   23.2878    0.0000 [ 0.7088; 0.8125 ] 
#>   eta2 =~ y21      0.5165      0.0570    9.0570    0.0000 [ 0.3942; 0.5988 ] 
#>   eta2 =~ y22      0.7554      0.0307   24.5890    0.0000 [ 0.7083; 0.8075 ] 
#>   eta2 =~ y23      0.7997      0.0357   22.4177    0.0000 [ 0.7365; 0.8559 ] 
#>   eta3 =~ y31      0.8223      0.0363   22.6741    0.0000 [ 0.7724; 0.9020 ] 
#>   eta3 =~ y32      0.6581      0.0443   14.8519    0.0000 [ 0.5461; 0.7235 ] 
#>   eta3 =~ y33      0.7474      0.0344   21.7359    0.0000 [ 0.6881; 0.8082 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0207   19.0917    0.0000 [ 0.3588; 0.4427 ] 
#>   eta1 <~ y12      0.3873      0.0181   21.4202    0.0000 [ 0.3599; 0.4191 ] 
#>   eta1 <~ y13      0.4542      0.0207   21.9125    0.0000 [ 0.4191; 0.4912 ] 
#>   eta2 <~ y21      0.3058      0.0293   10.4215    0.0000 [ 0.2407; 0.3558 ] 
#>   eta2 <~ y22      0.4473      0.0186   24.0628    0.0000 [ 0.4257; 0.4844 ] 
#>   eta2 <~ y23      0.4735      0.0235   20.1142    0.0000 [ 0.4302; 0.5148 ] 
#>   eta3 <~ y31      0.4400      0.0204   21.5665    0.0000 [ 0.4207; 0.4854 ] 
#>   eta3 <~ y32      0.3521      0.0194   18.1364    0.0000 [ 0.3032; 0.3822 ] 
#>   eta3 <~ y33      0.3999      0.0182   22.0254    0.0000 [ 0.3681; 0.4284 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0468   14.3555    0.0000 [ 0.5851; 0.7532 ] 
#>   eta3 ~ eta1       0.6634      0.0354   18.7164    0.0000 [ 0.5928; 0.7083 ] 
#>   eta3 ~ eta2       0.3052      0.0767    3.9780    0.0001 [ 0.1875; 0.4539 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0544    3.7657    0.0002 [ 0.1181; 0.3059 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03986099 16.634555  3.916520e-62
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03570776 18.183102  7.024506e-74
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03269286 23.287828 5.889562e-120
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05702245  9.057043  1.340351e-19
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03072050 24.589043 1.654616e-133
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03567108 22.417706 2.644694e-111
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03626499 22.674139 8.064478e-114
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04430888 14.851851  6.767375e-50
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03438662 21.735900 9.392457e-105
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6047135          0.7260998
#> 2          0.6068653          0.7113712
#> 3          0.7088101          0.8124983
#> 4          0.3941751          0.5988294
#> 5          0.7082854          0.8075037
#> 6          0.7364719          0.8558807
#> 7          0.7724387          0.9019700
#> 8          0.5460892          0.7234709
#> 9          0.6880979          0.8082091

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1097535634
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0468   14.3555    0.0000 [ 0.5518; 0.7936 ] 
#>   eta3 ~ eta1      0.4585      0.0734    6.2449    0.0000 [ 0.2752; 0.6549 ] 
#>   eta3 ~ eta2      0.3052      0.0767    3.9780    0.0001 [ 0.1006; 0.4973 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0399   16.6346    0.0000 [ 0.5570; 0.7632 ] 
#>   eta1 =~ y12      0.6493      0.0357   18.1831    0.0000 [ 0.5569; 0.7415 ] 
#>   eta1 =~ y13      0.7613      0.0327   23.2878    0.0000 [ 0.6739; 0.8429 ] 
#>   eta2 =~ y21      0.5165      0.0570    9.0570    0.0000 [ 0.3736; 0.6685 ] 
#>   eta2 =~ y22      0.7554      0.0307   24.5890    0.0000 [ 0.6673; 0.8262 ] 
#>   eta2 =~ y23      0.7997      0.0357   22.4177    0.0000 [ 0.7129; 0.8974 ] 
#>   eta3 =~ y31      0.8223      0.0363   22.6741    0.0000 [ 0.7197; 0.9072 ] 
#>   eta3 =~ y32      0.6581      0.0443   14.8519    0.0000 [ 0.5529; 0.7821 ] 
#>   eta3 =~ y33      0.7474      0.0344   21.7359    0.0000 [ 0.6632; 0.8411 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0207   19.0917    0.0000 [ 0.3422; 0.4493 ] 
#>   eta1 <~ y12      0.3873      0.0181   21.4202    0.0000 [ 0.3425; 0.4360 ] 
#>   eta1 <~ y13      0.4542      0.0207   21.9125    0.0000 [ 0.4009; 0.5081 ] 
#>   eta2 <~ y21      0.3058      0.0293   10.4215    0.0000 [ 0.2332; 0.3850 ] 
#>   eta2 <~ y22      0.4473      0.0186   24.0628    0.0000 [ 0.3943; 0.4904 ] 
#>   eta2 <~ y23      0.4735      0.0235   20.1142    0.0000 [ 0.4160; 0.5378 ] 
#>   eta3 <~ y31      0.4400      0.0204   21.5665    0.0000 [ 0.3812; 0.4867 ] 
#>   eta3 <~ y32      0.3521      0.0194   18.1364    0.0000 [ 0.3062; 0.4066 ] 
#>   eta3 <~ y33      0.3999      0.0182   22.0254    0.0000 [ 0.3544; 0.4483 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0468   14.3555    0.0000 [ 0.5518; 0.7936 ] 
#>   eta3 ~ eta1       0.6634      0.0354   18.7164    0.0000 [ 0.5745; 0.7578 ] 
#>   eta3 ~ eta2       0.3052      0.0767    3.9780    0.0001 [ 0.1006; 0.4973 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0544    3.7657    0.0002 [ 0.0604; 0.3418 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04676481 14.355526 9.837623e-47
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07342063  6.244931 4.239867e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07670914  3.978028 6.948916e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5517519          0.7935931          0.5807923          0.7645527
#> 2          0.2752323          0.6549223          0.3208256          0.6093290
#> 3          0.1005901          0.4972865          0.1482256          0.4496510
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5613116          0.7567317          0.5851187          0.7531627
#> 2          0.3196611          0.6138244          0.3341016          0.6104663
#> 3          0.1358898          0.4667720          0.1874617          0.4538622