Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1302145400
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0373   18.0097    0.0000 [ 0.5992; 0.7363 ] 
#>   eta3 ~ eta1      0.4585      0.1066    4.2999    0.0000 [ 0.2993; 0.6842 ] 
#>   eta3 ~ eta2      0.3052      0.1030    2.9616    0.0031 [ 0.0904; 0.4656 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0469   14.1257    0.0000 [ 0.5734; 0.7411 ] 
#>   eta1 =~ y12      0.6493      0.0463   14.0177    0.0000 [ 0.5564; 0.7201 ] 
#>   eta1 =~ y13      0.7613      0.0338   22.5136    0.0000 [ 0.7078; 0.8283 ] 
#>   eta2 =~ y21      0.5165      0.0531    9.7259    0.0000 [ 0.4191; 0.6070 ] 
#>   eta2 =~ y22      0.7554      0.0428   17.6381    0.0000 [ 0.6658; 0.8150 ] 
#>   eta2 =~ y23      0.7997      0.0345   23.1726    0.0000 [ 0.7532; 0.8752 ] 
#>   eta3 =~ y31      0.8223      0.0289   28.4358    0.0000 [ 0.7836; 0.8893 ] 
#>   eta3 =~ y32      0.6581      0.0486   13.5388    0.0000 [ 0.5764; 0.7414 ] 
#>   eta3 =~ y33      0.7474      0.0368   20.2953    0.0000 [ 0.6943; 0.8244 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0223   17.7726    0.0000 [ 0.3527; 0.4320 ] 
#>   eta1 <~ y12      0.3873      0.0237   16.3662    0.0000 [ 0.3381; 0.4209 ] 
#>   eta1 <~ y13      0.4542      0.0240   18.9052    0.0000 [ 0.4282; 0.5051 ] 
#>   eta2 <~ y21      0.3058      0.0265   11.5588    0.0000 [ 0.2601; 0.3554 ] 
#>   eta2 <~ y22      0.4473      0.0254   17.5834    0.0000 [ 0.3963; 0.4847 ] 
#>   eta2 <~ y23      0.4735      0.0189   25.0881    0.0000 [ 0.4387; 0.5050 ] 
#>   eta3 <~ y31      0.4400      0.0190   23.1749    0.0000 [ 0.4100; 0.4769 ] 
#>   eta3 <~ y32      0.3521      0.0190   18.4901    0.0000 [ 0.3188; 0.3819 ] 
#>   eta3 <~ y33      0.3999      0.0199   20.0702    0.0000 [ 0.3642; 0.4307 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0373   18.0097    0.0000 [ 0.5992; 0.7363 ] 
#>   eta3 ~ eta1       0.6634      0.0499   13.2989    0.0000 [ 0.5590; 0.7412 ] 
#>   eta3 ~ eta2       0.3052      0.1030    2.9616    0.0031 [ 0.0904; 0.4656 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0692    2.9618    0.0031 [ 0.0655; 0.3050 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04694057 14.125730  2.636524e-45
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04631835 14.017727  1.214418e-44
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03381720 22.513565 3.056758e-112
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05310102  9.725892  2.338468e-22
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04282706 17.638092  1.256316e-69
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03450909 23.172553 8.614190e-119
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02891696 28.435820 7.297673e-178
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04860597 13.538848  9.222350e-42
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03682753 20.295256  1.416159e-91
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5734226          0.7410659
#> 2          0.5564189          0.7201075
#> 3          0.7078098          0.8282671
#> 4          0.4191234          0.6070453
#> 5          0.6658473          0.8149570
#> 6          0.7531732          0.8752150
#> 7          0.7835719          0.8892846
#> 8          0.5763833          0.7414312
#> 9          0.6943463          0.8244057

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1302145400
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0373   18.0097    0.0000 [ 0.5716; 0.7644 ] 
#>   eta3 ~ eta1      0.4585      0.1066    4.2999    0.0000 [ 0.1864; 0.7378 ] 
#>   eta3 ~ eta2      0.3052      0.1030    2.9616    0.0031 [ 0.0369; 0.5697 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0469   14.1257    0.0000 [ 0.5482; 0.7909 ] 
#>   eta1 =~ y12      0.6493      0.0463   14.0177    0.0000 [ 0.5378; 0.7773 ] 
#>   eta1 =~ y13      0.7613      0.0338   22.5136    0.0000 [ 0.6696; 0.8445 ] 
#>   eta2 =~ y21      0.5165      0.0531    9.7259    0.0000 [ 0.3719; 0.6465 ] 
#>   eta2 =~ y22      0.7554      0.0428   17.6381    0.0000 [ 0.6497; 0.8712 ] 
#>   eta2 =~ y23      0.7997      0.0345   23.1726    0.0000 [ 0.7094; 0.8879 ] 
#>   eta3 =~ y31      0.8223      0.0289   28.4358    0.0000 [ 0.7425; 0.8920 ] 
#>   eta3 =~ y32      0.6581      0.0486   13.5388    0.0000 [ 0.5353; 0.7867 ] 
#>   eta3 =~ y33      0.7474      0.0368   20.2953    0.0000 [ 0.6490; 0.8394 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0223   17.7726    0.0000 [ 0.3395; 0.4546 ] 
#>   eta1 <~ y12      0.3873      0.0237   16.3662    0.0000 [ 0.3287; 0.4511 ] 
#>   eta1 <~ y13      0.4542      0.0240   18.9052    0.0000 [ 0.3862; 0.5104 ] 
#>   eta2 <~ y21      0.3058      0.0265   11.5588    0.0000 [ 0.2342; 0.3710 ] 
#>   eta2 <~ y22      0.4473      0.0254   17.5834    0.0000 [ 0.3855; 0.5170 ] 
#>   eta2 <~ y23      0.4735      0.0189   25.0881    0.0000 [ 0.4251; 0.5227 ] 
#>   eta3 <~ y31      0.4400      0.0190   23.1749    0.0000 [ 0.3899; 0.4881 ] 
#>   eta3 <~ y32      0.3521      0.0190   18.4901    0.0000 [ 0.3063; 0.4048 ] 
#>   eta3 <~ y33      0.3999      0.0199   20.0702    0.0000 [ 0.3483; 0.4514 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0373   18.0097    0.0000 [ 0.5716; 0.7644 ] 
#>   eta3 ~ eta1       0.6634      0.0499   13.2989    0.0000 [ 0.5362; 0.7942 ] 
#>   eta3 ~ eta2       0.3052      0.1030    2.9616    0.0031 [ 0.0369; 0.5697 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0692    2.9618    0.0031 [ 0.0243; 0.3819 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03727622 18.009698 1.635176e-72
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.10663186  4.299904 1.708722e-05
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.10303624  2.961590 3.060549e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.57162437          0.7643959          0.5947725          0.7412478
#> 2         0.18638547          0.7378253          0.2526026          0.6716081
#> 3         0.03690425          0.5697495          0.1008885          0.5057652
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.59404543          0.7430783         0.59916759          0.7363051
#> 2         0.26765335          0.6969241         0.29925324          0.6842432
#> 3         0.06005369          0.4800881         0.09036404          0.4656071