Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1507926919
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0396   16.9374    0.0000 [ 0.5979; 0.7429 ] 
#>   eta3 ~ eta1      0.4585      0.0793    5.7820    0.0000 [ 0.3356; 0.6106 ] 
#>   eta3 ~ eta2      0.3052      0.0941    3.2414    0.0012 [ 0.1462; 0.4673 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0453   14.6304    0.0000 [ 0.5868; 0.7411 ] 
#>   eta1 =~ y12      0.6493      0.0335   19.3532    0.0000 [ 0.5732; 0.6904 ] 
#>   eta1 =~ y13      0.7613      0.0414   18.3959    0.0000 [ 0.6737; 0.8221 ] 
#>   eta2 =~ y21      0.5165      0.0598    8.6303    0.0000 [ 0.4173; 0.6359 ] 
#>   eta2 =~ y22      0.7554      0.0392   19.2835    0.0000 [ 0.6859; 0.8536 ] 
#>   eta2 =~ y23      0.7997      0.0384   20.8335    0.0000 [ 0.7467; 0.8807 ] 
#>   eta3 =~ y31      0.8223      0.0305   26.9520    0.0000 [ 0.7755; 0.8813 ] 
#>   eta3 =~ y32      0.6581      0.0386   17.0284    0.0000 [ 0.5718; 0.7064 ] 
#>   eta3 =~ y33      0.7474      0.0453   16.4925    0.0000 [ 0.6650; 0.8130 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0227   17.4598    0.0000 [ 0.3667; 0.4489 ] 
#>   eta1 <~ y12      0.3873      0.0151   25.6269    0.0000 [ 0.3628; 0.4107 ] 
#>   eta1 <~ y13      0.4542      0.0251   18.0756    0.0000 [ 0.4115; 0.4978 ] 
#>   eta2 <~ y21      0.3058      0.0292   10.4595    0.0000 [ 0.2553; 0.3613 ] 
#>   eta2 <~ y22      0.4473      0.0265   16.8792    0.0000 [ 0.4080; 0.4987 ] 
#>   eta2 <~ y23      0.4735      0.0226   20.9789    0.0000 [ 0.4329; 0.5150 ] 
#>   eta3 <~ y31      0.4400      0.0202   21.7556    0.0000 [ 0.4094; 0.4917 ] 
#>   eta3 <~ y32      0.3521      0.0174   20.2064    0.0000 [ 0.3136; 0.3793 ] 
#>   eta3 <~ y33      0.3999      0.0202   19.7975    0.0000 [ 0.3681; 0.4336 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0396   16.9374    0.0000 [ 0.5979; 0.7429 ] 
#>   eta3 ~ eta1       0.6634      0.0348   19.0381    0.0000 [ 0.6025; 0.7233 ] 
#>   eta3 ~ eta2       0.3052      0.0941    3.2414    0.0012 [ 0.1462; 0.4673 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0675    3.0341    0.0024 [ 0.1043; 0.3298 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04532136 14.630406  1.797103e-48
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03354885 19.353212  1.915261e-83
#> 3 eta1 =~ y13  Common factor 0.7613458 0.04138669 18.395907  1.416693e-75
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05984183  8.630331  6.117485e-18
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03917273 19.283507  7.389275e-83
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03838359 20.833481  2.152340e-96
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03050891 26.952036 5.399308e-160
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03864536 17.028406  5.056266e-65
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04531912 16.492467  4.156242e-61
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5868254          0.7410967
#> 2          0.5731812          0.6903758
#> 3          0.6736773          0.8220990
#> 4          0.4172736          0.6359469
#> 5          0.6859171          0.8535984
#> 6          0.7467437          0.8807140
#> 7          0.7754976          0.8813449
#> 8          0.5718014          0.7063736
#> 9          0.6649776          0.8130411

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1507926919
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0396   16.9374    0.0000 [ 0.5631; 0.7681 ] 
#>   eta3 ~ eta1      0.4585      0.0793    5.7820    0.0000 [ 0.2518; 0.6619 ] 
#>   eta3 ~ eta2      0.3052      0.0941    3.2414    0.0012 [ 0.0586; 0.5455 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0453   14.6304    0.0000 [ 0.5452; 0.7796 ] 
#>   eta1 =~ y12      0.6493      0.0335   19.3532    0.0000 [ 0.5709; 0.7444 ] 
#>   eta1 =~ y13      0.7613      0.0414   18.3959    0.0000 [ 0.6554; 0.8695 ] 
#>   eta2 =~ y21      0.5165      0.0598    8.6303    0.0000 [ 0.3596; 0.6691 ] 
#>   eta2 =~ y22      0.7554      0.0392   19.2835    0.0000 [ 0.6523; 0.8549 ] 
#>   eta2 =~ y23      0.7997      0.0384   20.8335    0.0000 [ 0.7019; 0.9004 ] 
#>   eta3 =~ y31      0.8223      0.0305   26.9520    0.0000 [ 0.7425; 0.9002 ] 
#>   eta3 =~ y32      0.6581      0.0386   17.0284    0.0000 [ 0.5619; 0.7617 ] 
#>   eta3 =~ y33      0.7474      0.0453   16.4925    0.0000 [ 0.6371; 0.8714 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0227   17.4598    0.0000 [ 0.3344; 0.4516 ] 
#>   eta1 <~ y12      0.3873      0.0151   25.6269    0.0000 [ 0.3511; 0.4293 ] 
#>   eta1 <~ y13      0.4542      0.0251   18.0756    0.0000 [ 0.3871; 0.5170 ] 
#>   eta2 <~ y21      0.3058      0.0292   10.4595    0.0000 [ 0.2303; 0.3815 ] 
#>   eta2 <~ y22      0.4473      0.0265   16.8792    0.0000 [ 0.3786; 0.5157 ] 
#>   eta2 <~ y23      0.4735      0.0226   20.9789    0.0000 [ 0.4172; 0.5339 ] 
#>   eta3 <~ y31      0.4400      0.0202   21.7556    0.0000 [ 0.3843; 0.4889 ] 
#>   eta3 <~ y32      0.3521      0.0174   20.2064    0.0000 [ 0.3071; 0.3972 ] 
#>   eta3 <~ y33      0.3999      0.0202   19.7975    0.0000 [ 0.3491; 0.4536 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0396   16.9374    0.0000 [ 0.5631; 0.7681 ] 
#>   eta3 ~ eta1       0.6634      0.0348   19.0381    0.0000 [ 0.5669; 0.7470 ] 
#>   eta3 ~ eta2       0.3052      0.0941    3.2414    0.0012 [ 0.0586; 0.5455 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0675    3.0341    0.0024 [ 0.0255; 0.3747 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03963612 16.937417 2.383630e-64
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07929889  5.782007 7.381454e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09414046  3.241445 1.189253e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56310843          0.7680840          0.5877220          0.7434705
#> 2         0.25178035          0.6618694          0.3010240          0.6126258
#> 3         0.05862678          0.5454681          0.1170869          0.4870079
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5836055          0.7460542          0.5979493          0.7428543
#> 2          0.3054810          0.6172754          0.3356452          0.6106091
#> 3          0.1311901          0.4754000          0.1462456          0.4672864