Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1564251249
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0451   14.8928    0.0000 [ 0.6010; 0.7390 ] 
#>   eta3 ~ eta1      0.4585      0.0839    5.4637    0.0000 [ 0.3226; 0.6686 ] 
#>   eta3 ~ eta2      0.3052      0.0880    3.4669    0.0005 [ 0.1116; 0.4463 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0372   17.8373    0.0000 [ 0.5948; 0.7234 ] 
#>   eta1 =~ y12      0.6493      0.0413   15.7121    0.0000 [ 0.5882; 0.7206 ] 
#>   eta1 =~ y13      0.7613      0.0242   31.4747    0.0000 [ 0.7296; 0.8091 ] 
#>   eta2 =~ y21      0.5165      0.0645    8.0103    0.0000 [ 0.3916; 0.6388 ] 
#>   eta2 =~ y22      0.7554      0.0308   24.5017    0.0000 [ 0.7065; 0.8091 ] 
#>   eta2 =~ y23      0.7997      0.0331   24.1926    0.0000 [ 0.7315; 0.8548 ] 
#>   eta3 =~ y31      0.8223      0.0387   21.2608    0.0000 [ 0.7470; 0.8799 ] 
#>   eta3 =~ y32      0.6581      0.0360   18.3013    0.0000 [ 0.6033; 0.7303 ] 
#>   eta3 =~ y33      0.7474      0.0450   16.5935    0.0000 [ 0.6713; 0.8302 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0187   21.1328    0.0000 [ 0.3564; 0.4232 ] 
#>   eta1 <~ y12      0.3873      0.0191   20.2301    0.0000 [ 0.3543; 0.4160 ] 
#>   eta1 <~ y13      0.4542      0.0190   23.9408    0.0000 [ 0.4296; 0.4878 ] 
#>   eta2 <~ y21      0.3058      0.0311    9.8488    0.0000 [ 0.2448; 0.3591 ] 
#>   eta2 <~ y22      0.4473      0.0241   18.5651    0.0000 [ 0.3971; 0.4836 ] 
#>   eta2 <~ y23      0.4735      0.0219   21.6546    0.0000 [ 0.4399; 0.5117 ] 
#>   eta3 <~ y31      0.4400      0.0193   22.8425    0.0000 [ 0.4074; 0.4770 ] 
#>   eta3 <~ y32      0.3521      0.0142   24.7566    0.0000 [ 0.3288; 0.3817 ] 
#>   eta3 <~ y33      0.3999      0.0252   15.8481    0.0000 [ 0.3607; 0.4436 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0451   14.8928    0.0000 [ 0.6010; 0.7390 ] 
#>   eta3 ~ eta1       0.6634      0.0432   15.3701    0.0000 [ 0.5958; 0.7663 ] 
#>   eta3 ~ eta2       0.3052      0.0880    3.4669    0.0005 [ 0.1116; 0.4463 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0561    3.6500    0.0003 [ 0.0797; 0.2956 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03717326 17.837278  3.629491e-71
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04132339 15.712115  1.249404e-55
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02418913 31.474713 1.927470e-217
#> 4 eta2 =~ y21  Common factor 0.5164548 0.06447358  8.010333  1.143986e-15
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03082999 24.501718 1.416172e-132
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03305401 24.192635 2.659496e-129
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03867571 21.260823 2.617852e-100
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03595746 18.301316  8.077392e-75
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04504324 16.593482  7.768369e-62
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5947911          0.7233852
#> 2          0.5881617          0.7206028
#> 3          0.7295656          0.8091312
#> 4          0.3916463          0.6388025
#> 5          0.7064774          0.8090916
#> 6          0.7314666          0.8548056
#> 7          0.7470072          0.8798513
#> 8          0.6033309          0.7302703
#> 9          0.6713288          0.8302372

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1564251249
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0451   14.8928    0.0000 [ 0.5509; 0.7840 ] 
#>   eta3 ~ eta1      0.4585      0.0839    5.4637    0.0000 [ 0.2275; 0.6615 ] 
#>   eta3 ~ eta2      0.3052      0.0880    3.4669    0.0005 [ 0.0801; 0.5353 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0372   17.8373    0.0000 [ 0.5782; 0.7704 ] 
#>   eta1 =~ y12      0.6493      0.0413   15.7121    0.0000 [ 0.5453; 0.7590 ] 
#>   eta1 =~ y13      0.7613      0.0242   31.4747    0.0000 [ 0.6973; 0.8224 ] 
#>   eta2 =~ y21      0.5165      0.0645    8.0103    0.0000 [ 0.3534; 0.6869 ] 
#>   eta2 =~ y22      0.7554      0.0308   24.5017    0.0000 [ 0.6776; 0.8370 ] 
#>   eta2 =~ y23      0.7997      0.0331   24.1926    0.0000 [ 0.7083; 0.8793 ] 
#>   eta3 =~ y31      0.8223      0.0387   21.2608    0.0000 [ 0.7316; 0.9316 ] 
#>   eta3 =~ y32      0.6581      0.0360   18.3013    0.0000 [ 0.5582; 0.7442 ] 
#>   eta3 =~ y33      0.7474      0.0450   16.5935    0.0000 [ 0.6314; 0.8644 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0187   21.1328    0.0000 [ 0.3507; 0.4475 ] 
#>   eta1 <~ y12      0.3873      0.0191   20.2301    0.0000 [ 0.3365; 0.4355 ] 
#>   eta1 <~ y13      0.4542      0.0190   23.9408    0.0000 [ 0.4001; 0.4982 ] 
#>   eta2 <~ y21      0.3058      0.0311    9.8488    0.0000 [ 0.2292; 0.3897 ] 
#>   eta2 <~ y22      0.4473      0.0241   18.5651    0.0000 [ 0.3867; 0.5113 ] 
#>   eta2 <~ y23      0.4735      0.0219   21.6546    0.0000 [ 0.4143; 0.5274 ] 
#>   eta3 <~ y31      0.4400      0.0193   22.8425    0.0000 [ 0.3943; 0.4939 ] 
#>   eta3 <~ y32      0.3521      0.0142   24.7566    0.0000 [ 0.3111; 0.3847 ] 
#>   eta3 <~ y33      0.3999      0.0252   15.8481    0.0000 [ 0.3340; 0.4645 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0451   14.8928    0.0000 [ 0.5509; 0.7840 ] 
#>   eta3 ~ eta1       0.6634      0.0432   15.3701    0.0000 [ 0.5398; 0.7630 ] 
#>   eta3 ~ eta2       0.3052      0.0880    3.4669    0.0005 [ 0.0801; 0.5353 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0561    3.6500    0.0003 [ 0.0618; 0.3520 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04507773 14.892796 3.670924e-50
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08391828  5.463729 4.662348e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08801911  3.466874 5.265495e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55086750          0.7839841          0.5788602          0.7559913
#> 2         0.22750463          0.6614826          0.2796169          0.6093703
#> 3         0.08007216          0.5352573          0.1347310          0.4805985
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.57532480          0.7507934          0.6009971          0.7390198
#> 2         0.31194733          0.6920643          0.3225650          0.6685590
#> 3         0.05721632          0.4631221          0.1115927          0.4462921