Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 786534843
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0495   13.5594    0.0000 [ 0.5739; 0.7608 ] 
#>   eta3 ~ eta1      0.4585      0.0977    4.6933    0.0000 [ 0.3019; 0.6416 ] 
#>   eta3 ~ eta2      0.3052      0.1120    2.7256    0.0064 [ 0.0806; 0.4979 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0379   17.4967    0.0000 [ 0.5835; 0.7198 ] 
#>   eta1 =~ y12      0.6493      0.0327   19.8776    0.0000 [ 0.5892; 0.7095 ] 
#>   eta1 =~ y13      0.7613      0.0312   24.3942    0.0000 [ 0.6982; 0.8099 ] 
#>   eta2 =~ y21      0.5165      0.0563    9.1773    0.0000 [ 0.4320; 0.6280 ] 
#>   eta2 =~ y22      0.7554      0.0387   19.5211    0.0000 [ 0.6890; 0.8369 ] 
#>   eta2 =~ y23      0.7997      0.0332   24.0662    0.0000 [ 0.7428; 0.8554 ] 
#>   eta3 =~ y31      0.8223      0.0378   21.7595    0.0000 [ 0.7500; 0.8848 ] 
#>   eta3 =~ y32      0.6581      0.0425   15.4851    0.0000 [ 0.5697; 0.7315 ] 
#>   eta3 =~ y33      0.7474      0.0371   20.1281    0.0000 [ 0.7022; 0.8303 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0164   24.1558    0.0000 [ 0.3656; 0.4220 ] 
#>   eta1 <~ y12      0.3873      0.0178   21.7514    0.0000 [ 0.3594; 0.4231 ] 
#>   eta1 <~ y13      0.4542      0.0180   25.2539    0.0000 [ 0.4227; 0.4842 ] 
#>   eta2 <~ y21      0.3058      0.0257   11.8950    0.0000 [ 0.2593; 0.3466 ] 
#>   eta2 <~ y22      0.4473      0.0251   17.8151    0.0000 [ 0.3974; 0.4875 ] 
#>   eta2 <~ y23      0.4735      0.0218   21.7662    0.0000 [ 0.4373; 0.5153 ] 
#>   eta3 <~ y31      0.4400      0.0211   20.8920    0.0000 [ 0.3999; 0.4715 ] 
#>   eta3 <~ y32      0.3521      0.0206   17.0792    0.0000 [ 0.3113; 0.3887 ] 
#>   eta3 <~ y33      0.3999      0.0193   20.7236    0.0000 [ 0.3810; 0.4486 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0495   13.5594    0.0000 [ 0.5739; 0.7608 ] 
#>   eta3 ~ eta1       0.6634      0.0417   15.9056    0.0000 [ 0.5848; 0.7474 ] 
#>   eta3 ~ eta2       0.3052      0.1120    2.7256    0.0064 [ 0.0806; 0.4979 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0770    2.6607    0.0078 [ 0.0581; 0.3564 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03789678 17.496735  1.517197e-68
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03266383 19.877582  6.362699e-88
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03121010 24.394210 1.970140e-131
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05627505  9.177331  4.419065e-20
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03869603 19.521064  7.270871e-85
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03322765 24.066216 5.647375e-128
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03778938 21.759482 5.617987e-105
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04249693 15.485091  4.374428e-54
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03713333 20.128121  4.185683e-90
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5834741          0.7198487
#> 2          0.5891911          0.7095158
#> 3          0.6981876          0.8099420
#> 4          0.4319959          0.6280430
#> 5          0.6889558          0.8368845
#> 6          0.7427827          0.8554497
#> 7          0.7500224          0.8847508
#> 8          0.5697399          0.7315159
#> 9          0.7021653          0.8302698

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 786534843
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0495   13.5594    0.0000 [ 0.5446; 0.8006 ] 
#>   eta3 ~ eta1      0.4585      0.0977    4.6933    0.0000 [ 0.2036; 0.7088 ] 
#>   eta3 ~ eta2      0.3052      0.1120    2.7256    0.0064 [ 0.0138; 0.5927 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0379   17.4967    0.0000 [ 0.5697; 0.7657 ] 
#>   eta1 =~ y12      0.6493      0.0327   19.8776    0.0000 [ 0.5640; 0.7329 ] 
#>   eta1 =~ y13      0.7613      0.0312   24.3942    0.0000 [ 0.6842; 0.8456 ] 
#>   eta2 =~ y21      0.5165      0.0563    9.1773    0.0000 [ 0.3716; 0.6627 ] 
#>   eta2 =~ y22      0.7554      0.0387   19.5211    0.0000 [ 0.6510; 0.8511 ] 
#>   eta2 =~ y23      0.7997      0.0332   24.0662    0.0000 [ 0.7223; 0.8941 ] 
#>   eta3 =~ y31      0.8223      0.0378   21.7595    0.0000 [ 0.7346; 0.9300 ] 
#>   eta3 =~ y32      0.6581      0.0425   15.4851    0.0000 [ 0.5578; 0.7776 ] 
#>   eta3 =~ y33      0.7474      0.0371   20.1281    0.0000 [ 0.6453; 0.8373 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0164   24.1558    0.0000 [ 0.3542; 0.4389 ] 
#>   eta1 <~ y12      0.3873      0.0178   21.7514    0.0000 [ 0.3388; 0.4309 ] 
#>   eta1 <~ y13      0.4542      0.0180   25.2539    0.0000 [ 0.4074; 0.5004 ] 
#>   eta2 <~ y21      0.3058      0.0257   11.8950    0.0000 [ 0.2396; 0.3725 ] 
#>   eta2 <~ y22      0.4473      0.0251   17.8151    0.0000 [ 0.3784; 0.5082 ] 
#>   eta2 <~ y23      0.4735      0.0218   21.7662    0.0000 [ 0.4209; 0.5334 ] 
#>   eta3 <~ y31      0.4400      0.0211   20.8920    0.0000 [ 0.3871; 0.4960 ] 
#>   eta3 <~ y32      0.3521      0.0206   17.0792    0.0000 [ 0.3011; 0.4077 ] 
#>   eta3 <~ y33      0.3999      0.0193   20.7236    0.0000 [ 0.3433; 0.4431 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0495   13.5594    0.0000 [ 0.5446; 0.8006 ] 
#>   eta3 ~ eta1       0.6634      0.0417   15.9056    0.0000 [ 0.5523; 0.7679 ] 
#>   eta3 ~ eta2       0.3052      0.1120    2.7256    0.0064 [ 0.0138; 0.5927 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0770    2.6607    0.0078 [ 0.0048; 0.4030 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04951070 13.559360 6.974187e-42
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09769477  4.693258 2.688882e-06
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.11195715  2.725606 6.418350e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54460252          0.8006439         0.57534809          0.7698984
#> 2         0.20358069          0.7088029         0.26424800          0.6481356
#> 3         0.01376287          0.5927421         0.08328695          0.5232180
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.55721439          0.7705450         0.57394172          0.7607776
#> 2         0.27853418          0.6465338         0.30187647          0.6415808
#> 3         0.05312887          0.5155412         0.08064709          0.4978928