Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 2138507617
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0440   15.2707    0.0000 [ 0.6017; 0.7548 ] 
#>   eta3 ~ eta1      0.4585      0.0822    5.5773    0.0000 [ 0.3222; 0.5892 ] 
#>   eta3 ~ eta2      0.3052      0.0889    3.4328    0.0006 [ 0.1562; 0.4665 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0419   15.8275    0.0000 [ 0.5964; 0.7372 ] 
#>   eta1 =~ y12      0.6493      0.0362   17.9195    0.0000 [ 0.5902; 0.7217 ] 
#>   eta1 =~ y13      0.7613      0.0290   26.2249    0.0000 [ 0.7102; 0.8193 ] 
#>   eta2 =~ y21      0.5165      0.0443   11.6648    0.0000 [ 0.4297; 0.5970 ] 
#>   eta2 =~ y22      0.7554      0.0340   22.2233    0.0000 [ 0.6861; 0.8025 ] 
#>   eta2 =~ y23      0.7997      0.0439   18.1977    0.0000 [ 0.7369; 0.8776 ] 
#>   eta3 =~ y31      0.8223      0.0323   25.4734    0.0000 [ 0.7607; 0.8740 ] 
#>   eta3 =~ y32      0.6581      0.0471   13.9739    0.0000 [ 0.5725; 0.7316 ] 
#>   eta3 =~ y33      0.7474      0.0406   18.4191    0.0000 [ 0.6717; 0.8269 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0196   20.1874    0.0000 [ 0.3644; 0.4268 ] 
#>   eta1 <~ y12      0.3873      0.0174   22.2625    0.0000 [ 0.3578; 0.4150 ] 
#>   eta1 <~ y13      0.4542      0.0189   23.9801    0.0000 [ 0.4204; 0.4903 ] 
#>   eta2 <~ y21      0.3058      0.0266   11.4822    0.0000 [ 0.2660; 0.3668 ] 
#>   eta2 <~ y22      0.4473      0.0190   23.5140    0.0000 [ 0.4109; 0.4787 ] 
#>   eta2 <~ y23      0.4735      0.0211   22.4407    0.0000 [ 0.4364; 0.5094 ] 
#>   eta3 <~ y31      0.4400      0.0201   21.8657    0.0000 [ 0.3993; 0.4685 ] 
#>   eta3 <~ y32      0.3521      0.0205   17.1864    0.0000 [ 0.3087; 0.3852 ] 
#>   eta3 <~ y33      0.3999      0.0218   18.3050    0.0000 [ 0.3615; 0.4304 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0440   15.2707    0.0000 [ 0.6017; 0.7548 ] 
#>   eta3 ~ eta1       0.6634      0.0373   17.7831    0.0000 [ 0.6038; 0.7300 ] 
#>   eta3 ~ eta2       0.3052      0.0889    3.4328    0.0006 [ 0.1562; 0.4665 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0608    3.3689    0.0008 [ 0.1064; 0.3291 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04189364 15.82746  2.011858e-56
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03623308 17.91948  8.310220e-72
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02903146 26.22486 1.383770e-151
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04427483 11.66475  1.929689e-31
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03399075 22.22333 2.043428e-109
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04394312 18.19770  5.381909e-74
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03227989 25.47336 3.890654e-143
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04709268 13.97391  2.249269e-44
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04057883 18.41906  9.238733e-76
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5964146          0.7372320
#> 2          0.5901866          0.7216993
#> 3          0.7102030          0.8193204
#> 4          0.4297282          0.5969824
#> 5          0.6861289          0.8024989
#> 6          0.7368950          0.8775552
#> 7          0.7606714          0.8739742
#> 8          0.5725312          0.7316426
#> 9          0.6717193          0.8268578

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 2138507617
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0440   15.2707    0.0000 [ 0.5498; 0.7772 ] 
#>   eta3 ~ eta1      0.4585      0.0822    5.5773    0.0000 [ 0.2319; 0.6570 ] 
#>   eta3 ~ eta2      0.3052      0.0889    3.4328    0.0006 [ 0.0881; 0.5478 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0419   15.8275    0.0000 [ 0.5504; 0.7671 ] 
#>   eta1 =~ y12      0.6493      0.0362   17.9195    0.0000 [ 0.5583; 0.7456 ] 
#>   eta1 =~ y13      0.7613      0.0290   26.2249    0.0000 [ 0.6837; 0.8339 ] 
#>   eta2 =~ y21      0.5165      0.0443   11.6648    0.0000 [ 0.4024; 0.6314 ] 
#>   eta2 =~ y22      0.7554      0.0340   22.2233    0.0000 [ 0.6719; 0.8477 ] 
#>   eta2 =~ y23      0.7997      0.0439   18.1977    0.0000 [ 0.6860; 0.9133 ] 
#>   eta3 =~ y31      0.8223      0.0323   25.4734    0.0000 [ 0.7449; 0.9119 ] 
#>   eta3 =~ y32      0.6581      0.0471   13.9739    0.0000 [ 0.5297; 0.7732 ] 
#>   eta3 =~ y33      0.7474      0.0406   18.4191    0.0000 [ 0.6394; 0.8493 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0196   20.1874    0.0000 [ 0.3439; 0.4452 ] 
#>   eta1 <~ y12      0.3873      0.0174   22.2625    0.0000 [ 0.3453; 0.4353 ] 
#>   eta1 <~ y13      0.4542      0.0189   23.9801    0.0000 [ 0.4050; 0.5030 ] 
#>   eta2 <~ y21      0.3058      0.0266   11.4822    0.0000 [ 0.2363; 0.3740 ] 
#>   eta2 <~ y22      0.4473      0.0190   23.5140    0.0000 [ 0.3994; 0.4978 ] 
#>   eta2 <~ y23      0.4735      0.0211   22.4407    0.0000 [ 0.4177; 0.5269 ] 
#>   eta3 <~ y31      0.4400      0.0201   21.8657    0.0000 [ 0.3923; 0.4963 ] 
#>   eta3 <~ y32      0.3521      0.0205   17.1864    0.0000 [ 0.2969; 0.4028 ] 
#>   eta3 <~ y33      0.3999      0.0218   18.3050    0.0000 [ 0.3429; 0.4558 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0440   15.2707    0.0000 [ 0.5498; 0.7772 ] 
#>   eta3 ~ eta1       0.6634      0.0373   17.7831    0.0000 [ 0.5600; 0.7529 ] 
#>   eta3 ~ eta2       0.3052      0.0889    3.4328    0.0006 [ 0.0881; 0.5478 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0608    3.3689    0.0008 [ 0.0547; 0.3692 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04396221 15.270693 1.198923e-52
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08220919  5.577318 2.442553e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08889190  3.432834 5.973079e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54984993          0.7771977          0.5771499          0.7498976
#> 2         0.23190566          0.6570452          0.2829566          0.6059942
#> 3         0.08809563          0.5477943          0.1432965          0.4925935
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5748209          0.7593765          0.6016790          0.7547678
#> 2          0.2490102          0.6056718          0.3221979          0.5891633
#> 3          0.1551345          0.5290355          0.1562377          0.4664912