Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1988127708
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0453   14.8042    0.0000 [ 0.5747; 0.7439 ] 
#>   eta3 ~ eta1      0.4585      0.0986    4.6483    0.0000 [ 0.2395; 0.5914 ] 
#>   eta3 ~ eta2      0.3052      0.0974    3.1314    0.0017 [ 0.1574; 0.5210 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0410   16.1591    0.0000 [ 0.5955; 0.7515 ] 
#>   eta1 =~ y12      0.6493      0.0421   15.4254    0.0000 [ 0.5767; 0.7143 ] 
#>   eta1 =~ y13      0.7613      0.0318   23.9603    0.0000 [ 0.7108; 0.8192 ] 
#>   eta2 =~ y21      0.5165      0.0750    6.8860    0.0000 [ 0.3603; 0.6255 ] 
#>   eta2 =~ y22      0.7554      0.0404   18.7068    0.0000 [ 0.6602; 0.8272 ] 
#>   eta2 =~ y23      0.7997      0.0401   19.9272    0.0000 [ 0.6959; 0.8526 ] 
#>   eta3 =~ y31      0.8223      0.0317   25.9143    0.0000 [ 0.7660; 0.8786 ] 
#>   eta3 =~ y32      0.6581      0.0346   19.0130    0.0000 [ 0.6176; 0.7294 ] 
#>   eta3 =~ y33      0.7474      0.0398   18.7709    0.0000 [ 0.6760; 0.8094 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0228   17.3119    0.0000 [ 0.3596; 0.4478 ] 
#>   eta1 <~ y12      0.3873      0.0203   19.0362    0.0000 [ 0.3505; 0.4124 ] 
#>   eta1 <~ y13      0.4542      0.0201   22.6308    0.0000 [ 0.4149; 0.4833 ] 
#>   eta2 <~ y21      0.3058      0.0379    8.0773    0.0000 [ 0.2298; 0.3683 ] 
#>   eta2 <~ y22      0.4473      0.0286   15.6222    0.0000 [ 0.4042; 0.5010 ] 
#>   eta2 <~ y23      0.4735      0.0257   18.3930    0.0000 [ 0.4258; 0.5191 ] 
#>   eta3 <~ y31      0.4400      0.0184   23.9152    0.0000 [ 0.4031; 0.4699 ] 
#>   eta3 <~ y32      0.3521      0.0176   19.9902    0.0000 [ 0.3331; 0.3970 ] 
#>   eta3 <~ y33      0.3999      0.0177   22.5591    0.0000 [ 0.3684; 0.4251 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0453   14.8042    0.0000 [ 0.5747; 0.7439 ] 
#>   eta3 ~ eta1       0.6634      0.0454   14.6191    0.0000 [ 0.5459; 0.7354 ] 
#>   eta3 ~ eta2       0.3052      0.0974    3.1314    0.0017 [ 0.1574; 0.5210 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0678    3.0198    0.0025 [ 0.1077; 0.3631 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04103380 16.159116  9.795670e-59
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04209135 15.425448  1.103857e-53
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03177532 23.960288 7.218749e-127
#> 4 eta2 =~ y21  Common factor 0.5164548 0.07500115  6.885959  5.739958e-12
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04038043 18.706775  4.359743e-78
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04012916 19.927247  2.362009e-88
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03173068 25.914272 4.598721e-148
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03461161 19.012955  1.332328e-80
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03981816 18.770937  1.305627e-78
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5954941          0.7515138
#> 2          0.5766887          0.7142967
#> 3          0.7108050          0.8192062
#> 4          0.3603073          0.6255300
#> 5          0.6602094          0.8271523
#> 6          0.6959185          0.8526108
#> 7          0.7660403          0.8785532
#> 8          0.6176413          0.7294005
#> 9          0.6760180          0.8093528

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1988127708
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0453   14.8042    0.0000 [ 0.5502; 0.7847 ] 
#>   eta3 ~ eta1      0.4585      0.0986    4.6483    0.0000 [ 0.2107; 0.7208 ] 
#>   eta3 ~ eta2      0.3052      0.0974    3.1314    0.0017 [ 0.0423; 0.5463 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0410   16.1591    0.0000 [ 0.5505; 0.7627 ] 
#>   eta1 =~ y12      0.6493      0.0421   15.4254    0.0000 [ 0.5380; 0.7557 ] 
#>   eta1 =~ y13      0.7613      0.0318   23.9603    0.0000 [ 0.6779; 0.8422 ] 
#>   eta2 =~ y21      0.5165      0.0750    6.8860    0.0000 [ 0.3239; 0.7118 ] 
#>   eta2 =~ y22      0.7554      0.0404   18.7068    0.0000 [ 0.6577; 0.8665 ] 
#>   eta2 =~ y23      0.7997      0.0401   19.9272    0.0000 [ 0.7039; 0.9114 ] 
#>   eta3 =~ y31      0.8223      0.0317   25.9143    0.0000 [ 0.7450; 0.9091 ] 
#>   eta3 =~ y32      0.6581      0.0346   19.0130    0.0000 [ 0.5571; 0.7361 ] 
#>   eta3 =~ y33      0.7474      0.0398   18.7709    0.0000 [ 0.6494; 0.8553 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0228   17.3119    0.0000 [ 0.3357; 0.4539 ] 
#>   eta1 <~ y12      0.3873      0.0203   19.0362    0.0000 [ 0.3365; 0.4417 ] 
#>   eta1 <~ y13      0.4542      0.0201   22.6308    0.0000 [ 0.4049; 0.5087 ] 
#>   eta2 <~ y21      0.3058      0.0379    8.0773    0.0000 [ 0.2066; 0.4024 ] 
#>   eta2 <~ y22      0.4473      0.0286   15.6222    0.0000 [ 0.3724; 0.5205 ] 
#>   eta2 <~ y23      0.4735      0.0257   18.3930    0.0000 [ 0.4067; 0.5399 ] 
#>   eta3 <~ y31      0.4400      0.0184   23.9152    0.0000 [ 0.3954; 0.4905 ] 
#>   eta3 <~ y32      0.3521      0.0176   19.9902    0.0000 [ 0.3009; 0.3920 ] 
#>   eta3 <~ y33      0.3999      0.0177   22.5591    0.0000 [ 0.3573; 0.4490 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0453   14.8042    0.0000 [ 0.5502; 0.7847 ] 
#>   eta3 ~ eta1       0.6634      0.0454   14.6191    0.0000 [ 0.5452; 0.7799 ] 
#>   eta3 ~ eta2       0.3052      0.0974    3.1314    0.0017 [ 0.0423; 0.5463 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0678    3.0198    0.0025 [ 0.0213; 0.3722 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04534735 14.804247 1.375188e-49
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09864067  4.648253 3.347588e-06
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09744763  3.131437 1.739530e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5502006          0.7847116          0.5783608          0.7565514
#> 2          0.2107277          0.7208415          0.2719824          0.6595868
#> 3          0.0423195          0.5462636          0.1028333          0.4857498
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5680656          0.7709644          0.5746583          0.7439431
#> 2          0.2292835          0.6097910          0.2394955          0.5914224
#> 3          0.1444328          0.5269280          0.1573672          0.5209889