Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1459518157
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0458   14.6711    0.0000 [ 0.5857; 0.7474 ] 
#>   eta3 ~ eta1      0.4585      0.0793    5.7847    0.0000 [ 0.3390; 0.6128 ] 
#>   eta3 ~ eta2      0.3052      0.0835    3.6565    0.0003 [ 0.1976; 0.4376 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0365   18.1911    0.0000 [ 0.5850; 0.7223 ] 
#>   eta1 =~ y12      0.6493      0.0434   14.9636    0.0000 [ 0.5498; 0.7162 ] 
#>   eta1 =~ y13      0.7613      0.0307   24.7753    0.0000 [ 0.7022; 0.8042 ] 
#>   eta2 =~ y21      0.5165      0.0480   10.7638    0.0000 [ 0.4540; 0.6210 ] 
#>   eta2 =~ y22      0.7554      0.0392   19.2587    0.0000 [ 0.6682; 0.8010 ] 
#>   eta2 =~ y23      0.7997      0.0378   21.1349    0.0000 [ 0.7237; 0.8800 ] 
#>   eta3 =~ y31      0.8223      0.0336   24.5089    0.0000 [ 0.7530; 0.8734 ] 
#>   eta3 =~ y32      0.6581      0.0397   16.5883    0.0000 [ 0.5929; 0.7355 ] 
#>   eta3 =~ y33      0.7474      0.0405   18.4388    0.0000 [ 0.6723; 0.8084 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0199   19.8738    0.0000 [ 0.3556; 0.4288 ] 
#>   eta1 <~ y12      0.3873      0.0199   19.4182    0.0000 [ 0.3515; 0.4159 ] 
#>   eta1 <~ y13      0.4542      0.0203   22.3975    0.0000 [ 0.4172; 0.4892 ] 
#>   eta2 <~ y21      0.3058      0.0275   11.1084    0.0000 [ 0.2688; 0.3678 ] 
#>   eta2 <~ y22      0.4473      0.0198   22.5845    0.0000 [ 0.4097; 0.4767 ] 
#>   eta2 <~ y23      0.4735      0.0206   22.9386    0.0000 [ 0.4348; 0.5119 ] 
#>   eta3 <~ y31      0.4400      0.0204   21.5449    0.0000 [ 0.4017; 0.4764 ] 
#>   eta3 <~ y32      0.3521      0.0173   20.3705    0.0000 [ 0.3244; 0.3858 ] 
#>   eta3 <~ y33      0.3999      0.0209   19.1019    0.0000 [ 0.3521; 0.4272 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0458   14.6711    0.0000 [ 0.5857; 0.7474 ] 
#>   eta3 ~ eta1       0.6634      0.0412   16.1198    0.0000 [ 0.6080; 0.7591 ] 
#>   eta3 ~ eta2       0.3052      0.0835    3.6565    0.0003 [ 0.1976; 0.4376 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0573    3.5768    0.0003 [ 0.1162; 0.2981 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03645027 18.19108  6.072955e-74
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04339058 14.96357  1.270253e-50
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03073005 24.77528 1.655972e-135
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04798054 10.76384  5.100025e-27
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03922329 19.25865  1.194464e-82
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03783621 21.13488  3.801841e-99
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03355015 24.50890 1.187302e-132
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03967069 16.58829  8.469946e-62
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04053548 18.43876  6.419530e-76
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5849743          0.7222682
#> 2          0.5497662          0.7161855
#> 3          0.7022134          0.8042174
#> 4          0.4540387          0.6209800
#> 5          0.6682296          0.8010001
#> 6          0.7236797          0.8800393
#> 7          0.7530290          0.8733894
#> 8          0.5929440          0.7355110
#> 9          0.6722886          0.8083537

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1459518157
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0458   14.6711    0.0000 [ 0.5465; 0.7832 ] 
#>   eta3 ~ eta1      0.4585      0.0793    5.7847    0.0000 [ 0.2566; 0.6665 ] 
#>   eta3 ~ eta2      0.3052      0.0835    3.6565    0.0003 [ 0.0752; 0.5068 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0365   18.1911    0.0000 [ 0.5669; 0.7554 ] 
#>   eta1 =~ y12      0.6493      0.0434   14.9636    0.0000 [ 0.5371; 0.7614 ] 
#>   eta1 =~ y13      0.7613      0.0307   24.7753    0.0000 [ 0.6865; 0.8454 ] 
#>   eta2 =~ y21      0.5165      0.0480   10.7638    0.0000 [ 0.3837; 0.6318 ] 
#>   eta2 =~ y22      0.7554      0.0392   19.2587    0.0000 [ 0.6617; 0.8646 ] 
#>   eta2 =~ y23      0.7997      0.0378   21.1349    0.0000 [ 0.7094; 0.9051 ] 
#>   eta3 =~ y31      0.8223      0.0336   24.5089    0.0000 [ 0.7379; 0.9114 ] 
#>   eta3 =~ y32      0.6581      0.0397   16.5883    0.0000 [ 0.5432; 0.7484 ] 
#>   eta3 =~ y33      0.7474      0.0405   18.4388    0.0000 [ 0.6425; 0.8521 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0199   19.8738    0.0000 [ 0.3423; 0.4453 ] 
#>   eta1 <~ y12      0.3873      0.0199   19.4182    0.0000 [ 0.3354; 0.4385 ] 
#>   eta1 <~ y13      0.4542      0.0203   22.3975    0.0000 [ 0.4036; 0.5085 ] 
#>   eta2 <~ y21      0.3058      0.0275   11.1084    0.0000 [ 0.2279; 0.3702 ] 
#>   eta2 <~ y22      0.4473      0.0198   22.5845    0.0000 [ 0.3985; 0.5009 ] 
#>   eta2 <~ y23      0.4735      0.0206   22.9386    0.0000 [ 0.4222; 0.5289 ] 
#>   eta3 <~ y31      0.4400      0.0204   21.5449    0.0000 [ 0.3914; 0.4970 ] 
#>   eta3 <~ y32      0.3521      0.0173   20.3705    0.0000 [ 0.3035; 0.3929 ] 
#>   eta3 <~ y33      0.3999      0.0209   19.1019    0.0000 [ 0.3485; 0.4568 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0458   14.6711    0.0000 [ 0.5465; 0.7832 ] 
#>   eta3 ~ eta1       0.6634      0.0412   16.1198    0.0000 [ 0.5488; 0.7616 ] 
#>   eta3 ~ eta2       0.3052      0.0835    3.6565    0.0003 [ 0.0752; 0.5068 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0573    3.5768    0.0003 [ 0.0455; 0.3417 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04575876 14.67115 9.866495e-49
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07926225  5.78468 7.265055e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08345397  3.65652 2.556621e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54653148          0.7831699          0.5749471          0.7547543
#> 2         0.25664490          0.6665445          0.3058658          0.6173236
#> 3         0.07518604          0.5067628          0.1270100          0.4549389
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5816999          0.7676893          0.5857038          0.7473552
#> 2          0.3375974          0.6498008          0.3390089          0.6128317
#> 3          0.1220395          0.4414514          0.1976195          0.4376319