Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -194113587
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0429   15.6441    0.0000 [ 0.5963; 0.7508 ] 
#>   eta3 ~ eta1      0.4585      0.0896    5.1147    0.0000 [ 0.3041; 0.6198 ] 
#>   eta3 ~ eta2      0.3052      0.1035    2.9482    0.0032 [ 0.1179; 0.5161 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0413   16.0400    0.0000 [ 0.5951; 0.7298 ] 
#>   eta1 =~ y12      0.6493      0.0408   15.9292    0.0000 [ 0.5621; 0.7042 ] 
#>   eta1 =~ y13      0.7613      0.0346   22.0126    0.0000 [ 0.7105; 0.8301 ] 
#>   eta2 =~ y21      0.5165      0.0487   10.5976    0.0000 [ 0.4464; 0.6054 ] 
#>   eta2 =~ y22      0.7554      0.0367   20.5901    0.0000 [ 0.6552; 0.7938 ] 
#>   eta2 =~ y23      0.7997      0.0305   26.1786    0.0000 [ 0.7369; 0.8368 ] 
#>   eta3 =~ y31      0.8223      0.0291   28.2996    0.0000 [ 0.7671; 0.8827 ] 
#>   eta3 =~ y32      0.6581      0.0341   19.3028    0.0000 [ 0.5922; 0.7132 ] 
#>   eta3 =~ y33      0.7474      0.0341   21.9470    0.0000 [ 0.6912; 0.8005 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0226   17.4838    0.0000 [ 0.3629; 0.4349 ] 
#>   eta1 <~ y12      0.3873      0.0180   21.4853    0.0000 [ 0.3467; 0.4071 ] 
#>   eta1 <~ y13      0.4542      0.0226   20.1381    0.0000 [ 0.4286; 0.5068 ] 
#>   eta2 <~ y21      0.3058      0.0283   10.8199    0.0000 [ 0.2690; 0.3695 ] 
#>   eta2 <~ y22      0.4473      0.0186   24.0910    0.0000 [ 0.4132; 0.4744 ] 
#>   eta2 <~ y23      0.4735      0.0181   26.1260    0.0000 [ 0.4450; 0.5071 ] 
#>   eta3 <~ y31      0.4400      0.0157   27.9477    0.0000 [ 0.4157; 0.4680 ] 
#>   eta3 <~ y32      0.3521      0.0166   21.2313    0.0000 [ 0.3202; 0.3778 ] 
#>   eta3 <~ y33      0.3999      0.0146   27.4203    0.0000 [ 0.3763; 0.4284 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0429   15.6441    0.0000 [ 0.5963; 0.7508 ] 
#>   eta3 ~ eta1       0.6634      0.0318   20.8560    0.0000 [ 0.6104; 0.7234 ] 
#>   eta3 ~ eta2       0.3052      0.1035    2.9482    0.0032 [ 0.1179; 0.5161 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0742    2.7622    0.0057 [ 0.0821; 0.3359 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04133853 16.04000  6.715893e-58
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04076027 15.92919  3.974896e-57
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03458681 22.01261 2.180774e-107
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04873305 10.59763  3.056285e-26
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03668686 20.59014  3.363763e-94
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03054651 26.17857 4.662381e-151
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02905619 28.29956 3.499537e-176
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03409193 19.30277  5.090226e-83
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03405579 21.94705 9.241210e-107
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5951273          0.7298330
#> 2          0.5621095          0.7041771
#> 3          0.7104736          0.8301245
#> 4          0.4464391          0.6054369
#> 5          0.6552337          0.7937907
#> 6          0.7369353          0.8368332
#> 7          0.7671228          0.8827303
#> 8          0.5921652          0.7131892
#> 9          0.6912085          0.8005090

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -194113587
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0429   15.6441    0.0000 [ 0.5508; 0.7727 ] 
#>   eta3 ~ eta1      0.4585      0.0896    5.1147    0.0000 [ 0.2359; 0.6995 ] 
#>   eta3 ~ eta2      0.3052      0.1035    2.9482    0.0032 [ 0.0307; 0.5659 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0413   16.0400    0.0000 [ 0.5578; 0.7715 ] 
#>   eta1 =~ y12      0.6493      0.0408   15.9292    0.0000 [ 0.5592; 0.7700 ] 
#>   eta1 =~ y13      0.7613      0.0346   22.0126    0.0000 [ 0.6669; 0.8458 ] 
#>   eta2 =~ y21      0.5165      0.0487   10.5976    0.0000 [ 0.3812; 0.6332 ] 
#>   eta2 =~ y22      0.7554      0.0367   20.5901    0.0000 [ 0.6735; 0.8632 ] 
#>   eta2 =~ y23      0.7997      0.0305   26.1786    0.0000 [ 0.7264; 0.8844 ] 
#>   eta3 =~ y31      0.8223      0.0291   28.2996    0.0000 [ 0.7495; 0.8998 ] 
#>   eta3 =~ y32      0.6581      0.0341   19.3028    0.0000 [ 0.5712; 0.7475 ] 
#>   eta3 =~ y33      0.7474      0.0341   21.9470    0.0000 [ 0.6540; 0.8301 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0226   17.4838    0.0000 [ 0.3350; 0.4520 ] 
#>   eta1 <~ y12      0.3873      0.0180   21.4853    0.0000 [ 0.3472; 0.4404 ] 
#>   eta1 <~ y13      0.4542      0.0226   20.1381    0.0000 [ 0.3892; 0.5058 ] 
#>   eta2 <~ y21      0.3058      0.0283   10.8199    0.0000 [ 0.2250; 0.3712 ] 
#>   eta2 <~ y22      0.4473      0.0186   24.0910    0.0000 [ 0.4039; 0.4999 ] 
#>   eta2 <~ y23      0.4735      0.0181   26.1260    0.0000 [ 0.4265; 0.5203 ] 
#>   eta3 <~ y31      0.4400      0.0157   27.9477    0.0000 [ 0.4011; 0.4825 ] 
#>   eta3 <~ y32      0.3521      0.0166   21.2313    0.0000 [ 0.3104; 0.3962 ] 
#>   eta3 <~ y33      0.3999      0.0146   27.4203    0.0000 [ 0.3600; 0.4354 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0429   15.6441    0.0000 [ 0.5508; 0.7727 ] 
#>   eta3 ~ eta1       0.6634      0.0318   20.8560    0.0000 [ 0.5827; 0.7471 ] 
#>   eta3 ~ eta2       0.3052      0.1035    2.9482    0.0032 [ 0.0307; 0.5659 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0742    2.7622    0.0057 [ 0.0054; 0.3890 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04291288 15.644100 3.644886e-55
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08964482  5.114704 3.142328e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.10350333  2.948225 3.196042e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55078689          0.7727081          0.5774353          0.7460597
#> 2         0.23590929          0.6995017          0.2915777          0.6438333
#> 3         0.03067854          0.5659393          0.0949529          0.5016650
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5960784          0.7603730          0.5963112          0.7507592
#> 2          0.2430127          0.6295369          0.3040634          0.6198032
#> 3          0.1112521          0.5724199          0.1178869          0.5161360