Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1494725579
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0475   14.1322    0.0000 [ 0.5835; 0.7553 ] 
#>   eta3 ~ eta1      0.4585      0.0905    5.0668    0.0000 [ 0.2800; 0.6226 ] 
#>   eta3 ~ eta2      0.3052      0.1039    2.9361    0.0033 [ 0.0975; 0.5198 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0346   19.1456    0.0000 [ 0.5966; 0.7065 ] 
#>   eta1 =~ y12      0.6493      0.0379   17.1297    0.0000 [ 0.5604; 0.7017 ] 
#>   eta1 =~ y13      0.7613      0.0296   25.7534    0.0000 [ 0.7072; 0.8215 ] 
#>   eta2 =~ y21      0.5165      0.0531    9.7288    0.0000 [ 0.4199; 0.5996 ] 
#>   eta2 =~ y22      0.7554      0.0348   21.7271    0.0000 [ 0.6952; 0.8075 ] 
#>   eta2 =~ y23      0.7997      0.0417   19.1847    0.0000 [ 0.7341; 0.8697 ] 
#>   eta3 =~ y31      0.8223      0.0391   21.0488    0.0000 [ 0.7427; 0.8881 ] 
#>   eta3 =~ y32      0.6581      0.0427   15.4105    0.0000 [ 0.5698; 0.7416 ] 
#>   eta3 =~ y33      0.7474      0.0340   21.9624    0.0000 [ 0.6692; 0.8007 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0167   23.6357    0.0000 [ 0.3555; 0.4228 ] 
#>   eta1 <~ y12      0.3873      0.0191   20.2455    0.0000 [ 0.3508; 0.4189 ] 
#>   eta1 <~ y13      0.4542      0.0198   22.8813    0.0000 [ 0.4285; 0.4932 ] 
#>   eta2 <~ y21      0.3058      0.0270   11.3109    0.0000 [ 0.2611; 0.3501 ] 
#>   eta2 <~ y22      0.4473      0.0234   19.1263    0.0000 [ 0.4054; 0.4882 ] 
#>   eta2 <~ y23      0.4735      0.0214   22.1752    0.0000 [ 0.4454; 0.5065 ] 
#>   eta3 <~ y31      0.4400      0.0197   22.2890    0.0000 [ 0.3998; 0.4763 ] 
#>   eta3 <~ y32      0.3521      0.0201   17.5040    0.0000 [ 0.3130; 0.3866 ] 
#>   eta3 <~ y33      0.3999      0.0190   21.0916    0.0000 [ 0.3638; 0.4349 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0475   14.1322    0.0000 [ 0.5835; 0.7553 ] 
#>   eta3 ~ eta1       0.6634      0.0373   17.7880    0.0000 [ 0.5772; 0.7089 ] 
#>   eta3 ~ eta2       0.3052      0.1039    2.9361    0.0033 [ 0.0975; 0.5198 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0716    2.8610    0.0042 [ 0.0665; 0.3499 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03463302 19.145600  1.053178e-81
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03790361 17.129714  8.909114e-66
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02956296 25.753369 2.955185e-146
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05308516  9.728798  2.272615e-22
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03476700 21.727147 1.136483e-104
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04168239 19.184689  4.969042e-82
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03906535 21.048766  2.347107e-98
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04270266 15.410488  1.391560e-53
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03403202 21.962380 6.595291e-107
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5966400          0.7065017
#> 2          0.5603744          0.7016915
#> 3          0.7072060          0.8215395
#> 4          0.4198979          0.5995618
#> 5          0.6952231          0.8075134
#> 6          0.7340553          0.8697089
#> 7          0.7427474          0.8881366
#> 8          0.5698423          0.7416195
#> 9          0.6691509          0.8006808

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1494725579
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0475   14.1322    0.0000 [ 0.5418; 0.7874 ] 
#>   eta3 ~ eta1      0.4585      0.0905    5.0668    0.0000 [ 0.2427; 0.7107 ] 
#>   eta3 ~ eta2      0.3052      0.1039    2.9361    0.0033 [ 0.0170; 0.5545 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0346   19.1456    0.0000 [ 0.5779; 0.7570 ] 
#>   eta1 =~ y12      0.6493      0.0379   17.1297    0.0000 [ 0.5626; 0.7587 ] 
#>   eta1 =~ y13      0.7613      0.0296   25.7534    0.0000 [ 0.6727; 0.8256 ] 
#>   eta2 =~ y21      0.5165      0.0531    9.7288    0.0000 [ 0.3839; 0.6584 ] 
#>   eta2 =~ y22      0.7554      0.0348   21.7271    0.0000 [ 0.6685; 0.8483 ] 
#>   eta2 =~ y23      0.7997      0.0417   19.1847    0.0000 [ 0.6931; 0.9087 ] 
#>   eta3 =~ y31      0.8223      0.0391   21.0488    0.0000 [ 0.7205; 0.9225 ] 
#>   eta3 =~ y32      0.6581      0.0427   15.4105    0.0000 [ 0.5507; 0.7716 ] 
#>   eta3 =~ y33      0.7474      0.0340   21.9624    0.0000 [ 0.6610; 0.8370 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0167   23.6357    0.0000 [ 0.3544; 0.4409 ] 
#>   eta1 <~ y12      0.3873      0.0191   20.2455    0.0000 [ 0.3441; 0.4431 ] 
#>   eta1 <~ y13      0.4542      0.0198   22.8813    0.0000 [ 0.3946; 0.4973 ] 
#>   eta2 <~ y21      0.3058      0.0270   11.3109    0.0000 [ 0.2374; 0.3773 ] 
#>   eta2 <~ y22      0.4473      0.0234   19.1263    0.0000 [ 0.3861; 0.5070 ] 
#>   eta2 <~ y23      0.4735      0.0214   22.1752    0.0000 [ 0.4166; 0.5271 ] 
#>   eta3 <~ y31      0.4400      0.0197   22.2890    0.0000 [ 0.3876; 0.4897 ] 
#>   eta3 <~ y32      0.3521      0.0201   17.5040    0.0000 [ 0.3011; 0.4052 ] 
#>   eta3 <~ y33      0.3999      0.0190   21.0916    0.0000 [ 0.3509; 0.4489 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0475   14.1322    0.0000 [ 0.5418; 0.7874 ] 
#>   eta3 ~ eta1       0.6634      0.0373   17.7880    0.0000 [ 0.5701; 0.7630 ] 
#>   eta3 ~ eta2       0.3052      0.1039    2.9361    0.0033 [ 0.0170; 0.5545 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0716    2.8610    0.0042 [ 0.0047; 0.3750 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04750387 14.132183 2.405702e-45
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09049281  5.066776 4.046107e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.10392997  2.936123 3.323430e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54177682          0.7874400         0.57127617          0.7579407
#> 2         0.24272115          0.7106989         0.29891613          0.6545039
#> 3         0.01701426          0.5544814         0.08155356          0.4899421
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.57013023          0.7580044         0.58351479          0.7553025
#> 2         0.27534605          0.6272676         0.27995043          0.6225740
#> 3         0.08832088          0.5286000         0.09751209          0.5197821