Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -932815609
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0508   13.2155    0.0000 [ 0.5533; 0.7458 ] 
#>   eta3 ~ eta1      0.4585      0.0738    6.2115    0.0000 [ 0.3114; 0.5869 ] 
#>   eta3 ~ eta2      0.3052      0.0758    4.0279    0.0001 [ 0.2135; 0.4569 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0363   18.2808    0.0000 [ 0.6087; 0.7300 ] 
#>   eta1 =~ y12      0.6493      0.0363   17.8782    0.0000 [ 0.6022; 0.7220 ] 
#>   eta1 =~ y13      0.7613      0.0262   29.0651    0.0000 [ 0.7174; 0.8178 ] 
#>   eta2 =~ y21      0.5165      0.0507   10.1912    0.0000 [ 0.4261; 0.6001 ] 
#>   eta2 =~ y22      0.7554      0.0363   20.8325    0.0000 [ 0.6894; 0.8283 ] 
#>   eta2 =~ y23      0.7997      0.0363   22.0444    0.0000 [ 0.7440; 0.8716 ] 
#>   eta3 =~ y31      0.8223      0.0320   25.7229    0.0000 [ 0.7777; 0.8973 ] 
#>   eta3 =~ y32      0.6581      0.0411   16.0063    0.0000 [ 0.5662; 0.7198 ] 
#>   eta3 =~ y33      0.7474      0.0390   19.1513    0.0000 [ 0.6570; 0.8048 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0199   19.8846    0.0000 [ 0.3664; 0.4340 ] 
#>   eta1 <~ y12      0.3873      0.0173   22.3601    0.0000 [ 0.3575; 0.4190 ] 
#>   eta1 <~ y13      0.4542      0.0181   25.1514    0.0000 [ 0.4242; 0.4993 ] 
#>   eta2 <~ y21      0.3058      0.0267   11.4607    0.0000 [ 0.2643; 0.3438 ] 
#>   eta2 <~ y22      0.4473      0.0199   22.5177    0.0000 [ 0.4136; 0.4819 ] 
#>   eta2 <~ y23      0.4735      0.0217   21.8232    0.0000 [ 0.4469; 0.5207 ] 
#>   eta3 <~ y31      0.4400      0.0184   23.8530    0.0000 [ 0.4183; 0.4761 ] 
#>   eta3 <~ y32      0.3521      0.0201   17.5072    0.0000 [ 0.3145; 0.3874 ] 
#>   eta3 <~ y33      0.3999      0.0190   21.0384    0.0000 [ 0.3599; 0.4273 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0508   13.2155    0.0000 [ 0.5533; 0.7458 ] 
#>   eta3 ~ eta1       0.6634      0.0299   22.1660    0.0000 [ 0.6095; 0.7132 ] 
#>   eta3 ~ eta2       0.3052      0.0758    4.0279    0.0001 [ 0.2135; 0.4569 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0556    3.6833    0.0002 [ 0.1245; 0.3114 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03627147 18.28076  1.177806e-74
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03631667 17.87824  1.742623e-71
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02619446 29.06515 9.904109e-186
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05067666 10.19118  2.171175e-24
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03626000 20.83253  2.195346e-96
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03627513 22.04441 1.080904e-107
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03196672 25.72292 6.477813e-146
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04111321 16.00627  1.155389e-57
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03902727 19.15133  9.434820e-82
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6087487          0.7299808
#> 2          0.6022321          0.7219734
#> 3          0.7173822          0.8178099
#> 4          0.4261434          0.6000764
#> 5          0.6894211          0.8282518
#> 6          0.7439867          0.8715686
#> 7          0.7777338          0.8973314
#> 8          0.5661622          0.7198165
#> 9          0.6570242          0.8047518

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -932815609
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0508   13.2155    0.0000 [ 0.5451; 0.8078 ] 
#>   eta3 ~ eta1      0.4585      0.0738    6.2115    0.0000 [ 0.2889; 0.6706 ] 
#>   eta3 ~ eta2      0.3052      0.0758    4.0279    0.0001 [ 0.0747; 0.4664 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0363   18.2808    0.0000 [ 0.5653; 0.7528 ] 
#>   eta1 =~ y12      0.6493      0.0363   17.8782    0.0000 [ 0.5490; 0.7368 ] 
#>   eta1 =~ y13      0.7613      0.0262   29.0651    0.0000 [ 0.6922; 0.8277 ] 
#>   eta2 =~ y21      0.5165      0.0507   10.1912    0.0000 [ 0.3867; 0.6488 ] 
#>   eta2 =~ y22      0.7554      0.0363   20.8325    0.0000 [ 0.6659; 0.8534 ] 
#>   eta2 =~ y23      0.7997      0.0363   22.0444    0.0000 [ 0.6921; 0.8797 ] 
#>   eta3 =~ y31      0.8223      0.0320   25.7229    0.0000 [ 0.7309; 0.8962 ] 
#>   eta3 =~ y32      0.6581      0.0411   16.0063    0.0000 [ 0.5512; 0.7639 ] 
#>   eta3 =~ y33      0.7474      0.0390   19.1513    0.0000 [ 0.6557; 0.8576 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0199   19.8846    0.0000 [ 0.3452; 0.4481 ] 
#>   eta1 <~ y12      0.3873      0.0173   22.3601    0.0000 [ 0.3423; 0.4319 ] 
#>   eta1 <~ y13      0.4542      0.0181   25.1514    0.0000 [ 0.4105; 0.5039 ] 
#>   eta2 <~ y21      0.3058      0.0267   11.4607    0.0000 [ 0.2400; 0.3780 ] 
#>   eta2 <~ y22      0.4473      0.0199   22.5177    0.0000 [ 0.4016; 0.5043 ] 
#>   eta2 <~ y23      0.4735      0.0217   21.8232    0.0000 [ 0.4126; 0.5248 ] 
#>   eta3 <~ y31      0.4400      0.0184   23.8530    0.0000 [ 0.3878; 0.4832 ] 
#>   eta3 <~ y32      0.3521      0.0201   17.5072    0.0000 [ 0.3001; 0.4042 ] 
#>   eta3 <~ y33      0.3999      0.0190   21.0384    0.0000 [ 0.3560; 0.4543 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0508   13.2155    0.0000 [ 0.5451; 0.8078 ] 
#>   eta3 ~ eta1       0.6634      0.0299   22.1660    0.0000 [ 0.5851; 0.7398 ] 
#>   eta3 ~ eta2       0.3052      0.0758    4.0279    0.0001 [ 0.0747; 0.4664 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0556    3.6833    0.0002 [ 0.0389; 0.3265 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05079894 13.215501 7.140936e-40
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07381550  6.211524 5.247307e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07575930  4.027903 5.627652e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54507613          0.8077796          0.5766217          0.7762340
#> 2         0.28885908          0.6705912          0.3346976          0.6247526
#> 3         0.07466224          0.4664466          0.1217079          0.4194009
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5527275          0.7560233          0.5533127          0.7458406
#> 2          0.3059932          0.5909815          0.3113616          0.5868728
#> 3          0.1905451          0.4840629          0.2134640          0.4569432