Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -980269089
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0390   17.2018    0.0000 [ 0.5972; 0.7269 ] 
#>   eta3 ~ eta1      0.4585      0.0635    7.2181    0.0000 [ 0.3326; 0.5439 ] 
#>   eta3 ~ eta2      0.3052      0.0617    4.9435    0.0000 [ 0.2048; 0.4473 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0471   14.0732    0.0000 [ 0.5799; 0.7534 ] 
#>   eta1 =~ y12      0.6493      0.0429   15.1406    0.0000 [ 0.5719; 0.7106 ] 
#>   eta1 =~ y13      0.7613      0.0343   22.2250    0.0000 [ 0.7030; 0.8143 ] 
#>   eta2 =~ y21      0.5165      0.0605    8.5386    0.0000 [ 0.4032; 0.6175 ] 
#>   eta2 =~ y22      0.7554      0.0331   22.8101    0.0000 [ 0.6923; 0.8064 ] 
#>   eta2 =~ y23      0.7997      0.0364   21.9617    0.0000 [ 0.7195; 0.8585 ] 
#>   eta3 =~ y31      0.8223      0.0256   32.1402    0.0000 [ 0.7725; 0.8665 ] 
#>   eta3 =~ y32      0.6581      0.0427   15.4013    0.0000 [ 0.5795; 0.7261 ] 
#>   eta3 =~ y33      0.7474      0.0425   17.5780    0.0000 [ 0.6580; 0.8224 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0260   15.1927    0.0000 [ 0.3436; 0.4448 ] 
#>   eta1 <~ y12      0.3873      0.0218   17.7805    0.0000 [ 0.3559; 0.4195 ] 
#>   eta1 <~ y13      0.4542      0.0224   20.2480    0.0000 [ 0.4131; 0.4920 ] 
#>   eta2 <~ y21      0.3058      0.0307    9.9504    0.0000 [ 0.2456; 0.3655 ] 
#>   eta2 <~ y22      0.4473      0.0222   20.1364    0.0000 [ 0.4131; 0.4824 ] 
#>   eta2 <~ y23      0.4735      0.0229   20.6593    0.0000 [ 0.4361; 0.5164 ] 
#>   eta3 <~ y31      0.4400      0.0214   20.5249    0.0000 [ 0.4022; 0.4787 ] 
#>   eta3 <~ y32      0.3521      0.0177   19.9240    0.0000 [ 0.3231; 0.3898 ] 
#>   eta3 <~ y33      0.3999      0.0177   22.5407    0.0000 [ 0.3731; 0.4386 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0390   17.2018    0.0000 [ 0.5972; 0.7269 ] 
#>   eta3 ~ eta1       0.6634      0.0359   18.5009    0.0000 [ 0.5966; 0.7258 ] 
#>   eta3 ~ eta2       0.3052      0.0617    4.9435    0.0000 [ 0.2048; 0.4473 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0426    4.8067    0.0000 [ 0.1506; 0.3072 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04711580 14.073195  5.550389e-45
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04288326 15.140593  8.742134e-52
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03425634 22.224963 1.970458e-109
#> 4 eta2 =~ y21  Common factor 0.5164548 0.06048467  8.538607  1.358496e-17
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03311641 22.810071 3.642675e-115
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03641169 21.961733 6.689897e-107
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02558408 32.140192 1.210943e-226
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04272818 15.401285  1.604484e-53
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04252046 17.577986  3.632489e-69
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5799459          0.7533847
#> 2          0.5718795          0.7106188
#> 3          0.7029757          0.8142560
#> 4          0.4032009          0.6175021
#> 5          0.6923478          0.8063960
#> 6          0.7194679          0.8584537
#> 7          0.7724883          0.8665264
#> 8          0.5795131          0.7260703
#> 9          0.6580168          0.8224305

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -980269089
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0390   17.2018    0.0000 [ 0.5725; 0.7743 ] 
#>   eta3 ~ eta1      0.4585      0.0635    7.2181    0.0000 [ 0.3117; 0.6402 ] 
#>   eta3 ~ eta2      0.3052      0.0617    4.9435    0.0000 [ 0.1226; 0.4418 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0471   14.0732    0.0000 [ 0.5377; 0.7814 ] 
#>   eta1 =~ y12      0.6493      0.0429   15.1406    0.0000 [ 0.5370; 0.7587 ] 
#>   eta1 =~ y13      0.7613      0.0343   22.2250    0.0000 [ 0.6751; 0.8522 ] 
#>   eta2 =~ y21      0.5165      0.0605    8.5386    0.0000 [ 0.3557; 0.6685 ] 
#>   eta2 =~ y22      0.7554      0.0331   22.8101    0.0000 [ 0.6716; 0.8428 ] 
#>   eta2 =~ y23      0.7997      0.0364   21.9617    0.0000 [ 0.7106; 0.8989 ] 
#>   eta3 =~ y31      0.8223      0.0256   32.1402    0.0000 [ 0.7599; 0.8922 ] 
#>   eta3 =~ y32      0.6581      0.0427   15.4013    0.0000 [ 0.5430; 0.7639 ] 
#>   eta3 =~ y33      0.7474      0.0425   17.5780    0.0000 [ 0.6272; 0.8471 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0260   15.1927    0.0000 [ 0.3272; 0.4619 ] 
#>   eta1 <~ y12      0.3873      0.0218   17.7805    0.0000 [ 0.3313; 0.4439 ] 
#>   eta1 <~ y13      0.4542      0.0224   20.2480    0.0000 [ 0.3985; 0.5145 ] 
#>   eta2 <~ y21      0.3058      0.0307    9.9504    0.0000 [ 0.2239; 0.3828 ] 
#>   eta2 <~ y22      0.4473      0.0222   20.1364    0.0000 [ 0.3902; 0.5051 ] 
#>   eta2 <~ y23      0.4735      0.0229   20.6593    0.0000 [ 0.4166; 0.5351 ] 
#>   eta3 <~ y31      0.4400      0.0214   20.5249    0.0000 [ 0.3896; 0.5005 ] 
#>   eta3 <~ y32      0.3521      0.0177   19.9240    0.0000 [ 0.3070; 0.3984 ] 
#>   eta3 <~ y33      0.3999      0.0177   22.5407    0.0000 [ 0.3519; 0.4437 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0390   17.2018    0.0000 [ 0.5725; 0.7743 ] 
#>   eta3 ~ eta1       0.6634      0.0359   18.5009    0.0000 [ 0.5737; 0.7591 ] 
#>   eta3 ~ eta2       0.3052      0.0617    4.9435    0.0000 [ 0.1226; 0.4418 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0426    4.8067    0.0000 [ 0.0802; 0.3006 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03902685 17.201834 2.572582e-66
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.06352190  7.218090 5.272297e-13
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.06172816  4.943467 7.674546e-07
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5724705          0.7742953          0.5967057          0.7500601
#> 2          0.3117195          0.6402189          0.3511659          0.6007726
#> 3          0.1226257          0.4418489          0.1609582          0.4035165
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5939410          0.7446421          0.5972028          0.7268699
#> 2          0.3022713          0.5530247          0.3325760          0.5439046
#> 3          0.2010387          0.4650787          0.2048047          0.4473068