Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1962608932
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0433   15.5112    0.0000 [ 0.6140; 0.7493 ] 
#>   eta3 ~ eta1      0.4585      0.0796    5.7598    0.0000 [ 0.3238; 0.5987 ] 
#>   eta3 ~ eta2      0.3052      0.0820    3.7219    0.0002 [ 0.1630; 0.4579 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0431   15.3860    0.0000 [ 0.5838; 0.7296 ] 
#>   eta1 =~ y12      0.6493      0.0366   17.7392    0.0000 [ 0.5792; 0.7183 ] 
#>   eta1 =~ y13      0.7613      0.0306   24.8408    0.0000 [ 0.7217; 0.8273 ] 
#>   eta2 =~ y21      0.5165      0.0465   11.1174    0.0000 [ 0.4164; 0.5943 ] 
#>   eta2 =~ y22      0.7554      0.0438   17.2277    0.0000 [ 0.6285; 0.8081 ] 
#>   eta2 =~ y23      0.7997      0.0551   14.5246    0.0000 [ 0.6852; 0.9083 ] 
#>   eta3 =~ y31      0.8223      0.0322   25.5696    0.0000 [ 0.7729; 0.8831 ] 
#>   eta3 =~ y32      0.6581      0.0504   13.0504    0.0000 [ 0.5258; 0.7264 ] 
#>   eta3 =~ y33      0.7474      0.0360   20.7642    0.0000 [ 0.6758; 0.8115 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0208   19.0322    0.0000 [ 0.3550; 0.4281 ] 
#>   eta1 <~ y12      0.3873      0.0197   19.6245    0.0000 [ 0.3535; 0.4201 ] 
#>   eta1 <~ y13      0.4542      0.0218   20.8592    0.0000 [ 0.4250; 0.5007 ] 
#>   eta2 <~ y21      0.3058      0.0304   10.0677    0.0000 [ 0.2710; 0.3716 ] 
#>   eta2 <~ y22      0.4473      0.0252   17.7476    0.0000 [ 0.3937; 0.4841 ] 
#>   eta2 <~ y23      0.4735      0.0222   21.2880    0.0000 [ 0.4375; 0.5172 ] 
#>   eta3 <~ y31      0.4400      0.0173   25.4251    0.0000 [ 0.4150; 0.4759 ] 
#>   eta3 <~ y32      0.3521      0.0228   15.4580    0.0000 [ 0.2944; 0.3813 ] 
#>   eta3 <~ y33      0.3999      0.0211   18.9211    0.0000 [ 0.3691; 0.4445 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0433   15.5112    0.0000 [ 0.6140; 0.7493 ] 
#>   eta3 ~ eta1       0.6634      0.0383   17.3059    0.0000 [ 0.5890; 0.7250 ] 
#>   eta3 ~ eta2       0.3052      0.0820    3.7219    0.0002 [ 0.1630; 0.4579 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0551    3.7205    0.0002 [ 0.1109; 0.3218 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04309574 15.38597  2.033022e-53
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03660129 17.73921  2.088617e-70
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03064896 24.84084 3.248181e-136
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04645447 11.11744  1.031894e-28
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04384717 17.22774  1.644562e-66
#> 6 eta2 =~ y23  Common factor 0.7996637 0.05505597 14.52456  8.468602e-48
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03215844 25.56957 3.327303e-144
#> 8 eta3 =~ y32  Common factor 0.6580689 0.05042530 13.05037  6.323725e-39
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03599578 20.76422  9.119968e-96
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5837723          0.7296489
#> 2          0.5792356          0.7182844
#> 3          0.7217200          0.8272607
#> 4          0.4163873          0.5942545
#> 5          0.6284601          0.8080511
#> 6          0.6852046          0.9082698
#> 7          0.7728713          0.8831238
#> 8          0.5258082          0.7263733
#> 9          0.6757670          0.8114838

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1962608932
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0433   15.5112    0.0000 [ 0.5524; 0.7762 ] 
#>   eta3 ~ eta1      0.4585      0.0796    5.7598    0.0000 [ 0.2505; 0.6622 ] 
#>   eta3 ~ eta2      0.3052      0.0820    3.7219    0.0002 [ 0.0995; 0.5235 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0431   15.3860    0.0000 [ 0.5603; 0.7832 ] 
#>   eta1 =~ y12      0.6493      0.0366   17.7392    0.0000 [ 0.5653; 0.7546 ] 
#>   eta1 =~ y13      0.7613      0.0306   24.8408    0.0000 [ 0.6768; 0.8353 ] 
#>   eta2 =~ y21      0.5165      0.0465   11.1174    0.0000 [ 0.3867; 0.6269 ] 
#>   eta2 =~ y22      0.7554      0.0438   17.2277    0.0000 [ 0.6486; 0.8753 ] 
#>   eta2 =~ y23      0.7997      0.0551   14.5246    0.0000 [ 0.6621; 0.9468 ] 
#>   eta3 =~ y31      0.8223      0.0322   25.5696    0.0000 [ 0.7370; 0.9033 ] 
#>   eta3 =~ y32      0.6581      0.0504   13.0504    0.0000 [ 0.5414; 0.8021 ] 
#>   eta3 =~ y33      0.7474      0.0360   20.7642    0.0000 [ 0.6538; 0.8400 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0208   19.0322    0.0000 [ 0.3436; 0.4511 ] 
#>   eta1 <~ y12      0.3873      0.0197   19.6245    0.0000 [ 0.3392; 0.4413 ] 
#>   eta1 <~ y13      0.4542      0.0218   20.8592    0.0000 [ 0.3903; 0.5029 ] 
#>   eta2 <~ y21      0.3058      0.0304   10.0677    0.0000 [ 0.2208; 0.3779 ] 
#>   eta2 <~ y22      0.4473      0.0252   17.7476    0.0000 [ 0.3852; 0.5155 ] 
#>   eta2 <~ y23      0.4735      0.0222   21.2880    0.0000 [ 0.4185; 0.5336 ] 
#>   eta3 <~ y31      0.4400      0.0173   25.4251    0.0000 [ 0.3912; 0.4807 ] 
#>   eta3 <~ y32      0.3521      0.0228   15.4580    0.0000 [ 0.2986; 0.4164 ] 
#>   eta3 <~ y33      0.3999      0.0211   18.9211    0.0000 [ 0.3423; 0.4516 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0433   15.5112    0.0000 [ 0.5524; 0.7762 ] 
#>   eta3 ~ eta1       0.6634      0.0383   17.3059    0.0000 [ 0.5650; 0.7632 ] 
#>   eta3 ~ eta2       0.3052      0.0820    3.7219    0.0002 [ 0.0995; 0.5235 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0551    3.7205    0.0002 [ 0.0653; 0.3501 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04328064 15.511172 2.915068e-54
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07960479  5.759788 8.421941e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08198716  3.721938 1.976995e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55242338          0.7762464          0.5793002          0.7493696
#> 2         0.25054039          0.6622114          0.2999740          0.6127778
#> 3         0.09947117          0.5234625          0.1503842          0.4725494
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5786159          0.7635039          0.6140295          0.7493176
#> 2          0.2620480          0.6031963          0.3237721          0.5987392
#> 3          0.1457267          0.5132222          0.1630055          0.4578517