Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1646384428
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0453   14.8301    0.0000 [ 0.5982; 0.7482 ] 
#>   eta3 ~ eta1      0.4585      0.0779    5.8869    0.0000 [ 0.2622; 0.5685 ] 
#>   eta3 ~ eta2      0.3052      0.0768    3.9717    0.0001 [ 0.2219; 0.5046 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0330   20.1178    0.0000 [ 0.6095; 0.7230 ] 
#>   eta1 =~ y12      0.6493      0.0414   15.6813    0.0000 [ 0.5823; 0.7472 ] 
#>   eta1 =~ y13      0.7613      0.0356   21.3957    0.0000 [ 0.6992; 0.8134 ] 
#>   eta2 =~ y21      0.5165      0.0491   10.5170    0.0000 [ 0.4267; 0.6002 ] 
#>   eta2 =~ y22      0.7554      0.0424   17.8010    0.0000 [ 0.6665; 0.8169 ] 
#>   eta2 =~ y23      0.7997      0.0378   21.1352    0.0000 [ 0.7357; 0.8724 ] 
#>   eta3 =~ y31      0.8223      0.0336   24.4855    0.0000 [ 0.7703; 0.8722 ] 
#>   eta3 =~ y32      0.6581      0.0361   18.2134    0.0000 [ 0.5896; 0.7146 ] 
#>   eta3 =~ y33      0.7474      0.0412   18.1626    0.0000 [ 0.6785; 0.8061 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0205   19.3131    0.0000 [ 0.3600; 0.4349 ] 
#>   eta1 <~ y12      0.3873      0.0170   22.8332    0.0000 [ 0.3684; 0.4284 ] 
#>   eta1 <~ y13      0.4542      0.0181   25.0725    0.0000 [ 0.4180; 0.4828 ] 
#>   eta2 <~ y21      0.3058      0.0256   11.9617    0.0000 [ 0.2702; 0.3533 ] 
#>   eta2 <~ y22      0.4473      0.0240   18.6016    0.0000 [ 0.4093; 0.4867 ] 
#>   eta2 <~ y23      0.4735      0.0207   22.8615    0.0000 [ 0.4407; 0.4997 ] 
#>   eta3 <~ y31      0.4400      0.0181   24.2605    0.0000 [ 0.4110; 0.4738 ] 
#>   eta3 <~ y32      0.3521      0.0157   22.4699    0.0000 [ 0.3199; 0.3740 ] 
#>   eta3 <~ y33      0.3999      0.0218   18.3756    0.0000 [ 0.3653; 0.4372 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0453   14.8301    0.0000 [ 0.5982; 0.7482 ] 
#>   eta3 ~ eta1       0.6634      0.0419   15.8147    0.0000 [ 0.5871; 0.7415 ] 
#>   eta3 ~ eta2       0.3052      0.0768    3.9717    0.0001 [ 0.2219; 0.5046 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0552    3.7144    0.0002 [ 0.1431; 0.3262 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03295939 20.11778  5.156429e-90
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04140467 15.68127  2.031361e-55
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03558408 21.39569 1.465505e-101
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04910670 10.51699  7.213999e-26
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04243510 17.80101  6.940590e-71
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03783572 21.13515  3.779587e-99
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03358221 24.48550 2.108095e-132
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03613109 18.21337  4.042926e-74
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04115187 18.16258  1.021109e-73
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6095263          0.7229574
#> 2          0.5822693          0.7472054
#> 3          0.6991513          0.8134240
#> 4          0.4267493          0.6002369
#> 5          0.6665064          0.8168757
#> 6          0.7356818          0.8724343
#> 7          0.7703274          0.8722410
#> 8          0.5895600          0.7146116
#> 9          0.6785030          0.8060572

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1646384428
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0453   14.8301    0.0000 [ 0.5472; 0.7813 ] 
#>   eta3 ~ eta1      0.4585      0.0779    5.8869    0.0000 [ 0.2736; 0.6764 ] 
#>   eta3 ~ eta2      0.3052      0.0768    3.9717    0.0001 [ 0.0891; 0.4865 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0330   20.1178    0.0000 [ 0.5828; 0.7532 ] 
#>   eta1 =~ y12      0.6493      0.0414   15.6813    0.0000 [ 0.5335; 0.7476 ] 
#>   eta1 =~ y13      0.7613      0.0356   21.3957    0.0000 [ 0.6741; 0.8581 ] 
#>   eta2 =~ y21      0.5165      0.0491   10.5170    0.0000 [ 0.3882; 0.6421 ] 
#>   eta2 =~ y22      0.7554      0.0424   17.8010    0.0000 [ 0.6504; 0.8698 ] 
#>   eta2 =~ y23      0.7997      0.0378   21.1352    0.0000 [ 0.7050; 0.9006 ] 
#>   eta3 =~ y31      0.8223      0.0336   24.4855    0.0000 [ 0.7309; 0.9046 ] 
#>   eta3 =~ y32      0.6581      0.0361   18.2134    0.0000 [ 0.5712; 0.7580 ] 
#>   eta3 =~ y33      0.7474      0.0412   18.1626    0.0000 [ 0.6397; 0.8525 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0205   19.3131    0.0000 [ 0.3452; 0.4511 ] 
#>   eta1 <~ y12      0.3873      0.0170   22.8332    0.0000 [ 0.3383; 0.4260 ] 
#>   eta1 <~ y13      0.4542      0.0181   25.0725    0.0000 [ 0.4099; 0.5036 ] 
#>   eta2 <~ y21      0.3058      0.0256   11.9617    0.0000 [ 0.2379; 0.3701 ] 
#>   eta2 <~ y22      0.4473      0.0240   18.6016    0.0000 [ 0.3859; 0.5103 ] 
#>   eta2 <~ y23      0.4735      0.0207   22.8615    0.0000 [ 0.4197; 0.5268 ] 
#>   eta3 <~ y31      0.4400      0.0181   24.2605    0.0000 [ 0.3907; 0.4845 ] 
#>   eta3 <~ y32      0.3521      0.0157   22.4699    0.0000 [ 0.3153; 0.3963 ] 
#>   eta3 <~ y33      0.3999      0.0218   18.3756    0.0000 [ 0.3430; 0.4556 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0453   14.8301    0.0000 [ 0.5472; 0.7813 ] 
#>   eta3 ~ eta1       0.6634      0.0419   15.8147    0.0000 [ 0.5572; 0.7742 ] 
#>   eta3 ~ eta2       0.3052      0.0768    3.9717    0.0001 [ 0.0891; 0.4865 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0552    3.7144    0.0002 [ 0.0481; 0.3333 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04526819 14.830137 9.354241e-50
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07788616  5.886884 3.935449e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07683144  3.971696 7.136280e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54720986          0.7813114          0.5753209          0.7532004
#> 2         0.27360529          0.6763885          0.3219717          0.6280221
#> 3         0.08912484          0.4864537          0.1368363          0.4387423
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5914385          0.7592426          0.5982377          0.7482462
#> 2          0.2544731          0.6359291          0.2622490          0.5685163
#> 3          0.1283801          0.5321860          0.2219023          0.5045582