Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2011360324
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0469   14.3238    0.0000 [ 0.5730; 0.7527 ] 
#>   eta3 ~ eta1      0.4585      0.0903    5.0752    0.0000 [ 0.3129; 0.6228 ] 
#>   eta3 ~ eta2      0.3052      0.0921    3.3141    0.0009 [ 0.1701; 0.4851 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0297   22.2954    0.0000 [ 0.6095; 0.7034 ] 
#>   eta1 =~ y12      0.6493      0.0437   14.8485    0.0000 [ 0.5704; 0.7193 ] 
#>   eta1 =~ y13      0.7613      0.0349   21.8099    0.0000 [ 0.6945; 0.8176 ] 
#>   eta2 =~ y21      0.5165      0.0489   10.5693    0.0000 [ 0.4320; 0.5837 ] 
#>   eta2 =~ y22      0.7554      0.0396   19.0783    0.0000 [ 0.6879; 0.8288 ] 
#>   eta2 =~ y23      0.7997      0.0403   19.8632    0.0000 [ 0.7342; 0.8753 ] 
#>   eta3 =~ y31      0.8223      0.0309   26.5741    0.0000 [ 0.7660; 0.8735 ] 
#>   eta3 =~ y32      0.6581      0.0342   19.2605    0.0000 [ 0.5970; 0.7145 ] 
#>   eta3 =~ y33      0.7474      0.0299   24.9731    0.0000 [ 0.7001; 0.8081 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0214   18.4581    0.0000 [ 0.3575; 0.4279 ] 
#>   eta1 <~ y12      0.3873      0.0195   19.8715    0.0000 [ 0.3551; 0.4222 ] 
#>   eta1 <~ y13      0.4542      0.0170   26.7874    0.0000 [ 0.4193; 0.4888 ] 
#>   eta2 <~ y21      0.3058      0.0280   10.9303    0.0000 [ 0.2607; 0.3555 ] 
#>   eta2 <~ y22      0.4473      0.0186   24.1077    0.0000 [ 0.4127; 0.4762 ] 
#>   eta2 <~ y23      0.4735      0.0214   22.1267    0.0000 [ 0.4413; 0.5161 ] 
#>   eta3 <~ y31      0.4400      0.0153   28.7261    0.0000 [ 0.3985; 0.4615 ] 
#>   eta3 <~ y32      0.3521      0.0157   22.4773    0.0000 [ 0.3258; 0.3789 ] 
#>   eta3 <~ y33      0.3999      0.0170   23.5522    0.0000 [ 0.3757; 0.4387 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0469   14.3238    0.0000 [ 0.5730; 0.7527 ] 
#>   eta3 ~ eta1       0.6634      0.0427   15.5371    0.0000 [ 0.6046; 0.7506 ] 
#>   eta3 ~ eta2       0.3052      0.0921    3.3141    0.0009 [ 0.1701; 0.4851 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0617    3.3209    0.0009 [ 0.1170; 0.2987 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.02974027 22.29536 4.098769e-110
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04372693 14.84847  7.117769e-50
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03490831 21.80987 1.869915e-105
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04886376 10.56928  4.136555e-26
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03959402 19.07833  3.822835e-81
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04025858 19.86319  8.476074e-88
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03094280 26.57411 1.352323e-155
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03416675 19.26051  1.152465e-82
#> 9 eta3 =~ y33  Common factor 0.7474241 0.02992919 24.97308 1.199039e-137
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6094739          0.7033640
#> 2          0.5704397          0.7193238
#> 3          0.6944886          0.8176180
#> 4          0.4319811          0.5837244
#> 5          0.6879322          0.8287563
#> 6          0.7341848          0.8753294
#> 7          0.7660060          0.8735029
#> 8          0.5970062          0.7145463
#> 9          0.7000701          0.8080889

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2011360324
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0469   14.3238    0.0000 [ 0.5443; 0.7867 ] 
#>   eta3 ~ eta1      0.4585      0.0903    5.0752    0.0000 [ 0.2228; 0.6900 ] 
#>   eta3 ~ eta2      0.3052      0.0921    3.3141    0.0009 [ 0.0572; 0.5334 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0297   22.2954    0.0000 [ 0.5875; 0.7413 ] 
#>   eta1 =~ y12      0.6493      0.0437   14.8485    0.0000 [ 0.5349; 0.7610 ] 
#>   eta1 =~ y13      0.7613      0.0349   21.8099    0.0000 [ 0.6698; 0.8503 ] 
#>   eta2 =~ y21      0.5165      0.0489   10.5693    0.0000 [ 0.3955; 0.6482 ] 
#>   eta2 =~ y22      0.7554      0.0396   19.0783    0.0000 [ 0.6497; 0.8544 ] 
#>   eta2 =~ y23      0.7997      0.0403   19.8632    0.0000 [ 0.6897; 0.8979 ] 
#>   eta3 =~ y31      0.8223      0.0309   26.5741    0.0000 [ 0.7477; 0.9077 ] 
#>   eta3 =~ y32      0.6581      0.0342   19.2605    0.0000 [ 0.5642; 0.7409 ] 
#>   eta3 =~ y33      0.7474      0.0299   24.9731    0.0000 [ 0.6617; 0.8165 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0214   18.4581    0.0000 [ 0.3412; 0.4520 ] 
#>   eta1 <~ y12      0.3873      0.0195   19.8715    0.0000 [ 0.3370; 0.4378 ] 
#>   eta1 <~ y13      0.4542      0.0170   26.7874    0.0000 [ 0.4102; 0.4979 ] 
#>   eta2 <~ y21      0.3058      0.0280   10.9303    0.0000 [ 0.2380; 0.3827 ] 
#>   eta2 <~ y22      0.4473      0.0186   24.1077    0.0000 [ 0.3994; 0.4954 ] 
#>   eta2 <~ y23      0.4735      0.0214   22.1267    0.0000 [ 0.4168; 0.5275 ] 
#>   eta3 <~ y31      0.4400      0.0153   28.7261    0.0000 [ 0.4058; 0.4850 ] 
#>   eta3 <~ y32      0.3521      0.0157   22.4773    0.0000 [ 0.3108; 0.3918 ] 
#>   eta3 <~ y33      0.3999      0.0170   23.5522    0.0000 [ 0.3538; 0.4417 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0469   14.3238    0.0000 [ 0.5443; 0.7867 ] 
#>   eta3 ~ eta1       0.6634      0.0427   15.5371    0.0000 [ 0.5433; 0.7640 ] 
#>   eta3 ~ eta2       0.3052      0.0921    3.3141    0.0009 [ 0.0572; 0.5334 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0617    3.3209    0.0009 [ 0.0378; 0.3568 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04686837 14.323805 1.553778e-46
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09034332  5.075160 3.871708e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09207543  3.314143 9.192452e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54434156          0.7867183          0.5734463          0.7576136
#> 2         0.22278218          0.6899868          0.2788843          0.6338847
#> 3         0.05723608          0.5333982          0.1144138          0.4762204
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5309651          0.7609211          0.5729711          0.7526900
#> 2          0.2490065          0.6914655          0.3128895          0.6227816
#> 3          0.0690236          0.5486399          0.1700553          0.4851321