Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1775305145
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0456   14.7202    0.0000 [ 0.6056; 0.7608 ] 
#>   eta3 ~ eta1      0.4585      0.0848    5.4072    0.0000 [ 0.2702; 0.6230 ] 
#>   eta3 ~ eta2      0.3052      0.0948    3.2203    0.0013 [ 0.1360; 0.4882 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0509   13.0367    0.0000 [ 0.5539; 0.7387 ] 
#>   eta1 =~ y12      0.6493      0.0413   15.7119    0.0000 [ 0.5924; 0.7291 ] 
#>   eta1 =~ y13      0.7613      0.0384   19.8046    0.0000 [ 0.6921; 0.8335 ] 
#>   eta2 =~ y21      0.5165      0.0408   12.6669    0.0000 [ 0.4516; 0.5843 ] 
#>   eta2 =~ y22      0.7554      0.0351   21.4938    0.0000 [ 0.6811; 0.8032 ] 
#>   eta2 =~ y23      0.7997      0.0338   23.6412    0.0000 [ 0.7465; 0.8500 ] 
#>   eta3 =~ y31      0.8223      0.0328   25.0404    0.0000 [ 0.7599; 0.8757 ] 
#>   eta3 =~ y32      0.6581      0.0468   14.0735    0.0000 [ 0.5598; 0.7575 ] 
#>   eta3 =~ y33      0.7474      0.0412   18.1434    0.0000 [ 0.6487; 0.8050 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0251   15.7864    0.0000 [ 0.3413; 0.4363 ] 
#>   eta1 <~ y12      0.3873      0.0190   20.4131    0.0000 [ 0.3600; 0.4261 ] 
#>   eta1 <~ y13      0.4542      0.0261   17.3688    0.0000 [ 0.4076; 0.5101 ] 
#>   eta2 <~ y21      0.3058      0.0190   16.0981    0.0000 [ 0.2738; 0.3413 ] 
#>   eta2 <~ y22      0.4473      0.0156   28.6168    0.0000 [ 0.4170; 0.4705 ] 
#>   eta2 <~ y23      0.4735      0.0217   21.8286    0.0000 [ 0.4380; 0.5205 ] 
#>   eta3 <~ y31      0.4400      0.0215   20.5088    0.0000 [ 0.4092; 0.4885 ] 
#>   eta3 <~ y32      0.3521      0.0188   18.7495    0.0000 [ 0.3168; 0.3872 ] 
#>   eta3 <~ y33      0.3999      0.0219   18.2980    0.0000 [ 0.3603; 0.4424 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0456   14.7202    0.0000 [ 0.6056; 0.7608 ] 
#>   eta3 ~ eta1       0.6634      0.0407   16.2895    0.0000 [ 0.5974; 0.7279 ] 
#>   eta3 ~ eta2       0.3052      0.0948    3.2203    0.0013 [ 0.1360; 0.4882 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0640    3.2030    0.0014 [ 0.0838; 0.3346 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.05086177 13.03670  7.565808e-39
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04132391 15.71192  1.253307e-55
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03844286 19.80461  2.716650e-87
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04077203 12.66689  9.022140e-37
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03514451 21.49376 1.780901e-102
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03382502 23.64119 1.454345e-123
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03283806 25.04037 2.222781e-138
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04675947 14.07349  5.527401e-45
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04119534 18.14342  1.447489e-73
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5538556          0.7386545
#> 2          0.5924355          0.7291267
#> 3          0.6921284          0.8335311
#> 4          0.4515833          0.5843174
#> 5          0.6811423          0.8032036
#> 6          0.7465004          0.8499652
#> 7          0.7599485          0.8757367
#> 8          0.5597985          0.7574543
#> 9          0.6486857          0.8050374

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1775305145
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0456   14.7202    0.0000 [ 0.5416; 0.7774 ] 
#>   eta3 ~ eta1      0.4585      0.0848    5.4072    0.0000 [ 0.2479; 0.6864 ] 
#>   eta3 ~ eta2      0.3052      0.0948    3.2203    0.0013 [ 0.0463; 0.5363 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0509   13.0367    0.0000 [ 0.5358; 0.7988 ] 
#>   eta1 =~ y12      0.6493      0.0413   15.7119    0.0000 [ 0.5346; 0.7483 ] 
#>   eta1 =~ y13      0.7613      0.0384   19.8046    0.0000 [ 0.6681; 0.8669 ] 
#>   eta2 =~ y21      0.5165      0.0408   12.6669    0.0000 [ 0.4039; 0.6147 ] 
#>   eta2 =~ y22      0.7554      0.0351   21.4938    0.0000 [ 0.6692; 0.8509 ] 
#>   eta2 =~ y23      0.7997      0.0338   23.6412    0.0000 [ 0.7209; 0.8958 ] 
#>   eta3 =~ y31      0.8223      0.0328   25.0404    0.0000 [ 0.7319; 0.9017 ] 
#>   eta3 =~ y32      0.6581      0.0468   14.0735    0.0000 [ 0.5466; 0.7884 ] 
#>   eta3 =~ y33      0.7474      0.0412   18.1434    0.0000 [ 0.6472; 0.8603 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0251   15.7864    0.0000 [ 0.3330; 0.4625 ] 
#>   eta1 <~ y12      0.3873      0.0190   20.4131    0.0000 [ 0.3332; 0.4313 ] 
#>   eta1 <~ y13      0.4542      0.0261   17.3688    0.0000 [ 0.3893; 0.5245 ] 
#>   eta2 <~ y21      0.3058      0.0190   16.0981    0.0000 [ 0.2511; 0.3494 ] 
#>   eta2 <~ y22      0.4473      0.0156   28.6168    0.0000 [ 0.4074; 0.4883 ] 
#>   eta2 <~ y23      0.4735      0.0217   21.8286    0.0000 [ 0.4199; 0.5321 ] 
#>   eta3 <~ y31      0.4400      0.0215   20.5088    0.0000 [ 0.3786; 0.4896 ] 
#>   eta3 <~ y32      0.3521      0.0188   18.7495    0.0000 [ 0.3068; 0.4039 ] 
#>   eta3 <~ y33      0.3999      0.0219   18.2980    0.0000 [ 0.3443; 0.4574 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0456   14.7202    0.0000 [ 0.5416; 0.7774 ] 
#>   eta3 ~ eta1       0.6634      0.0407   16.2895    0.0000 [ 0.5541; 0.7647 ] 
#>   eta3 ~ eta2       0.3052      0.0948    3.2203    0.0013 [ 0.0463; 0.5363 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0640    3.2030    0.0014 [ 0.0269; 0.3577 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04560617 14.720232 4.780144e-49
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08479629  5.407156 6.403327e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09475824  3.220312 1.280510e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54157181          0.7774212          0.5698927          0.7491003
#> 2         0.24787969          0.6863982          0.3005372          0.6337407
#> 3         0.04625715          0.5362932          0.1051009          0.4774495
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5972507          0.7672280          0.6055668          0.7608080
#> 2          0.2542919          0.6258631          0.2701782          0.6229718
#> 3          0.1175194          0.4911747          0.1359616          0.4882384