Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1425910167
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0418   16.0510    0.0000 [ 0.5976; 0.7423 ] 
#>   eta3 ~ eta1      0.4585      0.0708    6.4737    0.0000 [ 0.3111; 0.5691 ] 
#>   eta3 ~ eta2      0.3052      0.0781    3.9089    0.0001 [ 0.1505; 0.4735 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0417   15.9125    0.0000 [ 0.5909; 0.7264 ] 
#>   eta1 =~ y12      0.6493      0.0347   18.6956    0.0000 [ 0.5855; 0.6973 ] 
#>   eta1 =~ y13      0.7613      0.0305   24.9374    0.0000 [ 0.7065; 0.8058 ] 
#>   eta2 =~ y21      0.5165      0.0559    9.2433    0.0000 [ 0.4037; 0.6121 ] 
#>   eta2 =~ y22      0.7554      0.0343   22.0118    0.0000 [ 0.7044; 0.8164 ] 
#>   eta2 =~ y23      0.7997      0.0368   21.7208    0.0000 [ 0.7344; 0.8641 ] 
#>   eta3 =~ y31      0.8223      0.0423   19.4337    0.0000 [ 0.7443; 0.8976 ] 
#>   eta3 =~ y32      0.6581      0.0486   13.5396    0.0000 [ 0.5485; 0.7147 ] 
#>   eta3 =~ y33      0.7474      0.0348   21.4896    0.0000 [ 0.6952; 0.8165 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0210   18.8177    0.0000 [ 0.3601; 0.4396 ] 
#>   eta1 <~ y12      0.3873      0.0187   20.7247    0.0000 [ 0.3505; 0.4217 ] 
#>   eta1 <~ y13      0.4542      0.0196   23.2095    0.0000 [ 0.4119; 0.4848 ] 
#>   eta2 <~ y21      0.3058      0.0286   10.7009    0.0000 [ 0.2426; 0.3561 ] 
#>   eta2 <~ y22      0.4473      0.0247   18.1260    0.0000 [ 0.4138; 0.4992 ] 
#>   eta2 <~ y23      0.4735      0.0198   23.9687    0.0000 [ 0.4431; 0.5110 ] 
#>   eta3 <~ y31      0.4400      0.0216   20.3506    0.0000 [ 0.4021; 0.4760 ] 
#>   eta3 <~ y32      0.3521      0.0215   16.3454    0.0000 [ 0.2959; 0.3753 ] 
#>   eta3 <~ y33      0.3999      0.0204   19.5808    0.0000 [ 0.3739; 0.4403 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0418   16.0510    0.0000 [ 0.5976; 0.7423 ] 
#>   eta3 ~ eta1       0.6634      0.0301   22.0543    0.0000 [ 0.5833; 0.7335 ] 
#>   eta3 ~ eta2       0.3052      0.0781    3.9089    0.0001 [ 0.1505; 0.4735 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0582    3.5197    0.0004 [ 0.1005; 0.3512 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04166963 15.912544  5.186259e-57
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03472891 18.695603  5.375877e-78
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03053023 24.937439 2.922533e-137
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05587322  9.243334  2.389344e-20
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03431740 22.011801 2.219981e-107
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03681553 21.720825 1.304150e-104
#> 7 eta3 =~ y31  Common factor 0.8222773 0.04231189 19.433720  4.002746e-84
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04860327 13.539603  9.128119e-42
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03478075 21.489590 1.948222e-102
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5909176          0.7263727
#> 2          0.5854549          0.6972870
#> 3          0.7064926          0.8058044
#> 4          0.4036549          0.6121032
#> 5          0.7043506          0.8164417
#> 6          0.7344454          0.8640764
#> 7          0.7442634          0.8976113
#> 8          0.5485013          0.7146920
#> 9          0.6952153          0.8164549

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1425910167
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0418   16.0510    0.0000 [ 0.5649; 0.7812 ] 
#>   eta3 ~ eta1      0.4585      0.0708    6.4737    0.0000 [ 0.2750; 0.6412 ] 
#>   eta3 ~ eta2      0.3052      0.0781    3.9089    0.0001 [ 0.1074; 0.5111 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0417   15.9125    0.0000 [ 0.5566; 0.7721 ] 
#>   eta1 =~ y12      0.6493      0.0347   18.6956    0.0000 [ 0.5686; 0.7482 ] 
#>   eta1 =~ y13      0.7613      0.0305   24.9374    0.0000 [ 0.6879; 0.8458 ] 
#>   eta2 =~ y21      0.5165      0.0559    9.2433    0.0000 [ 0.3949; 0.6838 ] 
#>   eta2 =~ y22      0.7554      0.0343   22.0118    0.0000 [ 0.6608; 0.8383 ] 
#>   eta2 =~ y23      0.7997      0.0368   21.7208    0.0000 [ 0.7056; 0.8960 ] 
#>   eta3 =~ y31      0.8223      0.0423   19.4337    0.0000 [ 0.7128; 0.9316 ] 
#>   eta3 =~ y32      0.6581      0.0486   13.5396    0.0000 [ 0.5330; 0.7843 ] 
#>   eta3 =~ y33      0.7474      0.0348   21.4896    0.0000 [ 0.6508; 0.8306 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0210   18.8177    0.0000 [ 0.3378; 0.4465 ] 
#>   eta1 <~ y12      0.3873      0.0187   20.7247    0.0000 [ 0.3403; 0.4370 ] 
#>   eta1 <~ y13      0.4542      0.0196   23.2095    0.0000 [ 0.4018; 0.5030 ] 
#>   eta2 <~ y21      0.3058      0.0286   10.7009    0.0000 [ 0.2431; 0.3909 ] 
#>   eta2 <~ y22      0.4473      0.0247   18.1260    0.0000 [ 0.3755; 0.5031 ] 
#>   eta2 <~ y23      0.4735      0.0198   23.9687    0.0000 [ 0.4186; 0.5208 ] 
#>   eta3 <~ y31      0.4400      0.0216   20.3506    0.0000 [ 0.3861; 0.4979 ] 
#>   eta3 <~ y32      0.3521      0.0215   16.3454    0.0000 [ 0.2987; 0.4101 ] 
#>   eta3 <~ y33      0.3999      0.0204   19.5808    0.0000 [ 0.3453; 0.4510 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0418   16.0510    0.0000 [ 0.5649; 0.7812 ] 
#>   eta3 ~ eta1       0.6634      0.0301   22.0543    0.0000 [ 0.5876; 0.7431 ] 
#>   eta3 ~ eta2       0.3052      0.0781    3.9089    0.0001 [ 0.1074; 0.5111 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0582    3.5197    0.0004 [ 0.0567; 0.3577 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04182509 16.050973 5.627451e-58
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07082650  6.473661 9.565618e-11
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07806593  3.908890 9.272129e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5648979          0.7811936          0.5908708          0.7552207
#> 2          0.2749747          0.6412493          0.3189571          0.5972669
#> 3          0.1073952          0.5111081          0.1558732          0.4626301
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5866109          0.7520213          0.5975828          0.7423396
#> 2          0.2835937          0.5736027          0.3110540          0.5690922
#> 3          0.1397619          0.4989676          0.1504618          0.4734967