Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -405822513
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0406   16.5268    0.0000 [ 0.5896; 0.7381 ] 
#>   eta3 ~ eta1      0.4585      0.0829    5.5287    0.0000 [ 0.3153; 0.6037 ] 
#>   eta3 ~ eta2      0.3052      0.0957    3.1889    0.0014 [ 0.1343; 0.4518 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0413   16.0357    0.0000 [ 0.6147; 0.7507 ] 
#>   eta1 =~ y12      0.6493      0.0425   15.2924    0.0000 [ 0.5676; 0.7028 ] 
#>   eta1 =~ y13      0.7613      0.0273   27.9107    0.0000 [ 0.7122; 0.8083 ] 
#>   eta2 =~ y21      0.5165      0.0485   10.6380    0.0000 [ 0.4270; 0.5896 ] 
#>   eta2 =~ y22      0.7554      0.0367   20.5946    0.0000 [ 0.6972; 0.8234 ] 
#>   eta2 =~ y23      0.7997      0.0331   24.1614    0.0000 [ 0.7271; 0.8474 ] 
#>   eta3 =~ y31      0.8223      0.0340   24.1514    0.0000 [ 0.7708; 0.8742 ] 
#>   eta3 =~ y32      0.6581      0.0407   16.1624    0.0000 [ 0.5997; 0.7277 ] 
#>   eta3 =~ y33      0.7474      0.0388   19.2544    0.0000 [ 0.6831; 0.8235 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0189   20.9220    0.0000 [ 0.3709; 0.4470 ] 
#>   eta1 <~ y12      0.3873      0.0213   18.1912    0.0000 [ 0.3363; 0.4206 ] 
#>   eta1 <~ y13      0.4542      0.0195   23.2520    0.0000 [ 0.4119; 0.4859 ] 
#>   eta2 <~ y21      0.3058      0.0258   11.8689    0.0000 [ 0.2588; 0.3565 ] 
#>   eta2 <~ y22      0.4473      0.0200   22.3995    0.0000 [ 0.4062; 0.4836 ] 
#>   eta2 <~ y23      0.4735      0.0192   24.7001    0.0000 [ 0.4422; 0.4963 ] 
#>   eta3 <~ y31      0.4400      0.0179   24.5417    0.0000 [ 0.4093; 0.4669 ] 
#>   eta3 <~ y32      0.3521      0.0187   18.8337    0.0000 [ 0.3182; 0.3836 ] 
#>   eta3 <~ y33      0.3999      0.0200   19.9731    0.0000 [ 0.3652; 0.4371 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0406   16.5268    0.0000 [ 0.5896; 0.7381 ] 
#>   eta3 ~ eta1       0.6634      0.0339   19.5665    0.0000 [ 0.5988; 0.7182 ] 
#>   eta3 ~ eta2       0.3052      0.0957    3.1889    0.0014 [ 0.1343; 0.4518 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0647    3.1648    0.0016 [ 0.0891; 0.3185 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04134973 16.03565  7.202489e-58
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04245754 15.29241  8.591648e-53
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02727789 27.91073 1.976840e-171
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04854790 10.63805  1.982446e-26
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03667884 20.59464  3.065350e-94
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03309681 24.16135 5.672972e-129
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03404673 24.15143 7.212154e-129
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04071596 16.16243  9.282982e-59
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03881828 19.25444  1.295716e-82
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6146667          0.7507075
#> 2          0.5676080          0.7027801
#> 3          0.7122026          0.8082772
#> 4          0.4269770          0.5896031
#> 5          0.6972024          0.8234397
#> 6          0.7270703          0.8474232
#> 7          0.7707769          0.8741606
#> 8          0.5996903          0.7277303
#> 9          0.6831073          0.8235427

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -405822513
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0406   16.5268    0.0000 [ 0.5637; 0.7738 ] 
#>   eta3 ~ eta1      0.4585      0.0829    5.5287    0.0000 [ 0.2586; 0.6875 ] 
#>   eta3 ~ eta2      0.3052      0.0957    3.1889    0.0014 [ 0.0388; 0.5336 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0413   16.0357    0.0000 [ 0.5418; 0.7556 ] 
#>   eta1 =~ y12      0.6493      0.0425   15.2924    0.0000 [ 0.5480; 0.7676 ] 
#>   eta1 =~ y13      0.7613      0.0273   27.9107    0.0000 [ 0.6942; 0.8352 ] 
#>   eta2 =~ y21      0.5165      0.0485   10.6380    0.0000 [ 0.3832; 0.6343 ] 
#>   eta2 =~ y22      0.7554      0.0367   20.5946    0.0000 [ 0.6544; 0.8441 ] 
#>   eta2 =~ y23      0.7997      0.0331   24.1614    0.0000 [ 0.7260; 0.8972 ] 
#>   eta3 =~ y31      0.8223      0.0340   24.1514    0.0000 [ 0.7316; 0.9077 ] 
#>   eta3 =~ y32      0.6581      0.0407   16.1624    0.0000 [ 0.5543; 0.7648 ] 
#>   eta3 =~ y33      0.7474      0.0388   19.2544    0.0000 [ 0.6415; 0.8423 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0189   20.9220    0.0000 [ 0.3392; 0.4370 ] 
#>   eta1 <~ y12      0.3873      0.0213   18.1912    0.0000 [ 0.3384; 0.4485 ] 
#>   eta1 <~ y13      0.4542      0.0195   23.2520    0.0000 [ 0.4064; 0.5074 ] 
#>   eta2 <~ y21      0.3058      0.0258   11.8689    0.0000 [ 0.2352; 0.3685 ] 
#>   eta2 <~ y22      0.4473      0.0200   22.3995    0.0000 [ 0.3925; 0.4958 ] 
#>   eta2 <~ y23      0.4735      0.0192   24.7001    0.0000 [ 0.4315; 0.5306 ] 
#>   eta3 <~ y31      0.4400      0.0179   24.5417    0.0000 [ 0.3944; 0.4871 ] 
#>   eta3 <~ y32      0.3521      0.0187   18.8337    0.0000 [ 0.3065; 0.4032 ] 
#>   eta3 <~ y33      0.3999      0.0200   19.9731    0.0000 [ 0.3472; 0.4508 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0406   16.5268    0.0000 [ 0.5637; 0.7738 ] 
#>   eta3 ~ eta1       0.6634      0.0339   19.5665    0.0000 [ 0.5772; 0.7525 ] 
#>   eta3 ~ eta2       0.3052      0.0957    3.1889    0.0014 [ 0.0388; 0.5336 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0647    3.1648    0.0016 [ 0.0244; 0.3592 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04062088 16.526807 2.352816e-61
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08293174  5.528725 3.225671e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09569117  3.188916 1.428073e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56373549          0.7738037         0.58896058          0.7485786
#> 2         0.25861759          0.6874937         0.31011723          0.6359941
#> 3         0.03878852          0.5336492         0.09821162          0.4742261
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5871632          0.7620147          0.5895984          0.7380581
#> 2          0.2940965          0.6413305          0.3152762          0.6037059
#> 3          0.1049289          0.4631753          0.1343231          0.4518179