Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -135949826
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0467   14.3767    0.0000 [ 0.5654; 0.7423 ] 
#>   eta3 ~ eta1      0.4585      0.0903    5.0801    0.0000 [ 0.2693; 0.5933 ] 
#>   eta3 ~ eta2      0.3052      0.0868    3.5173    0.0004 [ 0.1632; 0.5183 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0409   16.2145    0.0000 [ 0.5957; 0.7160 ] 
#>   eta1 =~ y12      0.6493      0.0342   18.9587    0.0000 [ 0.5927; 0.7172 ] 
#>   eta1 =~ y13      0.7613      0.0311   24.4858    0.0000 [ 0.7084; 0.8087 ] 
#>   eta2 =~ y21      0.5165      0.0576    8.9617    0.0000 [ 0.4100; 0.6059 ] 
#>   eta2 =~ y22      0.7554      0.0385   19.6027    0.0000 [ 0.6906; 0.8487 ] 
#>   eta2 =~ y23      0.7997      0.0421   18.9830    0.0000 [ 0.7277; 0.8815 ] 
#>   eta3 =~ y31      0.8223      0.0292   28.1552    0.0000 [ 0.7718; 0.8834 ] 
#>   eta3 =~ y32      0.6581      0.0416   15.8114    0.0000 [ 0.6008; 0.7538 ] 
#>   eta3 =~ y33      0.7474      0.0355   21.0752    0.0000 [ 0.6724; 0.8059 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0224   17.6598    0.0000 [ 0.3535; 0.4290 ] 
#>   eta1 <~ y12      0.3873      0.0167   23.2577    0.0000 [ 0.3578; 0.4167 ] 
#>   eta1 <~ y13      0.4542      0.0202   22.4513    0.0000 [ 0.4205; 0.4947 ] 
#>   eta2 <~ y21      0.3058      0.0295   10.3768    0.0000 [ 0.2435; 0.3557 ] 
#>   eta2 <~ y22      0.4473      0.0234   19.0938    0.0000 [ 0.4153; 0.4993 ] 
#>   eta2 <~ y23      0.4735      0.0254   18.6101    0.0000 [ 0.4363; 0.5228 ] 
#>   eta3 <~ y31      0.4400      0.0179   24.5345    0.0000 [ 0.4156; 0.4797 ] 
#>   eta3 <~ y32      0.3521      0.0182   19.3175    0.0000 [ 0.3232; 0.3972 ] 
#>   eta3 <~ y33      0.3999      0.0181   22.1309    0.0000 [ 0.3669; 0.4294 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0467   14.3767    0.0000 [ 0.5654; 0.7423 ] 
#>   eta3 ~ eta1       0.6634      0.0425   15.6145    0.0000 [ 0.5811; 0.7401 ] 
#>   eta3 ~ eta2       0.3052      0.0868    3.5173    0.0004 [ 0.1632; 0.5183 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0601    3.4083    0.0007 [ 0.1202; 0.3339 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04089357 16.214526  3.981401e-59
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03424699 18.958684  3.743953e-80
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03109338 24.485781 2.093717e-132
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05762891  8.961731  3.196204e-19
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03853479 19.602745  1.465030e-85
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04212523 18.983010  2.356999e-80
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02920518 28.155189 2.070309e-174
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04161992 15.811391  2.596701e-56
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03546466 21.075178  1.343987e-98
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5957427          0.7159526
#> 2          0.5927148          0.7172040
#> 3          0.7084265          0.8087466
#> 4          0.4099931          0.6058921
#> 5          0.6905682          0.8487135
#> 6          0.7276741          0.8815307
#> 7          0.7717907          0.8834342
#> 8          0.6007652          0.7537982
#> 9          0.6723790          0.8059450

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -135949826
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0467   14.3767    0.0000 [ 0.5549; 0.7964 ] 
#>   eta3 ~ eta1      0.4585      0.0903    5.0801    0.0000 [ 0.2307; 0.6975 ] 
#>   eta3 ~ eta2      0.3052      0.0868    3.5173    0.0004 [ 0.0719; 0.5206 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0409   16.2145    0.0000 [ 0.5585; 0.7700 ] 
#>   eta1 =~ y12      0.6493      0.0342   18.9587    0.0000 [ 0.5601; 0.7372 ] 
#>   eta1 =~ y13      0.7613      0.0311   24.4858    0.0000 [ 0.6804; 0.8412 ] 
#>   eta2 =~ y21      0.5165      0.0576    8.9617    0.0000 [ 0.3684; 0.6664 ] 
#>   eta2 =~ y22      0.7554      0.0385   19.6027    0.0000 [ 0.6501; 0.8493 ] 
#>   eta2 =~ y23      0.7997      0.0421   18.9830    0.0000 [ 0.6901; 0.9080 ] 
#>   eta3 =~ y31      0.8223      0.0292   28.1552    0.0000 [ 0.7502; 0.9012 ] 
#>   eta3 =~ y32      0.6581      0.0416   15.8114    0.0000 [ 0.5427; 0.7580 ] 
#>   eta3 =~ y33      0.7474      0.0355   21.0752    0.0000 [ 0.6627; 0.8461 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0224   17.6598    0.0000 [ 0.3386; 0.4545 ] 
#>   eta1 <~ y12      0.3873      0.0167   23.2577    0.0000 [ 0.3442; 0.4303 ] 
#>   eta1 <~ y13      0.4542      0.0202   22.4513    0.0000 [ 0.4016; 0.5062 ] 
#>   eta2 <~ y21      0.3058      0.0295   10.3768    0.0000 [ 0.2322; 0.3846 ] 
#>   eta2 <~ y22      0.4473      0.0234   19.0938    0.0000 [ 0.3856; 0.5067 ] 
#>   eta2 <~ y23      0.4735      0.0254   18.6101    0.0000 [ 0.4097; 0.5413 ] 
#>   eta3 <~ y31      0.4400      0.0179   24.5345    0.0000 [ 0.3946; 0.4873 ] 
#>   eta3 <~ y32      0.3521      0.0182   19.3175    0.0000 [ 0.3004; 0.3947 ] 
#>   eta3 <~ y33      0.3999      0.0181   22.1309    0.0000 [ 0.3562; 0.4497 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0467   14.3767    0.0000 [ 0.5549; 0.7964 ] 
#>   eta3 ~ eta1       0.6634      0.0425   15.6145    0.0000 [ 0.5549; 0.7746 ] 
#>   eta3 ~ eta2       0.3052      0.0868    3.5173    0.0004 [ 0.0719; 0.5206 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0601    3.4083    0.0007 [ 0.0453; 0.3561 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04669605 14.376663 7.250680e-47
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09025476  5.080139 3.771582e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08675814  3.517262 4.360230e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55488917          0.7963748          0.5838869          0.7673771
#> 2         0.23072143          0.6974681          0.2867686          0.6414210
#> 3         0.07192484          0.5205890          0.1258006          0.4667132
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5617506          0.7458630          0.5654071          0.7423416
#> 2          0.2282685          0.6279651          0.2692793          0.5932814
#> 3          0.1478368          0.5208898          0.1632050          0.5183169