Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2135914799
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0402   16.7006    0.0000 [ 0.6192; 0.7501 ] 
#>   eta3 ~ eta1      0.4585      0.0883    5.1951    0.0000 [ 0.3355; 0.6339 ] 
#>   eta3 ~ eta2      0.3052      0.0964    3.1665    0.0015 [ 0.0913; 0.4246 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0483   13.7158    0.0000 [ 0.5394; 0.7590 ] 
#>   eta1 =~ y12      0.6493      0.0385   16.8703    0.0000 [ 0.5793; 0.7084 ] 
#>   eta1 =~ y13      0.7613      0.0301   25.2660    0.0000 [ 0.7204; 0.8210 ] 
#>   eta2 =~ y21      0.5165      0.0517    9.9915    0.0000 [ 0.4514; 0.6276 ] 
#>   eta2 =~ y22      0.7554      0.0415   18.2031    0.0000 [ 0.6802; 0.8217 ] 
#>   eta2 =~ y23      0.7997      0.0338   23.6721    0.0000 [ 0.7506; 0.8594 ] 
#>   eta3 =~ y31      0.8223      0.0326   25.2006    0.0000 [ 0.7660; 0.8783 ] 
#>   eta3 =~ y32      0.6581      0.0418   15.7433    0.0000 [ 0.5766; 0.7144 ] 
#>   eta3 =~ y33      0.7474      0.0444   16.8298    0.0000 [ 0.6795; 0.8170 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0219   18.0284    0.0000 [ 0.3430; 0.4325 ] 
#>   eta1 <~ y12      0.3873      0.0199   19.4320    0.0000 [ 0.3477; 0.4199 ] 
#>   eta1 <~ y13      0.4542      0.0225   20.1653    0.0000 [ 0.4178; 0.5068 ] 
#>   eta2 <~ y21      0.3058      0.0258   11.8384    0.0000 [ 0.2713; 0.3706 ] 
#>   eta2 <~ y22      0.4473      0.0204   21.9372    0.0000 [ 0.4087; 0.4806 ] 
#>   eta2 <~ y23      0.4735      0.0255   18.5619    0.0000 [ 0.4339; 0.5118 ] 
#>   eta3 <~ y31      0.4400      0.0186   23.6297    0.0000 [ 0.4089; 0.4706 ] 
#>   eta3 <~ y32      0.3521      0.0202   17.4280    0.0000 [ 0.3124; 0.3896 ] 
#>   eta3 <~ y33      0.3999      0.0215   18.6439    0.0000 [ 0.3643; 0.4355 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0402   16.7006    0.0000 [ 0.6192; 0.7501 ] 
#>   eta3 ~ eta1       0.6634      0.0308   21.5489    0.0000 [ 0.6100; 0.7273 ] 
#>   eta3 ~ eta2       0.3052      0.0964    3.1665    0.0015 [ 0.0913; 0.4246 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0683    2.9997    0.0027 [ 0.0561; 0.2742 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04834335 13.715843  8.161211e-43
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03848649 16.870281  7.444179e-64
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03013323 25.265987 7.558149e-141
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05168963  9.991458  1.661201e-23
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04149783 18.203065  4.879949e-74
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03378082 23.672125 6.986230e-124
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03262932 25.200566 3.948832e-140
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04179996 15.743290  7.636741e-56
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04441085 16.829764  1.476888e-63
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5394312          0.7589983
#> 2          0.5792560          0.7084328
#> 3          0.7204272          0.8209635
#> 4          0.4513830          0.6275788
#> 5          0.6802036          0.8217137
#> 6          0.7505839          0.8594403
#> 7          0.7660102          0.8782950
#> 8          0.5765503          0.7144402
#> 9          0.6795176          0.8170235

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -2135914799
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0402   16.7006    0.0000 [ 0.5684; 0.7763 ] 
#>   eta3 ~ eta1      0.4585      0.0883    5.1951    0.0000 [ 0.2089; 0.6653 ] 
#>   eta3 ~ eta2      0.3052      0.0964    3.1665    0.0015 [ 0.0745; 0.5728 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0483   13.7158    0.0000 [ 0.5456; 0.7956 ] 
#>   eta1 =~ y12      0.6493      0.0385   16.8703    0.0000 [ 0.5438; 0.7428 ] 
#>   eta1 =~ y13      0.7613      0.0301   25.2660    0.0000 [ 0.6764; 0.8323 ] 
#>   eta2 =~ y21      0.5165      0.0517    9.9915    0.0000 [ 0.3709; 0.6382 ] 
#>   eta2 =~ y22      0.7554      0.0415   18.2031    0.0000 [ 0.6509; 0.8655 ] 
#>   eta2 =~ y23      0.7997      0.0338   23.6721    0.0000 [ 0.7122; 0.8869 ] 
#>   eta3 =~ y31      0.8223      0.0326   25.2006    0.0000 [ 0.7413; 0.9101 ] 
#>   eta3 =~ y32      0.6581      0.0418   15.7433    0.0000 [ 0.5532; 0.7693 ] 
#>   eta3 =~ y33      0.7474      0.0444   16.8298    0.0000 [ 0.6263; 0.8560 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0219   18.0284    0.0000 [ 0.3454; 0.4589 ] 
#>   eta1 <~ y12      0.3873      0.0199   19.4320    0.0000 [ 0.3341; 0.4371 ] 
#>   eta1 <~ y13      0.4542      0.0225   20.1653    0.0000 [ 0.3936; 0.5100 ] 
#>   eta2 <~ y21      0.3058      0.0258   11.8384    0.0000 [ 0.2343; 0.3679 ] 
#>   eta2 <~ y22      0.4473      0.0204   21.9372    0.0000 [ 0.3992; 0.5046 ] 
#>   eta2 <~ y23      0.4735      0.0255   18.5619    0.0000 [ 0.4101; 0.5420 ] 
#>   eta3 <~ y31      0.4400      0.0186   23.6297    0.0000 [ 0.3937; 0.4900 ] 
#>   eta3 <~ y32      0.3521      0.0202   17.4280    0.0000 [ 0.3018; 0.4063 ] 
#>   eta3 <~ y33      0.3999      0.0215   18.6439    0.0000 [ 0.3413; 0.4523 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0402   16.7006    0.0000 [ 0.5684; 0.7763 ] 
#>   eta3 ~ eta1       0.6634      0.0308   21.5489    0.0000 [ 0.5752; 0.7343 ] 
#>   eta3 ~ eta2       0.3052      0.0964    3.1665    0.0015 [ 0.0745; 0.5728 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0683    2.9997    0.0027 [ 0.0411; 0.3943 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04019817 16.700598 1.297582e-62
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08825796  5.195076 2.046368e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09636985  3.166458 1.543075e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56842798          0.7763102          0.5933906          0.7513476
#> 2         0.20885546          0.6652758          0.2636626          0.6104686
#> 3         0.07445345          0.5728239          0.1342980          0.5129794
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.6136679          0.7653512         0.61916264          0.7501242
#> 2          0.2404685          0.6561143         0.33549765          0.6338649
#> 3          0.0722890          0.5498025         0.09126963          0.4245733