Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 270651386
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0465   14.4244    0.0000 [ 0.5666; 0.7157 ] 
#>   eta3 ~ eta1      0.4585      0.0958    4.7883    0.0000 [ 0.3014; 0.6519 ] 
#>   eta3 ~ eta2      0.3052      0.1031    2.9585    0.0031 [ 0.1081; 0.4764 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0305   21.7587    0.0000 [ 0.6011; 0.7231 ] 
#>   eta1 =~ y12      0.6493      0.0371   17.5217    0.0000 [ 0.6016; 0.7190 ] 
#>   eta1 =~ y13      0.7613      0.0313   24.3075    0.0000 [ 0.7033; 0.8188 ] 
#>   eta2 =~ y21      0.5165      0.0541    9.5494    0.0000 [ 0.4157; 0.6173 ] 
#>   eta2 =~ y22      0.7554      0.0367   20.5862    0.0000 [ 0.6938; 0.8320 ] 
#>   eta2 =~ y23      0.7997      0.0439   18.2014    0.0000 [ 0.7227; 0.8882 ] 
#>   eta3 =~ y31      0.8223      0.0284   28.9642    0.0000 [ 0.7721; 0.8832 ] 
#>   eta3 =~ y32      0.6581      0.0421   15.6225    0.0000 [ 0.5833; 0.7169 ] 
#>   eta3 =~ y33      0.7474      0.0393   19.0181    0.0000 [ 0.6756; 0.8104 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0190   20.8404    0.0000 [ 0.3596; 0.4248 ] 
#>   eta1 <~ y12      0.3873      0.0189   20.4560    0.0000 [ 0.3569; 0.4148 ] 
#>   eta1 <~ y13      0.4542      0.0164   27.7259    0.0000 [ 0.4177; 0.4798 ] 
#>   eta2 <~ y21      0.3058      0.0292   10.4570    0.0000 [ 0.2536; 0.3521 ] 
#>   eta2 <~ y22      0.4473      0.0232   19.2514    0.0000 [ 0.4135; 0.4806 ] 
#>   eta2 <~ y23      0.4735      0.0214   22.1326    0.0000 [ 0.4373; 0.5069 ] 
#>   eta3 <~ y31      0.4400      0.0195   22.5200    0.0000 [ 0.4156; 0.4889 ] 
#>   eta3 <~ y32      0.3521      0.0190   18.4938    0.0000 [ 0.3132; 0.3787 ] 
#>   eta3 <~ y33      0.3999      0.0166   24.0535    0.0000 [ 0.3741; 0.4248 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0465   14.4244    0.0000 [ 0.5666; 0.7157 ] 
#>   eta3 ~ eta1       0.6634      0.0406   16.3193    0.0000 [ 0.5919; 0.7267 ] 
#>   eta3 ~ eta2       0.3052      0.1031    2.9585    0.0031 [ 0.1081; 0.4764 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0668    3.0685    0.0022 [ 0.0730; 0.3084 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03047373 21.758737 5.709881e-105
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03705566 17.521695  9.786554e-69
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03132139 24.307535 1.631760e-130
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05408216  9.549449  1.303884e-21
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03669383 20.586229  3.646719e-94
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04393414 18.201420  5.028692e-74
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02838945 28.964187 1.859968e-184
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04212306 15.622535  5.113193e-55
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03930077 19.018052  1.208932e-80
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6010720          0.7230519
#> 2          0.6015554          0.7189955
#> 3          0.7033453          0.8187574
#> 4          0.4157276          0.6173029
#> 5          0.6937814          0.8320238
#> 6          0.7227426          0.8881519
#> 7          0.7720998          0.8831584
#> 8          0.5833388          0.7169311
#> 9          0.6756097          0.8103616

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 270651386
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0465   14.4244    0.0000 [ 0.5644; 0.8050 ] 
#>   eta3 ~ eta1      0.4585      0.0958    4.7883    0.0000 [ 0.2011; 0.6963 ] 
#>   eta3 ~ eta2      0.3052      0.1031    2.9585    0.0031 [ 0.0432; 0.5766 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0305   21.7587    0.0000 [ 0.5847; 0.7423 ] 
#>   eta1 =~ y12      0.6493      0.0371   17.5217    0.0000 [ 0.5424; 0.7340 ] 
#>   eta1 =~ y13      0.7613      0.0313   24.3075    0.0000 [ 0.6762; 0.8382 ] 
#>   eta2 =~ y21      0.5165      0.0541    9.5494    0.0000 [ 0.3829; 0.6626 ] 
#>   eta2 =~ y22      0.7554      0.0367   20.5862    0.0000 [ 0.6563; 0.8460 ] 
#>   eta2 =~ y23      0.7997      0.0439   18.2014    0.0000 [ 0.6850; 0.9122 ] 
#>   eta3 =~ y31      0.8223      0.0284   28.9642    0.0000 [ 0.7421; 0.8889 ] 
#>   eta3 =~ y32      0.6581      0.0421   15.6225    0.0000 [ 0.5576; 0.7754 ] 
#>   eta3 =~ y33      0.7474      0.0393   19.0181    0.0000 [ 0.6513; 0.8545 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0190   20.8404    0.0000 [ 0.3510; 0.4491 ] 
#>   eta1 <~ y12      0.3873      0.0189   20.4560    0.0000 [ 0.3362; 0.4341 ] 
#>   eta1 <~ y13      0.4542      0.0164   27.7259    0.0000 [ 0.4143; 0.4991 ] 
#>   eta2 <~ y21      0.3058      0.0292   10.4570    0.0000 [ 0.2345; 0.3857 ] 
#>   eta2 <~ y22      0.4473      0.0232   19.2514    0.0000 [ 0.3850; 0.5051 ] 
#>   eta2 <~ y23      0.4735      0.0214   22.1326    0.0000 [ 0.4182; 0.5288 ] 
#>   eta3 <~ y31      0.4400      0.0195   22.5200    0.0000 [ 0.3838; 0.4848 ] 
#>   eta3 <~ y32      0.3521      0.0190   18.4938    0.0000 [ 0.3061; 0.4046 ] 
#>   eta3 <~ y33      0.3999      0.0166   24.0535    0.0000 [ 0.3584; 0.4444 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0465   14.4244    0.0000 [ 0.5644; 0.8050 ] 
#>   eta3 ~ eta1       0.6634      0.0406   16.3193    0.0000 [ 0.5568; 0.7670 ] 
#>   eta3 ~ eta2       0.3052      0.1031    2.9585    0.0031 [ 0.0432; 0.5766 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0668    3.0685    0.0022 [ 0.0405; 0.3858 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04654143 14.424426 3.632781e-47
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09575512  4.788326 1.681780e-06
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.10314311  2.958522 3.091184e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56435473          0.8050407          0.5932564          0.7761390
#> 2         0.20112281          0.6963142          0.2605856          0.6368514
#> 3         0.04319986          0.5765978          0.1072505          0.5125471
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5148273          0.7635805          0.5666203          0.7157486
#> 2          0.3001804          0.6597905          0.3013867          0.6519153
#> 3          0.1007406          0.4773263          0.1080611          0.4764425