Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1137170062
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0393   17.0721    0.0000 [ 0.5884; 0.7387 ] 
#>   eta3 ~ eta1      0.4585      0.0753    6.0928    0.0000 [ 0.2836; 0.5815 ] 
#>   eta3 ~ eta2      0.3052      0.0795    3.8393    0.0001 [ 0.1654; 0.4492 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0359   18.4793    0.0000 [ 0.5985; 0.7252 ] 
#>   eta1 =~ y12      0.6493      0.0346   18.7729    0.0000 [ 0.5839; 0.6996 ] 
#>   eta1 =~ y13      0.7613      0.0354   21.4854    0.0000 [ 0.7116; 0.8256 ] 
#>   eta2 =~ y21      0.5165      0.0436   11.8355    0.0000 [ 0.4345; 0.6041 ] 
#>   eta2 =~ y22      0.7554      0.0353   21.3945    0.0000 [ 0.6912; 0.8281 ] 
#>   eta2 =~ y23      0.7997      0.0404   19.8101    0.0000 [ 0.7051; 0.8428 ] 
#>   eta3 =~ y31      0.8223      0.0325   25.2991    0.0000 [ 0.7650; 0.8773 ] 
#>   eta3 =~ y32      0.6581      0.0499   13.1913    0.0000 [ 0.5437; 0.7325 ] 
#>   eta3 =~ y33      0.7474      0.0420   17.7909    0.0000 [ 0.6543; 0.7973 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0181   21.8778    0.0000 [ 0.3696; 0.4325 ] 
#>   eta1 <~ y12      0.3873      0.0163   23.8272    0.0000 [ 0.3541; 0.4168 ] 
#>   eta1 <~ y13      0.4542      0.0197   23.0731    0.0000 [ 0.4214; 0.4967 ] 
#>   eta2 <~ y21      0.3058      0.0247   12.3607    0.0000 [ 0.2619; 0.3512 ] 
#>   eta2 <~ y22      0.4473      0.0216   20.6963    0.0000 [ 0.4068; 0.4913 ] 
#>   eta2 <~ y23      0.4735      0.0187   25.3833    0.0000 [ 0.4390; 0.5014 ] 
#>   eta3 <~ y31      0.4400      0.0198   22.1726    0.0000 [ 0.4142; 0.4903 ] 
#>   eta3 <~ y32      0.3521      0.0204   17.2460    0.0000 [ 0.3170; 0.3910 ] 
#>   eta3 <~ y33      0.3999      0.0250   15.9969    0.0000 [ 0.3588; 0.4375 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0393   17.0721    0.0000 [ 0.5884; 0.7387 ] 
#>   eta3 ~ eta1       0.6634      0.0408   16.2678    0.0000 [ 0.5977; 0.7325 ] 
#>   eta3 ~ eta2       0.3052      0.0795    3.8393    0.0001 [ 0.1654; 0.4492 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0539    3.8040    0.0001 [ 0.1208; 0.3210 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03588167 18.47935  3.028308e-76
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03458596 18.77288  1.258792e-78
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03543545 21.48543 2.130831e-102
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04363590 11.83555  2.556680e-32
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03530751 21.39453 1.502263e-101
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04036650 19.81008  2.436847e-87
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03250229 25.29906 3.271088e-141
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04988650 13.19132  9.844459e-40
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04201162 17.79089  8.315098e-71
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5984717          0.7252319
#> 2          0.5838699          0.6996047
#> 3          0.7116060          0.8255621
#> 4          0.4345157          0.6040840
#> 5          0.6911655          0.8280902
#> 6          0.7051372          0.8428416
#> 7          0.7649726          0.8773184
#> 8          0.5436794          0.7325366
#> 9          0.6542881          0.7973064

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1137170062
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0393   17.0721    0.0000 [ 0.5702; 0.7735 ] 
#>   eta3 ~ eta1      0.4585      0.0753    6.0928    0.0000 [ 0.2567; 0.6458 ] 
#>   eta3 ~ eta2      0.3052      0.0795    3.8393    0.0001 [ 0.1070; 0.5180 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0359   18.4793    0.0000 [ 0.5708; 0.7564 ] 
#>   eta1 =~ y12      0.6493      0.0346   18.7729    0.0000 [ 0.5719; 0.7508 ] 
#>   eta1 =~ y13      0.7613      0.0354   21.4854    0.0000 [ 0.6706; 0.8539 ] 
#>   eta2 =~ y21      0.5165      0.0436   11.8355    0.0000 [ 0.4089; 0.6346 ] 
#>   eta2 =~ y22      0.7554      0.0353   21.3945    0.0000 [ 0.6654; 0.8480 ] 
#>   eta2 =~ y23      0.7997      0.0404   19.8101    0.0000 [ 0.7034; 0.9121 ] 
#>   eta3 =~ y31      0.8223      0.0325   25.2991    0.0000 [ 0.7370; 0.9051 ] 
#>   eta3 =~ y32      0.6581      0.0499   13.1913    0.0000 [ 0.5324; 0.7903 ] 
#>   eta3 =~ y33      0.7474      0.0420   17.7909    0.0000 [ 0.6569; 0.8741 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0181   21.8778    0.0000 [ 0.3455; 0.4390 ] 
#>   eta1 <~ y12      0.3873      0.0163   23.8272    0.0000 [ 0.3490; 0.4331 ] 
#>   eta1 <~ y13      0.4542      0.0197   23.0731    0.0000 [ 0.3995; 0.5013 ] 
#>   eta2 <~ y21      0.3058      0.0247   12.3607    0.0000 [ 0.2421; 0.3701 ] 
#>   eta2 <~ y22      0.4473      0.0216   20.6963    0.0000 [ 0.3877; 0.4995 ] 
#>   eta2 <~ y23      0.4735      0.0187   25.3833    0.0000 [ 0.4257; 0.5222 ] 
#>   eta3 <~ y31      0.4400      0.0198   22.1726    0.0000 [ 0.3821; 0.4847 ] 
#>   eta3 <~ y32      0.3521      0.0204   17.2460    0.0000 [ 0.2969; 0.4025 ] 
#>   eta3 <~ y33      0.3999      0.0250   15.9969    0.0000 [ 0.3396; 0.4689 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0393   17.0721    0.0000 [ 0.5702; 0.7735 ] 
#>   eta3 ~ eta1       0.6634      0.0408   16.2678    0.0000 [ 0.5561; 0.7670 ] 
#>   eta3 ~ eta2       0.3052      0.0795    3.8393    0.0001 [ 0.1070; 0.5180 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0539    3.8040    0.0001 [ 0.0711; 0.3496 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03932347 17.072082 2.395048e-65
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07525397  6.092792 1.109583e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07948023  3.839334 1.233686e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5701591          0.7735179          0.5945785          0.7490985
#> 2          0.2566718          0.6458428          0.3034036          0.5991110
#> 3          0.1069856          0.5180125          0.1563419          0.4686562
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5833067          0.7398739          0.5884255          0.7386954
#> 2          0.2752352          0.5879252          0.2836471          0.5815276
#> 3          0.1651373          0.4972794          0.1654071          0.4492444