Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1570964118
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0511   13.1440    0.0000 [ 0.5827; 0.7791 ] 
#>   eta3 ~ eta1      0.4585      0.0922    4.9723    0.0000 [ 0.3089; 0.6406 ] 
#>   eta3 ~ eta2      0.3052      0.0961    3.1756    0.0015 [ 0.1227; 0.4747 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0327   20.3066    0.0000 [ 0.5860; 0.7099 ] 
#>   eta1 =~ y12      0.6493      0.0401   16.1867    0.0000 [ 0.5809; 0.7113 ] 
#>   eta1 =~ y13      0.7613      0.0346   21.9819    0.0000 [ 0.7019; 0.8363 ] 
#>   eta2 =~ y21      0.5165      0.0545    9.4843    0.0000 [ 0.4203; 0.6234 ] 
#>   eta2 =~ y22      0.7554      0.0385   19.5978    0.0000 [ 0.6836; 0.8358 ] 
#>   eta2 =~ y23      0.7997      0.0423   18.9079    0.0000 [ 0.7086; 0.8651 ] 
#>   eta3 =~ y31      0.8223      0.0335   24.5300    0.0000 [ 0.7540; 0.8728 ] 
#>   eta3 =~ y32      0.6581      0.0354   18.5732    0.0000 [ 0.5888; 0.7081 ] 
#>   eta3 =~ y33      0.7474      0.0285   26.2619    0.0000 [ 0.6960; 0.8067 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0156   25.3937    0.0000 [ 0.3641; 0.4185 ] 
#>   eta1 <~ y12      0.3873      0.0220   17.5863    0.0000 [ 0.3458; 0.4218 ] 
#>   eta1 <~ y13      0.4542      0.0185   24.5117    0.0000 [ 0.4270; 0.4911 ] 
#>   eta2 <~ y21      0.3058      0.0280   10.9055    0.0000 [ 0.2532; 0.3636 ] 
#>   eta2 <~ y22      0.4473      0.0264   16.9195    0.0000 [ 0.4067; 0.5022 ] 
#>   eta2 <~ y23      0.4735      0.0207   22.8208    0.0000 [ 0.4368; 0.5139 ] 
#>   eta3 <~ y31      0.4400      0.0146   30.0512    0.0000 [ 0.4079; 0.4671 ] 
#>   eta3 <~ y32      0.3521      0.0188   18.7734    0.0000 [ 0.3231; 0.3821 ] 
#>   eta3 <~ y33      0.3999      0.0148   26.9861    0.0000 [ 0.3764; 0.4311 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0511   13.1440    0.0000 [ 0.5827; 0.7791 ] 
#>   eta3 ~ eta1       0.6634      0.0496   13.3631    0.0000 [ 0.5756; 0.7550 ] 
#>   eta3 ~ eta2       0.3052      0.0961    3.1756    0.0015 [ 0.1227; 0.4747 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0622    3.2953    0.0010 [ 0.0872; 0.3159 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03265287 20.306634  1.123456e-91
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04011178 16.186715  6.258153e-59
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03463512 21.981903 4.290912e-107
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05445361  9.484308  2.439955e-21
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03854460 19.597756  1.615952e-85
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04229260 18.907886  9.821518e-80
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03352135 24.529962 7.077703e-133
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03543108 18.573211  5.294056e-77
#> 9 eta3 =~ y33  Common factor 0.7474241 0.02846043 26.261870 5.231359e-152
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5860131          0.7098912
#> 2          0.5809028          0.7112626
#> 3          0.7018750          0.8362602
#> 4          0.4203073          0.6234462
#> 5          0.6836359          0.8357959
#> 6          0.7086130          0.8650880
#> 7          0.7539888          0.8728492
#> 8          0.5887626          0.7081028
#> 9          0.6960405          0.8067408

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1570964118
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0511   13.1440    0.0000 [ 0.5409; 0.8050 ] 
#>   eta3 ~ eta1      0.4585      0.0922    4.9723    0.0000 [ 0.2219; 0.6988 ] 
#>   eta3 ~ eta2      0.3052      0.0961    3.1756    0.0015 [ 0.0536; 0.5506 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0327   20.3066    0.0000 [ 0.5879; 0.7568 ] 
#>   eta1 =~ y12      0.6493      0.0401   16.1867    0.0000 [ 0.5463; 0.7537 ] 
#>   eta1 =~ y13      0.7613      0.0346   21.9819    0.0000 [ 0.6673; 0.8464 ] 
#>   eta2 =~ y21      0.5165      0.0545    9.4843    0.0000 [ 0.3728; 0.6544 ] 
#>   eta2 =~ y22      0.7554      0.0385   19.5978    0.0000 [ 0.6483; 0.8477 ] 
#>   eta2 =~ y23      0.7997      0.0423   18.9079    0.0000 [ 0.7059; 0.9246 ] 
#>   eta3 =~ y31      0.8223      0.0335   24.5300    0.0000 [ 0.7360; 0.9093 ] 
#>   eta3 =~ y32      0.6581      0.0354   18.5732    0.0000 [ 0.5782; 0.7614 ] 
#>   eta3 =~ y33      0.7474      0.0285   26.2619    0.0000 [ 0.6679; 0.8150 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0156   25.3937    0.0000 [ 0.3596; 0.4401 ] 
#>   eta1 <~ y12      0.3873      0.0220   17.5863    0.0000 [ 0.3295; 0.4434 ] 
#>   eta1 <~ y13      0.4542      0.0185   24.5117    0.0000 [ 0.4020; 0.4978 ] 
#>   eta2 <~ y21      0.3058      0.0280   10.9055    0.0000 [ 0.2309; 0.3759 ] 
#>   eta2 <~ y22      0.4473      0.0264   16.9195    0.0000 [ 0.3727; 0.5094 ] 
#>   eta2 <~ y23      0.4735      0.0207   22.8208    0.0000 [ 0.4276; 0.5349 ] 
#>   eta3 <~ y31      0.4400      0.0146   30.0512    0.0000 [ 0.4009; 0.4766 ] 
#>   eta3 <~ y32      0.3521      0.0188   18.7734    0.0000 [ 0.3087; 0.4057 ] 
#>   eta3 <~ y33      0.3999      0.0148   26.9861    0.0000 [ 0.3570; 0.4336 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0511   13.1440    0.0000 [ 0.5409; 0.8050 ] 
#>   eta3 ~ eta1       0.6634      0.0496   13.3631    0.0000 [ 0.5370; 0.7937 ] 
#>   eta3 ~ eta2       0.3052      0.0961    3.1756    0.0015 [ 0.0536; 0.5506 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0622    3.2953    0.0010 [ 0.0442; 0.3657 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.05107524 13.144008 1.842049e-39
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09221202  4.972310 6.615981e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09609130  3.175637 1.495078e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54088547          0.8050178          0.5726026          0.7733007
#> 2         0.22193968          0.6988082          0.2792023          0.6415456
#> 3         0.05362246          0.5505524          0.1132940          0.4908808
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5549425          0.7907535          0.5826754          0.7791012
#> 2          0.2960374          0.6652448          0.3088552          0.6405874
#> 3          0.1141573          0.4888108          0.1226502          0.4747058