Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 854889493
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0433   15.5133    0.0000 [ 0.5921; 0.7344 ] 
#>   eta3 ~ eta1      0.4585      0.0957    4.7905    0.0000 [ 0.2913; 0.6378 ] 
#>   eta3 ~ eta2      0.3052      0.1007    3.0303    0.0024 [ 0.1131; 0.4933 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0493   13.4426    0.0000 [ 0.5320; 0.7224 ] 
#>   eta1 =~ y12      0.6493      0.0413   15.7163    0.0000 [ 0.5843; 0.7160 ] 
#>   eta1 =~ y13      0.7613      0.0327   23.2508    0.0000 [ 0.7026; 0.8236 ] 
#>   eta2 =~ y21      0.5165      0.0549    9.4101    0.0000 [ 0.4013; 0.5868 ] 
#>   eta2 =~ y22      0.7554      0.0427   17.6900    0.0000 [ 0.6795; 0.8215 ] 
#>   eta2 =~ y23      0.7997      0.0402   19.8877    0.0000 [ 0.7274; 0.8606 ] 
#>   eta3 =~ y31      0.8223      0.0387   21.2321    0.0000 [ 0.7348; 0.8880 ] 
#>   eta3 =~ y32      0.6581      0.0441   14.9058    0.0000 [ 0.5540; 0.7331 ] 
#>   eta3 =~ y33      0.7474      0.0350   21.3569    0.0000 [ 0.6718; 0.8070 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0263   15.0423    0.0000 [ 0.3445; 0.4412 ] 
#>   eta1 <~ y12      0.3873      0.0176   22.0266    0.0000 [ 0.3587; 0.4228 ] 
#>   eta1 <~ y13      0.4542      0.0250   18.1807    0.0000 [ 0.4090; 0.5238 ] 
#>   eta2 <~ y21      0.3058      0.0262   11.6797    0.0000 [ 0.2585; 0.3471 ] 
#>   eta2 <~ y22      0.4473      0.0267   16.7522    0.0000 [ 0.4064; 0.5045 ] 
#>   eta2 <~ y23      0.4735      0.0238   19.8895    0.0000 [ 0.4348; 0.5143 ] 
#>   eta3 <~ y31      0.4400      0.0180   24.3883    0.0000 [ 0.4025; 0.4693 ] 
#>   eta3 <~ y32      0.3521      0.0208   16.9073    0.0000 [ 0.3136; 0.3959 ] 
#>   eta3 <~ y33      0.3999      0.0215   18.5871    0.0000 [ 0.3640; 0.4340 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0433   15.5133    0.0000 [ 0.5921; 0.7344 ] 
#>   eta3 ~ eta1       0.6634      0.0438   15.1513    0.0000 [ 0.5642; 0.7252 ] 
#>   eta3 ~ eta2       0.3052      0.1007    3.0303    0.0024 [ 0.1131; 0.4933 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0679    3.0180    0.0025 [ 0.0781; 0.3261 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04932611 13.442573  3.403918e-41
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04131248 15.716267  1.170182e-55
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03274488 23.250832 1.395232e-119
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05488294  9.410116  4.955896e-21
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04270131 17.690037  5.004266e-70
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04020888 19.887738  5.196662e-88
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03872810 21.232061 4.829697e-100
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04414856 14.905784  3.022429e-50
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03499683 21.356909 3.363411e-101
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5320007          0.7224473
#> 2          0.5843004          0.7159524
#> 3          0.7026484          0.8235647
#> 4          0.4012621          0.5867670
#> 5          0.6794704          0.8214771
#> 6          0.7273601          0.8606128
#> 7          0.7347619          0.8880478
#> 8          0.5539668          0.7330553
#> 9          0.6718119          0.8070218

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 854889493
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0433   15.5133    0.0000 [ 0.5645; 0.7883 ] 
#>   eta3 ~ eta1      0.4585      0.0957    4.7905    0.0000 [ 0.1997; 0.6946 ] 
#>   eta3 ~ eta2      0.3052      0.1007    3.0303    0.0024 [ 0.0600; 0.5807 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0493   13.4426    0.0000 [ 0.5511; 0.8062 ] 
#>   eta1 =~ y12      0.6493      0.0413   15.7163    0.0000 [ 0.5372; 0.7508 ] 
#>   eta1 =~ y13      0.7613      0.0327   23.2508    0.0000 [ 0.6714; 0.8407 ] 
#>   eta2 =~ y21      0.5165      0.0549    9.4101    0.0000 [ 0.3801; 0.6639 ] 
#>   eta2 =~ y22      0.7554      0.0427   17.6900    0.0000 [ 0.6494; 0.8702 ] 
#>   eta2 =~ y23      0.7997      0.0402   19.8877    0.0000 [ 0.6871; 0.8951 ] 
#>   eta3 =~ y31      0.8223      0.0387   21.2321    0.0000 [ 0.7280; 0.9283 ] 
#>   eta3 =~ y32      0.6581      0.0441   14.9058    0.0000 [ 0.5397; 0.7680 ] 
#>   eta3 =~ y33      0.7474      0.0350   21.3569    0.0000 [ 0.6635; 0.8445 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0263   15.0423    0.0000 [ 0.3360; 0.4720 ] 
#>   eta1 <~ y12      0.3873      0.0176   22.0266    0.0000 [ 0.3379; 0.4289 ] 
#>   eta1 <~ y13      0.4542      0.0250   18.1807    0.0000 [ 0.3848; 0.5140 ] 
#>   eta2 <~ y21      0.3058      0.0262   11.6797    0.0000 [ 0.2421; 0.3775 ] 
#>   eta2 <~ y22      0.4473      0.0267   16.7522    0.0000 [ 0.3810; 0.5191 ] 
#>   eta2 <~ y23      0.4735      0.0238   19.8895    0.0000 [ 0.4071; 0.5302 ] 
#>   eta3 <~ y31      0.4400      0.0180   24.3883    0.0000 [ 0.3942; 0.4875 ] 
#>   eta3 <~ y32      0.3521      0.0208   16.9073    0.0000 [ 0.2942; 0.4019 ] 
#>   eta3 <~ y33      0.3999      0.0215   18.5871    0.0000 [ 0.3455; 0.4568 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0433   15.5133    0.0000 [ 0.5645; 0.7883 ] 
#>   eta3 ~ eta1       0.6634      0.0438   15.1513    0.0000 [ 0.5507; 0.7771 ] 
#>   eta3 ~ eta2       0.3052      0.1007    3.0303    0.0024 [ 0.0600; 0.5807 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0679    3.0180    0.0025 [ 0.0413; 0.3923 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04327475 15.513282 2.820833e-54
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09571097  4.790535 1.663370e-06
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.10070058  3.030282 2.443257e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.56449133          0.7882839          0.5913644          0.7614108
#> 2         0.19966505          0.6946281          0.2591004          0.6351927
#> 3         0.05996878          0.5807353          0.1225027          0.5182014
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5702406          0.7344811          0.5920769          0.7344064
#> 2          0.2666805          0.6716103          0.2912873          0.6378229
#> 3          0.1080954          0.5016344          0.1131019          0.4933301