Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -153053878
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0448   14.9829    0.0000 [ 0.5994; 0.7576 ] 
#>   eta3 ~ eta1      0.4585      0.0863    5.3136    0.0000 [ 0.3027; 0.6015 ] 
#>   eta3 ~ eta2      0.3052      0.0874    3.4909    0.0005 [ 0.1397; 0.4485 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0397   16.7131    0.0000 [ 0.6055; 0.7401 ] 
#>   eta1 =~ y12      0.6493      0.0317   20.4963    0.0000 [ 0.5860; 0.7048 ] 
#>   eta1 =~ y13      0.7613      0.0306   24.8812    0.0000 [ 0.7025; 0.8091 ] 
#>   eta2 =~ y21      0.5165      0.0557    9.2774    0.0000 [ 0.4271; 0.6072 ] 
#>   eta2 =~ y22      0.7554      0.0387   19.5122    0.0000 [ 0.6954; 0.8320 ] 
#>   eta2 =~ y23      0.7997      0.0326   24.5629    0.0000 [ 0.7378; 0.8323 ] 
#>   eta3 =~ y31      0.8223      0.0309   26.6304    0.0000 [ 0.7675; 0.8750 ] 
#>   eta3 =~ y32      0.6581      0.0404   16.3041    0.0000 [ 0.5732; 0.7154 ] 
#>   eta3 =~ y33      0.7474      0.0390   19.1417    0.0000 [ 0.6718; 0.7950 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0170   23.2760    0.0000 [ 0.3651; 0.4310 ] 
#>   eta1 <~ y12      0.3873      0.0195   19.8455    0.0000 [ 0.3494; 0.4186 ] 
#>   eta1 <~ y13      0.4542      0.0183   24.7924    0.0000 [ 0.4186; 0.4885 ] 
#>   eta2 <~ y21      0.3058      0.0283   10.8055    0.0000 [ 0.2572; 0.3462 ] 
#>   eta2 <~ y22      0.4473      0.0230   19.4595    0.0000 [ 0.4121; 0.5033 ] 
#>   eta2 <~ y23      0.4735      0.0215   22.0109    0.0000 [ 0.4295; 0.5068 ] 
#>   eta3 <~ y31      0.4400      0.0190   23.1084    0.0000 [ 0.4088; 0.4766 ] 
#>   eta3 <~ y32      0.3521      0.0185   19.0355    0.0000 [ 0.3154; 0.3865 ] 
#>   eta3 <~ y33      0.3999      0.0187   21.3825    0.0000 [ 0.3645; 0.4279 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0448   14.9829    0.0000 [ 0.5994; 0.7576 ] 
#>   eta3 ~ eta1       0.6634      0.0422   15.7282    0.0000 [ 0.5899; 0.7383 ] 
#>   eta3 ~ eta2       0.3052      0.0874    3.4909    0.0005 [ 0.1397; 0.4485 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0590    3.4742    0.0005 [ 0.0915; 0.3000 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03967358 16.713134  1.051599e-62
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03167775 20.496343  2.320954e-93
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03059918 24.881250 1.187396e-136
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05566784  9.277435  1.736072e-20
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03871365 19.512180  8.651371e-85
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03255575 24.562904 3.148748e-133
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03087744 26.630362 3.022037e-156
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04036206 16.304145  9.221994e-60
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03904700 19.141652  1.136104e-81
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6054630          0.7401254
#> 2          0.5860495          0.7048258
#> 3          0.7025036          0.8090698
#> 4          0.4270589          0.6072069
#> 5          0.6954102          0.8319827
#> 6          0.7377777          0.8322675
#> 7          0.7674913          0.8750493
#> 8          0.5732475          0.7154456
#> 9          0.6718004          0.7949773

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -153053878
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0448   14.9829    0.0000 [ 0.5527; 0.7844 ] 
#>   eta3 ~ eta1      0.4585      0.0863    5.3136    0.0000 [ 0.2488; 0.6950 ] 
#>   eta3 ~ eta2      0.3052      0.0874    3.4909    0.0005 [ 0.0715; 0.5236 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0397   16.7131    0.0000 [ 0.5541; 0.7592 ] 
#>   eta1 =~ y12      0.6493      0.0317   20.4963    0.0000 [ 0.5675; 0.7314 ] 
#>   eta1 =~ y13      0.7613      0.0306   24.8812    0.0000 [ 0.6853; 0.8435 ] 
#>   eta2 =~ y21      0.5165      0.0557    9.2774    0.0000 [ 0.3699; 0.6577 ] 
#>   eta2 =~ y22      0.7554      0.0387   19.5122    0.0000 [ 0.6493; 0.8495 ] 
#>   eta2 =~ y23      0.7997      0.0326   24.5629    0.0000 [ 0.7267; 0.8951 ] 
#>   eta3 =~ y31      0.8223      0.0309   26.6304    0.0000 [ 0.7348; 0.8945 ] 
#>   eta3 =~ y32      0.6581      0.0404   16.3041    0.0000 [ 0.5575; 0.7663 ] 
#>   eta3 =~ y33      0.7474      0.0390   19.1417    0.0000 [ 0.6518; 0.8537 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0170   23.2760    0.0000 [ 0.3490; 0.4369 ] 
#>   eta1 <~ y12      0.3873      0.0195   19.8455    0.0000 [ 0.3378; 0.4388 ] 
#>   eta1 <~ y13      0.4542      0.0183   24.7924    0.0000 [ 0.4097; 0.5044 ] 
#>   eta2 <~ y21      0.3058      0.0283   10.8055    0.0000 [ 0.2311; 0.3774 ] 
#>   eta2 <~ y22      0.4473      0.0230   19.4595    0.0000 [ 0.3837; 0.5025 ] 
#>   eta2 <~ y23      0.4735      0.0215   22.0109    0.0000 [ 0.4238; 0.5350 ] 
#>   eta3 <~ y31      0.4400      0.0190   23.1084    0.0000 [ 0.3864; 0.4848 ] 
#>   eta3 <~ y32      0.3521      0.0185   19.0355    0.0000 [ 0.3063; 0.4020 ] 
#>   eta3 <~ y33      0.3999      0.0187   21.3825    0.0000 [ 0.3543; 0.4511 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0448   14.9829    0.0000 [ 0.5527; 0.7844 ] 
#>   eta3 ~ eta1       0.6634      0.0422   15.7282    0.0000 [ 0.5622; 0.7803 ] 
#>   eta3 ~ eta2       0.3052      0.0874    3.4909    0.0005 [ 0.0715; 0.5236 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0590    3.4742    0.0005 [ 0.0469; 0.3519 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04480657 14.982925 9.494501e-51
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08628997  5.313558 1.075051e-07
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08741251  3.490932 4.813385e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55269091          0.7844052          0.5805153          0.7565808
#> 2         0.24877036          0.6950134          0.3023554          0.6414283
#> 3         0.07152275          0.5235709          0.1258049          0.4692887
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.58159279          0.7795424          0.5994196          0.7576094
#> 2         0.29195781          0.6314567          0.3026696          0.6014996
#> 3         0.09808186          0.4493287          0.1396531          0.4484824