Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 368415455
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0497   13.5053    0.0000 [ 0.5478; 0.7558 ] 
#>   eta3 ~ eta1      0.4585      0.0781    5.8688    0.0000 [ 0.2870; 0.5723 ] 
#>   eta3 ~ eta2      0.3052      0.0849    3.5963    0.0003 [ 0.1662; 0.4776 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0425   15.6168    0.0000 [ 0.5597; 0.7293 ] 
#>   eta1 =~ y12      0.6493      0.0404   16.0723    0.0000 [ 0.5601; 0.6964 ] 
#>   eta1 =~ y13      0.7613      0.0345   22.0856    0.0000 [ 0.7095; 0.8440 ] 
#>   eta2 =~ y21      0.5165      0.0566    9.1175    0.0000 [ 0.3858; 0.5830 ] 
#>   eta2 =~ y22      0.7554      0.0315   24.0059    0.0000 [ 0.7104; 0.8127 ] 
#>   eta2 =~ y23      0.7997      0.0352   22.7188    0.0000 [ 0.7371; 0.8513 ] 
#>   eta3 =~ y31      0.8223      0.0342   24.0184    0.0000 [ 0.7592; 0.8813 ] 
#>   eta3 =~ y32      0.6581      0.0382   17.2233    0.0000 [ 0.5997; 0.7311 ] 
#>   eta3 =~ y33      0.7474      0.0380   19.6771    0.0000 [ 0.6844; 0.8091 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0216   18.2868    0.0000 [ 0.3534; 0.4216 ] 
#>   eta1 <~ y12      0.3873      0.0226   17.1736    0.0000 [ 0.3407; 0.4229 ] 
#>   eta1 <~ y13      0.4542      0.0204   22.2606    0.0000 [ 0.4358; 0.5001 ] 
#>   eta2 <~ y21      0.3058      0.0305   10.0151    0.0000 [ 0.2331; 0.3468 ] 
#>   eta2 <~ y22      0.4473      0.0197   22.7595    0.0000 [ 0.4286; 0.4943 ] 
#>   eta2 <~ y23      0.4735      0.0217   21.7804    0.0000 [ 0.4476; 0.5151 ] 
#>   eta3 <~ y31      0.4400      0.0174   25.2725    0.0000 [ 0.4110; 0.4687 ] 
#>   eta3 <~ y32      0.3521      0.0182   19.3458    0.0000 [ 0.3256; 0.3877 ] 
#>   eta3 <~ y33      0.3999      0.0176   22.7227    0.0000 [ 0.3728; 0.4275 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0497   13.5053    0.0000 [ 0.5478; 0.7558 ] 
#>   eta3 ~ eta1       0.6634      0.0398   16.6635    0.0000 [ 0.5887; 0.7229 ] 
#>   eta3 ~ eta2       0.3052      0.0849    3.5963    0.0003 [ 0.1662; 0.4776 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0526    3.8916    0.0001 [ 0.1074; 0.2977 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04245884 15.616767  5.597256e-55
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04039742 16.072264  3.992343e-58
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03447254 22.085575 4.349782e-108
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05664448  9.117478  7.689296e-20
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03146680 24.005866 2.414999e-127
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03519826 22.718845 2.917793e-114
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03423529 24.018415 1.785734e-127
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03820805 17.223307  1.775475e-66
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03798447 19.677098  3.388603e-86
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5596703          0.7292647
#> 2          0.5601309          0.6964417
#> 3          0.7094528          0.8440221
#> 4          0.3857844          0.5829564
#> 5          0.7103548          0.8127051
#> 6          0.7370581          0.8512545
#> 7          0.7592117          0.8812581
#> 8          0.5997203          0.7310924
#> 9          0.6844117          0.8091174

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 368415455
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0497   13.5053    0.0000 [ 0.5527; 0.8098 ] 
#>   eta3 ~ eta1      0.4585      0.0781    5.8688    0.0000 [ 0.2542; 0.6582 ] 
#>   eta3 ~ eta2      0.3052      0.0849    3.5963    0.0003 [ 0.0875; 0.5263 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0425   15.6168    0.0000 [ 0.5553; 0.7749 ] 
#>   eta1 =~ y12      0.6493      0.0404   16.0723    0.0000 [ 0.5552; 0.7641 ] 
#>   eta1 =~ y13      0.7613      0.0345   22.0856    0.0000 [ 0.6633; 0.8416 ] 
#>   eta2 =~ y21      0.5165      0.0566    9.1175    0.0000 [ 0.3988; 0.6917 ] 
#>   eta2 =~ y22      0.7554      0.0315   24.0059    0.0000 [ 0.6699; 0.8326 ] 
#>   eta2 =~ y23      0.7997      0.0352   22.7188    0.0000 [ 0.7183; 0.9004 ] 
#>   eta3 =~ y31      0.8223      0.0342   24.0184    0.0000 [ 0.7376; 0.9146 ] 
#>   eta3 =~ y32      0.6581      0.0382   17.2233    0.0000 [ 0.5558; 0.7534 ] 
#>   eta3 =~ y33      0.7474      0.0380   19.6771    0.0000 [ 0.6438; 0.8403 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0216   18.2868    0.0000 [ 0.3404; 0.4522 ] 
#>   eta1 <~ y12      0.3873      0.0226   17.1736    0.0000 [ 0.3346; 0.4513 ] 
#>   eta1 <~ y13      0.4542      0.0204   22.2606    0.0000 [ 0.3953; 0.5008 ] 
#>   eta2 <~ y21      0.3058      0.0305   10.0151    0.0000 [ 0.2383; 0.3962 ] 
#>   eta2 <~ y22      0.4473      0.0197   22.7595    0.0000 [ 0.3845; 0.4862 ] 
#>   eta2 <~ y23      0.4735      0.0217   21.7804    0.0000 [ 0.4131; 0.5256 ] 
#>   eta3 <~ y31      0.4400      0.0174   25.2725    0.0000 [ 0.3985; 0.4886 ] 
#>   eta3 <~ y32      0.3521      0.0182   19.3458    0.0000 [ 0.3045; 0.3986 ] 
#>   eta3 <~ y33      0.3999      0.0176   22.7227    0.0000 [ 0.3530; 0.4441 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0497   13.5053    0.0000 [ 0.5527; 0.8098 ] 
#>   eta3 ~ eta1       0.6634      0.0398   16.6635    0.0000 [ 0.5634; 0.7692 ] 
#>   eta3 ~ eta2       0.3052      0.0849    3.5963    0.0003 [ 0.0875; 0.5263 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0526    3.8916    0.0001 [ 0.0740; 0.3462 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04970900 13.505271 1.455781e-41
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07812625  5.868793 4.389796e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08485255  3.596252 3.228353e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55274146          0.8098083          0.5836102          0.7789396
#> 2         0.25416370          0.6581885          0.3026792          0.6096731
#> 3         0.08749231          0.5263018          0.1401847          0.4736093
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5364329          0.7690082          0.5478083          0.7558434
#> 2          0.2525687          0.6107041          0.2869747          0.5722975
#> 3          0.1349602          0.5290521          0.1661914          0.4775521