Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 521268663
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0356   18.8784    0.0000 [ 0.6256; 0.7372 ] 
#>   eta3 ~ eta1      0.4585      0.0945    4.8530    0.0000 [ 0.2696; 0.6254 ] 
#>   eta3 ~ eta2      0.3052      0.0917    3.3291    0.0009 [ 0.1707; 0.4652 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0395   16.7988    0.0000 [ 0.5785; 0.7271 ] 
#>   eta1 =~ y12      0.6493      0.0309   21.0380    0.0000 [ 0.5850; 0.6977 ] 
#>   eta1 =~ y13      0.7613      0.0255   29.9028    0.0000 [ 0.7225; 0.8095 ] 
#>   eta2 =~ y21      0.5165      0.0587    8.8016    0.0000 [ 0.4054; 0.6180 ] 
#>   eta2 =~ y22      0.7554      0.0346   21.8182    0.0000 [ 0.7052; 0.8129 ] 
#>   eta2 =~ y23      0.7997      0.0353   22.6556    0.0000 [ 0.7459; 0.8609 ] 
#>   eta3 =~ y31      0.8223      0.0313   26.3110    0.0000 [ 0.7522; 0.8705 ] 
#>   eta3 =~ y32      0.6581      0.0403   16.3254    0.0000 [ 0.5916; 0.7270 ] 
#>   eta3 =~ y33      0.7474      0.0342   21.8474    0.0000 [ 0.6825; 0.7942 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0179   22.1096    0.0000 [ 0.3647; 0.4294 ] 
#>   eta1 <~ y12      0.3873      0.0149   26.0797    0.0000 [ 0.3553; 0.4043 ] 
#>   eta1 <~ y13      0.4542      0.0187   24.2431    0.0000 [ 0.4218; 0.4823 ] 
#>   eta2 <~ y21      0.3058      0.0321    9.5389    0.0000 [ 0.2447; 0.3679 ] 
#>   eta2 <~ y22      0.4473      0.0201   22.2257    0.0000 [ 0.4158; 0.4826 ] 
#>   eta2 <~ y23      0.4735      0.0196   24.1272    0.0000 [ 0.4403; 0.5180 ] 
#>   eta3 <~ y31      0.4400      0.0179   24.5882    0.0000 [ 0.4101; 0.4699 ] 
#>   eta3 <~ y32      0.3521      0.0171   20.6153    0.0000 [ 0.3193; 0.3824 ] 
#>   eta3 <~ y33      0.3999      0.0181   22.0390    0.0000 [ 0.3667; 0.4258 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0356   18.8784    0.0000 [ 0.6256; 0.7372 ] 
#>   eta3 ~ eta1       0.6634      0.0439   15.1261    0.0000 [ 0.5951; 0.7425 ] 
#>   eta3 ~ eta2       0.3052      0.0917    3.3291    0.0009 [ 0.1707; 0.4652 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0648    3.1591    0.0016 [ 0.1048; 0.3173 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03947117 16.798842  2.488563e-63
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03086221 21.037959  2.947972e-98
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02546065 29.902841 1.807360e-196
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05867736  8.801603  1.348751e-18
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03462184 21.818239 1.557426e-105
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03529658 22.655555 1.229847e-113
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03125217 26.311046 1.433525e-152
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04030961 16.325362  6.515335e-60
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03421106 21.847441 8.221276e-106
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5785027          0.7271440
#> 2          0.5850149          0.6976541
#> 3          0.7225220          0.8094669
#> 4          0.4054101          0.6179504
#> 5          0.7052058          0.8128656
#> 6          0.7458652          0.8608506
#> 7          0.7522297          0.8705031
#> 8          0.5915854          0.7269502
#> 9          0.6824931          0.7942131

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 521268663
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0356   18.8784    0.0000 [ 0.5714; 0.7553 ] 
#>   eta3 ~ eta1      0.4585      0.0945    4.8530    0.0000 [ 0.2249; 0.7134 ] 
#>   eta3 ~ eta2      0.3052      0.0917    3.3291    0.0009 [ 0.0546; 0.5286 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0395   16.7988    0.0000 [ 0.5570; 0.7611 ] 
#>   eta1 =~ y12      0.6493      0.0309   21.0380    0.0000 [ 0.5751; 0.7347 ] 
#>   eta1 =~ y13      0.7613      0.0255   29.9028    0.0000 [ 0.6895; 0.8212 ] 
#>   eta2 =~ y21      0.5165      0.0587    8.8016    0.0000 [ 0.3658; 0.6693 ] 
#>   eta2 =~ y22      0.7554      0.0346   21.8182    0.0000 [ 0.6694; 0.8484 ] 
#>   eta2 =~ y23      0.7997      0.0353   22.6556    0.0000 [ 0.7146; 0.8971 ] 
#>   eta3 =~ y31      0.8223      0.0313   26.3110    0.0000 [ 0.7443; 0.9059 ] 
#>   eta3 =~ y32      0.6581      0.0403   16.3254    0.0000 [ 0.5583; 0.7667 ] 
#>   eta3 =~ y33      0.7474      0.0342   21.8474    0.0000 [ 0.6608; 0.8377 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0179   22.1096    0.0000 [ 0.3486; 0.4411 ] 
#>   eta1 <~ y12      0.3873      0.0149   26.0797    0.0000 [ 0.3537; 0.4305 ] 
#>   eta1 <~ y13      0.4542      0.0187   24.2431    0.0000 [ 0.4036; 0.5005 ] 
#>   eta2 <~ y21      0.3058      0.0321    9.5389    0.0000 [ 0.2217; 0.3875 ] 
#>   eta2 <~ y22      0.4473      0.0201   22.2257    0.0000 [ 0.3943; 0.4983 ] 
#>   eta2 <~ y23      0.4735      0.0196   24.1272    0.0000 [ 0.4232; 0.5247 ] 
#>   eta3 <~ y31      0.4400      0.0179   24.5882    0.0000 [ 0.3926; 0.4851 ] 
#>   eta3 <~ y32      0.3521      0.0171   20.6153    0.0000 [ 0.3085; 0.3968 ] 
#>   eta3 <~ y33      0.3999      0.0181   22.0390    0.0000 [ 0.3516; 0.4455 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0356   18.8784    0.0000 [ 0.5714; 0.7553 ] 
#>   eta3 ~ eta1       0.6634      0.0439   15.1261    0.0000 [ 0.5488; 0.7756 ] 
#>   eta3 ~ eta2       0.3052      0.0917    3.3291    0.0009 [ 0.0546; 0.5286 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0648    3.1591    0.0016 [ 0.0253; 0.3607 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03556085 18.878440 1.715792e-79
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09447987  4.852957 1.216341e-06
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09166259  3.329069 8.713669e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.57142941          0.7553301          0.5935123          0.7332472
#> 2         0.22485343          0.7134500          0.2835243          0.6547791
#> 3         0.05461537          0.5286425          0.1115368          0.4717211
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5872458          0.7402569          0.6256438          0.7371907
#> 2          0.2481953          0.6387162          0.2695917          0.6253797
#> 3          0.1355689          0.5063331          0.1706602          0.4651713