Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -278488666
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0442   15.1911    0.0000 [ 0.6093; 0.7751 ] 
#>   eta3 ~ eta1      0.4585      0.0938    4.8871    0.0000 [ 0.3316; 0.6481 ] 
#>   eta3 ~ eta2      0.3052      0.0999    3.0538    0.0023 [ 0.1203; 0.4518 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0341   19.4534    0.0000 [ 0.6094; 0.7232 ] 
#>   eta1 =~ y12      0.6493      0.0355   18.2683    0.0000 [ 0.5838; 0.7274 ] 
#>   eta1 =~ y13      0.7613      0.0343   22.1784    0.0000 [ 0.6944; 0.8075 ] 
#>   eta2 =~ y21      0.5165      0.0561    9.2023    0.0000 [ 0.4174; 0.6013 ] 
#>   eta2 =~ y22      0.7554      0.0338   22.3305    0.0000 [ 0.6907; 0.8097 ] 
#>   eta2 =~ y23      0.7997      0.0320   25.0115    0.0000 [ 0.7340; 0.8443 ] 
#>   eta3 =~ y31      0.8223      0.0373   22.0239    0.0000 [ 0.7585; 0.8864 ] 
#>   eta3 =~ y32      0.6581      0.0452   14.5436    0.0000 [ 0.5765; 0.7282 ] 
#>   eta3 =~ y33      0.7474      0.0329   22.7498    0.0000 [ 0.7106; 0.8311 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0208   19.0614    0.0000 [ 0.3561; 0.4286 ] 
#>   eta1 <~ y12      0.3873      0.0181   21.3550    0.0000 [ 0.3580; 0.4220 ] 
#>   eta1 <~ y13      0.4542      0.0182   25.0101    0.0000 [ 0.4188; 0.4855 ] 
#>   eta2 <~ y21      0.3058      0.0255   11.9999    0.0000 [ 0.2671; 0.3461 ] 
#>   eta2 <~ y22      0.4473      0.0247   18.1225    0.0000 [ 0.4063; 0.4935 ] 
#>   eta2 <~ y23      0.4735      0.0197   24.0312    0.0000 [ 0.4345; 0.5103 ] 
#>   eta3 <~ y31      0.4400      0.0216   20.3990    0.0000 [ 0.3940; 0.4804 ] 
#>   eta3 <~ y32      0.3521      0.0192   18.3410    0.0000 [ 0.3142; 0.3826 ] 
#>   eta3 <~ y33      0.3999      0.0167   23.9406    0.0000 [ 0.3726; 0.4323 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0442   15.1911    0.0000 [ 0.6093; 0.7751 ] 
#>   eta3 ~ eta1       0.6634      0.0454   14.6256    0.0000 [ 0.5844; 0.7370 ] 
#>   eta3 ~ eta2       0.3052      0.0999    3.0538    0.0023 [ 0.1203; 0.4518 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0640    3.1986    0.0014 [ 0.0877; 0.2979 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03408506 19.453389  2.727913e-84
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03554124 18.268297  1.479983e-74
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03432828 22.178384 5.553832e-109
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05612232  9.202307  3.503469e-20
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03382769 22.330454 1.870170e-110
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03197188 25.011473 4.586589e-138
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03733560 22.023950 1.697990e-107
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04524804 14.543589  6.413929e-48
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03285403 22.749848 1.439988e-114
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6094253          0.7232117
#> 2          0.5837825          0.7273699
#> 3          0.6944061          0.8075424
#> 4          0.4174093          0.6012809
#> 5          0.6907242          0.8097062
#> 6          0.7340395          0.8443425
#> 7          0.7584636          0.8864041
#> 8          0.5764562          0.7281963
#> 9          0.7106018          0.8310987

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -278488666
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0442   15.1911    0.0000 [ 0.5464; 0.7750 ] 
#>   eta3 ~ eta1      0.4585      0.0938    4.8871    0.0000 [ 0.2111; 0.6963 ] 
#>   eta3 ~ eta2      0.3052      0.0999    3.0538    0.0023 [ 0.0561; 0.5728 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0341   19.4534    0.0000 [ 0.5683; 0.7445 ] 
#>   eta1 =~ y12      0.6493      0.0355   18.2683    0.0000 [ 0.5551; 0.7389 ] 
#>   eta1 =~ y13      0.7613      0.0343   22.1784    0.0000 [ 0.6784; 0.8560 ] 
#>   eta2 =~ y21      0.5165      0.0561    9.2023    0.0000 [ 0.3695; 0.6597 ] 
#>   eta2 =~ y22      0.7554      0.0338   22.3305    0.0000 [ 0.6708; 0.8457 ] 
#>   eta2 =~ y23      0.7997      0.0320   25.0115    0.0000 [ 0.7272; 0.8925 ] 
#>   eta3 =~ y31      0.8223      0.0373   22.0239    0.0000 [ 0.7252; 0.9183 ] 
#>   eta3 =~ y32      0.6581      0.0452   14.5436    0.0000 [ 0.5472; 0.7812 ] 
#>   eta3 =~ y33      0.7474      0.0329   22.7498    0.0000 [ 0.6545; 0.8244 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0208   19.0614    0.0000 [ 0.3388; 0.4461 ] 
#>   eta1 <~ y12      0.3873      0.0181   21.3550    0.0000 [ 0.3401; 0.4339 ] 
#>   eta1 <~ y13      0.4542      0.0182   25.0101    0.0000 [ 0.4119; 0.5058 ] 
#>   eta2 <~ y21      0.3058      0.0255   11.9999    0.0000 [ 0.2371; 0.3689 ] 
#>   eta2 <~ y22      0.4473      0.0247   18.1225    0.0000 [ 0.3814; 0.5090 ] 
#>   eta2 <~ y23      0.4735      0.0197   24.0312    0.0000 [ 0.4248; 0.5267 ] 
#>   eta3 <~ y31      0.4400      0.0216   20.3990    0.0000 [ 0.3848; 0.4963 ] 
#>   eta3 <~ y32      0.3521      0.0192   18.3410    0.0000 [ 0.3068; 0.4061 ] 
#>   eta3 <~ y33      0.3999      0.0167   23.9406    0.0000 [ 0.3534; 0.4398 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0442   15.1911    0.0000 [ 0.5464; 0.7750 ] 
#>   eta3 ~ eta1       0.6634      0.0454   14.6256    0.0000 [ 0.5469; 0.7814 ] 
#>   eta3 ~ eta2       0.3052      0.0999    3.0538    0.0023 [ 0.0561; 0.5728 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0640    3.1986    0.0014 [ 0.0449; 0.3761 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04419242 15.191143 4.048004e-52
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09382054  4.887062 1.023520e-06
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09992484  3.053806 2.259579e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54644506          0.7749833          0.5738880          0.7475403
#> 2         0.21107672          0.6962636          0.2693382          0.6380021
#> 3         0.05607337          0.5728282          0.1181255          0.5107761
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.60160464          0.7802958          0.6093327          0.7751440
#> 2         0.28355466          0.6803752          0.3315829          0.6481086
#> 3         0.03174041          0.4524056          0.1202946          0.4518432