Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1289374850
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0411   16.3312    0.0000 [ 0.5754; 0.7558 ] 
#>   eta3 ~ eta1      0.4585      0.0767    5.9799    0.0000 [ 0.3100; 0.5722 ] 
#>   eta3 ~ eta2      0.3052      0.0844    3.6146    0.0003 [ 0.1584; 0.4555 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0372   17.8260    0.0000 [ 0.6001; 0.7225 ] 
#>   eta1 =~ y12      0.6493      0.0319   20.3241    0.0000 [ 0.5915; 0.7081 ] 
#>   eta1 =~ y13      0.7613      0.0258   29.5578    0.0000 [ 0.7076; 0.7994 ] 
#>   eta2 =~ y21      0.5165      0.0459   11.2640    0.0000 [ 0.4116; 0.5894 ] 
#>   eta2 =~ y22      0.7554      0.0306   24.6946    0.0000 [ 0.7137; 0.8111 ] 
#>   eta2 =~ y23      0.7997      0.0374   21.3618    0.0000 [ 0.7225; 0.8541 ] 
#>   eta3 =~ y31      0.8223      0.0282   29.1248    0.0000 [ 0.7758; 0.8742 ] 
#>   eta3 =~ y32      0.6581      0.0307   21.4384    0.0000 [ 0.6079; 0.7104 ] 
#>   eta3 =~ y33      0.7474      0.0364   20.5249    0.0000 [ 0.6675; 0.8105 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0208   18.9850    0.0000 [ 0.3647; 0.4332 ] 
#>   eta1 <~ y12      0.3873      0.0168   23.0544    0.0000 [ 0.3562; 0.4129 ] 
#>   eta1 <~ y13      0.4542      0.0132   34.3923    0.0000 [ 0.4298; 0.4715 ] 
#>   eta2 <~ y21      0.3058      0.0234   13.0435    0.0000 [ 0.2565; 0.3388 ] 
#>   eta2 <~ y22      0.4473      0.0211   21.2214    0.0000 [ 0.4032; 0.4794 ] 
#>   eta2 <~ y23      0.4735      0.0184   25.6846    0.0000 [ 0.4364; 0.5065 ] 
#>   eta3 <~ y31      0.4400      0.0167   26.3282    0.0000 [ 0.4127; 0.4782 ] 
#>   eta3 <~ y32      0.3521      0.0143   24.7010    0.0000 [ 0.3322; 0.3794 ] 
#>   eta3 <~ y33      0.3999      0.0172   23.2503    0.0000 [ 0.3634; 0.4323 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0411   16.3312    0.0000 [ 0.5754; 0.7558 ] 
#>   eta3 ~ eta1       0.6634      0.0369   17.9611    0.0000 [ 0.6125; 0.7139 ] 
#>   eta3 ~ eta2       0.3052      0.0844    3.6146    0.0003 [ 0.1584; 0.4555 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0567    3.6161    0.0003 [ 0.1065; 0.3009 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03719679 17.82600  4.440986e-71
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03194628 20.32405  7.879194e-92
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02575784 29.55783 5.208832e-192
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04585001 11.26401  1.975702e-29
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03058913 24.69464 1.221113e-134
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03743437 21.36175 3.032138e-101
#> 7 eta3 =~ y31  Common factor 0.8222773 0.02823293 29.12476 1.744431e-186
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03069577 21.43843 5.856152e-102
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03641548 20.52490  1.290318e-93
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6001040          0.7224916
#> 2          0.5915127          0.7081204
#> 3          0.7076096          0.7994494
#> 4          0.4115771          0.5894162
#> 5          0.7137459          0.8110646
#> 6          0.7224932          0.8540626
#> 7          0.7758486          0.8742226
#> 8          0.6078905          0.7103971
#> 9          0.6674619          0.8104785

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1289374850
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0411   16.3312    0.0000 [ 0.5743; 0.7869 ] 
#>   eta3 ~ eta1      0.4585      0.0767    5.9799    0.0000 [ 0.2681; 0.6647 ] 
#>   eta3 ~ eta2      0.3052      0.0844    3.6146    0.0003 [ 0.0762; 0.5128 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0372   17.8260    0.0000 [ 0.5653; 0.7577 ] 
#>   eta1 =~ y12      0.6493      0.0319   20.3241    0.0000 [ 0.5624; 0.7276 ] 
#>   eta1 =~ y13      0.7613      0.0258   29.5578    0.0000 [ 0.6975; 0.8307 ] 
#>   eta2 =~ y21      0.5165      0.0459   11.2640    0.0000 [ 0.3945; 0.6316 ] 
#>   eta2 =~ y22      0.7554      0.0306   24.6946    0.0000 [ 0.6770; 0.8352 ] 
#>   eta2 =~ y23      0.7997      0.0374   21.3618    0.0000 [ 0.6980; 0.8916 ] 
#>   eta3 =~ y31      0.8223      0.0282   29.1248    0.0000 [ 0.7421; 0.8881 ] 
#>   eta3 =~ y32      0.6581      0.0307   21.4384    0.0000 [ 0.5795; 0.7382 ] 
#>   eta3 =~ y33      0.7474      0.0364   20.5249    0.0000 [ 0.6567; 0.8450 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0208   18.9850    0.0000 [ 0.3417; 0.4495 ] 
#>   eta1 <~ y12      0.3873      0.0168   23.0544    0.0000 [ 0.3423; 0.4292 ] 
#>   eta1 <~ y13      0.4542      0.0132   34.3923    0.0000 [ 0.4228; 0.4911 ] 
#>   eta2 <~ y21      0.3058      0.0234   13.0435    0.0000 [ 0.2452; 0.3665 ] 
#>   eta2 <~ y22      0.4473      0.0211   21.2214    0.0000 [ 0.3956; 0.5046 ] 
#>   eta2 <~ y23      0.4735      0.0184   25.6846    0.0000 [ 0.4259; 0.5212 ] 
#>   eta3 <~ y31      0.4400      0.0167   26.3282    0.0000 [ 0.3939; 0.4803 ] 
#>   eta3 <~ y32      0.3521      0.0143   24.7010    0.0000 [ 0.3166; 0.3903 ] 
#>   eta3 <~ y33      0.3999      0.0172   23.2503    0.0000 [ 0.3583; 0.4473 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0411   16.3312    0.0000 [ 0.5743; 0.7869 ] 
#>   eta3 ~ eta1       0.6634      0.0369   17.9611    0.0000 [ 0.5717; 0.7627 ] 
#>   eta3 ~ eta2       0.3052      0.0844    3.6146    0.0003 [ 0.0762; 0.5128 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0567    3.6161    0.0003 [ 0.0543; 0.3473 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04110739 16.331210 5.919826e-60
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07667473  5.979894 2.232827e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08442144  3.614617 3.007925e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5742793          0.7868635          0.5998065          0.7613363
#> 2          0.2681452          0.6646636          0.3157593          0.6170495
#> 3          0.0761896          0.5127696          0.1286143          0.4603449
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5650570          0.7557980          0.5754140          0.7557946
#> 2          0.2666879          0.5729657          0.3100275          0.5721893
#> 3          0.1321731          0.4587227          0.1583933          0.4554527