Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1125157913
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0442   15.1948    0.0000 [ 0.5991; 0.7343 ] 
#>   eta3 ~ eta1      0.4585      0.0855    5.3600    0.0000 [ 0.3003; 0.6174 ] 
#>   eta3 ~ eta2      0.3052      0.0901    3.3886    0.0007 [ 0.1092; 0.4901 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0459   14.4406    0.0000 [ 0.5675; 0.7367 ] 
#>   eta1 =~ y12      0.6493      0.0407   15.9455    0.0000 [ 0.5667; 0.7237 ] 
#>   eta1 =~ y13      0.7613      0.0348   21.8509    0.0000 [ 0.6957; 0.8299 ] 
#>   eta2 =~ y21      0.5165      0.0567    9.1051    0.0000 [ 0.4042; 0.5873 ] 
#>   eta2 =~ y22      0.7554      0.0369   20.4922    0.0000 [ 0.6776; 0.8299 ] 
#>   eta2 =~ y23      0.7997      0.0456   17.5508    0.0000 [ 0.7112; 0.8904 ] 
#>   eta3 =~ y31      0.8223      0.0311   26.4697    0.0000 [ 0.7592; 0.8764 ] 
#>   eta3 =~ y32      0.6581      0.0388   16.9443    0.0000 [ 0.5962; 0.7333 ] 
#>   eta3 =~ y33      0.7474      0.0431   17.3431    0.0000 [ 0.6780; 0.8353 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0245   16.1234    0.0000 [ 0.3503; 0.4369 ] 
#>   eta1 <~ y12      0.3873      0.0225   17.2513    0.0000 [ 0.3387; 0.4283 ] 
#>   eta1 <~ y13      0.4542      0.0196   23.1210    0.0000 [ 0.4234; 0.4908 ] 
#>   eta2 <~ y21      0.3058      0.0271   11.2704    0.0000 [ 0.2547; 0.3421 ] 
#>   eta2 <~ y22      0.4473      0.0256   17.4852    0.0000 [ 0.3929; 0.4965 ] 
#>   eta2 <~ y23      0.4735      0.0259   18.2677    0.0000 [ 0.4393; 0.5375 ] 
#>   eta3 <~ y31      0.4400      0.0183   24.0737    0.0000 [ 0.4058; 0.4727 ] 
#>   eta3 <~ y32      0.3521      0.0190   18.5415    0.0000 [ 0.3228; 0.3845 ] 
#>   eta3 <~ y33      0.3999      0.0196   20.3649    0.0000 [ 0.3656; 0.4356 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0442   15.1948    0.0000 [ 0.5991; 0.7343 ] 
#>   eta3 ~ eta1       0.6634      0.0418   15.8659    0.0000 [ 0.5670; 0.7299 ] 
#>   eta3 ~ eta2       0.3052      0.0901    3.3886    0.0007 [ 0.1092; 0.4901 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0607    3.3755    0.0007 [ 0.0739; 0.3322 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04591711 14.440585  2.873930e-47
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04071852 15.945521  3.060783e-57
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03484271 21.850935 7.615754e-106
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05672153  9.105094  8.619188e-20
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03686223 20.492185  2.527891e-93
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04556271 17.550836  5.861141e-69
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03106483 26.469722 2.163640e-154
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03883709 16.944343  2.118896e-64
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04309638 17.343083  2.224789e-67
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5675187          0.7367379
#> 2          0.5667227          0.7237273
#> 3          0.6956899          0.8298527
#> 4          0.4042491          0.5872810
#> 5          0.6775712          0.8299476
#> 6          0.7112171          0.8904196
#> 7          0.7591838          0.8763526
#> 8          0.5961911          0.7332556
#> 9          0.6780120          0.8352685

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -1125157913
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0442   15.1948    0.0000 [ 0.5598; 0.7883 ] 
#>   eta3 ~ eta1      0.4585      0.0855    5.3600    0.0000 [ 0.2486; 0.6910 ] 
#>   eta3 ~ eta2      0.3052      0.0901    3.3886    0.0007 [ 0.0641; 0.5298 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0459   14.4406    0.0000 [ 0.5421; 0.7796 ] 
#>   eta1 =~ y12      0.6493      0.0407   15.9455    0.0000 [ 0.5488; 0.7594 ] 
#>   eta1 =~ y13      0.7613      0.0348   21.8509    0.0000 [ 0.6709; 0.8511 ] 
#>   eta2 =~ y21      0.5165      0.0567    9.1051    0.0000 [ 0.3799; 0.6732 ] 
#>   eta2 =~ y22      0.7554      0.0369   20.4922    0.0000 [ 0.6611; 0.8517 ] 
#>   eta2 =~ y23      0.7997      0.0456   17.5508    0.0000 [ 0.6761; 0.9117 ] 
#>   eta3 =~ y31      0.8223      0.0311   26.4697    0.0000 [ 0.7379; 0.8986 ] 
#>   eta3 =~ y32      0.6581      0.0388   16.9443    0.0000 [ 0.5482; 0.7490 ] 
#>   eta3 =~ y33      0.7474      0.0431   17.3431    0.0000 [ 0.6396; 0.8625 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0245   16.1234    0.0000 [ 0.3305; 0.4574 ] 
#>   eta1 <~ y12      0.3873      0.0225   17.2513    0.0000 [ 0.3318; 0.4479 ] 
#>   eta1 <~ y13      0.4542      0.0196   23.1210    0.0000 [ 0.4027; 0.5043 ] 
#>   eta2 <~ y21      0.3058      0.0271   11.2704    0.0000 [ 0.2417; 0.3820 ] 
#>   eta2 <~ y22      0.4473      0.0256   17.4852    0.0000 [ 0.3807; 0.5130 ] 
#>   eta2 <~ y23      0.4735      0.0259   18.2677    0.0000 [ 0.4022; 0.5363 ] 
#>   eta3 <~ y31      0.4400      0.0183   24.0737    0.0000 [ 0.3935; 0.4880 ] 
#>   eta3 <~ y32      0.3521      0.0190   18.5415    0.0000 [ 0.3005; 0.3987 ] 
#>   eta3 <~ y33      0.3999      0.0196   20.3649    0.0000 [ 0.3540; 0.4555 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0442   15.1948    0.0000 [ 0.5598; 0.7883 ] 
#>   eta3 ~ eta1       0.6634      0.0418   15.8659    0.0000 [ 0.5623; 0.7785 ] 
#>   eta3 ~ eta2       0.3052      0.0901    3.3886    0.0007 [ 0.0641; 0.5298 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0607    3.3755    0.0007 [ 0.0436; 0.3575 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04418173 15.194818 3.827230e-52
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08554215  5.360010 8.321729e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.09005300  3.388573 7.025741e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.55983825          0.7883212          0.5872746          0.7608849
#> 2         0.24863876          0.6910144          0.3017594          0.6378938
#> 3         0.06414057          0.5298438          0.1200624          0.4739220
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5749047          0.7436366          0.5991365          0.7343029
#> 2          0.2808721          0.6455809          0.3003112          0.6173734
#> 3          0.0664866          0.5019419          0.1091866          0.4901082