Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1390748584
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0384   17.4904    0.0000 [ 0.6022; 0.7300 ] 
#>   eta3 ~ eta1      0.4585      0.0992    4.6204    0.0000 [ 0.2887; 0.6159 ] 
#>   eta3 ~ eta2      0.3052      0.1018    2.9975    0.0027 [ 0.1095; 0.4703 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0447   14.8203    0.0000 [ 0.5816; 0.7372 ] 
#>   eta1 =~ y12      0.6493      0.0374   17.3824    0.0000 [ 0.5677; 0.7222 ] 
#>   eta1 =~ y13      0.7613      0.0294   25.8845    0.0000 [ 0.7238; 0.8193 ] 
#>   eta2 =~ y21      0.5165      0.0513   10.0590    0.0000 [ 0.4251; 0.5808 ] 
#>   eta2 =~ y22      0.7554      0.0353   21.4258    0.0000 [ 0.6902; 0.8217 ] 
#>   eta2 =~ y23      0.7997      0.0392   20.4186    0.0000 [ 0.7229; 0.8652 ] 
#>   eta3 =~ y31      0.8223      0.0422   19.4664    0.0000 [ 0.7563; 0.8952 ] 
#>   eta3 =~ y32      0.6581      0.0374   17.5743    0.0000 [ 0.6002; 0.7336 ] 
#>   eta3 =~ y33      0.7474      0.0428   17.4604    0.0000 [ 0.6832; 0.8213 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0235   16.8365    0.0000 [ 0.3555; 0.4314 ] 
#>   eta1 <~ y12      0.3873      0.0206   18.7924    0.0000 [ 0.3461; 0.4256 ] 
#>   eta1 <~ y13      0.4542      0.0195   23.2784    0.0000 [ 0.4172; 0.4925 ] 
#>   eta2 <~ y21      0.3058      0.0281   10.8658    0.0000 [ 0.2640; 0.3589 ] 
#>   eta2 <~ y22      0.4473      0.0220   20.3376    0.0000 [ 0.4134; 0.4944 ] 
#>   eta2 <~ y23      0.4735      0.0207   22.9177    0.0000 [ 0.4372; 0.5062 ] 
#>   eta3 <~ y31      0.4400      0.0205   21.4432    0.0000 [ 0.4042; 0.4733 ] 
#>   eta3 <~ y32      0.3521      0.0165   21.3615    0.0000 [ 0.3333; 0.3852 ] 
#>   eta3 <~ y33      0.3999      0.0234   17.1017    0.0000 [ 0.3576; 0.4341 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0384   17.4904    0.0000 [ 0.6022; 0.7300 ] 
#>   eta3 ~ eta1       0.6634      0.0449   14.7828    0.0000 [ 0.5738; 0.7408 ] 
#>   eta3 ~ eta2       0.3052      0.1018    2.9975    0.0027 [ 0.1095; 0.4703 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0718    2.8550    0.0043 [ 0.0792; 0.3369 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04474069 14.82029  1.083193e-49
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03735260 17.38240  1.121515e-67
#> 3 eta1 =~ y13  Common factor 0.7613458 0.02941319 25.88451 9.952540e-148
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05134277 10.05896  8.388128e-24
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03525598 21.42580 7.679975e-102
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03916344 20.41863  1.142267e-92
#> 7 eta3 =~ y31  Common factor 0.8222773 0.04224085 19.46640  2.116297e-84
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03744502 17.57427  3.878539e-69
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04280684 17.46039  2.869733e-68
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5815841          0.7371780
#> 2          0.5677390          0.7221640
#> 3          0.7238102          0.8193091
#> 4          0.4251030          0.5807851
#> 5          0.6902165          0.8216580
#> 6          0.7228838          0.8652338
#> 7          0.7563079          0.8952490
#> 8          0.6001917          0.7336007
#> 9          0.6831992          0.8212756

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1390748584
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0384   17.4904    0.0000 [ 0.5782; 0.7767 ] 
#>   eta3 ~ eta1      0.4585      0.0992    4.6204    0.0000 [ 0.2025; 0.7157 ] 
#>   eta3 ~ eta2      0.3052      0.1018    2.9975    0.0027 [ 0.0492; 0.5757 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0447   14.8203    0.0000 [ 0.5564; 0.7878 ] 
#>   eta1 =~ y12      0.6493      0.0374   17.3824    0.0000 [ 0.5498; 0.7430 ] 
#>   eta1 =~ y13      0.7613      0.0294   25.8845    0.0000 [ 0.6805; 0.8327 ] 
#>   eta2 =~ y21      0.5165      0.0513   10.0590    0.0000 [ 0.3974; 0.6629 ] 
#>   eta2 =~ y22      0.7554      0.0353   21.4258    0.0000 [ 0.6641; 0.8465 ] 
#>   eta2 =~ y23      0.7997      0.0392   20.4186    0.0000 [ 0.6998; 0.9023 ] 
#>   eta3 =~ y31      0.8223      0.0422   19.4664    0.0000 [ 0.7161; 0.9345 ] 
#>   eta3 =~ y32      0.6581      0.0374   17.5743    0.0000 [ 0.5428; 0.7365 ] 
#>   eta3 =~ y33      0.7474      0.0428   17.4604    0.0000 [ 0.6376; 0.8590 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0235   16.8365    0.0000 [ 0.3403; 0.4618 ] 
#>   eta1 <~ y12      0.3873      0.0206   18.7924    0.0000 [ 0.3323; 0.4388 ] 
#>   eta1 <~ y13      0.4542      0.0195   23.2784    0.0000 [ 0.4007; 0.5016 ] 
#>   eta2 <~ y21      0.3058      0.0281   10.8658    0.0000 [ 0.2389; 0.3844 ] 
#>   eta2 <~ y22      0.4473      0.0220   20.3376    0.0000 [ 0.3864; 0.5002 ] 
#>   eta2 <~ y23      0.4735      0.0207   22.9177    0.0000 [ 0.4169; 0.5238 ] 
#>   eta3 <~ y31      0.4400      0.0205   21.4432    0.0000 [ 0.3929; 0.4990 ] 
#>   eta3 <~ y32      0.3521      0.0165   21.3615    0.0000 [ 0.3033; 0.3885 ] 
#>   eta3 <~ y33      0.3999      0.0234   17.1017    0.0000 [ 0.3437; 0.4646 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0384   17.4904    0.0000 [ 0.5782; 0.7767 ] 
#>   eta3 ~ eta1       0.6634      0.0449   14.7828    0.0000 [ 0.5543; 0.7863 ] 
#>   eta3 ~ eta2       0.3052      0.1018    2.9975    0.0027 [ 0.0492; 0.5757 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0718    2.8550    0.0043 [ 0.0257; 0.3968 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03838289 17.490435 1.694586e-68
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.09923477  4.620425 3.829554e-06
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.10180218  2.997491 2.722118e-03
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.57821059          0.7767052          0.6020459          0.7528699
#> 2         0.20248499          0.7156712          0.2641086          0.6540475
#> 3         0.04921735          0.5756807          0.1124353          0.5124628
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.60195214          0.7501610          0.6022276          0.7299535
#> 2         0.27036978          0.6528874          0.2886776          0.6159359
#> 3         0.07802854          0.4784723          0.1095251          0.4702761