Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1695953273
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0397   16.9187    0.0000 [ 0.6134; 0.7419 ] 
#>   eta3 ~ eta1      0.4585      0.0759    6.0439    0.0000 [ 0.3382; 0.5913 ] 
#>   eta3 ~ eta2      0.3052      0.0803    3.8011    0.0001 [ 0.1610; 0.4330 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0444   14.9452    0.0000 [ 0.5892; 0.7497 ] 
#>   eta1 =~ y12      0.6493      0.0326   19.8872    0.0000 [ 0.5961; 0.7100 ] 
#>   eta1 =~ y13      0.7613      0.0344   22.1416    0.0000 [ 0.6733; 0.8217 ] 
#>   eta2 =~ y21      0.5165      0.0572    9.0235    0.0000 [ 0.3953; 0.6208 ] 
#>   eta2 =~ y22      0.7554      0.0409   18.4673    0.0000 [ 0.6961; 0.8244 ] 
#>   eta2 =~ y23      0.7997      0.0360   22.2152    0.0000 [ 0.7332; 0.8591 ] 
#>   eta3 =~ y31      0.8223      0.0369   22.3017    0.0000 [ 0.7610; 0.9060 ] 
#>   eta3 =~ y32      0.6581      0.0434   15.1522    0.0000 [ 0.5623; 0.7128 ] 
#>   eta3 =~ y33      0.7474      0.0402   18.6015    0.0000 [ 0.6653; 0.7989 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0230   17.1657    0.0000 [ 0.3565; 0.4235 ] 
#>   eta1 <~ y12      0.3873      0.0183   21.1143    0.0000 [ 0.3591; 0.4273 ] 
#>   eta1 <~ y13      0.4542      0.0197   22.9977    0.0000 [ 0.4242; 0.4834 ] 
#>   eta2 <~ y21      0.3058      0.0313    9.7740    0.0000 [ 0.2350; 0.3696 ] 
#>   eta2 <~ y22      0.4473      0.0213   21.0373    0.0000 [ 0.3946; 0.4741 ] 
#>   eta2 <~ y23      0.4735      0.0206   22.9549    0.0000 [ 0.4404; 0.5125 ] 
#>   eta3 <~ y31      0.4400      0.0188   23.4172    0.0000 [ 0.4172; 0.4845 ] 
#>   eta3 <~ y32      0.3521      0.0209   16.8799    0.0000 [ 0.3128; 0.3845 ] 
#>   eta3 <~ y33      0.3999      0.0211   18.9294    0.0000 [ 0.3545; 0.4299 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0397   16.9187    0.0000 [ 0.6134; 0.7419 ] 
#>   eta3 ~ eta1       0.6634      0.0439   15.0959    0.0000 [ 0.5823; 0.7379 ] 
#>   eta3 ~ eta2       0.3052      0.0803    3.8011    0.0001 [ 0.1610; 0.4330 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0553    3.7012    0.0002 [ 0.1001; 0.3062 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04436668 14.945223  1.673312e-50
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03264802 19.887205  5.252229e-88
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03438529 22.141616 1.256493e-108
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05723434  9.023513  1.821517e-19
#> 5 eta2 =~ y22  Common factor 0.7553877 0.04090404 18.467314  3.784506e-76
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03599619 22.215231 2.447199e-109
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03687053 22.301748 3.553497e-110
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04343060 15.152194  7.327874e-52
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04018085 18.601501  3.124338e-77
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5892442          0.7496570
#> 2          0.5961432          0.7100344
#> 3          0.6732659          0.8217107
#> 4          0.3952803          0.6207611
#> 5          0.6960797          0.8244159
#> 6          0.7331896          0.8591295
#> 7          0.7610303          0.9059798
#> 8          0.5622788          0.7128343
#> 9          0.6653400          0.7988733

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1695953273
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0397   16.9187    0.0000 [ 0.5618; 0.7670 ] 
#>   eta3 ~ eta1      0.4585      0.0759    6.0439    0.0000 [ 0.2544; 0.6468 ] 
#>   eta3 ~ eta2      0.3052      0.0803    3.8011    0.0001 [ 0.1072; 0.5223 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0444   14.9452    0.0000 [ 0.5530; 0.7825 ] 
#>   eta1 =~ y12      0.6493      0.0326   19.8872    0.0000 [ 0.5570; 0.7258 ] 
#>   eta1 =~ y13      0.7613      0.0344   22.1416    0.0000 [ 0.6732; 0.8510 ] 
#>   eta2 =~ y21      0.5165      0.0572    9.0235    0.0000 [ 0.3592; 0.6552 ] 
#>   eta2 =~ y22      0.7554      0.0409   18.4673    0.0000 [ 0.6526; 0.8641 ] 
#>   eta2 =~ y23      0.7997      0.0360   22.2152    0.0000 [ 0.7034; 0.8895 ] 
#>   eta3 =~ y31      0.8223      0.0369   22.3017    0.0000 [ 0.7257; 0.9164 ] 
#>   eta3 =~ y32      0.6581      0.0434   15.1522    0.0000 [ 0.5437; 0.7683 ] 
#>   eta3 =~ y33      0.7474      0.0402   18.6015    0.0000 [ 0.6517; 0.8595 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0230   17.1657    0.0000 [ 0.3398; 0.4589 ] 
#>   eta1 <~ y12      0.3873      0.0183   21.1143    0.0000 [ 0.3360; 0.4309 ] 
#>   eta1 <~ y13      0.4542      0.0197   22.9977    0.0000 [ 0.4045; 0.5066 ] 
#>   eta2 <~ y21      0.3058      0.0313    9.7740    0.0000 [ 0.2218; 0.3836 ] 
#>   eta2 <~ y22      0.4473      0.0213   21.0373    0.0000 [ 0.3973; 0.5072 ] 
#>   eta2 <~ y23      0.4735      0.0206   22.9549    0.0000 [ 0.4215; 0.5282 ] 
#>   eta3 <~ y31      0.4400      0.0188   23.4172    0.0000 [ 0.3894; 0.4866 ] 
#>   eta3 <~ y32      0.3521      0.0209   16.8799    0.0000 [ 0.2962; 0.4040 ] 
#>   eta3 <~ y33      0.3999      0.0211   18.9294    0.0000 [ 0.3485; 0.4578 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0397   16.9187    0.0000 [ 0.5618; 0.7670 ] 
#>   eta3 ~ eta1       0.6634      0.0439   15.0959    0.0000 [ 0.5466; 0.7738 ] 
#>   eta3 ~ eta2       0.3052      0.0803    3.8011    0.0001 [ 0.1072; 0.5223 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0553    3.7012    0.0002 [ 0.0665; 0.3527 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03968005 16.918663 3.277856e-64
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07586290  6.043887 1.504453e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08027904  3.801131 1.440373e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5617807          0.7669835          0.5864215          0.7423426
#> 2          0.2544323          0.6467524          0.3015423          0.5996424
#> 3          0.1071874          0.5223453          0.1570398          0.4724929
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5810633          0.7627060          0.6133624          0.7418642
#> 2          0.3317689          0.5928439          0.3381850          0.5912962
#> 3          0.1119416          0.4700521          0.1610449          0.4330048