Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1623780465
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0410   16.3667    0.0000 [ 0.6141; 0.7490 ] 
#>   eta3 ~ eta1      0.4585      0.0782    5.8642    0.0000 [ 0.3104; 0.5940 ] 
#>   eta3 ~ eta2      0.3052      0.0758    4.0263    0.0001 [ 0.1607; 0.4306 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0369   17.9861    0.0000 [ 0.5904; 0.7210 ] 
#>   eta1 =~ y12      0.6493      0.0451   14.3874    0.0000 [ 0.5488; 0.7082 ] 
#>   eta1 =~ y13      0.7613      0.0356   21.4004    0.0000 [ 0.7042; 0.8284 ] 
#>   eta2 =~ y21      0.5165      0.0472   10.9432    0.0000 [ 0.3979; 0.5949 ] 
#>   eta2 =~ y22      0.7554      0.0340   22.2135    0.0000 [ 0.7042; 0.8225 ] 
#>   eta2 =~ y23      0.7997      0.0391   20.4505    0.0000 [ 0.7236; 0.8625 ] 
#>   eta3 =~ y31      0.8223      0.0319   25.7527    0.0000 [ 0.7477; 0.8559 ] 
#>   eta3 =~ y32      0.6581      0.0427   15.4054    0.0000 [ 0.5739; 0.7085 ] 
#>   eta3 =~ y33      0.7474      0.0312   23.9481    0.0000 [ 0.6821; 0.7870 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0204   19.3897    0.0000 [ 0.3619; 0.4310 ] 
#>   eta1 <~ y12      0.3873      0.0220   17.6008    0.0000 [ 0.3443; 0.4199 ] 
#>   eta1 <~ y13      0.4542      0.0222   20.5002    0.0000 [ 0.4275; 0.5031 ] 
#>   eta2 <~ y21      0.3058      0.0247   12.3740    0.0000 [ 0.2534; 0.3445 ] 
#>   eta2 <~ y22      0.4473      0.0189   23.6938    0.0000 [ 0.4189; 0.4973 ] 
#>   eta2 <~ y23      0.4735      0.0209   22.7095    0.0000 [ 0.4396; 0.5095 ] 
#>   eta3 <~ y31      0.4400      0.0173   25.4224    0.0000 [ 0.4146; 0.4675 ] 
#>   eta3 <~ y32      0.3521      0.0182   19.3501    0.0000 [ 0.3222; 0.3795 ] 
#>   eta3 <~ y33      0.3999      0.0187   21.3867    0.0000 [ 0.3622; 0.4326 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0410   16.3667    0.0000 [ 0.6141; 0.7490 ] 
#>   eta3 ~ eta1       0.6634      0.0405   16.3639    0.0000 [ 0.5850; 0.7436 ] 
#>   eta3 ~ eta2       0.3052      0.0758    4.0263    0.0001 [ 0.1607; 0.4306 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0543    3.7743    0.0002 [ 0.1168; 0.3082 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03686574 17.98607  2.505062e-72
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04512825 14.38739  6.209215e-47
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03557627 21.40038 1.325226e-101
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04719428 10.94317  7.165338e-28
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03400576 22.21352 2.542161e-109
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03910250 20.45045  5.952162e-93
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03192981 25.75266 3.010072e-146
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04271679 15.40539  1.505735e-53
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03121017 23.94809 9.672961e-127
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5904357          0.7210138
#> 2          0.5487814          0.7081550
#> 3          0.7041579          0.8284442
#> 4          0.3978596          0.5948861
#> 5          0.7041605          0.8224876
#> 6          0.7236187          0.8624785
#> 7          0.7477126          0.8559275
#> 8          0.5739390          0.7084767
#> 9          0.6820740          0.7869622

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 1623780465
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0410   16.3667    0.0000 [ 0.5539; 0.7661 ] 
#>   eta3 ~ eta1      0.4585      0.0782    5.8642    0.0000 [ 0.2622; 0.6666 ] 
#>   eta3 ~ eta2      0.3052      0.0758    4.0263    0.0001 [ 0.1037; 0.4956 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0369   17.9861    0.0000 [ 0.5751; 0.7658 ] 
#>   eta1 =~ y12      0.6493      0.0451   14.3874    0.0000 [ 0.5489; 0.7823 ] 
#>   eta1 =~ y13      0.7613      0.0356   21.4004    0.0000 [ 0.6718; 0.8557 ] 
#>   eta2 =~ y21      0.5165      0.0472   10.9432    0.0000 [ 0.4066; 0.6507 ] 
#>   eta2 =~ y22      0.7554      0.0340   22.2135    0.0000 [ 0.6691; 0.8450 ] 
#>   eta2 =~ y23      0.7997      0.0391   20.4505    0.0000 [ 0.7042; 0.9064 ] 
#>   eta3 =~ y31      0.8223      0.0319   25.7527    0.0000 [ 0.7481; 0.9132 ] 
#>   eta3 =~ y32      0.6581      0.0427   15.4054    0.0000 [ 0.5505; 0.7714 ] 
#>   eta3 =~ y33      0.7474      0.0312   23.9481    0.0000 [ 0.6741; 0.8355 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0204   19.3897    0.0000 [ 0.3401; 0.4456 ] 
#>   eta1 <~ y12      0.3873      0.0220   17.6008    0.0000 [ 0.3336; 0.4474 ] 
#>   eta1 <~ y13      0.4542      0.0222   20.5002    0.0000 [ 0.3899; 0.5045 ] 
#>   eta2 <~ y21      0.3058      0.0247   12.3740    0.0000 [ 0.2457; 0.3735 ] 
#>   eta2 <~ y22      0.4473      0.0189   23.6938    0.0000 [ 0.3938; 0.4915 ] 
#>   eta2 <~ y23      0.4735      0.0209   22.7095    0.0000 [ 0.4171; 0.5249 ] 
#>   eta3 <~ y31      0.4400      0.0173   25.4224    0.0000 [ 0.3942; 0.4837 ] 
#>   eta3 <~ y32      0.3521      0.0182   19.3501    0.0000 [ 0.3024; 0.3965 ] 
#>   eta3 <~ y33      0.3999      0.0187   21.3867    0.0000 [ 0.3504; 0.4471 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0410   16.3667    0.0000 [ 0.5539; 0.7661 ] 
#>   eta3 ~ eta1       0.6634      0.0405   16.3639    0.0000 [ 0.5570; 0.7667 ] 
#>   eta3 ~ eta2       0.3052      0.0758    4.0263    0.0001 [ 0.1037; 0.4956 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0543    3.7743    0.0002 [ 0.0571; 0.3378 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04101825 16.366699 3.306669e-60
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07818688  5.864242 4.511895e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07578866  4.026343 5.665115e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5539399          0.7660632          0.5794118          0.7405913
#> 2          0.2622324          0.6665708          0.3107856          0.6180177
#> 3          0.1037028          0.4956390          0.1507667          0.4485751
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.6021155          0.7671376          0.6141467          0.7490093
#> 2          0.3094561          0.6109788          0.3104086          0.5939810
#> 3          0.1547024          0.4622177          0.1607342          0.4305942