Summarize model

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1680748694
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0459   14.6337    0.0000 [ 0.5899; 0.7406 ] 
#>   eta3 ~ eta1      0.4585      0.0755    6.0764    0.0000 [ 0.3192; 0.5636 ] 
#>   eta3 ~ eta2      0.3052      0.0795    3.8395    0.0001 [ 0.1609; 0.4401 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0406   16.3373    0.0000 [ 0.6004; 0.7303 ] 
#>   eta1 =~ y12      0.6493      0.0350   18.5480    0.0000 [ 0.5918; 0.7076 ] 
#>   eta1 =~ y13      0.7613      0.0350   21.7589    0.0000 [ 0.6981; 0.8239 ] 
#>   eta2 =~ y21      0.5165      0.0519    9.9521    0.0000 [ 0.4234; 0.5984 ] 
#>   eta2 =~ y22      0.7554      0.0343   22.0171    0.0000 [ 0.6903; 0.8118 ] 
#>   eta2 =~ y23      0.7997      0.0354   22.6026    0.0000 [ 0.7354; 0.8435 ] 
#>   eta3 =~ y31      0.8223      0.0343   24.0052    0.0000 [ 0.7702; 0.8957 ] 
#>   eta3 =~ y32      0.6581      0.0341   19.2705    0.0000 [ 0.6088; 0.7212 ] 
#>   eta3 =~ y33      0.7474      0.0481   15.5550    0.0000 [ 0.6433; 0.8331 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0217   18.2436    0.0000 [ 0.3568; 0.4331 ] 
#>   eta1 <~ y12      0.3873      0.0195   19.8763    0.0000 [ 0.3550; 0.4263 ] 
#>   eta1 <~ y13      0.4542      0.0181   25.0506    0.0000 [ 0.4208; 0.4952 ] 
#>   eta2 <~ y21      0.3058      0.0296   10.3441    0.0000 [ 0.2593; 0.3647 ] 
#>   eta2 <~ y22      0.4473      0.0201   22.2609    0.0000 [ 0.4201; 0.4882 ] 
#>   eta2 <~ y23      0.4735      0.0160   29.6477    0.0000 [ 0.4445; 0.4971 ] 
#>   eta3 <~ y31      0.4400      0.0225   19.5539    0.0000 [ 0.4044; 0.4831 ] 
#>   eta3 <~ y32      0.3521      0.0143   24.7060    0.0000 [ 0.3259; 0.3794 ] 
#>   eta3 <~ y33      0.3999      0.0208   19.1964    0.0000 [ 0.3616; 0.4375 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0459   14.6337    0.0000 [ 0.5899; 0.7406 ] 
#>   eta3 ~ eta1       0.6634      0.0352   18.8241    0.0000 [ 0.5955; 0.7149 ] 
#>   eta3 ~ eta2       0.3052      0.0795    3.8395    0.0001 [ 0.1609; 0.4401 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0584    3.5060    0.0005 [ 0.1019; 0.2968 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.04058624 16.337305  5.356873e-60
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03500523 18.548026  8.460241e-77
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03499011 21.758888 5.691215e-105
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05189410  9.952091  2.469389e-23
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03430909 22.017127 1.973857e-107
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03537923 22.602633 4.082907e-113
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03425416 24.005180 2.455146e-127
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03414896 19.270542  9.493739e-83
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04805056 15.554951  1.472672e-54
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6003732          0.7303080
#> 2          0.5917579          0.7076048
#> 3          0.6980918          0.8238894
#> 4          0.4234415          0.5984160
#> 5          0.6902679          0.8118260
#> 6          0.7354069          0.8434892
#> 7          0.7701805          0.8957038
#> 8          0.6088218          0.7212185
#> 9          0.6432944          0.8330867

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = 1680748694
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0459   14.6337    0.0000 [ 0.5497; 0.7869 ] 
#>   eta3 ~ eta1      0.4585      0.0755    6.0764    0.0000 [ 0.2818; 0.6720 ] 
#>   eta3 ~ eta2      0.3052      0.0795    3.8395    0.0001 [ 0.0779; 0.4889 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0406   16.3373    0.0000 [ 0.5528; 0.7627 ] 
#>   eta1 =~ y12      0.6493      0.0350   18.5480    0.0000 [ 0.5634; 0.7444 ] 
#>   eta1 =~ y13      0.7613      0.0350   21.7589    0.0000 [ 0.6762; 0.8571 ] 
#>   eta2 =~ y21      0.5165      0.0519    9.9521    0.0000 [ 0.3866; 0.6549 ] 
#>   eta2 =~ y22      0.7554      0.0343   22.0171    0.0000 [ 0.6630; 0.8404 ] 
#>   eta2 =~ y23      0.7997      0.0354   22.6026    0.0000 [ 0.7187; 0.9016 ] 
#>   eta3 =~ y31      0.8223      0.0343   24.0052    0.0000 [ 0.7240; 0.9012 ] 
#>   eta3 =~ y32      0.6581      0.0341   19.2705    0.0000 [ 0.5667; 0.7433 ] 
#>   eta3 =~ y33      0.7474      0.0481   15.5550    0.0000 [ 0.6226; 0.8711 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0217   18.2436    0.0000 [ 0.3352; 0.4473 ] 
#>   eta1 <~ y12      0.3873      0.0195   19.8763    0.0000 [ 0.3386; 0.4393 ] 
#>   eta1 <~ y13      0.4542      0.0181   25.0506    0.0000 [ 0.4092; 0.5029 ] 
#>   eta2 <~ y21      0.3058      0.0296   10.3441    0.0000 [ 0.2299; 0.3828 ] 
#>   eta2 <~ y22      0.4473      0.0201   22.2609    0.0000 [ 0.3900; 0.4939 ] 
#>   eta2 <~ y23      0.4735      0.0160   29.6477    0.0000 [ 0.4353; 0.5179 ] 
#>   eta3 <~ y31      0.4400      0.0225   19.5539    0.0000 [ 0.3806; 0.4970 ] 
#>   eta3 <~ y32      0.3521      0.0143   24.7060    0.0000 [ 0.3171; 0.3908 ] 
#>   eta3 <~ y33      0.3999      0.0208   19.1964    0.0000 [ 0.3498; 0.4575 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0459   14.6337    0.0000 [ 0.5497; 0.7869 ] 
#>   eta3 ~ eta1       0.6634      0.0352   18.8241    0.0000 [ 0.5746; 0.7568 ] 
#>   eta3 ~ eta2       0.3052      0.0795    3.8395    0.0001 [ 0.0779; 0.4889 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0584    3.5060    0.0005 [ 0.0377; 0.3399 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04587582 14.633709 1.711943e-48
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07545664  6.076427 1.228901e-09
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07947778  3.839452 1.233092e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54968974          0.7869336          0.5781781          0.7584452
#> 2         0.28181319          0.6720323          0.3286709          0.6251747
#> 3         0.07788176          0.4888960          0.1272365          0.4395412
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5814920          0.7746239          0.5898822          0.7405955
#> 2          0.2020299          0.6005729          0.3191856          0.5635552
#> 3          0.1212473          0.5635491          0.1608982          0.4400906