Summarize model

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -1326160720
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0433   15.5093    0.0000 [ 0.5841; 0.7438 ] 
#>   eta3 ~ eta1      0.4585      0.0833    5.5062    0.0000 [ 0.3264; 0.6172 ] 
#>   eta3 ~ eta2      0.3052      0.0859    3.5508    0.0004 [ 0.1327; 0.4407 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0378   17.5233    0.0000 [ 0.6058; 0.7275 ] 
#>   eta1 =~ y12      0.6493      0.0353   18.4185    0.0000 [ 0.5936; 0.7067 ] 
#>   eta1 =~ y13      0.7613      0.0310   24.5793    0.0000 [ 0.7073; 0.8111 ] 
#>   eta2 =~ y21      0.5165      0.0485   10.6595    0.0000 [ 0.4322; 0.5965 ] 
#>   eta2 =~ y22      0.7554      0.0337   22.4319    0.0000 [ 0.7027; 0.8159 ] 
#>   eta2 =~ y23      0.7997      0.0355   22.5239    0.0000 [ 0.7398; 0.8614 ] 
#>   eta3 =~ y31      0.8223      0.0313   26.2826    0.0000 [ 0.7723; 0.8800 ] 
#>   eta3 =~ y32      0.6581      0.0419   15.6898    0.0000 [ 0.5874; 0.7189 ] 
#>   eta3 =~ y33      0.7474      0.0333   22.4672    0.0000 [ 0.6714; 0.8040 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0208   19.0041    0.0000 [ 0.3577; 0.4267 ] 
#>   eta1 <~ y12      0.3873      0.0171   22.6343    0.0000 [ 0.3623; 0.4125 ] 
#>   eta1 <~ y13      0.4542      0.0174   26.1325    0.0000 [ 0.4230; 0.4826 ] 
#>   eta2 <~ y21      0.3058      0.0218   14.0129    0.0000 [ 0.2700; 0.3448 ] 
#>   eta2 <~ y22      0.4473      0.0197   22.6692    0.0000 [ 0.4075; 0.4793 ] 
#>   eta2 <~ y23      0.4735      0.0210   22.5802    0.0000 [ 0.4294; 0.5051 ] 
#>   eta3 <~ y31      0.4400      0.0163   26.9262    0.0000 [ 0.4076; 0.4696 ] 
#>   eta3 <~ y32      0.3521      0.0195   18.0256    0.0000 [ 0.3211; 0.3822 ] 
#>   eta3 <~ y33      0.3999      0.0188   21.2657    0.0000 [ 0.3563; 0.4277 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0433   15.5093    0.0000 [ 0.5841; 0.7438 ] 
#>   eta3 ~ eta1       0.6634      0.0403   16.4640    0.0000 [ 0.5833; 0.7431 ] 
#>   eta3 ~ eta2       0.3052      0.0859    3.5508    0.0004 [ 0.1327; 0.4407 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0617    3.3208    0.0009 [ 0.0900; 0.3080 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03783933 17.52330  9.514863e-69
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03525143 18.41849  9.337697e-76
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03097505 24.57933 2.101894e-133
#> 4 eta2 =~ y21  Common factor 0.5164548 0.04845037 10.65946  1.575069e-26
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03367477 22.43186 1.924397e-111
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03550295 22.52386 2.422959e-112
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03128599 26.28261 3.031561e-152
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04194251 15.68978  1.776592e-55
#> 9 eta3 =~ y33  Common factor 0.7474241 0.03326728 22.46725 8.681062e-112
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6057843          0.7275234
#> 2          0.5935561          0.7067190
#> 3          0.7072609          0.8111216
#> 4          0.4321872          0.5965003
#> 5          0.7026668          0.8159312
#> 6          0.7397602          0.8614276
#> 7          0.7723261          0.8799588
#> 8          0.5873622          0.7188946
#> 9          0.6714090          0.8039584

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = -1326160720
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0433   15.5093    0.0000 [ 0.5439; 0.7678 ] 
#>   eta3 ~ eta1      0.4585      0.0833    5.5062    0.0000 [ 0.2355; 0.6661 ] 
#>   eta3 ~ eta2      0.3052      0.0859    3.5508    0.0004 [ 0.0939; 0.5383 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0378   17.5233    0.0000 [ 0.5585; 0.7542 ] 
#>   eta1 =~ y12      0.6493      0.0353   18.4185    0.0000 [ 0.5566; 0.7389 ] 
#>   eta1 =~ y13      0.7613      0.0310   24.5793    0.0000 [ 0.6842; 0.8444 ] 
#>   eta2 =~ y21      0.5165      0.0485   10.6595    0.0000 [ 0.3785; 0.6291 ] 
#>   eta2 =~ y22      0.7554      0.0337   22.4319    0.0000 [ 0.6658; 0.8400 ] 
#>   eta2 =~ y23      0.7997      0.0355   22.5239    0.0000 [ 0.7141; 0.8977 ] 
#>   eta3 =~ y31      0.8223      0.0313   26.2826    0.0000 [ 0.7363; 0.8981 ] 
#>   eta3 =~ y32      0.6581      0.0419   15.6898    0.0000 [ 0.5465; 0.7635 ] 
#>   eta3 =~ y33      0.7474      0.0333   22.4672    0.0000 [ 0.6703; 0.8423 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0208   19.0041    0.0000 [ 0.3393; 0.4470 ] 
#>   eta1 <~ y12      0.3873      0.0171   22.6343    0.0000 [ 0.3439; 0.4324 ] 
#>   eta1 <~ y13      0.4542      0.0174   26.1325    0.0000 [ 0.4128; 0.5027 ] 
#>   eta2 <~ y21      0.3058      0.0218   14.0129    0.0000 [ 0.2441; 0.3569 ] 
#>   eta2 <~ y22      0.4473      0.0197   22.6692    0.0000 [ 0.3972; 0.4992 ] 
#>   eta2 <~ y23      0.4735      0.0210   22.5802    0.0000 [ 0.4255; 0.5339 ] 
#>   eta3 <~ y31      0.4400      0.0163   26.9262    0.0000 [ 0.3950; 0.4795 ] 
#>   eta3 <~ y32      0.3521      0.0195   18.0256    0.0000 [ 0.3001; 0.4011 ] 
#>   eta3 <~ y33      0.3999      0.0188   21.2657    0.0000 [ 0.3559; 0.4532 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0433   15.5093    0.0000 [ 0.5439; 0.7678 ] 
#>   eta3 ~ eta1       0.6634      0.0403   16.4640    0.0000 [ 0.5543; 0.7626 ] 
#>   eta3 ~ eta2       0.3052      0.0859    3.5508    0.0004 [ 0.0939; 0.5383 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0617    3.3208    0.0009 [ 0.0481; 0.3672 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04328597 15.509262 3.003115e-54
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.08327028  5.506247 3.665631e-08
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08593858  3.550805 3.840553e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1         0.54394282          0.7677934          0.5708229          0.7409133
#> 2         0.23550715          0.6661340          0.2872170          0.6144242
#> 3         0.09385147          0.5382773          0.1472183          0.4849104
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1         0.56380440          0.7462325          0.5841434          0.7438155
#> 2         0.28803710          0.6514039          0.3263971          0.6172350
#> 3         0.07594834          0.4722850          0.1327072          0.4407466