Summarize model

Usage

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object

An R object of class cSEMResults resulting from a call to csem().

.alpha

An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci

A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...

Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -335691483
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0389   17.2592    0.0000 [ 0.6107; 0.7490 ] 
#>   eta3 ~ eta1      0.4585      0.0747    6.1377    0.0000 [ 0.2918; 0.6114 ] 
#>   eta3 ~ eta2      0.3052      0.0838    3.6402    0.0003 [ 0.1254; 0.4443 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0362   18.2972    0.0000 [ 0.6072; 0.7148 ] 
#>   eta1 =~ y12      0.6493      0.0406   15.9747    0.0000 [ 0.5773; 0.7061 ] 
#>   eta1 =~ y13      0.7613      0.0329   23.1752    0.0000 [ 0.6966; 0.8210 ] 
#>   eta2 =~ y21      0.5165      0.0580    8.9006    0.0000 [ 0.4195; 0.6173 ] 
#>   eta2 =~ y22      0.7554      0.0288   26.2361    0.0000 [ 0.6908; 0.7937 ] 
#>   eta2 =~ y23      0.7997      0.0455   17.5560    0.0000 [ 0.7175; 0.8595 ] 
#>   eta3 =~ y31      0.8223      0.0319   25.7778    0.0000 [ 0.7567; 0.8781 ] 
#>   eta3 =~ y32      0.6581      0.0394   16.7210    0.0000 [ 0.5933; 0.7267 ] 
#>   eta3 =~ y33      0.7474      0.0436   17.1423    0.0000 [ 0.6693; 0.8141 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0193   20.4431    0.0000 [ 0.3643; 0.4346 ] 
#>   eta1 <~ y12      0.3873      0.0208   18.6527    0.0000 [ 0.3426; 0.4168 ] 
#>   eta1 <~ y13      0.4542      0.0196   23.2303    0.0000 [ 0.4205; 0.4922 ] 
#>   eta2 <~ y21      0.3058      0.0305   10.0367    0.0000 [ 0.2614; 0.3644 ] 
#>   eta2 <~ y22      0.4473      0.0199   22.4268    0.0000 [ 0.4001; 0.4746 ] 
#>   eta2 <~ y23      0.4735      0.0213   22.2457    0.0000 [ 0.4275; 0.5066 ] 
#>   eta3 <~ y31      0.4400      0.0189   23.2699    0.0000 [ 0.4055; 0.4735 ] 
#>   eta3 <~ y32      0.3521      0.0198   17.7436    0.0000 [ 0.3130; 0.3878 ] 
#>   eta3 <~ y33      0.3999      0.0207   19.3261    0.0000 [ 0.3656; 0.4452 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0389   17.2592    0.0000 [ 0.6107; 0.7490 ] 
#>   eta3 ~ eta1       0.6634      0.0303   21.8712    0.0000 [ 0.6091; 0.7103 ] 
#>   eta3 ~ eta2       0.3052      0.0838    3.6402    0.0003 [ 0.1254; 0.4443 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0595    3.4430    0.0006 [ 0.0897; 0.3319 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03623897 18.297151  8.719042e-75
#> 2 eta1 =~ y12  Common factor 0.6492779 0.04064414 15.974701  1.917711e-57
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03285169 23.175242 8.092721e-119
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05802502  8.900553  5.556894e-19
#> 5 eta2 =~ y22  Common factor 0.7553877 0.02879196 26.236065 1.030881e-151
#> 6 eta2 =~ y23  Common factor 0.7996637 0.04554927 17.556015  5.350208e-69
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03189871 25.777758 1.574921e-146
#> 8 eta3 =~ y32  Common factor 0.6580689 0.03935587 16.720986  9.218215e-63
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04360127 17.142255  7.181025e-66
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.6072395          0.7147613
#> 2          0.5772564          0.7060937
#> 3          0.6965836          0.8210160
#> 4          0.4195360          0.6173093
#> 5          0.6907539          0.7936915
#> 6          0.7174953          0.8595119
#> 7          0.7567206          0.8780928
#> 8          0.5932787          0.7266837
#> 9          0.6692544          0.8141325

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize
#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                  = Ok
#> 	Number of observations             = 500
#> 	Weight estimator                   = PLS-PM
#> 	Inner weighting scheme             = "path"
#> 	Type of indicator correlation      = Pearson
#> 	Path model estimator               = OLS
#> 	Second-order approach              = NA
#> 	Type of path model                 = Linear
#> 	Disattenuated                      = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                    = "bootstrap"
#> 	Number of resamples                = 40
#> 	Number of admissible results       = 40
#> 	Approach to handle inadmissibles   = "drop"
#> 	Sign change option                 = "none"
#> 	Random seed                        = -335691483
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0389   17.2592    0.0000 [ 0.5610; 0.7621 ] 
#>   eta3 ~ eta1      0.4585      0.0747    6.1377    0.0000 [ 0.2529; 0.6392 ] 
#>   eta3 ~ eta2      0.3052      0.0838    3.6402    0.0003 [ 0.1017; 0.5352 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0362   18.2972    0.0000 [ 0.5712; 0.7586 ] 
#>   eta1 =~ y12      0.6493      0.0406   15.9747    0.0000 [ 0.5451; 0.7553 ] 
#>   eta1 =~ y13      0.7613      0.0329   23.1752    0.0000 [ 0.6807; 0.8506 ] 
#>   eta2 =~ y21      0.5165      0.0580    8.9006    0.0000 [ 0.3497; 0.6498 ] 
#>   eta2 =~ y22      0.7554      0.0288   26.2361    0.0000 [ 0.6878; 0.8367 ] 
#>   eta2 =~ y23      0.7997      0.0455   17.5560    0.0000 [ 0.6771; 0.9126 ] 
#>   eta3 =~ y31      0.8223      0.0319   25.7778    0.0000 [ 0.7425; 0.9074 ] 
#>   eta3 =~ y32      0.6581      0.0394   16.7210    0.0000 [ 0.5546; 0.7581 ] 
#>   eta3 =~ y33      0.7474      0.0436   17.1423    0.0000 [ 0.6268; 0.8523 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0193   20.4431    0.0000 [ 0.3448; 0.4449 ] 
#>   eta1 <~ y12      0.3873      0.0208   18.6527    0.0000 [ 0.3325; 0.4399 ] 
#>   eta1 <~ y13      0.4542      0.0196   23.2303    0.0000 [ 0.4040; 0.5051 ] 
#>   eta2 <~ y21      0.3058      0.0305   10.0367    0.0000 [ 0.2206; 0.3782 ] 
#>   eta2 <~ y22      0.4473      0.0199   22.4268    0.0000 [ 0.4039; 0.5070 ] 
#>   eta2 <~ y23      0.4735      0.0213   22.2457    0.0000 [ 0.4204; 0.5305 ] 
#>   eta3 <~ y31      0.4400      0.0189   23.2699    0.0000 [ 0.3947; 0.4925 ] 
#>   eta3 <~ y32      0.3521      0.0198   17.7436    0.0000 [ 0.3018; 0.4044 ] 
#>   eta3 <~ y33      0.3999      0.0207   19.3261    0.0000 [ 0.3444; 0.4514 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0389   17.2592    0.0000 [ 0.5610; 0.7621 ] 
#>   eta3 ~ eta1       0.6634      0.0303   21.8712    0.0000 [ 0.5782; 0.7350 ] 
#>   eta3 ~ eta2       0.3052      0.0838    3.6402    0.0003 [ 0.1017; 0.5352 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0595    3.4430    0.0006 [ 0.0567; 0.3644 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates
#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.03889712 17.259208 9.541089e-67
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.07470343  6.137694 8.372813e-10
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.08382739  3.640232 2.723927e-04
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5609935          0.7621474          0.5851482          0.7379928
#> 2          0.2528884          0.6392124          0.2992784          0.5928224
#> 3          0.1016555          0.5351634          0.1537113          0.4831076
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.6101446          0.7509432          0.6106794          0.7489856
#> 2          0.2855900          0.6181835          0.2917638          0.6113738
#> 3          0.1151228          0.4888258          0.1253961          0.4442702