Summarize model

The summary is mainly focused on estimated parameters. For quality criteria such as the average variance extracted (AVE), reliability estimates, effect size estimates etc., use assess().

summarize(
 .object = NULL, 
 .alpha  = 0.05,
 .ci     = NULL,
 ...
 )

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.ci
A vector of character strings naming the confidence interval to compute. For possible choices see infer().

...
Further arguments to summarize(). Currently ignored.

Value

An object of class cSEMSummarize. A cSEMSummarize object has the same structure as the cSEMResults object with a couple differences:

Elements $Path_estimates, $Loadings_estimates, $Weight_estimates, $Weight_estimates, and $Residual_correlation are standardized data frames instead of matrices.

Data frames $Effect_estimates, $Indicator_correlation, and $Exo_construct_correlation are added to $Estimates.

The data frame format is usually much more convenient if users intend to present the results in e.g., a paper or a presentation.

Details

If .object contains resamples, standard errors, t-values and p-values (assuming estimates are standard normally distributed) are printed as well. By default the percentile confidence interval is given as well. For other confidence intervals use the .ci argument. See infer() for possible choices and a description.

See also

Examples

## Take a look at the dataset
#?threecommonfactors

## Specify the (correct) model
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# (Reflective) measurement model
eta1 =~ y11 + y12 + y13
eta2 =~ y21 + y22 + y23
eta3 =~ y31 + y32 + y33
"

## Estimate
res <- csem(threecommonfactors, model, .resample_method = "bootstrap", .R = 40)

## Postestimation
res_summarize <- summarize(res)
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 162653002
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_percentile   
#>   Path           Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1      0.6713      0.0413   16.2614    0.0000 [ 0.6053; 0.7362 ] 
#>   eta3 ~ eta1      0.4585      0.0618    7.4193    0.0000 [ 0.3749; 0.5828 ] 
#>   eta3 ~ eta2      0.3052      0.0700    4.3580    0.0000 [ 0.1730; 0.4362 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_percentile   
#>   Loading        Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 =~ y11      0.6631      0.0391   16.9679    0.0000 [ 0.5868; 0.7349 ] 
#>   eta1 =~ y12      0.6493      0.0350   18.5695    0.0000 [ 0.5761; 0.7112 ] 
#>   eta1 =~ y13      0.7613      0.0370   20.5545    0.0000 [ 0.6969; 0.8257 ] 
#>   eta2 =~ y21      0.5165      0.0507   10.1918    0.0000 [ 0.3997; 0.5964 ] 
#>   eta2 =~ y22      0.7554      0.0338   22.3760    0.0000 [ 0.6881; 0.8291 ] 
#>   eta2 =~ y23      0.7997      0.0344   23.2261    0.0000 [ 0.7198; 0.8597 ] 
#>   eta3 =~ y31      0.8223      0.0375   21.9469    0.0000 [ 0.7557; 0.8729 ] 
#>   eta3 =~ y32      0.6581      0.0408   16.1363    0.0000 [ 0.5968; 0.7314 ] 
#>   eta3 =~ y33      0.7474      0.0435   17.1866    0.0000 [ 0.6738; 0.8054 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_percentile   
#>   Weight         Estimate  Std. error   t-stat.   p-value         95%        
#>   eta1 <~ y11      0.3956      0.0227   17.3921    0.0000 [ 0.3567; 0.4420 ] 
#>   eta1 <~ y12      0.3873      0.0191   20.2993    0.0000 [ 0.3506; 0.4174 ] 
#>   eta1 <~ y13      0.4542      0.0177   25.6679    0.0000 [ 0.4328; 0.4856 ] 
#>   eta2 <~ y21      0.3058      0.0245   12.4815    0.0000 [ 0.2576; 0.3453 ] 
#>   eta2 <~ y22      0.4473      0.0205   21.7883    0.0000 [ 0.4132; 0.4971 ] 
#>   eta2 <~ y23      0.4735      0.0216   21.9221    0.0000 [ 0.4438; 0.5171 ] 
#>   eta3 <~ y31      0.4400      0.0184   23.9697    0.0000 [ 0.4066; 0.4679 ] 
#>   eta3 <~ y32      0.3521      0.0193   18.2518    0.0000 [ 0.3149; 0.3923 ] 
#>   eta3 <~ y33      0.3999      0.0229   17.4335    0.0000 [ 0.3545; 0.4423 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_percentile   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta2 ~ eta1       0.6713      0.0413   16.2614    0.0000 [ 0.6053; 0.7362 ] 
#>   eta3 ~ eta1       0.6634      0.0379   17.5072    0.0000 [ 0.6084; 0.7476 ] 
#>   eta3 ~ eta2       0.3052      0.0700    4.3580    0.0000 [ 0.1730; 0.4362 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_percentile   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         95%        
#>   eta3 ~ eta1          0.2049      0.0439    4.6615    0.0000 [ 0.1188; 0.2732 ] 
#> ________________________________________________________________________________

# Extract e.g. the loadings
res_summarize$Estimates$Loading_estimates

#>          Name Construct_type  Estimate    Std_err   t_stat       p_value
#> 1 eta1 =~ y11  Common factor 0.6630699 0.03907790 16.96790  1.419148e-64
#> 2 eta1 =~ y12  Common factor 0.6492779 0.03496467 18.56954  5.668174e-77
#> 3 eta1 =~ y13  Common factor 0.7613458 0.03704030 20.55452  7.011328e-94
#> 4 eta2 =~ y21  Common factor 0.5164548 0.05067344 10.19182  2.156790e-24
#> 5 eta2 =~ y22  Common factor 0.7553877 0.03375876 22.37605 6.735896e-111
#> 6 eta2 =~ y23  Common factor 0.7996637 0.03442952 23.22611 2.480913e-119
#> 7 eta3 =~ y31  Common factor 0.8222773 0.03746673 21.94687 9.277277e-107
#> 8 eta3 =~ y32  Common factor 0.6580689 0.04078197 16.13627  1.418613e-58
#> 9 eta3 =~ y33  Common factor 0.7474241 0.04348881 17.18658  3.346783e-66
#>   CI_percentile.95%L CI_percentile.95%U
#> 1          0.5868075          0.7348934
#> 2          0.5760859          0.7111879
#> 3          0.6969159          0.8257267
#> 4          0.3997408          0.5964299
#> 5          0.6881148          0.8290820
#> 6          0.7198206          0.8596593
#> 7          0.7556622          0.8729142
#> 8          0.5968447          0.7314467
#> 9          0.6738069          0.8053909

## By default only the 95% percentile confidence interval is printed. User
## can have several confidence interval computed, however, only the first
## will be printed.

res_summarize <- summarize(res, .ci = c("CI_standard_t", "CI_percentile"), 
                           .alpha = c(0.05, 0.01))
res_summarize

#> ________________________________________________________________________________
#> ----------------------------------- Overview -----------------------------------
#> 
#> 	General information:
#> 	------------------------
#> 	Estimation status                = Ok
#> 	Number of observations           = 500
#> 	Weight estimator                 = PLS-PM
#> 	Inner weighting scheme           = "path"
#> 	Type of indicator correlation    = Pearson
#> 	Path model estimator             = OLS
#> 	Second-order approach            = NA
#> 	Type of path model               = Linear
#> 	Disattenuated                    = Yes (PLSc)
#> 
#> 	Resample information:
#> 	---------------------
#> 	Resample method                  = "bootstrap"
#> 	Number of resamples              = 40
#> 	Number of admissible results     = 40
#> 	Approach to handle inadmissibles = "drop"
#> 	Sign change option               = "none"
#> 	Random seed                      = 162653002
#> 
#> 	Construct details:
#> 	------------------
#> 	Name  Modeled as     Order         Mode      
#> 
#> 	eta1  Common factor  First order   "modeA"   
#> 	eta2  Common factor  First order   "modeA"   
#> 	eta3  Common factor  First order   "modeA"   
#> 
#> ----------------------------------- Estimates ----------------------------------By default, only one confidence interval supplied to `.ci` is printed.
#> Use `xxx` to print all confidence intervals (not yet implemented).
#> 
#> 
#> 
#> Estimated path coefficients:
#> ============================
#>                                                              CI_standard_t   
#>   Path           Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1      0.6713      0.0413   16.2614    0.0000 [ 0.5624; 0.7759 ] 
#>   eta3 ~ eta1      0.4585      0.0618    7.4193    0.0000 [ 0.2761; 0.5957 ] 
#>   eta3 ~ eta2      0.3052      0.0700    4.3580    0.0000 [ 0.1424; 0.5045 ] 
#> 
#> Estimated loadings:
#> ===================
#>                                                              CI_standard_t   
#>   Loading        Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 =~ y11      0.6631      0.0391   16.9679    0.0000 [ 0.5618; 0.7639 ] 
#>   eta1 =~ y12      0.6493      0.0350   18.5695    0.0000 [ 0.5701; 0.7509 ] 
#>   eta1 =~ y13      0.7613      0.0370   20.5545    0.0000 [ 0.6628; 0.8543 ] 
#>   eta2 =~ y21      0.5165      0.0507   10.1918    0.0000 [ 0.3887; 0.6507 ] 
#>   eta2 =~ y22      0.7554      0.0338   22.3760    0.0000 [ 0.6738; 0.8484 ] 
#>   eta2 =~ y23      0.7997      0.0344   23.2261    0.0000 [ 0.7162; 0.8942 ] 
#>   eta3 =~ y31      0.8223      0.0375   21.9469    0.0000 [ 0.7272; 0.9210 ] 
#>   eta3 =~ y32      0.6581      0.0408   16.1363    0.0000 [ 0.5475; 0.7584 ] 
#>   eta3 =~ y33      0.7474      0.0435   17.1866    0.0000 [ 0.6408; 0.8657 ] 
#> 
#> Estimated weights:
#> ==================
#>                                                              CI_standard_t   
#>   Weight         Estimate  Std. error   t-stat.   p-value         99%        
#>   eta1 <~ y11      0.3956      0.0227   17.3921    0.0000 [ 0.3345; 0.4521 ] 
#>   eta1 <~ y12      0.3873      0.0191   20.2993    0.0000 [ 0.3427; 0.4414 ] 
#>   eta1 <~ y13      0.4542      0.0177   25.6679    0.0000 [ 0.4045; 0.4960 ] 
#>   eta2 <~ y21      0.3058      0.0245   12.4815    0.0000 [ 0.2419; 0.3686 ] 
#>   eta2 <~ y22      0.4473      0.0205   21.7883    0.0000 [ 0.3932; 0.4993 ] 
#>   eta2 <~ y23      0.4735      0.0216   21.9221    0.0000 [ 0.4162; 0.5279 ] 
#>   eta3 <~ y31      0.4400      0.0184   23.9697    0.0000 [ 0.3929; 0.4878 ] 
#>   eta3 <~ y32      0.3521      0.0193   18.2518    0.0000 [ 0.2991; 0.3988 ] 
#>   eta3 <~ y33      0.3999      0.0229   17.4335    0.0000 [ 0.3431; 0.4617 ] 
#> 
#> ------------------------------------ Effects -----------------------------------
#> 
#> Estimated total effects:
#> ========================
#>                                                               CI_standard_t   
#>   Total effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta2 ~ eta1       0.6713      0.0413   16.2614    0.0000 [ 0.5624; 0.7759 ] 
#>   eta3 ~ eta1       0.6634      0.0379   17.5072    0.0000 [ 0.5554; 0.7514 ] 
#>   eta3 ~ eta2       0.3052      0.0700    4.3580    0.0000 [ 0.1424; 0.5045 ] 
#> 
#> Estimated indirect effects:
#> ===========================
#>                                                                  CI_standard_t   
#>   Indirect effect    Estimate  Std. error   t-stat.   p-value         99%        
#>   eta3 ~ eta1          0.2049      0.0439    4.6615    0.0000 [ 0.1038; 0.3311 ] 
#> ________________________________________________________________________________

# Extract the loading including both confidence intervals
res_summarize$Estimates$Path_estimates

#>          Name Construct_type  Estimate    Std_err    t_stat      p_value
#> 1 eta2 ~ eta1  Common factor 0.6713334 0.04128394 16.261368 1.855480e-59
#> 2 eta3 ~ eta1  Common factor 0.4585068 0.06179914  7.419307 1.177349e-13
#> 3 eta3 ~ eta2  Common factor 0.3051511 0.07002151  4.357963 1.312788e-05
#>   CI_standard_t.99%L CI_standard_t.99%U CI_standard_t.95%L CI_standard_t.95%U
#> 1          0.5623903          0.7758876          0.5880271          0.7502507
#> 2          0.2761464          0.5957366          0.3145229          0.5573601
#> 3          0.1424141          0.5045258          0.1858966          0.4610432
#>   CI_percentile.99%L CI_percentile.99%U CI_percentile.95%L CI_percentile.95%U
#> 1          0.5609948          0.7427477          0.6052876          0.7361881
#> 2          0.3499260          0.5926782          0.3748842          0.5828459
#> 3          0.1686009          0.4613791          0.1729925          0.4361836