Estimates the distance between multiple groups.

The geodesic distance (dG) and the Euclidean distance (dE) is used to assess group differences. Permutation is used to generate a reference distribtuion.

testMGD(
 .object             = args_default()$.object,
 .alpha              = args_default()$.alpha,
 .drop_inadmissibles = args_default()$.drop_inadmissibles,
 .R                  = args_default()$.R,
 .saturated          = args_default()$.saturated,
 .type_vcv           = args_default()$.type_vcv
 .verbose            = args_default()$.verbose
 )

# S3 method for cSEMResults_default
testMGD(.object = args_default()$.object,
  .alpha = args_default()$.alpha,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .R = args_default()$.R, .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv,
  .verbose = args_default()$.verbose)

# S3 method for cSEMResults_multi
testMGD(.object = args_default()$.object,
  .alpha = args_default()$.alpha,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .R = args_default()$.R, .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv,
  .verbose = args_default()$.verbose)

# S3 method for cSEMResults_2ndorder
testMGD(.object = args_default()$.object,
  .alpha = args_default()$.alpha,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .R = args_default()$.R, .saturated = args_default()$.saturated,
  .type_vcv = args_default()$.type_vcv,
  .verbose = args_default()$.verbose)

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.handle_inadmissibles
Character string. How should inadmissible results be treated? One of "drop", "ignore", or "replace". If "drop", all replications/resamples yielding an inadmissible result will be dropped (i.e. the number of results returned will be less than .R). For "ignore" all results are returned even if they are inadmissible (i.e. number of results returned = .R). For "replace" resampling continues until there are exactly .R admissible solutions. Defaults to "drop".

.R
Integer. The number of bootstrap replications. Defaults to 499.

.saturated
Logical. Should a saturated structural model be used? Defaults to FALSE.

.type_vcv
Character string. Indicates which model-implied correlation matrix is calcuted One of "indicator" or "construct". Defaults to "indicator".

.verbose
Logical. Should information be printed to the console? Defaults to TRUE.

Value

A standardized list of class cSEMTest. Technically cSEMTest is a named list containing the following list elements:

$Test_statistic

The value of test statistic(s).

$Critical_value

The critical value(s).

$Decision

The test decision. One of: Reject or Do not reject

$Number_admissibles

The number of admissible runs. See verify() for what constitues and inadmissible run.

Methods (by class)

cSEMResults_default: (TODO)

cSEMResults_multi: (TODO)

cSEMResults_2ndorder: (TODO)

See also

Examples

# NOT RUN {
require(cSEM)
data(satisfaction)

model <- "
# Structural model
QUAL ~ EXPE
EXPE ~ IMAG
SAT  ~ IMAG + EXPE + QUAL + VAL
LOY  ~ IMAG + SAT
VAL  ~ EXPE + QUAL

# Measurement model

EXPE <~ expe1 + expe2 + expe3 + expe4 + expe5
IMAG <~ imag1 + imag2 + imag3 + imag4 + imag5
LOY  =~ loy1  + loy2  + loy3  + loy4
QUAL =~ qual1 + qual2 + qual3 + qual4 + qual5
SAT  <~ sat1  + sat2  + sat3  + sat4
VAL  <~ val1  + val2  + val3  + val4
"

listData <- list(satisfaction[-3,], satisfaction[-5, ], satisfaction[-10, ])
out.cSEM <- csem(listData, model)

testMGD(.object = out.cSEM, .R = 20, .type_vcv= 'construct')
# }# NOT RUN {
# }