Test measurement invariance of composites

This functions performs the test for measurement invariance of composites proposed by Henseler et al. (2016).

testMICOM(
 .object               = args_default()$.object,
 .alpha                = args_default()$.alpha,
 .handle_inadmissibles = args_default()$.handle_inadmissibles,
 .R                    = args_default()$.R,
 .verbose              = args_default()$.verbose
 )

  # S3 method for cSEMResults_default
testMICOM(.object = args_default()$.object,
  .alpha = args_default()$.alpha,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .R = args_default()$.R, .verbose = args_default()$.verbose)

# S3 method for cSEMResults_multi
testMICOM(.object = args_default()$.object,
  .alpha = args_default()$.alpha,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .R = args_default()$.R, .verbose = args_default()$.verbose)

  # S3 method for cSEMResults_2ndorder
testMICOM(.object = args_default()$.object,
  .alpha = args_default()$.alpha,
  .handle_inadmissibles = args_default()$.handle_inadmissibles,
  .R = args_default()$.R, .verbose = args_default()$.verbose)

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

.alpha
An integer or a numeric vector of significance levels. Defaults to 0.05.

.handle_inadmissibles
Character string. How should inadmissible results be treated? One of "drop", "ignore", or "replace". If "drop", all replications/resamples yielding an inadmissible result will be dropped (i.e. the number of results returned will be less than .R). For "ignore" all results are returned even if they are inadmissible (i.e. number of results returned = .R). For "replace" resampling continues until there are exactly .R admissible solutions. Defaults to "drop".

.R
Integer. The number of bootstrap replications. Defaults to 499.

.verbose
Logical. Should information be printed to the console? Defaults to TRUE.

Value

An object of class "testMICOM"

An object of class "testMICOM" is a named list containing the following list element:

Step2

A list containing the results of the test for compositional invariance (Step 2).

Step3

A list containing the results of the test for mean and variance equality (Step 3).

Meta_information

A list of additional information on the test.

Details

The test is only meaningful for composite models.

If more than two groups are to be compared issues related to multiple testing should be taken into account.

The number of permutation runs defaults to args_default()$.R for performance reasons. According to Henseler et al. (2016) the number of permutations should be at least 5000 for assessment to be reliable.

Methods (by class)

cSEMResults_default: (TODO)

cSEMResults_multi: (TODO)

cSEMResults_2ndorder: (TODO)

Examples

# NOT RUN {
# TODO
# }