Verify admissibility

Verify admissibility of the estimated quantities for a given model.

verify(.object)

# S3 method for cSEMResults_default
verify(.object)

# S3 method for cSEMResults_multi
verify(.object)

# S3 method for cSEMResults_2ndorder
verify(.object)

Arguments

.object
An R object of class cSEMResults resulting from a call to csem().

Value

A logical vector indicating which (if any) problem occurred. A TRUE indicates that the problem has occured. For models containg second order constructs a list of two such vectors (one for the first and one for the second stage) is returned. Status codes are:

1: The algorithm has not converged.

2: At least one absolute standardized loading estimate is larger than 1,

3: The construct VCV is not positive semi-definite.

4: At least one proxy reliability is larger than 1.

5: The model-implied indicator VCV is not positive semi-definite. This is only checked for linear models.

Details

Results based on an estimated model exhibiting one of the following defects are deemed inadmissible: non-convergence, loadings and/or proxy reliabilities larger than 1, a construct VCV and/or a model-implied VCV matrix that is not positive (semi-)definite.

For models containing second order constructs estimation is by default done in a two/three stage procedure. In this case both the first and the second/third stage model quantities are checked.

Methods (by class)

cSEMResults_default: (TODO)

cSEMResults_multi: (TODO)

cSEMResults_2ndorder: (TODO)

See also